% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{adas}
\alias{adas}
\title{ADAS-Cognitive}
\format{
A data frame with 11584 rows and 211 variables
}
\usage{
data(adas)
}
\description{
Raw Alzheimer's Disease Assessment Scale-Cognitive (ADAS-Cog) data from the electronic case report form (eCRF). More information is available at \url{http://adni.loni.usc.edu/data-samples/data-faq/}.
}
\details{
A dataset with variables as follows:

\itemize{
  \item DONE N NA Was Assessment/Procedure Done?
  \item NDREASON N NA If No, Reason Not Done:
  \item DATE D YYYY-MM-DD Date Collected
  \item SOURCE N NA Information Source
  \item Q1TRIT N NA Trial 1 Total
  \item Q1TR2T N NA Trial 2 Total
  \item Q1TRT N NA Trial 3 Total
  \item Q5NAME1 N NA Flower
  \item Q5NAME2 N NA Bed
  \item Q5NAME3 N NA Whistle
  \item Q5NAME4 N NA Pencil
  \item Q5NAME5 N NA Rattle
  \item Q5NAME6 N NA Mask
  \item Q5NAME7 N NA Scissors
  \item Q5NAME8 N NA Comb
  \item Q5NAME9 N NA Wallet
  \item Q5NAME10 N NA Harmonica
  \item Q5NAME11 N NA Stethoscope
  \item Q5NAME12 N NA Tongs
  \item Q5FINGER N NA Naming Fingers
  \item Q5SCORE_CUE N  ADAS Score Component Q5 Naming Objects and Fingers (with cue when necessary) 
  \item TIMEEND T HHMM Time Ended:
  \item TIMEBEGAN T HHMM Time Began:
  \item TOTSCORE N NA Total Score (ADAS 11)
  \item WORDLIST N NA Word Recall and Recognition Tasks List Number
  \item Q1UNABLE N NA 1. Word Recall
  \item Q1TR1 T NA Trial 1
  \item Q1TR2 T NA Trial 2
  \item Q1TR3 T NA Trial 3
  \item Q1SCORE N NA Score Component
  \item Q2UNABLE N NA 2. Commands
  \item Q2TASK T NA <!--2. Commands-->
  \item Q2SCORE N NA Score Component
  \item Q3UNABLE N NA 3. Constructional Praxis
  \item Q3TASK1 N NA Circle
  \item Q3TASK2 N NA Two overlapping rectangles
  \item Q3TASK3 N NA Rhombus(Diamond)
  \item Q3TASK4 N NA Cube
  \item Q3SCORE N NA Score Component
  \item Q4UNABLE N NA 4.  Delayed Word Recall
  \item Q4TASK T NA <!--4.  Delayed Word Recall-->
  \item Q4SCORE N NA Score Component
  \item Q5UNABLE N NA 5. Naming Objects and Fingers
  \item Q5TASK T NA <!--5. Naming Objects and Fingers-->
  \item Q5SCORE N NA Score Component
  \item Q6UNABLE N NA 6. Ideational Praxis
  \item Q6TASK T NA <!--6. Ideational Praxis-->
  \item Q6SCORE N NA Score Component
  \item Q7UNABLE N NA 7. Orientation
  \item Q7TASK T NA <!--7. Orientation-->
  \item Q7SCORE N NA Score Component
  \item Q8UNABLE N NA 8. Word Recognition
  \item Q8WORD1 N NA Word 1
  \item Q8WORD1R T NA <!--Reminder given?-->
  \item Q8WORD2 N NA Word 2
  \item Q8WORD2R T NA <!--Reminder given?-->
  \item Q8WORD3 N NA Word 3
  \item Q8WORD3R T NA <!--Reminder given?-->
  \item Q8WORD4 N NA Word 4
  \item Q8WORD4R T NA <!--Reminder given?-->
  \item Q8WORD5 N NA Word 5
  \item Q8WORD5R T NA <!--Reminder given?-->
  \item Q8WORD6 N NA Word 6
  \item Q8WORD6R T NA <!--Reminder given?-->
  \item Q8WORD7 N NA Word 7
  \item Q8WORD7R T NA <!--Reminder given?-->
  \item Q8WORD8 N NA Word 8
  \item Q8WORD8R T NA <!--Reminder given?-->
  \item Q8WORD9 N NA Word 9
  \item Q8WORD9R T NA <!--Reminder given?-->
  \item Q8WORD10 N NA Word 10
  \item Q8WORD10R T NA <!--Reminder given?-->
  \item Q8WORD11 N NA Word 11
  \item Q8WORD11R T NA <!--Reminder given?-->
  \item Q8WORD12 N NA Word 12
  \item Q8WORD12R T NA <!--Reminder given?-->
  \item Q8WORD13 N NA Word 13
  \item Q8WORD13R T NA <!--Reminder given?-->
  \item Q8WORD14 N NA Word 14
  \item Q8WORD14R T NA <!--Reminder given?-->
  \item Q8WORD15 N NA Word 15
  \item Q8WORD15R T NA <!--Reminder given?-->
  \item Q8WORD16 N NA Word 16
  \item Q8WORD16R T NA <!--Reminder given?-->
  \item Q8WORD17 N NA Word 17
  \item Q8WORD17R T NA <!--Reminder given?-->
  \item Q8WORD18 N NA Word 18
  \item Q8WORD18R T NA <!--Reminder given?-->
  \item Q8WORD19 N NA Word 19
  \item Q8WORD19R T NA <!--Reminder given?-->
  \item Q8WORD20 N NA Word 20
  \item Q8WORD20R T NA <!--Reminder given?-->
  \item Q8WORD21 N NA Word 21
  \item Q8WORD21R T NA <!--Reminder given?-->
  \item Q8WORD22 N NA Word 22
  \item Q8WORD22R T NA <!--Reminder given?-->
  \item Q8WORD23 N NA Word 23
  \item Q8WORD23R T NA <!--Reminder given?-->
  \item Q8WORD24 N NA Word 24
  \item Q8WORD24R T NA <!--Reminder given?-->
  \item Q8SCORE N NA Score Component
  \item Q9TASK N NA 9. Remembering Test Instructions (this item is calculated based on responses recorded in Item 8.)
  \item Q9SCORE N NA Score Component
  \item Q10TASK N NA 10. Comprehension of Spoken Language
  \item Q10SCORE N NA Score Component
  \item Q11TASK N NA 11. Word-finding Difficulty
  \item Q11SCORE N NA Score Component
  \item Q12TASK N NA 12. Language
  \item Q12SCORE N NA Score Component
  \item Q13UNABLE N NA 13. Number Cancellation (Time limit: 45 seconds)
  \item Q13TASKA N NA 13.a Number of targets hit (range = 0-49)
  \item Q13TASKB N NA 13b. Number of errors
  \item Q13TASKC N NA 13c. Number of times reminded of task
  \item Q13SCORE N NA Score Component
  \item TOTAL13 N NA Total Score (ADAS 13)
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item COT1LIST T  Trial 1
  \item COT2LIST T  Trial 2
  \item COT3LIST T  Trial 3
  \item COT1SCOR N  Total <!--Trial 1-->
  \item COT2SCOR N  Total <!--Trial 2-->
  \item COT3SCOR N  Total <!--Trial 1-->
  \item COP1COMP N  If any trial not administered, check one:
  \item COCOMND T  Check each command performed correctly.
  \item COP2COMP N  If task not administered, check one:
  \item COCONSTR T  Check each figure drawn correctly.
  \item COP3COMP N  If task not administered, check one:
  \item COT4LIST T  Check each word correctly recalled.
  \item COT4TOTL N  Total
  \item COP4COMP N  If task not completed, check one:
  \item CONAME T  Check each object and finger named correctly.
  \item COP5COMP N  If task not administered, check one:
  \item COIDEA T  Check each step completed correctly.
  \item COP6COMP N  If task not administered, check one:
  \item COORIEN T  Check each item answered correctly.
  \item COP7COMP N  If task not administered, check one:
  \item CO8NURSE N  NURSE - Participant Response
  \item CO8MAGAZ N  MAGAZINE - Participant Response
  \item CO8WIZRD N  WIZARD - Participant Response
  \item CO8WIZRR T  Reminder given?
  \item CO8VAN N  VAN - Participant Response
  \item CO8VANR T  Reminder given?
  \item CO8LEPRD N  LEOPARD - Participant Response
  \item CO8LEPRR T  Reminder given?
  \item CO8SALE N  SALE - Participant Response
  \item CO8SALER T  Reminder given?
  \item CO8SEA N  SEA - Participant Response
  \item CO8SEAR T  Reminder given?
  \item CO8TRAIN N  TRAIN - Participant Response
  \item CO8TRAIR T  Reminder given?
  \item CO8COIN N  COIN - Participant Response
  \item CO8COINR T  Reminder given?
  \item CO8SHIP N  SHIP - Participant Response
  \item CO8SHIPR T  Reminder given?
  \item CO8INST N  INSTITUTION - Participant Response
  \item CO8INSTR T  Reminder given?
  \item CO8MAP N  MAP - Participant Response
  \item CO8MAPR T  Reminder given?
  \item CO8AXE N  AXE - Participant Response
  \item CO8AXER T  Reminder given?
  \item CO8BOARD N  BOARD - Participant Response
  \item CO8BOARR T  Reminder given?
  \item CO8CARRT N  CARROT - Participant Response
  \item CO8CARRR T  Reminder given?
  \item CO8MILK N  MILK - Participant Response
  \item CO8MILKR T  Reminder given?
  \item CO8VOL N  VOLUME - Participant Response
  \item CO8VOLR T  Reminder given?
  \item CO8FORST N  FOREST - Participant Response
  \item CO8FORSR T  Reminder given?
  \item CO8ANCHR N  ANCHOR - Participant Response
  \item CO8ANCRR T  Reminder given?
  \item CO8GEM N  GEM - Participant Response
  \item CO8GEMR T  Reminder given?
  \item CO8CAT N  CAT - Participant Response
  \item CO8CATR T  Reminder given?
  \item CO8FUND N  FUND - Participant Response
  \item CO8FUNDR T  Reminder given?
  \item CO8EDGE N  EDGE - Participant Response
  \item CO8EDGER T  Reminder given?
  \item CO8CAKE N  CAKE - Participant Response
  \item CO8CAKER T  Reminder given?
  \item COP8COMP N  If task not administered, check one:
  \item COINSTRC N  Check level of impairment.  Base your answer only on observations made during Word Recognition task (Question #8).
  \item COCOMPRE N  Check level of impairment.
  \item COWRDFND N  Check one response
  \item COLANG N  Check level of impairment.
  \item CONMCXLA N  14a. Number of targets hit (range = 0-49)
  \item CONMCXLB N  14b. Number of errors
  \item CONMCXLC N  14c. Number of times reminded of task
  \item COP14CMP N  If task not administered, check one:
  \item USERDATE2 S  Date record last updated
}
}
\references{
Rosen WG, Mohs RC, & Davis KL. (1984). A new rating scale for Alzheimer's disease. \emph{The American Journal of Psychiatry}. 141 (11), 1356-64. \url{http://www.ncbi.nlm.nih.gov/pubmed/6496779}
}
\seealso{
\code{\link{adasscores}}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
