% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{addcomm}
\alias{addcomm}
\title{Additional Comments}
\format{
A data frame with 6229 rows and 10 variables
}
\usage{
data(addcomm)
}
\description{
See \url{http://adni.loni.usc.edu}.
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item RECNO N  Record Number
  \item EXAMDATE D  Examination Date
  \item CAAPPLY N  This Additional Comment applies to:
  \item USERDATE2 S  Date record last updated
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
