% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{admcba}
\alias{admcba}
\title{AD Metabolomics Consortium Bile Acids}
\format{
A data frame with 866 rows and 53 variables
}
\usage{
data(admcba)
}
\description{
AD Metabolomics Consortium Bile Acids
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item Sample.Type -4 -4 Study samples are Sample.  Biocrates QCs are QC Level 1, 2 or 3.
  \item Species -4 -4 Species of sample
  \item Material -4 -4 For this study the samples are serum
  \item Well.Position -4 -4 Well number in the 96-well plate from which the sample was injected
  \item Sample.Volume -4 -4 The volume of sample injected in microliters
  \item Run.Number -4 -4 Generated by the Biocrates MetIDQ software for every plate indicating the run number for a particular plate.  For this study, all plates were injected only once.  Therefore, all Run Numbers are 1.
  \item Injection.Number -4 -4 Generated by the Biocrates MetIDQ software for every sample on a plate indicating whether a sample was injected more than once.  All study samples were injected only once for every plate.  Therefore, all Injection Numbers for study samples is 1.  QC samples are injected three times for every plate resulting in a 1, 2, or 3 listing.
  \item CA -4 -4 Cholic acid [M]
  \item CA.Status -4 -4 Status of Cholic acid result
  \item CDCA -4 -4 Chenodeoxycholic acid [M]
  \item CDCA.Status -4 -4 Status of Chenodeoxycholic acid result
  \item DCA -4 -4 Deoxycholic acid [M]
  \item DCA.Status -4 -4 Status of Deoxycholic acid result
  \item GCA -4 -4 Glycocholic acid [M]
  \item GCA.Status -4 -4 Status of Glycocholic acid result
  \item GCDCA -4 -4 Glycochenodeoxycholic acid [M]
  \item GCDCA.Status -4 -4 Status of Glycochenodeoxycholic acid result
  \item GDCA -4 -4 Glycodeoxycholic acid [M]
  \item GDCA.Status -4 -4 Status of Glycodeoxycholic acid
  \item GLCA -4 -4 Glycolithocholic acid [M]
  \item GLCA.Status -4 -4 Status of Glycolithocholic acid result
  \item GLCAS -4 -4 Glycolithocholic acid sulphate [M]
  \item GLCAS.Status -4 -4 Status of Glycolithocholic acid sulphate
  \item GUDCA -4 -4 Glycoursodeoxycholic acid [M]
  \item GUDCA.Status -4 -4 Status of Glycoursodeoxycholic acid result
  \item HDCA -4 -4 Hyodeoxycholic acid [M]
  \item HDCA.Status -4 -4 Status of Hyodeoxycholic acid result
  \item LCA -4 -4 Lithocholic acid [M]
  \item LCA.Status -4 -4 Status of Lithocholic acid result
  \item MCA.a. -4 -4 -4
  \item MCA.a.Status -4 -4 -4
  \item MCA.b. -4 -4 -4
  \item MCA.b.Status -4 -4 -4
  \item MCA.o. -4 -4 -4
  \item MCA.o.Status -4 -4 -4
  \item TCA -4 -4 Taurocholic acid [M]
  \item TCA.Status -4 -4 Status of Taurocholic acid result
  \item TCDCA -4 -4 Taurochenodeoxycholic acid [M]
  \item TCDCA.Status -4 -4 Status of Taurochenodeoxycholic acid result
  \item TDCA -4 -4 Taurodeoxycholic acid [M]
  \item TDCA.Status -4 -4 Status of Taurodeoxycholic acid result
  \item TLCA -4 -4 Taurolithocholic acid [M]
  \item TLCA.Status -4 -4 Status of Taurolithocholic acid result
  \item TLCAS -4 -4 Taurolithocholic acid sulphate [M]
  \item TLCAS.Status -4 -4 Status of Taurolithocholic acid sulphate result
  \item TMCA.a.b. -4 -4 -4
  \item TMCA.a.b.Status -4 -4 -4
  \item TUDCA -4 -4 Tauroursodeoxycholic acid [M]
  \item TUDCA.Status -4 -4 Status ofTauroursodeoxycholic acid result
  \item UDCA -4 -4 Ursodeoxycholic acid [M]
  \item UDCA.Status -4 -4 Status of Ursodeoxycholic acid result
}
}
\examples{
\dontrun{
describe(admcba)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/ADMCBA/ADMC\%20Bile_Acids_Method_Description.pdf}
}
\keyword{datasets}
