% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{admcbapp}
\alias{admcbapp}
\title{AD Metabolomics Consortium Bile Acids - Post-processed}
\format{
A data frame with 1671 rows and 58 variables
}
\usage{
data(admcbapp)
}
\description{
AD Metabolomics Consortium Bile Acids - Post-processed
}
\details{
A dataset with variables as follows:

\itemize{
  \item ORIGPROT -4 -4 Original study protocol
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item PLATFORM_ID -4 -4 BILEACID: Biocrates Bile Acid Kit
  \item SAMPLE_ID -4 -4 -4
  \item PLATE_ID -4 -4 Plate bar code number from which the sample was injected
  \item SAMPLE_BAR_CODE -4 -4 -4
  \item SPECIES -4 -4 Species of sample
  \item MATERIAL -4 -4 For this study the samples are serum
  \item WELL_POSITION -4 -4 Well number in the 96-well plate from which the sample was injected
  \item SAMPLE_VOLUME -4 -4 The volume of sample injected in microliters
  \item SUBJECT_FLAG -4 -4 A flag shows whether a sample were excluded during preprocessing steps; 0: sample were not excluded. 1: sample were excluded
  \item CA -4 -4 Cholic acid [M]
  \item CDCA -4 -4 Chenodeoxycholic acid [M]
  \item DCA -4 -4 Deoxycholic acid [M]
  \item GCA -4 -4 Glycocholic acid [M]
  \item GCDCA -4 -4 Glycochenodeoxycholic acid [M]
  \item GDCA -4 -4 Glycodeoxycholic acid [M]
  \item GLCA -4 -4 Glycolithocholic acid [M]
  \item GUDCA -4 -4 Glycoursodeoxycholic acid [M]
  \item TCA -4 -4 Taurocholic acid [M]
  \item TCDCA -4 -4 Taurochenodeoxycholic acid [M]
  \item TDCA -4 -4 Taurodeoxycholic acid [M]
  \item TLCA -4 -4 Taurolithocholic acid [M]
  \item TMCA_A_B -4 -4 -4
  \item TUDCA -4 -4 Tauroursodeoxycholic acid [M]
  \item UDCA -4 -4 Ursodeoxycholic acid [M]
  \item CA.CDCA -4 -4 CA:CDCA ratio
  \item DCA.CA -4 -4 DCA:CA ratio [unitless]
  \item GLCA.CDCA -4 -4 GLCA:CDCA ratio  [unitless]
  \item GDCA.CA -4 -4 GDCA:CA ratio  [unitless]
  \item GDCA.DCA -4 -4 GDCA:DCA ratio  [unitless]
  \item TDCA.CA -4 -4 TDCA:CA ratio  [unitless]
  \item TLCA.CDCA -4 -4 TLCA:CDCA ratio  [unitless]
  \item TDCA.DCA -4 -4 TDCA:DCA ratio  [unitless]
  \item CA_LOGTRANSFORMFLAG -4 -4 Flag showing whether Cholic acid levels were log2 transformed
  \item CDCA_LOGTRANSFORMFLAG -4 -4 Flag showing whether Chenodeoxycholic acid levels were log2 transformed
  \item DCA_LOGTRANSFORMFLAG -4 -4 Flag showing whether Deoxycholic acid levels were log2 transformed
  \item GCA_LOGTRANSFORMFLAG -4 -4 Flag showing whether Glycocholic acid levels were log2 transformed
  \item GCDCA_LOGTRANSFORMFLAG -4 -4 Flag showing whether Glycochenodeoxycholic acid levels were log2 transformed
  \item GDCA_LOGTRANSFORMFLAG -4 -4 Flag showing whether Glycodeoxycholic acid levels were log2 transformed
  \item GLCA_LOGTRANSFORMFLAG -4 -4 Flag showing whether Glycolithocholic acid levels were log2 transformed
  \item GUDCA_LOGTRANSFORMFLAG -4 -4 Flag showing whether Glycoursodeoxycholic acid levels were log2 transformed
  \item TCA_LOGTRANSFORMFLAG -4 -4 Flag showing whether Taurocholic acid levels were log2 transformed
  \item TCDCA_LOGTRANSFORMFLAG -4 -4 Flag showing whether Taurochenodeoxycholic acid levels were log2 transformed
  \item TDCA_LOGTRANSFORMFLAG -4 -4 Flag showing whether Taurodeoxycholic acid levels were log2 transformed
  \item TLCA_LOGTRANSFORMFLAG -4 -4 Flag showing whether Taurolithocholic acid levels were log2 transformed
  \item TMCA_A_B_LOGTRANSFORMFLAG -4 -4 -4
  \item TUDCA_LOGTRANSFORMFLAG -4 -4 Flag showing whether Tauroursodeoxycholic acid levels were log2 transformed
  \item UDCA_LOGTRANSFORMFLAG -4 -4 Flag showing whether Ursodeoxycholic acid levels were log2 transformed
  \item CA.CDCA_LOGTRANSFORMFLAG -4 -4 Flag showing whether CA:CDCA ratio values were log2 transformed
  \item DCA.CA_LOGTRANSFORMFLAG -4 -4 Flag showing whether DCA:CA ratio values were log2 transformed
  \item GLCA.CDCA_LOGTRANSFORMFLAG -4 -4 Flag showing whether GLCA:CDCA ratio values were log2 transformed
  \item GDCA.CA_LOGTRANSFORMFLAG -4 -4 Flag showing whether GDCA:CA ratio values were log2 transformed
  \item GDCA.DCA_LOGTRANSFORMFLAG -4 -4 Flag showing whether GDCA:DCA ratio values were log2 transformed
  \item TDCA.CA_LOGTRANSFORMFLAG -4 -4 Flag showing whether TDCA:CA ratio values were log2 transformed
  \item TLCA.CDCA_LOGTRANSFORMFLAG -4 -4 Flag showing whether TLCA:CDCA ratio values were log2 transformed
  \item TDCA.DCA_LOGTRANSFORMFLAG -4 -4 Flag showing whether TDCA:DCA ratio values were log2 transformed
}
}
\examples{
\dontrun{
describe(admcbapp)
}
}
\keyword{datasets}
