% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{admcbarcelonapurine}
\alias{admcbarcelonapurine}
\title{AD Metabolomics Consortium Barcelona Purines}
\format{
A data frame with 890 rows and 12 variables
}
\usage{
data(admcbarcelonapurine)
}
\description{
AD Metabolomics Consortium Barcelona Purines
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item RECNO -4 -4 Record number (2 indicates replicate sample)
  \item Comments -4 -4 -4
  \item L.Glutamine -4 -4 L-Glutamine
  \item Taurine -4 -4 Taurine
  \item Glycine -4 -4 Glycine
  \item Inosine -4 -4 Inosine
  \item Hypoxanthine -4 -4 Hypoxanthine
  \item Xanthine -4 -4 Xanthine
  \item Xanthosine -4 -4 Xanthosine
  \item Uric.acid -4 -4 Uric acid
}
}
\examples{
\dontrun{
describe(admcbarcelonapurine)
}
}
\keyword{datasets}
