% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{admcdukep180fiaadni2go}
\alias{admcdukep180fiaadni2go}
\title{ADMC Duke p180 Flow Injection Analysis [ADNI GO,2]}
\format{
A data frame with 941 rows and 167 variables
}
\usage{
data(admcdukep180fiaadni2go)
}
\description{
ADMC Duke p180 Flow Injection Analysis [ADNI GO,2]
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item Customer.Sample.Identification -4 -4 -4
  \item OTHERSPECID -4 -4 -4
  \item SITE -4 -4 -4
  \item PRIM -4 -4 -4
  \item DER -4 -4 -4
  \item ADDV -4 -4 -4
  \item GRP -4 -4 -4
  \item PROT -4 -4 -4
  \item VOL -4 -4 -4
  \item DRAWDTE -4 -4 -4
  \item DRAWTIME -4 -4 -4
  \item RECDTE -4 -4 -4
  \item CONDITION -4 -4 -4
  \item Plate.Bar.Code -4 -4 Unique identifier for every plate generated by the Biocrates MetIDQ software
  \item Sample.Bar.Code -4 -4 Unique identifier for every sample registered in the Biocrates MetIDQ software
  \item Sample.Type -4 -4 -4
  \item Sample.Identification -4 -4 Unique identifier for every sample entered in the Duke Proteomics Sample Sumbission System.
  \item Species -4 -4 Species of sample origin
  \item Material -4 -4 For this study the samples are serum
  \item Well.Position -4 -4 Well number in the 96-well plate from which the sample was injected
  \item Sample.Volume -4 -4 The volume of sample injected in microliters
  \item Run.Number -4 -4 -4
  \item Injection.Number -4 -4 -4
  \item C0 -4 -4 Carnitine
  \item C10 -4 -4 Decanoylcarnitine
  \item C10.1 -4 -4 Decenoylcarnitine
  \item C10.2 -4 -4 Decadienylcarnitine
  \item C12 -4 -4 Dodecanoylcarnitine
  \item C12.DC -4 -4 Dodecanedioylcarnitine
  \item C12.1 -4 -4 Dodecenoylcarnitine
  \item C14 -4 -4 Tetradecanoylcarnitine
  \item C14.1 -4 -4 Tetradecenoylcarnitine
  \item C14.1.OH -4 -4 Hydroxytetradecenoylcarnitine
  \item C14.2 -4 -4 Tetradecadienylcarnitine
  \item C14.2.OH -4 -4 Hydroxytetradecadienylcarnitine
  \item C16 -4 -4 Hexadecanoylcarnitine
  \item C16.OH -4 -4 Hydroxyhexadecanoylcarnitine
  \item C16.1 -4 -4 Hexadecenoylcarnitine
  \item C16.1.OH -4 -4 Hydroxyhexadecenoylcarnitine
  \item C16.2 -4 -4 Hexadecadienylcarnitine
  \item C16.2.OH -4 -4 Hydroxyhexadecadienylcarnitine
  \item C18 -4 -4 Octadecanoylcarnitine
  \item C18.1 -4 -4 Octadecenoylcarnitine
  \item C18.1.OH -4 -4 Hydroxyoctadecenoylcarnitine
  \item C18.2 -4 -4 Octadecadienylcarnitine
  \item C2 -4 -4 Acetylcarnitine
  \item C3 -4 -4 Propionylcarnitine
  \item C5.OH..C3.DC.M. -4 -4 Hydroxyvalerylcarnitine (Methylmalonylcarnitine)
  \item C3.OH -4 -4 Hydroxypropionylcarnitine
  \item C3.1 -4 -4 Propenylcarnitine
  \item C4 -4 -4 Butyrylcarnitine (Isobutyrylcarnitine)
  \item C3.DC..C4.OH. -4 -4 Malonylcarnitine (Hydroxybutyrylcarnitine)
  \item C4.1 -4 -4 Butenylcarnitine
  \item C5 -4 -4 Isovalerylcarnitine (Methylbutylcarnitine, Valerylcarnitine)
  \item C5.DC..C6.OH. -4 -4 Glutarylcarnitine (Hydroxyhexanoylcarnitine)
  \item C5.M.DC -4 -4 Methylglutarylcarnitine
  \item C5.1 -4 -4 Tiglylcarnitine
  \item C5.1.DC -4 -4 Glutaconylcarnitine
  \item C6..C4.1.DC. -4 -4 Hexanoylcarnitine (Fumarylcarnitine)
  \item C6.1 -4 -4 Hexenoylcarnitine
  \item C7.DC -4 -4 Pimelylcarnitine
  \item C8 -4 -4 Octanoylcarnitine
  \item C9 -4 -4 Nonanoylcarnitine
  \item lysoPC.a.C14.0 -4 -4 lysoPhosphatidylcholine a C14:0
  \item lysoPC.a.C16.0 -4 -4 lysoPhosphatidylcholine a C16:0
  \item lysoPC.a.C16.1 -4 -4 lysoPhosphatidylcholine a C16:1
  \item lysoPC.a.C17.0 -4 -4 lysoPhosphatidylcholine a C17:0
  \item lysoPC.a.C18.0 -4 -4 lysoPhosphatidylcholine a C18:0
  \item lysoPC.a.C18.1 -4 -4 lysoPhosphatidylcholine a C18:1
  \item lysoPC.a.C18.2 -4 -4 lysoPhosphatidylcholine a C18:2
  \item lysoPC.a.C20.3 -4 -4 lysoPhosphatidylcholine a C20:3
  \item lysoPC.a.C20.4 -4 -4 lysoPhosphatidylcholine a C20:4
  \item lysoPC.a.C24.0 -4 -4 lysoPhosphatidylcholine a C24:0
  \item lysoPC.a.C26.0 -4 -4 lysoPhosphatidylcholine a C26:0
  \item lysoPC.a.C26.1 -4 -4 lysoPhosphatidylcholine a C26:1
  \item lysoPC.a.C28.0 -4 -4 lysoPhosphatidylcholine a C28:0
  \item lysoPC.a.C28.1 -4 -4 lysoPhosphatidylcholine a C28:1
  \item PC.aa.C24.0 -4 -4 PC aa C24:0
  \item PC.aa.C26.0 -4 -4 PC aa C26:0
  \item PC.aa.C28.1 -4 -4 PC aa C28:1
  \item PC.aa.C30.0 -4 -4 PC aa C30:0
  \item PC.aa.C32.0 -4 -4 PC aa C32:0
  \item PC.aa.C32.1 -4 -4 PC aa C32:1
  \item PC.aa.C32.3 -4 -4 PC aa C32:3
  \item PC.aa.C34.1 -4 -4 PC aa C34:1
  \item PC.aa.C34.2 -4 -4 PC aa C34:2
  \item PC.aa.C34.3 -4 -4 PC aa C34:3
  \item PC.aa.C34.4 -4 -4 PC aa C34:4
  \item PC.aa.C36.0 -4 -4 PC aa C36:0
  \item PC.aa.C36.1 -4 -4 PC aa C36:1
  \item PC.aa.C36.2 -4 -4 PC aa C36:2
  \item PC.aa.C36.3 -4 -4 PC aa C36:3
  \item PC.aa.C36.4 -4 -4 PC aa C36:4
  \item PC.aa.C36.5 -4 -4 PC aa C36:5
  \item PC.aa.C36.6 -4 -4 PC aa C36:6
  \item PC.aa.C38.0 -4 -4 PC aa C38:0
  \item PC.aa.C38.3 -4 -4 PC aa C38:3
  \item PC.aa.C38.4 -4 -4 PC aa C38:4
  \item PC.aa.C38.5 -4 -4 PC aa C38:5
  \item PC.aa.C38.6 -4 -4 PC aa C38:6
  \item PC.aa.C40.1 -4 -4 PC aa C40:1
  \item PC.aa.C40.2 -4 -4 PC aa C40:2
  \item PC.aa.C40.3 -4 -4 PC aa C40:3
  \item PC.aa.C40.4 -4 -4 PC aa C40:4
  \item PC.aa.C40.5 -4 -4 PC aa C40:5
  \item PC.aa.C40.6 -4 -4 PC aa C40:6
  \item PC.aa.C42.0 -4 -4 PC aa C42:0
  \item PC.aa.C42.1 -4 -4 PC aa C42:1
  \item PC.aa.C42.2 -4 -4 PC aa C42:2
  \item PC.aa.C42.4 -4 -4 PC aa C42:4
  \item PC.aa.C42.5 -4 -4 PC aa C42:5
  \item PC.aa.C42.6 -4 -4 PC aa C42:6
  \item PC.ae.C30.0 -4 -4 PC ae C30:0
  \item PC.ae.C30.1 -4 -4 PC ae C30:1
  \item PC.ae.C30.2 -4 -4 PC ae C30:2
  \item PC.ae.C32.1 -4 -4 PC ae C32:1
  \item PC.ae.C32.2 -4 -4 PC ae C32:2
  \item PC.ae.C34.0 -4 -4 PC ae C34:0
  \item PC.ae.C34.1 -4 -4 PC ae C34:1
  \item PC.ae.C34.2 -4 -4 PC ae C34:2
  \item PC.ae.C34.3 -4 -4 PC ae C34:3
  \item PC.ae.C36.0 -4 -4 PC ae C36:0
  \item PC.ae.C36.1 -4 -4 PC ae C36:1
  \item PC.ae.C36.2 -4 -4 PC ae C36:2
  \item PC.ae.C36.3 -4 -4 PC ae C36:3
  \item PC.ae.C36.4 -4 -4 PC ae C36:4
  \item PC.ae.C36.5 -4 -4 PC ae C36:5
  \item PC.ae.C38.0 -4 -4 PC ae C38:0
  \item PC.ae.C38.1 -4 -4 PC ae C38:1
  \item PC.ae.C38.2 -4 -4 PC ae C38:2
  \item PC.ae.C38.3 -4 -4 PC ae C38:3
  \item PC.ae.C38.4 -4 -4 PC ae C38:4
  \item PC.ae.C38.5 -4 -4 PC ae C38:5
  \item PC.ae.C38.6 -4 -4 PC ae C38:6
  \item PC.ae.C40.1 -4 -4 PC ae C40:1
  \item PC.ae.C40.2 -4 -4 PC ae C40:2
  \item PC.ae.C40.3 -4 -4 PC ae C40:3
  \item PC.ae.C40.4 -4 -4 PC ae C40:4
  \item PC.ae.C40.5 -4 -4 PC ae C40:5
  \item PC.ae.C40.6 -4 -4 PC ae C40:6
  \item PC.ae.C42.0 -4 -4 PC ae C42:0
  \item PC.ae.C42.1 -4 -4 PC ae C42:1
  \item PC.ae.C42.2 -4 -4 PC ae C42:2
  \item PC.ae.C42.3 -4 -4 PC ae C42:3
  \item PC.ae.C42.4 -4 -4 PC ae C42:4
  \item PC.ae.C42.5 -4 -4 PC ae C42:5
  \item PC.ae.C44.3 -4 -4 PC ae C44:3
  \item PC.ae.C44.4 -4 -4 PC ae C44:4
  \item PC.ae.C44.5 -4 -4 PC ae C44:5
  \item PC.ae.C44.6 -4 -4 PC ae C44:6
  \item SM..OH..C14.1 -4 -4 SM (OH) C14:1
  \item SM..OH..C16.1 -4 -4 SM (OH) C16:1
  \item SM..OH..C22.1 -4 -4 SM (OH) C22:1
  \item SM..OH..C22.2 -4 -4 SM (OH) C22:2
  \item SM..OH..C24.1 -4 -4 SM (OH) C24:1
  \item SM.C16.0 -4 -4 SM C16:0
  \item SM.C16.1 -4 -4 SM C16:1
  \item SM.C18.0 -4 -4 SM C18:0
  \item SM.C18.1 -4 -4 SM C18:1
  \item SM.C20.2 -4 -4 SM C20:2
  \item SM.C24.0 -4 -4 SM C24:0
  \item SM.C24.1 -4 -4 SM C24:1
  \item SM.C26.0 -4 -4 SM C26:0
  \item SM.C26.1 -4 -4 SM C26:1
}
}
\examples{
\dontrun{
describe(admcdukep180fiaadni2go)
}
}
\keyword{datasets}
