% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{admcdukep180uplcadni2go}
\alias{admcdukep180uplcadni2go}
\title{ADMC Duke p180 Ultra Performance Liquid Chromatography [ADNI GO,2]}
\format{
A data frame with 941 rows and 67 variables
}
\usage{
data(admcdukep180uplcadni2go)
}
\description{
ADMC Duke p180 Ultra Performance Liquid Chromatography [ADNI GO,2]
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item Customer.Sample.Identification -4 -4 -4
  \item OTHERSPECID -4 -4 -4
  \item SITE -4 -4 -4
  \item PRIM -4 -4 -4
  \item DER -4 -4 -4
  \item ADDV -4 -4 -4
  \item GRP -4 -4 -4
  \item PROT -4 -4 -4
  \item VOL -4 -4 -4
  \item DRAWDTE -4 -4 -4
  \item DRAWTIME -4 -4 -4
  \item RECDTE -4 -4 -4
  \item CONDITION -4 -4 -4
  \item Plate.Bar.Code -4 -4 Unique identifier for every plate generated by the Biocrates MetIDQ software
  \item Sample.Bar.Code -4 -4 Unique identifier for every sample registered in the Biocrates MetIDQ software
  \item Sample.Type -4 -4 -4
  \item Sample.Identification -4 -4 Unique identifier for every sample entered in the Duke Proteomics Sample Sumbission System.
  \item Species -4 -4 Species of sample origin
  \item Material -4 -4 For this study the samples are serum
  \item Well.Position -4 -4 Well number in the 96-well plate from which the sample was injected
  \item Sample.Volume -4 -4 The volume of sample injected in microliters
  \item Run.Number -4 -4 -4
  \item Injection.Number -4 -4 -4
  \item Ala -4 -4 Alanine
  \item Arg -4 -4 Arginine
  \item Asn -4 -4 Asparagine
  \item Asp -4 -4 Aspartic Acid
  \item Cit -4 -4 Citrulline
  \item Gln -4 -4 Glutamine
  \item Glu -4 -4 Glutamic Acid
  \item Gly -4 -4 Glycine
  \item His -4 -4 Histidine
  \item Ile -4 -4 Isoleucine
  \item Lys -4 -4 Lysine
  \item Met -4 -4 Methionine
  \item Orn -4 -4 Ornithine
  \item Phe -4 -4 Phenylalanine
  \item Pro -4 -4 Proline
  \item Ser -4 -4 Serine
  \item Thr -4 -4 Threonine
  \item Trp -4 -4 Tryptophan
  \item Tyr -4 -4 Tyrosine
  \item Val -4 -4 Valine
  \item Ac.Orn -4 -4 Acetylornithine
  \item ADMA -4 -4 Asymmetric dimethylarginine
  \item alpha.AAA -4 -4 alpha-Aminoadipic acid
  \item c4.OH.Pro -4 -4 c4-OH-proline 
  \item Carnosine -4 -4 Carnosine
  \item Creatinine -4 -4 Creatinine
  \item DOPA -4 -4 Dihydroxyphenylalanine
  \item Dopamine -4 -4 Dopamine
  \item Histamine -4 -4 Histamine
  \item Kynurenine -4 -4 Kynurenine
  \item Met.SO -4 -4 Methionine-Sulfoxide
  \item Nitro.Tyr -4 -4 Nitrotyrosine
  \item PEA -4 -4 Phenylethylamine
  \item Putrescine -4 -4 Putrescine
  \item Sarcosine -4 -4 Sarcosine
  \item Serotonin -4 -4 Serotonin
  \item Spermidine -4 -4 Spermidine
  \item Spermine -4 -4 Spermine
  \item t4.OH.Pro -4 -4 trans isomer of 4-hydroxyproline 
  \item Taurine -4 -4 Taurine
  \item SDMA -4 -4 Symmetric dimethylarginine
}
}
\examples{
\dontrun{
describe(admcdukep180uplcadni2go)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/ADMCDUKEP180UPLCADNI2GO/}
}
\keyword{datasets}
