% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{admcgctof}
\alias{admcgctof}
\title{ADMC Metabolomic Analysis by Gas chromatography time of flight mass spectrometry (GCTOF) Data [ADNI1]}
\format{
A data frame with 916 rows and 401 variables
}
\usage{
data(admcgctof)
}
\description{
ADMC Metabolomic Analysis by Gas chromatography time of flight mass spectrometry (GCTOF) Data [ADNI1]
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item Z.C30.FAME.INTERNAL.STANDARD -4 -4 -4
  \item Z.C28.FAME.INTERNAL.STANDARD -4 -4 -4
  \item Z.C26.FAME.INTERNAL.STANDARD -4 -4 -4
  \item Z.C24.FAME.INTERNAL.STANDARD -4 -4 -4
  \item Z.C22.FAME.INTERNAL.STANDARD -4 -4 -4
  \item Z.C20.FAME.INTERNAL.STANDARD -4 -4 -4
  \item Z.C18.FAME.INTERNAL.STANDARD -4 -4 -4
  \item Z.C16.FAME.INTERNAL.STANDARD -4 -4 -4
  \item Z.C14.FAME.INTERNAL.STANDARD -4 -4 -4
  \item Z.C12.FAME.INTERNAL.STANDARD -4 -4 -4
  \item Z.C10.FAME.INTERNAL.STANDARD -4 -4 -4
  \item Z.C09.FAME.INTERNAL.STANDARD -4 -4 -4
  \item Z.C08.FAME.INTERNAL.STANDARD -4 -4 -4
  \item XYLULOSE.NIST -4 -4 -4
  \item XYLOSE -4 -4 -4
  \item XYLITOL -4 -4 -4
  \item XANTHINE -4 -4 -4
  \item VALINE -4 -4 -4
  \item URIDINE -4 -4 -4
  \item URIC.ACID -4 -4 -4
  \item UREA -4 -4 -4
  \item UDP.GLUCURONIC.ACID -4 -4 -4
  \item TYROSINE -4 -4 -4
  \item TRYPTOPHAN -4 -4 -4
  \item TRANS.4.HYDROXYPROLINE -4 -4 -4
  \item TOCOPHEROL.GAMMA. -4 -4 -4
  \item TOCOPHEROL.ALPHA. -4 -4 -4
  \item THREONINE -4 -4 -4
  \item THREONIC.ACID -4 -4 -4
  \item THREITOL -4 -4 -4
  \item TAURINE -4 -4 -4
  \item TARTARIC.ACID -4 -4 -4
  \item TAGATOSE -4 -4 -4
  \item SUCROSE -4 -4 -4
  \item SUCCINIC.ACID -4 -4 -4
  \item STEARIC.ACID -4 -4 -4
  \item SHIKIMIC.ACID -4 -4 -4
  \item SERINE -4 -4 -4
  \item SALICYLIC.ACID -4 -4 -4
  \item RIBOSE -4 -4 -4
  \item RIBONIC.ACID -4 -4 -4
  \item RIBITOL -4 -4 -4
  \item QUINIC.ACID -4 -4 -4
  \item PYRUVIC.ACID -4 -4 -4
  \item PYROPHOSPHATE -4 -4 -4
  \item PSEUDO.URIDINE -4 -4 -4
  \item PROLINE -4 -4 -4
  \item PIPECOLINIC.ACID -4 -4 -4
  \item PHOSPHATE -4 -4 -4
  \item PHENYLETHYLAMINE -4 -4 -4
  \item PHENYLALANINE -4 -4 -4
  \item PELARGONIC.ACID -4 -4 -4
  \item PARABANIC.ACID.NIST -4 -4 -4
  \item PALMITOLEIC.ACID -4 -4 -4
  \item PALMITIC.ACID -4 -4 -4
  \item OXOPROLINE -4 -4 -4
  \item ORNITHINE -4 -4 -4
  \item OLEIC.ACID -4 -4 -4
  \item OCTADECANOL -4 -4 -4
  \item NAPROXEN -4 -4 -4
  \item N.METHYLALANINE -4 -4 -4
  \item N.ACETYLMANNOSAMINE -4 -4 -4
  \item N.ACETYLGLUTAMATE -4 -4 -4
  \item MYRISTIC.ACID -4 -4 -4
  \item MYO.INOSITOL -4 -4 -4
  \item METHIONINE.SULFOXIDE -4 -4 -4
  \item METHANOLPHOSPHATE -4 -4 -4
  \item MANNOSE -4 -4 -4
  \item MALTOSE -4 -4 -4
  \item MALIC.ACID -4 -4 -4
  \item MALEIMIDE -4 -4 -4
  \item LYXITOL -4 -4 -4
  \item LYSINE -4 -4 -4
  \item LINOLEIC.ACID -4 -4 -4
  \item LEVOGLUCOSAN -4 -4 -4
  \item LEUCINE -4 -4 -4
  \item LAURIC.ACID -4 -4 -4
  \item LACTIC.ACID -4 -4 -4
  \item ISOTHREONIC.ACID -4 -4 -4
  \item ISOLEUCINE -4 -4 -4
  \item ISOHEXONIC.ACID -4 -4 -4
  \item ISODEOXYTETRONIC.ACID.NIST -4 -4 -4
  \item ISOCITRIC.ACID -4 -4 -4
  \item INOSINE -4 -4 -4
  \item INDOLE.3.LACTATE -4 -4 -4
  \item INDOLE.3.ACETATE -4 -4 -4
  \item IBUPROFEN -4 -4 -4
  \item HYPOXANTHINE -4 -4 -4
  \item HYDROXYPROLINE.DIPEPTIDE.NIST -4 -4 -4
  \item HYDROXYLAMINE -4 -4 -4
  \item HISTIDINE -4 -4 -4
  \item HEXURONIC.ACID2 -4 -4 -4
  \item HEXITOL -4 -4 -4
  \item HEPTADECANOIC.ACID -4 -4 -4
  \item GLYCOLIC.ACID -4 -4 -4
  \item GLYCINE -4 -4 -4
  \item GLYCEROL.ALPHA.PHOSPHATE -4 -4 -4
  \item GLYCEROL.3.GALACTOSIDE -4 -4 -4
  \item GLYCEROL -4 -4 -4
  \item GLYCERIC.ACID -4 -4 -4
  \item GLUTAMINE -4 -4 -4
  \item GLUTAMIC.ACID -4 -4 -4
  \item GLUCOSE -4 -4 -4
  \item GLUCONIC.ACID -4 -4 -4
  \item GLUCOHEPTULOSE -4 -4 -4
  \item FUMARIC.ACID -4 -4 -4
  \item FUCOSE -4 -4 -4
  \item FRUCTOSE -4 -4 -4
  \item ETHANOLAMINE -4 -4 -4
  \item ERYTHRITOL -4 -4 -4
  \item EPSILON.CAPROLACTAM -4 -4 -4
  \item DODECANOL -4 -4 -4
  \item DIGLYCEROL -4 -4 -4
  \item DEOXYPENTITOL -4 -4 -4
  \item CYSTINE -4 -4 -4
  \item CYSTEINE -4 -4 -4
  \item CREATININE -4 -4 -4
  \item CONDURITOL.BETA.EPOXIDE -4 -4 -4
  \item CITRULLINE -4 -4 -4
  \item CITRIC.ACID -4 -4 -4
  \item CHOLESTEROL -4 -4 -4
  \item CELLOBIOSE -4 -4 -4
  \item CAPRYLIC.ACID -4 -4 -4
  \item CAPRIC.ACID -4 -4 -4
  \item CAFFEINE -4 -4 -4
  \item BETA.SITOSTEROL -4 -4 -4
  \item BETA.ALANINE -4 -4 -4
  \item BENZYLAMINE -4 -4 -4
  \item BENZYLALCOHOL -4 -4 -4
  \item BENZOIC.ACID -4 -4 -4
  \item BEHENIC.ACID -4 -4 -4
  \item ASPARTIC.ACID -4 -4 -4
  \item ASPARAGINE -4 -4 -4
  \item ARACHIDONIC.ACID -4 -4 -4
  \item ARACHIDIC.ACID -4 -4 -4
  \item ARABITOL -4 -4 -4
  \item AMINOMALONATE -4 -4 -4
  \item ALTROSE -4 -4 -4
  \item ALLANTOIC.ACID -4 -4 -4
  \item ALANINE.ALANINE -4 -4 -4
  \item ALANINE -4 -4 -4
  \item ADIPIC.ACID -4 -4 -4
  \item ACONITIC.ACID -4 -4 -4
  \item ACETAMINOPHEN -4 -4 -4
  \item X5.METHOXYTRYPTAMINE -4 -4 -4
  \item X5.HYDROXYNORVALINE.NIST -4 -4 -4
  \item X3.HYDROXYBUTYRIC.ACID -4 -4 -4
  \item X3.AMINOISOBUTYRIC.ACID -4 -4 -4
  \item X2.3.DIHYDROXYBUTANOIC.ACID.NIST -4 -4 -4
  \item X2.HYDROXYVALERIC.ACID -4 -4 -4
  \item X2.HYDROXYHIPPURIC.ACID -4 -4 -4
  \item X2.HYDROXYGLUTARIC.ACID -4 -4 -4
  \item X2.HYDROXYBUTANOIC.ACID -4 -4 -4
  \item X2.DEOXYTETRONIC.ACID -4 -4 -4
  \item X1.5.ANHYDROGLUCITOL -4 -4 -4
  \item X1.MONOSTEARIN -4 -4 -4
  \item X1.MONOPALMITIN -4 -4 -4
  \item X1.MONOOLEIN -4 -4 -4
  \item X102978 -4 -4 -4
  \item X16591 -4 -4 -4
  \item X130462 -4 -4 -4
  \item X130461 -4 -4 -4
  \item X4879 -4 -4 -4
  \item X130479 -4 -4 -4
  \item X130465 -4 -4 -4
  \item X130466 -4 -4 -4
  \item X130478 -4 -4 -4
  \item X153 -4 -4 -4
  \item X1862 -4 -4 -4
  \item X68 -4 -4 -4
  \item X490 -4 -4 -4
  \item X130463 -4 -4 -4
  \item X125662 -4 -4 -4
  \item X130470 -4 -4 -4
  \item X134 -4 -4 -4
  \item X130483 -4 -4 -4
  \item X105164 -4 -4 -4
  \item X130468 -4 -4 -4
  \item X85651 -4 -4 -4
  \item X108052 -4 -4 -4
  \item X12768 -4 -4 -4
  \item X1675 -4 -4 -4
  \item X6491 -4 -4 -4
  \item X130464 -4 -4 -4
  \item X1700 -4 -4 -4
  \item X31559 -4 -4 -4
  \item X16757 -4 -4 -4
  \item X130713 -4 -4 -4
  \item X124293 -4 -4 -4
  \item X18157 -4 -4 -4
  \item X4550 -4 -4 -4
  \item X3902 -4 -4 -4
  \item X131089 -4 -4 -4
  \item X39 -4 -4 -4
  \item X106936 -4 -4 -4
  \item X124880 -4 -4 -4
  \item X4871 -4 -4 -4
  \item X87756 -4 -4 -4
  \item X103476 -4 -4 -4
  \item X4533 -4 -4 -4
  \item X130716 -4 -4 -4
  \item X3286 -4 -4 -4
  \item X25321 -4 -4 -4
  \item X9320 -4 -4 -4
  \item X88583 -4 -4 -4
  \item X257 -4 -4 -4
  \item X6767 -4 -4 -4
  \item X458 -4 -4 -4
  \item X1909 -4 -4 -4
  \item X98 -4 -4 -4
  \item X21502 -4 -4 -4
  \item X43763 -4 -4 -4
  \item X137 -4 -4 -4
  \item X22735 -4 -4 -4
  \item X130396 -4 -4 -4
  \item X473 -4 -4 -4
  \item X462 -4 -4 -4
  \item X1873 -4 -4 -4
  \item X31756 -4 -4 -4
  \item X1941 -4 -4 -4
  \item X494 -4 -4 -4
  \item X105122 -4 -4 -4
  \item X3258 -4 -4 -4
  \item X168 -4 -4 -4
  \item X54 -4 -4 -4
  \item X11831 -4 -4 -4
  \item X12330 -4 -4 -4
  \item X915 -4 -4 -4
  \item X1709 -4 -4 -4
  \item X131325 -4 -4 -4
  \item X573 -4 -4 -4
  \item X812 -4 -4 -4
  \item X87877 -4 -4 -4
  \item X307 -4 -4 -4
  \item X109401 -4 -4 -4
  \item X89383 -4 -4 -4
  \item X110501 -4 -4 -4
  \item X3143 -4 -4 -4
  \item X9489 -4 -4 -4
  \item X61 -4 -4 -4
  \item X4887 -4 -4 -4
  \item X4605 -4 -4 -4
  \item X112556 -4 -4 -4
  \item X4929 -4 -4 -4
  \item X4898 -4 -4 -4
  \item X4937 -4 -4 -4
  \item X5396 -4 -4 -4
  \item X4577 -4 -4 -4
  \item X119167 -4 -4 -4
  \item X4536 -4 -4 -4
  \item X4863 -4 -4 -4
  \item X43729 -4 -4 -4
  \item X1704 -4 -4 -4
  \item X11523 -4 -4 -4
  \item X97743 -4 -4 -4
  \item X120562 -4 -4 -4
  \item X228 -4 -4 -4
  \item X1981 -4 -4 -4
  \item X101742 -4 -4 -4
  \item X97708 -4 -4 -4
  \item X8270 -4 -4 -4
  \item X592 -4 -4 -4
  \item X119023 -4 -4 -4
  \item X45378 -4 -4 -4
  \item X7403 -4 -4 -4
  \item X5290 -4 -4 -4
  \item X121467 -4 -4 -4
  \item X4579 -4 -4 -4
  \item X1872 -4 -4 -4
  \item X33282 -4 -4 -4
  \item X119026 -4 -4 -4
  \item X5930 -4 -4 -4
  \item X4264 -4 -4 -4
  \item X4548 -4 -4 -4
  \item X4945 -4 -4 -4
  \item X120526 -4 -4 -4
  \item X4600 -4 -4 -4
  \item X110008 -4 -4 -4
  \item X106629 -4 -4 -4
  \item X1912 -4 -4 -4
  \item X17664 -4 -4 -4
  \item X2430 -4 -4 -4
  \item X2011 -4 -4 -4
  \item X236 -4 -4 -4
  \item X46292 -4 -4 -4
  \item X6381 -4 -4 -4
  \item X110 -4 -4 -4
  \item X18488 -4 -4 -4
  \item X41833 -4 -4 -4
  \item X5259 -4 -4 -4
  \item X6324 -4 -4 -4
  \item X1875 -4 -4 -4
  \item X5471 -4 -4 -4
  \item X97891 -4 -4 -4
  \item X99 -4 -4 -4
  \item X107960 -4 -4 -4
  \item X109923 -4 -4 -4
  \item X62 -4 -4 -4
  \item X104901 -4 -4 -4
  \item X31359 -4 -4 -4
  \item X2039 -4 -4 -4
  \item X4546 -4 -4 -4
  \item X41808 -4 -4 -4
  \item X31664 -4 -4 -4
  \item X5990 -4 -4 -4
  \item X50367 -4 -4 -4
  \item X126423 -4 -4 -4
  \item X110013 -4 -4 -4
  \item X125664 -4 -4 -4
  \item X10176 -4 -4 -4
  \item X41811 -4 -4 -4
  \item X1064 -4 -4 -4
  \item X4263 -4 -4 -4
  \item X49400 -4 -4 -4
  \item X124996 -4 -4 -4
  \item X448 -4 -4 -4
  \item X5276 -4 -4 -4
  \item X16788 -4 -4 -4
  \item X4986 -4 -4 -4
  \item X5269 -4 -4 -4
  \item X22863 -4 -4 -4
  \item X100666 -4 -4 -4
  \item X20903 -4 -4 -4
  \item X121002 -4 -4 -4
  \item X119129 -4 -4 -4
  \item X1029 -4 -4 -4
  \item X2936 -4 -4 -4
  \item X117141 -4 -4 -4
  \item X124963 -4 -4 -4
  \item X47358 -4 -4 -4
  \item X91 -4 -4 -4
  \item X103102 -4 -4 -4
  \item X46138 -4 -4 -4
  \item X44509 -4 -4 -4
  \item X17651 -4 -4 -4
  \item X443 -4 -4 -4
  \item X5346 -4 -4 -4
  \item X1996 -4 -4 -4
  \item X62263 -4 -4 -4
  \item X87822 -4 -4 -4
  \item X129225 -4 -4 -4
  \item X657 -4 -4 -4
  \item X122191 -4 -4 -4
  \item X4746 -4 -4 -4
  \item X47221 -4 -4 -4
  \item X610256 -4 -4 -4
  \item X4609 -4 -4 -4
  \item X3228 -4 -4 -4
  \item X1878 -4 -4 -4
  \item X2900 -4 -4 -4
  \item X46131 -4 -4 -4
  \item X4292 -4 -4 -4
  \item X110612 -4 -4 -4
  \item X109386 -4 -4 -4
  \item X21885 -4 -4 -4
  \item X6646 -4 -4 -4
  \item X17068 -4 -4 -4
  \item X2476 -4 -4 -4
  \item X47 -4 -4 -4
  \item X46362 -4 -4 -4
  \item X4793 -4 -4 -4
  \item X112264 -4 -4 -4
  \item X121890 -4 -4 -4
  \item X3781 -4 -4 -4
  \item X110321 -4 -4 -4
  \item X119066 -4 -4 -4
  \item X5085 -4 -4 -4
  \item X103390 -4 -4 -4
  \item X21666 -4 -4 -4
  \item X1922 -4 -4 -4
  \item X130475 -4 -4 -4
  \item X31285 -4 -4 -4
  \item X120744 -4 -4 -4
  \item X4529 -4 -4 -4
  \item X479 -4 -4 -4
  \item X124283 -4 -4 -4
  \item X98027 -4 -4 -4
  \item X118071 -4 -4 -4
  \item X892 -4 -4 -4
  \item X31357 -4 -4 -4
  \item X108309 -4 -4 -4
  \item X34135 -4 -4 -4
  \item X4901 -4 -4 -4
  \item X6300 -4 -4 -4
  \item X23635 -4 -4 -4
  \item X47170 -4 -4 -4
  \item X5121 -4 -4 -4
  \item X33993 -4 -4 -4
  \item X95410 -4 -4 -4
  \item X18176 -4 -4 -4
  \item X42357 -4 -4 -4
  \item X7444 -4 -4 -4
  \item X102223 -4 -4 -4
  \item X2061 -4 -4 -4
  \item X18403 -4 -4 -4
  \item X120781 -4 -4 -4
  \item X2674 -4 -4 -4
  \item X453 -4 -4 -4
  \item X106427 -4 -4 -4
}
}
\examples{
\dontrun{
describe(admcgctof)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/ADMCGCTOF/ADNI_Methods_GC-TOF_MS_metabolomics_UC_Davis_10-2016.pdf}
}
\keyword{datasets}
