% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{admcgutmetaboliteslong}
\alias{admcgutmetaboliteslong}
\title{ADMC U Hawaii UPLC-MS/MS Gut Metabolites Serum Longitudinal}
\format{
A data frame with 4229 rows and 112 variables
}
\usage{
data(admcgutmetaboliteslong)
}
\description{
ADMC U Hawaii UPLC-MS/MS Gut Metabolites Serum Longitudinal
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item DRAWDATE -4 -4 -4
  \item SAMPLE_ID -4 -4 -4
  \item VID -4 -4 -4
  \item DRAWTIME -4 -4 -4
  \item L.HISTIDINE -4 M L-Histidine
  \item L.LYSINE -4 M L-Lysine
  \item L.SERINE -4 M L-Serine
  \item BETA.ALANINE -4 M Beta Alanine
  \item L.ALANINE -4 M L-Alanine
  \item L.TYROSINE -4 M L-Tyrosine
  \item L.ASPARAGINE -4 M L-Asparagine
  \item L.ASPARTIC.ACID -4 M L-Aspartic acid
  \item HIPPURIC.ACID -4 M Hippuric acid
  \item L.MALIC.ACID -4 M L-Malic acid
  \item ADIPIC.ACID -4 M Adipic acid
  \item INDOLEACETIC.ACID -4 M Indoleacetic acid
  \item X3.INDOLEPROPIONIC.ACID -4 M 3-Indolepropionic acid
  \item X3.METHYL.2.OXOVALERIC.ACID -4 M 3-Methyl-2-oxovaleric acid
  \item GLYCINE -4 M Glycine
  \item GLYCERIC.ACID -4 M Glyceric acid
  \item L.PROLINE -4 M L-Proline
  \item L.ALPHA.AMINOBUTYRIC.ACID -4 M L-Alpha aminobutyric acid
  \item GLYCOLIC.ACID -4 M Glycolic acid
  \item P.CRESOL.SULFATE -4 M p-Cresol sulfate
  \item PYROGLUTAMIC.ACID -4 M Pyroglutamic acid
  \item L.VALINE -4 M L-Valine
  \item L.METHIONINE -4 M L-Methionine
  \item L.LEUCINE -4 M L-Leucine
  \item L.TRYPTOPHAN -4 M L-Tryptophan
  \item CIS.ACONITIC_ACID -4 M cis-Aconitic_acid
  \item METHYLSUCCINIC.ACID -4 M Methylsuccinic acid
  \item ISOCITRIC.ACID -4 M Isocitric acid
  \item CITRIC.ACID -4 M Citric acid
  \item OXOGLUTARIC.ACID -4 M Oxoglutaric acid
  \item C1.0 -4 M Formic acid
  \item X3.HYDROXYBUTYRIC.ACID -4 M 3-Hydroxybutyric acid
  \item C2.0 -4 M Acetic acid
  \item C3.0 -4 M Propionic acid
  \item X3.HYDROXYISOVALERIC.ACID -4 M 3-Hydroxyisovaleric acid
  \item X2.HYDROXYBUTYRIC.ACID -4 M 2-Hydroxybutyric acid
  \item ISOBUTYRIC.ACID -4 M Isobutyric acid
  \item C4.0 -4 M Butyric acid
  \item SUCCINIC.ACID -4 M Succinic acid
  \item C5.0 -4 M Valeric acid
  \item SUBERIC.ACID -4 M Suberic acid
  \item X....2.METHYLPENTANOIC.ACID -4 M ()-2-Methylpentanoic acid
  \item C6.0 -4 M Hexanoic acid
  \item C7.0 -4 M Heptanoic acid
  \item C8.0 -4 M Caprylic acid
  \item C11.1.CIS.10. -4 M Undecylenic acid
  \item C10.0 -4 M Decanoic acid
  \item C9.0 -4 M Pelargonic acid
  \item C11.0 -4 M Undecanoic acid
  \item C12.0 -4 M Dodecanoic acid
  \item C13.0 -4 M Tridecanoic acid
  \item C20.5.CIS.5.8.11.14.17. -4 M Eicosapentaenoic acid (EPA)
  \item C15.0 -4 M Pentadecanoic acid
  \item C18.3.CIS.9.12.15. -4 M Alpha-Linolenic acid
  \item C18.3.CIS.6.9.12. -4 M Gamma-Linolenic acid
  \item C14.0 -4 M Myristic acid
  \item C22.6.CIS.4.7.10.13.16.19. -4 M Docosahexaenoic acid
  \item C20.2.CIS.11.14. -4 M Eicosadienoic acid
  \item C16.1.CIS.9. -4 M Palmitoleic acid
  \item C20.4.CIS.5.8.11.14. -4 M Arachidonic acid
  \item C22.5.CIS.7.10.13.16.19. -4 M Docosapentaenoic acid
  \item C22.5.CIS.4.7.10.13.16. -4 M Docosapentaenoic acid 
  \item C18.2.CIS.9.12. -4 M Linoleic acid
  \item C17.0 -4 M Heptadecanoic acid
  \item C20.3.CIS.8.11.14. -4 M 8,11,14-Eicosatrienoic acid
  \item C17.1.CIS.10. -4 M 10Z-Heptadecenoic acid
  \item C19.2.CIS.10.13. -4 M 10,13-Nonadecadienoic Acid
  \item C16.0 -4 M Palmitic acid
  \item C22.4.CIS.7.10.13.16. -4 M docosatetraenoic acid (Adrenic acid)
  \item C18.1.CIS.9. -4 M Oleic acid
  \item C18.0 -4 M Stearic acid
  \item X12.KETOLCA -4 nM 12-Ketolithocholic acid /12-Ketodeoxycholic acid
  \item X3.DHCA -4 nM 3-Dehydrocholic Acid/ 3-Oxocholic Acid
  \item X7.KETOLCA -4 nM Nutriacholic acid / 7-Ketolithocholic Acid
  \item X7.DHCA -4 nM 7-Dehydrocholic Acid
  \item ALLOLCA -4 nM Allolithocholic acid (Isoallolithocholic acid)
  \item APOCA -4 nM Apocholic Acid
  \item CA -4 nM Cholic Acid
  \item CDCA -4 nM Chenodeoxycholic Acid
  \item DCA -4 nM Deoxycholic Acid
  \item DEHYDROLCA -4 nM Dehydrolithocholic Acid
  \item GCA -4 nM Glycocholic Acid
  \item GCDCA -4 nM Glycochenodeoxycholic Acid
  \item GDCA -4 nM Glycodeoxycholic Acid
  \item GHCA -4 nM Glycohyocholic Acid
  \item GHDCA -4 nM Glycohyodeoxycholic Acid
  \item GLCA -4 nM Glycolithocholic Acid
  \item GUDCA -4 nM Glycoursodeoxycholic Acid
  \item HCA -4 nM -Muricholic Acid/Hyocholic Acid
  \item HDCA -4 nM -Hyodeoxycholic Acid 
  \item ISOLCA -4 nM Isolithocholic acid
  \item LCA -4 nM Lithocholic Acid
  \item LCA_3S -4 nM Lithocholic Acid 3-Sulfate
  \item MUROCA -4 nM Murocholic Acid
  \item NORCA -4 nM Norcholic acid
  \item NORDCA -4 nM 3,12-Dihydroxynorcholanate\23-Nordeoxycholic acid
  \item TCA -4 nM Taurocholic Acid
  \item TCDCA -4 nM Taurochenodeoxycholic acid
  \item TDCA -4 nM Taurodeoxycholic Acid
  \item TUDCA -4 nM Tauroursodeoxycholic Acid
  \item TAMCA -4 nM Tauro--muricholic Acid 
  \item UCA -4 nM Ursocholic Acid
  \item UDCA -4 nM Ursodeoxycholic Acid
  \item BUDCA -4 nM 3-Ursodeoxycholic Acid (Isoursodeoxycholic acid)
}
}
\examples{
\dontrun{
describe(admcgutmetaboliteslong)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/ADMCGUTMETABOLITESLONG/UHawaii_methods_Human_serum_Rima_MicrobiomeMetabolites[Updated_for_20-10-2020].pdf}
}
\keyword{datasets}
