% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{admchivalmetab}
\alias{admchivalmetab}
\title{Targeted UHPLC-MS analysis of High-Value Metabolites in Serum Samples}
\format{
A data frame with 829 rows and 22 variables
}
\usage{
data(admchivalmetab)
}
\description{
Targeted UHPLC-MS analysis of High-Value Metabolites in Serum Samples
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item RECNO -4 -4 Record number (2 indicates replicate sample)
  \item Samples.Metabolites -4 -4 -4
  \item Comments -4 -4 -4
  \item ADMA -4 -4 -4
  \item Betaine -4 -4 -4
  \item Choline -4 -4 -4
  \item Creatinine -4 -4 -4
  \item Glutamic.acid -4 -4 -4
  \item Leucine -4 -4 -4
  \item Methionine -4 -4 -4
  \item Phenylalanine -4 -4 -4
  \item Serine -4 -4 -4
  \item Taurine -4 -4 -4
  \item TMAO -4 -4 -4
  \item Tryptophan -4 -4 -4
  \item Tyrosine -4 -4 -4
  \item Glutamine -4 -4 -4
  \item beta.hydroxybutyrate -4 -4 -4
  \item Glucose -4 -4 -4
  \item Lactate -4 -4 -4
}
}
\examples{
\dontrun{
describe(admchivalmetab)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/ADMCHIVALMETAB/ADMC_high_value_high_abundance\%20metabolites.pdf}
}
\keyword{datasets}
