% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{admclipidomicsmeiklelablong}
\alias{admclipidomicsmeiklelablong}
\title{ADMC Lipidomics Meikle Lab Longitudinal Data Matrix}
\format{
A data frame with 4873 rows and 788 variables
}
\usage{
data(admclipidomicsmeiklelablong)
}
\description{
ADMC Lipidomics Meikle Lab Longitudinal Data Matrix
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item GUSPECID -4 -4 -4
  \item COHORT -4 -4 Cohort. ADNI1 longitudinal or ADNI2/GO
  \item SAMPLE.ID -4 -4 -4
  \item SPH.D18.1. -4 -4 Sph(d18:1). Lipid metabolite [nM]
  \item SPH.D18.2. -4 -4 Sph(d18:2). Lipid metabolite [nM]
  \item S1P.D16.1. -4 -4 S1P(d16:1). Lipid metabolite [nM]
  \item S1P.D18.0. -4 -4 S1P(d18:0). Lipid metabolite [nM]
  \item S1P.D18.1. -4 -4 S1P(d18:1). Lipid metabolite [nM]
  \item S1P.D18.2. -4 -4 S1P(d18:2). Lipid metabolite [nM]
  \item DHCER.D18.0.16.0. -4 -4 dhCer(d18:0/16:0). Lipid metabolite [nM]
  \item DHCER.D18.0.18.0. -4 -4 dhCer(d18:0/18:0). Lipid metabolite [nM]
  \item DHCER.D18.0.20.0. -4 -4 dhCer(d18:0/20:0). Lipid metabolite [nM]
  \item DHCER.D18.0.22.0. -4 -4 dhCer(d18:0/22:0). Lipid metabolite [nM]
  \item DHCER.D18.0.24.0. -4 -4 dhCer(d18:0/24:0). Lipid metabolite [nM]
  \item DHCER.D18.0.24.1. -4 -4 dhCer(d18:0/24:1). Lipid metabolite [nM]
  \item CER.D16.1.16.0. -4 -4 Cer(d16:1/16:0). Lipid metabolite [nM]
  \item CER.D16.1.18.0. -4 -4 Cer(d16:1/18:0). Lipid metabolite [nM]
  \item CER.D16.1.20.0. -4 -4 Cer(d16:1/20:0). Lipid metabolite [nM]
  \item CER.D16.1.22.0. -4 -4 Cer(d16:1/22:0). Lipid metabolite [nM]
  \item CER.D16.1.23.0. -4 -4 Cer(d16:1/23:0). Lipid metabolite [nM]
  \item CER.D16.1.24.0. -4 -4 Cer(d16:1/24:0). Lipid metabolite [nM]
  \item CER.D16.1.24.1. -4 -4 Cer(d16:1/24:1). Lipid metabolite [nM]
  \item CER.D17.1.16.0. -4 -4 Cer(d17:1/16:0). Lipid metabolite [nM]
  \item CER.D17.1.18.0. -4 -4 Cer(d17:1/18:0). Lipid metabolite [nM]
  \item CER.D17.1.20.0. -4 -4 Cer(d17:1/20:0). Lipid metabolite [nM]
  \item CER.D17.1.22.0. -4 -4 Cer(d17:1/22:0). Lipid metabolite [nM]
  \item CER.D17.1.23.0. -4 -4 Cer(d17:1/23:0). Lipid metabolite [nM]
  \item CER.D17.1.24.0. -4 -4 Cer(d17:1/24:0). Lipid metabolite [nM]
  \item CER.D17.1.24.1. -4 -4 Cer(d17:1/24:1). Lipid metabolite [nM]
  \item CER.D18.1.14.0. -4 -4 Cer(d18:1/14:0). Lipid metabolite [nM]
  \item CER.D18.1.16.0. -4 -4 Cer(d18:1/16:0). Lipid metabolite [nM]
  \item CER.D18.1.18.0. -4 -4 Cer(d18:1/18:0). Lipid metabolite [nM]
  \item CER.D18.1.19.0. -4 -4 Cer(d18:1/19:0). Lipid metabolite [nM]
  \item CER.D18.1.20.0. -4 -4 Cer(d18:1/20:0). Lipid metabolite [nM]
  \item CER.D18.1.21.0. -4 -4 Cer(d18:1/21:0). Lipid metabolite [nM]
  \item CER.D18.1.22.0. -4 -4 Cer(d18:1/22:0). Lipid metabolite [nM]
  \item CER.D18.1.23.0. -4 -4 Cer(d18:1/23:0). Lipid metabolite [nM]
  \item CER.D18.1.24.0. -4 -4 Cer(d18:1/24:0). Lipid metabolite [nM]
  \item CER.D18.1.24.1. -4 -4 Cer(d18:1/24:1). Lipid metabolite [nM]
  \item CER.D18.1.26.0. -4 -4 Cer(d18:1/26:0). Lipid metabolite [nM]
  \item CER.D18.2.14.0. -4 -4 Cer(d18:2/14:0). Lipid metabolite [nM]
  \item CER.D18.2.16.0. -4 -4 Cer(d18:2/16:0). Lipid metabolite [nM]
  \item CER.D18.2.18.0. -4 -4 Cer(d18:2/18:0). Lipid metabolite [nM]
  \item CER.D18.2.20.0. -4 -4 Cer(d18:2/20:0). Lipid metabolite [nM]
  \item CER.D18.2.21.0. -4 -4 Cer(d18:2/21:0). Lipid metabolite [nM]
  \item CER.D18.2.22.0. -4 -4 Cer(d18:2/22:0). Lipid metabolite [nM]
  \item CER.D18.2.23.0. -4 -4 Cer(d18:2/23:0). Lipid metabolite [nM]
  \item CER.D18.2.24.0. -4 -4 Cer(d18:2/24:0). Lipid metabolite [nM]
  \item CER.D18.2.24.1. -4 -4 Cer(d18:2/24:1). Lipid metabolite [nM]
  \item CER.D18.2.26.0. -4 -4 Cer(d18:2/26:0). Lipid metabolite [nM]
  \item CER.D19.1.16.0. -4 -4 Cer(d19:1/16:0). Lipid metabolite [nM]
  \item CER.D19.1.18.0. -4 -4 Cer(d19:1/18:0). Lipid metabolite [nM]
  \item CER.D19.1.20.0. -4 -4 Cer(d19:1/20:0). Lipid metabolite [nM]
  \item CER.D19.1.22.0. -4 -4 Cer(d19:1/22:0). Lipid metabolite [nM]
  \item CER.D19.1.23.0. -4 -4 Cer(d19:1/23:0). Lipid metabolite [nM]
  \item CER.D19.1.24.0. -4 -4 Cer(d19:1/24:0). Lipid metabolite [nM]
  \item CER.D19.1.24.1. -4 -4 Cer(d19:1/24:1). Lipid metabolite [nM]
  \item CER.D19.1.26.0. -4 -4 Cer(d19:1/26:0). Lipid metabolite [nM]
  \item CER.D20.1.22.0. -4 -4 Cer(d20:1/22:0). Lipid metabolite [nM]
  \item CER.D20.1.23.0. -4 -4 Cer(d20:1/23:0). Lipid metabolite [nM]
  \item CER.D20.1.24.0. -4 -4 Cer(d20:1/24:0). Lipid metabolite [nM]
  \item CER.D20.1.24.1. -4 -4 Cer(d20:1/24:1). Lipid metabolite [nM]
  \item CER.M18.0.20.0. -4 -4 Cer(m18:0/20:0). Lipid metabolite [nM]
  \item CER.M18.0.22.0. -4 -4 Cer(m18:0/22:0). Lipid metabolite [nM]
  \item CER.M18.0.23.0. -4 -4 Cer(m18:0/23:0). Lipid metabolite [nM]
  \item CER.M18.0.24.0. -4 -4 Cer(m18:0/24:0). Lipid metabolite [nM]
  \item CER.M18.0.24.1. -4 -4 Cer(m18:0/24:1). Lipid metabolite [nM]
  \item CER.M18.1.18.0. -4 -4 Cer(m18:1/18:0). Lipid metabolite [nM]
  \item CER.M18.1.20.0. -4 -4 Cer(m18:1/20:0). Lipid metabolite [nM]
  \item CER.M18.1.22.0. -4 -4 Cer(m18:1/22:0). Lipid metabolite [nM]
  \item CER.M18.1.23.0. -4 -4 Cer(m18:1/23:0). Lipid metabolite [nM]
  \item CER.M18.1.24.0. -4 -4 Cer(m18:1/24:0). Lipid metabolite [nM]
  \item CER.M18.1.24.1. -4 -4 Cer(m18:1/24:1). Lipid metabolite [nM]
  \item CER1P.D18.1.16.0. -4 -4 Cer1P(d18:1/16:0). Lipid metabolite [nM]
  \item HEXCER.D16.1.18.0. -4 -4 HexCer(d16:1/18:0). Lipid metabolite [nM]
  \item HEXCER.D16.1.20.0. -4 -4 HexCer(d16:1/20:0). Lipid metabolite [nM]
  \item HEXCER.D16.1.22.0. -4 -4 HexCer(d16:1/22:0). Lipid metabolite [nM]
  \item HEXCER.D16.1.24.0. -4 -4 HexCer(d16:1/24:0). Lipid metabolite [nM]
  \item HEXCER.D18.1.16.0. -4 -4 HexCer(d18:1/16:0). Lipid metabolite [nM]
  \item HEXCER.D18.1.18.0. -4 -4 HexCer(d18:1/18:0). Lipid metabolite [nM]
  \item HEXCER.D18.1.20.0. -4 -4 HexCer(d18:1/20:0). Lipid metabolite [nM]
  \item HEXCER.D18.1.22.0. -4 -4 HexCer(d18:1/22:0). Lipid metabolite [nM]
  \item HEXCER.D18.1.24.0. -4 -4 HexCer(d18:1/24:0). Lipid metabolite [nM]
  \item HEXCER.D18.1.24.1. -4 -4 HexCer(d18:1/24:1). Lipid metabolite [nM]
  \item HEXCER.D18.2.18.0. -4 -4 HexCer(d18:2/18:0). Lipid metabolite [nM]
  \item HEXCER.D18.2.20.0. -4 -4 HexCer(d18:2/20:0). Lipid metabolite [nM]
  \item HEXCER.D18.2.22.0. -4 -4 HexCer(d18:2/22:0). Lipid metabolite [nM]
  \item HEXCER.D18.2.24.0. -4 -4 HexCer(d18:2/24:0). Lipid metabolite [nM]
  \item HEX2CER.D16.1.16.0. -4 -4 Hex2Cer(d16:1/16:0). Lipid metabolite [nM]
  \item HEX2CER.D16.1.24.1. -4 -4 Hex2Cer(d16:1/24:1). Lipid metabolite [nM]
  \item HEX2CER.D18.1.16.0. -4 -4 Hex2Cer(d18:1/16:0). Lipid metabolite [nM]
  \item HEX2CER.D18.1.20.0. -4 -4 Hex2Cer(d18:1/20:0). Lipid metabolite [nM]
  \item HEX2CER.D18.1.22.0. -4 -4 Hex2Cer(d18:1/22:0). Lipid metabolite [nM]
  \item HEX2CER.D18.1.24.0. -4 -4 Hex2Cer(d18:1/24:0). Lipid metabolite [nM]
  \item HEX2CER.D18.1.24.1. -4 -4 Hex2Cer(d18:1/24:1). Lipid metabolite [nM]
  \item HEX2CER.D18.2.16.0. -4 -4 Hex2Cer(d18:2/16:0). Lipid metabolite [nM]
  \item HEX2CER.D18.2.24.1. -4 -4 Hex2Cer(d18:2/24:1). Lipid metabolite [nM]
  \item HEX3CER.D18.1.16.0. -4 -4 Hex3Cer(d18:1/16:0). Lipid metabolite [nM]
  \item HEX3CER.D18.1.18.0. -4 -4 Hex3Cer(d18:1/18:0). Lipid metabolite [nM]
  \item HEX3CER.D18.1.20.0. -4 -4 Hex3Cer(d18:1/20:0). Lipid metabolite [nM]
  \item HEX3CER.D18.1.22.0. -4 -4 Hex3Cer(d18:1/22:0). Lipid metabolite [nM]
  \item HEX3CER.D18.1.24.0. -4 -4 Hex3Cer(d18:1/24:0). Lipid metabolite [nM]
  \item HEX3CER.D18.1.24.1. -4 -4 Hex3Cer(d18:1/24:1). Lipid metabolite [nM]
  \item GM3.D18.1.16.0. -4 -4 GM3(d18:1/16:0). Lipid metabolite [nM]
  \item GM3.D18.1.18.0. -4 -4 GM3(d18:1/18:0). Lipid metabolite [nM]
  \item GM3.D18.1.20.0. -4 -4 GM3(d18:1/20:0). Lipid metabolite [nM]
  \item GM3.D18.1.22.0. -4 -4 GM3(d18:1/22:0). Lipid metabolite [nM]
  \item GM3.D18.1.24.0. -4 -4 GM3(d18:1/24:0). Lipid metabolite [nM]
  \item GM3.D18.1.24.1. -4 -4 GM3(d18:1/24:1). Lipid metabolite [nM]
  \item GM1.D18.1.16.0. -4 -4 GM1(d18:1/16:0). Lipid metabolite [nM]
  \item SULFATIDE..D18.1..16.0.OH.. -4 -4 Sulfatide (d18:1:/16:0(OH)). Lipid metabolite [nM]
  \item SULFATIDE..D18.1..16.0. -4 -4 Sulfatide (d18:1:/16:0). Lipid metabolite [nM]
  \item SULFATIDE..D18.1..24.0.OH.. -4 -4 Sulfatide (d18:1:/24:0(OH)). Lipid metabolite [nM]
  \item SULFATIDE..D18.1..24.0. -4 -4 Sulfatide (d18:1:/24:0). Lipid metabolite [nM]
  \item SULFATIDE..D18.1..24.1.OH.. -4 -4 Sulfatide (d18:1:/24:1(OH)). Lipid metabolite [nM]
  \item SULFATIDE..D18.1..24.1. -4 -4 Sulfatide (d18:1:/24:1). Lipid metabolite [nM]
  \item SM.34.3. -4 -4 SM(34:3). Lipid metabolite [nM]
  \item SM.35.2...B. -4 -4 SM(35:2) (b). Lipid metabolite [nM]
  \item SM.37.1. -4 -4 SM(37:1). Lipid metabolite [nM]
  \item SM.37.2. -4 -4 SM(37:2). Lipid metabolite [nM]
  \item SM.38.3...A. -4 -4 SM(38:3) (a). Lipid metabolite [nM]
  \item SM.38.3...B. -4 -4 SM(38:3) (b). Lipid metabolite [nM]
  \item SM.40.3...A. -4 -4 SM(40:3) (a). Lipid metabolite [nM]
  \item SM.40.3...B. -4 -4 SM(40:3) (b). Lipid metabolite [nM]
  \item SM.41.0. -4 -4 SM(41:0). Lipid metabolite [nM]
  \item SM.41.1...A. -4 -4 SM(41:1) (a). Lipid metabolite [nM]
  \item SM.43.1. -4 -4 SM(43:1). Lipid metabolite [nM]
  \item SM.43.2...B. -4 -4 SM(43:2) (b). Lipid metabolite [nM]
  \item SM.43.2...C. -4 -4 SM(43:2) (c). Lipid metabolite [nM]
  \item SM.44.1. -4 -4 SM(44:1). Lipid metabolite [nM]
  \item SM.44.2. -4 -4 SM(44:2). Lipid metabolite [nM]
  \item SM.44.3. -4 -4 SM(44:3). Lipid metabolite [nM]
  \item SM.D16.1.19.0. -4 -4 SM(d16:1/19:0). Lipid metabolite [nM]
  \item SM.D16.1.23.0..SM.D17.1.22.0. -4 -4 SM(d16:1/23:0)/SM(d17:1/22:0). Lipid metabolite [nM]
  \item SM.D16.1.24.1. -4 -4 SM(d16:1/24:1). Lipid metabolite [nM]
  \item SM.D17.1.14.0. -4 -4 SM(d17:1/14:0). Lipid metabolite [nM]
  \item SM.D17.1.16.0. -4 -4 SM(d17:1/16:0). Lipid metabolite [nM]
  \item SM.D17.1.24.1. -4 -4 SM(d17:1/24:1). Lipid metabolite [nM]
  \item SM.D18.0.14.0. -4 -4 SM(d18:0/14:0). Lipid metabolite [nM]
  \item SM.D18.0.16.0. -4 -4 SM(d18:0/16:0). Lipid metabolite [nM]
  \item SM.D18.0.22.0. -4 -4 SM(d18:0/22:0). Lipid metabolite [nM]
  \item SM.D18.1.14.0..SM.D16.1.16.0. -4 -4 SM(d18:1/14:0)/SM(d16:1/16:0). Lipid metabolite [nM]
  \item SM.D18.1.16.0. -4 -4 SM(d18:1/16:0). Lipid metabolite [nM]
  \item SM.D18.1.17.0..SM.D17.1.18.0. -4 -4 SM(d18:1/17:0)/SM(d17:1/18:0). Lipid metabolite [nM]
  \item SM.D18.1.18.0..SM.D16.1.20.0. -4 -4 SM(d18:1/18:0)/SM(d16:1/20:0). Lipid metabolite [nM]
  \item SM.D18.1.20.0..SM.D16.1.22.0. -4 -4 SM(d18:1/20:0)/SM(d16:1/22:0). Lipid metabolite [nM]
  \item SM.D18.1.22.0..SM.D16.1.24.0. -4 -4 SM(d18:1/22:0)/SM(d16:1/24:0). Lipid metabolite [nM]
  \item SM.D18.1.23.0..SM.D17.1.24.0. -4 -4 SM(d18:1/23:0)/SM(d17:1/24:0). Lipid metabolite [nM]
  \item SM.D18.1.24.0. -4 -4 SM(d18:1/24:0). Lipid metabolite [nM]
  \item SM.D18.1.24.1. -4 -4 SM(d18:1/24:1). Lipid metabolite [nM]
  \item SM.D18.2.14.0. -4 -4 SM(d18:2/14:0). Lipid metabolite [nM]
  \item SM.D18.2.16.0. -4 -4 SM(d18:2/16:0). Lipid metabolite [nM]
  \item SM.D18.2.17.0. -4 -4 SM(d18:2/17:0). Lipid metabolite [nM]
  \item SM.D18.2.18.0. -4 -4 SM(d18:2/18:0). Lipid metabolite [nM]
  \item SM.D18.2.18.1. -4 -4 SM(d18:2/18:1). Lipid metabolite [nM]
  \item SM.D18.2.20.0. -4 -4 SM(d18:2/20:0). Lipid metabolite [nM]
  \item SM.D18.2.22.0. -4 -4 SM(d18:2/22:0). Lipid metabolite [nM]
  \item SM.D18.2.23.0. -4 -4 SM(d18:2/23:0). Lipid metabolite [nM]
  \item SM.D18.2.24.0. -4 -4 SM(d18:2/24:0). Lipid metabolite [nM]
  \item SM.D19.1.24.1. -4 -4 SM(d19:1/24:1). Lipid metabolite [nM]
  \item PA.34.1. -4 -4 PA(34:1). Lipid metabolite [nM]
  \item PA.36.1. -4 -4 PA(36:1). Lipid metabolite [nM]
  \item PA.36.2. -4 -4 PA(36:2). Lipid metabolite [nM]
  \item PA.36.3. -4 -4 PA(36:3). Lipid metabolite [nM]
  \item PA.36.4. -4 -4 PA(36:4). Lipid metabolite [nM]
  \item PA.40.6. -4 -4 PA(40:6). Lipid metabolite [nM]
  \item PC.14.0_16.0. -4 -4 PC(14:0_16:0). Lipid metabolite [nM]
  \item PC.14.0_20.4. -4 -4 PC(14:0_20:4). Lipid metabolite [nM]
  \item PC.14.0_22.6. -4 -4 PC(14:0_22:6). Lipid metabolite [nM]
  \item PC.15.MHDA_18.1. -4 -4 PC(15-MHDA_18:1). Lipid metabolite [nM]
  \item PC.15.MHDA_18.2. -4 -4 PC(15-MHDA_18:2). Lipid metabolite [nM]
  \item PC.15.MHDA_20.4. -4 -4 PC(15-MHDA_20:4). Lipid metabolite [nM]
  \item PC.15.MHDA_22.6. -4 -4 PC(15-MHDA_22:6). Lipid metabolite [nM]
  \item PC.15.0_20.3. -4 -4 PC(15:0_20:3). Lipid metabolite [nM]
  \item PC.15.0_20.4. -4 -4 PC(15:0_20:4). Lipid metabolite [nM]
  \item PC.15.0_22.6. -4 -4 PC(15:0_22:6). Lipid metabolite [nM]
  \item PC.16.0_16.0. -4 -4 PC(16:0_16:0). Lipid metabolite [nM]
  \item PC.16.0_18.0. -4 -4 PC(16:0_18:0). Lipid metabolite [nM]
  \item PC.16.0_18.1. -4 -4 PC(16:0_18:1). Lipid metabolite [nM]
  \item PC.16.0_18.2. -4 -4 PC(16:0_18:2). Lipid metabolite [nM]
  \item PC.16.0_18.3...A. -4 -4 PC(16:0_18:3) (a). Lipid metabolite [nM]
  \item PC.16.0_18.3...B. -4 -4 PC(16:0_18:3) (b). Lipid metabolite [nM]
  \item PC.16.0_20.3...A. -4 -4 PC(16:0_20:3) (a). Lipid metabolite [nM]
  \item PC.16.0_20.3...B. -4 -4 PC(16:0_20:3) (b). Lipid metabolite [nM]
  \item PC.16.0_20.4. -4 -4 PC(16:0_20:4). Lipid metabolite [nM]
  \item PC.16.0_20.5. -4 -4 PC(16:0_20:5). Lipid metabolite [nM]
  \item PC.16.0_22.6. -4 -4 PC(16:0_22:6). Lipid metabolite [nM]
  \item PC.16.1_18.2. -4 -4 PC(16:1_18:2). Lipid metabolite [nM]
  \item PC.16.1_20.4. -4 -4 PC(16:1_20:4). Lipid metabolite [nM]
  \item PC.16.1_22.6. -4 -4 PC(16:1_22:6). Lipid metabolite [nM]
  \item PC.17.0_18.1. -4 -4 PC(17:0_18:1). Lipid metabolite [nM]
  \item PC.17.0_18.2. -4 -4 PC(17:0_18:2). Lipid metabolite [nM]
  \item PC.17.0_20.4. -4 -4 PC(17:0_20:4). Lipid metabolite [nM]
  \item PC.17.0_22.6. -4 -4 PC(17:0_22:6). Lipid metabolite [nM]
  \item PC.17.1_18.2. -4 -4 PC(17:1_18:2). Lipid metabolite [nM]
  \item PC.18.0_18.1. -4 -4 PC(18:0_18:1). Lipid metabolite [nM]
  \item PC.18.0_18.2. -4 -4 PC(18:0_18:2). Lipid metabolite [nM]
  \item PC.18.0_20.3. -4 -4 PC(18:0_20:3). Lipid metabolite [nM]
  \item PC.18.0_20.4. -4 -4 PC(18:0_20:4). Lipid metabolite [nM]
  \item PC.18.0_22.4. -4 -4 PC(18:0_22:4). Lipid metabolite [nM]
  \item PC.18.0_22.5...N3..PC.20.1_20.4. -4 -4 PC(18:0_22:5) (n3)/PC(20:1_20:4). Lipid metabolite [nM]
  \item PC.18.0_22.5...N6. -4 -4 PC(18:0_22:5) (n6). Lipid metabolite [nM]
  \item PC.18.0_22.6. -4 -4 PC(18:0_22:6). Lipid metabolite [nM]
  \item PC.18.1_18.1. -4 -4 PC(18:1_18:1). Lipid metabolite [nM]
  \item PC.18.1_18.2. -4 -4 PC(18:1_18:2). Lipid metabolite [nM]
  \item PC.18.1_20.3. -4 -4 PC(18:1_20:3). Lipid metabolite [nM]
  \item PC.18.1_22.6...A. -4 -4 PC(18:1_22:6) (a). Lipid metabolite [nM]
  \item PC.18.1_22.6...B. -4 -4 PC(18:1_22:6) (b). Lipid metabolite [nM]
  \item PC.18.2_18.2. -4 -4 PC(18:2_18:2). Lipid metabolite [nM]
  \item PC.18.2_20.5. -4 -4 PC(18:2_20:5). Lipid metabolite [nM]
  \item PC.20.0_20.4. -4 -4 PC(20:0_20:4). Lipid metabolite [nM]
  \item PC.28.0. -4 -4 PC(28:0). Lipid metabolite [nM]
  \item PC.31.0...A. -4 -4 PC(31:0) (a). Lipid metabolite [nM]
  \item PC.31.0...B. -4 -4 PC(31:0) (b). Lipid metabolite [nM]
  \item PC.32.1. -4 -4 PC(32:1). Lipid metabolite [nM]
  \item PC.32.2. -4 -4 PC(32:2). Lipid metabolite [nM]
  \item PC.33.0...A. -4 -4 PC(33:0) (a). Lipid metabolite [nM]
  \item PC.33.0...B. -4 -4 PC(33:0) (b). Lipid metabolite [nM]
  \item PC.33.1. -4 -4 PC(33:1). Lipid metabolite [nM]
  \item PC.33.2. -4 -4 PC(33:2). Lipid metabolite [nM]
  \item PC.34.5. -4 -4 PC(34:5). Lipid metabolite [nM]
  \item PC.35.5. -4 -4 PC(35:5). Lipid metabolite [nM]
  \item PC.36.6...A. -4 -4 PC(36:6) (a). Lipid metabolite [nM]
  \item PC.38.2. -4 -4 PC(38:2). Lipid metabolite [nM]
  \item PC.38.4...B. -4 -4 PC(38:4) (b). Lipid metabolite [nM]
  \item PC.38.5...A. -4 -4 PC(38:5) (a). Lipid metabolite [nM]
  \item PC.38.5...B. -4 -4 PC(38:5) (b). Lipid metabolite [nM]
  \item PC.38.6...A. -4 -4 PC(38:6) (a). Lipid metabolite [nM]
  \item PC.38.7..C. -4 -4 PC(38:7)(c). Lipid metabolite [nM]
  \item PC.39.5..A. -4 -4 PC(39:5)(a). Lipid metabolite [nM]
  \item PC.39.5..B. -4 -4 PC(39:5)(b). Lipid metabolite [nM]
  \item PC.40.7...A. -4 -4 PC(40:7) (a). Lipid metabolite [nM]
  \item PC.40.8. -4 -4 PC(40:8). Lipid metabolite [nM]
  \item PC.44.12. -4 -4 PC(44:12). Lipid metabolite [nM]
  \item PC.O.16.0.16.0. -4 -4 PC(O-16:0/16:0). Lipid metabolite [nM]
  \item PC.O.16.0.20.3. -4 -4 PC(O-16:0/20:3). Lipid metabolite [nM]
  \item PC.O.16.0.20.4. -4 -4 PC(O-16:0/20:4). Lipid metabolite [nM]
  \item PC.O.16.0.22.6. -4 -4 PC(O-16:0/22:6). Lipid metabolite [nM]
  \item PC.O.18.0.18.1. -4 -4 PC(O-18:0/18:1). Lipid metabolite [nM]
  \item PC.O.18.0.18.2. -4 -4 PC(O-18:0/18:2). Lipid metabolite [nM]
  \item PC.O.18.0.20.4. -4 -4 PC(O-18:0/20:4). Lipid metabolite [nM]
  \item PC.O.18.0.22.6. -4 -4 PC(O-18:0/22:6). Lipid metabolite [nM]
  \item PC.O.18.1.18.1. -4 -4 PC(O-18:1/18:1). Lipid metabolite [nM]
  \item PC.O.18.1.18.2. -4 -4 PC(O-18:1/18:2). Lipid metabolite [nM]
  \item PC.O.32.1. -4 -4 PC(O-32:1). Lipid metabolite [nM]
  \item PC.O.32.2. -4 -4 PC(O-32:2). Lipid metabolite [nM]
  \item PC.O.34.1. -4 -4 PC(O-34:1). Lipid metabolite [nM]
  \item PC.O.34.2. -4 -4 PC(O-34:2). Lipid metabolite [nM]
  \item PC.O.34.4. -4 -4 PC(O-34:4). Lipid metabolite [nM]
  \item PC.O.35.4. -4 -4 PC(O-35:4). Lipid metabolite [nM]
  \item PC.O.36.0. -4 -4 PC(O-36:0). Lipid metabolite [nM]
  \item PC.O.36.5. -4 -4 PC(O-36:5). Lipid metabolite [nM]
  \item PC.O.38.5. -4 -4 PC(O-38:5). Lipid metabolite [nM]
  \item PC.O.40.5. -4 -4 PC(O-40:5). Lipid metabolite [nM]
  \item PC.P.15.0.20.4...A. -4 -4 PC(P-15:0/20:4) (a). Lipid metabolite [nM]
  \item PC.P.15.0.20.4...B. -4 -4 PC(P-15:0/20:4) (b). Lipid metabolite [nM]
  \item PC.P.16.0.14.0. -4 -4 PC(P-16:0/14:0). Lipid metabolite [nM]
  \item PC.P.16.0.16.0. -4 -4 PC(P-16:0/16:0). Lipid metabolite [nM]
  \item PC.P.16.0.16.1. -4 -4 PC(P-16:0/16:1). Lipid metabolite [nM]
  \item PC.P.16.0.18.0. -4 -4 PC(P-16:0/18:0). Lipid metabolite [nM]
  \item PC.P.16.0.18.1. -4 -4 PC(P-16:0/18:1). Lipid metabolite [nM]
  \item PC.P.16.0.18.2. -4 -4 PC(P-16:0/18:2). Lipid metabolite [nM]
  \item PC.P.16.0.18.3. -4 -4 PC(P-16:0/18:3). Lipid metabolite [nM]
  \item PC.P.16.0.20.4. -4 -4 PC(P-16:0/20:4). Lipid metabolite [nM]
  \item PC.P.16.0.20.5. -4 -4 PC(P-16:0/20:5). Lipid metabolite [nM]
  \item PC.P.16.0.22.6. -4 -4 PC(P-16:0/22:6). Lipid metabolite [nM]
  \item PC.P.17.0.20.4...A. -4 -4 PC(P-17:0/20:4) (a). Lipid metabolite [nM]
  \item PC.P.17.0.20.4...B. -4 -4 PC(P-17:0/20:4) (b). Lipid metabolite [nM]
  \item PC.P.18.0.18.2. -4 -4 PC(P-18:0/18:2). Lipid metabolite [nM]
  \item PC.P.18.0.20.4. -4 -4 PC(P-18:0/20:4). Lipid metabolite [nM]
  \item PC.P.18.0.22.5. -4 -4 PC(P-18:0/22:5). Lipid metabolite [nM]
  \item PC.P.18.0.22.6. -4 -4 PC(P-18:0/22:6). Lipid metabolite [nM]
  \item PC.P.18.1.18.1. -4 -4 PC(P-18:1/18:1). Lipid metabolite [nM]
  \item PC.P.18.1.22.6. -4 -4 PC(P-18:1/22:6). Lipid metabolite [nM]
  \item PC.P.20.0.20.4. -4 -4 PC(P-20:0/20:4). Lipid metabolite [nM]
  \item PC.P.35.2..A. -4 -4 PC(P-35:2)(a). Lipid metabolite [nM]
  \item PC.P.35.2..B. -4 -4 PC(P-35:2)(b). Lipid metabolite [nM]
  \item PC.P.36.3. -4 -4 PC(P-36:3). Lipid metabolite [nM]
  \item PC.P.38.5...A. -4 -4 PC(P-38:5) (a). Lipid metabolite [nM]
  \item PC.P.38.5...B. -4 -4 PC(P-38:5) (b). Lipid metabolite [nM]
  \item LPC.14.0...SN1. -4 -4 LPC(14:0) [sn1]. Lipid metabolite [nM]
  \item LPC.14.0...SN2. -4 -4 LPC(14:0) [sn2]. Lipid metabolite [nM]
  \item LPC.15.MHDA...SN1...104_SN1. -4 -4 LPC(15-MHDA) [sn1] [104_sn1]. Lipid metabolite [nM]
  \item LPC.15.MHDA...SN1....LPC.17.0...SN2. -4 -4 LPC(15-MHDA) [sn1] / LPC(17:0) [sn2]. Lipid metabolite [nM]
  \item LPC.15.MHDA...SN2. -4 -4 LPC(15-MHDA) [sn2]. Lipid metabolite [nM]
  \item LPC.15.0...SN1. -4 -4 LPC(15:0) [sn1]. Lipid metabolite [nM]
  \item LPC.15.0...SN2. -4 -4 LPC(15:0) [sn2]. Lipid metabolite [nM]
  \item LPC.16.0...SN1. -4 -4 LPC(16:0) [sn1]. Lipid metabolite [nM]
  \item LPC.16.0...SN2. -4 -4 LPC(16:0) [sn2]. Lipid metabolite [nM]
  \item LPC.16.1...SN1. -4 -4 LPC(16:1) [sn1]. Lipid metabolite [nM]
  \item LPC.16.1...SN2. -4 -4 LPC(16:1) [sn2]. Lipid metabolite [nM]
  \item LPC.17.0...SN1. -4 -4 LPC(17:0) [sn1]. Lipid metabolite [nM]
  \item LPC.17.1...A...SN1...104_SN1. -4 -4 LPC(17:1) (a) [sn1] [104_sn1]. Lipid metabolite [nM]
  \item LPC.17.1...SN1...A....LPC.17.1...SN2...B. -4 -4 LPC(17:1) [sn1] (a) / LPC(17:1) [sn2] (b). Lipid metabolite [nM]
  \item LPC.17.1...SN1...B. -4 -4 LPC(17:1) [sn1] (b). Lipid metabolite [nM]
  \item LPC.17.1...SN2...A. -4 -4 LPC(17:1) [sn2] (a). Lipid metabolite [nM]
  \item LPC.18.0...SN1. -4 -4 LPC(18:0) [sn1]. Lipid metabolite [nM]
  \item LPC.18.0...SN2. -4 -4 LPC(18:0) [sn2]. Lipid metabolite [nM]
  \item LPC.18.1...SN1. -4 -4 LPC(18:1) [sn1]. Lipid metabolite [nM]
  \item LPC.18.1...SN2. -4 -4 LPC(18:1) [sn2]. Lipid metabolite [nM]
  \item LPC.18.2...SN1. -4 -4 LPC(18:2) [sn1]. Lipid metabolite [nM]
  \item LPC.18.2...SN2. -4 -4 LPC(18:2) [sn2]. Lipid metabolite [nM]
  \item LPC.18.3...A...SN1...104_SN1. -4 -4 LPC(18:3) (a) [sn1] [104_sn1]. Lipid metabolite [nM]
  \item LPC.18.3...SN1...A..LPC.18.3...SN2...B. -4 -4 LPC(18:3) [sn1] (a)/LPC(18:3) [sn2] (b). Lipid metabolite [nM]
  \item LPC.18.3...SN1...B. -4 -4 LPC(18:3) [sn1] (b). Lipid metabolite [nM]
  \item LPC.18.3...SN2...A. -4 -4 LPC(18:3) [sn2] (a). Lipid metabolite [nM]
  \item LPC.19.0...A...SN1...104_SN1. -4 -4 LPC(19:0) (a) [sn1] [104_sn1]. Lipid metabolite [nM]
  \item LPC.19.0...SN1...A....LPC.19.0...SN2...B. -4 -4 LPC(19:0) [sn1] (a) / LPC(19:0) [sn2] (b). Lipid metabolite [nM]
  \item LPC.19.0...SN1...B. -4 -4 LPC(19:0) [sn1] (b). Lipid metabolite [nM]
  \item LPC.19.0...SN2...A. -4 -4 LPC(19:0) [sn2] (a). Lipid metabolite [nM]
  \item LPC.19.1...A. -4 -4 LPC(19:1) (a). Lipid metabolite [nM]
  \item LPC.19.1...B. -4 -4 LPC(19:1) (b). Lipid metabolite [nM]
  \item LPC.19.1...C. -4 -4 LPC(19:1) (c). Lipid metabolite [nM]
  \item LPC.20.0...SN1. -4 -4 LPC(20:0) [sn1]. Lipid metabolite [nM]
  \item LPC.20.0...SN2. -4 -4 LPC(20:0) [sn2]. Lipid metabolite [nM]
  \item LPC.20.1...SN1. -4 -4 LPC(20:1) [sn1]. Lipid metabolite [nM]
  \item LPC.20.1...SN2. -4 -4 LPC(20:1) [sn2]. Lipid metabolite [nM]
  \item LPC.20.2...SN1. -4 -4 LPC(20:2) [sn1]. Lipid metabolite [nM]
  \item LPC.20.2...SN2. -4 -4 LPC(20:2) [sn2]. Lipid metabolite [nM]
  \item LPC.20.3...SN1. -4 -4 LPC(20:3) [sn1]. Lipid metabolite [nM]
  \item LPC.20.3...SN2. -4 -4 LPC(20:3) [sn2]. Lipid metabolite [nM]
  \item LPC.20.4...SN1. -4 -4 LPC(20:4) [sn1]. Lipid metabolite [nM]
  \item LPC.20.4...SN2. -4 -4 LPC(20:4) [sn2]. Lipid metabolite [nM]
  \item LPC.20.5...SN1. -4 -4 LPC(20:5) [sn1]. Lipid metabolite [nM]
  \item LPC.20.5...SN2. -4 -4 LPC(20:5) [sn2]. Lipid metabolite [nM]
  \item LPC.22.0...SN1. -4 -4 LPC(22:0) [sn1]. Lipid metabolite [nM]
  \item LPC.22.0...SN2. -4 -4 LPC(22:0) [sn2]. Lipid metabolite [nM]
  \item LPC.22.1...SN1. -4 -4 LPC(22:1) [sn1]. Lipid metabolite [nM]
  \item LPC.22.1...SN2. -4 -4 LPC(22:1) [sn2]. Lipid metabolite [nM]
  \item LPC.22.4...SN1. -4 -4 LPC(22:4) [sn1]. Lipid metabolite [nM]
  \item LPC.22.4...SN2. -4 -4 LPC(22:4) [sn2]. Lipid metabolite [nM]
  \item LPC.22.5...N3...SN1...104_SN1. -4 -4 LPC(22:5) (n3) [sn1] [104_sn1]. Lipid metabolite [nM]
  \item LPC.22.5...SN1...N3..LPC.22.5...SN2...N6. -4 -4 LPC(22:5) [sn1] (n3)/LPC(22:5) [sn2] (n6). Lipid metabolite [nM]
  \item LPC.22.5...SN1...N6. -4 -4 LPC(22:5) [sn1] (n6). Lipid metabolite [nM]
  \item LPC.22.5...SN2...N3. -4 -4 LPC(22:5) [sn2] (n3). Lipid metabolite [nM]
  \item LPC.22.6...SN1. -4 -4 LPC(22:6) [sn1]. Lipid metabolite [nM]
  \item LPC.22.6...SN2. -4 -4 LPC(22:6) [sn2]. Lipid metabolite [nM]
  \item LPC.24.0...SN1. -4 -4 LPC(24:0) [sn1]. Lipid metabolite [nM]
  \item LPC.24.0...SN2. -4 -4 LPC(24:0) [sn2]. Lipid metabolite [nM]
  \item LPC.26.0...SN1. -4 -4 LPC(26:0) [sn1]. Lipid metabolite [nM]
  \item LPC.26.0...SN2. -4 -4 LPC(26:0) [sn2]. Lipid metabolite [nM]
  \item LPC.O.16.0. -4 -4 LPC(O-16:0). Lipid metabolite [nM]
  \item LPC.O.18.0. -4 -4 LPC(O-18:0). Lipid metabolite [nM]
  \item LPC.O.18.1. -4 -4 LPC(O-18:1). Lipid metabolite [nM]
  \item LPC.O.20.0. -4 -4 LPC(O-20:0). Lipid metabolite [nM]
  \item LPC.O.20.1. -4 -4 LPC(O-20:1). Lipid metabolite [nM]
  \item LPC.O.22.0. -4 -4 LPC(O-22:0). Lipid metabolite [nM]
  \item LPC.O.22.1. -4 -4 LPC(O-22:1). Lipid metabolite [nM]
  \item LPC.O.24.0. -4 -4 LPC(O-24:0). Lipid metabolite [nM]
  \item LPC.O.24.1. -4 -4 LPC(O-24:1). Lipid metabolite [nM]
  \item LPC.O.24.2. -4 -4 LPC(O-24:2). Lipid metabolite [nM]
  \item LPC.P.16.0. -4 -4 LPC(P-16:0). Lipid metabolite [nM]
  \item LPC.P.17.0...A. -4 -4 LPC(P-17:0) (a). Lipid metabolite [nM]
  \item LPC.P.17.0...B. -4 -4 LPC(P-17:0) (b). Lipid metabolite [nM]
  \item LPC.P.18.0. -4 -4 LPC(P-18:0). Lipid metabolite [nM]
  \item LPC.P.18.1. -4 -4 LPC(P-18:1). Lipid metabolite [nM]
  \item LPC.P.20.0. -4 -4 LPC(P-20:0). Lipid metabolite [nM]
  \item PE.15.MHDA_18.1. -4 -4 PE(15-MHDA_18:1). Lipid metabolite [nM]
  \item PE.15.MHDA_18.2. -4 -4 PE(15-MHDA_18:2). Lipid metabolite [nM]
  \item PE.15.MHDA_20.4. -4 -4 PE(15-MHDA_20:4). Lipid metabolite [nM]
  \item PE.15.MHDA_22.6. -4 -4 PE(15-MHDA_22:6). Lipid metabolite [nM]
  \item PE.16.0_16.0. -4 -4 PE(16:0_16:0). Lipid metabolite [nM]
  \item PE.16.0_16.1. -4 -4 PE(16:0_16:1). Lipid metabolite [nM]
  \item PE.16.0_18.1. -4 -4 PE(16:0_18:1). Lipid metabolite [nM]
  \item PE.16.0_18.2. -4 -4 PE(16:0_18:2). Lipid metabolite [nM]
  \item PE.16.0_18.3...A. -4 -4 PE(16:0_18:3) (a). Lipid metabolite [nM]
  \item PE.16.0_18.3...B. -4 -4 PE(16:0_18:3) (b). Lipid metabolite [nM]
  \item PE.16.0_20.3. -4 -4 PE(16:0_20:3). Lipid metabolite [nM]
  \item PE.16.0_20.4. -4 -4 PE(16:0_20:4). Lipid metabolite [nM]
  \item PE.16.0_20.5. -4 -4 PE(16:0_20:5). Lipid metabolite [nM]
  \item PE.16.0_22.6. -4 -4 PE(16:0_22:6). Lipid metabolite [nM]
  \item PE.16.1_18.2. -4 -4 PE(16:1_18:2). Lipid metabolite [nM]
  \item PE.16.1_20.4. -4 -4 PE(16:1_20:4). Lipid metabolite [nM]
  \item PE.17.0_18.1. -4 -4 PE(17:0_18:1). Lipid metabolite [nM]
  \item PE.17.0_18.2. -4 -4 PE(17:0_18:2). Lipid metabolite [nM]
  \item PE.17.0_20.4. -4 -4 PE(17:0_20:4). Lipid metabolite [nM]
  \item PE.17.0_22.6. -4 -4 PE(17:0_22:6). Lipid metabolite [nM]
  \item PE.18.0_18.1. -4 -4 PE(18:0_18:1). Lipid metabolite [nM]
  \item PE.18.0_18.2. -4 -4 PE(18:0_18:2). Lipid metabolite [nM]
  \item PE.18.0_20.3...A. -4 -4 PE(18:0_20:3) (a). Lipid metabolite [nM]
  \item PE.18.0_20.3...B. -4 -4 PE(18:0_20:3) (b). Lipid metabolite [nM]
  \item PE.18.0_20.4. -4 -4 PE(18:0_20:4). Lipid metabolite [nM]
  \item PE.18.0_22.4. -4 -4 PE(18:0_22:4). Lipid metabolite [nM]
  \item PE.18.0_22.5...N3. -4 -4 PE(18:0_22:5) (n3). Lipid metabolite [nM]
  \item PE.18.0_22.5...N6. -4 -4 PE(18:0_22:5) (n6). Lipid metabolite [nM]
  \item PE.18.0_22.6. -4 -4 PE(18:0_22:6). Lipid metabolite [nM]
  \item PE.18.1_18.1. -4 -4 PE(18:1_18:1). Lipid metabolite [nM]
  \item PE.18.1_18.2. -4 -4 PE(18:1_18:2). Lipid metabolite [nM]
  \item PE.18.1_22.6...A. -4 -4 PE(18:1_22:6) (a). Lipid metabolite [nM]
  \item PE.18.1_22.6...B. -4 -4 PE(18:1_22:6) (b). Lipid metabolite [nM]
  \item PE.20.0_20.4. -4 -4 PE(20:0_20:4). Lipid metabolite [nM]
  \item PE.36.0. -4 -4 PE(36:0). Lipid metabolite [nM]
  \item PE.38.5...A. -4 -4 PE(38:5) (a). Lipid metabolite [nM]
  \item PE.38.5...B. -4 -4 PE(38:5) (b). Lipid metabolite [nM]
  \item PE.O.16.0.18.2. -4 -4 PE(O-16:0/18:2). Lipid metabolite [nM]
  \item PE.O.16.0.20.3. -4 -4 PE(O-16:0/20:3). Lipid metabolite [nM]
  \item PE.O.16.0.20.4. -4 -4 PE(O-16:0/20:4). Lipid metabolite [nM]
  \item PE.O.16.0.22.4. -4 -4 PE(O-16:0/22:4). Lipid metabolite [nM]
  \item PE.O.16.0.22.6. -4 -4 PE(O-16:0/22:6). Lipid metabolite [nM]
  \item PE.O.18.0.20.4. -4 -4 PE(O-18:0/20:4). Lipid metabolite [nM]
  \item PE.O.18.0.22.5. -4 -4 PE(O-18:0/22:5). Lipid metabolite [nM]
  \item PE.O.18.0.22.6. -4 -4 PE(O-18:0/22:6). Lipid metabolite [nM]
  \item PE.O.18.1.18.2. -4 -4 PE(O-18:1/18:2). Lipid metabolite [nM]
  \item PE.O.18.1.22.6. -4 -4 PE(O-18:1/22:6). Lipid metabolite [nM]
  \item PE.O.34.1. -4 -4 PE(O-34:1). Lipid metabolite [nM]
  \item PE.O.36.5. -4 -4 PE(O-36:5). Lipid metabolite [nM]
  \item PE.O.38.5...A. -4 -4 PE(O-38:5) (a). Lipid metabolite [nM]
  \item PE.O.38.5...B. -4 -4 PE(O-38:5) (b). Lipid metabolite [nM]
  \item PE.P.15.0.20.4...A. -4 -4 PE(P-15:0/20:4) (a). Lipid metabolite [nM]
  \item PE.P.15.0.20.4...B. -4 -4 PE(P-15:0/20:4) (b). Lipid metabolite [nM]
  \item PE.P.15.0.22.6...A. -4 -4 PE(P-15:0/22:6) (a). Lipid metabolite [nM]
  \item PE.P.15.0.22.6...B. -4 -4 PE(P-15:0/22:6) (b). Lipid metabolite [nM]
  \item PE.P.16.0.18.1. -4 -4 PE(P-16:0/18:1). Lipid metabolite [nM]
  \item PE.P.16.0.18.2. -4 -4 PE(P-16:0/18:2). Lipid metabolite [nM]
  \item PE.P.16.0.18.3. -4 -4 PE(P-16:0/18:3). Lipid metabolite [nM]
  \item PE.P.16.0.20.3...A. -4 -4 PE(P-16:0/20:3) (a). Lipid metabolite [nM]
  \item PE.P.16.0.20.3...B. -4 -4 PE(P-16:0/20:3) (b). Lipid metabolite [nM]
  \item PE.P.16.0.20.4. -4 -4 PE(P-16:0/20:4). Lipid metabolite [nM]
  \item PE.P.16.0.20.5. -4 -4 PE(P-16:0/20:5). Lipid metabolite [nM]
  \item PE.P.16.0.22.4. -4 -4 PE(P-16:0/22:4). Lipid metabolite [nM]
  \item PE.P.16.0.22.5...N3. -4 -4 PE(P-16:0/22:5) (n3). Lipid metabolite [nM]
  \item PE.P.16.0.22.5...N6. -4 -4 PE(P-16:0/22:5) (n6). Lipid metabolite [nM]
  \item PE.P.16.0.22.6. -4 -4 PE(P-16:0/22:6). Lipid metabolite [nM]
  \item PE.P.17.0.20.4...A. -4 -4 PE(P-17:0/20:4) (a). Lipid metabolite [nM]
  \item PE.P.17.0.20.4...B. -4 -4 PE(P-17:0/20:4) (b). Lipid metabolite [nM]
  \item PE.P.17.0.22.6...A. -4 -4 PE(P-17:0/22:6) (a). Lipid metabolite [nM]
  \item PE.P.17.0.22.6...B. -4 -4 PE(P-17:0/22:6) (b). Lipid metabolite [nM]
  \item PE.P.18.0.18.1. -4 -4 PE(P-18:0/18:1). Lipid metabolite [nM]
  \item PE.P.18.0.18.2. -4 -4 PE(P-18:0/18:2). Lipid metabolite [nM]
  \item PE.P.18.0.18.3. -4 -4 PE(P-18:0/18:3). Lipid metabolite [nM]
  \item PE.P.18.0.20.3...A. -4 -4 PE(P-18:0/20:3) (a). Lipid metabolite [nM]
  \item PE.P.18.0.20.3...B. -4 -4 PE(P-18:0/20:3) (b). Lipid metabolite [nM]
  \item PE.P.18.0.20.4. -4 -4 PE(P-18:0/20:4). Lipid metabolite [nM]
  \item PE.P.18.0.20.5. -4 -4 PE(P-18:0/20:5). Lipid metabolite [nM]
  \item PE.P.18.0.22.4. -4 -4 PE(P-18:0/22:4). Lipid metabolite [nM]
  \item PE.P.18.0.22.5...N3. -4 -4 PE(P-18:0/22:5) (n3). Lipid metabolite [nM]
  \item PE.P.18.0.22.5...N6. -4 -4 PE(P-18:0/22:5) (n6). Lipid metabolite [nM]
  \item PE.P.18.0.22.6. -4 -4 PE(P-18:0/22:6). Lipid metabolite [nM]
  \item PE.P.18.1.18.1...A. -4 -4 PE(P-18:1/18:1) (a). Lipid metabolite [nM]
  \item PE.P.18.1.18.1...B. -4 -4 PE(P-18:1/18:1) (b). Lipid metabolite [nM]
  \item PE.P.18.1.18.2...A. -4 -4 PE(P-18:1/18:2) (a). Lipid metabolite [nM]
  \item PE.P.18.1.18.2...B. -4 -4 PE(P-18:1/18:2) (b). Lipid metabolite [nM]
  \item PE.P.18.1.18.3. -4 -4 PE(P-18:1/18:3). Lipid metabolite [nM]
  \item PE.P.18.1.20.3...A. -4 -4 PE(P-18:1/20:3) (a). Lipid metabolite [nM]
  \item PE.P.18.1.20.3...B. -4 -4 PE(P-18:1/20:3) (b). Lipid metabolite [nM]
  \item PE.P.18.1.20.4...A. -4 -4 PE(P-18:1/20:4) (a). Lipid metabolite [nM]
  \item PE.P.18.1.20.4...B. -4 -4 PE(P-18:1/20:4) (b). Lipid metabolite [nM]
  \item PE.P.18.1.20.5...A. -4 -4 PE(P-18:1/20:5) (a). Lipid metabolite [nM]
  \item PE.P.18.1.20.5...B. -4 -4 PE(P-18:1/20:5) (b). Lipid metabolite [nM]
  \item PE.P.18.1.22.4. -4 -4 PE(P-18:1/22:4). Lipid metabolite [nM]
  \item PE.P.18.1.22.5...A. -4 -4 PE(P-18:1/22:5) (a). Lipid metabolite [nM]
  \item PE.P.18.1.22.5...B. -4 -4 PE(P-18:1/22:5) (b). Lipid metabolite [nM]
  \item PE.P.18.1.22.6...A. -4 -4 PE(P-18:1/22:6) (a). Lipid metabolite [nM]
  \item PE.P.18.1.22.6...B. -4 -4 PE(P-18:1/22:6) (b). Lipid metabolite [nM]
  \item PE.P.19.0.20.4...A. -4 -4 PE(P-19:0/20:4) (a). Lipid metabolite [nM]
  \item PE.P.19.0.20.4...B. -4 -4 PE(P-19:0/20:4) (b). Lipid metabolite [nM]
  \item PE.P.20.0.18.1. -4 -4 PE(P-20:0/18:1). Lipid metabolite [nM]
  \item PE.P.20.0.18.2. -4 -4 PE(P-20:0/18:2). Lipid metabolite [nM]
  \item PE.P.20.0.20.4. -4 -4 PE(P-20:0/20:4). Lipid metabolite [nM]
  \item PE.P.20.0.22.6. -4 -4 PE(P-20:0/22:6). Lipid metabolite [nM]
  \item PE.P.20.1.20.4. -4 -4 PE(P-20:1/20:4). Lipid metabolite [nM]
  \item PE.P.20.1.22.6. -4 -4 PE(P-20:1/22:6). Lipid metabolite [nM]
  \item LPE.16.0...SN1. -4 -4 LPE(16:0) [sn1]. Lipid metabolite [nM]
  \item LPE.16.0...SN2. -4 -4 LPE(16:0) [sn2]. Lipid metabolite [nM]
  \item LPE.18.0...SN1. -4 -4 LPE(18:0) [sn1]. Lipid metabolite [nM]
  \item LPE.18.0...SN2. -4 -4 LPE(18:0) [sn2]. Lipid metabolite [nM]
  \item LPE.18.1...SN1. -4 -4 LPE(18:1) [sn1]. Lipid metabolite [nM]
  \item LPE.18.1...SN2. -4 -4 LPE(18:1) [sn2]. Lipid metabolite [nM]
  \item LPE.18.2...SN1. -4 -4 LPE(18:2) [sn1]. Lipid metabolite [nM]
  \item LPE.18.2...SN2. -4 -4 LPE(18:2) [sn2]. Lipid metabolite [nM]
  \item LPE.20.4...SN1. -4 -4 LPE(20:4) [sn1]. Lipid metabolite [nM]
  \item LPE.20.4...SN2. -4 -4 LPE(20:4) [sn2]. Lipid metabolite [nM]
  \item LPE.22.6...SN1. -4 -4 LPE(22:6) [sn1]. Lipid metabolite [nM]
  \item LPE.22.6...SN2. -4 -4 LPE(22:6) [sn2]. Lipid metabolite [nM]
  \item LPE.P.16.0. -4 -4 LPE(P-16:0). Lipid metabolite [nM]
  \item LPE.P.18.0. -4 -4 LPE(P-18:0). Lipid metabolite [nM]
  \item LPE.P.18.1. -4 -4 LPE(P-18:1). Lipid metabolite [nM]
  \item LPE.P.20.0. -4 -4 LPE(P-20:0). Lipid metabolite [nM]
  \item PI..38.5...B. -4 -4 PI (38:5) (b). Lipid metabolite [nM]
  \item PI.15.MHDA_18.1..PI.17.0_18.1. -4 -4 PI(15-MHDA_18:1)/PI(17:0_18:1). Lipid metabolite [nM]
  \item PI.15.MHDA_18.2..PI.17.0_18.2. -4 -4 PI(15-MHDA_18:2)/PI(17:0_18:2). Lipid metabolite [nM]
  \item PI.15.MHDA_20.4..PI.17.0_20.4. -4 -4 PI(15-MHDA_20:4)/PI(17:0_20:4). Lipid metabolite [nM]
  \item PI.16.0_16.1. -4 -4 PI(16:0_16:1). Lipid metabolite [nM]
  \item PI.16.0_20.3...A. -4 -4 PI(16:0_20:3) (a). Lipid metabolite [nM]
  \item PI.16.0_20.3...B. -4 -4 PI(16:0_20:3) (b). Lipid metabolite [nM]
  \item PI.16.0_20.4. -4 -4 PI(16:0_20:4). Lipid metabolite [nM]
  \item PI.16.0.16.0. -4 -4 PI(16:0/16:0). Lipid metabolite [nM]
  \item PI.18.0_18.1. -4 -4 PI(18:0_18:1). Lipid metabolite [nM]
  \item PI.18.0_20.2. -4 -4 PI(18:0_20:2). Lipid metabolite [nM]
  \item PI.18.0_20.3...A. -4 -4 PI(18:0_20:3) (a). Lipid metabolite [nM]
  \item PI.18.0_20.3...B. -4 -4 PI(18:0_20:3) (b). Lipid metabolite [nM]
  \item PI.18.0_20.4. -4 -4 PI(18:0_20:4). Lipid metabolite [nM]
  \item PI.18.0_22.4. -4 -4 PI(18:0_22:4). Lipid metabolite [nM]
  \item PI.18.0_22.5...N3. -4 -4 PI(18:0_22:5) (n3). Lipid metabolite [nM]
  \item PI.18.0_22.5...N6. -4 -4 PI(18:0_22:5) (n6). Lipid metabolite [nM]
  \item PI.18.0_22.6. -4 -4 PI(18:0_22:6). Lipid metabolite [nM]
  \item PI.18.1_18.2. -4 -4 PI(18:1_18:2). Lipid metabolite [nM]
  \item PI.20.0_20.4. -4 -4 PI(20:0_20:4). Lipid metabolite [nM]
  \item PI.34.0. -4 -4 PI(34:0). Lipid metabolite [nM]
  \item PI.34.1. -4 -4 PI(34:1). Lipid metabolite [nM]
  \item PI.36.2. -4 -4 PI(36:2). Lipid metabolite [nM]
  \item PI.37.6. -4 -4 PI(37:6). Lipid metabolite [nM]
  \item PI.38.5...A. -4 -4 PI(38:5) (a). Lipid metabolite [nM]
  \item PI.38.6. -4 -4 PI(38:6). Lipid metabolite [nM]
  \item PI.39.6. -4 -4 PI(39:6). Lipid metabolite [nM]
  \item PIP1.38.4. -4 -4 PIP1(38:4). Lipid metabolite [nM]
  \item LPI.18.0...SN1. -4 -4 LPI(18:0) [sn1]. Lipid metabolite [nM]
  \item LPI.18.0...SN2. -4 -4 LPI(18:0) [sn2]. Lipid metabolite [nM]
  \item LPI.18.1...SN1. -4 -4 LPI(18:1) [sn1]. Lipid metabolite [nM]
  \item LPI.18.1...SN2. -4 -4 LPI(18:1) [sn2]. Lipid metabolite [nM]
  \item LPI.18.2...SN1. -4 -4 LPI(18:2) [sn1]. Lipid metabolite [nM]
  \item LPI.18.2...SN2. -4 -4 LPI(18:2) [sn2]. Lipid metabolite [nM]
  \item LPI.20.4...SN1. -4 -4 LPI(20:4) [sn1]. Lipid metabolite [nM]
  \item LPI.20.4...SN2. -4 -4 LPI(20:4) [sn2]. Lipid metabolite [nM]
  \item PS.36.1. -4 -4 PS(36:1). Lipid metabolite [nM]
  \item PS.36.2. -4 -4 PS(36:2). Lipid metabolite [nM]
  \item PS.38.3. -4 -4 PS(38:3). Lipid metabolite [nM]
  \item PS.38.4. -4 -4 PS(38:4). Lipid metabolite [nM]
  \item PS.40.5. -4 -4 PS(40:5). Lipid metabolite [nM]
  \item PS.40.6. -4 -4 PS(40:6). Lipid metabolite [nM]
  \item PG.34.1. -4 -4 PG(34:1). Lipid metabolite [nM]
  \item PG.36.1. -4 -4 PG(36:1). Lipid metabolite [nM]
  \item PG.36.2. -4 -4 PG(36:2). Lipid metabolite [nM]
  \item CE.14.0. -4 -4 CE(14:0). Lipid metabolite [nM]
  \item CE.15.0. -4 -4 CE(15:0). Lipid metabolite [nM]
  \item CE.16.0. -4 -4 CE(16:0). Lipid metabolite [nM]
  \item CE.16.1. -4 -4 CE(16:1). Lipid metabolite [nM]
  \item CE.16.2. -4 -4 CE(16:2). Lipid metabolite [nM]
  \item CE.17.0. -4 -4 CE(17:0). Lipid metabolite [nM]
  \item CE.17.1. -4 -4 CE(17:1). Lipid metabolite [nM]
  \item CE.18.0. -4 -4 CE(18:0). Lipid metabolite [nM]
  \item CE.18.1. -4 -4 CE(18:1). Lipid metabolite [nM]
  \item CE.18.2. -4 -4 CE(18:2). Lipid metabolite [nM]
  \item CE.18.3. -4 -4 CE(18:3). Lipid metabolite [nM]
  \item CE.20.0. -4 -4 CE(20:0). Lipid metabolite [nM]
  \item CE.20.1. -4 -4 CE(20:1). Lipid metabolite [nM]
  \item CE.20.2. -4 -4 CE(20:2). Lipid metabolite [nM]
  \item CE.20.3. -4 -4 CE(20:3). Lipid metabolite [nM]
  \item CE.20.4. -4 -4 CE(20:4). Lipid metabolite [nM]
  \item CE.20.5. -4 -4 CE(20:5). Lipid metabolite [nM]
  \item CE.22.0. -4 -4 CE(22:0). Lipid metabolite [nM]
  \item CE.22.1. -4 -4 CE(22:1). Lipid metabolite [nM]
  \item CE.22.4. -4 -4 CE(22:4). Lipid metabolite [nM]
  \item CE.22.5. -4 -4 CE(22:5). Lipid metabolite [nM]
  \item CE.22.6. -4 -4 CE(22:6). Lipid metabolite [nM]
  \item CE.24.0. -4 -4 CE(24:0). Lipid metabolite [nM]
  \item CE.24.1. -4 -4 CE(24:1). Lipid metabolite [nM]
  \item CE.24.4. -4 -4 CE(24:4). Lipid metabolite [nM]
  \item CE.24.5. -4 -4 CE(24:5). Lipid metabolite [nM]
  \item CE.24.6. -4 -4 CE(24:6). Lipid metabolite [nM]
  \item COH -4 -4 COH. Lipid metabolite [nM]
  \item DE.16.0. -4 -4 DE(16:0). Lipid metabolite [nM]
  \item DE.18.1. -4 -4 DE(18:1). Lipid metabolite [nM]
  \item DE.18.2. -4 -4 DE(18:2). Lipid metabolite [nM]
  \item DE.20.4. -4 -4 DE(20:4). Lipid metabolite [nM]
  \item DE.20.5. -4 -4 DE(20:5). Lipid metabolite [nM]
  \item DE.22.6. -4 -4 DE(22:6). Lipid metabolite [nM]
  \item DEDE.18.2. -4 -4 deDE(18:2). Lipid metabolite [nM]
  \item DEDE.20.4. -4 -4 deDE(20:4). Lipid metabolite [nM]
  \item METHYL.CE.18.0. -4 -4 methyl-CE(18:0). Lipid metabolite [nM]
  \item METHYL.CE.18.1. -4 -4 methyl-CE(18:1). Lipid metabolite [nM]
  \item METHYL.CE.18.2. -4 -4 methyl-CE(18:2). Lipid metabolite [nM]
  \item METHYL.CE.20.4. -4 -4 methyl-CE(20:4). Lipid metabolite [nM]
  \item METHYL.CE.22.6. -4 -4 methyl-CE(22:6). Lipid metabolite [nM]
  \item METHYL.DE.18.1. -4 -4 methyl-DE(18:1). Lipid metabolite [nM]
  \item METHYL.DE.18.2. -4 -4 methyl-DE(18:2). Lipid metabolite [nM]
  \item DIMETHYL.CE.18.1. -4 -4 dimethyl-CE(18:1). Lipid metabolite [nM]
  \item DIMETHYL.CE.18.2. -4 -4 dimethyl-CE(18:2). Lipid metabolite [nM]
  \item DIMETHYL.CE.20.4. -4 -4 dimethyl-CE(20:4). Lipid metabolite [nM]
  \item DIMETHYL.CE.22.6. -4 -4 dimethyl-CE(22:6). Lipid metabolite [nM]
  \item FA.14.0. -4 -4 FA(14:0). Lipid metabolite [nM]
  \item FA.16.0. -4 -4 FA(16:0). Lipid metabolite [nM]
  \item FA.16.1. -4 -4 FA(16:1). Lipid metabolite [nM]
  \item FA.17.0. -4 -4 FA(17:0). Lipid metabolite [nM]
  \item FA.17.1. -4 -4 FA(17:1). Lipid metabolite [nM]
  \item FA.18.0. -4 -4 FA(18:0). Lipid metabolite [nM]
  \item FA.18.1. -4 -4 FA(18:1). Lipid metabolite [nM]
  \item FA.18.2. -4 -4 FA(18:2). Lipid metabolite [nM]
  \item FA.18.3. -4 -4 FA(18:3). Lipid metabolite [nM]
  \item FA.20.2. -4 -4 FA(20:2). Lipid metabolite [nM]
  \item FA.20.3. -4 -4 FA(20:3). Lipid metabolite [nM]
  \item FA.20.4. -4 -4 FA(20:4). Lipid metabolite [nM]
  \item FA.20.5. -4 -4 FA(20:5). Lipid metabolite [nM]
  \item FA.22.4. -4 -4 FA(22:4). Lipid metabolite [nM]
  \item FA.22.5. -4 -4 FA(22:5). Lipid metabolite [nM]
  \item FA.22.6. -4 -4 FA(22:6). Lipid metabolite [nM]
  \item AC.12.0. -4 -4 AC(12:0). Lipid metabolite [nM]
  \item AC.12.1. -4 -4 AC(12:1). Lipid metabolite [nM]
  \item AC.13.0. -4 -4 AC(13:0). Lipid metabolite [nM]
  \item AC.14.0. -4 -4 AC(14:0). Lipid metabolite [nM]
  \item AC.14.1. -4 -4 AC(14:1). Lipid metabolite [nM]
  \item AC.14.2. -4 -4 AC(14:2). Lipid metabolite [nM]
  \item AC.15.0...A. -4 -4 AC(15:0) (a). Lipid metabolite [nM]
  \item AC.15.0...B. -4 -4 AC(15:0) (b). Lipid metabolite [nM]
  \item AC.16.0. -4 -4 AC(16:0). Lipid metabolite [nM]
  \item AC.16.1. -4 -4 AC(16:1). Lipid metabolite [nM]
  \item AC.17.0...A. -4 -4 AC(17:0) (a). Lipid metabolite [nM]
  \item AC.17.0...B. -4 -4 AC(17:0) (b). Lipid metabolite [nM]
  \item AC.18.0. -4 -4 AC(18:0). Lipid metabolite [nM]
  \item AC.18.1. -4 -4 AC(18:1). Lipid metabolite [nM]
  \item AC.18.2. -4 -4 AC(18:2). Lipid metabolite [nM]
  \item AC.18.3. -4 -4 AC(18:3). Lipid metabolite [nM]
  \item AC.20.3...A. -4 -4 AC(20:3) (a). Lipid metabolite [nM]
  \item AC.20.3...B. -4 -4 AC(20:3) (b). Lipid metabolite [nM]
  \item AC.20.4. -4 -4 AC(20:4). Lipid metabolite [nM]
  \item AC.20.5. -4 -4 AC(20:5). Lipid metabolite [nM]
  \item AC.22.5. -4 -4 AC(22:5). Lipid metabolite [nM]
  \item AC.22.6. -4 -4 AC(22:6). Lipid metabolite [nM]
  \item AC.24.0. -4 -4 AC(24:0). Lipid metabolite [nM]
  \item AC.24.1. -4 -4 AC(24:1). Lipid metabolite [nM]
  \item AC.26.0. -4 -4 AC(26:0). Lipid metabolite [nM]
  \item AC.26.1. -4 -4 AC(26:1). Lipid metabolite [nM]
  \item AC.14.0..OH -4 -4 AC(14:0)-OH. Lipid metabolite [nM]
  \item AC.14.1..OH -4 -4 AC(14:1)-OH. Lipid metabolite [nM]
  \item AC.16.0..OH -4 -4 AC(16:0)-OH. Lipid metabolite [nM]
  \item AC.16.1..OH -4 -4 AC(16:1)-OH. Lipid metabolite [nM]
  \item AC.18.0..OH -4 -4 AC(18:0)-OH. Lipid metabolite [nM]
  \item AC.18.1..OH -4 -4 AC(18:1)-OH. Lipid metabolite [nM]
  \item AC.20.3..OH -4 -4 AC(20:3)-OH. Lipid metabolite [nM]
  \item AC.22.5..OH -4 -4 AC(22:5)-OH. Lipid metabolite [nM]
  \item AC.24.1..OH -4 -4 AC(24:1)-OH. Lipid metabolite [nM]
  \item CA -4 -4 CA. Lipid metabolite [nM]
  \item DXCA -4 -4 dxCA. Lipid metabolite [nM]
  \item DG.14.0_16.0. -4 -4 DG(14:0_16:0). Lipid metabolite [nM]
  \item DG.14.0_18.2. -4 -4 DG(14:0_18:2). Lipid metabolite [nM]
  \item DG.16.0_16.0. -4 -4 DG(16:0_16:0). Lipid metabolite [nM]
  \item DG.16.0_16.1. -4 -4 DG(16:0_16:1). Lipid metabolite [nM]
  \item DG.16.0_18.1. -4 -4 DG(16:0_18:1). Lipid metabolite [nM]
  \item DG.16.0_18.2. -4 -4 DG(16:0_18:2). Lipid metabolite [nM]
  \item DG.16.0_20.4. -4 -4 DG(16:0_20:4). Lipid metabolite [nM]
  \item DG.16.0_22.5. -4 -4 DG(16:0_22:5). Lipid metabolite [nM]
  \item DG.16.0_22.6. -4 -4 DG(16:0_22:6). Lipid metabolite [nM]
  \item DG.16.1_18.1. -4 -4 DG(16:1_18:1). Lipid metabolite [nM]
  \item DG.18.0_18.1. -4 -4 DG(18:0_18:1). Lipid metabolite [nM]
  \item DG.18.0_18.2. -4 -4 DG(18:0_18:2). Lipid metabolite [nM]
  \item DG.18.0_20.4. -4 -4 DG(18:0_20:4). Lipid metabolite [nM]
  \item DG.18.1_18.1. -4 -4 DG(18:1_18:1). Lipid metabolite [nM]
  \item DG.18.1_18.2. -4 -4 DG(18:1_18:2). Lipid metabolite [nM]
  \item DG.18.1_18.3. -4 -4 DG(18:1_18:3). Lipid metabolite [nM]
  \item DG.18.1_20.3. -4 -4 DG(18:1_20:3). Lipid metabolite [nM]
  \item DG.18.1_20.4. -4 -4 DG(18:1_20:4). Lipid metabolite [nM]
  \item DG.18.1_20.5. -4 -4 DG(18:1_20:5). Lipid metabolite [nM]
  \item DG.18.1_22.5. -4 -4 DG(18:1_22:5). Lipid metabolite [nM]
  \item DG.18.1_22.6. -4 -4 DG(18:1_22:6). Lipid metabolite [nM]
  \item DG.18.2_18.2. -4 -4 DG(18:2_18:2). Lipid metabolite [nM]
  \item DG.18.2_20.4. -4 -4 DG(18:2_20:4). Lipid metabolite [nM]
  \item DG.18.2_22.6. -4 -4 DG(18:2_22:6). Lipid metabolite [nM]
  \item TG.48.0...SIM. -4 -4 TG(48:0) [SIM]. Lipid metabolite [nM]
  \item TG.48.1...SIM. -4 -4 TG(48:1) [SIM]. Lipid metabolite [nM]
  \item TG.48.2...SIM. -4 -4 TG(48:2) [SIM]. Lipid metabolite [nM]
  \item TG.48.3...SIM. -4 -4 TG(48:3) [SIM]. Lipid metabolite [nM]
  \item TG.50.0...SIM. -4 -4 TG(50:0) [SIM]. Lipid metabolite [nM]
  \item TG.50.1...SIM. -4 -4 TG(50:1) [SIM]. Lipid metabolite [nM]
  \item TG.50.2...SIM. -4 -4 TG(50:2) [SIM]. Lipid metabolite [nM]
  \item TG.50.3...SIM. -4 -4 TG(50:3) [SIM]. Lipid metabolite [nM]
  \item TG.50.4...SIM. -4 -4 TG(50:4) [SIM]. Lipid metabolite [nM]
  \item TG.51.1...SIM. -4 -4 TG(51:1) [SIM]. Lipid metabolite [nM]
  \item TG.51.2...SIM. -4 -4 TG(51:2) [SIM]. Lipid metabolite [nM]
  \item TG.52.1...SIM. -4 -4 TG(52:1) [SIM]. Lipid metabolite [nM]
  \item TG.52.2...SIM. -4 -4 TG(52:2) [SIM]. Lipid metabolite [nM]
  \item TG.52.3...SIM. -4 -4 TG(52:3) [SIM]. Lipid metabolite [nM]
  \item TG.52.4...SIM. -4 -4 TG(52:4) [SIM]. Lipid metabolite [nM]
  \item TG.52.5...SIM. -4 -4 TG(52:5) [SIM]. Lipid metabolite [nM]
  \item TG.53.2...SIM. -4 -4 TG(53:2) [SIM]. Lipid metabolite [nM]
  \item TG.54.0...SIM. -4 -4 TG(54:0) [SIM]. Lipid metabolite [nM]
  \item TG.54.1...SIM. -4 -4 TG(54:1) [SIM]. Lipid metabolite [nM]
  \item TG.54.2...SIM. -4 -4 TG(54:2) [SIM]. Lipid metabolite [nM]
  \item TG.54.3...SIM. -4 -4 TG(54:3) [SIM]. Lipid metabolite [nM]
  \item TG.54.4...SIM. -4 -4 TG(54:4) [SIM]. Lipid metabolite [nM]
  \item TG.54.5...SIM. -4 -4 TG(54:5) [SIM]. Lipid metabolite [nM]
  \item TG.54.6...SIM. -4 -4 TG(54:6) [SIM]. Lipid metabolite [nM]
  \item TG.54.7...SIM. -4 -4 TG(54:7) [SIM]. Lipid metabolite [nM]
  \item TG.56.6...SIM. -4 -4 TG(56:6) [SIM]. Lipid metabolite [nM]
  \item TG.56.7...SIM. -4 -4 TG(56:7) [SIM]. Lipid metabolite [nM]
  \item TG.56.8...SIM. -4 -4 TG(56:8) [SIM]. Lipid metabolite [nM]
  \item TG.56.9...SIM. -4 -4 TG(56:9) [SIM]. Lipid metabolite [nM]
  \item TG.58.10...SIM. -4 -4 TG(58:10) [SIM]. Lipid metabolite [nM]
  \item TG.58.8...SIM. -4 -4 TG(58:8) [SIM]. Lipid metabolite [nM]
  \item TG.58.9...SIM. -4 -4 TG(58:9) [SIM]. Lipid metabolite [nM]
  \item TG.48.0...NL.16.0. -4 -4 TG(48:0) [NL-16:0]. Lipid metabolite [nM]
  \item TG.48.0...NL.18.0. -4 -4 TG(48:0) [NL-18:0]. Lipid metabolite [nM]
  \item TG.48.1...NL.16.1. -4 -4 TG(48:1) [NL-16:1]. Lipid metabolite [nM]
  \item TG.48.1...NL.18.1. -4 -4 TG(48:1) [NL-18:1]. Lipid metabolite [nM]
  \item TG.48.2...NL.14.0. -4 -4 TG(48:2) [NL-14:0]. Lipid metabolite [nM]
  \item TG.48.2...NL.14.1. -4 -4 TG(48:2) [NL-14:1]. Lipid metabolite [nM]
  \item TG.48.2...NL.16.1. -4 -4 TG(48:2) [NL-16:1]. Lipid metabolite [nM]
  \item TG.48.2...NL.18.2. -4 -4 TG(48:2) [NL-18:2]. Lipid metabolite [nM]
  \item TG.48.3...NL.14.0. -4 -4 TG(48:3) [NL-14:0]. Lipid metabolite [nM]
  \item TG.48.3...NL.16.1. -4 -4 TG(48:3) [NL-16:1]. Lipid metabolite [nM]
  \item TG.48.3...NL.18.3. -4 -4 TG(48:3) [NL-18:3]. Lipid metabolite [nM]
  \item TG.49.1...NL.16.1. -4 -4 TG(49:1) [NL-16:1]. Lipid metabolite [nM]
  \item TG.49.1...NL.17.1. -4 -4 TG(49:1) [NL-17:1]. Lipid metabolite [nM]
  \item TG.50.0...NL.18.0. -4 -4 TG(50:0) [NL-18:0]. Lipid metabolite [nM]
  \item TG.50.1...NL.14.0. -4 -4 TG(50:1) [NL-14:0]. Lipid metabolite [nM]
  \item TG.50.1...NL.16.0. -4 -4 TG(50:1) [NL-16:0]. Lipid metabolite [nM]
  \item TG.50.1...NL.18.1. -4 -4 TG(50:1) [NL-18:1]. Lipid metabolite [nM]
  \item TG.50.2...NL.14.0. -4 -4 TG(50:2) [NL-14:0]. Lipid metabolite [nM]
  \item TG.50.2...NL.16.1. -4 -4 TG(50:2) [NL-16:1]. Lipid metabolite [nM]
  \item TG.50.2...NL.18.1. -4 -4 TG(50:2) [NL-18:1]. Lipid metabolite [nM]
  \item TG.50.2...NL.18.2. -4 -4 TG(50:2) [NL-18:2]. Lipid metabolite [nM]
  \item TG.50.3...NL.14.0. -4 -4 TG(50:3) [NL-14:0]. Lipid metabolite [nM]
  \item TG.50.3...NL.14.1. -4 -4 TG(50:3) [NL-14:1]. Lipid metabolite [nM]
  \item TG.50.3...NL.16.1. -4 -4 TG(50:3) [NL-16:1]. Lipid metabolite [nM]
  \item TG.50.3...NL.18.2. -4 -4 TG(50:3) [NL-18:2]. Lipid metabolite [nM]
  \item TG.50.3...NL.18.3. -4 -4 TG(50:3) [NL-18:3]. Lipid metabolite [nM]
  \item TG.50.4...NL.14.0. -4 -4 TG(50:4) [NL-14:0]. Lipid metabolite [nM]
  \item TG.50.4...NL.18.3. -4 -4 TG(50:4) [NL-18:3]. Lipid metabolite [nM]
  \item TG.50.4...NL.20.4. -4 -4 TG(50:4) [NL-20:4]. Lipid metabolite [nM]
  \item TG.51.0...NL.16.0. -4 -4 TG(51:0) [NL-16:0]. Lipid metabolite [nM]
  \item TG.51.1...NL.17.0. -4 -4 TG(51:1) [NL-17:0]. Lipid metabolite [nM]
  \item TG.51.2...NL.15.0. -4 -4 TG(51:2) [NL-15:0]. Lipid metabolite [nM]
  \item TG.51.2...NL.17.0. -4 -4 TG(51:2) [NL-17:0]. Lipid metabolite [nM]
  \item TG.51.2...NL.17.1. -4 -4 TG(51:2) [NL-17:1]. Lipid metabolite [nM]
  \item TG.52.1...NL.18.0. -4 -4 TG(52:1) [NL-18:0]. Lipid metabolite [nM]
  \item TG.52.1...NL.18.1. -4 -4 TG(52:1) [NL-18:1]. Lipid metabolite [nM]
  \item TG.52.2...NL.16.0. -4 -4 TG(52:2) [NL-16:0]. Lipid metabolite [nM]
  \item TG.52.2...NL.18.2. -4 -4 TG(52:2) [NL-18:2]. Lipid metabolite [nM]
  \item TG.52.3...NL.16.1. -4 -4 TG(52:3) [NL-16:1]. Lipid metabolite [nM]
  \item TG.52.3...NL.18.2. -4 -4 TG(52:3) [NL-18:2]. Lipid metabolite [nM]
  \item TG.52.4...NL.16.1. -4 -4 TG(52:4) [NL-16:1]. Lipid metabolite [nM]
  \item TG.52.4...NL.18.2. -4 -4 TG(52:4) [NL-18:2]. Lipid metabolite [nM]
  \item TG.52.4...NL.18.3. -4 -4 TG(52:4) [NL-18:3]. Lipid metabolite [nM]
  \item TG.52.5...NL.18.3. -4 -4 TG(52:5) [NL-18:3]. Lipid metabolite [nM]
  \item TG.52.5...NL.20.4. -4 -4 TG(52:5) [NL-20:4]. Lipid metabolite [nM]
  \item TG.52.5...NL.20.5. -4 -4 TG(52:5) [NL-20:5]. Lipid metabolite [nM]
  \item TG.53.2...NL.17.1. -4 -4 TG(53:2) [NL-17:1]. Lipid metabolite [nM]
  \item TG.53.2...NL.18.1. -4 -4 TG(53:2) [NL-18:1]. Lipid metabolite [nM]
  \item TG.54.0...NL.18.0. -4 -4 TG(54:0) [NL-18:0]. Lipid metabolite [nM]
  \item TG.54.1...NL.18.1. -4 -4 TG(54:1) [NL-18:1]. Lipid metabolite [nM]
  \item TG.54.2...NL.18.0. -4 -4 TG(54:2) [NL-18:0]. Lipid metabolite [nM]
  \item TG.54.2...NL.20.1. -4 -4 TG(54:2) [NL-20:1]. Lipid metabolite [nM]
  \item TG.54.3...NL.18.1. -4 -4 TG(54:3) [NL-18:1]. Lipid metabolite [nM]
  \item TG.54.3...NL.18.2. -4 -4 TG(54:3) [NL-18:2]. Lipid metabolite [nM]
  \item TG.54.4...NL.18.2. -4 -4 TG(54:4) [NL-18:2]. Lipid metabolite [nM]
  \item TG.54.4...NL.20.3. -4 -4 TG(54:4) [NL-20:3]. Lipid metabolite [nM]
  \item TG.54.5...NL.18.3. -4 -4 TG(54:5) [NL-18:3]. Lipid metabolite [nM]
  \item TG.54.5...NL.20.4. -4 -4 TG(54:5) [NL-20:4]. Lipid metabolite [nM]
  \item TG.54.6...NL.18.3. -4 -4 TG(54:6) [NL-18:3]. Lipid metabolite [nM]
  \item TG.54.6...NL.20.4. -4 -4 TG(54:6) [NL-20:4]. Lipid metabolite [nM]
  \item TG.54.6...NL.20.5. -4 -4 TG(54:6) [NL-20:5]. Lipid metabolite [nM]
  \item TG.54.6...NL.22.6. -4 -4 TG(54:6) [NL-22:6]. Lipid metabolite [nM]
  \item TG.54.7...NL.20.5. -4 -4 TG(54:7) [NL-20:5]. Lipid metabolite [nM]
  \item TG.54.7...NL.22.6. -4 -4 TG(54:7) [NL-22:6]. Lipid metabolite [nM]
  \item TG.56.6...NL.20.4. -4 -4 TG(56:6) [NL-20:4]. Lipid metabolite [nM]
  \item TG.56.6...NL.22.5. -4 -4 TG(56:6) [NL-22:5]. Lipid metabolite [nM]
  \item TG.56.7...NL.20.4. -4 -4 TG(56:7) [NL-20:4]. Lipid metabolite [nM]
  \item TG.56.7...NL.20.5. -4 -4 TG(56:7) [NL-20:5]. Lipid metabolite [nM]
  \item TG.56.7...NL.22.5. -4 -4 TG(56:7) [NL-22:5]. Lipid metabolite [nM]
  \item TG.56.7...NL.22.6. -4 -4 TG(56:7) [NL-22:6]. Lipid metabolite [nM]
  \item TG.56.8...NL.20.4. -4 -4 TG(56:8) [NL-20:4]. Lipid metabolite [nM]
  \item TG.56.8...NL.20.5. -4 -4 TG(56:8) [NL-20:5]. Lipid metabolite [nM]
  \item TG.56.8...NL.22.6. -4 -4 TG(56:8) [NL-22:6]. Lipid metabolite [nM]
  \item TG.56.9...NL.22.6. -4 -4 TG(56:9) [NL-22:6]. Lipid metabolite [nM]
  \item TG.58.10...NL.22.6. -4 -4 TG(58:10) [NL-22:6]. Lipid metabolite [nM]
  \item TG.58.8...NL.22.6. -4 -4 TG(58:8) [NL-22:6]. Lipid metabolite [nM]
  \item TG.58.9...NL.22.6. -4 -4 TG(58:9) [NL-22:6]. Lipid metabolite [nM]
  \item TG.O.50.1...NL.15.0. -4 -4 TG(O-50:1) [NL-15:0]. Lipid metabolite [nM]
  \item TG.O.50.1...NL.16.0. -4 -4 TG(O-50:1) [NL-16:0]. Lipid metabolite [nM]
  \item TG.O.50.1...NL.17.1. -4 -4 TG(O-50:1) [NL-17:1]. Lipid metabolite [nM]
  \item TG.O.50.1...NL.18.1. -4 -4 TG(O-50:1) [NL-18:1]. Lipid metabolite [nM]
  \item TG.O.50.2...NL.18.1. -4 -4 TG(O-50:2) [NL-18:1]. Lipid metabolite [nM]
  \item TG.O.50.2...NL.18.2. -4 -4 TG(O-50:2) [NL-18:2]. Lipid metabolite [nM]
  \item TG.O.50.3...NL.18.2. -4 -4 TG(O-50:3) [NL-18:2]. Lipid metabolite [nM]
  \item TG.O.52.0...NL.16.0. -4 -4 TG(O-52:0) [NL-16:0]. Lipid metabolite [nM]
  \item TG.O.52.1...NL.16.0. -4 -4 TG(O-52:1) [NL-16:0]. Lipid metabolite [nM]
  \item TG.O.52.1...NL.18.1. -4 -4 TG(O-52:1) [NL-18:1]. Lipid metabolite [nM]
  \item TG.O.52.2...NL.16.0. -4 -4 TG(O-52:2) [NL-16:0]. Lipid metabolite [nM]
  \item TG.O.52.2...NL.17.1. -4 -4 TG(O-52:2) [NL-17:1]. Lipid metabolite [nM]
  \item TG.O.52.2...NL.18.1. -4 -4 TG(O-52:2) [NL-18:1]. Lipid metabolite [nM]
  \item TG.O.54.2...NL.17.1. -4 -4 TG(O-54:2) [NL-17:1]. Lipid metabolite [nM]
  \item TG.O.54.2...NL.18.1. -4 -4 TG(O-54:2) [NL-18:1]. Lipid metabolite [nM]
  \item TG.O.54.3...NL.17.1. -4 -4 TG(O-54:3) [NL-17:1]. Lipid metabolite [nM]
  \item TG.O.54.3...NL.18.1. -4 -4 TG(O-54:3) [NL-18:1]. Lipid metabolite [nM]
  \item TG.O.54.4...NL.17.1. -4 -4 TG(O-54:4) [NL-17:1]. Lipid metabolite [nM]
  \item TG.O.54.4...NL.18.2. -4 -4 TG(O-54:4) [NL-18:2]. Lipid metabolite [nM]
  \item UBIQUINONE -4 -4 Ubiquinone. Lipid metabolite [nM]
  \item CE.18.2....OH. -4 -4 CE(18:2) [+OH]. Lipid metabolite [nM]
  \item CE.20.4....OH. -4 -4 CE(20:4) [+OH]. Lipid metabolite [nM]
  \item CE.22.6....OH. -4 -4 CE(22:6) [+OH]. Lipid metabolite [nM]
  \item LPC.18.2....OH. -4 -4 LPC(18:2) [+OH]. Lipid metabolite [nM]
  \item LPC.20.4....OH. -4 -4 LPC(20:4) [+OH]. Lipid metabolite [nM]
  \item LPC.22.6....OH. -4 -4 LPC(22:6) [+OH]. Lipid metabolite [nM]
  \item PC.34.2....OH. -4 -4 PC(34:2) [+OH]. Lipid metabolite [nM]
  \item PC.36.4....OH. -4 -4 PC(36:4) [+OH]. Lipid metabolite [nM]
}
}
\examples{
\dontrun{
describe(admclipidomicsmeiklelablong)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/ADMCLIPIDOMICSMEIKLELABLONG/meiklelab_methods_[Updated_for_20-10-2020].pdf}
}
\keyword{datasets}
