% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{admcm2oveba}
\alias{admcm2oveba}
\title{ADMC M2OVE-AD ADNIGO-2 Bile Acids}
\format{
A data frame with 940 rows and 33 variables
}
\usage{
data(admcm2oveba)
}
\description{
ADMC M2OVE-AD ADNIGO-2 Bile Acids
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item CUSTOMER.SAMPLE.IDENTIFICATION -4 -4 This is the number listed on every tube of the study samples
  \item PLATE.BAR.CODE -4 -4 Unique identifier for every plate generated by the Biocrates MetIDQ software
  \item SAMPLE.BAR.CODE -4 -4 Unique identifier for every sample registered in the Biocrates MetIDQ software
  \item SAMPLE.TYPE -4 -4 Study samples are Sample.  Biocrates QCs are QC Level 1, 2 or 3.
  \item SAMPLE.IDENTIFICATION -4 -4 Unique identifier for every sample entered in the Duke Proteomics Sample Sumbission System
  \item SPECIES -4 -4 Species of sample
  \item MATERIAL -4 -4 For this study the samples are serum
  \item WELL.POSITION -4 -4 Well number in the 96-well plate from which the sample was injected
  \item SAMPLE.VOLUME -4 -4 The volume of sample injected in microliters
  \item RUN.NUMBER -4 -4 Generated by the Biocrates MetIDQ software for every plate indicating the run number for a particular plate.  For this study, all plates were injected only once.  Therefore, all Run Numbers are 1.
  \item INJECTION.NUMBER -4 -4 Generated by the Biocrates MetIDQ software for every sample on a plate indicating whether a sample was injected more than once.  All study samples were injected only once for every plate.  Therefore, all Injection Numbers for study samples is 1.  QC samples are injected three times for every plate resulting in a 1, 2, or 3 listing.
  \item CA -4 -4 Cholic acid [M]
  \item CDCA -4 -4 Chenodeoxycholic acid [M]
  \item DCA -4 -4 Deoxycholic acid [M]
  \item GCA -4 -4 Glycocholic acid [M]
  \item GCDCA -4 -4 Glycochenodeoxycholic acid [M]
  \item GDCA -4 -4 Glycodeoxycholic acid [M]
  \item GLCA -4 -4 Glycolithocholic acid [M]
  \item GUDCA -4 -4 Glycoursodeoxycholic acid [M]
  \item HDCA -4 -4 Hyodeoxycholic acid [M]
  \item LCA -4 -4 Lithocholic acid [M]
  \item MCA.A. -4 -4 ?-Muricholic acid [M]
  \item MCA.B. -4 -4 Muricholic acid, beta [M]
  \item MCA.O. -4 -4 Muricholic acid, omega [M]
  \item TCA -4 -4 Taurocholic acid [M]
  \item TCDCA -4 -4 Taurochenodeoxycholic acid [M]
  \item TDCA -4 -4 Taurodeoxycholic acid [M]
  \item TLCA -4 -4 Taurolithocholic acid [M]
  \item TMCA.A.B. -4 -4 Tauromuricholic acid (alpha+beta) [M]
  \item TUDCA -4 -4 Tauroursodeoxycholic acid [M]
  \item UDCA -4 -4 Ursodeoxycholic acid [M]
}
}
\examples{
\dontrun{
describe(admcm2oveba)
}
}
\keyword{datasets}
