% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{admcpdinesipcisratios}
\alias{admcpdinesipcisratios}
\title{AD Metabolomics Consortium Phenomenome Phosphatidylcholine (PtdCho), Lysophosphatidylcholine (LPtdCho) Flow Injection Negative Ionization LC-MS/MS Internal Standard Ratios}
\format{
A data frame with 829 rows and 38 variables
}
\usage{
data(admcpdinesipcisratios)
}
\description{
AD Metabolomics Consortium Phenomenome Phosphatidylcholine (PtdCho), Lysophosphatidylcholine (LPtdCho) Flow Injection Negative Ionization LC-MS/MS Internal Standard Ratios
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item RECNO -4 -4 Record number (2 indicates replicate sample)
  \item PC.16.0_18.0 -4 -4 PtdCho 16:0/18:0
  \item PC.18.0_18.0 -4 -4 PtdCho 18:0/18:0
  \item PC.16.0_18.1 -4 -4 PtdCho 16:0/18:1
  \item PC.18.0_18.1 -4 -4 PtdCho 18:0/18:1
  \item PC.18.1_18.1 -4 -4 PtdCho 18:1/18:1
  \item PC.16.0_18.2 -4 -4 PtdCho 16:0/18:2
  \item PC.18.0_18.2 -4 -4 PtdCho 18:0/18:2
  \item PC.18.1_18.2 -4 -4 PtdCho 18:1/18:2
  \item PC.16.0_18.3 -4 -4 PtdCho 16:0/18:3
  \item PC.18.0_18.3 -4 -4 PtdCho 18:0/18:3
  \item PC.18.1_18.3 -4 -4 PtdCho 18:1/18:3
  \item PC.16.0_20.4 -4 -4 PtdCho 16:0/20:4
  \item PC.18.0_20.4 -4 -4 PtdCho 18:0/20:4
  \item PC.18.1_20.4 -4 -4 PtdCho 18:1/20:4
  \item PC.16.0_20.5 -4 -4 PtdCho 16:0/20:5
  \item PC.18.0_20.5 -4 -4 PtdCho 18:0/20:5
  \item PC.18.1_20.5 -4 -4 PtdCho 18:1/20:5
  \item PC.16.0_22.4 -4 -4 PtdCho 16:0/22:4
  \item PC.18.0_22.4 -4 -4 PtdCho 18:0/22:4
  \item PC.18.1_22.4 -4 -4 PtdCho 18:1/22:4
  \item PC.16.0_22.6 -4 -4 PtdCho 16:0/22:6
  \item PC.16.1_22.6 -4 -4 PtdCho 16:1/22:6
  \item PC.18.0_22.6 -4 -4 PtdCho 18:0/22:6
  \item PC.18.1_22.6 -4 -4 PtdCho 18:1/22:6
  \item LPC.16.0 -4 -4 LPtdCho 16:0/OH
  \item LPC.16.1 -4 -4 LPtdCho 16:1/OH
  \item LPC.18.0 -4 -4 LPtdCho 18:0/OH
  \item LPC.18.1 -4 -4 LPtdCho 18:1/OH
  \item LPC.18.2 -4 -4 LPtdCho 18:2/OH
  \item LPC.18.3 -4 -4 LysoPtdCho 18:3
  \item LPC.20.4 -4 -4 LPtdCho 20:4/OH
  \item LPC.20.5 -4 -4 LysoPtdCho 20:5
  \item LPC.22.6 -4 -4 LPtdCho 22:6/OH
  \item LPC.22.4 -4 -4 LysoPtdCho 22:4
  \item update_stamp -4 -4 -4
}
}
\examples{
\dontrun{
describe(admcpdinesipcisratios)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/ADMCPDINESIPCISRATIOS/ADMC_DMC_Phenomenome_Phosphatidylcholine_Lysophosphatidylcholine_Flow_Injection_Negative_Ionization_LC-MS-MS_Protocol.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADMCPDINESIPCISRATIOS/ADMC_Phenomenome_Phosphatidylethanolamine_Plasmenylethanolamine_Flow_Injection_LC-MS-MS_Protocol.pdf}
}
\keyword{datasets}
