% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{admcpdipepeakheights}
\alias{admcpdipepeakheights}
\title{AD Metabolomics Consortium Phenomenome Phosphatidylethanolamine (PtdEtn), Plasmenylethanolamine (PlsEtn) Flow Injection LC-MS/MS Internal Standard Ratios}
\format{
A data frame with 1734 rows and 40 variables
}
\usage{
data(admcpdipepeakheights)
}
\description{
AD Metabolomics Consortium Phenomenome Phosphatidylethanolamine (PtdEtn), Plasmenylethanolamine (PlsEtn) Flow Injection LC-MS/MS Internal Standard Ratios
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item BATCH.ID -4 -4 -4
  \item PDI.ID -4 -4 -4
  \item C13.PE.P.16.0.22.6 -4 -4 13C Labeled PlsEtn 16:0/22:6
  \item C13.PE.16.0.22.6 -4 -4 13C Labeled PtdEtn 16:0/22:6
  \item PE.16.0.18.3 -4 -4 PtdEtn 16:0/18:3
  \item PE.16.0.22.6 -4 -4 PtdEtn 16:0/22:6
  \item PE.18.0.20.5 -4 -4 PtdEtn 18:0/20:5
  \item PE.18.0.22.4 -4 -4 PtdEtn 18:0/22:4
  \item PE.P.16.0.18.2 -4 -4 PlsEtn 16:0/18:2
  \item PE.P.18.0.20.5 -4 -4 PlsEtn 18:0/20:5
  \item PE.P.16.0.22.4 -4 -4 PlsEtn 16:0/22:4
  \item PE.P.16.0.22.6 -4 -4 PlsEtn 16:0/22:6
  \item SAMPLE.TYPE -4 -4 -4
  \item PE.16.0.18.0 -4 -4 PtdEtn 16:0/18:0
  \item PE.16.0.18.1 -4 -4 PtdEtn 16:0/18:1
  \item PE.16.0.18.2 -4 -4 PtdEtn 16:0/18:2
  \item PE.16.0.20.4 -4 -4 PtdEtn 16:0/20:4
  \item PE.16.0.22.4 -4 -4 PtdEtn 16:0/22:4
  \item PE.16.0.24.6 -4 -4 PtdEtn 16:0/24:6
  \item PE.18.0.18.1 -4 -4 PtdEtn 18:0/18:1
  \item PE.18.0.18.2 -4 -4 PtdEtn 18:0/18:2
  \item PE.18.0.18.3 -4 -4 PtdEtn 18:0/18:3
  \item PE.18.0.20.4 -4 -4 PtdEtn 18:0/20:4
  \item PE.18.0.22.6 -4 -4 PtdEtn 18:0/22:6
  \item PE.18.0.24.6 -4 -4 PtdEtn 18:0/24:6
  \item PE.P.16.0.18.0 -4 -4 PlsEtn 16:0/18:0
  \item PE.P.16.0.18.1 -4 -4 PlsEtn 16:0/18:1
  \item PE.P.16.0.18.3 -4 -4 PlsEtn 16:0/18:3
  \item PE.P.16.0.20.4 -4 -4 PlsEtn 16:0/20:4
  \item PE.P.18.0.18.0 -4 -4 PlsEtn 18:0/18:0
  \item PE.P.18.0.18.1 -4 -4 PlsEtn 18:0/18:1
  \item PE.P.18.0.18.2 -4 -4 PlsEtn 18:0/18:2
  \item PE.P.18.0.18.3 -4 -4 PlsEtn 18:0/18:3
  \item PE.P.18.0.22.4 -4 -4 PlsEtn 18:0/22:4
  \item PE.P.18.0.22.6 -4 -4 PlsEtn 18:0/22:6
  \item PE.16.0.20.5 -4 -4 PtdEtn 16:0/20:5
  \item PE.P.16.0.20.5 -4 -4 PlsEtn 16:0/20:5
  \item PE.P.18.0.20.4 -4 -4 PlsEtn 18:0/20:4
}
}
\examples{
\dontrun{
describe(admcpdipepeakheights)
}
}
\keyword{datasets}
