% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{admcpdipesipcisratios}
\alias{admcpdipesipcisratios}
\title{AD Metabolomics Consortium Phenomenome Phosphatidylcholine (PtdCho), Lysophosphatidylcholine (LPtdCho), Sphingomyeline (SM) Flow Injection Positive Ionization LC-MS/MS Internal Standard Ratios}
\format{
A data frame with 829 rows and 36 variables
}
\usage{
data(admcpdipesipcisratios)
}
\description{
AD Metabolomics Consortium Phenomenome Phosphatidylcholine (PtdCho), Lysophosphatidylcholine (LPtdCho), Sphingomyeline (SM) Flow Injection Positive Ionization LC-MS/MS Internal Standard Ratios
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item RECNO -4 -4 Record number (2 indicates replicate sample)
  \item PC.34.3 -4 -4 diacyl-PtdCho  16:0/18:3; 14:0/20:3; 16:1/18:2
  \item PC.34.0 -4 -4 diacyl-PtdCho 16:0/18:0; 14:0/20:0
  \item PC.34.1 -4 -4 diacyl PtdCho 16:0/18:1; 14:0/20:1; 16:1/18:0
  \item PC.34.2 -4 -4 diacyl PtdCho 16:0/18:2; 14:0/20:2; 16:1/18:1
  \item PC.36.4 -4 -4 diacyl PtdCho 16:0/20:4; 18:1/18:3; 14:0/22:4; 18:2/18:2; 16:1/20:3
  \item PC.36.5 -4 -4 diacyl PtdCho 16:0/20:5; 14:0/22:5; 18:2/18:3; 18:2/18:3
  \item PC.38.4 -4 -4 diacyl PtdCho 16:0/22:4, 18:0/20:4; 14:0/24:4; 18:1/20:3
  \item PC.38.6 -4 -4 diacyl PtdCho 16:0/22:6, 18:1/20:5; 14:0/24:6; 18:2/20:4
  \item PC.38.7 -4 -4 diacyl PtdCho 16:1/22:6; 14:1/24:6; 18:2/20:5
  \item PC.36.2 -4 -4 diacyl PtdCho 18:0/18:2; 18:1/18:1; 14:0/22:2; 16:0/20:2; 16:1/20:1
  \item PC.36.3 -4 -4 diacyl PtdCho 18:0/18:3, 18:1/18:2; 14:0/22:3; 16:0/20:3; 16:1/20:2
  \item PC.38.5 -4 -4 diacyl PtdCho 18:0/20:5, 18:1/20:4; 14:0/24:5; 16:0/22:5; 16:1/22:4
  \item PC.40.4 -4 -4 diacyl PtdCho 18:0/22:4; 16:0/24:4; 16:1/24:3; 20:0/20:4
  \item PC.40.6 -4 -4 diacyl PtdCho 18:0/22:6; 16:0/24:6; 18:1/22:5
  \item PC.40.5 -4 -4 diacyl PtdCho 18:1/22:4; 16:0/24:5; 18:0/22:5; 20:0/20:5; 20:1/20:4
  \item PC.40.7 -4 -4 diacyl PtdCho 18:1/22:6; 16:0/24:6
  \item PC.36.6 -4 -4 diacyl PtdCho 14:0/22:6; 16:1/20:5; 18:3/18:3
  \item PC.38.0 -4 -4 diacyl PtdCho 18:0/20:0;16:0/22:0
  \item PC.40.1 -4 -4 diacyl PtdCho 18:0/22:1; 18:1/22:0; 16:0/24:1; 16:1/24:0
  \item PC.40.2 -4 -4 diacyl PtdCho 18:1/22:1; 18:0/22:2; 16:0/24:2; 16:1/24:1
  \item LPC.16.0 -4 -4 LPtdCho 16:0/OH
  \item LPC.16.1 -4 -4 LPtdCho 16:1/OH
  \item LPC.18.0 -4 -4 LPtdCho 18:0/OH
  \item LPC.18.1 -4 -4 LPtdCho 18:1/OH
  \item LPC.18.2 -4 -4 LPtdCho 18:2/OH
  \item LPC.20.4 -4 -4 LPtdCho 20:4/OH
  \item LPC.22.6 -4 -4 LPtdCho 22:6/OH
  \item SM.34.1 -4 -4 SphMyelin (d18:1/16:0), SphMyelin (d18:0/16:1)
  \item SM.36.1 -4 -4 SphMyelin (d18:1/18:0), SphMyelin (d18:0/18:1)
  \item SM.36.2 -4 -4 SphMyelin (d18:1/18:1)
  \item SM.42.1 -4 -4 SphMyelin (d18:1/24:0)
  \item SM.42.2 -4 -4 SphMyelin (d18:1/24:1)
  \item update_stamp -4 -4 -4
}
}
\examples{
\dontrun{
describe(admcpdipesipcisratios)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/ADMCPDIPESIPCISRATIOS/ADMC_DMC_Phenomenome_Phosphatidylcholine_Lysophosphatidylcholine_Flow_Injection_Negative_Ionization_LC-MS-MS_Protocol.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADMCPDIPESIPCISRATIOS/ADMC_Phenomenome_Phosphatidylethanolamine_Plasmenylethanolamine_Flow_Injection_LC-MS-MS_Protocol.pdf}
}
\keyword{datasets}
