% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{adni1lipidomicsrader}
\alias{adni1lipidomicsrader}
\title{Rader Lab ADNI 1 Lipids CAD ALZ}
\format{
A data frame with 792 rows and 9 variables
}
\usage{
data(adni1lipidomicsrader)
}
\description{
Rader Lab ADNI 1 Lipids CAD ALZ
}
\details{
A dataset with variables as follows:

\itemize{
  \item ORIGPROT -4 -4 Original study protocol
  \item COLPROT -4 -4 Study protocol of data collection
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item CHOL -4 mg/dl 
  \item HDL -4 mg/dl 
  \item TG -4 mg/dl 
  \item APOA1 -4 mg/dl 
  \item APOE -4 mg/dl 
}
}
\examples{
\dontrun{
describe(adni1lipidomicsrader)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/ADNI1LIPIDOMICSRADER/ADNI1_Methods_Lipid_and_ApoLPs_Testing_Rader_Lab_06_21_18.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNI1LIPIDOMICSRADER/COBAS_C311_Methods_REF_RANGES_2_5_2018.xlsx}
}
\keyword{datasets}
