% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{adni3_urmc_labdata}
\alias{adni3_urmc_labdata}
\title{ADNI3 URMC Lab Data}
\format{
A data frame with 58784 rows and 24 variables
}
\usage{
data(adni3_urmc_labdata)
}
\description{
ADNI3 URMC Lab Data
}
\details{
A dataset with variables as follows:

\itemize{
  \item COLPROT -4 -4 Study protocol of data collection
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item SAMPLEDATE Date  -4 Date of Sample Collection
  \item SAMPLETIME Time -4 Time of Sample Collection
  \item RECDATE Date -4 Date of Sample Receipt
  \item RECTIME Time -4 Time of Sample Receipt
  \item KITID Text -4 Specimen Collection Kit Identifier
  \item TESTID Text -4 Individual Test Identifier
  \item TESTNAME Text -4 Individual Test Name
  \item RESULTVALUE -4 -4 -4
  \item FLAGS Text -4 Result Flag 
  \item LOWERRANGE -4 -4 -4
  \item UPPERRANGE -4 -4 -4
  \item UNITS -4 -4 -4
  \item RESULTVALUECONV Text -4 Result Value in Conventional Units
  \item LOWERRANGECONV Text -4 Lower Limit of Normal in Conventional Units
  \item UPPERRANGECONV Text -4 Upper Limit of Normal in Conventional Units
  \item UNITSCONV Text -4 Result Units in Conventional Units
  \item RESULTVALUESI Text -4 Result Value in SI Units
  \item LOWERRANGESI Text -4 Lower Limit of Normal in SI Units
  \item UPPERRANGESI Text -4 Upper Limit of Normal in SI Units
  \item UNITSSI Text -4 Result Units in SI Units
}
}
\examples{
\dontrun{
describe(adni3_urmc_labdata)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/ADNI3_URMC_LABDATA/Cystatin.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNI3_URMC_LABDATA/Homocysteine.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNI3_URMC_LABDATA/Insert.A1CX3.0207559674190c513.V2.en.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNI3_URMC_LABDATA/Insert.ALB2.0005166861190c701.V7.en.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNI3_URMC_LABDATA/Insert.ALP2.0005166888190c701.V8.en.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNI3_URMC_LABDATA/Insert.ALT.0005850797190c701.V8.en.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNI3_URMC_LABDATA/Insert.AST.0005850819190c701.V8.en.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNI3_URMC_LABDATA/Insert.BILT3.0205795419190c701.V9.en.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNI3_URMC_LABDATA/Insert.CA2.0005168449190c701.V8.en.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNI3_URMC_LABDATA/Insert.CHOL2.0005168538190c701.V12.en.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNI3_URMC_LABDATA/Insert.CO2-L.0005446376191c701.V12.en.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNI3_URMC_LABDATA/Insert.CREP2.0205168589190c701.V1.en.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNI3_URMC_LABDATA/Insert.DBILI.0605975921190c701.V7.en.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNI3_URMC_LABDATA/Insert.Elecsys\%20TSH.08443432501.V2.en.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNI3_URMC_LABDATA/Insert.Elecsys\%20Vitamin\%20B12\%20II.07028121501.V3.en.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNI3_URMC_LABDATA/Insert.GLUC3.0205168791190c701.V9.en.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNI3_URMC_LABDATA/Insert.ISE\%20indirect\%20Na-K-Cl\%20for\%20Gen.2.0210825468001c503.V2.en.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNI3_URMC_LABDATA/Insert.TP2.0005171385190c701.V12.en.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNI3_URMC_LABDATA/Roche\%20COBAS\%20UREAL.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNI3_URMC_LABDATA/xn-1000-brochure_mkt-10-1215_rev3.pdf}
}
\keyword{datasets}
