% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{adni_dian_comparison}
\alias{adni_dian_comparison}
\title{The DIAN-ADNI Comparison Study}
\format{
A data frame with 3253 rows and 133 variables
}
\usage{
data(adni_dian_comparison)
}
\description{
The DIAN-ADNI Comparison Study
}
\details{
A dataset with variables as follows:

\itemize{
  \item COLPROT -4 -4 Study protocol of data collection
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item VISITAGE -4 -4 -4
  \item CDRGLOB -4 -4 -4
  \item CSFDATE -4 -4 -4
  \item EDUC -4 -4 -4
  \item MRI_SCANDATE -4 -4 -4
  \item FLUF -4 -4 -4
  \item WORDIM -4 -4 -4
  \item LOGIMEM -4 -4 -4
  \item DIGIF -4 -4 -4
  \item DIGIFLEN -4 -4 -4
  \item DIGIB -4 -4 -4
  \item DIGIBLEN -4 -4 -4
  \item ANIMALS -4 -4 -4
  \item VEG -4 -4 -4
  \item TRAILA -4 -4 -4
  \item TRAILARR -4 -4 -4
  \item TRAILB -4 -4 -4
  \item TRAILBRR -4 -4 -4
  \item WAIS -4 -4 -4
  \item MEMUNITS -4 -4 -4
  \item BOSTON -4 -4 -4
  \item WORDDEL -4 -4 -4
  \item MR_TOTV_WMHYPOINTENSITIES -4 -4 -4
  \item MR_TOTV_INTRACRANIAL -4 -4 -4
  \item MR_TOTV_HIPPOCAMPUS -4 -4 -4
  \item FDG_FSUVR_RSF_TOT_CTX_PRECUNEUS -4 -4 -4
  \item FDG_FSUVR_RSF_TOT_HIPPOCAMPUS -4 -4 -4
  \item FDG_FSUVR_RSF_TOT_CORTMEAN -4 -4 -4
  \item PIB_FSUVR_RSF_TOT_CTX_PRECUNEUS -4 -4 -4
  \item PIB_FSUVR_RSF_TOT_HIPPOCAMPUS -4 -4 -4
  \item PIB_FSUVR_RSF_TOT_CORTMEAN -4 -4 -4
  \item BIRTHMO -4 -4 -4
  \item BIRTHYR -4 -4 -4
  \item HISPANIC -4 -4 -4
  \item RACE -4 -4 -4
  \item PRIMLANG -4 -4 -4
  \item MARISTAT -4 -4 -4
  \item HANDED -4 -4 -4
  \item MOMDEM -4 -4 -4
  \item MOMAUTO -4 -4 -4
  \item DADDEM -4 -4 -4
  \item DADAUTO -4 -4 -4
  \item ABRUPT -4 -4 -4
  \item STEPWISE -4 -4 -4
  \item SOMATIC -4 -4 -4
  \item EMOT -4 -4 -4
  \item HXHYPER -4 -4 -4
  \item HXSTROKE -4 -4 -4
  \item FOCLSYM -4 -4 -4
  \item FOCLSIGN -4 -4 -4
  \item HACHIN -4 -4 -4
  \item CDRSUM -4 -4 -4
  \item GDS -4 -4 -4
  \item BILLS -4 -4 -4
  \item TAXES -4 -4 -4
  \item SHOPPING -4 -4 -4
  \item GAMES -4 -4 -4
  \item STOVE -4 -4 -4
  \item MEALPREP -4 -4 -4
  \item EVENTS -4 -4 -4
  \item PAYATTN -4 -4 -4
  \item REMDATES -4 -4 -4
  \item TRAVEL -4 -4 -4
  \item DECSUB -4 -4 -4
  \item VASC -4 -4 -4
  \item ALCDEM -4 -4 -4
  \item FTD -4 -4 -4
  \item PPAPH -4 -4 -4
  \item PSP -4 -4 -4
  \item CORT -4 -4 -4
  \item HUNT -4 -4 -4
  \item PRION -4 -4 -4
  \item DEP -4 -4 -4
  \item PARK -4 -4 -4
  \item HYCEPH -4 -4 -4
  \item COGOTH -4 -4 -4
  \item COGOTHX -4 -4 -4
  \item GENDER -4 -4 -4
  \item CSF_ELC_AB42 -4 -4 -4
  \item CSF_ELC_PTAU -4 -4 -4
  \item CSF_ELC_TAU -4 -4 -4
  \item CSF_ELC_AB40 -4 -4 -4
  \item CSF_ELC_AB4240 -4 -4 -4
  \item MSP_AB38 -4 -4 -4
  \item MSP_AB40 -4 -4 -4
  \item MSP_AB42 -4 -4 -4
  \item CSF_DATE_BL -4 -4 -4
  \item MRI_DATE_BL -4 -4 -4
  \item CSF_CDR_BL -4 -4 -4
  \item MRI_CDR_BL -4 -4 -4
  \item CSF_CDRSB_BL -4 -4 -4
  \item MRI_CDRSB_BL -4 -4 -4
  \item VISIT_DATE_BL -4 -4 -4
  \item MR_TOTT_PRECUNEUS -4 -4 -4
  \item LAST_VISIT_DATE -4 -4 -4
  \item TIME_COGNITIVE_FOLLOWUP -4 -4 -4
  \item TIME_CSF_FOLLOWUP -4 -4 -4
  \item TIME_MRI_FOLLOWUP -4 -4 -4
  \item CDR_BASELINE -4 -4 -4
  \item GROUP_CSF -4 -4 -4
  \item GROUP_MRI -4 -4 -4
  \item COHORT -4 -4 -4
  \item PET_TYPE -4 -4 -4
  \item PIB_MSUVR_TOT_PRECUNEUS -4 -4 -4
  \item PIB_MSUVR_TOT_CORTMEAN -4 -4 -4
  \item FDG_MSUVR_TOT_PRECUNEUS -4 -4 -4
  \item FDG_MSUVR_TOT_CORTMEAN -4 -4 -4
  \item AV45_FSUVR_RSF_TOT_CTX_PRECUNEUS -4 -4 -4
  \item AV45_FSUVR_RSF_TOT_HIPPOCAMPUS -4 -4 -4
  \item AV45_FSUVR_RSF_TOT_CORTMEAN -4 -4 -4
  \item AV45_MSUVR_TOT_PRECUNEUS -4 -4 -4
  \item AV45_MSUVR_TOT_CORTMEAN -4 -4 -4
  \item PET_DATE_BL -4 -4 -4
  \item PET_CDR_BL -4 -4 -4
  \item PET_CDRSB_BL -4 -4 -4
  \item TIME_PET_FOLLOWUP -4 -4 -4
  \item ADGROUP -4 -4 -4
  \item ADGROUP_BL -4 -4 -4
  \item GROUP_PET -4 -4 -4
  \item PET_SCANDATE -4 -4 -4
  \item ORIGIN -4 -4 -4
  \item DIAN_ID -4 -4 -4
  \item DIAN_APOE -4 -4 -4
  \item DIAN_GROUP -4 -4 -4
  \item DIAN_CDRSB_BL -4 -4 -4
  \item DIAN_YEARS_BL -4 -4 -4
  \item DIAN_MMSE -4 -4 -4
  \item NEYO -4 -4 -4
}
}
\examples{
\dontrun{
describe(adni_dian_comparison)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/ADNI_DIAN_COMPARISON/ADNI-DIAN_Study_Summary_20190227.doc}
}
\keyword{datasets}
