% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADNIMERGE-package.R
\docType{package}
\name{ADNIMERGE-package}
\alias{ADNIMERGE-package}
\title{ADNIMERGE: Clinical and biomarker data from the Alzheimer's Disease Neuroimaging Initiative (ADNI)}
\seealso{
\code{\url{https://groups.google.com/forum/#!forum/adni-data}}
}
\author{
ATRI Biostatistics
}
\keyword{package}
