% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{adnimerge}
\alias{adnimerge}
\title{Key ADNI tables merged into one table}
\format{
A data frame with 16443 rows and 115 variables
}
\usage{
data(adnimerge)
}
\description{
Key ADNI tables merged into one table
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID  NA Participant roster ID
  \item COLPROT  NA Study protocol of data collection
  \item ORIGPROT  NA Original study protocol
  \item PTID  NA Original study protocol
  \item SITE  NA Site
  \item VISCODE  NA Visit code
  \item EXAMDATE  NA Date
  \item DX.bl  NA Baseline Dx
  \item AGE  NA Age
  \item PTGENDER  NA Sex
  \item PTEDUCAT  NA Education
  \item PTETHCAT  NA Ethnicity
  \item PTRACCAT  NA Race
  \item PTMARRY  NA Marital
  \item APOE4  NA Number of APOEe4 alleles
  \item FDG  NA FDG-PET metaROI
  \item PIB  NA Average PIB SUVR of frontal cortex, anterior cingulate, precuneus cortex, and parietal cortex
  \item AV45  NA AV45 ratio (cortical grey matter/whole cerebellum) Summary florbetapir cortical SUVR normalized by whole cerebellum. See Jagust lab PDF on LONI for details and cutoff info
  \item FBB  NA FBB ratio (cortical grey matter/whole cerebellum) Summary florbetaben cortical SUVR normalized by whole cerebellum. See Jagust lab PDF on LONI for details and cutoff info
  \item ABETA  NA CSF ABETA
  \item TAU  NA CSF TAU
  \item PTAU  NA CSF PTAU
  \item CDRSB  NA CDR-SB
  \item ADAS11  NA ADAS 11
  \item ADAS13  NA ADAS 13 (including Delayed Word Recall and Number Cancellation)
  \item ADASQ4  NA ADAS Delayed Word Recall
  \item MMSE  NA MMSE
  \item RAVLT.immediate  NA RAVLT Immediate (sum of 5 trials)
  \item RAVLT.learning  NA RAVLT Learning (trial 5 - trial 1)
  \item RAVLT.forgetting  NA RAVLT Forgetting (trial 5 - delayed)
  \item RAVLT.perc.forgetting  NA RAVLT Percent Forgetting
  \item LDELTOTAL  NA Logical Memory - Delayed Recall
  \item DIGITSCOR  NA Digit Symbol Substitution
  \item TRABSCOR  NA Trails B
  \item FAQ  NA FAQ
  \item MOCA  NA MOCA
  \item EcogPtMem  NA Pt ECog - Mem
  \item EcogPtLang  NA Pt ECog - Lang
  \item EcogPtVisspat  NA Pt ECog - Vis/Spat
  \item EcogPtPlan  NA Pt ECog - Plan
  \item EcogPtOrgan  NA Pt ECog - Organ
  \item EcogPtDivatt  NA Pt ECog - Div atten
  \item EcogPtTotal  NA Pt ECog - Total
  \item EcogSPMem  NA SP ECog - Mem
  \item EcogSPLang  NA SP ECog - Lang
  \item EcogSPVisspat  NA SP ECog - Vis/Spat
  \item EcogSPPlan  NA SP ECog - Plan
  \item EcogSPOrgan  NA SP ECog - Organ
  \item EcogSPDivatt  NA SP ECog - Div atten
  \item EcogSPTotal  NA SP ECog - Total
  \item FLDSTRENG  NA MRI Field Strength
  \item FSVERSION  NA FreeSurfer Software Version
  \item IMAGEUID  NA LONI Image ID
  \item Ventricles  NA UCSF Ventricles
  \item Hippocampus  NA UCSF Hippocampus
  \item WholeBrain  NA UCSF WholeBrain
  \item Entorhinal  NA UCSF Entorhinal
  \item Fusiform  NA UCSF Fusiform
  \item MidTemp  NA UCSF Med Temp
  \item ICV  NA UCSF ICV
  \item DX  NA Diagnosis
  \item mPACCdigit  NA ADNI modified Preclinical Alzheimer's Cognitive Composite (PACC) with Digit Symbol Substitution
  \item mPACCtrailsB  NA ADNI modified Preclinical Alzheimer's Cognitive Composite (PACC) with Trails B
  \item EXAMDATE.bl  NA Date
  \item CDRSB.bl  NA CDR-SB
  \item ADAS11.bl  NA ADAS 11
  \item ADAS13.bl  NA ADAS 13 (including Delayed Word Recall and Number Cancellation)
  \item ADASQ4.bl  NA ADAS Delayed Word Recall
  \item MMSE.bl  NA MMSE
  \item RAVLT.immediate.bl  NA RAVLT Immediate (sum of 5 trials)
  \item RAVLT.learning.bl  NA RAVLT Learning (trial 5 - trial 1)
  \item RAVLT.forgetting.bl  NA RAVLT Forgetting (trial 5 - delayed)
  \item RAVLT.perc.forgetting.bl  NA RAVLT Percent Forgetting
  \item LDELTOTAL.bl  NA Logical Memory - Delayed Recall
  \item DIGITSCOR.bl  NA Digit Symbol Substitution
  \item TRABSCOR.bl  NA Trails B
  \item FAQ.bl  NA FAQ
  \item mPACCdigit.bl  NA ADNI modified Preclinical Alzheimer's Cognitive Composite (PACC) with Digit Symbol Substitution
  \item mPACCtrailsB.bl  NA ADNI modified Preclinical Alzheimer's Cognitive Composite (PACC) with Trails B
  \item FLDSTRENG.bl  NA MRI Field Strength
  \item FSVERSION.bl  NA FreeSurfer Software Version
  \item IMAGEUID.bl  NA LONI Image ID
  \item Ventricles.bl  NA UCSF Ventricles
  \item Hippocampus.bl  NA UCSF Hippocampus
  \item WholeBrain.bl  NA UCSF WholeBrain
  \item Entorhinal.bl  NA UCSF Entorhinal
  \item Fusiform.bl  NA UCSF Fusiform
  \item MidTemp.bl  NA UCSF Med Temp
  \item ICV.bl  NA UCSF ICV
  \item MOCA.bl  NA MOCA
  \item EcogPtMem.bl  NA Pt ECog - Mem
  \item EcogPtLang.bl  NA Pt ECog - Lang
  \item EcogPtVisspat.bl  NA Pt ECog - Vis/Spat
  \item EcogPtPlan.bl  NA Pt ECog - Plan
  \item EcogPtOrgan.bl  NA Pt ECog - Organ
  \item EcogPtDivatt.bl  NA Pt ECog - Div atten
  \item EcogPtTotal.bl  NA Pt ECog - Total
  \item EcogSPMem.bl  NA SP ECog - Mem
  \item EcogSPLang.bl  NA SP ECog - Lang
  \item EcogSPVisspat.bl  NA SP ECog - Vis/Spat
  \item EcogSPPlan.bl  NA SP ECog - Plan
  \item EcogSPOrgan.bl  NA SP ECog - Organ
  \item EcogSPDivatt.bl  NA SP ECog - Div atten
  \item EcogSPTotal.bl  NA SP ECog - Total
  \item ABETA.bl  NA CSF ABETA
  \item TAU.bl  NA CSF TAU
  \item PTAU.bl  NA CSF PTAU
  \item FDG.bl  NA FDG-PET metaROI
  \item PIB.bl  NA Average PIB SUVR of frontal cortex, anterior cingulate, precuneus cortex, and parietal cortex
  \item AV45.bl  NA AV45 ratio (cortical grey matter/whole cerebellum) Summary florbetapir cortical SUVR normalized by whole cerebellum. See Jagust lab PDF on LONI for details and cutoff info
  \item FBB.bl  NA FBB ratio (cortical grey matter/whole cerebellum) Summary florbetaben cortical SUVR normalized by whole cerebellum. See Jagust lab PDF on LONI for details and cutoff info
  \item Years.bl  NA 
  \item Month.bl  NA 
  \item Month  NA Months since baseline
  \item M  NA Month since baseline
}
}
\examples{
dd <- adnimerge
covariates <- c('CDRSB', 'ADAS11', 'ADAS13', 'MMSE', 'RAVLT.immediate', 'FAQ', 'Ventricles', 'Hippocampus', 'WholeBrain', 'Entorhinal', 'MOCA')

dd$Age.c <- dd$AGE - mean(dd$AGE[dd$VISCODE == 'bl'], na.rm = TRUE)

dd[, paste(covariates, '.bl.c', sep='')] <- sweep(dd[, paste(covariates, '.bl', sep='')], 2, 
                                                  colMeans(dd[, paste(covariates, '.bl', sep='')], na.rm=TRUE), '-')

dd[, paste(c('EXAMDATE', covariates), '.ch', sep='')] <- dd[, c('EXAMDATE', covariates)] - 
  dd[, paste(c('EXAMDATE', covariates), '.bl', sep='')]

dd$Dx <- unlist(lapply(strsplit(as.character(dd$DX), ' to ', fixed = TRUE), function(x) x[length(x)]))

label(dd$Age.c) <- 'Centered Age'

}
\seealso{
\code{\link{adnimerger}}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
