% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adnimerger.R
\name{adnimerger}
\alias{adnimerger}
\title{Function to merge commonly used ADNI data tables}
\usage{
adnimerger()
}
\arguments{
\item{csf}{use Roche Elecsys ('Roche', default) or INNO-BIA AlzBio3 immunoassay ('INNO')}
}
\value{
Returns an analysis-ready dataset.
}
\description{
This function merges commonly used ADNI data tables which come packaged in ADNIMERGE.
}
