% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{adninightingale}
\alias{adninightingale}
\title{ADMC Nightingale Platform  NMR Analysis of Lipoproteins and Metabolites [ADNI 1, GO, 2]}
\format{
A data frame with 1697 rows and 242 variables
}
\usage{
data(adninightingale)
}
\description{
ADMC Nightingale Platform  NMR Analysis of Lipoproteins and Metabolites [ADNI 1, GO, 2]
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item SAMPLEID -4  Sample ID
  \item GLOBAL.SPEC.ID -4 -4 -4
  \item TAG_LOW_GLUCOSE -4  Low glucose
  \item TAG_HIGH_PYRUVATE -4  High pyruvate
  \item TAG_HIGH_LACTATE -4  High lactate
  \item TAG_UNEXPECTED_AMINO_ACID_SIGNALS -4  Unexpected amino acid signals
  \item TAG_LOW_GLUTAMINE_HIGH_GLUTAMATE -4  Low glutamine / high glutamate
  \item TAG_LOW_PROTEIN_CONTENT -4  Low protein content
  \item TAG_PLASMA_SAMPLE -4  Plasma sample
  \item TAG_HIGH_ETHANOL -4  High ethanol
  \item XXL.VLDL.P -4 mol/l Concentration of chylomicrons and extremely large VLDL particles
  \item XXL.VLDL.L -4 mmol/l Total lipids in chylomicrons and extremely large VLDL 
  \item XXL.VLDL.PL -4 mmol/l Phospholipids in chylomicrons and extremely large VLDL 
  \item XXL.VLDL.C -4 mmol/l Total cholesterol in chylomicrons and extremely large VLDL 
  \item XXL.VLDL.CE -4 mmol/l Cholesterol esters in chylomicrons and extremely large VLDL 
  \item XXL.VLDL.FC -4 mmol/l Free cholesterol in chylomicrons and extremely large VLDL 
  \item XXL.VLDL.TG -4 mmol/l Triglycerides in chylomicrons and extremely large VLDL 
  \item XL.VLDL.P -4 mol/l Concentration of very large VLDL particles
  \item XL.VLDL.L -4 mmol/l Total lipids in very large VLDL 
  \item XL.VLDL.PL -4 mmol/l Phospholipids in very large VLDL 
  \item XL.VLDL.C -4 mmol/l Total cholesterol in very large VLDL 
  \item XL.VLDL.CE -4 mmol/l Cholesterol esters in very large VLDL 
  \item XL.VLDL.FC -4 mmol/l Free cholesterol in very large VLDL 
  \item XL.VLDL.TG -4 mmol/l Triglycerides in very large VLDL 
  \item L.VLDL.P -4 mol/l Concentration of large VLDL particles
  \item L.VLDL.L -4 mmol/l Total lipids in large VLDL 
  \item L.VLDL.PL -4 mmol/l Phospholipids in large VLDL 
  \item L.VLDL.C -4 mmol/l Total cholesterol in large VLDL 
  \item L.VLDL.CE -4 mmol/l Cholesterol esters in large VLDL 
  \item L.VLDL.FC -4 mmol/l Free cholesterol in large VLDL 
  \item L.VLDL.TG -4 mmol/l Triglycerides in large VLDL 
  \item M.VLDL.P -4 mol/l Concentration of medium VLDL particles
  \item M.VLDL.L -4 mmol/l Total lipids in medium VLDL 
  \item M.VLDL.PL -4 mmol/l Phospholipids in medium VLDL 
  \item M.VLDL.C -4 mmol/l Total cholesterol in medium VLDL 
  \item M.VLDL.CE -4 mmol/l Cholesterol esters in medium VLDL 
  \item M.VLDL.FC -4 mmol/l Free cholesterol in medium VLDL 
  \item M.VLDL.TG -4 mmol/l Triglycerides in medium VLDL 
  \item S.VLDL.P -4 mol/l Concentration of small VLDL particles
  \item S.VLDL.L -4 mmol/l Total lipids in small VLDL 
  \item S.VLDL.PL -4 mmol/l Phospholipids in small VLDL 
  \item S.VLDL.C -4 mmol/l Total cholesterol in small VLDL 
  \item S.VLDL.CE -4 mmol/l Cholesterol esters in small VLDL 
  \item S.VLDL.FC -4 mmol/l Free cholesterol in small VLDL 
  \item S.VLDL.TG -4 mmol/l Triglycerides in small VLDL 
  \item XS.VLDL.P -4 mol/l Concentration of very small VLDL particles
  \item XS.VLDL.L -4 mmol/l Total lipids in very small VLDL 
  \item XS.VLDL.PL -4 mmol/l Phospholipids in very small VLDL 
  \item XS.VLDL.C -4 mmol/l Total cholesterol in very small VLDL 
  \item XS.VLDL.CE -4 mmol/l Cholesterol esters in very small VLDL 
  \item XS.VLDL.FC -4 mmol/l Free cholesterol in very small VLDL 
  \item XS.VLDL.TG -4 mmol/l Triglycerides in very small VLDL 
  \item IDL.P -4 mol/l Concentration of IDL particles
  \item IDL.L -4 mmol/l Total lipids in IDL 
  \item IDL.PL -4 mmol/l Phospholipids in IDL 
  \item IDL.C -4 mmol/l Total cholesterol in IDL 
  \item IDL.CE -4 mmol/l Cholesterol esters in IDL 
  \item IDL.FC -4 mmol/l Free cholesterol in IDL 
  \item IDL.TG -4 mmol/l Triglycerides in IDL 
  \item L.LDL.P -4 mol/l Concentration of large LDL particles
  \item L.LDL.L -4 mmol/l Total lipids in large LDL 
  \item L.LDL.PL -4 mmol/l Phospholipids in large LDL 
  \item L.LDL.C -4 mmol/l Total cholesterol in large LDL 
  \item L.LDL.CE -4 mmol/l Cholesterol esters in large LDL 
  \item L.LDL.FC -4 mmol/l Free cholesterol in large LDL 
  \item L.LDL.TG -4 mmol/l Triglycerides in large LDL 
  \item M.LDL.P -4 mol/l Concentration of medium LDL particles
  \item M.LDL.L -4 mmol/l Total lipids in medium LDL 
  \item M.LDL.PL -4 mmol/l Phospholipids in medium LDL 
  \item M.LDL.C -4 mmol/l Total cholesterol in medium LDL 
  \item M.LDL.CE -4 mmol/l Cholesterol esters in medium LDL 
  \item M.LDL.FC -4 mmol/l Free cholesterol in medium LDL 
  \item M.LDL.TG -4 mmol/l Triglycerides in medium LDL 
  \item S.LDL.P -4 mol/l Concentration of small LDL particles
  \item S.LDL.L -4 mmol/l Total lipids in small LDL 
  \item S.LDL.PL -4 mmol/l Phospholipids in small LDL 
  \item S.LDL.C -4 mmol/l Total cholesterol in small LDL 
  \item S.LDL.CE -4 mmol/l Cholesterol esters in small LDL 
  \item S.LDL.FC -4 mmol/l Free cholesterol in small LDL 
  \item S.LDL.TG -4 mmol/l Triglycerides in small LDL 
  \item XL.HDL.P -4 mol/l Concentration of very large HDL particles
  \item XL.HDL.L -4 mmol/l Total lipids in very large HDL 
  \item XL.HDL.PL -4 mmol/l Phospholipids in very large HDL 
  \item XL.HDL.C -4 mmol/l Total cholesterol in very large HDL 
  \item XL.HDL.CE -4 mmol/l Cholesterol esters in very large HDL 
  \item XL.HDL.FC -4 mmol/l Free cholesterol in very large HDL 
  \item XL.HDL.TG -4 mmol/l Triglycerides in very large HDL 
  \item L.HDL.P -4 mol/l Concentration of large HDL particles
  \item L.HDL.L -4 mmol/l Total lipids in large HDL 
  \item L.HDL.PL -4 mmol/l Phospholipids in large HDL 
  \item L.HDL.C -4 mmol/l Total cholesterol in large HDL 
  \item L.HDL.CE -4 mmol/l Cholesterol esters in large HDL 
  \item L.HDL.FC -4 mmol/l Free cholesterol in large HDL 
  \item L.HDL.TG -4 mmol/l Triglycerides in large HDL 
  \item M.HDL.P -4 mol/l Concentration of medium HDL particles
  \item M.HDL.L -4 mmol/l Total lipids in medium HDL 
  \item M.HDL.PL -4 mmol/l Phospholipids in medium HDL 
  \item M.HDL.C -4 mmol/l Total cholesterol in medium HDL 
  \item M.HDL.CE -4 mmol/l Cholesterol esters in medium HDL 
  \item M.HDL.FC -4 mmol/l Free cholesterol in medium HDL 
  \item M.HDL.TG -4 mmol/l Triglycerides in medium HDL 
  \item S.HDL.P -4 mol/l Concentration of small HDL particles
  \item S.HDL.L -4 mmol/l Total lipids in small HDL 
  \item S.HDL.PL -4 mmol/l Phospholipids in small HDL 
  \item S.HDL.C -4 mmol/l Total cholesterol in small HDL 
  \item S.HDL.CE -4 mmol/l Cholesterol esters in small HDL 
  \item S.HDL.FC -4 mmol/l Free cholesterol in small HDL 
  \item S.HDL.TG -4 mmol/l Triglycerides in small HDL 
  \item XXL.VLDL.PL_. -4 % Phospholipids to total lipds ratio in chylomicrons and extremely large VLDL 
  \item XXL.VLDL.C_. -4 % Total cholesterol to total lipids ratio in chylomicrons and extremely large VLDL 
  \item XXL.VLDL.CE_. -4 % Cholesterol esters to total lipids ratio in chylomicrons and extremely large VLDL 
  \item XXL.VLDL.FC_. -4 % Free cholesterol to total lipids ratio in chylomicrons and extremely large VLDL 
  \item XXL.VLDL.TG_. -4 % Triglycerides to total lipids ratio in chylomicrons and extremely large VLDL 
  \item XL.VLDL.PL_. -4 % Phospholipids to total lipds ratio in very large VLDL 
  \item XL.VLDL.C_. -4 % Total cholesterol to total lipids ratio in very large VLDL 
  \item XL.VLDL.CE_. -4 % Cholesterol esters to total lipids ratio in very large VLDL 
  \item XL.VLDL.FC_. -4 % Free cholesterol to total lipids ratio in very large VLDL 
  \item XL.VLDL.TG_. -4 % Triglycerides to total lipids ratio in very large VLDL 
  \item L.VLDL.PL_. -4 % Phospholipids to total lipds ratio in large VLDL 
  \item L.VLDL.C_. -4 % Total cholesterol to total lipids ratio in large VLDL 
  \item L.VLDL.CE_. -4 % Cholesterol esters to total lipids ratio in large VLDL 
  \item L.VLDL.FC_. -4 % Free cholesterol to total lipids ratio in large VLDL 
  \item L.VLDL.TG_. -4 % Triglycerides to total lipids ratio in large VLDL 
  \item M.VLDL.PL_. -4 % Phospholipids to total lipds ratio in medium VLDL 
  \item M.VLDL.C_. -4 % Total cholesterol to total lipids ratio in medium VLDL 
  \item M.VLDL.CE_. -4 % Cholesterol esters to total lipids ratio in medium VLDL 
  \item M.VLDL.FC_. -4 % Free cholesterol to total lipids ratio in medium VLDL 
  \item M.VLDL.TG_. -4 % Triglycerides to total lipids ratio in medium VLDL 
  \item S.VLDL.PL_. -4 % Phospholipids to total lipds ratio in small VLDL 
  \item S.VLDL.C_. -4 % Total cholesterol to total lipids ratio in small VLDL 
  \item S.VLDL.CE_. -4 % Cholesterol esters to total lipids ratio in small VLDL 
  \item S.VLDL.FC_. -4 % Free cholesterol to total lipids ratio in small VLDL 
  \item S.VLDL.TG_. -4 % Triglycerides to total lipids ratio in small VLDL 
  \item XS.VLDL.PL_. -4 % Phospholipids to total lipds ratio in very small VLDL 
  \item XS.VLDL.C_. -4 % Total cholesterol to total lipids ratio in very small VLDL 
  \item XS.VLDL.CE_. -4 % Cholesterol esters to total lipids ratio in very small VLDL 
  \item XS.VLDL.FC_. -4 % Free cholesterol to total lipids ratio in very small VLDL 
  \item XS.VLDL.TG_. -4 % Triglycerides to total lipids ratio in very small VLDL 
  \item IDL.PL_. -4 % Phospholipids to total lipds ratio in IDL 
  \item IDL.C_. -4 % Total cholesterol to total lipids ratio in IDL 
  \item IDL.CE_. -4 % Cholesterol esters to total lipids ratio in IDL 
  \item IDL.FC_. -4 % Free cholesterol to total lipids ratio in IDL 
  \item IDL.TG_. -4 % Triglycerides to total lipids ratio in IDL 
  \item L.LDL.PL_. -4 % Phospholipids to total lipds ratio in large LDL 
  \item L.LDL.C_. -4 % Total cholesterol to total lipids ratio in large LDL 
  \item L.LDL.CE_. -4 % Cholesterol esters to total lipids ratio in large LDL 
  \item L.LDL.FC_. -4 % Free cholesterol to total lipids ratio in large LDL 
  \item L.LDL.TG_. -4 % Triglycerides to total lipids ratio in large LDL 
  \item M.LDL.PL_. -4 % Phospholipids to total lipds ratio in medium LDL 
  \item M.LDL.C_. -4 % Total cholesterol to total lipids ratio in medium LDL 
  \item M.LDL.CE_. -4 % Cholesterol esters to total lipids ratio in medium LDL 
  \item M.LDL.FC_. -4 % Free cholesterol to total lipids ratio in medium LDL 
  \item M.LDL.TG_. -4 % Triglycerides to total lipids ratio in medium LDL 
  \item S.LDL.PL_. -4 % Phospholipids to total lipds ratio in small LDL 
  \item S.LDL.C_. -4 % Total cholesterol to total lipids ratio in small LDL 
  \item S.LDL.CE_. -4 % Cholesterol esters to total lipids ratio in small LDL 
  \item S.LDL.FC_. -4 % Free cholesterol to total lipids ratio in small LDL 
  \item S.LDL.TG_. -4 % Triglycerides to total lipids ratio in small LDL 
  \item XL.HDL.PL_. -4 % Phospholipids to total lipds ratio in very large HDL 
  \item XL.HDL.C_. -4 % Total cholesterol to total lipids ratio in very large HDL 
  \item XL.HDL.CE_. -4 % Cholesterol esters to total lipids ratio in very large HDL 
  \item XL.HDL.FC_. -4 % Free cholesterol to total lipids ratio in very large HDL 
  \item XL.HDL.TG_. -4 % Triglycerides to total lipids ratio in very large HDL 
  \item L.HDL.PL_. -4 % Phospholipids to total lipds ratio in large HDL 
  \item L.HDL.C_. -4 % Total cholesterol to total lipids ratio in large HDL 
  \item L.HDL.CE_. -4 % Cholesterol esters to total lipids ratio in large HDL 
  \item L.HDL.FC_. -4 % Free cholesterol to total lipids ratio in large HDL 
  \item L.HDL.TG_. -4 % Triglycerides to total lipids ratio in large HDL 
  \item M.HDL.PL_. -4 % Phospholipids to total lipds ratio in medium HDL 
  \item M.HDL.C_. -4 % Total cholesterol to total lipids ratio in medium HDL 
  \item M.HDL.CE_. -4 % Cholesterol esters to total lipids ratio in medium HDL 
  \item M.HDL.FC_. -4 % Free cholesterol to total lipids ratio in medium HDL 
  \item M.HDL.TG_. -4 % Triglycerides to total lipids ratio in medium HDL 
  \item S.HDL.PL_. -4 % Phospholipids to total lipds ratio in small HDL 
  \item S.HDL.C_. -4 % Total cholesterol to total lipids ratio in small HDL 
  \item S.HDL.CE_. -4 % Cholesterol esters to total lipids ratio in small HDL 
  \item S.HDL.FC_. -4 % Free cholesterol to total lipids ratio in small HDL 
  \item S.HDL.TG_. -4 % Triglycerides to total lipids ratio in small HDL 
  \item VLDL.D -4 nm Mean diameter for VLDL particles
  \item LDL.D -4 nm Mean diameter for LDL particles
  \item HDL.D -4 nm Mean diameter for HDL particles
  \item SERUM.C -4 mmol/l Serum total cholesterol
  \item VLDL.C -4 mmol/l Total cholesterol in VLDL
  \item REMNANT.C -4 mmol/l Remnant cholesterol (non-HDL, non-LDL -cholesterol)
  \item LDL.C -4 mmol/l Total cholesterol in LDL
  \item HDL.C -4 mmol/l Total cholesterol in HDL
  \item HDL2.C -4 mmol/l Total cholesterol in HDL2
  \item HDL3.C -4 mmol/l Total cholesterol in HDL3
  \item ESTC -4 mmol/l Esterified cholesterol
  \item FREEC -4 mmol/l Free cholesterol
  \item SERUM.TG -4 mmol/l Serum total triglycerides
  \item VLDL.TG -4 mmol/l Triglycerides in VLDL
  \item LDL.TG -4 mmol/l Triglycerides in LDL
  \item HDL.TG -4 mmol/l Triglycerides in HDL
  \item TOTPG -4 mmol/l Total phosphoglycerides
  \item TG.PG -4  Ratio of triglycerides to phosphoglycerides
  \item PC -4 mmol/l Phosphatidylcholine and other cholines
  \item SM -4 mmol/l Sphingomyelins
  \item TOTCHO -4 mmol/l Total cholines
  \item APOA1 -4 g/l Apolipoprotein A-I
  \item APOB -4 g/l Apolipoprotein B
  \item APOB.APOA1 -4  Ratio of apolipoprotein B to apolipoprotein A-I
  \item TOTFA -4 mmol/l Total fatty acids
  \item UNSAT -4  Estimated degree of unsaturation
  \item DHA -4 mmol/l 22:6, docosahexaenoic acid
  \item LA -4 mmol/l 18:2, linoleic acid
  \item FAW3 -4 mmol/l Omega-3 fatty acids
  \item FAW6 -4 mmol/l Omega-6 fatty acids
  \item PUFA -4 mmol/l Polyunsaturated fatty acids
  \item MUFA -4 mmol/l Monounsaturated fatty acids; 16:1, 18:1
  \item SFA -4 mmol/l Saturated fatty acids
  \item DHA.FA -4 % Ratio of 22:6 docosahexaenoic acid to total fatty acids
  \item LA.FA -4 % Ratio of 18:2 linoleic acid to total fatty acids
  \item FAW3.FA -4 % Ratio of omega-3 fatty acids to total fatty acids
  \item FAW6.FA -4 % Ratio of omega-6 fatty acids to total fatty acids
  \item PUFA.FA -4 % Ratio of polyunsaturated fatty acids to total fatty acids
  \item MUFA.FA -4 % Ratio of monounsaturated fatty acids to total fatty acids
  \item SFA.FA -4 % Ratio of saturated fatty acids to total fatty acids
  \item GLC -4 mmol/l Glucose
  \item LAC -4 mmol/l Lactate
  \item PYR -4 mmol/l Pyruvate
  \item CIT -4 mmol/l Citrate
  \item GLOL -4 mmol/l Glycerol
  \item ALA -4 mmol/l Alanine
  \item GLN -4 mmol/l Glutamine
  \item GLY -4 mmol/l Glycine
  \item HIS -4 mmol/l Histidine
  \item ILE -4 mmol/l Isoleucine
  \item LEU -4 mmol/l Leucine
  \item VAL -4 mmol/l Valine
  \item PHE -4 mmol/l Phenylalanine
  \item TYR -4 mmol/l Tyrosine
  \item ACE -4 mmol/l Acetate
  \item ACACE -4 mmol/l Acetoacetate
  \item BOHBUT -4 mmol/l 3-hydroxybutyrate
  \item CREA -4 mmol/l Creatinine
  \item ALB -4 signal area Albumin
  \item GP -4 mmol/l Glycoprotein acetyls, mainly a1-acid glycoprotein
}
}
\examples{
\dontrun{
describe(adninightingale)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/ADNINIGHTINGALE/ADMC_Nightingale_Platform-Methods_Description.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNINIGHTINGALE/ADMC_Nightingale_Platform-Report.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNINIGHTINGALE/ADMC_Nightingale_Platform-Result_file_description.pdf}
}
\keyword{datasets}
