% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{adninightingale2}
\alias{adninightingale2}
\title{ADMC Nightingale Platform NMR Post-Unblinding Re-Analysis of Lipoproteins and Metabolites}
\format{
A data frame with 1697 rows and 277 variables
}
\usage{
data(adninightingale2)
}
\description{
ADMC Nightingale Platform NMR Post-Unblinding Re-Analysis of Lipoproteins and Metabolites
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item SAMPLEID -4 -4 -4
  \item GLOBAL.SPEC.ID -4 -4 -4
  \item EDTA_PLASMA -4 -4 Detected EDTA anticoagulant. This indicates a EDTA plasma sample. EDTA anticoagulant interferes with the quantification of glycerol, glycine and pyruvate. Also, metabolites may behave differently (mainly in absolute concentration terms, but potentially also in association terms) in serum and EDTA plasma due to the different biological matrices.
  \item CITRATE_PLASMA -4 -4 A non-physiological citrate concentration. This may result from the sample being citrate plasma.
  \item LOW_ETHANOL -4 -4 This arises potentially due to disinfectant contaminations in the blood donation/collection process and/or contamination in air from lab handling, such as cleaning of well-plates. In higher concentrations, ethanol prohibits the quantification of glycerol, and therefore the sample has been flagged with a "TAG", and no concentration has been provided for the sample. In lower concentrations, it may interfere with the reliable quantification of glycerol and result in artificially distorted glycerol levels. We recommend caution when interpreting absolute glycerol concentrations. It is also good to note that high alcohol in the sample might artificially increase the acetate concentration.
  \item MEDIUM_ETHANOL -4 -4 This arises potentially due to disinfectant contaminations in the blood donation/collection process and/or contamination in air from lab handling, such as cleaning of well-plates. In higher concentrations, ethanol prohibits the quantification of glycerol and 3-hydroxybutyrate, and therefore the sample has been flagged with a "TAG", and no concentration has been provided for the sample. In lower concentrations, it may interfere with the reliable quantification of glycerol and 3-hydroxybutyrate and result in artificially distorted glycerol and 3-hydroxybutyrate levels. We recommend caution when interpreting absolute glycerol and 3-hydroxybutyrate concentrations. It is also good to note that high alcohol in the sample might artificially increase the acetate concentration.
  \item HIGH_ETHANOL -4 -4 This arises potentially due to disinfectant contaminations in the blood donation/collection process and/or contamination in air from lab handling, such as cleaning of well-plates. In higher concentrations, ethanol prohibits the quantification of glycerol, 3-hydroxybutyrate and valine, and therefore the sample has been flagged with a "TAG", and no concentration has been provided for the sample. In lower concentrations, it may interfere with the reliable quantification of glycerol, 3-hydroxybutyrate and valine and result in artificially distorted glycerol, 3-hydroxybutyrate and valine levels. We recommend caution when interpreting absolute glycerol, 3-hydroxybutyrate and valine concentrations. It is also good to note that high alcohol in the sample might artificially increase the acetate concentration.
  \item ISOPROPYL_ALCOHOL -4 -4 Isopropyl alcohol signals detected. In higher concentrations, this compound prohibits the quantification of creatinine, and therefore the sample has been flagged with a "TAG", and no concentration has been provided for the sample. In lower concentrations, it may interfere with the reliable quantification of creatinine and result in artificially distorted creatinine levels. We recommend caution when interpreting absolute creatinine concentrations.
  \item N_METHYL_2_PYRROLIDONE -4 -4 A compound that can be used in sample collection processes. In higher concentrations, it prohibits the quantification of glutamine, and therefore the sample has been flagged with a "TAG", and no concentration has been provided for the sample. In lower concentrations, it may interfere with the reliable quantification of glutamine and result in artificially distorted glutamine levels. We recommend caution when interpreting absolute glutamine concentrations.
  \item POLYSACCHARIDES -4 -4 Polysaccharide or some other type of carbohydrate signals detected. Most commonly observed along with other signs of suboptimal sample quality, but may also indicate presence of heparin. In higher concentrations, this compound prohibits the quantification of creatinine, and therefore the sample has been flagged with a "TAG", and no concentration has been provided for the sample. In lower concentrations, it may interfere with the reliable quantification of creatinine and result in artificially distorted creatinine levels. We recommend caution when interpreting absolute creatinine concentrations.
  \item AMINOCAPROIC_ACID -4 -4 Aminocaproic acid detected in the sample; this compound is an antifibrinolytic agent. In higher concentrations, it prohibits the quantification of acetone, and therefore the sample has been flagged with a "TAG", and no concentration has been provided for the sample.
  \item LOW_GLUCOSE -4 -4 Low glucose is tagged since the physiological concentration is commonly well controlled, particularly in the low concentration end. However, glucose is metabolised to pyruvate and lactate if the blood sample is kept with the cells during the sample collection process, in particular at room temperature but also to some extent at fridge temperature. The low glucose tag thus commonly reflects extended time periods prior to preparation to serum/plasma. As a result of the glycolysis, high lactate and/or high pyruvate is commonly observed along with low glucose levels.
  \item HIGH_LACTATE -4 -4 Low glucose is tagged since the physiological concentration is commonly well controlled, particularly in the low concentration end. However, glucose is metabolised to pyruvate and lactate if the blood sample is kept with the cells during the sample collection process, in particular at room temperature but also to some extent at fridge temperature. The low glucose tag thus commonly reflects extended time periods prior to preparation to serum/plasma. As a result of the glycolysis, high lactate and/or high pyruvate is commonly observed along with low glucose levels.
  \item HIGH_PYRUVATE -4 -4 Low glucose is tagged since the physiological concentration is commonly well controlled, particularly in the low concentration end. However, glucose is metabolised to pyruvate and lactate if the blood sample is kept with the cells during the sample collection process, in particular at room temperature but also to some extent at fridge temperature. The low glucose tag thus commonly reflects extended time periods prior to preparation to serum/plasma. As a result of the glycolysis, high lactate and/or high pyruvate is commonly observed along with low glucose levels.
  \item LOW_GLUTAMINE_OR_HIGH_GLUTAMATE -4 -4 In these samples most of the glutamine appears to have degraded into glutamate. This can take place, e.g. during the sample collection process, if the sample is kept at room temperature for a prolonged period of time, there are multiple freeze-thaw cycles, or there are oxidative conditions / oxidation occurs in the sample. In these samples glutamine is not quantified. In higher concentrations, glutamate prohibits the quantification of pyruvate, and therefore the sample has been flagged with a "TAG", and no concentration has been provided for the sample. In lower concentrations, glutamate may interfere with the reliable quantification of pyruvate and result in artificially distorted pyruvate levels. We recommend caution when interpreting absolute pyruvate concentrations.
  \item GLUCONOLACTONE -4 -4 Gluconolactone is an oxidation product of glucose and is not normally detected in samples. Interferes with the quantification of creatinine.
  \item LOW_PROTEIN -4 -4 Unexpected protein features have been identified during the analysis. This quality observation is related to unexpectedly low total protein concentration in the sample and may be an indication of potential sample dilution. Sample dilution may occur due to water contamination, or prolonged storage in freezer in combination with loosely closed tube caps, which exposes the sample to collect ice and increases the chances of dilution.In case there is reason to assume sample dilution, we recommend caution when analyzing absolute concentrations.
  \item UNEXPECTED_AMINO_ACID_SIGNALS -4 -4 These are signals that we do not typically observe in normal samples. It is likely that they relate to protein breakage due to non optimal storage.Interferes with the reliable quantification of citrate and glycerol, and may result in artificially distorted metabolite concentrations.
  \item UNIDENTIFIED_MACROMOLECULES -4 -4 Signals from abnormal macromolecule(s)/protein(s).Interferes with the quantification of phenylalanine, acetate, acetone, glutamine and pyruvate.
  \item UNIDENTIFIED_SMALL_MOLECULE_A -4 -4 Untypical signals from unknown small molecule(s); potentially xenobiotics.
  \item UNIDENTIFIED_SMALL_MOLECULE_B -4 -4 Untypical signals from unknown small molecule(s); potentially xenobiotics. Interferes with the quantification of acetate.
  \item UNIDENTIFIED_SMALL_MOLECULE_C -4 -4 Unidentified signals at 2.27 ppm and various other spectral areas. Interferes with the reliable quantification of albumin, creatinine, glycerol, glycine and acetone.
  \item BELOW_LIMIT_OF_QUANTIFICATION -4 -4 Below Limit-Of-Quantification (LOQ) means that the concentration of the given biomarker is smaller than the range where the Nightingale platform offers highly accurate quantification. For instance, the LOQ for lipoprotein lipid measures is < 0.01 mmol/L. The estimated biomarker concentration is reported for use in certain epidemiological analyses, but it may be advisable to conduct sensitivity analyses with all values below LOQ excluded.
  \item TOTAL_C -4 mmol/l Total cholesterol
  \item NON_HDL_C -4 mmol/l Total cholesterol minus HDL-C
  \item REMNANT_C -4 mmol/l Remnant cholesterol (non-HDL, non-LDL -cholesterol)
  \item VLDL_C -4 mmol/l VLDL cholesterol
  \item CLINICAL_LDL_C -4 mmol/l Clinical LDL cholesterol
  \item LDL_C -4 mmol/l LDL cholesterol
  \item HDL_C -4 mmol/l HDL cholesterol
  \item TOTAL_TG -4 mmol/l Total triglycerides
  \item VLDL_TG -4 mmol/l Triglycerides in VLDL
  \item LDL_TG -4 mmol/l Triglycerides in LDL
  \item HDL_TG -4 mmol/l Triglycerides in HDL
  \item TOTAL_PL -4 mmol/l Total phospholipids in lipoprotein particles
  \item VLDL_PL -4 mmol/l Phospholipids in VLDL
  \item LDL_PL -4 mmol/l Phospholipids in LDL
  \item HDL_PL -4 mmol/l Phospholipids in HDL
  \item TOTAL_CE -4 mmol/l Total esterified cholesterol
  \item VLDL_CE -4 mmol/l Cholesteryl esters in VLDL
  \item LDL_CE -4 mmol/l Cholesteryl esters in LDL
  \item HDL_CE -4 mmol/l Cholesteryl esters in HDL
  \item TOTAL_FC -4 mmol/l Total free cholesterol
  \item VLDL_FC -4 mmol/l Free cholesterol in VLDL
  \item LDL_FC -4 mmol/l Free cholesterol in LDL
  \item HDL_FC -4 mmol/l Free cholesterol in HDL
  \item TOTAL_L -4 mmol/l Total lipids in lipoprotein particles
  \item VLDL_L -4 mmol/l Total lipids in VLDL
  \item LDL_L -4 mmol/l Total lipids in LDL
  \item HDL_L -4 mmol/l Total lipids in HDL
  \item TOTAL_P -4 mmol/l Total concentration of lipoprotein particles
  \item VLDL_P -4 mmol/l Concentration of VLDL particles
  \item LDL_P -4 mmol/l Concentration of LDL particles
  \item HDL_P -4 mmol/l Concentration of HDL particles
  \item VLDL_SIZE -4 nm Average diameter for VLDL particles
  \item LDL_SIZE -4 nm Average diameter for LDL particles
  \item HDL_SIZE -4 nm Average diameter for HDL particles
  \item PHOSPHOGLYC -4 mmol/l Phosphoglycerides
  \item TG_BY_PG -4 ratio Ratio of triglycerides to phosphoglycerides
  \item CHOLINES -4 mmol/l Total cholines
  \item PHOSPHATIDYLC -4 mmol/l Phosphatidylcholines
  \item SPHINGOMYELINS -4 mmol/l Sphingomyelins
  \item APOB -4 g/l Apolipoprotein B
  \item APOA1 -4 g/l Apolipoprotein A1
  \item APOB_BY_APOA1 -4 ratio Ratio of apolipoprotein B to apolipoprotein A1
  \item TOTAL_FA -4 mmol/l Total fatty acids
  \item UNSATURATION -4 degree Degree of unsaturation
  \item OMEGA_3 -4 mmol/l Omega-3 fatty acids
  \item OMEGA_6 -4 mmol/l Omega-6 fatty acids
  \item PUFA -4 mmol/l Polyunsaturated fatty acids
  \item MUFA -4 mmol/l Monounsaturated fatty acids
  \item SFA -4 mmol/l Saturated fatty acids
  \item LA -4 mmol/l Linoleic acid
  \item DHA -4 mmol/l Docosahexaenoic acid
  \item OMEGA_3_PCT -4 % Ratio of omega-3 fatty acids to total fatty acids
  \item OMEGA_6_PCT -4 % Ratio of omega-6 fatty acids to total fatty acids
  \item PUFA_PCT -4 % Ratio of polyunsaturated fatty acids to total fatty acids
  \item MUFA_PCT -4 % Ratio of monounsaturated fatty acids to total fatty acids
  \item SFA_PCT -4 % Ratio of saturated fatty acids to total fatty acids
  \item LA_PCT -4 % Ratio of linoleic acid to total fatty acids
  \item DHA_PCT -4 % Ratio of docosahexaenoic acid to total fatty acids
  \item PUFA_BY_MUFA -4 ratio Ratio of polyunsaturated fatty acids to monounsaturated fatty acids
  \item OMEGA_6_BY_OMEGA_3 -4 ratio Ratio of omega-6 fatty acids to omega-3 fatty acids
  \item ALA -4 mmol/l Alanine
  \item GLN -4 mmol/l Glutamine
  \item GLY -4 mmol/l Glycine
  \item HIS -4 mmol/l Histidine
  \item TOTAL_BCAA -4 mmol/l Total concentration of branched-chain amino acids (leucine + isoleucine + valine)
  \item ILE -4 mmol/l Isoleucine
  \item LEU -4 mmol/l Leucine
  \item VAL -4 mmol/l Valine
  \item PHE -4 mmol/l Phenylalanine
  \item TYR -4 mmol/l Tyrosine
  \item GLUCOSE -4 mmol/l Glucose
  \item LACTATE -4 mmol/l Lactate
  \item PYRUVATE -4 mmol/l Pyruvate
  \item CITRATE -4 mmol/l Citrate
  \item GLYCEROL -4 mmol/l Glycerol
  \item BOHBUTYRATE -4 mmol/l 3-Hydroxybutyrate
  \item ACETATE -4 mmol/l Acetate
  \item ACETOACETATE -4 mmol/l Acetoacetate
  \item ACETONE -4 mmol/l Acetone
  \item CREATININE -4 ?mol/l Creatinine
  \item ALBUMIN -4 g/l Albumin
  \item GLYCA -4 mmol/l Glycoprotein acetyls
  \item XXL_VLDL_P -4 mmol/l Concentration of chylomicrons and extremely large VLDL particles
  \item XXL_VLDL_L -4 mmol/l Total lipids in chylomicrons and extremely large VLDL
  \item XXL_VLDL_PL -4 mmol/l Phospholipids in chylomicrons and extremely large VLDL
  \item XXL_VLDL_C -4 mmol/l Cholesterol in chylomicrons and extremely large VLDL
  \item XXL_VLDL_CE -4 mmol/l Cholesteryl esters in chylomicrons and extremely large VLDL
  \item XXL_VLDL_FC -4 mmol/l Free cholesterol in chylomicrons and extremely large VLDL
  \item XXL_VLDL_TG -4 mmol/l Triglycerides in chylomicrons and extremely large VLDL
  \item XL_VLDL_P -4 mmol/l Concentration of very large VLDL particles
  \item XL_VLDL_L -4 mmol/l Total lipids in very large VLDL
  \item XL_VLDL_PL -4 mmol/l Phospholipids in very large VLDL
  \item XL_VLDL_C -4 mmol/l Cholesterol in very large VLDL
  \item XL_VLDL_CE -4 mmol/l Cholesteryl esters in very large VLDL
  \item XL_VLDL_FC -4 mmol/l Free cholesterol in very large VLDL
  \item XL_VLDL_TG -4 mmol/l Triglycerides in very large VLDL
  \item L_VLDL_P -4 mmol/l Concentration of large VLDL particles
  \item L_VLDL_L -4 mmol/l Total lipids in large VLDL
  \item L_VLDL_PL -4 mmol/l Phospholipids in large VLDL
  \item L_VLDL_C -4 mmol/l Cholesterol in large VLDL
  \item L_VLDL_CE -4 mmol/l Cholesteryl esters in large VLDL
  \item L_VLDL_FC -4 mmol/l Free cholesterol in large VLDL
  \item L_VLDL_TG -4 mmol/l Triglycerides in large VLDL
  \item M_VLDL_P -4 mmol/l Concentration of medium VLDL particles
  \item M_VLDL_L -4 mmol/l Total lipids in medium VLDL
  \item M_VLDL_PL -4 mmol/l Phospholipids in medium VLDL
  \item M_VLDL_C -4 mmol/l Cholesterol in medium VLDL
  \item M_VLDL_CE -4 mmol/l Cholesteryl esters in medium VLDL
  \item M_VLDL_FC -4 mmol/l Free cholesterol in medium VLDL
  \item M_VLDL_TG -4 mmol/l Triglycerides in medium VLDL
  \item S_VLDL_P -4 mmol/l Concentration of small VLDL particles
  \item S_VLDL_L -4 mmol/l Total lipids in small VLDL
  \item S_VLDL_PL -4 mmol/l Phospholipids in small VLDL
  \item S_VLDL_C -4 mmol/l Cholesterol in small VLDL
  \item S_VLDL_CE -4 mmol/l Cholesteryl esters in small VLDL
  \item S_VLDL_FC -4 mmol/l Free cholesterol in small VLDL
  \item S_VLDL_TG -4 mmol/l Triglycerides in small VLDL
  \item XS_VLDL_P -4 mmol/l Concentration of very small VLDL particles
  \item XS_VLDL_L -4 mmol/l Total lipids in very small VLDL
  \item XS_VLDL_PL -4 mmol/l Phospholipids in very small VLDL
  \item XS_VLDL_C -4 mmol/l Cholesterol in very small VLDL
  \item XS_VLDL_CE -4 mmol/l Cholesteryl esters in very small VLDL
  \item XS_VLDL_FC -4 mmol/l Free cholesterol in very small VLDL
  \item XS_VLDL_TG -4 mmol/l Triglycerides in very small VLDL
  \item IDL_P -4 mmol/l Concentration of IDL particles
  \item IDL_L -4 mmol/l Total lipids in IDL
  \item IDL_PL -4 mmol/l Phospholipids in IDL
  \item IDL_C -4 mmol/l Cholesterol in IDL
  \item IDL_CE -4 mmol/l Cholesteryl esters in IDL
  \item IDL_FC -4 mmol/l Free cholesterol in IDL
  \item IDL_TG -4 mmol/l Triglycerides in IDL
  \item L_LDL_P -4 mmol/l Concentration of large LDL particles
  \item L_LDL_L -4 mmol/l Total lipids in large LDL
  \item L_LDL_PL -4 mmol/l Phospholipids in large LDL
  \item L_LDL_C -4 mmol/l Cholesterol in large LDL
  \item L_LDL_CE -4 mmol/l Cholesteryl esters in large LDL
  \item L_LDL_FC -4 mmol/l Free cholesterol in large LDL
  \item L_LDL_TG -4 mmol/l Triglycerides in large LDL
  \item M_LDL_P -4 mmol/l Concentration of medium LDL particles
  \item M_LDL_L -4 mmol/l Total lipids in medium LDL
  \item M_LDL_PL -4 mmol/l Phospholipids in medium LDL
  \item M_LDL_C -4 mmol/l Cholesterol in medium LDL
  \item M_LDL_CE -4 mmol/l Cholesteryl esters in medium LDL
  \item M_LDL_FC -4 mmol/l Free cholesterol in medium LDL
  \item M_LDL_TG -4 mmol/l Triglycerides in medium LDL
  \item S_LDL_P -4 mmol/l Concentration of small LDL particles
  \item S_LDL_L -4 mmol/l Total lipids in small LDL
  \item S_LDL_PL -4 mmol/l Phospholipids in small LDL
  \item S_LDL_C -4 mmol/l Cholesterol in small LDL
  \item S_LDL_CE -4 mmol/l Cholesteryl esters in small LDL
  \item S_LDL_FC -4 mmol/l Free cholesterol in small LDL
  \item S_LDL_TG -4 mmol/l Triglycerides in small LDL
  \item XL_HDL_P -4 mmol/l Concentration of very large HDL particles
  \item XL_HDL_L -4 mmol/l Total lipids in very large HDL
  \item XL_HDL_PL -4 mmol/l Phospholipids in very large HDL
  \item XL_HDL_C -4 mmol/l Cholesterol in very large HDL
  \item XL_HDL_CE -4 mmol/l Cholesteryl esters in very large HDL
  \item XL_HDL_FC -4 mmol/l Free cholesterol in very large HDL
  \item XL_HDL_TG -4 mmol/l Triglycerides in very large HDL
  \item L_HDL_P -4 mmol/l Concentration of large HDL particles
  \item L_HDL_L -4 mmol/l Total lipids in large HDL
  \item L_HDL_PL -4 mmol/l Phospholipids in large HDL
  \item L_HDL_C -4 mmol/l Cholesterol in large HDL
  \item L_HDL_CE -4 mmol/l Cholesteryl esters in large HDL
  \item L_HDL_FC -4 mmol/l Free cholesterol in large HDL
  \item L_HDL_TG -4 mmol/l Triglycerides in large HDL
  \item M_HDL_P -4 mmol/l Concentration of medium HDL particles
  \item M_HDL_L -4 mmol/l Total lipids in medium HDL
  \item M_HDL_PL -4 mmol/l Phospholipids in medium HDL
  \item M_HDL_C -4 mmol/l Cholesterol in medium HDL
  \item M_HDL_CE -4 mmol/l Cholesteryl esters in medium HDL
  \item M_HDL_FC -4 mmol/l Free cholesterol in medium HDL
  \item M_HDL_TG -4 mmol/l Triglycerides in medium HDL
  \item S_HDL_P -4 mmol/l Concentration of small HDL particles
  \item S_HDL_L -4 mmol/l Total lipids in small HDL
  \item S_HDL_PL -4 mmol/l Phospholipids in small HDL
  \item S_HDL_C -4 mmol/l Cholesterol in small HDL
  \item S_HDL_CE -4 mmol/l Cholesteryl esters in small HDL
  \item S_HDL_FC -4 mmol/l Free cholesterol in small HDL
  \item S_HDL_TG -4 mmol/l Triglycerides in small HDL
  \item XXL_VLDL_PL_PCT -4 % Phospholipids to total lipids ratio in chylomicrons and extremely large VLDL
  \item XXL_VLDL_C_PCT -4 % Cholesterol to total lipids ratio in chylomicrons and extremely large VLDL
  \item XXL_VLDL_CE_PCT -4 % Cholesteryl esters to total lipids ratio in chylomicrons and extremely large VLDL
  \item XXL_VLDL_FC_PCT -4 % Free cholesterol to total lipids ratio in chylomicrons and extremely large VLDL
  \item XXL_VLDL_TG_PCT -4 % Triglycerides to total lipids ratio in chylomicrons and extremely large VLDL
  \item XL_VLDL_PL_PCT -4 % Phospholipids to total lipids ratio in very large VLDL
  \item XL_VLDL_C_PCT -4 % Cholesterol to total lipids ratio in very large VLDL
  \item XL_VLDL_CE_PCT -4 % Cholesteryl esters to total lipids ratio in very large VLDL
  \item XL_VLDL_FC_PCT -4 % Free cholesterol to total lipids ratio in very large VLDL
  \item XL_VLDL_TG_PCT -4 % Triglycerides to total lipids ratio in very large VLDL
  \item L_VLDL_PL_PCT -4 % Phospholipids to total lipids ratio in large VLDL
  \item L_VLDL_C_PCT -4 % Cholesterol to total lipids ratio in large VLDL
  \item L_VLDL_CE_PCT -4 % Cholesteryl esters to total lipids ratio in large VLDL
  \item L_VLDL_FC_PCT -4 % Free cholesterol to total lipids ratio in large VLDL
  \item L_VLDL_TG_PCT -4 % Triglycerides to total lipids ratio in large VLDL
  \item M_VLDL_PL_PCT -4 % Phospholipids to total lipids ratio in medium VLDL
  \item M_VLDL_C_PCT -4 % Cholesterol to total lipids ratio in medium VLDL
  \item M_VLDL_CE_PCT -4 % Cholesteryl esters to total lipids ratio in medium VLDL
  \item M_VLDL_FC_PCT -4 % Free cholesterol to total lipids ratio in medium VLDL
  \item M_VLDL_TG_PCT -4 % Triglycerides to total lipids ratio in medium VLDL
  \item S_VLDL_PL_PCT -4 % Phospholipids to total lipids ratio in small VLDL
  \item S_VLDL_C_PCT -4 % Cholesterol to total lipids ratio in small VLDL
  \item S_VLDL_CE_PCT -4 % Cholesteryl esters to total lipids ratio in small VLDL
  \item S_VLDL_FC_PCT -4 % Free cholesterol to total lipids ratio in small VLDL
  \item S_VLDL_TG_PCT -4 % Triglycerides to total lipids ratio in small VLDL
  \item XS_VLDL_PL_PCT -4 % Phospholipids to total lipids ratio in very small VLDL
  \item XS_VLDL_C_PCT -4 % Cholesterol to total lipids ratio in very small VLDL
  \item XS_VLDL_CE_PCT -4 % Cholesteryl esters to total lipids ratio in very small VLDL
  \item XS_VLDL_FC_PCT -4 % Free cholesterol to total lipids ratio in very small VLDL
  \item XS_VLDL_TG_PCT -4 % Triglycerides to total lipids ratio in very small VLDL
  \item IDL_PL_PCT -4 % Phospholipids to total lipids ratio in IDL
  \item IDL_C_PCT -4 % Cholesterol to total lipids ratio in IDL
  \item IDL_CE_PCT -4 % Cholesteryl esters to total lipids ratio in IDL
  \item IDL_FC_PCT -4 % Free cholesterol to total lipids ratio in IDL
  \item IDL_TG_PCT -4 % Triglycerides to total lipids ratio in IDL
  \item L_LDL_PL_PCT -4 % Phospholipids to total lipids ratio in large LDL
  \item L_LDL_C_PCT -4 % Cholesterol to total lipids ratio in large LDL
  \item L_LDL_CE_PCT -4 % Cholesteryl esters to total lipids ratio in large LDL
  \item L_LDL_FC_PCT -4 % Free cholesterol to total lipids ratio in large LDL
  \item L_LDL_TG_PCT -4 % Triglycerides to total lipids ratio in large LDL
  \item M_LDL_PL_PCT -4 % Phospholipids to total lipids ratio in medium LDL
  \item M_LDL_C_PCT -4 % Cholesterol to total lipids ratio in medium LDL
  \item M_LDL_CE_PCT -4 % Cholesteryl esters to total lipids ratio in medium LDL
  \item M_LDL_FC_PCT -4 % Free cholesterol to total lipids ratio in medium LDL
  \item M_LDL_TG_PCT -4 % Triglycerides to total lipids ratio in medium LDL
  \item S_LDL_PL_PCT -4 % Phospholipids to total lipids ratio in small LDL
  \item S_LDL_C_PCT -4 % Cholesterol to total lipids ratio in small LDL
  \item S_LDL_CE_PCT -4 % Cholesteryl esters to total lipids ratio in small LDL
  \item S_LDL_FC_PCT -4 % Free cholesterol to total lipids ratio in small LDL
  \item S_LDL_TG_PCT -4 % Triglycerides to total lipids ratio in small LDL
  \item XL_HDL_PL_PCT -4 % Phospholipids to total lipids ratio in very large HDL
  \item XL_HDL_C_PCT -4 % Cholesterol to total lipids ratio in very large HDL
  \item XL_HDL_CE_PCT -4 % Cholesteryl esters to total lipids ratio in very large HDL
  \item XL_HDL_FC_PCT -4 % Free cholesterol to total lipids ratio in very large HDL
  \item XL_HDL_TG_PCT -4 % Triglycerides to total lipids ratio in very large HDL
  \item L_HDL_PL_PCT -4 % Phospholipids to total lipids ratio in large HDL
  \item L_HDL_C_PCT -4 % Cholesterol to total lipids ratio in large HDL
  \item L_HDL_CE_PCT -4 % Cholesteryl esters to total lipids ratio in large HDL
  \item L_HDL_FC_PCT -4 % Free cholesterol to total lipids ratio in large HDL
  \item L_HDL_TG_PCT -4 % Triglycerides to total lipids ratio in large HDL
  \item M_HDL_PL_PCT -4 % Phospholipids to total lipids ratio in medium HDL
  \item M_HDL_C_PCT -4 % Cholesterol to total lipids ratio in medium HDL
  \item M_HDL_CE_PCT -4 % Cholesteryl esters to total lipids ratio in medium HDL
  \item M_HDL_FC_PCT -4 % Free cholesterol to total lipids ratio in medium HDL
  \item M_HDL_TG_PCT -4 % Triglycerides to total lipids ratio in medium HDL
  \item S_HDL_PL_PCT -4 % Phospholipids to total lipids ratio in small HDL
  \item S_HDL_C_PCT -4 % Cholesterol to total lipids ratio in small HDL
  \item S_HDL_CE_PCT -4 % Cholesteryl esters to total lipids ratio in small HDL
  \item S_HDL_FC_PCT -4 % Free cholesterol to total lipids ratio in small HDL
  \item S_HDL_TG_PCT -4 % Triglycerides to total lipids ratio in small HDL
}
}
\examples{
\dontrun{
describe(adninightingale2)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/ADNINIGHTINGALE2/10.1515_ijb-2020-0107.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNINIGHTINGALE2/ADMC_Nightingale_Platform-Methods_Description-2020-update.pdf}, \url{https://adni.bitbucket.io/reference/docs/ADNINIGHTINGALE2/Nightingale_2020_quantification.pdf}
}
\keyword{datasets}
