% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{adninightingalelong}
\alias{adninightingalelong}
\title{ADMC Nightingale Platform  NMR Analysis of Lipoproteins and Metabolites Longitudinal}
\format{
A data frame with 4891 rows and 279 variables
}
\usage{
data(adninightingalelong)
}
\description{
ADMC Nightingale Platform  NMR Analysis of Lipoproteins and Metabolites Longitudinal
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item GUSPECID -4 -4 -4
  \item VID -4 -4 -4
  \item DRAWTIME -4 -4 -4
  \item PROT -4 -4 -4
  \item EDTA_PLASMA -4 -4 -4
  \item CITRATE_PLASMA -4 -4 -4
  \item LOW_ETHANOL -4 -4 -4
  \item MEDIUM_ETHANOL -4 -4 -4
  \item HIGH_ETHANOL -4 -4 -4
  \item ISOPROPYL_ALCOHOL -4 -4 -4
  \item N_METHYL_2_PYRROLIDONE -4 -4 -4
  \item POLYSACCHARIDES -4 -4 -4
  \item AMINOCAPROIC_ACID -4 -4 -4
  \item LOW_GLUCOSE -4 -4 -4
  \item HIGH_LACTATE -4 -4 -4
  \item HIGH_PYRUVATE -4 -4 -4
  \item LOW_GLUTAMINE_OR_HIGH_GLUTAMATE -4 -4 -4
  \item GLUCONOLACTONE -4 -4 -4
  \item LOW_PROTEIN -4 -4 -4
  \item UNEXPECTED_AMINO_ACID_SIGNALS -4 -4 -4
  \item UNIDENTIFIED_MACROMOLECULES -4 -4 -4
  \item UNIDENTIFIED_SMALL_MOLECULE_A -4 -4 -4
  \item UNIDENTIFIED_SMALL_MOLECULE_B -4 -4 -4
  \item UNIDENTIFIED_SMALL_MOLECULE_C -4 -4 -4
  \item BELOW_LIMIT_OF_QUANTIFICATION -4 -4 -4
  \item TOTAL_C -4 -4 -4
  \item NON_HDL_C -4 -4 -4
  \item REMNANT_C -4 mmol/l Remnant cholesterol (non-HDL, non-LDL -cholesterol)
  \item VLDL_C -4 mmol/l Total cholesterol in VLDL
  \item CLINICAL_LDL_C -4 -4 -4
  \item LDL_C -4 mmol/l Total cholesterol in LDL
  \item HDL_C -4 mmol/l Total cholesterol in HDL
  \item TOTAL_TG -4 -4 -4
  \item VLDL_TG -4 mmol/l Triglycerides in VLDL
  \item LDL_TG -4 mmol/l Triglycerides in LDL
  \item HDL_TG -4 mmol/l Triglycerides in HDL
  \item TOTAL_PL -4 -4 -4
  \item VLDL_PL -4 -4 -4
  \item LDL_PL -4 -4 -4
  \item HDL_PL -4 -4 -4
  \item TOTAL_CE -4 -4 -4
  \item VLDL_CE -4 -4 -4
  \item LDL_CE -4 -4 -4
  \item HDL_CE -4 -4 -4
  \item TOTAL_FC -4 -4 -4
  \item VLDL_FC -4 -4 -4
  \item LDL_FC -4 -4 -4
  \item HDL_FC -4 -4 -4
  \item TOTAL_L -4 -4 -4
  \item VLDL_L -4 -4 -4
  \item LDL_L -4 -4 -4
  \item HDL_L -4 -4 -4
  \item TOTAL_P -4 -4 -4
  \item VLDL_P -4 -4 -4
  \item LDL_P -4 -4 -4
  \item HDL_P -4 -4 -4
  \item VLDL_SIZE -4 -4 -4
  \item LDL_SIZE -4 -4 -4
  \item HDL_SIZE -4 -4 -4
  \item PHOSPHOGLYC -4 -4 -4
  \item TG_BY_PG -4 -4 -4
  \item CHOLINES -4 -4 -4
  \item PHOSPHATIDYLC -4 -4 -4
  \item SPHINGOMYELINS -4 -4 -4
  \item APOB -4 g/l Apolipoprotein B
  \item APOA1 -4 g/l Apolipoprotein A-I
  \item APOB_BY_APOA1 -4 -4 -4
  \item TOTAL_FA -4 -4 -4
  \item UNSATURATION -4 -4 -4
  \item OMEGA_3 -4 -4 -4
  \item OMEGA_6 -4 -4 -4
  \item PUFA -4 mmol/l Polyunsaturated fatty acids
  \item MUFA -4 mmol/l Monounsaturated fatty acids; 16:1, 18:1
  \item SFA -4 mmol/l Saturated fatty acids
  \item LA -4 mmol/l 18:2, linoleic acid
  \item DHA -4 mmol/l 22:6, docosahexaenoic acid
  \item OMEGA_3_PCT -4 -4 -4
  \item OMEGA_6_PCT -4 -4 -4
  \item PUFA_PCT -4 -4 -4
  \item MUFA_PCT -4 -4 -4
  \item SFA_PCT -4 -4 -4
  \item LA_PCT -4 -4 -4
  \item DHA_PCT -4 -4 -4
  \item PUFA_BY_MUFA -4 -4 -4
  \item OMEGA_6_BY_OMEGA_3 -4 -4 -4
  \item ALA -4 mmol/l Alanine
  \item GLN -4 mmol/l Glutamine
  \item GLY -4 mmol/l Glycine
  \item HIS -4 mmol/l Histidine
  \item TOTAL_BCAA -4 -4 -4
  \item ILE -4 mmol/l Isoleucine
  \item LEU -4 mmol/l Leucine
  \item VAL -4 mmol/l Valine
  \item PHE -4 mmol/l Phenylalanine
  \item TYR -4 mmol/l Tyrosine
  \item GLUCOSE -4 -4 -4
  \item LACTATE -4 -4 -4
  \item PYRUVATE -4 -4 -4
  \item CITRATE -4 -4 -4
  \item GLYCEROL -4 -4 -4
  \item BOHBUTYRATE -4 -4 -4
  \item ACETATE -4 -4 -4
  \item ACETOACETATE -4 -4 -4
  \item ACETONE -4 -4 -4
  \item CREATININE -4 -4 -4
  \item ALBUMIN -4 -4 -4
  \item GLYCA -4 -4 -4
  \item XXL_VLDL_P -4 mol/l Concentration of chylomicrons and extremely large VLDL particles
  \item XXL_VLDL_L -4 mmol/l Total lipids in chylomicrons and extremely large VLDL 
  \item XXL_VLDL_PL -4 mmol/l Phospholipids in chylomicrons and extremely large VLDL 
  \item XXL_VLDL_C -4 mmol/l Total cholesterol in chylomicrons and extremely large VLDL 
  \item XXL_VLDL_CE -4 mmol/l Cholesterol esters in chylomicrons and extremely large VLDL 
  \item XXL_VLDL_FC -4 mmol/l Free cholesterol in chylomicrons and extremely large VLDL 
  \item XXL_VLDL_TG -4 mmol/l Triglycerides in chylomicrons and extremely large VLDL 
  \item XL_VLDL_P -4 mol/l Concentration of very large VLDL particles
  \item XL_VLDL_L -4 mmol/l Total lipids in very large VLDL 
  \item XL_VLDL_PL -4 mmol/l Phospholipids in very large VLDL 
  \item XL_VLDL_C -4 mmol/l Total cholesterol in very large VLDL 
  \item XL_VLDL_CE -4 mmol/l Cholesterol esters in very large VLDL 
  \item XL_VLDL_FC -4 mmol/l Free cholesterol in very large VLDL 
  \item XL_VLDL_TG -4 mmol/l Triglycerides in very large VLDL 
  \item L_VLDL_P -4 mol/l Concentration of large VLDL particles
  \item L_VLDL_L -4 mmol/l Total lipids in large VLDL 
  \item L_VLDL_PL -4 mmol/l Phospholipids in large VLDL 
  \item L_VLDL_C -4 mmol/l Total cholesterol in large VLDL 
  \item L_VLDL_CE -4 mmol/l Cholesterol esters in large VLDL 
  \item L_VLDL_FC -4 mmol/l Free cholesterol in large VLDL 
  \item L_VLDL_TG -4 mmol/l Triglycerides in large VLDL 
  \item M_VLDL_P -4 mol/l Concentration of medium VLDL particles
  \item M_VLDL_L -4 mmol/l Total lipids in medium VLDL 
  \item M_VLDL_PL -4 mmol/l Phospholipids in medium VLDL 
  \item M_VLDL_C -4 mmol/l Total cholesterol in medium VLDL 
  \item M_VLDL_CE -4 mmol/l Cholesterol esters in medium VLDL 
  \item M_VLDL_FC -4 mmol/l Free cholesterol in medium VLDL 
  \item M_VLDL_TG -4 mmol/l Triglycerides in medium VLDL 
  \item S_VLDL_P -4 mol/l Concentration of small VLDL particles
  \item S_VLDL_L -4 mmol/l Total lipids in small VLDL 
  \item S_VLDL_PL -4 mmol/l Phospholipids in small VLDL 
  \item S_VLDL_C -4 mmol/l Total cholesterol in small VLDL 
  \item S_VLDL_CE -4 mmol/l Cholesterol esters in small VLDL 
  \item S_VLDL_FC -4 mmol/l Free cholesterol in small VLDL 
  \item S_VLDL_TG -4 mmol/l Triglycerides in small VLDL 
  \item XS_VLDL_P -4 mol/l Concentration of very small VLDL particles
  \item XS_VLDL_L -4 mmol/l Total lipids in very small VLDL 
  \item XS_VLDL_PL -4 mmol/l Phospholipids in very small VLDL 
  \item XS_VLDL_C -4 mmol/l Total cholesterol in very small VLDL 
  \item XS_VLDL_CE -4 mmol/l Cholesterol esters in very small VLDL 
  \item XS_VLDL_FC -4 mmol/l Free cholesterol in very small VLDL 
  \item XS_VLDL_TG -4 mmol/l Triglycerides in very small VLDL 
  \item IDL_P -4 mol/l Concentration of IDL particles
  \item IDL_L -4 mmol/l Total lipids in IDL 
  \item IDL_PL -4 mmol/l Phospholipids in IDL 
  \item IDL_C -4 mmol/l Total cholesterol in IDL 
  \item IDL_CE -4 mmol/l Cholesterol esters in IDL 
  \item IDL_FC -4 mmol/l Free cholesterol in IDL 
  \item IDL_TG -4 mmol/l Triglycerides in IDL 
  \item L_LDL_P -4 mol/l Concentration of large LDL particles
  \item L_LDL_L -4 mmol/l Total lipids in large LDL 
  \item L_LDL_PL -4 mmol/l Phospholipids in large LDL 
  \item L_LDL_C -4 mmol/l Total cholesterol in large LDL 
  \item L_LDL_CE -4 mmol/l Cholesterol esters in large LDL 
  \item L_LDL_FC -4 mmol/l Free cholesterol in large LDL 
  \item L_LDL_TG -4 mmol/l Triglycerides in large LDL 
  \item M_LDL_P -4 mol/l Concentration of medium LDL particles
  \item M_LDL_L -4 mmol/l Total lipids in medium LDL 
  \item M_LDL_PL -4 mmol/l Phospholipids in medium LDL 
  \item M_LDL_C -4 mmol/l Total cholesterol in medium LDL 
  \item M_LDL_CE -4 mmol/l Cholesterol esters in medium LDL 
  \item M_LDL_FC -4 mmol/l Free cholesterol in medium LDL 
  \item M_LDL_TG -4 mmol/l Triglycerides in medium LDL 
  \item S_LDL_P -4 mol/l Concentration of small LDL particles
  \item S_LDL_L -4 mmol/l Total lipids in small LDL 
  \item S_LDL_PL -4 mmol/l Phospholipids in small LDL 
  \item S_LDL_C -4 mmol/l Total cholesterol in small LDL 
  \item S_LDL_CE -4 mmol/l Cholesterol esters in small LDL 
  \item S_LDL_FC -4 mmol/l Free cholesterol in small LDL 
  \item S_LDL_TG -4 mmol/l Triglycerides in small LDL 
  \item XL_HDL_P -4 mol/l Concentration of very large HDL particles
  \item XL_HDL_L -4 mmol/l Total lipids in very large HDL 
  \item XL_HDL_PL -4 mmol/l Phospholipids in very large HDL 
  \item XL_HDL_C -4 mmol/l Total cholesterol in very large HDL 
  \item XL_HDL_CE -4 mmol/l Cholesterol esters in very large HDL 
  \item XL_HDL_FC -4 mmol/l Free cholesterol in very large HDL 
  \item XL_HDL_TG -4 mmol/l Triglycerides in very large HDL 
  \item L_HDL_P -4 mol/l Concentration of large HDL particles
  \item L_HDL_L -4 mmol/l Total lipids in large HDL 
  \item L_HDL_PL -4 mmol/l Phospholipids in large HDL 
  \item L_HDL_C -4 mmol/l Total cholesterol in large HDL 
  \item L_HDL_CE -4 mmol/l Cholesterol esters in large HDL 
  \item L_HDL_FC -4 mmol/l Free cholesterol in large HDL 
  \item L_HDL_TG -4 mmol/l Triglycerides in large HDL 
  \item M_HDL_P -4 mol/l Concentration of medium HDL particles
  \item M_HDL_L -4 mmol/l Total lipids in medium HDL 
  \item M_HDL_PL -4 mmol/l Phospholipids in medium HDL 
  \item M_HDL_C -4 mmol/l Total cholesterol in medium HDL 
  \item M_HDL_CE -4 mmol/l Cholesterol esters in medium HDL 
  \item M_HDL_FC -4 mmol/l Free cholesterol in medium HDL 
  \item M_HDL_TG -4 mmol/l Triglycerides in medium HDL 
  \item S_HDL_P -4 mol/l Concentration of small HDL particles
  \item S_HDL_L -4 mmol/l Total lipids in small HDL 
  \item S_HDL_PL -4 mmol/l Phospholipids in small HDL 
  \item S_HDL_C -4 mmol/l Total cholesterol in small HDL 
  \item S_HDL_CE -4 mmol/l Cholesterol esters in small HDL 
  \item S_HDL_FC -4 mmol/l Free cholesterol in small HDL 
  \item S_HDL_TG -4 mmol/l Triglycerides in small HDL 
  \item XXL_VLDL_PL_PCT -4 -4 -4
  \item XXL_VLDL_C_PCT -4 -4 -4
  \item XXL_VLDL_CE_PCT -4 -4 -4
  \item XXL_VLDL_FC_PCT -4 -4 -4
  \item XXL_VLDL_TG_PCT -4 -4 -4
  \item XL_VLDL_PL_PCT -4 -4 -4
  \item XL_VLDL_C_PCT -4 -4 -4
  \item XL_VLDL_CE_PCT -4 -4 -4
  \item XL_VLDL_FC_PCT -4 -4 -4
  \item XL_VLDL_TG_PCT -4 -4 -4
  \item L_VLDL_PL_PCT -4 -4 -4
  \item L_VLDL_C_PCT -4 -4 -4
  \item L_VLDL_CE_PCT -4 -4 -4
  \item L_VLDL_FC_PCT -4 -4 -4
  \item L_VLDL_TG_PCT -4 -4 -4
  \item M_VLDL_PL_PCT -4 -4 -4
  \item M_VLDL_C_PCT -4 -4 -4
  \item M_VLDL_CE_PCT -4 -4 -4
  \item M_VLDL_FC_PCT -4 -4 -4
  \item M_VLDL_TG_PCT -4 -4 -4
  \item S_VLDL_PL_PCT -4 -4 -4
  \item S_VLDL_C_PCT -4 -4 -4
  \item S_VLDL_CE_PCT -4 -4 -4
  \item S_VLDL_FC_PCT -4 -4 -4
  \item S_VLDL_TG_PCT -4 -4 -4
  \item XS_VLDL_PL_PCT -4 -4 -4
  \item XS_VLDL_C_PCT -4 -4 -4
  \item XS_VLDL_CE_PCT -4 -4 -4
  \item XS_VLDL_FC_PCT -4 -4 -4
  \item XS_VLDL_TG_PCT -4 -4 -4
  \item IDL_PL_PCT -4 -4 -4
  \item IDL_C_PCT -4 -4 -4
  \item IDL_CE_PCT -4 -4 -4
  \item IDL_FC_PCT -4 -4 -4
  \item IDL_TG_PCT -4 -4 -4
  \item L_LDL_PL_PCT -4 -4 -4
  \item L_LDL_C_PCT -4 -4 -4
  \item L_LDL_CE_PCT -4 -4 -4
  \item L_LDL_FC_PCT -4 -4 -4
  \item L_LDL_TG_PCT -4 -4 -4
  \item M_LDL_PL_PCT -4 -4 -4
  \item M_LDL_C_PCT -4 -4 -4
  \item M_LDL_CE_PCT -4 -4 -4
  \item M_LDL_FC_PCT -4 -4 -4
  \item M_LDL_TG_PCT -4 -4 -4
  \item S_LDL_PL_PCT -4 -4 -4
  \item S_LDL_C_PCT -4 -4 -4
  \item S_LDL_CE_PCT -4 -4 -4
  \item S_LDL_FC_PCT -4 -4 -4
  \item S_LDL_TG_PCT -4 -4 -4
  \item XL_HDL_PL_PCT -4 -4 -4
  \item XL_HDL_C_PCT -4 -4 -4
  \item XL_HDL_CE_PCT -4 -4 -4
  \item XL_HDL_FC_PCT -4 -4 -4
  \item XL_HDL_TG_PCT -4 -4 -4
  \item L_HDL_PL_PCT -4 -4 -4
  \item L_HDL_C_PCT -4 -4 -4
  \item L_HDL_CE_PCT -4 -4 -4
  \item L_HDL_FC_PCT -4 -4 -4
  \item L_HDL_TG_PCT -4 -4 -4
  \item M_HDL_PL_PCT -4 -4 -4
  \item M_HDL_C_PCT -4 -4 -4
  \item M_HDL_CE_PCT -4 -4 -4
  \item M_HDL_FC_PCT -4 -4 -4
  \item M_HDL_TG_PCT -4 -4 -4
  \item S_HDL_PL_PCT -4 -4 -4
  \item S_HDL_C_PCT -4 -4 -4
  \item S_HDL_CE_PCT -4 -4 -4
  \item S_HDL_FC_PCT -4 -4 -4
  \item S_HDL_TG_PCT -4 -4 -4
}
}
\examples{
\dontrun{
describe(adninightingalelong)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/ADNINIGHTINGALELONG/ADNINIGHTINGALELONG_METHODS.pdf}
}
\keyword{datasets}
