% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{adrc_emory_dde}
\alias{adrc_emory_dde}
\title{DDE analysis summary}
\format{
A data frame with 218 rows and 10 variables
}
\usage{
data(adrc_emory_dde)
}
\description{
DDE analysis summary
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VERSION -4  Version number
  \item RUNDATE -4  Analysis date
  \item STATUS -4  Analysis status
  \item ANALYSIS_BATCH -4  Batch number
  \item DDE_RECOVER_13C_PERCENT -4 percent Recovery percent
  \item DDE_PLASMA_CONCENTRATION -4 ng/L Plasma concentration
  \item DETECTION_LIMIT -4  Above detection limit
}
}
\examples{
\dontrun{
describe(adrc_emory_dde)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/ADRC_EMORY_DDE/Levey_Walker_ADNI_Methods_Submission_v2.pdf}
}
\keyword{datasets}
