% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{adsxlist}
\alias{adsxlist}
\title{Diagnosis and Symptoms Checklist}
\format{
A data frame with 4884 rows and 36 variables
}
\usage{
data(adsxlist)
}
\description{
See \url{http://adni.loni.usc.edu}.
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item AXNAUSEA N  1. Nausea
  \item AXVOMIT N  2. Vomiting
  \item AXDIARRH N  3. Diarrhea
  \item AXCONSTP N  4. Constipation
  \item AXABDOMN N  5. Abdominal discomfort
  \item AXSWEATN N  6. Sweating
  \item AXDIZZY N  7. Dizziness
  \item AXENERGY N  8. Low energy
  \item AXDROWSY N  9. Drowsiness
  \item AXVISION N  10. Blurred vision
  \item AXHDACHE N  11. Headache
  \item AXDRYMTH N  12. Dry mouth
  \item AXBREATH N  13. Shortness of breath
  \item AXCOUGH N  14. Coughing
  \item AXPALPIT N  15. Palpitations
  \item AXCHEST N  16. Chest pain
  \item AXURNDIS N  17. Urinary discomfort (e.g., burning)
  \item AXURNFRQ N  18. Urinary frequency
  \item AXANKLE N  19. Ankle swelling
  \item AXMUSCLE N  20. Muscloskeletal pain
  \item AXRASH N  21. Rash
  \item AXINSOMN N  22. Insomnia
  \item AXDPMOOD N  23. Depressed mood
  \item AXCRYING N  24. Crying
  \item AXELMOOD N  25. Elevated mood
  \item AXWANDER N  26. Wandering
  \item AXFALL N  27. Fall
  \item AXOTHER N  28. Other/Diagnosis
  \item USERDATE2 S  Date record last updated
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
