% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{adverse}
\alias{adverse}
\title{Adverse Events Log}
\format{
A data frame with 3063 rows and 79 variables
}
\usage{
data(adverse)
}
\description{
Adverse Events Log
}
\details{
A dataset with variables as follows:

\itemize{
  \item AENUMBER N NA Adverse event number
  \item AECHANGE N NA Is this event a:
  \item AEPENUM N NA Initial Health Assessment Condition Number:
  \item AECOMPOUND N NA Is this event part of a compound event?
  \item AEHONSDT D YYYY-MM-DD Event onset date
  \item AEOUTCOME N NA Event outcome
  \item AEHCSDT D YYYY-MM-DD Event cease date
  \item AEHDTHDT D YYYY-MM-DD Date of death
  \item AEHDBRN N NA Was a brain autopsy performed?
  \item AEHCHRON N NA Initial Chronicity
  \item AEHSEVR N NA Initial Severity
  \item AERELAD N NA Related to symptoms associated with the development of early stage AD or progression of AD
  \item AERELCM N NA Related to concomitant therapy
  \item AERELFLRBTBN N NA Related to Florbetaben tracer
  \item AERELFLRBPR N NA Related to Florbetapir tracer
  \item AEHIMG N NA Related to PET/MRI imaging procedure
  \item AERELTAU N NA Related to AV-1451 (Tau) tracer
  \item AEHLUMB N NA Related to lumbar puncture
  \item AERELCOVID N NA Related to COVID-19 illness
  \item AERELPAN N NA Related to COVID-19 pandemic disruption
  \item AERELATESP N NA Related to other study procedure(s)
  \item AEHCMEDS N NA Was treatment (medication) for AE required?
  \item AESERIOUS N NA Was event serious at any time during the trial?
  \item AESERDATE D YYYY-MM-DD Date event became serious:
  \item AEDTSITE D YYYY-MM-DD Date site became aware that the event is serious:
  \item SAELIFE N NA Is this an SAE because event was life-threatening?
  \item SAEHOSPIT N NA Is this an SAE because it lead to  hospitalization?
  \item SAEPROLONG N NA Is this an SAE because it lead to  prolongation of hospitalization?
  \item SAEDEATH N NA Is this an SAE because it lead to  death?
  \item SAECONGEN N NA Is this an SAE because it lead to a congenital anomaly or birth defect?
  \item SAEDISAB N NA Is this an SAE because it caused  persistent or significant disability/incapacity?
  \item SAEOTHER N NA Is this an SAE because it lead to another medically important condition?
  \item AECHANGE1 N NA Is there a change to report?
  \item AEDATE1 D YYYY-MM-DD Change 1.  Date
  \item AESEV1 N NA Change 1: Severity
  \item AECHRON1 N NA Change 1: Chronicity
  \item AECHANGE2 N NA Is there a 2nd change to report?
  \item AEDATE2 D YYYY-MM-DD Change 2.  Date
  \item AESEV2 N NA Change 2: Severity
  \item AECHRON2 N NA Change 2: Chronicity
  \item AECHANGE3 N NA Is there a 3rd change to report?
  \item AEDATE3 D YYYY-MM-DD Change 3.  Date
  \item AESEV3 N NA Change 3: Severity
  \item AECHRON3 N NA Change 3: Chronicity
  \item AECHANGE4 N NA Is there a 4th change to report?
  \item AEDATE4 D YYYY-MM-DD Change 4.  Date
  \item AESEV4 N NA Change 4: Severity
  \item AECHRON4 N NA Change 4: Chronicity
  \item AECHANGE5 N NA Is there a 5th change to report?
  \item AEDATE5 D YYYY-MM-DD Change 5.  Date
  \item AESEV5 N NA Change 5: Severity
  \item AECHRON5 N NA Change 5: Chronicity
  \item AECHANGE6 N NA Is there a 6th change to report?
  \item AEDATE6 D YYYY-MM-DD Change 6.  Date
  \item AESEV6 N NA Change 6: Severity
  \item AECHRON6 N NA Change 6: Chronicity
  \item AECHANGE7 N NA Is there a 7th change to report?
  \item AEDATE7 D YYYY-MM-DD Change 7.  Date
  \item AESEV7 N NA Change 7: Severity
  \item AECHRON7 N NA Change 7: Chronicity
  \item AECHANGE8 N NA Is there a 8th change to report?
  \item AEDATE8 D YYYY-MM-DD Change 8.  Date
  \item AESEV8 N NA Change 8: Severity
  \item AECHRON8 N NA Change 8: Chronicity
  \item AECHANGE9 N NA Is there a 9th change to report?
  \item AEDATE9 D YYYY-MM-DD Change 9.  Date
  \item AESEV9 N NA Change 9: Severity
  \item AECHRON9 N NA Change 9: Chronicity
  \item AECHANGE10 N NA Is there a 10th change to report?
  \item AEDATE10 D YYYY-MM-DD Change 10.  Date
  \item AESEV10 N NA Change 10: Severity
  \item AECHRON10 N NA Change 10: Chronicity
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
}
}
\examples{
\dontrun{
describe(adverse)
}
}
\keyword{datasets}
