% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{aecheck}
\alias{aecheck}
\title{Adverse Event Checklist}
\format{
A data frame with 5750 rows and 36 variables
}
\usage{
data(aecheck)
}
\description{
See \url{http://adni.loni.usc.edu}.
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
  \item AENAUSEA N  1. Nausea
  \item AEVOMIT N  2. Vomiting
  \item AEDIARRH N  3. Diarrhea
  \item AECONSTP N  4. Constipation
  \item AEABDOMN N  5. Abdominal discomfort
  \item AESWEATN N  6. Sweating
  \item AEDIZZY N  7. Dizziness
  \item AEENERGY N  8. Low energy
  \item AEDROWSY N  9. Drowsiness
  \item AEVISION N  10. Blurred vision
  \item AEHDACHE N  11. Headache
  \item AEDRYMTH N  12. Dry mouth
  \item AEBREATH N  13. Shortness of breath
  \item AECOUGH N  14. Coughing
  \item AEPALPIT N  15. Palpitations
  \item AECHEST N  16. Chest pain
  \item AEURNDIS N  17. Urinary discomfort (e.g., burning)
  \item AEURNFRQ N  18. Urinary frequency
  \item AEANKLE N  19. Ankle swelling
  \item AEMUSCLE N  20. Musculoskeletal pain
  \item AERASH N  21. Rash
  \item AEINSOMN N  22. Insomnia
  \item AEDPMOOD N  23. Depressed mood
  \item AECRYING N  24. Crying
  \item AEELMOOD N  25. Elevated mood
  \item AEWANDER N  26. Wandering
  \item AEFALL N  27. Fall
  \item AEOTHER N  28. Other Symptoms
  \item AESPECIF T  If Other, specify:
}
}
\keyword{datasets}
