% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{agrocampus}
\alias{agrocampus}
\title{Institut Agro-ecole interne Agrocampus Ouest lipids in red blood cells}
\format{
A data frame with 849 rows and 38 variables
}
\usage{
data(agrocampus)
}
\description{
Institut Agro-ecole interne Agrocampus Ouest lipids in red blood cells
}
\details{
A dataset with variables as follows:

\itemize{
  \item ORIGPROT -4 -4 Original study protocol
  \item COLPROT -4 -4 Study protocol of data collection
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item SPECIMEN.BAR.CODE -4 -4 Specimen Bar Code
  \item SEQUENCE.NUMBER -4 -4 Sequence Number
  \item SPECIMEN.TYPE -4 -4 Specimen Type
  \item SPECIMEN.QUANTITY -4 -4 Specimen Quantity
  \item UOM -4 -4 UOM
  \item BOX.NAME -4 -4 Box Name
  \item POSITION -4 -4 Position
  \item KIT.NUMBER -4 -4 Kit Number
  \item LAB.NUMBER -4 -4 Lab number
  \item MASSE.MG -4 -4 masse (mg)
  \item C140 -4 -4 C14:0 (% of total fatty acids)
  \item C160 -4 -4 C16:0 (% of total fatty acids)
  \item C161N7 -4 -4 C16:1 n-7 (% of total fatty acids)
  \item C180 -4 -4 C18:0 (% of total fatty acids)
  \item C181N9 -4 -4 C18:1 n-9 (% of total fatty acids)
  \item C181N7 -4 -4 C18:1 n-7 (% of total fatty acids)
  \item C182N6 -4 -4 C18:2 n-6 (% of total fatty acids)
  \item C183N3PLUSC200 -4 -4 C18:3 n-3 + C20:0 (% of total fatty acids)
  \item C201N9 -4 -4 C20:1 n-9 (% of total fatty acids)
  \item C203N6 -4 -4 C20:3 n-6 (% of total fatty acids)
  \item C204N6 -4 -4 C20:4 n-6 (% of total fatty acids)
  \item C205N3 -4 -4 C20:5 n-3 (% of total fatty acids)
  \item C224N6 -4 -4 C22:4 n-6 (% of total fatty acids)
  \item C225N6 -4 -4 C22:5 n-6 (% of total fatty acids)
  \item C225N3 -4 -4 C22:5 n-3 (% of total fatty acids)
  \item C226N3 -4 -4 C22:6 n-3 (% of total fatty acids)
  \item TOTAL -4 -4 Total (g/g)
  \item SATURES -4 -4 Satures (% of total fatty acids)
  \item MONO -4 -4 Mono (% of total fatty acids)
  \item N6 -4 -4 n-6 (% of total fatty acids)
  \item N3 -4 -4 n-3 (% of total fatty acids)
  \item C205PLUSC226N3 -4 -4 C20:5 + C22:6 n-3 (% of total fatty acids)
  \item C205PLUSC225PLUSC226N3 -4 -4 C20:5 + C22:5 +C22:6 n-3 (% of total fatty acids)
}
}
\examples{
\dontrun{
describe(agrocampus)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/AGROCAMPUS/ADNI_Methods_Agrocampus_june7.pdf}
}
\keyword{datasets}
