% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{aliquot_list}
\alias{aliquot_list}
\title{Aliquot count in the LDMS database}
\format{
A data frame with 10431 rows and 22 variables
}
\usage{
data(aliquot_list)
}
\description{
Aliquot count in the LDMS database
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4  Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 MM/DD/YYYY Examination Date
  \item VID -4  Visit number from sample tube
  \item PLA -4  Count of pristine Plasma aliquots in storage
  \item PLAVOL -4 mL Total volume of received Plasma
  \item PLA_SENT -4  Count of Plasma aliquots shipped out to other investigators
  \item PLA_USED -4  Count of Plasma aliquots used for analysis at Biomarker Core Lab
  \item SER -4  Count of pristine Serum aliquots in storage
  \item SERVOL -4 mL Total volume of received Serum
  \item SER_SENT -4  Count of Serum aliquots shipped out to other investigators
  \item SER_USED -4  Count of Serum aliquots used for analysis at Biomarker Core Lab
  \item CSF -4  Count of pristine CSF aliquots in storage
  \item CSFVOL -4 mL Total volume of received CSF
  \item CSF_SENT -4  Count of CSF aliquots shipped out to other investigators
  \item CSF_USED -4  Count of CSF aliquots used for analysis at Biomarker Core Lab
  \item URN -4  Count of pristine Urine aliquots in storage
  \item URNVOL -4 mL Total volume of received Urine
  \item URN_SENT -4  Count of Urine aliquots shipped out to other investigators
  \item URN_USED -4  Count of Urine aliquots used for analysis at Biomarker Core Lab
  \item QUERYDATE -4 MM/DD/YYYY Date of LDMS database query
  \item USED_PLA -4  Count of Plasma aliquots used for analysis at Biomarker Core Lab
  \item USED_SER -4  Count of Serum aliquots used for analysis at Biomarker Core Lab
  \item USED_CSF -4  Count of CSF aliquots used for analysis at Biomarker Core Lab
  \item USED_URN -4  Count of Urine aliquots used for analysis at Biomarker Core Lab
  \item SENT_PLA -4  Count of Plasma aliquots shipped out to other investigators
  \item SENT_SER -4  Count of Serum aliquots shipped out to other investigators
  \item SENT_CSF -4  Count of CSF aliquots shipped out to other investigators
  \item SENT_URN -4  Count of Urine aliquots shipped out to other investigators
}
}
\examples{
\dontrun{
describe(aliquot_list)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/ALIQUOT_LIST/CSF\%20Aliquot\%20Inventory\%20Description\%203feb2015.pdf}, \url{https://adni.bitbucket.io/reference/docs/ALIQUOT_LIST/PLA\%20Aliquot\%20Inventory\%20Description\%203feb2015.pdf}
}
\keyword{datasets}
