% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{amymeta}
\alias{amymeta}
\title{Amyloid PET Scan Information}
\format{
A data frame with 1918 rows and 31 variables
}
\usage{
data(amymeta)
}
\description{
Amyloid PET Scan Information
}
\details{
A dataset with variables as follows:

\itemize{
  \item DONE N NA Was the scan conducted?
  \item NDREASON N NA If No, reason not done
  \item RADTRACER N NA If scan not done, was radiotracer administered?
  \item TRACERTYPE N NA Amyloid Imaging radiotracer used
  \item SCANDATE D YYYY-MM-DD Scan Date/Dose Date
  \item QCTIME T HHMM Time of today's Scanner QC
  \item ASSAYTIME T HHMM Time of dose assay
  \item DOSEFLRBTBN N mCi Florbetaben (Neuraceq):  Injected dose assay
  \item DOSEFLRBTPR N mCi Florbetapir (Amyvid):  Injected dose assay
  \item INJTIME T HHMM Time of injection
  \item EARLYADD N NA Does the study include an 'early add-on' scan (0-20) min?
  \item EARLYTIME T NA If yes, Early add on sequence start time: [H H M M]
  \item EARLYNO N NA If no, why?
  \item EARLYOTHER T NA If other, specify
  \item SCANTIME T HHMM Emission Scan Start Time
  \item VARIAT N NA Any variations from protocol during uptake?
  \item PROTID T NA Predefined Acquisition Protocol Name
  \item MOTION N NA Participant motion problems
  \item OTHERVAR N NA Other protocol variations
  \item RECON N NA Select reconstruction used
  \item SHARP N NA For Phillips scanners:  was "smooth" parameter set to "sharp"?
  \item ARCHIVE N NA Data archived locally?
  \item LONI N NA Data transferred to LONI?
  \item TRANDATE D YYYY-MM-DD Transfer Date
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
}
}
\examples{
\dontrun{
describe(amymeta)
}
}
\keyword{datasets}
