% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{amyqc}
\alias{amyqc}
\title{Amyloid PET QC}
\format{
A data frame with 1827 rows and 20 variables
}
\usage{
data(amyqc)
}
\description{
Amyloid PET QC
}
\details{
A dataset with variables as follows:

\itemize{
  \item SCANDATE D YYYY-MM-DD Scan Date
  \item LONIUPDT D YYYY-MM-DD Date uploaded to LONI
  \item LONIUID T NA LONI Unique Series ID
  \item REVDT D YYYY-MM-DD Date Reviewed
  \item ALLFRAME N NA Were all frames acceptable?
  \item UNUSABL T NA If No, indicate which frames were unacceptable
  \item UNRSN N NA If No, indicate why frames were unacceptable
  \item UNRSNSPEC T NA If Other, Specify
  \item SCANQLTY N NA Scan Pass QC?
  \item REPROCREQ N NA Reprocessing requested?
  \item PROCERR T NA If Yes, select processing error(s)
  \item PROCERRSPEC T NA If Other, Specify
  \item RESCANREQ N NA Rescan requested?
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
}
}
\examples{
\dontrun{
describe(amyqc)
}
}
\keyword{datasets}
