% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{apoe}
\alias{apoe}
\title{ApoE Genotyping - Draw Data}
\format{
A data frame with 1513 rows and 21 variables
}
\usage{
data(apoe)
}
\description{
See \url{http://adni.loni.usc.edu}.
}
\details{
A dataset with variables as follows:

\itemize{
  \item APCOLLECT N NA Was DNA sample collected?
  \item RNACOLL N NA Was RNA sample collected?
  \item RNADATE D NA Date of RNA collection:
  \item RNATIME N NA Time of RNA collection:
  \item RNAVOL N NA Total Volume of Blood Drawn into 3 x 2.5 mL PAXgene RNA Tubes
  \item RNASHIP N NA Was the same shipment date and Fedex Tracking Number used to ship the RNA sample?
  \item RNAFEDEXDT D NA Date Fedexed
  \item RNAFEDEX N NA Fedex Tracking Number
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item APTIME T  Time of Blood Draw
  \item APFEDEXDT D  Date Fedexed
  \item APFEDEX N  Fedex Tracking Number
  \item APVOLUME N  Volume of Blood Drawn into Lavendar Top Tube
  \item APLICENSE N  6 digit License Plate Number
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
