% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{apoe3}
\alias{apoe3}
\title{ADNI3 APOE Genotypes}
\format{
A data frame with 637 rows and 4 variables
}
\usage{
data(apoe3)
}
\description{
ADNI3 APOE Genotypes
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item APGEN1 -4 -4 Genotype - Allele 1
  \item APGEN2 -4 -4 Genotype - Allele 2
}
}
\examples{
\dontrun{
describe(apoe3)
}
}
\keyword{datasets}
