% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{apoego2}
\alias{apoego2}
\title{ADNI GO & 2 APOE Genotypes}
\format{
A data frame with 908 rows and 14 variables
}
\usage{
data(apoego2)
}
\description{
ADNI GO & 2 APOE Genotypes
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N NA Participant roster ID
  \item SITEID N NA Site ID
  \item VISCODE T NA Visit code
  \item USERDATE S NA Date record created
  \item USERDATE2 S NA Date record last updated
  \item APTESTDT D NA Date Test Performed
  \item APGEN1 N NA Genotype - Allele 1
  \item APGEN2 N NA Genotype - Allele 2
  \item APVOLUME N NA Volume of Blood Shipped in Lavendar Top Tube
  \item APRECEIVE N NA Sample recieved within 24 hours of blood draw?
  \item APAMBTEMP N NA Sample shipped at ambient temperature?
  \item APRESAMP N NA Request Resample?
  \item APUSABLE N NA Sample Useable?
}
}
\examples{
\dontrun{
describe(apoego2)
}
}
\keyword{datasets}
