% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{apoeres}
\alias{apoeres}
\title{ApoE Genotyping - Results}
\format{
A data frame with 1159 rows and 15 variables
}
\usage{
data(apoeres)
}
\description{
See \url{http://adni.loni.usc.edu}.
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item APTESTDT D  Date Test Performed
  \item APGEN1 N  Genotype - Allele 1
  \item APGEN2 N  Genotype - Allele 2
  \item APVOLUME N  Volume of Blood Shipped in Lavendar Top Tube
  \item APRECEIVE N  Sample recieved within 24 hours of blood draw?
  \item APAMBTEMP N  Sample shipped at ambient temperature?
  \item APRESAMP N  Request Resample?
  \item APUSABLE N  Sample Useable?
  \item USERDATE2 S  Date record last updated
}
}
\examples{
apoeres$ApoE4 <- apply(apoeres[, c('APGEN1', 'APGEN2')]==4, 1, sum)>=1 
}
\keyword{datasets}
