% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{arm}
\alias{arm}
\title{Diagnosis and scan category assignment at screen}
\format{
A data frame with 4730 rows and 13 variables
}
\usage{
data(arm)
}
\description{
See \url{http://adni.loni.usc.edu}.
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item PTNO T  4 digit Participant Number
  \item TYPE T  ID Type
  \item ARM N  Diagnosis/Scan Category Assignment
  \item ENROLLED N  Enrollment Status
  \item USERDATE S  Date record created
  \item RANDDATE S  Randomization Date
  \item MAPPDATE S  Monitor Approval Date
  \item USERDATE2 S  Date record last updated
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
