% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{av45follow}
\alias{av45follow}
\title{AV-45 24-48 Hour Follow-Up}
\format{
A data frame with 1745 rows and 12 variables
}
\usage{
data(av45follow)
}
\description{
See \url{http://adni.loni.usc.edu}.
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N NA Participant roster ID
  \item SITEID N NA Site ID
  \item VISCODE T NA Visit code
  \item USERDATE S NA Date record created
  \item USERDATE2 S NA Date record last updated
  \item CONTACT N NA Was 24-48 hours post imaging follow-up telephone contact made?
  \item CALLDATE D NA Date of telephone contact:
  \item CALLTIME N NA Time of telephone contact:
  \item CONTACTED N NA Person who was contacted:
  \item CALLAE N NA Were any Adverse Events reported?
}
}
\keyword{datasets}
