% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{av45meta}
\alias{av45meta}
\title{AV-45 PET Scan Information}
\format{
A data frame with 3400 rows and 52 variables
}
\usage{
data(av45meta)
}
\description{
See \url{http://adni.loni.usc.edu}.
}
\details{
A dataset with variables as follows:

\itemize{
  \item RADTRACER N  If no, was radiotracer administered?
  \item PMAODONE N  Was scan started at time of injection?
  \item PMAOND N  If no, why?
  \item PMAONDOTH T  If other, specify:
  \item PMAOTIME T HH:MM:SS Early add on sequence start time:
  \item PMAOCOM T  Target start time for the early add on sequence is at the point of injection.  If more than 10 seconds after time of Florbetapir F18 injection, then please explain:
  \item AORECON N  Were all reconstruction parameters the same for the early add on sequence as the standard scan?
  \item AORECONDIF T  If no, please explain:
  \item RID N NA Participant roster ID
  \item SITEID N NA Site ID
  \item VISCODE T NA Visit code
  \item USERDATE S NA Date record created
  \item USERDATE2 S NA Date record last updated
  \item RECNO N NA 
  \item PMCONDCT N NA Was the scan conducted?
  \item PMREASON N NA Reason why the scan was not conducted:
  \item EXAMDATE D NA Examination Date
  \item PMSCANNER T NA GE
  \item PMGEMODEL N NA If GE, Scanner Model:
  \item PMSIEMENS T NA Siemens
  \item PMSIEMODEL N NA If Siemens, Scanner Model:
  \item PMPHILLIPS T NA Phillips
  \item PMPHMODEL N NA If Phillips, Scanner Model:
  \item PMQCTIME T NA Time of today's Scanner QC
  \item PMFDGTIME T NA Time of AV-45 dose assay
  \item PMFDGDOS N NA AV-45 dose assay
  \item PMFDGVOL N NA AV-45 Volume
  \item PMINJTIME T NA Time of AV-45 injection
  \item PMSCTIME T NA Emission Scan Start Time
  \item PMVARIAT N NA Any variations from protocol during AV-45 uptake?
  \item PMFRAME N NA Which framing rate was used?
  \item PMMOTION N NA Subject motion problems:
  \item PMMALFUN N NA Scanner malfunction
  \item PMOTHER N NA Other protocol variations:
  \item PMRECON N NA Check which of the following reconstructions was used:
  \item PMSUBSET N NA # subsets:
  \item PMSUBSPE N NA If Other, specify
  \item PMITERAT N NA # iterations:
  \item PMITERSPE N NA If Other, specify:
  \item PMRAMLA N NA If 3D-Ramla, Lambda=
  \item PMFILTER T NA If 3D Back-Projection, Ramp filter?
  \item PMMODEON N NA If FORE/2D-OSEM select one of the following:
  \item PMSMOOTH T NA No post-process smoothing:
  \item PMATTEN N NA Attenuation Correction:
  \item PMTRNSFR N NA Was data transferred to LONI within 24 hours of scan?
  \item PMTRNDATE D NA Transfer Date
  \item PMARCHIVE N NA Was all raw PET data archived locally to be able to do complete reconstruction of PET Scan if needed?
  \item PMLPDONE N NA Was a Lumbar Puncture completed prior to the AV-45 scan?
  \item PMLPINTER N NA If Yes, What was the interval between LP and AV-45?
  \item PMSHARP T NA Was "Smooth" parameter set to "Sharp"?
}
}
\keyword{datasets}
