% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{av45qc}
\alias{av45qc}
\title{AV-45 PET QC Tracking}
\format{
A data frame with 2707 rows and 34 variables
}
\usage{
data(av45qc)
}
\description{
See \url{http://adni.loni.usc.edu}.
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N NA Participant roster ID
  \item SITEID N NA Site ID
  \item VISCODE T NA Visit code
  \item USERDATE S NA Date record created
  \item USERDATE2 S NA Date record last updated
  \item RECNO N NA 
  \item EXAMDATE D NA Examination Date
  \item PQFILENO N NA Upload Number
  \item PQDATTRA D NA Date images uploaded to LONI
  \item LONIUID T NA LONI Unique Series ID
  \item PQDATE D NA Date reviewed
  \item PQACQFR N NA Acquisition
  \item PQCORALG N NA Correct Reconstruction Algorithm?
  \item PQRECON N NA Check which of the following reconstruction algorithms was used
  \item PQCORNUM N NA If FORE/2D-OSEM, OSEM3D, or 3D Iterative, correct # subsets and iterations?
  \item PQCORFIL N NA Correct Filters/TRIM?
  \item PQCORTHI N NA Correct Slice Thickness (PET/CT systems)?
  \item PQVOXEL N NA Acceptable Voxel Size?
  \item PQALLFRAME N NA Were all frames acceptable?
  \item PQUNUSABLE T NA Indicate which frames were unacceptable:
  \item PQREASON T NA Indicate why frames were unacceptable:
  \item PQREASSP T NA If Other, Specify:
  \item PASS N NA Pass QC?
  \item PQPROERR T NA A. Processing Error(s)
  \item PQERRSP T NA Specify
  \item PQACTION N NA B. QC outcome
  \item PQISSUES T NA Imaging issues
  \item PQISSOTH T NA Other, specify
  \item QUARANTINE N NA Release from quarantine?
  \item PAYSITE N NA Pay Site?
  \item SCAN N NA Scan
  \item PQNONESP T NA If unusable, reason for not rescanning
}
}
\keyword{datasets}
