% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{av45vitals}
\alias{av45vitals}
\title{AV-45 Pre and Post Injection Vitals}
\format{
A data frame with 1802 rows and 24 variables
}
\usage{
data(av45vitals)
}
\description{
See \url{http://adni.loni.usc.edu}.
}
\details{
A dataset with variables as follows:

\itemize{
  \item RECNO N  
  \item RID N NA Participant roster ID
  \item SITEID N NA Site ID
  \item VISCODE T NA Visit code
  \item USERDATE S NA Date record created
  \item USERDATE2 S NA Date record last updated
  \item SCANDONE N NA Was scan conducted?
  \item SCANDATE D NA AV-45 Scan Date:
  \item PREHEART N NA Heart Rate (bpm):
  \item PRERESP N NA Respiration (per minute):
  \item PRESYSTBP N NA Systolic Blood Pressure (mmHg):
  \item PREDIABP N NA Diastolic Blood Pressure (mmHg):
  \item PRETEMP N NA Temperature:
  \item POSTHEART N NA Heart Rate (bpm):
  \item POSTRESP N NA Respiration (per minute):
  \item POSTSYSTBP N NA Systolic Blood Pressure (mmHg):
  \item POSTDIABP N NA Diastolic Blood Pressure (mmHg):
  \item POSTTEMP N NA Temperature:
  \item PRETEMPU N NA Temperature Units:
  \item POSTTEMPU N NA Temperature Units:
  \item PRETEMPS N NA Temperature Source:
  \item POSTTEMPS N NA Temperature Source:
}
}
\keyword{datasets}
