% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{avgjacob}
\alias{avgjacob}
\title{Average Jacobian - Temporal (Paul Thompson's Lab)}
\format{
A data frame with 3760 rows and 11 variables
}
\usage{
data(avgjacob)
}
\description{
Average Jacobian - Temporal (Paul Thompson's Lab)
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N -4 Participant roster ID
  \item SITEID N -4 Site ID
  \item VISCODE T -4 Visit code
  \item USERDATE S -4 Date record created
  \item RECNO N -4 
  \item EXAMDATE D -4 Examination Date
  \item IMAGEUID N -4 imageUID
  \item MEAN.ATROPHY.RATE.TEMPORAL.ROI N -4 
  \item PERCENT.CHANGE.TEMPORAL.ROI N -4 
  \item PERCENT.CHANGE.STAT.ROI N -4 
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/AVGJACOB/LONI_TBM_Protocol.pdf}
}
\keyword{datasets}
