% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{baceappbeta}
\alias{baceappbeta}
\title{Biomarkers Consortium ADNI CSF BACE activity and sAPPbeta}
\format{
A data frame with 402 rows and 12 variables
}
\usage{
data(baceappbeta)
}
\description{
Biomarkers Consortium ADNI CSF BACE activity and sAPPbeta
}
\details{
A dataset with variables as follows:

\itemize{
  \item USERDATE -4  Date record created
  \item GROUP -4  Group
  \item RID -4  Participant roster ID
  \item LICENSEPLATE -4  License Plate #
  \item SPECIMENID -4  Specimen ID
  \item GLOBALSPECID -4  Global Spec ID
  \item VISCODE -4  Visit code
  \item SPECDATE -4  Spec Date
  \item RECEIVEDDATE -4  Received Date
  \item CSFBACE -4 pM CSF BACE
  \item CSFSAPPBETA -4 pM CSF sAPP Beta
}
}
\examples{
\dontrun{
describe(baceappbeta)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/BACEAPPBETA/Biomarkers_Consortium_Data_Primer_for_ADNI_CSF_BACE_activity_and_sAPPbeta_2012April10.pdf}
}
\keyword{datasets}
