% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{backmeds}
\alias{backmeds}
\title{Key Background Medications}
\format{
A data frame with 12095 rows and 8 variables
}
\usage{
data(backmeds)
}
\description{
Key Background Medications
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N NA Participant roster ID
  \item SITEID N NA Site ID
  \item VISCODE T NA Visit code
  \item USERDATE S NA Date record created
  \item USERDATE2 S NA Date record last updated
  \item KEYMED T NA At this visit, is participant on any of the following medication?
}
}
\keyword{datasets}
