% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{baimrinmrc}
\alias{baimrinmrc}
\title{Banner Alzheimer's Institute MRI NMRC Summaries}
\format{
A data frame with 726 rows and 10 variables
}
\usage{
data(baimrinmrc)
}
\description{
Banner Alzheimer's Institute MRI NMRC Summaries
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item VERSION -4  Version of Table
  \item LONIUID -4 -4 -4
  \item RUNDATE -4  DATE analysis performed
  \item STATUS -4  Complete/Incomplete
  \item IPCA -4 Percentage  percent annualized whole brain atrophy detected by Iterative Principal component Analysis (IPCA) 
  \item IPCA2 -4 unitless whole brain atrophy measured from T1 weighted MRI using iterative pincipal component analysis 
}
}
\examples{
\dontrun{
describe(baimrinmrc)
}
}
\keyword{datasets}
