% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{baipetnmrc}
\alias{baipetnmrc}
\title{Banner Alzheimer's Institute PET NMRC Summaries}
\format{
A data frame with 5598 rows and 343 variables
}
\usage{
data(baipetnmrc)
}
\description{
Banner Alzheimer's Institute PET NMRC Summaries
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4  Participant roster ID
  \item VISCODE -4  Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4  Examination Date
  \item VERSION -4  Version of Table
  \item LONIUID -4  LONI Unique Image Identifier
  \item RUNDATE -4  DATE analysis performed
  \item STATUS -4  Complete/Incomplete
  \item MODALITY -4  AV45/FDG/TAU
  \item HIPPL01 -4 Normalized Counts Globally normalized CMRgl from Hippocampus_L
  \item FRTSUPL01 -4 Normalized Counts Globally normalized CMRgl from Frontal_sup_L
  \item FRTMIDL01 -4 Normalized Counts Globally normalized CMRgl from Frontal_mid_L
  \item PARAHIPL01 -4 Normalized Counts Globally normalized CMRgl from ParaHippocampal_L
  \item FUSFRML01 -4 Normalized Counts Globally normalized CMRgl from Fusiform_L
  \item OCCMIDL01 -4 Normalized Counts Globally normalized CMRgl from Occipital_mid_L
  \item ANGULL01 -4 Normalized Counts Globally normalized CMRgl from Angular_L
  \item PARIINFL01 -4 Normalized Counts Globally normalized CMRgl from Parietal_Inf_L
  \item SUPMRGL01 -4 Normalized Counts Globally normalized CMRgl from SupraMarginal_L
  \item TMPMIDL01 -4 Normalized Counts Globally normalized CMRgl from Temporal _mid_L
  \item PRECUNL01 -4 Normalized Counts Globally normalized CMRgl from Precuneus_L
  \item CINGPSTL01 -4 Normalized Counts Globally normalized CMRgl from Cingulum_Post_L
  \item PARAHIPR01 -4 Normalized Counts Globally normalized CMRgl from ParaHippocampal_R
  \item FRTSUPR01 -4 Normalized Counts Globally normalized CMRgl from Frontal_Sup_R
  \item OCCMIDR01 -4 Normalized Counts Globally normalized CMRgl from Occipital_mid_R
  \item FUSFRMR01 -4 Normalized Counts Globally normalized CMRgl from Fusiform_R
  \item FRTMIDR01 -4 Normalized Counts Globally normalized CMRgl from Frontal_mid_R
  \item ANGULR01 -4 Normalized Counts Globally normalized CMRgl from Angular_R
  \item PARIINFR01 -4 Normalized Counts Globally normalized CMRgl from Parietal_Inf_R
  \item TMPMIDR01 -4 Normalized Counts Globally normalized CMRgl from Temporal_mid_R
  \item SUPMRGR01 -4 Normalized Counts Globally normalized CMRgl from SupraMarginal_R
  \item PRCUNSR01 -4 Normalized Counts Globally normalized CMRgl from Prcuneus_R
  \item HIPPR01 -4 Normalized Counts Globally normalized CMRgl from Hippocampus_R
  \item LINGUALL01 -4 Normalized Counts Globally normalized CMRgl from Lingual_L
  \item LINGUALR01 -4 Normalized Counts Globally normalized CMRgl from Lingual_R
  \item CINGPSTR01 -4 Normalized Counts Globally normalized CMRgl from Cingulum_Post_R
  \item FRTINFL01 -4 Normalized Counts Globally normalized CMRgl from Frontal_Inf_L
  \item FRTINFR01 -4 Normalized Counts Globally normalized CMRgl from Frontal_Inf_R
  \item PARISUPL01 -4 Normalized Counts Globally normalized CMRgl from Parietal_Sup_L
  \item PARISUPR01 -4 Normalized Counts Globally normalized CMRgl from Parietal_Sup_R
  \item INSULAL01 -4 Normalized Counts Globally normalized CMRgl from Insula_L
  \item INSULAR01 -4 Normalized Counts Globally normalized CMRgl from Insula_R
  \item CINGANTL01 -4 Normalized Counts Globally normalized CMRgl from Cingulum_Ant_L
  \item CINGANTR01 -4 Normalized Counts Globally normalized CMRgl from Cingulum_Ant_R
  \item CINGMIDL01 -4 Normalized Counts Globally normalized CMRgl from Cingulum_Mid_L
  \item CINGMIDR01 -4 Normalized Counts Globally normalized CMRgl from Cingulum_Mid_R
  \item TMPSUPL01 -4 Normalized Counts Globally normalized CMRgl from Temporal _Sup_L
  \item TMPSUPR01 -4 Normalized Counts Globally normalized CMRgl from Temporal _Sup_R
  \item TMPINFL01 -4 Normalized Counts Globally normalized CMRgl from Temporal_Inf_L
  \item TMPINFR01 -4 Normalized Counts Globally normalized CMRgl from Temporal_Inf_R
  \item FRTSUPL02 -4 Normalized Counts Globally normalized CMRgl from Frontal_Sup_L
  \item FRTMIDL02 -4 Normalized Counts Globally normalized CMRgl from Frontal_Mid_L
  \item CINGPSTL02 -4 Normalized Counts Globally normalized CMRgl from Cingulum_Post_L
  \item CINGMIDL02 -4 Normalized Counts Globally normalized CMRgl from Cingulum_Mid_L
  \item FRTSMEDL02 -4 Normalized Counts Globally normalized CMRgl from Frontal_Sup_Medial_L
  \item FUSFRML02 -4 Normalized Counts Globally normalized CMRgl from Fusiform_L
  \item TMPMIDL02 -4 Normalized Counts Globally normalized CMRgl from Temporal_Mid_L
  \item OCCMIDL02 -4 Normalized Counts Globally normalized CMRgl from Occipital_Mid_L
  \item TMPINFL02 -4 Normalized Counts Globally normalized CMRgl from Temporal_Inf_L
  \item PRECUNL02 -4 Normalized Counts Globally normalized CMRgl from Precuneus_L
  \item CINGANTL02 -4 Normalized Counts Globally normalized CMRgl from Cingulum_Ant_L
  \item FRTMIDOR02 -4 Normalized Counts Globally normalized CMRgl from Frontal_Mid_Orb_R
  \item FRTMIDR02 -4 Normalized Counts Globally normalized CMRgl from Frontal_Mid_R
  \item FRTSUPR02 -4 Normalized Counts Globally normalized CMRgl from Frontal_Sup_R
  \item ANGULR02 -4 Normalized Counts Globally normalized CMRgl from Angular_R
  \item ANGULL02 -4 Normalized Counts Globally normalized CMRgl from Angular_L
  \item PARIINFR02 -4 Normalized Counts Globally normalized CMRgl from Parietal_Inf_R
  \item TMPMIDR02 -4 Normalized Counts Globally normalized CMRgl from Temporal_Mid_R
  \item INSULAR02 -4 Normalized Counts Globally normalized CMRgl from Insula_R
  \item TMPPOSR02 -4 Normalized Counts Globally normalized CMRgl from Temporal_Pole_Sup_R
  \item CINGANTR02 -4 Normalized Counts Globally normalized CMRgl from Cingulum_Ant_R
  \item CINGMIDR02 -4 Normalized Counts Globally normalized CMRgl from Cingulum_Mid_R
  \item PRECUNR02 -4 Normalized Counts Globally normalized CMRgl from Precuneus_R
  \item RECTUSR02 -4 Normalized Counts Globally normalized CMRgl from Rectus_R
  \item CINGPSTR02 -4 Normalized Counts Globally normalized CMRgl from Cingulum_Post_R
  \item PARIINFL02 -4 Normalized Counts Globally normalized CMRgl from Parietal_Inf_L
  \item INSULAL02 -4 Normalized Counts Globally normalized CMRgl from Insula_L
  \item FRTINFL02 -4 Normalized Counts Globally normalized CMRgl from Frontal_Inf_L
  \item FRTINFR02 -4 Normalized Counts Globally normalized CMRgl from Frontal_Inf_R
  \item OCCMIDR02 -4 Normalized Counts Globally normalized CMRgl from Occipital_Mid_R
  \item FRTMIDOL02 -4 Normalized Counts Globally normalized CMRgl from Frontal_Mid_Orb_Right
  \item FRTSMEDR02 -4 Normalized Counts Globally normalized CMRgl from Frontal_Sup_Medial_R
  \item FUSFMR02 -4 Normalized Counts Globally normalized CMRgl from Fusiform_R
  \item RECTUSL02 -4 Normalized Counts Globally normalized CMRgl from Rectus_L
  \item TMPPOSL02 -4 Normalized Counts Globally normalized CMRgl from Temporal_Pole_Sup_L
  \item PARISUPL02 -4 Normalized Counts Globally normalized CMRgl from Parietal_Sup_L
  \item PARISUPR02 -4 Normalized Counts Globally normalized CMRgl from Parietal_Sup_L
  \item SUPMRGL02 -4 Normalized Counts Globally normalized CMRgl from SupraMarginal_L
  \item SUPMRGR02 -4 Normalized Counts Globally normalized CMRgl from SupraMarginal_R
  \item TMPINFR02 -4 Normalized Counts Globally normalized CMRgl from Temporal_Inf_L
  \item FRTMIDL03 -4 Normalized Counts Frontal_Mid_L
  \item FRTMIDR03 -4 Normalized Counts Frontal_Mid_R
  \item FRTSUPL03 -4 Normalized Counts Frontal_Sup_L
  \item FRTSUPR03 -4 Normalized Counts Frontal_Sup_R
  \item FRTSMEDL03 -4 Normalized Counts Frontal_Sup_Medial_L
  \item FRTSMEDR03 -4 Normalized Counts Frontal_Sup_Medial_R
  \item TMPSUPL03 -4 Normalized Counts Temporal_Sup_L
  \item TMPSUPR03 -4 Normalized Counts Temporal_Sup_R
  \item TMPMIDL03 -4 Normalized Counts Temporal_Mid_L
  \item TMPMIDR03 -4 Normalized Counts Temporal_Mid_R
  \item TMPINFL03 -4 Normalized Counts Temporal_Inf_L
  \item TMPINFR03 -4 Normalized Counts Temporal_Inf_R
  \item FUSFRML03 -4 Normalized Counts Fusiform_L
  \item FUSFRMR03 -4 Normalized Counts Fusiform_R
  \item LINGUALL03 -4 Normalized Counts Lingual_L
  \item LINGUALR03 -4 Normalized Counts Lingual_R
  \item PARAHIPL03 -4 Normalized Counts ParaHippocampal_L
  \item PARAHIPR03 -4 Normalized Counts ParaHippocampal_R
  \item PARISUPL03 -4 Normalized Counts Parietal_Sup_L
  \item PARISUPR03 -4 Normalized Counts Parietal_Sup_R
  \item PARIINFL03 -4 Normalized Counts Parietal_Inf_L
  \item PARIINFR03 -4 Normalized Counts Parietal_Inf_R
  \item PRECUNL03 -4 Normalized Counts Precuneus_L
  \item PRECUNR03 -4 Normalized Counts Precuneus_R
  \item CINGANT03 -4 Normalized Counts Cingulum_Ant
  \item CINGMID03 -4 Normalized Counts Cingulum_Mid
  \item CINGPST03 -4 Normalized Counts Cingulum_Post
  \item ANGULR03 -4 Normalized Counts Angular_R
  \item OCCMIDL03 -4 Normalized Counts Occipital_Mid_L
  \item OCCMIDR03 -4 Normalized Counts Occipital_Mid_R
  \item OCCSUPL03 -4 Normalized Counts Occipital_Sup_L
  \item OCCSUPR03 -4 Normalized Counts Occipital_Sup_R
  \item OCCINFL03 -4 Normalized Counts Occipital_Inf_L
  \item OCCINFR03 -4 Normalized Counts Occipital_Inf_R
  \item THALAML03 -4 Normalized Counts Thalamus_L
  \item THALAMR03 -4 Normalized Counts Thalamus_R
  \item FRTMEDL03 -4 Normalized Counts Frontal_Medial_Left
  \item FRTMEDR03 -4 Normalized Counts Frontal_Medial_Right
  \item SUPMRGL03 -4 Normalized Counts SupraMarginal_L
  \item SUPMRGR03 -4 Normalized Counts SupraMarginal_R
  \item FRTMIDL04 -4 Normalized Counts Frontal_Mid_L
  \item FRTMIDR04 -4 Normalized Counts Frontal_Mid_R
  \item FRTSUPL04 -4 Normalized Counts Frontal_Sup_L
  \item FRTSUPR04 -4 Normalized Counts Frontal_Sup_R
  \item FRTSMEDL04 -4 Normalized Counts Frontal_Sup_Medial_L
  \item FRTSMEDR04 -4 Normalized Counts Frontal_Sup_Medial_R
  \item TMPSUPL04 -4 Normalized Counts Temporal_Sup_L
  \item TMPSUPR04 -4 Normalized Counts Temporal_Sup_R
  \item TMPMIDL04 -4 Normalized Counts Temporal_Mid_L
  \item TMPMIDR04 -4 Normalized Counts Temporal_Mid_R
  \item TMPINFL04 -4 Normalized Counts Temporal_Inf_L
  \item TMPINFR04 -4 Normalized Counts Temporal_Inf_R
  \item FUSFRML04 -4 Normalized Counts Fusiform_L
  \item FUSFRMR04 -4 Normalized Counts Fusiform_R
  \item LINGUALL04 -4 Normalized Counts Lingual_L
  \item LINGUALR04 -4 Normalized Counts Lingual_R
  \item PARAHIPL04 -4 Normalized Counts ParaHippocampal_L
  \item PARAHIPR04 -4 Normalized Counts ParaHippocampal_R
  \item PARISUPL04 -4 Normalized Counts Parietal_Sup_L
  \item PARISUPR04 -4 Normalized Counts Parietal_Sup_R
  \item PARIINFL04 -4 Normalized Counts Parietal_Inf_L
  \item PARIINFR04 -4 Normalized Counts Parietal_Inf_R
  \item PRECUNL04 -4 Normalized Counts Precuneus_L
  \item PRECUNR04 -4 Normalized Counts Precuneus_R
  \item CINGANT04 -4 Normalized Counts Cingulum_Ant
  \item CINGMID04 -4 Normalized Counts Cingulum_Mid
  \item CINGPST04 -4 Normalized Counts Cingulum_Post
  \item OCCMIDL04 -4 Normalized Counts Occipital_Mid_L
  \item OCCMIDR04 -4 Normalized Counts Occipital_Mid_R
  \item OCCSUPL04 -4 Normalized Counts Occipital_Sup_L
  \item OCCSUPR04 -4 Normalized Counts Occipital_Sup_R
  \item OCCINFL04 -4 Normalized Counts Occipital_Inf_L
  \item OCCINFR04 -4 Normalized Counts Occipital_Inf_R
  \item THALAML04 -4 Normalized Counts Thalamus_L
  \item THALAMR04 -4 Normalized Counts Thalamus_R
  \item FRTMEDL04 -4 Normalized Counts Frontal_Medial_Left
  \item FRTMEDR04 -4 Normalized Counts Frontal_Medial_Right
  \item SUPMRGL04 -4 Normalized Counts SupraMarginal_L
  \item SUPMRGR04 -4 Normalized Counts SupraMarginal_R
  \item ANGULL04 -4 Normalized Counts Angular_L
  \item ANGULR04 -4 Normalized Counts Angular_R
  \item FRTMIDL05 -4 Normalized Counts Frontal_Mid_L
  \item FRTMIDR05 -4 Normalized Counts Frontal_Mid_R
  \item FRTSUPL05 -4 Normalized Counts Frontal_Sup_L
  \item FRTSUPR05 -4 Normalized Counts Frontal_Sup_R
  \item FRTSMEDL05 -4 Normalized Counts Frontal_Sup_Medial_L
  \item FRTSMEDR05 -4 Normalized Counts Frontal_Sup_Medial_R
  \item TMPSUPL05 -4 Normalized Counts Temporal_Sup_L
  \item TMPSUPR05 -4 Normalized Counts Temporal_Sup_R
  \item TMPMIDL05 -4 Normalized Counts Temporal_Mid_L
  \item TMPMIDR05 -4 Normalized Counts Temporal_Mid_R
  \item TMPINFL05 -4 Normalized Counts Temporal_Inf_L
  \item TMPINFR05 -4 Normalized Counts Temporal_Inf_R
  \item FUSFRML05 -4 Normalized Counts Fusiform_L
  \item FUSFRMR05 -4 Normalized Counts Fusiform_R
  \item PARAHIPL05 -4 Normalized Counts ParaHippocampal_L
  \item PARAHIPR05 -4 Normalized Counts ParaHippocampal_R
  \item PARISUPL05 -4 Normalized Counts Parietal_Sup_L
  \item PARISUPR05 -4 Normalized Counts Parietal_Sup_R
  \item PARIINFL05 -4 Normalized Counts Parietal_Inf_L
  \item PARIINFR05 -4 Normalized Counts Parietal_Inf_R
  \item PRECUNL05 -4 Normalized Counts Precuneus_L
  \item PRECUNR05 -4 Normalized Counts Precuneus_R
  \item CINGANT05 -4 Normalized Counts Cingulum_Ant
  \item CINGMID05 -4 Normalized Counts Cingulum_Mid
  \item CINGPST05 -4 Normalized Counts Cingulum_Post
  \item OCCMIDL05 -4 Normalized Counts Occipital_Mid_L
  \item OCCMIDR05 -4 Normalized Counts Occipital_Mid_R
  \item OCCSUPL05 -4 Normalized Counts Occipital_Sup_L
  \item OCCSUPR05 -4 Normalized Counts Occipital_Sup_R
  \item OCCINFL05 -4 Normalized Counts Occipital_Inf_L
  \item OCCINFR05 -4 Normalized Counts Occipital_Inf_R
  \item THALAML05 -4 Normalized Counts Thalamus_L
  \item THALAMR05 -4 Normalized Counts Thalamus_R
  \item FRTMEDL05 -4 Normalized Counts Frontal_Medial_Left
  \item FRTMEDR05 -4 Normalized Counts Frontal_Medial_Right
  \item ANGULR05 -4 Normalized Counts Angular_Right
  \item ANGULL05 -4 Normalized Counts Angular_Left
  \item CALCARL05 -4 Normalized Counts Calcarine_L
  \item CALCARR05 -4 Normalized Counts Calcarine_R
  \item LINGUALL05 -4 Normalized Counts Lingual_L
  \item LINGUALR05 -4 Normalized Counts Lingual_R
  \item CEREB8L05 -4 Normalized Counts Cerebelum_8_L
  \item CEREB8R05 -4 Normalized Counts Cerebelum_8_R
  \item CERBCR2L05 -4 Normalized Counts Cerebelum_Crus2_L
  \item CERBCR2R05 -4 Normalized Counts Cerebelum_Crus2_R
  \item SUPMRGL05 -4 Normalized Counts SupraMarginal_L
  \item SUPMRGR05 -4 Normalized Counts SupraMarginal_R
  \item FRTMIDL06 -4 Normalized Counts Frontal_Mid_L
  \item FRTMIDR06 -4 Normalized Counts Frontal_Mid_R
  \item FRTSUPL06 -4 Normalized Counts Frontal_Sup_L
  \item FRTSUPR06 -4 Normalized Counts Frontal_Sup_R
  \item FRTSMEDL06 -4 Normalized Counts Frontal_Sup_Medial_L
  \item FRTSMEDR06 -4 Normalized Counts Frontal_Sup_Medial_R
  \item TMPSUPL06 -4 Normalized Counts Temporal_Sup_L
  \item TMPSUPR06 -4 Normalized Counts Temporal_Sup_R
  \item TMPMIDL06 -4 Normalized Counts Temporal_Mid_L
  \item TMPMIDR06 -4 Normalized Counts Temporal_Mid_R
  \item TMPINFL06 -4 Normalized Counts Temporal_Inf_L
  \item TMPINFR06 -4 Normalized Counts Temporal_Inf_R
  \item FUSFRML06 -4 Normalized Counts Fusiform_L
  \item FUSFRMR06 -4 Normalized Counts Fusiform_R
  \item LINGUALL06 -4 Normalized Counts Lingual_L
  \item LINGUALR06 -4 Normalized Counts Lingual_R
  \item PARAHIPL06 -4 Normalized Counts ParaHippocampal_L
  \item PARAHIPR06 -4 Normalized Counts ParaHippocampal_R
  \item PARISUPL06 -4 Normalized Counts Parietal_Sup_L
  \item PARISUPR06 -4 Normalized Counts Parietal_Sup_R
  \item PARIINFL06 -4 Normalized Counts Parietal_Inf_L
  \item PARIINFR06 -4 Normalized Counts Parietal_Inf_R
  \item PRECUNL06 -4 Normalized Counts Precuneus_L
  \item PRECUNR06 -4 Normalized Counts Precuneus_R
  \item CINGANT06 -4 Normalized Counts Cingulum_Ant
  \item CINGMID06 -4 Normalized Counts Cingulum_Mid
  \item CINGPST06 -4 Normalized Counts Cingulum_Post
  \item OCCMIDL06 -4 Normalized Counts Occipital_Mid_L
  \item OCCMIDR06 -4 Normalized Counts Occipital_Mid_R
  \item OCCSUPL06 -4 Normalized Counts Occipital_Sup_L
  \item OCCSUPR06 -4 Normalized Counts Occipital_Sup_R
  \item OCCINFL06 -4 Normalized Counts Occipital_Inf_L
  \item OCCINFR06 -4 Normalized Counts Occipital_Inf_R
  \item THALAML06 -4 Normalized Counts Thalamus_L
  \item THALAMR06 -4 Normalized Counts Thalamus_R
  \item FRTMEDL06 -4 Normalized Counts Frontal_Medial_Left
  \item FRTMEDR06 -4 Normalized Counts Frontal_Medial_Right
  \item HIPPL06 -4 Normalized Counts Hippocampus_L
  \item HIPPR06 -4 Normalized Counts Hippocampus_R
  \item INDGLBNDX -4 Percentage Individualized Global Index
  \item INDTEPNDX -4 Percentage Individualized Temporal index 
  \item PRECUNL07 -4 Normalized Counts Precuneus_L
  \item CINGPST07 -4 Normalized Counts Cingulum_Post
  \item PARIINFR07 -4 Normalized Counts Parietal_Inf_R
  \item TMPMIDL07 -4 Normalized Counts Temporal_Mid_L
  \item CUNEUSL08 -4 Normalized Counts Cuneus_L
  \item TMPMIDR08 -4 Normalized Counts Temporal_Mid_R
  \item TMPINFL09 -4 Normalized Counts Temporal_Inf_L
  \item CINGPST09 -4 Normalized Counts Cingulum_Post
  \item PRECUNL09 -4 Normalized Counts Precuneus_L
  \item PARIINFR09 -4 Normalized Counts Parietal_Inf_R
  \item TMPINFL10 -4 Normalized Counts Temporal_Inf_L
  \item PRECUNL10 -4 Normalized Counts Precuneus_L
  \item TMPINFL11 -4 Normalized Counts Temporal_Inf_L
  \item OCCMIDL11 -4 Normalized Counts Occipital_Mid_L
  \item CINGPST11 -4 Normalized Counts Cingulum_Post_L
  \item CINGMID11 -4 Normalized Counts Cingulum_Post_R/Cingulum_Mid_R
  \item TMPPOSR12 -4 Normalized Counts Temporal_Pole_Sup_R
  \item CINGPSTR12 -4 Normalized Counts Cingulum_Post_R/Cingulum_Mid_R
  \item FROITRN001 -4 Normalized Counts Training group, all subjects
  \item FROITRN002 -4 Normalized Counts Training group, exclude HRRT and BioGraph HiRez
  \item FROITST001 -4 Normalized Counts Testing group, all scans
  \item FROITST002 -4 Normalized Counts Testing group, exclude HRRT and BioGraph HiRez
  \item FROIWHL001 -4 Normalized Counts All subjects
  \item FROIWHL002 -4 Normalized Counts Exclude HRRT and BioGraph HiRez
  \item FRPMSK001 -4 Normalized Counts Exclude HRRT and BioGraph HiRez
  \item FRPMSK002 -4 Normalized Counts All subjects
  \item FRTMSK003 -4 Normalized Counts Exclude HRRT and BioGraph HiRez
  \item FRTMSK004 -4 Normalized Counts All subjects
  \item FRTMSK005 -4 Normalized Counts All subjects
  \item FRTMSK006 -4 Normalized Counts All subjects
  \item FRTMSK007 -4 Normalized Counts All subjects
  \item FRTMSK008 -4 Normalized Counts Exclude HRRT and BioGraph HiRez
  \item FRTMSK009 -4 Normalized Counts All subjects
  \item FRTMSK010 -4 Normalized Counts Exclude HRRT and BioGraph HiRez
  \item FRTMSK011 -4 Normalized Counts All subjects
  \item HCISUB1 -4 unitless hypometabolic convergence index for 74 AD subjects
  \item HCI99 -4 unitless hypometabolic convergence index for normal controls, stable MCI, MCI converter and ADs
  \item HCI -4 unitless hypometabolic convergence index for ADNI1/ADNI2/ADNIGO subjects
  \item NL1HIPP -4 Normalized Counts difference of globally normalized CMRgl from Baseline to 12 months and 12 months to 24 months from Hippocampus for 58 AD patients
  \item NL1PTEMP -4 Normalized Counts difference of globally normalized CMRgl from Baseline to 12 months and 12 months to 24 months from the parietotemporal region for 58 AD patients
  \item NL1PSTCG -4 Normalized Counts difference of globally normalized CMRgl from Baseline to 12 months and 12 months to 24 months from the posterior cingulate for for 58 AD patients
  \item NL2PSTCG -4 Normalized Counts difference of globally normalized CMRgl from Baseline to 12 months and 12 months to 24 months from Posterior Cingulate in 129 MCIs.
  \item NL2PARI -4 Normalized Counts difference of globally normalized CMRgl from 12 months to 24 months from the inferior parietal region in 129 MCI subjects
  \item NL3FRONT -4 Normalized Counts difference of globally normalized CMRgl from Baseline to 12 months and 12 months to 24 months from superior frontal region of 74 normal controls.
  \item DECLADFR -4 Normalized Counts difference of globally normalized CMRgl from Baseline to 12 months from the superior frontal region of 74 AD patients
  \item DECLADTMP -4 Normalized Counts CMRgl from Baseline from the inferior temporal region of 74 AD patients
  \item LONGHCI -4 unitless Extension of HCI to longitudinal data
  \item BSLVBM -4 unitless VBManalysis of structural MRI data
  \item ACI -4 unitless amyloid convergence index for AV-45 subjects
  \item MCSUVR -4 unitless mean cortical standard uptake value ratio
  \item FNMPREC -4 unitless precuneus FDG measure
  \item FNAPREC -4 unitless precuneus FDG measure
  \item FNAPC -4 unitless posterior cingulate FDG measure
  \item FEAPREC -4 unitless precuneus FDG measure
  \item FMAPREC -4 unitless precuneus FDG measure
  \item FMAPC -4 unitless posterior cingulate FDG measure
  \item AVNEFRONT -4 unitless frontal AV-45 measure
  \item AVNMFRONT -4 unitless frontal AV-45 measure
  \item AVNAFRONT -4 unitless frontal AV-45 measure
  \item AVEMTEMP -4 unitless temporal AV-45 measure
  \item AVEATEMP -4 unitless temporal AV-45 measure 
  \item AVMAOCCIP -4 unitless occipital AV-45 measure
  \item FEAPC2 -4 unitless posterior cingulate FDG measure
  \item FEMPC2 -4 unitless posterior cingulate FDG measure
  \item FMAPREC2 -4 unitless precuneus FDG measure
  \item FNAPARI2 -4 unitless parietal FDG measure
  \item FNEHC2 -4 unitless hippocampal FDG measure
  \item FNMTEMP2 -4 unitless temporal FDG measure
  \item AVEATEMP2 -4 unitless temporal AV-45 measure
  \item AVEMTEMP2 -4 unitless temporal AV-45 measure
  \item AVMAFRNT2 -4 unitless frontal AV-45 measure
  \item AVNATEMP2 -4 unitless temporal AV-45 measure
  \item AVNEFRNT2 -4 unitless frontal AV-45 measure
  \item AVNMPREC2 -4 unitless precuneus AV-45 measure
  \item MMPLS -4 unitless AVfdgPLS
  \item ICA -4 unitless AV_ICA
  \item PCA2 -4 unitless AV_SSM
  \item PLSFDG01 -4 unitless PLS for FDG PET
  \item HCI001 -4 unitless Hypometabolic Convergence Index
  \item MCSUVRWM -4 unitless Mean Cortical SUVR Corpus Callosum and Centrum Semiovale Combined
  \item MCSUVRWM2 -4 unitless Mean Cortical SUVR Corpus Callosum and Centrum Semiovale Combined
  \item HCI_2014 -4 unitless Hypometabolic Convergence Index
  \item SROI -4 unitless statistical region of interest (sROI) developed and validated for optimal longitudinal FDG-PET CMRgl changes
  \item MCSUVRCERE -4 unitless Mean Cortical SUVR with Cerebellum as reference region
  \item MCSUVRWM3 -4 unitless Mean Cortical SUVR with Corpus Callosum and Centrum Semiovale Combined as reference region
}
}
\examples{
\dontrun{
describe(baipetnmrc)
}
}
\keyword{datasets}
