% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{baipetnmrcav45}
\alias{baipetnmrcav45}
\title{Banner Alzheimer's Institute PET NMRC AV45 Summaries}
\format{
A data frame with 2697 rows and 10 variables
}
\usage{
data(baipetnmrcav45)
}
\description{
Banner Alzheimer's Institute PET NMRC AV45 Summaries
}
\details{
A dataset with variables as follows:

\itemize{
  \item ORIGPROT -4 -4 Original study protocol
  \item RID N  Participant roster ID
  \item VISCODE T  Visit code
  \item VISCODE2 T -4 Translated visit code
  \item EXAMDATE D  Examination Date
  \item LONIUID N  LONI Unique Image Identifier
  \item RUNDATE D  DATE analysis performed
  \item STATUS T  Complete/Incomplete
  \item MODALITY T  The type of pet scan used for analysis
  \item MCSUVRWM N unitless Mean Cortical SUVR with Corpus Callosum and Centrum Semiovale Combined as reference region
  \item MCSUVRCERE N unitless Mean Cortical SUVR with Cerebellum as reference region
}
}
\examples{
\dontrun{
describe(baipetnmrcav45)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/BAIPETNMRCAV45/Banner_SUVRwithCerebralWMRefROI_forLongitudinalFlorbetapirPET\%202016\%2008.pdf}, \url{https://adni.bitbucket.io/reference/docs/BAIPETNMRCAV45/Banner_SUVRwithCerebralWMRefROI_forLongitudinalFlorbetapirPET.pdf}
}
\keyword{datasets}
