% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{baipetnmrcftp}
\alias{baipetnmrcftp}
\title{Banner Alzheimer's Institute PET NMRC Flortaucipir (F-AV1451) Summaries}
\format{
A data frame with 1157 rows and 12 variables
}
\usage{
data(baipetnmrcftp)
}
\description{
Banner Alzheimer's Institute PET NMRC Flortaucipir (F-AV1451) Summaries
}
\details{
A dataset with variables as follows:

\itemize{
  \item ORIGPROT -4 -4 Original study protocol
  \item COLPROT -4 -4 Study protocol of data collection
  \item RID N  Participant roster ID
  \item VISCODE T  Visit code
  \item VISCODE2 T -4 Translated visit code
  \item EXAMDATE D  Examination Date
  \item LONIUID N  LONI Unique Image Identifier
  \item RUNDATE D  DATE analysis performed
  \item STATUS T  Complete/Incomplete
  \item MODALITY T  The type of pet scan used for analysis
  \item ENTORHINAL_SUVR N unitless Entorhinal SUVR with Cerebellum Crus I as reference region 
  \item INFERIOR_TEMPORAL_SUVR N unitless Inferior temporal SUVR with Cerebellum Crus I as reference region 
  \item TAU_METAROI N unitless Tau metaROI with Cerebellum Crus as reference region 
}
}
\examples{
\dontrun{
describe(baipetnmrcftp)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/BAIPETNMRCFTP/BAI-PET-NMRC-Flortaucipir-Processing-Methods.pdf}
}
\keyword{datasets}
