% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{batemanlab}
\alias{batemanlab}
\title{Bateman Lab Plasma Abeta42/Abeta40 Ratio as a Predictor of Brain Amyloidosis}
\format{
A data frame with 622 rows and 72 variables
}
\usage{
data(batemanlab)
}
\description{
Bateman Lab Plasma Abeta42/Abeta40 Ratio as a Predictor of Brain Amyloidosis
}
\details{
A dataset with variables as follows:

\itemize{
  \item ASSAY -4 na Experiment file name
  \item ASSAY_FLUID -4 na Main biological matrix of experiment
  \item ASSAY_IP_PROTOCOL -4 na Sample processing protocol
  \item BATCH_N -4 na Indicates the batch in which the sample was processed (prepped) for LC-MSMS
  \item FILE_NAME_NE -4 na Exact file name of LC-MSMS rum without extention
  \item INJECTION -4 na injection (technical replicate)  1-st denoted as 'a'; second as 'b' ei.
  \item INSTRUMENT -4 na Masspectrometer
  \item LC.MS_SAMPLE_ID -4 na Internal ID of LS_MS sample to control for technical replicates. It serves as internal blinding ID and for QC purpocess
  \item MS_RUN_DATE -4 na date when sequense get started to run on LC-MSMS
  \item QC_FLAGS -4 na None or one or many QC flags assigned by QC comitte when sample does not meet respective quality criterias
  \item QC_STATUS -4 na Status of samples assign by quality control comitte (Passed; Failed). If failed, samples should not be used for further analysis
  \item SAMPLE_TYPE -4 na Type of LC_MS sample
  \item SUBJECT -4 na Subject or biosample ID or blinded ID. For Clinical samples this column should be key for unblinding
  \item PEAK.AREA_ABETA38_N14.B07 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b7 for the LysN C-terminal peptide from endogenous Aβ38
  \item PEAK.AREA_ABETA38_N14.B08 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b8 for the LysN C-terminal peptide from endogenous Aβ38
  \item PEAK.AREA_ABETA38_N14.B09 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b9 for the LysN C-terminal peptide from endogenous Aβ38
  \item PEAK.AREA_ABETA38_N15.B07 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b7 for the LysN C-terminal peptide from 15N labled Aβ38 internal standard
  \item PEAK.AREA_ABETA38_N15.B08 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b8 for the LysN C-terminal peptide from 15N labled Aβ38 internal standard
  \item PEAK.AREA_ABETA38_N15.B09 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b9 for the LysN C-terminal peptide from 15N labled Aβ38 internal standard
  \item PEAK.AREA_ABETA40_N14.B05 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b5 for the LysN C-terminal peptide from endogenous Aβ40
  \item PEAK.AREA_ABETA40_N14.B07 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b7 for the LysN C-terminal peptide from endogenous Aβ40
  \item PEAK.AREA_ABETA40_N14.B08 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b8 for the LysN C-terminal peptide from endogenous Aβ40
  \item PEAK.AREA_ABETA40_N14.B09 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b9 for the LysN C-terminal peptide from endogenous Aβ40
  \item PEAK.AREA_ABETA40_N14.B10 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b10 for the LysN C-terminal peptide from endogenous Aβ40
  \item PEAK.AREA_ABETA40_N14.B11 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b11 for the LysN C-terminal peptide from endogenous Aβ40
  \item PEAK.AREA_ABETA40_N15.B05 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b5 for the LysN C-terminal peptide from 15N labled Aβ40 internal standard
  \item PEAK.AREA_ABETA40_N15.B07 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b7 for the LysN C-terminal peptide from 15N labled Aβ40 internal standard
  \item PEAK.AREA_ABETA40_N15.B08 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b8 for the LysN C-terminal peptide from 15N labled Aβ40 internal standard
  \item PEAK.AREA_ABETA40_N15.B09 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b9 for the LysN C-terminal peptide from 15N labled Aβ40 internal standard
  \item PEAK.AREA_ABETA40_N15.B10 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b10 for the LysN C-terminal peptide from 15N labled Aβ40 internal standard
  \item PEAK.AREA_ABETA40_N15.B11 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b11 for the LysN C-terminal peptide from 15N labled Aβ40 internal standard
  \item PEAK.AREA_ABETA42_N14.B08 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b8 for the LysN C-terminal peptide from endogenous Aβ42
  \item PEAK.AREA_ABETA42_N14.B09 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b9 for the LysN C-terminal peptide from endogenous Aβ42
  \item PEAK.AREA_ABETA42_N14.B10 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b10 for the LysN C-terminal peptide from endogenous Aβ42
  \item PEAK.AREA_ABETA42_N14.B11 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b11 for the LysN C-terminal peptide from endogenous Aβ42
  \item PEAK.AREA_ABETA42_N14.B12 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b12 for the LysN C-terminal peptide from endogenous Aβ42
  \item PEAK.AREA_ABETA42_N14.B13 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b13 for the LysN C-terminal peptide from endogenous Aβ42
  \item PEAK.AREA_ABETA42_N15.B08 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b8 for the LysN C-terminal peptide from 15N labled Aβ42 internal standard
  \item PEAK.AREA_ABETA42_N15.B09 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b9 for the LysN C-terminal peptide from 15N labled Aβ42 internal standard
  \item PEAK.AREA_ABETA42_N15.B10 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b10 for the LysN C-terminal peptide from 15N labled Aβ42 internal standard
  \item PEAK.AREA_ABETA42_N15.B11 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b11 for the LysN C-terminal peptide from 15N labled Aβ42 internal standard
  \item PEAK.AREA_ABETA42_N15.B12 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b12 for the LysN C-terminal peptide from 15N labled Aβ42 internal standard
  \item PEAK.AREA_ABETA42_N15.B13 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b13 for the LysN C-terminal peptide from 15N labled Aβ42 internal standard
  \item PEAK.AREA_ABETAMD_N14.B04 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b4 for the LysN Mid-Domain from endogenous Aβ
  \item PEAK.AREA_ABETAMD_N14.B05 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b5 for the LysN Mid-Domain from endogenous Aβ
  \item PEAK.AREA_ABETAMD_N14.B06 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b6 for the LysN Mid-Domain from endogenous Aβ
  \item PEAK.AREA_ABETAMD_N14.B07 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b7 for the LysN Mid-Domain from endogenous Aβ
  \item PEAK.AREA_ABETAMD_N14.B08 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b8 for the LysN Mid-Domain from endogenous Aβ
  \item PEAK.AREA_ABETAMD_N14.B09 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b9 for the LysN Mid-Domain from endogenous Aβ
  \item PEAK.AREA_ABETAMD_N14.B10 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b10 for the LysN Mid-Domain from endogenous Aβ
  \item PEAK.AREA_ABETAMD_N14.B11 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b11 for the LysN Mid-Domain from endogenous Aβ
  \item PEAK.AREA_ABETAMD_N15.B04 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b4 for the LysN Mid-Domain from 15N labled Aβ internal standard
  \item PEAK.AREA_ABETAMD_N15.B05 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b5 for the LysN Mid-Domain from 15N labled Aβ internal standard
  \item PEAK.AREA_ABETAMD_N15.B06 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b6 for the LysN Mid-Domain from 15N labled Aβ internal standard
  \item PEAK.AREA_ABETAMD_N15.B07 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b7 for the LysN Mid-Domain from 15N labled Aβ internal standard
  \item PEAK.AREA_ABETAMD_N15.B08 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b8 for the LysN Mid-Domain from 15N labled Aβ internal standard
  \item PEAK.AREA_ABETAMD_N15.B09 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b9 for the LysN Mid-Domain from 15N labled Aβ internal standard
  \item PEAK.AREA_ABETAMD_N15.B10 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b10 for the LysN Mid-Domain from 15N labled Aβ internal standard
  \item PEAK.AREA_ABETAMD_N15.B11 -4 na Integrated Peak Area for chromatograms derived from MS/MS transition ion b11 for the LysN Mid-Domain from 15N labled Aβ internal standard
  \item PEAK.AREA_ABETA40_N14.TOUSE -4 na Summed intergrated peak Area across all quantitative MS/MS transitions for peptide specific to endogenous Aβ40
  \item PEAK.AREA_ABETA40_N15.TOUSE -4 na Summed intergrated peak Area across all quantitative MS/MS transitions for peptide specific to 15N labeled Aβ40 internal standard
  \item PEAK.AREA_ABETA42_N14.TOUSE -4 na Summed intergrated peak Area across all quantitative MS/MS transitions for peptide specific to endogenous Aβ42
  \item PEAK.AREA_ABETA42_N15.TOUSE -4 na Summed intergrated peak Area across all quantitative MS/MS transitions for peptide specific to 15N labeled Aβ42 internal standard
  \item RATIO_ABETA42.40_BY_ISTD.TOUSE -4 na Final Calculated 42/40 Ratio
}
}
\examples{
\dontrun{
describe(batemanlab)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/BATEMANLAB/Batemanlab_ADNI_Methods_20190612.pdf}
}
\keyword{datasets}
