% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{bhr}
\alias{bhr}
\title{Brain Health Registry}
\format{
A data frame with 1093 rows and 11 variables
}
\usage{
data(bhr)
}
\description{
Brain Health Registry
}
\details{
A dataset with variables as follows:

\itemize{
  \item DONE N NA Was BHR brochure given to participant and BHR Study Partner Flyer given to the study partner?
  \item NDREASON N NA If No, reason not done:
  \item EXAMDATE D YYYY-MM-DD Examination Date
  \item WRITTEN N NA If Yes, was the ADNI Online ID written down on the participant's brochure?
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
}
}
\examples{
\dontrun{
describe(bhr)
}
}
\keyword{datasets}
