% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{biomark}
\alias{biomark}
\title{Biomarker Samples}
\format{
A data frame with 12548 rows and 56 variables
}
\usage{
data(biomark)
}
\description{
Biomarker Samples
}
\details{
A dataset with variables as follows:

\itemize{
  \item BILPOTPROC N NA Check if any of the following was performed
  \item NEEDLESIZE N  Needle Gauge
  \item COLTUBETYP N  Type of collection tube used
  \item SHPTUBETYP N  Type of tube used for shipping
  \item TUBEMIN N minutes If collected in polystyrene and shipped in polypropylene, please provide estimated amount of time CSF remained in polystyrene collection tube
  \item INTERSPACE N  LP performed at the
  \item PTPOSITION N  Patient Position
  \item BLREASON N  If no, please provide reason why blood was not collected
  \item BICSFFAST N  Was participant fasting for at least 6 hours?
  \item NEELDEOTH T  accident
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item RECNO N  
  \item BIBLOOD N  Which of the following was collected at this visit?
  \item BIURINE N  
  \item BICSF N  
  \item BINONE N  
  \item BINEEDLE N  Needle Used:
  \item BIMETHOD N  Method of Collection:
  \item BIFAST N  Overnight fast from midnight?
  \item EXAMDATE D  Examination Date
  \item BITIME T  Time of Collection
  \item BIREDTIME T  Time Collected  <!--Red Top-->
  \item BIREDAMT N  Amount Collected  <!--Serum-->
  \item BIREDCENT T  Centrifuged Time  <!--Serum-->
  \item BIREDTRNS T  Transfer Time  <!--Serum-->
  \item BIREDVOL N  Volume of Serum Transferred  <!--Serum-->
  \item BIREDFROZ T  Time Frozen  <!--Serum-->
  \item BILAVTIME T  Time Collected <!--Plasma-->
  \item BILAVAMT N  Amount Collected  <!--Plasma-->
  \item BILAVCENT T  Centrifuged Time <!--Plasma-->
  \item BILAVTRNS T  Transfer Time  <!--Plasma-->
  \item BILAVVOL N  Volume of Plasma Transferred  <!--Plasma-->
  \item BILAVFROZ T  Time Frozen  <!--Plasma-->
  \item BIURITIME T  Time Collected  <!--Urine-->
  \item BIURIAMT N  Amount Collected  <!--Urine-->
  \item BIURITRNS T  Transfer Time  <!--Urine-->
  \item BIURIVOL N  Volume of Urine Transfered  <!--Urine-->
  \item BIURIFROZ T  Time Frozen  <!--Urine-->
  \item BICSFTIME T  Time Collected  <!--CSF-->
  \item BICSFAMT N  Amount Collected  <!--CSF-->
  \item BICSFTRNS T  Transfer Time  <!--CSF-->
  \item BICSFVOL N  Volume of CSF Transferred  <!--CSF-->
  \item BICSFFROZ T  Time Frozen  <!--CSF-->
  \item BILPPATCH N  Check if any of the following was performed:
  \item BILPFLURO N  
  \item BILPSPFILM N  
  \item BILPPADATE D  Date of Blood Patch
  \item BILPFLDATE D  Date of Fluoroscopy
  \item BILPSPDATE D  Date of Spine Film
  \item BILPSPCOM T  If Spine Film performed, but no CSF was collected, provide explanation
  \item USERDATE2 S  Date record last updated
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
