% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{biomarkqc}
\alias{biomarkqc}
\title{Biomarker Samples Tracking and QC}
\format{
A data frame with 2702 rows and 15 variables
}
\usage{
data(biomarkqc)
}
\description{
Biomarker Samples Tracking and QC
}
\details{
A dataset with variables as follows:

\itemize{
  \item DRAWDATE D YYYY-MM-DD Sample Collection Date
  \item RECPLASMA N NA Plasma received?
  \item PLASMAQC N NA If yes, quality
  \item RECSERUM N NA Serum received?
  \item SERUMQC N NA If yes, quality
  \item RECCSF N NA CSF received?
  \item CSFQC N NA If yes, quality
  \item REDRAW N NA Notify the Monitor (e.g. to request redraw or other quality issues)?
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
}
}
\examples{
\dontrun{
describe(biomarkqc)
}
}
\keyword{datasets}
