% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{blchange}
\alias{blchange}
\title{Diagnostic Summary - Baseline Changes}
\format{
A data frame with 13525 rows and 23 variables
}
\usage{
data(blchange)
}
\description{
Diagnostic Summary - Baseline Changes
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item BCPREDX N  Pre-visit Diagnosis
  \item BCADAS N  1. Clinically relevant worsening on ADAS?
  \item BCMMSE N  2. Clinically relevant worsening on MMSE?
  \item BCMMSREC N  3. Clinically relevant worsening on MMSE recall?
  \item BCNMMMS N  4. Clinically relevant worsening on non-memory MMSE items?
  \item BCNEUPSY N  5. Clinically relevant worsening in memory on neuropsych testing?
  \item BCNONMEM N  6. Clinically relevant impairment/worsening in non-memory cognitive domains on neuropsych testing?
  \item BCFAQ N  7. Clinically relevant worsening in activities of daily living (FAQ)?
  \item BCCDR N  8. Clinically relevant deterioration on CDR Sum of Boxes or Overall CDR rating?
  \item BCDEPRES N  9. Clinically relevant depression based on clinical judgement or GDS?
  \item BCSTROKE N  10. Did subject have a stroke?
  \item BCDELIR N  11. Is there evidence of a delirium (medication effect, toxic or metabolic encephalopathy)?
  \item BCEXTCIR N  12. Has extenuating circumstance (such as a physical health problem, change in residence, charge in support network, death of a family member, etc.) contributed to a change in the subject's cognitive or functional performance?
  \item BCCORADL N  13. Is the change in clinical status corroborated by informant report of changes in ADL?
  \item BCCORCOG N  14. Is the change in clinical status corroborated by informant report of changes in cognition?
  \item USERDATE2 S  Date record last updated
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
