% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{blennow_lab_csf_gap_43}
\alias{blennow_lab_csf_gap_43}
\title{Blennow Lab CSF GAP 43}
\format{
A data frame with 1268 rows and 7 variables
}
\usage{
data(blennow_lab_csf_gap_43)
}
\description{
Blennow Lab CSF GAP 43
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item UGOT.NUMBER -4  Participant ID Blennow Lab
  \item SAMPLE.ID -4  Participant ID ADNI
  \item GAP.43 -4 pg/mL CSF GAP-43 ELISA
}
}
\examples{
\dontrun{
describe(blennow_lab_csf_gap_43)
}
}
\keyword{datasets}
