% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{blennowcsfnfl}
\alias{blennowcsfnfl}
\title{Blennow Lab ADNI1 CSF NFL}
\format{
A data frame with 416 rows and 7 variables
}
\usage{
data(blennowcsfnfl)
}
\description{
Blennow Lab ADNI1 CSF NFL
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item USERDATE -4 -4 Date record created
  \item EXAMDATE -4 -4 Examination Date
  \item VOLUME -4  Volume
  \item CSFNFL -4 ng/L -4
}
}
\examples{
\dontrun{
describe(blennowcsfnfl)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/BLENNOWCSFNFL/Blennow\%20Lab\%20-\%20ADNI-1\%20-\%20Method\%20-\%20CSF\%20NFL.pdf}
}
\keyword{datasets}
