% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{blennowcsfng}
\alias{blennowcsfng}
\title{Blennow Lab ADNI1 CSF Ng}
\format{
A data frame with 416 rows and 6 variables
}
\usage{
data(blennowcsfng)
}
\description{
Blennow Lab ADNI1 CSF Ng
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4  Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item USERDATE -4 -4 Date record created
  \item EXAMDATE -4  Examination Date
  \item CSFNG -4 pg/mL CSFNG
}
}
\examples{
\dontrun{
describe(blennowcsfng)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/BLENNOWCSFNG/Blennow\%20Lab\%20-\%20ADNI-1\%20-\%20Method\%20-\%20CSF\%20Ng.pdf}
}
\keyword{datasets}
