% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{blennowplasmanfl}
\alias{blennowplasmanfl}
\title{Blennow Lab ADNI1 Plasma neurofilament light (NFL)}
\format{
A data frame with 598 rows and 7 variables
}
\usage{
data(blennowplasmanfl)
}
\description{
Blennow Lab ADNI1 Plasma neurofilament light (NFL)
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item BOX_POS -4  Box position
  \item VOL -4  Sample volume
  \item PLASMA_NFL -4 pg/ml Plasma NFL
}
}
\examples{
\dontrun{
describe(blennowplasmanfl)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/BLENNOWPLASMANFL/Blennow\%20Lab\%20-\%20ADNI-1\%20-\%20Method\%20-\%20Plasma\%20NFL.pdf}
}
\keyword{datasets}
