% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{blennowplasmanfllong}
\alias{blennowplasmanfllong}
\title{Blennow Lab ADNI1-2 Plasma neurofilament light (NFL) longitudinal}
\format{
A data frame with 3762 rows and 9 variables
}
\usage{
data(blennowplasmanfllong)
}
\description{
Blennow Lab ADNI1-2 Plasma neurofilament light (NFL) longitudinal
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4  Participant roster ID
  \item VISCODE -4  Visit code
  \item VISCODE2 -4  Translated visit code
  \item USERDATE -4  Date record created
  \item EXAMDATE -4  Examination Date
  \item RECNO -4 -4 -4
  \item DRAW.DATE -4 -4 -4
  \item DRAW.TIME -4 -4 -4
  \item PLASMA_NFL -4 pg/ml Plasma neurofilament light (NFL)
}
}
\examples{
\dontrun{
describe(blennowplasmanfllong)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/BLENNOWPLASMANFLLONG/Blennow_Lab_-_ADNI-1-2_-_Method_-_Plasma_NFL_longitudinal_UA.pdf}
}
\keyword{datasets}
