% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{blennowplasmatau}
\alias{blennowplasmatau}
\title{Blennow Lab ADNI1 Plasma Tau}
\format{
A data frame with 598 rows and 7 variables
}
\usage{
data(blennowplasmatau)
}
\description{
Blennow Lab ADNI1 Plasma Tau
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item USERDATE -4 -4 Date record created
  \item EXAMDATE -4 -4 Examination Date
  \item VOL -4  Volume
  \item PLASMATAU -4 pg/mL -4
}
}
\examples{
\dontrun{
describe(blennowplasmatau)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/BLENNOWPLASMATAU/Blennow\%20Lab\%20-\%20ADNI-1\%20-\%20Method\%20-\%20Plasma\%20tau.pdf}
}
\keyword{datasets}
