% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{blscheck}
\alias{blscheck}
\title{Baseline Symptoms Checklist}
\format{
A data frame with 2550 rows and 36 variables
}
\usage{
data(blscheck)
}
\description{
Baseline Symptoms Checklist
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item BCNAUSEA N  1. Nausea
  \item BCVOMIT N  2. Vomiting
  \item BCDIARRH N  3. Diarrhea
  \item BCCONSTP N  4. Constipation
  \item BCABDOMN N  5. Abdominal discomfort
  \item BCSWEATN N  6. Sweating
  \item BCDIZZY N  7. Dizziness
  \item BCENERGY N  8. Low energy
  \item BCDROWSY N  9. Drowsiness
  \item BCVISION N  10. Blurred vision
  \item BCHDACHE N  11. Headache
  \item BCDRYMTH N  12. Dry mouth
  \item BCBREATH N  13. Shortness of breath
  \item BCCOUGH N  14. Coughing
  \item BCPALPIT N  15. Palpitations
  \item BCCHEST N  16. Chest pain
  \item BCURNDIS N  17. Urinary discomfort (e.g., burning)
  \item BCURNFRQ N  18. Urinary frequency
  \item BCANKLE N  19. Ankle swelling
  \item BCMUSCLE N  20. Musculoskeletal pain
  \item BCRASH N  21. Rash
  \item BCINSOMN N  22. Insomnia
  \item BCDPMOOD N  23. Depressed mood
  \item BCCRYING N  24. Crying
  \item BCELMOOD N  25. Elevated mood
  \item BCWANDER N  26. Wandering
  \item BCFALL N  27. Fall
  \item BCOTHER N  28. Other
  \item USERDATE2 S  Date record last updated
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
