% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{bsi}
\alias{bsi}
\title{Boundary Shift Integral Summaries (Nick Fox's lab)}
\format{
A data frame with 3602 rows and 18 variables
}
\usage{
data(bsi)
}
\description{
Boundary Shift Integral Summaries (Nick Fox's lab)
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item RECNO N  
  \item EXAMDATE D  Examination Date
  \item IMAGEUID N  imageUID
  \item FLDSTRENG N  Field Strength
  \item BBSI N ml Boundary Shift Integral from baseline
  \item DBCBBSI N ml Differential Bias Corrected Boundary Shift Integral from baseline
  \item VBSI N  Ventricular BSI
  \item BRAINVOL N ml Segmented brain volume. Not to be used as a measure of intracranial volume.
  \item VENTVOL N ml Segmented ventricular volume
  \item REGRATING N  
  \item VENTACCEPT N  A boolean to indicate whether or not we have confidence in the VBSI value. Similar purpose to the REGRATING.
  \item KMNDBCBBSI N  An improved version of the Differential Bias Corrected Boundary Shift Integral from baseline, which: a) Has a new intensity normalisation step that normalises scan intensity by matching the intensities by tissue type (cerebrospinal fluid, grey matter, white mater) instead of being just based on the average whole brain intensity. This helps reduce artefactual change in the BSI that might be caused by change in contrasts between scans (e.g scanner changes) b) Automatically calculates the BSI window based on k-means clustering of tissue intensities for each scan pair
  \item KMNREGRATING N  
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
