% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{cbb_results}
\alias{cbb_results}
\title{Cogstate Battery Results}
\format{
A data frame with 22197 rows and 43 variables
}
\usage{
data(cbb_results)
}
\description{
Cogstate Battery Results
}
\details{
A dataset with variables as follows:

\itemize{
  \item ProtocolID -4 Text/Number Study reference (i.e., ADNI2 or ADNI3).
  \item RID -4 Number Participant roster ID
  \item SessionID -4 Number Identification number for the session. 
  \item EXAMDATE -4 MM/DD/YYYY Examination Date
  \item TestTime -4 hh:mm:ss Time at the start of the test session.
  \item Visit -4 Text Visit label. The visit options are:In Clinic = indicates the session was completed in the clinic.Remote = indicates the session was completed remotely (i.e., at-home)Unknown = it is not known whether the session was completed in the clinic or remotely (Unknown visit labels may change to In Clinic/ Remote over time as data is queried and clarified by sites).
  \item SessionAttempt -4 Number The cumulative total of attempts of the Cogstate battery (including the current assessment).
  \item SessionDuration -4 Seconds (ADNI2) or milliseconds (ADNI3) The total duration of the current session.
  \item SessionCompletionPass -4 Text Session completion criteria.(i) Yes = all tests in session met completion criteria. If a test is attempted multiple times within the session, TestCompletionPass must = Yes for at least one of the attempts.(ii) No = at least one test in the session did not meet completion criteria. If a test is attempted multiple times within the session, TestCompletionPass = No for all attempts.(iii) If one or more tests in the battery do not have any data for the real test (and therefore, the test does not appear in the extract), SessionCompletionPass = No.
  \item SessionPerformancePass -4 Text Session performance criteria.(i) Yes = all tests in the session met the performance check. If a test is attempted multiple times within the session, TestPerformancePass must = Yes for the most recent attempt that passed completion.(ii) No = at least one test in the session did not meet performance check. If a test is attempted multiple times within the session, TestPerformancePass = No for the most recent attempt that passed completion.(iii) If TestCompletionPass = No for all attempts for a particular test, SessionPerformancePass = No.(iv) If one or more tests in the battery do not have any data for the real test (and therefore, the test does not appear in the extract), SessionPerformancePass = No.
  \item SessionIntegrityPass -4 Text Session integrity criteria.(i) Yes = all tests in the session met the integrity check. If a test is attempted multiple times within the session, TestIntegrityPass must = Yes for the most recent attempt that passed completion.(ii) No = at least one test in the session did not meet the integrity check. If a test is attempted multiple times within the session, TestIntegrityPass = No for the most recent attempt that passed completion.(iii) If TestCompletionPass = No for all attempts of a test that contributes to the integrity checks, SessionIntegrityPass = No.(iv) If one or more tests in the battery that contribute to the integrity checks do not have any data for the real test (and therefore, the test does not appear in the extract), SessionIntegrityPass = No.
  \item TestCode -4 Text Code name for the test.
  \item TestAttempt -4 Number Tests in the Cogstate battery have been configured to repeat if the test completion and/or performance criteria are not met. TestAttempt indicates the attempted number of the test in the current session. Tests can repeat a maximum of 3 times.
  \item TestCompletionScore -4 Number The completion score for the test [(TotalTrials/required correct responses) x 100; required correct responses is 35 for Detection, 30 for Identification, 80 for One Card Learning and 31 for One Back].
  \item TestPerformanceScore -4 Number The performance score for the test [reflects raw accuracy of performance; (TotalCorrect/TotalResponse) x 100]
  \item TestCompletionPass -4 Text Test completion criteria (Yes = met completion check; No = did not meet completion check).
  \item TestPerformancePass -4 Text Test performance criteria (Yes = met performance check; No = did not meet performance check).
  \item TestIntegrityPass -4 Text Test integrity criteria (Yes = met integrity check; No = did not meet integrity check).
  \item TestDuration -4 Milliseconds The total time (ms) from test start until the end of the test.
  \item ReactionTime -4 Number (log10 milliseconds) Mean of log10 transformed reaction times of correct responses.
  \item RawReactionTime -4 Milliseconds Mean of reaction times of correct responses.
  \item RTVariability -4 Number (log10 milliseconds) Standard deviation of log10 transformed reaction times of correct responses.
  \item RawRTVariability -4 Milliseconds Standard deviation of reaction times of correct responses.
  \item Accuracy -4 Number (arcsine square root proportion correct) Accuracy of performance (arcsine square root proportion correct)
  \item RawAccuracy -4 Number Accuracy of performance (proportion of correct responses).
  \item TotalCorrect -4 Number Number of correct responses.
  \item TotalCorrectExclPant -4 Number Number of correct responses after Pant (post-anticipatory) responses have been excluded.
  \item TotalErrors -4 Number Number of errors.
  \item LegalErrors -4 Number Number of legal errors (errors that are not breaking rules of test).
  \item RuleBreakErrors -4 Number Number of errors that break the rules of the tests.
  \item TotalAnticipatory -4 Number Number of Anticipatory responses (i.e., responding before the card turns face-up).
  \item TotalPost -4 Number Number of Post responses (i.e., responding during the response feedback period and before the beginning of the next trial).
  \item TotalMaxOut -4 Number Number of MaxOut responses (i.e., when a response is not provided within the maximum trial duration).
  \item TotalResponses -4 Number Number of responses provided.
  \item TotalTrials -4 Number Number of trials completed.
  \item StandardScoreZ -4 Number Standardized z-score for the primary outcome measure for each test. For the One Back test, Accuracy is used for this outcome.
  \item StandardScoreT -4 Number Standardized T-score (mean = 100, SD = 10) for the primary outcome measure for each test. For the One Back test, Accuracy is used for this outcome.
  \item AltStandardScoreZ -4 Number Standardized z-score for One Back using Reaction Time as the outcome measure.
  \item AltStandardScoreT -4 Number Standardized T-score (mean = 100, SD = 10) for One Back using Reaction Time as the outcome measure.
  \item Psyattstdscr -4 Number The score on the Psychomotor Function/Attention composite, incorporating ReactionTime as the outcome measure for the Detection and Identification tests, relative to age-matched normative data.
  \item LearnWMStdScr -4 Number The score on the Learning/Working Memory composite, incorporating Accuracy as the outcome measures for the One Card Learning and One Back tests, relative to age-matched normative data.
  \item AltLearnWMStdScr -4 Number The score on the Learning/Working Memory composite, incorporating Accuracy as the outcome measure for the One Card Learning test and ReactionTime as the outcome measure for the One Back test, relative to age-matched normative data.
}
}
\examples{
\dontrun{
describe(cbb_results)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/CBB_RESULTS/ADNI_Cogstate_Data_Dictionary.pdf}
}
\keyword{datasets}
