% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{cbbcomp}
\alias{cbbcomp}
\title{Cogstate Brief Battery}
\format{
A data frame with 2807 rows and 13 variables
}
\usage{
data(cbbcomp)
}
\description{
Cogstate Brief Battery
}
\details{
A dataset with variables as follows:

\itemize{
  \item ONLINEID T NA Online ID
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
  \item RECNO N  
  \item CBBLOC N  Administration Type:
  \item DONE N  Was assessment/procedure done?
  \item NDREASON N  If No, reason not done:
  \item EXAMDATE D MM/DD/YYYY Examination Date
}
}
\examples{
\dontrun{
describe(cbbcomp)
}
}
\keyword{datasets}
