% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{cci}
\alias{cci}
\title{Cognitive Change Index}
\format{
A data frame with 1224 rows and 30 variables
}
\usage{
data(cci)
}
\description{
Cognitive Change Index
}
\details{
A dataset with variables as follows:

\itemize{
  \item DATE D NA Date collected
  \item SOURCE N NA Information Source
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
  \item CCI1 N  1. Recalling information when I really try
  \item CCI2 N  2. Remembering names and faces of new people I meet
  \item CCI3 N  3. Remembering things that have happened recently
  \item CCI4 N  4. Recalling conversations a few days later
  \item CCI5 N  5. Remembering where things are usually kept
  \item CCI6 N  6. Remembering new information told to me
  \item CCI7 N  7. Remembering where I placed familiar objects
  \item CCI8 N  8. Remembering what I intended to do
  \item CCI9 N  9. Remembering names of family members and friends
  \item CCI10 N  10. Remembering without notes and reminders
  \item CCI11 N  11. People who know me would find that my memory is
  \item CCI12 N  12. Remembering things compared to my age group
  \item CCI13 N  13. Making decisions about everyday matters
  \item CCI14 N  14. Reasoning through a complicated problem
  \item CCI15 N  15. Focusing on goals and carrying out a plan
  \item CCI16 N  16. Shifting easily from one activity to the next
  \item CCI17 N  17. Organizing my daily activities
  \item CCI18 N  18. Understanding conversations
  \item CCI19 N  19. Expressing myself when speaking
  \item CCI20 N  20. Following a story in a book, movie or TV
  \item CCI12TOT N  CCI Index:  Sum of first 12 items
}
}
\examples{
\dontrun{
describe(cci)
}
}
\keyword{datasets}
