% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{cdr}
\alias{cdr}
\title{Clinical Dementia Rating}
\format{
A data frame with 13050 rows and 20 variables
}
\usage{
data(cdr)
}
\description{
Clinical Dementia Rating
}
\details{
A dataset with variables as follows:

\itemize{
  \item CDDATE D NA Date collected
  \item SPID N NA Study Partner ID:
  \item CDSOB N NA Sum of Boxes
  \item CDRSB N  CDR-SB
  \item CDVERSION N  Version
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item CDSOURCE N  Information Source
  \item CDMEMORY N  Score <!--Memory-->
  \item CDORIENT N  Score <!--Orientation-->
  \item CDJUDGE N  Score <!--Judgment and Problem Solving-->
  \item CDCOMMUN N  Score <!--Community Affairs-->
  \item CDHOME N  Score <!--Home and Hobbies-->
  \item CDCARE N  Score <!--Personal Care-->
  \item CDGLOBAL N  Global CDR
  \item USERDATE2 S  Date record last updated
}
}
\examples{

cdr$CDRSB = apply(cdr[, c('CDMEMORY', 'CDORIENT', 'CDJUDGE', 'CDCOMMUN', 'CDHOME', 'CDCARE')], 1, sum)

}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
