% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{cellimort}
\alias{cellimort}
\title{Cells For Immortalization Speciman Collection}
\format{
A data frame with 990 rows and 14 variables
}
\usage{
data(cellimort)
}
\description{
Cells For Immortalization Speciman Collection
}
\details{
A dataset with variables as follows:

\itemize{
  \item CLCOLL N NA Was sample collected?
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item CLTIME T  Time of Blood Draw
  \item CLFEDEXDT D  Date Fedexed
  \item CLFEDEX N  Fedex Tracking Number
  \item CLVOLUME N  Volume of Blood Shipped
  \item CLLICENSE N  6 digit License Plate Number
  \item USERDATE2 S  Date record last updated
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
