% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{clielg}
\alias{clielg}
\title{Clinician Verification}
\format{
A data frame with 4754 rows and 22 variables
}
\usage{
data(clielg)
}
\description{
Clinician Verification
}
\details{
A dataset with variables as follows:

\itemize{
  \item PIN T NA I certify that I am using my own personal username and pin and understand that by using my username and pin, I am authorizing this electronic signature to be the legally binding equivalent of my handwritten signature executed on paper. - Enter Your PIN -
  \item DIAGNOSIS N  Specify diagnostic category:
  \item RID N NA Participant roster ID
  \item SITEID N NA Site ID
  \item VISCODE T NA Visit code
  \item USERDATE S NA Date record created
  \item USERDATE2 S NA Date record last updated
  \item INCLUSION N NA Does the participant fail to meet any of the inclusion criteria for new EMCIs? <a target="_blank" href="/docs/studydocs/Procedures Manual/ADNIGO Inclusion, Exclusion Criteria_current.pdf"> link to protocol inclusion/exclusion criteria</a>
  \item FAILINCLU T NA If yes, which criteria does the participant fail to meet:
  \item EXCLUSION N NA Does the participant meet any exclusion criteria for new EMCIs? <a target="_blank" href="/docs/studydocs/Procedures Manual/ADNIGO Inclusion, Exclusion Criteria_current.pdf"> link to protocol inclusion/exclusion criteria</a>
  \item FAILEXCLU T NA If yes, which criteria does the participant meet:
  \item CENROLL N NA Based on my review:
  \item CRAND N NA Does the participant continue to meet all the screening criteria?
  \item CDATE D NA Date of Approval or Denial
  \item CPASSWD T NA I certify that I am using my own personal username and password and understand that by using my username and password, I am authorizing this electronic signature to be the legally binding equivalent of my handwritten signature executed on paper. Enter Your Password:
  \item CEMRI N NA Does Screening MRI have evidence of infection, infarction, other focal lesions, or multiple lacunes in critical memory structures?
  \item CONTINCLU N NA Does the participant fail to meet any of the inclusion criteria for ADNI1 continuing participants? <a target="_blank" href="/docs/studydocs/Procedures Manual/ADNIGO Inclusion, Exclusion Criteria_current.pdf"> link to protocol inclusion/exclusion criteria</a>
  \item CONTFAILI T NA If yes, which criteria does the participant fail to meet:
  \item CONTEXCLU N NA Does the participant meet any of the exclusion criteria for ADNI1 continuing participants? <a target="_blank" href="/docs/studydocs/Procedures Manual/ADNIGO Inclusion, Exclusion Criteria_current.pdf"> link to protocol inclusion/exclusion criteria</a>
  \item CONTFAILE T NA If yes, which criteria does the participant meet:
}
}
\keyword{datasets}
