% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{clirev}
\alias{clirev}
\title{Clinician Review}
\format{
A data frame with 7213 rows and 10 variables
}
\usage{
data(clirev)
}
\description{
Clinician Review
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N NA Participant roster ID
  \item SITEID N NA Site ID
  \item VISCODE T NA Visit code
  \item USERDATE S NA Date record created
  \item USERDATE2 S NA Date record last updated
  \item CLIPASS N NA If the visit is complete, please indicate below.
  \item CLIDATE D NA Approval Date
  \item CLIPASSWD T NA I certify that I am using my own personal username and password and understand that by using my username and password, I am authorizing this electronic signature to be the legally binding equivalent of my handwritten signature executed on paper. Enter Your Password:
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
