% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{consents}
\alias{consents}
\title{Consents}
\format{
A data frame with 7729 rows and 80 variables
}
\usage{
data(consents)
}
\description{
Consents
}
\details{
A dataset with variables as follows:

\itemize{
  \item PROTVERSION N NA 2.  Indicate Version of Protocol
  \item CONSENTVERSION N NA 3.  Indicate Coordinating Center Consent template version signed:
  \item AMY N NA 3d.  Consent to Amyloid PET scan procedures
  \item AMYSAME N NA Separate consent document used from that reported in 3a (overall study participation)?
  \item AMYIRB D YYYY-MM-DD Consent IRB approval date:
  \item AMYVRS T NA Site version of consent:
  \item AMYDET T NA Reason for nonconsent:
  \item TAU N NA 3e. Consent to Tau PET scan procedures
  \item TAUSAME N NA Separate consent document used from that reported in 3a (overall study participation)?
  \item TAUIRB D YYYY-MM-DD Consent IRB approval date:
  \item TAUVRS T NA Site version of consent:
  \item TAUDET T NA Reason for nonconsent:
  \item FUTURE N NA 3g. Consent to contact for future studies
  \item FUTPROCDET T NA Reason for nonconsent:
  \item FITBIR N NA 3h. Consent to allow Tau PET scans to be uploaded to FITBIR
  \item FITPROCDET T NA Reason for nonconsent:
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
  \item RECNO N  
  \item INCDATE D  1.  Date consent(s) signed:
  \item INCVERSION N  2.  Indicate ADCS template version signed:
  \item OVERALL N  3a. Overall study participation
  \item OVERALLDET T  Reason for nonconsent:  <!--Overall study participation-->
  \item MRI N  3b. MRI imaging
  \item MRIDET T  Reason for nonconsent:
  \item FDG N  3c. FDG imaging
  \item FDGDET T  Reason for nonconsent:
  \item AV45 N  3d. Florbetapir F18 PET imaging standard sequence
  \item AV45DET T  Reason for nonconsent:
  \item LPPROC N  3f. LP procedure
  \item LPPROCDET T  Reason for nonconsent:
  \item CLINSS N  3g. store and share samples/clinical data
  \item CLINSSDET T  Reason for nonconsent:
  \item DNACOLL N  3h. DNA sample collection
  \item DNACOLLDET T  Reason for nonconsent:
  \item DNASS N  3i. store and share DNA
  \item DNASSDET T  Reason for nonconsent:
  \item RNACOLL N  3j. RNA sample collection
  \item RNACOLLDET T  Reason for nonconsent:
  \item RNASS N  3k. store and share RNA
  \item RNASSDET T  Reason for nonconsent:
  \item OVERALLIRB D  Consent IRB approval date: <!--Overall study participation-->
  \item OVERALLVRS T  Site version of consent:  <!--Overall study participation-->
  \item MRISAME T  Check if:
  \item MRIIRB D  Consent IRB approval date:
  \item MRIVRS T  Site version of consent:
  \item FDGSAME T  Check if:
  \item FDGIRB D  Consent IRB approval date:
  \item FDGVRS T  Site version of consent:
  \item AV45SAME T  Check if:
  \item AV45IRB D  Consent IRB approval date:
  \item AV45VRS T  Site version of consent:
  \item LPSAME T  Check if:
  \item LPIRB D  Consent IRB approval date:
  \item LPVRS T  Site version of consent:
  \item CLINSAME T  Check if:
  \item CLINIRB D  Consent IRB approval date:
  \item CLINVRS T  Site version of consent:
  \item DNACSAME T  Check if:
  \item DNACIRB D  Consent IRB approval date:
  \item DNACVRS T  Site version of consent:
  \item DNASSAME T  Check if:
  \item DNASIRB D  Consent IRB approval date:
  \item DNASVRS T  Site version of consent:
  \item RNACSAME T  Check if:
  \item RNACIRB D  Consent IRB approval date:
  \item RNACVRS T  Site version of consent:
  \item RNASSAME T  Check if:
  \item RNASIRB D  Consent IRB approval date:
  \item RNASVRS T  Site version of consent:
  \item FBPAO N  3e. Florbetapir F18 PET imaging early add-on sequence
  \item FBPAOSAME T  Check if:
  \item FBPAOIRB D  Consent IRB approval date:
  \item FBPAOVRS T  Site version of consent:
  \item FBPAODET T  Reason for nonconsent:
}
}
\keyword{datasets}
