% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{conversions}
\alias{conversions}
\title{Conversions Confirmed by Conversion Committee}
\format{
A data frame with 19 rows and 5 variables
}
\usage{
data(conversions)
}
\description{
Conversions Confirmed by Conversion Committee
}
\details{
A dataset with variables as follows:

\itemize{
  \item USERDATE D -4 Date record created
  \item RID N -4 Participant roster ID
  \item VISCODE T -4 Visit code
  \item TYPE T -4 Conversion/Reversion Type
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
