% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{crossval}
\alias{crossval}
\title{Cross-Validation}
\format{
A data frame with 822 rows and 5 variables
}
\usage{
data(crossval)
}
\description{
Leave k-out cross-vaiidation, stratified by diagnosis, study arm, and age, will be used by the groups conducting voxel-based analyses. A document describing the details of the implementation of the scheme and a table of the randomized assignments can be found below. The table contains a column of participant identifiers (RID), strata (GROUP), randomized set assignements (SET), and a training set indicator (TRAINING).
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item GROUP -4 -4 Strata
  \item SET -4 -4 Randomized set assignments
  \item TRAINING -4 -4 Training set indicator
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/CROSSVAL/cross-validation.pdf}
}
\keyword{datasets}
