% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{csfalphasyn}
\alias{csfalphasyn}
\title{CSF Alpha-synuclein, Jing Zhang Lab, University of Washington}
\format{
A data frame with 390 rows and 9 variables
}
\usage{
data(csfalphasyn)
}
\description{
CSF Alpha-synuclein, Jing Zhang Lab, University of Washington
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N -4 Participant roster ID
  \item EXAMDATE D -4 Examination Date
  \item Protocol N -4 Protocol
  \item Box N -4 Box
  \item Luminex.batch.Number N -4 Luminex batch Number
  \item Luminex.testing.Date D -4 Luminex testing Date
  \item alpha.syn N ng/ml alpha syn
  \item Hemoglobin N ng/ml Hemoglobin
}
}
\examples{
\dontrun{
describe(csfalphasyn)
}
}
\keyword{datasets}
