% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{csfc3fh}
\alias{csfc3fh}
\title{CSF Complement 3 & Factor H, Jing Zhang Lab, University of Washington}
\format{
A data frame with 390 rows and 5 variables
}
\usage{
data(csfc3fh)
}
\description{
CSF Complement 3 & Factor H, Jing Zhang Lab, University of Washington
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item EXAMDATE -4 -4 Examination Date
  \item FH -4 ng/ml factor H
  \item C3 -4 ng/ml complement 3
}
}
\examples{
\dontrun{
describe(csfc3fh)
}
}
\keyword{datasets}
