% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{csfhc}
\alias{csfhc}
\title{Baseline CSF homocysteine data.  ADNI Biomarker Core laboratory.}
\format{
A data frame with 416 rows and 7 variables
}
\usage{
data(csfhc)
}
\description{
Baseline CSF homocysteine data.  ADNI Biomarker Core laboratory.
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N -4 Participant roster ID
  \item SITEID N -4 Site ID
  \item VISCODE T -4 Visit code
  \item EXAMDATE D -4 Examination Date
  \item USERDATE D -4 Date record created
  \item HC N micromoles/L CSF Homocysteine
}
}
\keyword{datasets}
