% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{csfmeth}
\alias{csfmeth}
\title{Method of CSF Collection}
\format{
A data frame with 1165 rows and 15 variables
}
\usage{
data(csfmeth)
}
\description{
Method of CSF Collection
}
\details{
A dataset with variables as follows:

\itemize{
  \item INTERSPACE N NA LP was performed at the:
  \item PTPOSITION N NA Patient Position:
  \item DONE N NA Was CSF collected?
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item NEEDLESIZE N  Needle used:
  \item COLTUBETYP N  Type of collection tube used
  \item SHPTUBETYP N  Type of tube used for shipping
  \item TUBEMIN N  If collected in polystyrene and shipped in polypropylene, please provide estimated amount of time CSF remained in collection tube
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
