% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{csfmrm}
\alias{csfmrm}
\title{Biomarkers Consortium CSF Proteomics MRM data set}
\format{
A data frame with 304 rows and 323 variables
}
\usage{
data(csfmrm)
}
\description{
Biomarkers Consortium CSF Proteomics MRM data set
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item A1AT.AVLTIDEK -4 -4 Alpha-1-antitrypsin 
  \item A1AT.LSITGTYDLK -4 -4 Alpha-1-antitrypsin 
  \item A1AT.SVLGQLGITK -4 -4 Alpha-1-antitrypsin 
  \item A1BG.NGVAQEPVHLDSPAIK -4 -4 Alpha-1B-glycoprotein 
  \item A1BG.SGLSTGWTQLSK -4 -4 Alpha-1B-glycoprotein 
  \item A2GL.DLLLPQPDLR -4 -4 Leucine-rich alpha-2-glycoprotein 
  \item A2GL.VAAGAFQGLR -4 -4 Leucine-rich alpha-2-glycoprotein 
  \item A4.LVFFAEDVGSNK -4 -4 Amyloid beta A4 protein 
  \item A4.THPHFVIPYR -4 -4 Amyloid beta A4 protein 
  \item A4.WYFDVTEGK -4 -4 Amyloid beta A4 protein 
  \item AACT.ADLSGITGAR -4 -4 Alpha-1-antichymotrypsin 
  \item AACT.EIGELYLPK -4 -4 Alpha-1-antichymotrypsin 
  \item AACT.NLAVSQVVHK -4 -4 Alpha-1-antichymotrypsin 
  \item AATC.IVASTLSNPELFEEWTGNVK -4 -4 Aspartate aminotransferase, cytoplasmic 
  \item AATC.LALGDDSPALK -4 -4 Aspartate aminotransferase, cytoplasmic 
  \item AATC.NLDYVATSIHEAVTK -4 -4 Aspartate aminotransferase, cytoplasmic 
  \item AATM.FVTVQTISGTGALR -4 -4 Aspartate aminotransferase, mitochondrial 
  \item AFAM.DADPDTFFAK -4 -4 Afamin 
  \item AFAM.FLVNLVK -4 -4 Afamin 
  \item AFAM.LPNNVLQEK -4 -4 Afamin 
  \item ALDOA.ALQASALK -4 -4 Fructose-bisphosphate aldolase A 
  \item ALDOA.QLLLTADDR -4 -4 Fructose-bisphosphate aldolase A 
  \item AMBP.AFIQLWAFDAVK -4 -4 Protein AMBP 
  \item AMBP.ETLLQDFR -4 -4 Protein AMBP 
  \item AMBP.FLYHK -4 -4 Protein AMBP 
  \item AMD.IPVDEEAFVIDFKPR -4 -4 Peptidyl-glycine alpha-amidating monooxygenase 
  \item AMD.IVQFSPSGK -4 -4 Peptidyl-glycine alpha-amidating monooxygenase 
  \item AMD.NGQWTLIGR -4 -4 Peptidyl-glycine alpha-amidating monooxygenase 
  \item APLP2.HYQHVLAVDPEK -4 -4 Amyloid-like protein 2 
  \item APLP2.WYFDLSK -4 -4 Amyloid-like protein 2 
  \item APOB.IAELSATAQEIIK -4 -4 Apolipoprotein B-100 
  \item APOB.SVSLPSLDPASAK -4 -4 Apolipoprotein B-100 
  \item APOB.TGISPLALIK -4 -4 Apolipoprotein B-100 
  \item APOD.VLNQELR -4 -4 Apolipoprotein D 
  \item APOE.AATVGSLAGQPLQER -4 -4 Apolipoprotein E 
  \item APOE.CLAVYQAGAR -4 -4 Apolipoprotein E 
  \item APOE.LAVYQAGAR -4 -4 Apolipoprotein E 
  \item APOE.LGADMEDVR -4 -4 Apolipoprotein E 
  \item APOE.LGPLVEQGR -4 -4 Apolipoprotein E 
  \item B2MG.VEHSDLSFSK -4 -4 Beta-2-microglobulin 
  \item B2MG.VNHVTLSQPK -4 -4 Beta-2-microglobulin 
  \item B3GN1.EPGEFALLR -4 -4 N-acetyllactosaminide beta-1,3-N-acetylglucosaminyltransferase 
  \item B3GN1.TALASGGVLDASGDYR -4 -4 N-acetyllactosaminide beta-1,3-N-acetylglucosaminyltransferase 
  \item B3GN1.YEAAVPDPR -4 -4 N-acetyllactosaminide beta-1,3-N-acetylglucosaminyltransferase 
  \item BACE1.SIVDSGTTNLR -4 -4 Beta-secretase 1 
  \item BASP1.ETPAATEAPSSTPK -4 -4 Brain acid soluble protein 1 
  \item BTD.LSSGLVTAALYGR -4 -4 Biotinidase 
  \item BTD.SHLIIAQVAK -4 -4 Biotinidase 
  \item C1QB.LEQGENVFLQATDK -4 -4 Complement C1q subcomponent subunit B 
  \item C1QB.VPGLYYFTYHASSR -4 -4 Complement C1q subcomponent subunit B 
  \item CA2D1.FVVTDGGITR -4 -4 Voltage-dependent calcium channel subunit alpha-2/delta-1 
  \item CA2D1.IKPVFIEDANFGR -4 -4 Voltage-dependent calcium channel subunit alpha-2/delta-1 
  \item CA2D1.TASGVNQLVDIYEK -4 -4 Voltage-dependent calcium channel subunit alpha-2/delta-1 
  \item CAD13.DIQGSLQDIFK -4 -4 Cadherin-13 
  \item CAD13.INENTGSVSVTR -4 -4 Cadherin-13 
  \item CAD13.YEVSSPYFK -4 -4 Cadherin-13 
  \item CADM3.EGSVPPLK -4 -4 Cell adhesion molecule 3 
  \item CADM3.GNPVPQQYLWEK -4 -4 Cell adhesion molecule 3 
  \item CADM3.SLVTVLGIPQKPIITGYK -4 -4 Cell adhesion molecule 3 
  \item CAH1.VLDALQAIK -4 -4 Carbonic anhydrase 1 
  \item CAH1.YSSLAEAASK -4 -4 Carbonic anhydrase 1 
  \item CATA.LFAYPDTHR -4 -4 Catalase 
  \item CATD.LVDQNIFSFYLSR -4 -4 Cathepsin D 
  \item CATD.VSTLPAITLK -4 -4 Cathepsin D 
  \item CATD.YSQAVPAVTEGPIPEVLK -4 -4 Cathepsin D 
  \item CATL1.VFQEPLFYEAPR -4 -4 Cathepsin L1 
  \item CCKN.AHLGALLAR -4 -4 Cholecystokinin 
  \item CCKN.NLQNLDPSHR -4 -4 Cholecystokinin 
  \item CD14.AFPALTSLDLSDNPGLGER -4 -4 Monocyte differentiation antigen CD14 
  \item CD14.FPAIQNLALR -4 -4 Monocyte differentiation antigen CD14 
  \item CD14.SWLAELQQWLKPGLK -4 -4 Monocyte differentiation antigen CD14 
  \item CD59.AGLQVYNK -4 -4 CD59 glycoprotein 
  \item CERU.IYHSHIDAPK -4 -4 Ceruloplasmin 
  \item CERU.NNEGTYYSPNYNPQSR -4 -4 Ceruloplasmin 
  \item CFAB.DAQYAPGYDK -4 -4 Complement factor B 
  \item CFAB.VSEADSSNADWVTK -4 -4 Complement factor B 
  \item CFAB.YGLVTYATYPK -4 -4 Complement factor B 
  \item CH3L1.ILGQQVPYATK -4 -4 Chitinase-3-like protein 1 
  \item CH3L1.SFTLASSETGVGAPISGPGIPGR -4 -4 Chitinase-3-like protein 1 
  \item CH3L1.VTIDSSYDIAK -4 -4 Chitinase-3-like protein 1 
  \item CLUS.IDSLLENDR -4 -4 Clusterin 
  \item CLUS.SGSGLVGR -4 -4 Clusterin 
  \item CLUS.VTTVASHTSDSDVPSGVTEVVVK -4 -4 Clusterin 
  \item CMGA.EDSLEAGLPLQVR -4 -4 Chromogranin-A 
  \item CMGA.SEALAVDGAGKPGAEEAQDPEGK -4 -4 Chromogranin-A 
  \item CMGA.SGEATDGARPQALPEPMQESK -4 -4 Chromogranin-A 
  \item CMGA.SGELEQEEER -4 -4 Chromogranin-A 
  \item CMGA.YPGPQAEGDSEGLSQGLVDR -4 -4 Chromogranin-A 
  \item CNDP1.ALEQDLPVNIK -4 -4 Beta-Ala-His dipeptidase 
  \item CNDP1.VFQYIDLHQDEFVQTLK -4 -4 Beta-Ala-His dipeptidase 
  \item CNDP1.WNYIEGTK -4 -4 Beta-Ala-His dipeptidase 
  \item CNTN1.DGEYVVEVR -4 -4 Contactin-1 
  \item CNTN1.TTKPYPADIVVQFK -4 -4 Contactin-1 
  \item CNTN2.IIVQAQPEWLK -4 -4 Contactin-2 
  \item CNTN2.TTGPGGDGIPAEVHIVR -4 -4 Contactin-2 
  \item CNTN2.VIASNILGTGEPSGPSSK -4 -4 Contactin-2 
  \item CNTP2.HELQHPIIAR -4 -4 Contactin-associated protein-like 2 
  \item CNTP2.VDNAPDQQNSHPDLAQEEIR -4 -4 Contactin-associated protein-like 2 
  \item CNTP2.YSSSDWVTQYR -4 -4 Contactin-associated protein-like 2 
  \item CO2.DFHINLFR -4 -4 Complement C2 
  \item CO2.HAIILLTDGK -4 -4 Complement C2 
  \item CO2.SSGQWQTPGATR -4 -4 Complement C2 
  \item CO3.IHWESASLLR -4 -4 Complement C3 
  \item CO3.LSINTHPSQKPLSITVR -4 -4 Complement C3 
  \item CO3.TELRPGETLNVNFLLR -4 -4 Complement C3 
  \item CO3.TGLQEVEVK -4 -4 Complement C3 
  \item CO3.VPVAVQGEDTVQSLTQGDGVAK -4 -4 Complement C3 
  \item CO4A.DHAVDLIQK -4 -4 Complement C4-A 
  \item CO4A.GSFEFPVGDAVSK -4 -4 Complement C4-A 
  \item CO4A.LGQYASPTAK -4 -4 Complement C4-A 
  \item CO4A.NVNFQK -4 -4 Complement C4-A 
  \item CO4A.VLSLAQEQVGGSPEK -4 -4 Complement C4-A 
  \item CO4A.VTASDPLDTLGSEGALSPGGVASLLR -4 -4 Complement C4-A 
  \item CO5.DINYVNPVIK -4 -4 Complement C5 
  \item CO5.TLLPVSKPEIR -4 -4 Complement C5 
  \item CO5.VFQFLEK -4 -4 Complement C5 
  \item CO6.ALNHLPLEYNSALYSR -4 -4 Complement component C6 
  \item CO6.SEYGAALAWEK -4 -4 Complement component C6 
  \item CO8B.IPGIFELGISSQSDR -4 -4 Complement component C8 beta chain 
  \item CO8B.SDLEVAHYK -4 -4 Complement component C8 beta chain 
  \item CO8B.YEFILK -4 -4 Complement component C8 beta chain 
  \item COCH.GVISNSGGPVR -4 -4 Cochlin 
  \item CRP.ESDTSYVSLK -4 -4 C-reactive protein 
  \item CSTN1.GNLAGLTLR -4 -4 Calsyntenin-1 
  \item CSTN1.IHGQNVPFDAVVVDK -4 -4 Calsyntenin-1 
  \item CSTN1.IPDGVVSVSPK -4 -4 Calsyntenin-1 
  \item CSTN3.ATGEGLIR -4 -4 Calsyntenin-3 
  \item CSTN3.ESLLLDTTSLQQR -4 -4 Calsyntenin-3 
  \item CUTA.TQSSLVPALTDFVR -4 -4 Protein CutA 
  \item CYTC.ALDFAVGEYNK -4 -4 Cystatin-C 
  \item DAG1.GVHYISVSATR -4 -4 Dystroglycan 
  \item DAG1.LVPVVNNR -4 -4 Dystroglycan 
  \item DAG1.VTIPTDLIASSGDIIK -4 -4 Dystroglycan 
  \item DIAC.ATYIQNYR -4 -4 Di-N-acetylchitobiase 
  \item ENOG.GNPTVEVDLYTAK -4 -4 Gamma-enolase 
  \item ENOG.LGAEVYHTLK -4 -4 Gamma-enolase 
  \item ENPP2.SYPEILTLK -4 -4 Ectonucleotide pyrophosphatase/phosphodiesterase family member 2 
  \item ENPP2.WWGGQPLWITATK -4 -4 Ectonucleotide pyrophosphatase/phosphodiesterase family member 2 
  \item EXTL2.VIVVWNNIGEK -4 -4 Exostosin-like 2 
  \item FABPH.SIVTLDGGK -4 -4 Fatty acid-binding protein, heart 
  \item FABPH.SLGVGFATR -4 -4 Fatty acid-binding protein, heart 
  \item FAM3C.GINVALANGK -4 -4 Protein FAM3C 
  \item FAM3C.SALDTAAR -4 -4 Protein FAM3C 
  \item FAM3C.SPFEQHIK -4 -4 Protein FAM3C 
  \item FAM3C.TGEVLDTK -4 -4 Protein FAM3C 
  \item FBLN1.AITPPHPASQANIIFDITEGNLR -4 -4 Fibulin-1 
  \item FBLN1.IIEVEEEQEDPYLNDR -4 -4 Fibulin-1 
  \item FBLN1.TGYYFDGISR -4 -4 Fibulin-1 
  \item FBLN3.IPSNPSHR -4 -4 EGF-containing fibulin-like extracellular matrix protein 1 
  \item FBLN3.LTIIVGPFSF -4 -4 EGF-containing fibulin-like extracellular matrix protein 1 
  \item FBLN3.SGNENGEFYLR -4 -4 EGF-containing fibulin-like extracellular matrix protein 1 
  \item FETUA.AHYDLR -4 -4 Alpha-2-HS-glycoprotein 
  \item FETUA.FSVVYAK -4 -4 Alpha-2-HS-glycoprotein 
  \item FETUA.HTLNQIDEVK -4 -4 Alpha-2-HS-glycoprotein 
  \item FMOD.YLPFVPSR -4 -4 Fibromodulin 
  \item GFAP.ALAAELNQLR -4 -4 Glial fibrillary acidic protein 
  \item GOLM1.DQLVIPDGQEEEQEAAGEGR -4 -4 Golgi membrane protein 1 
  \item GOLM1.QQLQALSEPQPR -4 -4 Golgi membrane protein 1 
  \item GRIA4.EYPGSETPPK -4 -4 Glutamate receptor 4 
  \item GRIA4.LQNILEQIVSVGK -4 -4 Glutamate receptor 4 
  \item GRIA4.NTDQEYTAFR -4 -4 Glutamate receptor 4 
  \item HBA.FLASVSTVLTSK -4 -4 Hemoglobin subunit alpha 
  \item HBA.TYFPHFDLSHGSAQVK -4 -4 Hemoglobin subunit alpha 
  \item HBA.VGAHAGEYGAEALER -4 -4 Hemoglobin subunit alpha 
  \item HBB.EFTPPVQAAYQK -4 -4 Hemoglobin subunit beta 
  \item HBB.SAVTALWGK -4 -4 Hemoglobin subunit beta 
  \item HBB.VNVDEVGGEALGR -4 -4 Hemoglobin subunit beta 
  \item HEMO.NFPSPVDAAFR -4 -4 Hemopexin 
  \item HEMO.QGHNSVFLIK -4 -4 Hemopexin 
  \item HEMO.SGAQATWTELPWPHEK -4 -4 Hemopexin 
  \item I18BP.LWEGSTSR -4 -4 Interleukin-18-binding protein 
  \item IBP2.HGLYNLK -4 -4 Insulin-like growth factor-binding protein 2 
  \item IBP2.LIQGAPTIR -4 -4 Insulin-like growth factor-binding protein 2 
  \item IGSF8.DTQFSYAVFK -4 -4 Immunoglobulin superfamily member 8 
  \item IGSF8.LQGDAVVLK -4 -4 Immunoglobulin superfamily member 8 
  \item IGSF8.VVAGEVQVQR -4 -4 Immunoglobulin superfamily member 8 
  \item ITIH1.EVAFDLEIPK -4 -4 Inter-alpha-trypsin inhibitor heavy chain H1 
  \item ITIH1.QYYEGSEIVVAGR -4 -4 Inter-alpha-trypsin inhibitor heavy chain H1 
  \item ITIH5.SYLEITPSR -4 -4 Inter-alpha-trypsin inhibitor heavy chain H5 
  \item KAIN.FYYLIASETPGK -4 -4 Kallistatin 
  \item KAIN.LGFTDLFSK -4 -4 Kallistatin 
  \item KAIN.VGSALFLSHNLK -4 -4 Kallistatin 
  \item KAIN.WADLSGITK -4 -4 Kallistatin 
  \item KLK10.ALQLPYR -4 -4 Kallikrein-10 
  \item KLK11.LPHTLR -4 -4 Kallikrein-11 
  \item KLK6.ESSQEQSSVVR -4 -4 Kallikrein-6 
  \item KLK6.LSELIQPLPLER -4 -4 Kallikrein-6 
  \item KLK6.YTNWIQK -4 -4 Kallikrein-6 
  \item KNG1.DIPTNSPELEETLTHTITK -4 -4 Kininogen-1 
  \item KNG1.QVVAGLNFR -4 -4 Kininogen-1 
  \item KNG1.TVGSDTFYSFK -4 -4 Kininogen-1 
  \item KPYM.LDIDSPPITAR -4 -4 Pyruvate kinase isozymes M1/M2 
  \item L1CAM.AQLLVVGSPGPVPR -4 -4 Neural cell adhesion molecule L1 
  \item L1CAM.LVLSDLHLLTQSQVR -4 -4 Neural cell adhesion molecule L1 
  \item L1CAM.WRPVDLAQVK -4 -4 Neural cell adhesion molecule L1 
  \item LAMB2.AQGIAQGAIR -4 -4 Laminin subunit beta-2 
  \item LPHN1.LVVSQLNPYTLR -4 -4 Latrophilin-1 
  \item LPHN1.SGETVINTANYHDTSPYR -4 -4 Latrophilin-1 
  \item LRC4B.HLEILQLSK -4 -4 Leucine-rich repeat-containing protein 4B 
  \item LRC4B.LTTVPTQAFEYLSK -4 -4 Leucine-rich repeat-containing protein 4B 
  \item LTBP2.EQDAPVAGLQPVER -4 -4 Latent-transforming growth factor beta-binding protein 2 
  \item MIME.ESAYLYAR -4 -4 Mimecan 
  \item MIME.ETVIIPNEK -4 -4 Mimecan 
  \item MIME.LEGNPIVLGK -4 -4 Mimecan 
  \item MOG.VVHLYR -4 -4 Myelin-oligodendrocyte glycoprotein 
  \item MUC18.EVTVPVFYPTEK -4 -4 Cell surface glycoprotein MUC18 
  \item MUC18.GATLALTQVTPQDER -4 -4 Cell surface glycoprotein MUC18 
  \item NBL1.LALFPDK -4 -4 Neuroblastoma suppressor of tumorigenicity 1 
  \item NCAM1.AGEQDATIHLK -4 -4 Neural cell adhesion molecule 1 
  \item NCAM1.GLGEISAASEFK -4 -4 Neural cell adhesion molecule 1 
  \item NCAM2.ASGSPEPAISWFR -4 -4 Neural cell adhesion molecule 2 
  \item NCAM2.IIELSQTTAK -4 -4 Neural cell adhesion molecule 2 
  \item NCAN.APVLELEK -4 -4 Neurocan core protein 
  \item NCAN.LSSAIIAAPR -4 -4 Neurocan core protein 
  \item NEGR1.SSIIFAGGDK -4 -4 Neuronal growth regulator 1 
  \item NEGR1.VVVNFAPTIQEIK -4 -4 Neuronal growth regulator 1 
  \item NEGR1.WSVDPR -4 -4 Neuronal growth regulator 1 
  \item NELL2.AFLFQDTPR -4 -4 Protein kinase C-binding protein NELL2 
  \item NELL2.FTGSSWIK -4 -4 Protein kinase C-binding protein NELL2 
  \item NELL2.SALAYVDGK -4 -4 Protein kinase C-binding protein NELL2 
  \item NEO1.DVVASLVSTR -4 -4 Neogenin 
  \item NEUS.ALGITEIFIK -4 -4 Neuroserpin 
  \item NEUS.QEVPLATLEPLVK -4 -4 Neuroserpin 
  \item NGF.SAPAAAIAAR -4 -4 Beta-nerve growth factor 
  \item NICA.ALADVATVLGR -4 -4 Nicastrin 
  \item NICA.APDVTTLPR -4 -4 Nicastrin 
  \item NPTX1.FQLTFPLR -4 -4 Neuronal pentraxin-1 
  \item NPTX1.LENLEQYSR -4 -4 Neuronal pentraxin-1 
  \item NPTX2.LESLEHQLR -4 -4 Neuronal pentraxin-2 
  \item NPTX2.TESTLNALLQR -4 -4 Neuronal pentraxin-2 
  \item NPTXR.ELDVLQGR -4 -4 Neuronal pentraxin receptor 
  \item NPTXR.LVEAFGGATK -4 -4 Neuronal pentraxin receptor 
  \item NRCAM.SLPSEASEQYLTK -4 -4 Neuronal cell adhesion molecule 
  \item NRCAM.VFNTPEGVPSAPSSLK -4 -4 Neuronal cell adhesion molecule 
  \item NRCAM.YIVSGTPTFVPYLIK -4 -4 Neuronal cell adhesion molecule 
  \item NRX1A.DLFIDGQSK -4 -4 Neurexin-1 
  \item NRX1A.ITTQITAGAR -4 -4 Neurexin-1 
  \item NRX1A.SDLYIGGVAK -4 -4 Neurexin-1 
  \item NRX2A.AIVADPVTFK -4 -4 Neurexin-2 
  \item NRX2A.LGERPPALLGSQGLR -4 -4 Neurexin-2 
  \item NRX2A.LSALTLSTVK -4 -4 Neurexin-2 
  \item NRX3A.IYGEVVFK -4 -4 Neurexin-3 
  \item NRX3A.SDLSFQFK -4 -4 Neurexin-3 
  \item OSTP.AIPVAQDLNAPSDWDSR -4 -4 Osteopontin 
  \item PCSK1.ALAHLLEAER -4 -4 ProSAAS 
  \item PCSK1.GEAAGAVQELAR -4 -4 ProSAAS 
  \item PCSK1.NSDPALGLDDDPDAPAAQLAR -4 -4 ProSAAS 
  \item PDYN.FLPSISTK -4 -4 Proenkephalin-B 
  \item PDYN.LSGSFLK -4 -4 Proenkephalin-B 
  \item PDYN.SVGEGPYSELAK -4 -4 Proenkephalin-B 
  \item PEDF.DTDTGALLFIGK -4 -4 Pigment epithelium-derived factor 
  \item PEDF.SSFVAPLEK -4 -4 Pigment epithelium-derived factor 
  \item PEDF.TVQAVLTVPK -4 -4 Pigment epithelium-derived factor 
  \item PGRP2.AGLLRPDYALLGHR -4 -4 N-acetylmuramoyl-L-alanine amidase 
  \item PGRP2.TFTLLDPK -4 -4 N-acetylmuramoyl-L-alanine amidase 
  \item PIMT.VQLVVGDGR -4 -4 Protein-L-isoaspartate(D-aspartate) O-methyltransferase 
  \item PLDX1.LYGPSEPHSR -4 -4 Plexin domain-containing protein 1 
  \item PLMN.EAQLPVIENK -4 -4 Plasminogen 
  \item PLMN.HSIFTPETNPR -4 -4 Plasminogen 
  \item PLMN.LSSPAVITDK -4 -4 Plasminogen 
  \item PPN.VHQSPDGTLLIYNLR -4 -4 Papilin 
  \item PRDX1.DISLSDYK -4 -4 Peroxiredoxin-1 
  \item PRDX1.LVQAFQFTDK -4 -4 Peroxiredoxin-1 
  \item PRDX2.GLFIIDGK -4 -4 Peroxiredoxin-2 
  \item PRDX2.IGKPAPDFK -4 -4 Peroxiredoxin-2 
  \item PRDX3.HLSVNDLPVGR -4 -4 Thioredoxin-dependent peroxide reductase, mitochondrial 
  \item PRDX6.LIALSIDSVEDHLAWSK -4 -4 Peroxiredoxin-6 
  \item PRDX6.LSILYPATTGR -4 -4 Peroxiredoxin-6 
  \item PTGDS.AQGFTEDTIVFLPQTDK -4 -4 Prostaglandin-H2 D-isomerase 
  \item PTGDS.WFSAGLASNSSWLR -4 -4 Prostaglandin-H2 D-isomerase 
  \item PTPRN.AEAPALFSR -4 -4 Receptor-type tyrosine-protein phosphatase-like N 
  \item PTPRN.LAAVLAGYGVELR -4 -4 Receptor-type tyrosine-protein phosphatase-like N 
  \item PTPRN.SELEAQTGLQILQTGVGQR -4 -4 Receptor-type tyrosine-protein phosphatase-like N 
  \item PVRL1.ITQVTWQK -4 -4 Poliovirus receptor-related protein 1 
  \item SCG1.GEAGAPGEEDIQGPTK -4 -4 Secretogranin-1 
  \item SCG1.HLEEPGETQNAFLNER -4 -4 Secretogranin-1 
  \item SCG1.NYLNYGEEGAPGK -4 -4 Secretogranin-1 
  \item SCG1.SSQGGSLPSEEK -4 -4 Secretogranin-1 
  \item SCG2.ALEYIENLR -4 -4 Secretogranin-2 
  \item SCG2.IILEALR -4 -4 Secretogranin-2 
  \item SCG2.VLEYLNQEK -4 -4 Secretogranin-2 
  \item SCG3.ELSAERPLNEQIAEAEEDK -4 -4 Secretogranin-3 
  \item SCG3.FQDDPDGLHQLDGTPLTAEDIVHK -4 -4 Secretogranin-3 
  \item SCG3.LNVEDVDSTK -4 -4 Secretogranin-3 
  \item SE6L1.ETGTPIWTSR -4 -4 Seizure 6-like protein 
  \item SE6L1.SPTNTISVYFR -4 -4 Seizure 6-like protein 
  \item SIAE.ELSNTAAYQSVR -4 -4 Sialate O-acetylesterase 
  \item SLIK1.SLPVDVFAGVSLSK -4 -4 SLIT and NTRK-like protein 1 
  \item SODC.GDGPVQGIINFEQK -4 -4 Superoxide dismutase [Cu-Zn] 
  \item SODC.HVGDLGNVTADK -4 -4 Superoxide dismutase [Cu-Zn] 
  \item SODC.TLVVHEK -4 -4 Superoxide dismutase [Cu-Zn] 
  \item SODE.AGLAASLAGPHSIVGR -4 -4 Extracellular superoxide dismutase [Cu-Zn] 
  \item SODE.AVVVHAGEDDLGR -4 -4 Extracellular superoxide dismutase [Cu-Zn] 
  \item SODE.VTGVVLFR -4 -4 Extracellular superoxide dismutase [Cu-Zn] 
  \item SORC1.TIAVYEEFR -4 -4 VPS10 domain-containing receptor SorCS1 
  \item SPON1.VTLSAAPPSYFR -4 -4 Spondin-1 
  \item SPRL1.HIQETEWQSQEGK -4 -4 SPARC-like protein 1 
  \item SPRL1.HSASDDYFIPSQAFLEAER -4 -4 SPARC-like protein 1 
  \item SPRL1.VLTHSELAPLR -4 -4 SPARC-like protein 1 
  \item TGFB1.LLAPSDSPEWLSFDVTGVVR -4 -4 Transforming growth factor beta-1 
  \item THRB.ETAASLLQAGYK -4 -4 Prothrombin 
  \item THRB.YGFYTHVFR -4 -4 Prothrombin 
  \item TIMP1.GFQALGDAADIR -4 -4 Metalloproteinase inhibitor 1 
  \item TIMP1.SEEFLIAGK -4 -4 Metalloproteinase inhibitor 1 
  \item TNR21.ASNLIGTYR -4 -4 Tumor necrosis factor receptor superfamily member 21 
  \item TRFM.ADTDGGLIFR -4 -4 Melanotransferrin 
  \item TTHY.TSESGELHGLTTEEEFVEGIYK -4 -4 Transthyretin 
  \item TTHY.VEIDTK -4 -4 Transthyretin 
  \item UBB.ESTLHLVLR -4 -4 Polyubiquitin-B 
  \item UBB.TITLEVEPSDTIENVK -4 -4 Polyubiquitin-B 
  \item UBB.TLSDYNIQK -4 -4 Polyubiquitin-B 
  \item VASN.NLHDLDVSDNQLER -4 -4 Vasorin 
  \item VASN.SLTLGIEPVSPTSLR -4 -4 Vasorin 
  \item VASN.YLQGSSVQLR -4 -4 Vasorin 
  \item VGF.AYQGVAAPFPK -4 -4 Neurosecretory protein VGF 
  \item VGF.NSEPQDEGELFQGVDPR -4 -4 Neurosecretory protein VGF 
  \item VGF.THLGEALAPLSK -4 -4 Neurosecretory protein VGF 
  \item VTDB.EFSHLGK -4 -4 Vitamin D-binding protein 
  \item VTDB.HLSLLTTLSNR -4 -4 Vitamin D-binding protein 
  \item VTDB.VPTADLEDVLPLAEDITNILSK -4 -4 Vitamin D-binding protein 
}
}
\examples{
\dontrun{
describe(csfmrm)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/CSFMRM/04\%20Apr\%202014\%20Bimarkers\%20Consortium\%20MRM\%20Data\%20Primer.pdf}
}
\keyword{datasets}
