% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{csfproteomics}
\alias{csfproteomics}
\title{FNIH Biomarkers Consortium CSF Proteomics Project}
\format{
A data frame with 750 rows and 523 variables
}
\usage{
data(csfproteomics)
}
\description{
FNIH Biomarkers Consortium CSF Proteomics Project
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4  Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4  Examination Date
  \item VID -4 -4 -4
  \item SAMPLE.NAME -4  Caprion sample ID
  \item INSTRUMENT -4  Caprion MS instrument
  \item RUN.NUMBER -4  Sample Analysis Run Number
  \item SUBRUN -4  Subrun Index to accommodate segmented runs
  \item MS.PLATE.INDEX -4  MS Plate ID
  \item RUN.INJECTION.NUMBER -4  MS Run number
  \item PLATE.COLUMN -4  MS Plate Column
  \item PLATE.ROW -4  MS Plate Row
  \item DEPLETION.BATCH -4  Sample Randomization Depletion Column Number
  \item RANDOMIZATION.BLOCKS -4  Sample Randomization Group for depletion order and plating
  \item PINK -4  Sample color notation
  \item CMGA -4 log2 scale/unitless Normalized quantitative data for protein CGMA
  \item FABPH -4 log2 scale/unitless Normalized quantitative data for protein FABPH
  \item NPTX2 -4 log2 scale/unitless Normalized quantitative data for protein NPTX2
  \item SCG2 -4 log2 scale/unitless Normalized quantitative data for protein SCG2
  \item VGF -4 log2 scale/unitless Normalized quantitative data for protein VGF
  \item A1AT_HUMAN.LSITGTYDLK.Y2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (A1AT_HUMAN-LSITGTYDLK-y2+)
  \item A1AT_HUMAN.LSITGTYDLK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (A1AT_HUMAN-LSITGTYDLK-y7+)
  \item A1AT_HUMAN.SVLGQLGITK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (A1AT_HUMAN-SVLGQLGITK-y7+)
  \item A1AT_HUMAN.SVLGQLGITK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (A1AT_HUMAN-SVLGQLGITK-y8+)
  \item A1BG_HUMAN.NGVAQEPVHLDSPAIK.B3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (A1BG_HUMAN-NGVAQEPVHLDSPAIK-b3+)
  \item A1BG_HUMAN.NGVAQEPVHLDSPAIK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (A1BG_HUMAN-NGVAQEPVHLDSPAIK-y7+)
  \item A1BG_HUMAN.SGLSTGWTQLSK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (A1BG_HUMAN-SGLSTGWTQLSK-y7+)
  \item A1BG_HUMAN.SGLSTGWTQLSK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (A1BG_HUMAN-SGLSTGWTQLSK-y8+)
  \item A2AP_HUMAN.DFLQSLK.B2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (A2AP_HUMAN-DFLQSLK-b2+)
  \item A2AP_HUMAN.DFLQSLK.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (A2AP_HUMAN-DFLQSLK-y5+)
  \item A2AP_HUMAN.LGNQEPGGQTALK.B5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (A2AP_HUMAN-LGNQEPGGQTALK-b5+)
  \item A2AP_HUMAN.LGNQEPGGQTALK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (A2AP_HUMAN-LGNQEPGGQTALK-y8+)
  \item A2GL_HUMAN.DLLLPQPDLR.B3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (A2GL_HUMAN-DLLLPQPDLR-b3+)
  \item A2GL_HUMAN.DLLLPQPDLR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (A2GL_HUMAN-DLLLPQPDLR-y6+)
  \item A2GL_HUMAN.VAAGAFQGLR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (A2GL_HUMAN-VAAGAFQGLR-y7+)
  \item A2GL_HUMAN.VAAGAFQGLR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (A2GL_HUMAN-VAAGAFQGLR-y8+)
  \item A4_HUMAN.LVFFAEDVGSNK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (A4_HUMAN-LVFFAEDVGSNK-y8+)
  \item A4_HUMAN.LVFFAEDVGSNK.Y9. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (A4_HUMAN-LVFFAEDVGSNK-y9+)
  \item A4_HUMAN.THPHFVIPYR.B4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (A4_HUMAN-THPHFVIPYR-b4+)
  \item A4_HUMAN.THPHFVIPYR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (A4_HUMAN-THPHFVIPYR-y4+)
  \item AACT_HUMAN.ADLSGITGAR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (AACT_HUMAN-ADLSGITGAR-y6+)
  \item AACT_HUMAN.ADLSGITGAR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (AACT_HUMAN-ADLSGITGAR-y7+)
  \item AACT_HUMAN.EIGELYLPK.Y2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (AACT_HUMAN-EIGELYLPK-y2+)
  \item AACT_HUMAN.EIGELYLPK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (AACT_HUMAN-EIGELYLPK-y7+)
  \item AACT_HUMAN.NLAVSQVVHK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (AACT_HUMAN-NLAVSQVVHK-y7+)
  \item AACT_HUMAN.NLAVSQVVHK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (AACT_HUMAN-NLAVSQVVHK-y8+)
  \item AATC_HUMAN.IVASTLSNPELFEEWTGNVK.B3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (AATC_HUMAN-IVASTLSNPELFEEWTGNVK-b3+)
  \item AATC_HUMAN.IVASTLSNPELFEEWTGNVK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (AATC_HUMAN-IVASTLSNPELFEEWTGNVK-y6+)
  \item AATC_HUMAN.LALGDDSPALK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (AATC_HUMAN-LALGDDSPALK-y8+)
  \item AATC_HUMAN.LALGDDSPALK.Y9. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (AATC_HUMAN-LALGDDSPALK-y9+)
  \item AFAM_HUMAN.DADPDTFFAK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (AFAM_HUMAN-DADPDTFFAK-y7+)
  \item AFAM_HUMAN.DADPDTFFAK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (AFAM_HUMAN-DADPDTFFAK-y8+)
  \item AFAM_HUMAN.FLVNLVK.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (AFAM_HUMAN-FLVNLVK-y5+)
  \item ALDOA_HUMAN.ADDGRPFPQVIK.B9. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (ALDOA_HUMAN-ADDGRPFPQVIK-b9+)
  \item ALDOA_HUMAN.ADDGRPFPQVIK.Y2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (ALDOA_HUMAN-ADDGRPFPQVIK-y2+)
  \item AMBP_HUMAN.AFIQLWAFDAVK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (AMBP_HUMAN-AFIQLWAFDAVK-y7+)
  \item AMBP_HUMAN.AFIQLWAFDAVK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (AMBP_HUMAN-AFIQLWAFDAVK-y8+)
  \item AMBP_HUMAN.ETLLQDFR.Y2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (AMBP_HUMAN-ETLLQDFR-y2+)
  \item AMBP_HUMAN.ETLLQDFR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (AMBP_HUMAN-ETLLQDFR-y4+)
  \item AMD_HUMAN.IPVDEEAFVIDFKPR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (AMD_HUMAN-IPVDEEAFVIDFKPR-y5+)
  \item AMD_HUMAN.IPVDEEAFVIDFKPR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (AMD_HUMAN-IPVDEEAFVIDFKPR-y6+)
  \item AMD_HUMAN.IVQFSPSGK.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (AMD_HUMAN-IVQFSPSGK-y4+)
  \item AMD_HUMAN.IVQFSPSGK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (AMD_HUMAN-IVQFSPSGK-y7+)
  \item ANT3_HUMAN.DDLYVSDAFHK.B3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (ANT3_HUMAN-DDLYVSDAFHK-b3+)
  \item ANT3_HUMAN.DDLYVSDAFHK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (ANT3_HUMAN-DDLYVSDAFHK-y7+)
  \item ANT3_HUMAN.FATTFYQHLADSK.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (ANT3_HUMAN-FATTFYQHLADSK-y4+)
  \item ANT3_HUMAN.FATTFYQHLADSK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (ANT3_HUMAN-FATTFYQHLADSK-y6+)
  \item APOB_HUMAN.SVSLPSLDPASAK.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (APOB_HUMAN-SVSLPSLDPASAK-y5+)
  \item APOB_HUMAN.SVSLPSLDPASAK.Y9. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (APOB_HUMAN-SVSLPSLDPASAK-y9+)
  \item APOB_HUMAN.TGISPLALIK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (APOB_HUMAN-TGISPLALIK-y6+)
  \item APOB_HUMAN.TGISPLALIK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (APOB_HUMAN-TGISPLALIK-y7+)
  \item APOD_HUMAN.IPTTFENGR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (APOD_HUMAN-IPTTFENGR-y6+)
  \item APOD_HUMAN.IPTTFENGR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (APOD_HUMAN-IPTTFENGR-y7+)
  \item APOD_HUMAN.NPNLPPETVDSLK.B4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (APOD_HUMAN-NPNLPPETVDSLK-b4+)
  \item APOD_HUMAN.NPNLPPETVDSLK.Y9. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (APOD_HUMAN-NPNLPPETVDSLK-y9+)
  \item APOD_HUMAN.VLNQELR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (APOD_HUMAN-VLNQELR-y5+)
  \item APOD_HUMAN.VLNQELR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (APOD_HUMAN-VLNQELR-y6+)
  \item APOE_HUMAN.AATVGSLAGQPLQER.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (APOE_HUMAN-AATVGSLAGQPLQER-y7+)
  \item APOE_HUMAN.AATVGSLAGQPLQER.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (APOE_HUMAN-AATVGSLAGQPLQER-y8+)
  \item APOE_HUMAN.LAVYQAGAR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (APOE_HUMAN-LAVYQAGAR-y5+)
  \item APOE_HUMAN.LAVYQAGAR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (APOE_HUMAN-LAVYQAGAR-y6+)
  \item APOE_HUMAN.LGADMEDVR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (APOE_HUMAN-LGADMEDVR-y5+)
  \item APOE_HUMAN.LGADMEDVR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (APOE_HUMAN-LGADMEDVR-y8+)
  \item APOE_HUMAN.LGPLVEQGR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (APOE_HUMAN-LGPLVEQGR-y5+)
  \item APOE_HUMAN.LGPLVEQGR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (APOE_HUMAN-LGPLVEQGR-y6+)
  \item C1QB_HUMAN.IAFSATR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (C1QB_HUMAN-IAFSATR-y4+)
  \item C1QB_HUMAN.IAFSATR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (C1QB_HUMAN-IAFSATR-y6+)
  \item C1QB_HUMAN.TINVPLR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (C1QB_HUMAN-TINVPLR-y4+)
  \item C1QB_HUMAN.TINVPLR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (C1QB_HUMAN-TINVPLR-y5+)
  \item C1QC_HUMAN.FNAVLTNPQGDYDTSTGK.B2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (C1QC_HUMAN-FNAVLTNPQGDYDTSTGK-b2+)
  \item C1QC_HUMAN.FNAVLTNPQGDYDTSTGK.B3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (C1QC_HUMAN-FNAVLTNPQGDYDTSTGK-b3+)
  \item C1QC_HUMAN.FQSVFTVTR.B2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (C1QC_HUMAN-FQSVFTVTR-b2+)
  \item C1QC_HUMAN.FQSVFTVTR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (C1QC_HUMAN-FQSVFTVTR-y7+)
  \item CA2D1_HUMAN.FVVTDGGITR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CA2D1_HUMAN-FVVTDGGITR-y7+)
  \item CA2D1_HUMAN.FVVTDGGITR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CA2D1_HUMAN-FVVTDGGITR-y8+)
  \item CA2D1_HUMAN.IKPVFIEDANFGR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CA2D1_HUMAN-IKPVFIEDANFGR-y6+)
  \item CA2D1_HUMAN.IKPVFIEDANFGR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CA2D1_HUMAN-IKPVFIEDANFGR-y7+)
  \item CA2D1_HUMAN.TASGVNQLVDIYEK.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CA2D1_HUMAN-TASGVNQLVDIYEK-y3+)
  \item CA2D1_HUMAN.TASGVNQLVDIYEK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CA2D1_HUMAN-TASGVNQLVDIYEK-y7+)
  \item CAD13_HUMAN.DIQGSLQDIFK.Y2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CAD13_HUMAN-DIQGSLQDIFK-y2+)
  \item CAD13_HUMAN.DIQGSLQDIFK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CAD13_HUMAN-DIQGSLQDIFK-y8+)
  \item CAD13_HUMAN.INENTGSVSVTR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CAD13_HUMAN-INENTGSVSVTR-y4+)
  \item CAD13_HUMAN.INENTGSVSVTR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CAD13_HUMAN-INENTGSVSVTR-y7+)
  \item CAD13_HUMAN.YEVSSPYFK.B2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CAD13_HUMAN-YEVSSPYFK-b2+)
  \item CAD13_HUMAN.YEVSSPYFK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CAD13_HUMAN-YEVSSPYFK-y7+)
  \item CAH1_HUMAN.YSSLAEAASK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CAH1_HUMAN-YSSLAEAASK-y6+)
  \item CAH1_HUMAN.YSSLAEAASK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CAH1_HUMAN-YSSLAEAASK-y8+)
  \item CATL1_HUMAN.VFQEPLFYEAPR.B3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CATL1_HUMAN-VFQEPLFYEAPR-b3+)
  \item CATL1_HUMAN.VFQEPLFYEAPR.B4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CATL1_HUMAN-VFQEPLFYEAPR-b4+)
  \item CATS_HUMAN.GIDSDASYPYK.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CATS_HUMAN-GIDSDASYPYK-y3+)
  \item CATS_HUMAN.GIDSDASYPYK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CATS_HUMAN-GIDSDASYPYK-y8+)
  \item CATS_HUMAN.GPVSVGVDAR.Y2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CATS_HUMAN-GPVSVGVDAR-y2+)
  \item CATS_HUMAN.GPVSVGVDAR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CATS_HUMAN-GPVSVGVDAR-y7+)
  \item CATS_HUMAN.HPSFFLYR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CATS_HUMAN-HPSFFLYR-y6+)
  \item CATS_HUMAN.HPSFFLYR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CATS_HUMAN-HPSFFLYR-y7+)
  \item CCKN_HUMAN.AHLGALLAR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CCKN_HUMAN-AHLGALLAR-y6+)
  \item CCKN_HUMAN.AHLGALLAR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CCKN_HUMAN-AHLGALLAR-y7+)
  \item CD14_HUMAN.AFPALTSLDLSDNPGLGER.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CD14_HUMAN-AFPALTSLDLSDNPGLGER-y7+)
  \item CD14_HUMAN.AFPALTSLDLSDNPGLGER.Y9. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CD14_HUMAN-AFPALTSLDLSDNPGLGER-y9+)
  \item CD44_HUMAN.YGFIEGHVVIPR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CD44_HUMAN-YGFIEGHVVIPR-y4+)
  \item CD44_HUMAN.YGFIEGHVVIPR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CD44_HUMAN-YGFIEGHVVIPR-y5+)
  \item CD59_HUMAN.AGLQVYNK.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CD59_HUMAN-AGLQVYNK-y4+)
  \item CD59_HUMAN.AGLQVYNK.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CD59_HUMAN-AGLQVYNK-y5+)
  \item CERU_HUMAN.IYHSHIDAPK.Y2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CERU_HUMAN-IYHSHIDAPK-y2+)
  \item CERU_HUMAN.IYHSHIDAPK.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CERU_HUMAN-IYHSHIDAPK-y3+)
  \item CERU_HUMAN.NNEGTYYSPNYNPQSR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CERU_HUMAN-NNEGTYYSPNYNPQSR-y4+)
  \item CERU_HUMAN.NNEGTYYSPNYNPQSR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CERU_HUMAN-NNEGTYYSPNYNPQSR-y8+)
  \item CFAI_HUMAN.HGNTDSEGIVEVK.B9. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CFAI_HUMAN-HGNTDSEGIVEVK-b9+)
  \item CFAI_HUMAN.HGNTDSEGIVEVK.Y2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CFAI_HUMAN-HGNTDSEGIVEVK-y2+)
  \item CFAI_HUMAN.VFSLQWGEVK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CFAI_HUMAN-VFSLQWGEVK-y6+)
  \item CFAI_HUMAN.VFSLQWGEVK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CFAI_HUMAN-VFSLQWGEVK-y8+)
  \item CH3L1_HUMAN.ILGQQVPYATK.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CH3L1_HUMAN-ILGQQVPYATK-y5+)
  \item CH3L1_HUMAN.ILGQQVPYATK.Y9. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CH3L1_HUMAN-ILGQQVPYATK-y9+)
  \item CH3L1_HUMAN.VTIDSSYDIAK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CH3L1_HUMAN-VTIDSSYDIAK-y7+)
  \item CH3L1_HUMAN.VTIDSSYDIAK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CH3L1_HUMAN-VTIDSSYDIAK-y8+)
  \item CLUS_HUMAN.IDSLLENDR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CLUS_HUMAN-IDSLLENDR-y5+)
  \item CLUS_HUMAN.IDSLLENDR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CLUS_HUMAN-IDSLLENDR-y7+)
  \item CLUS_HUMAN.SGSGLVGR.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CLUS_HUMAN-SGSGLVGR-y3+)
  \item CLUS_HUMAN.SGSGLVGR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CLUS_HUMAN-SGSGLVGR-y5+)
  \item CNTN1_HUMAN.DGEYVVEVR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CNTN1_HUMAN-DGEYVVEVR-y4+)
  \item CNTN1_HUMAN.DGEYVVEVR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CNTN1_HUMAN-DGEYVVEVR-y6+)
  \item CNTN1_HUMAN.TTKPYPADIVVQFK.B8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CNTN1_HUMAN-TTKPYPADIVVQFK-b8+)
  \item CNTN1_HUMAN.TTKPYPADIVVQFK.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CNTN1_HUMAN-TTKPYPADIVVQFK-y5+)
  \item CNTN2_HUMAN.IIVQAQPEWLK.B3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CNTN2_HUMAN-IIVQAQPEWLK-b3+)
  \item CNTN2_HUMAN.IIVQAQPEWLK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CNTN2_HUMAN-IIVQAQPEWLK-y6+)
  \item CNTN2_HUMAN.TTGPGGDGIPAEVHIVR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CNTN2_HUMAN-TTGPGGDGIPAEVHIVR-y4+)
  \item CNTN2_HUMAN.TTGPGGDGIPAEVHIVR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CNTN2_HUMAN-TTGPGGDGIPAEVHIVR-y8+)
  \item CNTN2_HUMAN.VIASNILGTGEPSGPSSK.B6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CNTN2_HUMAN-VIASNILGTGEPSGPSSK-b6+)
  \item CNTN2_HUMAN.VIASNILGTGEPSGPSSK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CNTN2_HUMAN-VIASNILGTGEPSGPSSK-y7+)
  \item CNTP2_HUMAN.YSSSDWVTQYR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CNTP2_HUMAN-YSSSDWVTQYR-y4+)
  \item CNTP2_HUMAN.YSSSDWVTQYR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CNTP2_HUMAN-YSSSDWVTQYR-y6+)
  \item CO2_HUMAN.HAIILLTDGK.B4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO2_HUMAN-HAIILLTDGK-b4+)
  \item CO2_HUMAN.HAIILLTDGK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO2_HUMAN-HAIILLTDGK-y8+)
  \item CO2_HUMAN.SSGQWQTPGATR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO2_HUMAN-SSGQWQTPGATR-y5+)
  \item CO2_HUMAN.SSGQWQTPGATR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO2_HUMAN-SSGQWQTPGATR-y7+)
  \item CO3_HUMAN.IHWESASLLR.B2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO3_HUMAN-IHWESASLLR-b2+)
  \item CO3_HUMAN.IHWESASLLR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO3_HUMAN-IHWESASLLR-y8+)
  \item CO3_HUMAN.LSINTHPSQKPLSITVR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO3_HUMAN-LSINTHPSQKPLSITVR-y5+)
  \item CO3_HUMAN.LSINTHPSQKPLSITVR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO3_HUMAN-LSINTHPSQKPLSITVR-y7+)
  \item CO3_HUMAN.TELRPGETLNVNFLLR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO3_HUMAN-TELRPGETLNVNFLLR-y5+)
  \item CO3_HUMAN.TELRPGETLNVNFLLR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO3_HUMAN-TELRPGETLNVNFLLR-y7+)
  \item CO3_HUMAN.TGLQEVEVK.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO3_HUMAN-TGLQEVEVK-y5+)
  \item CO3_HUMAN.TGLQEVEVK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO3_HUMAN-TGLQEVEVK-y6+)
  \item CO3_HUMAN.VPVAVQGEDTVQSLTQGDGVAK.B4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO3_HUMAN-VPVAVQGEDTVQSLTQGDGVAK-b4+)
  \item CO3_HUMAN.VPVAVQGEDTVQSLTQGDGVAK.B5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO3_HUMAN-VPVAVQGEDTVQSLTQGDGVAK-b5+)
  \item CO4A_HUMAN.DHAVDLIQK.B2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO4A_HUMAN-DHAVDLIQK-b2+)
  \item CO4A_HUMAN.DHAVDLIQK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO4A_HUMAN-DHAVDLIQK-y7+)
  \item CO4A_HUMAN.LGQYASPTAK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO4A_HUMAN-LGQYASPTAK-y6+)
  \item CO4A_HUMAN.LGQYASPTAK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO4A_HUMAN-LGQYASPTAK-y7+)
  \item CO4A_HUMAN.NVNFQK.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO4A_HUMAN-NVNFQK-y3+)
  \item CO4A_HUMAN.NVNFQK.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO4A_HUMAN-NVNFQK-y4+)
  \item CO4A_HUMAN.VLSLAQEQVGGSPEK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO4A_HUMAN-VLSLAQEQVGGSPEK-y6+)
  \item CO4A_HUMAN.VLSLAQEQVGGSPEK.Y9. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO4A_HUMAN-VLSLAQEQVGGSPEK-y9+)
  \item CO5_HUMAN.DINYVNPVIK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO5_HUMAN-DINYVNPVIK-y6+)
  \item CO5_HUMAN.DINYVNPVIK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO5_HUMAN-DINYVNPVIK-y8+)
  \item CO5_HUMAN.TLLPVSKPEIR.Y2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO5_HUMAN-TLLPVSKPEIR-y2+)
  \item CO5_HUMAN.TLLPVSKPEIR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO5_HUMAN-TLLPVSKPEIR-y7+)
  \item CO5_HUMAN.VFQFLEK.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO5_HUMAN-VFQFLEK-y5+)
  \item CO5_HUMAN.VFQFLEK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO5_HUMAN-VFQFLEK-y6+)
  \item CO6_HUMAN.ALNHLPLEYNSALYSR.B5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO6_HUMAN-ALNHLPLEYNSALYSR-b5+)
  \item CO6_HUMAN.ALNHLPLEYNSALYSR.B8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO6_HUMAN-ALNHLPLEYNSALYSR-b8+)
  \item CO6_HUMAN.SEYGAALAWEK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO6_HUMAN-SEYGAALAWEK-y6+)
  \item CO6_HUMAN.SEYGAALAWEK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO6_HUMAN-SEYGAALAWEK-y8+)
  \item CO7_HUMAN.AASGTQNNVLR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO7_HUMAN-AASGTQNNVLR-y5+)
  \item CO7_HUMAN.AASGTQNNVLR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO7_HUMAN-AASGTQNNVLR-y6+)
  \item CO7_HUMAN.ELSHLPSLYDYSAYR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO7_HUMAN-ELSHLPSLYDYSAYR-y4+)
  \item CO7_HUMAN.ELSHLPSLYDYSAYR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO7_HUMAN-ELSHLPSLYDYSAYR-y6+)
  \item CO7_HUMAN.LSGNVLSYTFQVK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO7_HUMAN-LSGNVLSYTFQVK-y7+)
  \item CO7_HUMAN.LSGNVLSYTFQVK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO7_HUMAN-LSGNVLSYTFQVK-y8+)
  \item CO7_HUMAN.SYTSHTNEIHK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO7_HUMAN-SYTSHTNEIHK-y6+)
  \item CO7_HUMAN.SYTSHTNEIHK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO7_HUMAN-SYTSHTNEIHK-y7+)
  \item CO8B_HUMAN.IPGIFELGISSQSDR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO8B_HUMAN-IPGIFELGISSQSDR-y8+)
  \item CO8B_HUMAN.IPGIFELGISSQSDR.Y9. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO8B_HUMAN-IPGIFELGISSQSDR-y9+)
  \item CO8B_HUMAN.SDLEVAHYK.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO8B_HUMAN-SDLEVAHYK-y5+)
  \item CO8B_HUMAN.SDLEVAHYK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO8B_HUMAN-SDLEVAHYK-y6+)
  \item CO8B_HUMAN.YEFILK.B2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO8B_HUMAN-YEFILK-b2+)
  \item CO8B_HUMAN.YEFILK.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO8B_HUMAN-YEFILK-y4+)
  \item CO9_HUMAN.LSPIYNLVPVK.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO9_HUMAN-LSPIYNLVPVK-y3+)
  \item CO9_HUMAN.LSPIYNLVPVK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO9_HUMAN-LSPIYNLVPVK-y7+)
  \item CO9_HUMAN.TSNFNAAISLK.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO9_HUMAN-TSNFNAAISLK-y3+)
  \item CO9_HUMAN.TSNFNAAISLK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO9_HUMAN-TSNFNAAISLK-y7+)
  \item CO9_HUMAN.VVEESELAR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO9_HUMAN-VVEESELAR-y6+)
  \item CO9_HUMAN.VVEESELAR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CO9_HUMAN-VVEESELAR-y7+)
  \item COCH_HUMAN.GVISNSGGPVR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (COCH_HUMAN-GVISNSGGPVR-y7+)
  \item COCH_HUMAN.GVISNSGGPVR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (COCH_HUMAN-GVISNSGGPVR-y8+)
  \item CPVL_HUMAN.FLSLPEVR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CPVL_HUMAN-FLSLPEVR-y4+)
  \item CPVL_HUMAN.FLSLPEVR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CPVL_HUMAN-FLSLPEVR-y6+)
  \item CPVL_HUMAN.GGGHILPYDQPLR.B6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CPVL_HUMAN-GGGHILPYDQPLR-b6+)
  \item CPVL_HUMAN.GGGHILPYDQPLR.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CPVL_HUMAN-GGGHILPYDQPLR-y3+)
  \item CPVL_HUMAN.NNDFYVTGESYAGK.B3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CPVL_HUMAN-NNDFYVTGESYAGK-b3+)
  \item CPVL_HUMAN.NNDFYVTGESYAGK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CPVL_HUMAN-NNDFYVTGESYAGK-y8+)
  \item CRP_HUMAN.ESDTSYVSLK.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CRP_HUMAN-ESDTSYVSLK-y3+)
  \item CRP_HUMAN.ESDTSYVSLK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CRP_HUMAN-ESDTSYVSLK-y6+)
  \item CSTN1_HUMAN.GNLAGLTLR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CSTN1_HUMAN-GNLAGLTLR-y5+)
  \item CSTN1_HUMAN.GNLAGLTLR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CSTN1_HUMAN-GNLAGLTLR-y6+)
  \item CSTN1_HUMAN.IHGQNVPFDAVVVDK.B6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CSTN1_HUMAN-IHGQNVPFDAVVVDK-b6+)
  \item CSTN1_HUMAN.IHGQNVPFDAVVVDK.Y9. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CSTN1_HUMAN-IHGQNVPFDAVVVDK-y9+)
  \item CSTN1_HUMAN.IPDGVVSVSPK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CSTN1_HUMAN-IPDGVVSVSPK-y6+)
  \item CSTN1_HUMAN.IPDGVVSVSPK.Y9. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CSTN1_HUMAN-IPDGVVSVSPK-y9+)
  \item CSTN3_HUMAN.ESLLLDTTSLQQR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CSTN3_HUMAN-ESLLLDTTSLQQR-y7+)
  \item CSTN3_HUMAN.ESLLLDTTSLQQR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (CSTN3_HUMAN-ESLLLDTTSLQQR-y8+)
  \item DIAC_HUMAN.ATYIQNYR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (DIAC_HUMAN-ATYIQNYR-y4+)
  \item DIAC_HUMAN.ATYIQNYR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (DIAC_HUMAN-ATYIQNYR-y6+)
  \item DIAC_HUMAN.ETTDSFHR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (DIAC_HUMAN-ETTDSFHR-y4+)
  \item DIAC_HUMAN.ETTDSFHR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (DIAC_HUMAN-ETTDSFHR-y6+)
  \item DNS2A_HUMAN.AFQPLVK.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (DNS2A_HUMAN-AFQPLVK-y4+)
  \item DNS2A_HUMAN.AFQPLVK.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (DNS2A_HUMAN-AFQPLVK-y5+)
  \item DNS2A_HUMAN.SLQPLYR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (DNS2A_HUMAN-SLQPLYR-y4+)
  \item DNS2A_HUMAN.SLQPLYR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (DNS2A_HUMAN-SLQPLYR-y5+)
  \item DNS2A_HUMAN.YLDESSGGWR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (DNS2A_HUMAN-YLDESSGGWR-y7+)
  \item DNS2A_HUMAN.YLDESSGGWR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (DNS2A_HUMAN-YLDESSGGWR-y8+)
  \item EFNA1_HUMAN.FTPFTLGK.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (EFNA1_HUMAN-FTPFTLGK-y4+)
  \item EFNA1_HUMAN.FTPFTLGK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (EFNA1_HUMAN-FTPFTLGK-y6+)
  \item EFNA1_HUMAN.LAADDPEVR.Y2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (EFNA1_HUMAN-LAADDPEVR-y2+)
  \item EFNA1_HUMAN.LAADDPEVR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (EFNA1_HUMAN-LAADDPEVR-y8+)
  \item ENOG_HUMAN.GNPTVEVDLYTAK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (ENOG_HUMAN-GNPTVEVDLYTAK-y7+)
  \item ENOG_HUMAN.GNPTVEVDLYTAK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (ENOG_HUMAN-GNPTVEVDLYTAK-y8+)
  \item ENOG_HUMAN.LGAEVYHTLK.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (ENOG_HUMAN-LGAEVYHTLK-y5+)
  \item ENOG_HUMAN.LGAEVYHTLK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (ENOG_HUMAN-LGAEVYHTLK-y6+)
  \item ENPP2_HUMAN.SYPEILTLK.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (ENPP2_HUMAN-SYPEILTLK-y5+)
  \item ENPP2_HUMAN.SYPEILTLK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (ENPP2_HUMAN-SYPEILTLK-y7+)
  \item EXTL2_HUMAN.VIVVWNNIGEK.B3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (EXTL2_HUMAN-VIVVWNNIGEK-b3+)
  \item EXTL2_HUMAN.VIVVWNNIGEK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (EXTL2_HUMAN-VIVVWNNIGEK-y7+)
  \item FAM3C_HUMAN.GINVALANGK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FAM3C_HUMAN-GINVALANGK-y6+)
  \item FAM3C_HUMAN.GINVALANGK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FAM3C_HUMAN-GINVALANGK-y8+)
  \item FAM3C_HUMAN.TGEVLDTK.Y2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FAM3C_HUMAN-TGEVLDTK-y2+)
  \item FAM3C_HUMAN.TGEVLDTK.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FAM3C_HUMAN-TGEVLDTK-y5+)
  \item FBLN1_HUMAN.AITPPHPASQANIIFDITEGNLR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FBLN1_HUMAN-AITPPHPASQANIIFDITEGNLR-y6+)
  \item FBLN1_HUMAN.AITPPHPASQANIIFDITEGNLR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FBLN1_HUMAN-AITPPHPASQANIIFDITEGNLR-y8+)
  \item FBLN1_HUMAN.EFTRPEEIIFLR.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FBLN1_HUMAN-EFTRPEEIIFLR-y3+)
  \item FBLN1_HUMAN.EFTRPEEIIFLR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FBLN1_HUMAN-EFTRPEEIIFLR-y4+)
  \item FBLN1_HUMAN.GYHLNEEGTR.B3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FBLN1_HUMAN-GYHLNEEGTR-b3+)
  \item FBLN1_HUMAN.GYHLNEEGTR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FBLN1_HUMAN-GYHLNEEGTR-y7+)
  \item FBLN1_HUMAN.TGYYFDGISR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FBLN1_HUMAN-TGYYFDGISR-y6+)
  \item FBLN1_HUMAN.TGYYFDGISR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FBLN1_HUMAN-TGYYFDGISR-y7+)
  \item FETUA_HUMAN.FSVVYAK.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FETUA_HUMAN-FSVVYAK-y5+)
  \item FETUA_HUMAN.FSVVYAK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FETUA_HUMAN-FSVVYAK-y6+)
  \item FMOD_HUMAN.SLILLDLSYNHLR.B3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FMOD_HUMAN-SLILLDLSYNHLR-b3+)
  \item FMOD_HUMAN.SLILLDLSYNHLR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FMOD_HUMAN-SLILLDLSYNHLR-y7+)
  \item FMOD_HUMAN.YLPFVPSR.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FMOD_HUMAN-YLPFVPSR-y3+)
  \item FMOD_HUMAN.YLPFVPSR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FMOD_HUMAN-YLPFVPSR-y6+)
  \item FUCO_HUMAN.DGLIVPIFQER.B3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FUCO_HUMAN-DGLIVPIFQER-b3+)
  \item FUCO_HUMAN.DGLIVPIFQER.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FUCO_HUMAN-DGLIVPIFQER-y6+)
  \item FUCO_HUMAN.DLVGELGTALR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FUCO_HUMAN-DLVGELGTALR-y5+)
  \item FUCO_HUMAN.DLVGELGTALR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FUCO_HUMAN-DLVGELGTALR-y8+)
  \item FUCO_HUMAN.YVVLTTK.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FUCO_HUMAN-YVVLTTK-y3+)
  \item FUCO_HUMAN.YVVLTTK.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (FUCO_HUMAN-YVVLTTK-y5+)
  \item GNS_HUMAN.IQEPNTFPAILR.B3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (GNS_HUMAN-IQEPNTFPAILR-b3+)
  \item GNS_HUMAN.IQEPNTFPAILR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (GNS_HUMAN-IQEPNTFPAILR-y5+)
  \item GNS_HUMAN.SDVLVEYQGEGR.B3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (GNS_HUMAN-SDVLVEYQGEGR-b3+)
  \item GNS_HUMAN.SDVLVEYQGEGR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (GNS_HUMAN-SDVLVEYQGEGR-y7+)
  \item GNS_HUMAN.TIDPELLGK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (GNS_HUMAN-TIDPELLGK-y6+)
  \item GNS_HUMAN.TIDPELLGK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (GNS_HUMAN-TIDPELLGK-y7+)
  \item HAVR2_HUMAN.DVNYWTSR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HAVR2_HUMAN-DVNYWTSR-y4+)
  \item HAVR2_HUMAN.DVNYWTSR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HAVR2_HUMAN-DVNYWTSR-y6+)
  \item HAVR2_HUMAN.TDERDVNYWTSR.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HAVR2_HUMAN-TDERDVNYWTSR-y3+)
  \item HAVR2_HUMAN.TDERDVNYWTSR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HAVR2_HUMAN-TDERDVNYWTSR-y4+)
  \item HAVR2_HUMAN.YWLNGDFR.B2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HAVR2_HUMAN-YWLNGDFR-b2+)
  \item HAVR2_HUMAN.YWLNGDFR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HAVR2_HUMAN-YWLNGDFR-y6+)
  \item HBA_HUMAN.FLASVSTVLTSK.B2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HBA_HUMAN-FLASVSTVLTSK-b2+)
  \item HBA_HUMAN.FLASVSTVLTSK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HBA_HUMAN-FLASVSTVLTSK-y7+)
  \item HBA_HUMAN.TYFPHFDLSHGSAQVK.B2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HBA_HUMAN-TYFPHFDLSHGSAQVK-b2+)
  \item HBA_HUMAN.TYFPHFDLSHGSAQVK.B3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HBA_HUMAN-TYFPHFDLSHGSAQVK-b3+)
  \item HBB_HUMAN.EFTPPVQAAYQK.B3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HBB_HUMAN-EFTPPVQAAYQK-b3+)
  \item HBB_HUMAN.EFTPPVQAAYQK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HBB_HUMAN-EFTPPVQAAYQK-y8+)
  \item HBB_HUMAN.SAVTALWGK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HBB_HUMAN-SAVTALWGK-y6+)
  \item HBB_HUMAN.SAVTALWGK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HBB_HUMAN-SAVTALWGK-y7+)
  \item HBB_HUMAN.VNVDEVGGEALGR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HBB_HUMAN-VNVDEVGGEALGR-y8+)
  \item HBB_HUMAN.VNVDEVGGEALGR.Y9. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HBB_HUMAN-VNVDEVGGEALGR-y9+)
  \item HEMO_HUMAN.NFPSPVDAAFR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HEMO_HUMAN-NFPSPVDAAFR-y7+)
  \item HEMO_HUMAN.NFPSPVDAAFR.Y9. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HEMO_HUMAN-NFPSPVDAAFR-y9+)
  \item HEMO_HUMAN.QGHNSVFLIK.Y2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HEMO_HUMAN-QGHNSVFLIK-y2+)
  \item HEMO_HUMAN.QGHNSVFLIK.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HEMO_HUMAN-QGHNSVFLIK-y4+)
  \item HEMO_HUMAN.SGAQATWTELPWPHEK.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HEMO_HUMAN-SGAQATWTELPWPHEK-y4+)
  \item HEMO_HUMAN.SGAQATWTELPWPHEK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HEMO_HUMAN-SGAQATWTELPWPHEK-y6+)
  \item HEP2_HUMAN.FAFNLYR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HEP2_HUMAN-FAFNLYR-y5+)
  \item HEP2_HUMAN.FAFNLYR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HEP2_HUMAN-FAFNLYR-y6+)
  \item HEP2_HUMAN.IAIDLFK.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HEP2_HUMAN-IAIDLFK-y5+)
  \item HEP2_HUMAN.IAIDLFK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HEP2_HUMAN-IAIDLFK-y6+)
  \item HEP2_HUMAN.NFGYTLR.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HEP2_HUMAN-NFGYTLR-y3+)
  \item HEP2_HUMAN.NFGYTLR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HEP2_HUMAN-NFGYTLR-y5+)
  \item HRG_HUMAN.DGYLFQLLR.B3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HRG_HUMAN-DGYLFQLLR-b3+)
  \item HRG_HUMAN.DGYLFQLLR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (HRG_HUMAN-DGYLFQLLR-y5+)
  \item IC1_HUMAN.FQPTLLTLPR.B2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (IC1_HUMAN-FQPTLLTLPR-b2+)
  \item IC1_HUMAN.FQPTLLTLPR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (IC1_HUMAN-FQPTLLTLPR-y8+)
  \item IC1_HUMAN.LLDSLPSDTR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (IC1_HUMAN-LLDSLPSDTR-y5+)
  \item IC1_HUMAN.LLDSLPSDTR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (IC1_HUMAN-LLDSLPSDTR-y8+)
  \item IGSF8_HUMAN.DTQFSYAVFK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (IGSF8_HUMAN-DTQFSYAVFK-y6+)
  \item IGSF8_HUMAN.DTQFSYAVFK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (IGSF8_HUMAN-DTQFSYAVFK-y7+)
  \item IGSF8_HUMAN.LQGDAVVLK.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (IGSF8_HUMAN-LQGDAVVLK-y3+)
  \item IGSF8_HUMAN.LQGDAVVLK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (IGSF8_HUMAN-LQGDAVVLK-y7+)
  \item IGSF8_HUMAN.VVAGEVQVQR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (IGSF8_HUMAN-VVAGEVQVQR-y7+)
  \item IGSF8_HUMAN.VVAGEVQVQR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (IGSF8_HUMAN-VVAGEVQVQR-y8+)
  \item IPSP_HUMAN.FSIEGSYQLEK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (IPSP_HUMAN-FSIEGSYQLEK-y7+)
  \item IPSP_HUMAN.FSIEGSYQLEK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (IPSP_HUMAN-FSIEGSYQLEK-y8+)
  \item IPSP_HUMAN.GFQQLLQELNQPR.Y2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (IPSP_HUMAN-GFQQLLQELNQPR-y2+)
  \item IPSP_HUMAN.GFQQLLQELNQPR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (IPSP_HUMAN-GFQQLLQELNQPR-y7+)
  \item IPSP_HUMAN.TLYLADTFPTNFR.B3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (IPSP_HUMAN-TLYLADTFPTNFR-b3+)
  \item IPSP_HUMAN.TLYLADTFPTNFR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (IPSP_HUMAN-TLYLADTFPTNFR-y5+)
  \item ITIH1_HUMAN.EVAFDLEIPK.Y2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (ITIH1_HUMAN-EVAFDLEIPK-y2+)
  \item ITIH1_HUMAN.EVAFDLEIPK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (ITIH1_HUMAN-EVAFDLEIPK-y6+)
  \item ITIH1_HUMAN.QYYEGSEIVVAGR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (ITIH1_HUMAN-QYYEGSEIVVAGR-y4+)
  \item ITIH1_HUMAN.QYYEGSEIVVAGR.Y9. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (ITIH1_HUMAN-QYYEGSEIVVAGR-y9+)
  \item KAIN_HUMAN.WADLSGITK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (KAIN_HUMAN-WADLSGITK-y7+)
  \item KAIN_HUMAN.WADLSGITK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (KAIN_HUMAN-WADLSGITK-y8+)
  \item KLK6_HUMAN.ESSQEQSSVVR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (KLK6_HUMAN-ESSQEQSSVVR-y6+)
  \item KLK6_HUMAN.ESSQEQSSVVR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (KLK6_HUMAN-ESSQEQSSVVR-y7+)
  \item KLK6_HUMAN.LSELIQPLPLER.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (KLK6_HUMAN-LSELIQPLPLER-y6+)
  \item KLK6_HUMAN.LSELIQPLPLER.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (KLK6_HUMAN-LSELIQPLPLER-y7+)
  \item KLK6_HUMAN.YTNWIQK.B2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (KLK6_HUMAN-YTNWIQK-b2+)
  \item KLK6_HUMAN.YTNWIQK.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (KLK6_HUMAN-YTNWIQK-y5+)
  \item KNG1_HUMAN.DIPTNSPELEETLTHTITK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (KNG1_HUMAN-DIPTNSPELEETLTHTITK-y7+)
  \item KNG1_HUMAN.DIPTNSPELEETLTHTITK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (KNG1_HUMAN-DIPTNSPELEETLTHTITK-y8+)
  \item KNG1_HUMAN.QVVAGLNFR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (KNG1_HUMAN-QVVAGLNFR-y6+)
  \item KNG1_HUMAN.QVVAGLNFR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (KNG1_HUMAN-QVVAGLNFR-y7+)
  \item KNG1_HUMAN.TVGSDTFYSFK.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (KNG1_HUMAN-TVGSDTFYSFK-y3+)
  \item KNG1_HUMAN.TVGSDTFYSFK.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (KNG1_HUMAN-TVGSDTFYSFK-y4+)
  \item KPYM_HUMAN.LDIDSPPITAR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (KPYM_HUMAN-LDIDSPPITAR-y6+)
  \item KPYM_HUMAN.LDIDSPPITAR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (KPYM_HUMAN-LDIDSPPITAR-y8+)
  \item LAMP2_HUMAN.GILTVDELLAIR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (LAMP2_HUMAN-GILTVDELLAIR-y7+)
  \item LAMP2_HUMAN.GILTVDELLAIR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (LAMP2_HUMAN-GILTVDELLAIR-y8+)
  \item LAMP2_HUMAN.IPLNDLFR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (LAMP2_HUMAN-IPLNDLFR-y5+)
  \item LAMP2_HUMAN.IPLNDLFR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (LAMP2_HUMAN-IPLNDLFR-y6+)
  \item LAMP2_HUMAN.YLDFVFAVK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (LAMP2_HUMAN-YLDFVFAVK-y6+)
  \item LAMP2_HUMAN.YLDFVFAVK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (LAMP2_HUMAN-YLDFVFAVK-y7+)
  \item LIGO1_HUMAN.LLWVFR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (LIGO1_HUMAN-LLWVFR-y4+)
  \item LIGO1_HUMAN.LLWVFR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (LIGO1_HUMAN-LLWVFR-y5+)
  \item LSR_HUMAN.LLEEAVR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (LSR_HUMAN-LLEEAVR-y5+)
  \item LSR_HUMAN.LLEEAVR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (LSR_HUMAN-LLEEAVR-y6+)
  \item LSR_HUMAN.SGDLPYDGR.B3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (LSR_HUMAN-SGDLPYDGR-b3+)
  \item LSR_HUMAN.SGDLPYDGR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (LSR_HUMAN-SGDLPYDGR-y5+)
  \item LYSC_HUMAN.ATNYNAGDR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (LYSC_HUMAN-ATNYNAGDR-y5+)
  \item LYSC_HUMAN.ATNYNAGDR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (LYSC_HUMAN-ATNYNAGDR-y7+)
  \item LYSC_HUMAN.STDYGIFQINSR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (LYSC_HUMAN-STDYGIFQINSR-y6+)
  \item LYSC_HUMAN.STDYGIFQINSR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (LYSC_HUMAN-STDYGIFQINSR-y8+)
  \item LYSC_HUMAN.WESGYNTR.B2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (LYSC_HUMAN-WESGYNTR-b2+)
  \item LYSC_HUMAN.WESGYNTR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (LYSC_HUMAN-WESGYNTR-y6+)
  \item MUC18_HUMAN.EVTVPVFYPTEK.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (MUC18_HUMAN-EVTVPVFYPTEK-y4+)
  \item MUC18_HUMAN.EVTVPVFYPTEK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (MUC18_HUMAN-EVTVPVFYPTEK-y8+)
  \item MUC18_HUMAN.GATLALTQVTPQDER.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (MUC18_HUMAN-GATLALTQVTPQDER-y5+)
  \item MUC18_HUMAN.GATLALTQVTPQDER.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (MUC18_HUMAN-GATLALTQVTPQDER-y6+)
  \item NCAN_HUMAN.APVLELEK.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NCAN_HUMAN-APVLELEK-y5+)
  \item NCAN_HUMAN.APVLELEK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NCAN_HUMAN-APVLELEK-y6+)
  \item NCAN_HUMAN.LSSAIIAAPR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NCAN_HUMAN-LSSAIIAAPR-y5+)
  \item NCAN_HUMAN.LSSAIIAAPR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NCAN_HUMAN-LSSAIIAAPR-y8+)
  \item NEGR1_HUMAN.SSIIFAGGDK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NEGR1_HUMAN-SSIIFAGGDK-y6+)
  \item NEGR1_HUMAN.SSIIFAGGDK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NEGR1_HUMAN-SSIIFAGGDK-y7+)
  \item NEGR1_HUMAN.VVVNFAPTIQEIK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NEGR1_HUMAN-VVVNFAPTIQEIK-y7+)
  \item NEGR1_HUMAN.VVVNFAPTIQEIK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NEGR1_HUMAN-VVVNFAPTIQEIK-y8+)
  \item NELL2_HUMAN.AFLFQDTPR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NELL2_HUMAN-AFLFQDTPR-y6+)
  \item NELL2_HUMAN.AFLFQDTPR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NELL2_HUMAN-AFLFQDTPR-y7+)
  \item NELL2_HUMAN.FTGSSWIK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NELL2_HUMAN-FTGSSWIK-y6+)
  \item NELL2_HUMAN.FTGSSWIK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NELL2_HUMAN-FTGSSWIK-y7+)
  \item NELL2_HUMAN.SALAYVDGK.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NELL2_HUMAN-SALAYVDGK-y3+)
  \item NELL2_HUMAN.SALAYVDGK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NELL2_HUMAN-SALAYVDGK-y6+)
  \item NEUS_HUMAN.ALGITEIFIK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NEUS_HUMAN-ALGITEIFIK-y6+)
  \item NEUS_HUMAN.ALGITEIFIK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NEUS_HUMAN-ALGITEIFIK-y8+)
  \item NEUS_HUMAN.QEVPLATLEPLVK.B2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NEUS_HUMAN-QEVPLATLEPLVK-b2+)
  \item NEUS_HUMAN.QEVPLATLEPLVK.B3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NEUS_HUMAN-QEVPLATLEPLVK-b3+)
  \item NPTX1_HUMAN.FQLTFPLR.B2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NPTX1_HUMAN-FQLTFPLR-b2+)
  \item NPTX1_HUMAN.FQLTFPLR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NPTX1_HUMAN-FQLTFPLR-y6+)
  \item NPTX1_HUMAN.LENLEQYSR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NPTX1_HUMAN-LENLEQYSR-y5+)
  \item NPTX1_HUMAN.LENLEQYSR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NPTX1_HUMAN-LENLEQYSR-y7+)
  \item NPTXR_HUMAN.ELDVLQGR.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NPTXR_HUMAN-ELDVLQGR-y3+)
  \item NPTXR_HUMAN.ELDVLQGR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NPTXR_HUMAN-ELDVLQGR-y6+)
  \item NPTXR_HUMAN.LVEAFGGATK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NPTXR_HUMAN-LVEAFGGATK-y7+)
  \item NPTXR_HUMAN.LVEAFGGATK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NPTXR_HUMAN-LVEAFGGATK-y8+)
  \item NRCAM_HUMAN.SLPSEASEQYLTK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NRCAM_HUMAN-SLPSEASEQYLTK-y7+)
  \item NRCAM_HUMAN.SLPSEASEQYLTK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NRCAM_HUMAN-SLPSEASEQYLTK-y8+)
  \item NRCAM_HUMAN.VFNTPEGVPSAPSSLK.B4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NRCAM_HUMAN-VFNTPEGVPSAPSSLK-b4+)
  \item NRCAM_HUMAN.VFNTPEGVPSAPSSLK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NRCAM_HUMAN-VFNTPEGVPSAPSSLK-y8+)
  \item NRCAM_HUMAN.YIVSGTPTFVPYLIK.B2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NRCAM_HUMAN-YIVSGTPTFVPYLIK-b2+)
  \item NRCAM_HUMAN.YIVSGTPTFVPYLIK.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NRCAM_HUMAN-YIVSGTPTFVPYLIK-y5+)
  \item NRX1A_HUMAN.DLFIDGQSK.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NRX1A_HUMAN-DLFIDGQSK-y5+)
  \item NRX1A_HUMAN.DLFIDGQSK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NRX1A_HUMAN-DLFIDGQSK-y6+)
  \item NRX1A_HUMAN.ITTQITAGAR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NRX1A_HUMAN-ITTQITAGAR-y5+)
  \item NRX1A_HUMAN.ITTQITAGAR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NRX1A_HUMAN-ITTQITAGAR-y8+)
  \item NRX1A_HUMAN.SDLYIGGVAK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NRX1A_HUMAN-SDLYIGGVAK-y6+)
  \item NRX1A_HUMAN.SDLYIGGVAK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NRX1A_HUMAN-SDLYIGGVAK-y7+)
  \item NRX2A_HUMAN.AIVADPVTFK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NRX2A_HUMAN-AIVADPVTFK-y7+)
  \item NRX2A_HUMAN.AIVADPVTFK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NRX2A_HUMAN-AIVADPVTFK-y8+)
  \item NRX2A_HUMAN.LGERPPALLGSQGLR.B8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NRX2A_HUMAN-LGERPPALLGSQGLR-b8+)
  \item NRX2A_HUMAN.LGERPPALLGSQGLR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NRX2A_HUMAN-LGERPPALLGSQGLR-y6+)
  \item NRX2A_HUMAN.LSALTLSTVK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NRX2A_HUMAN-LSALTLSTVK-y6+)
  \item NRX2A_HUMAN.LSALTLSTVK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (NRX2A_HUMAN-LSALTLSTVK-y8+)
  \item OSTP_HUMAN.AIPVAQDLNAPSDWDSR.Y2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (OSTP_HUMAN-AIPVAQDLNAPSDWDSR-y2+)
  \item OSTP_HUMAN.AIPVAQDLNAPSDWDSR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (OSTP_HUMAN-AIPVAQDLNAPSDWDSR-y7+)
  \item OSTP_HUMAN.YPDAVATWLNPDPSQK.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (OSTP_HUMAN-YPDAVATWLNPDPSQK-y4+)
  \item OSTP_HUMAN.YPDAVATWLNPDPSQK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (OSTP_HUMAN-YPDAVATWLNPDPSQK-y6+)
  \item PCSK1_HUMAN.ALAHLLEAER.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PCSK1_HUMAN-ALAHLLEAER-y5+)
  \item PCSK1_HUMAN.ALAHLLEAER.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PCSK1_HUMAN-ALAHLLEAER-y6+)
  \item PDYN_HUMAN.FLPSISTK.B2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PDYN_HUMAN-FLPSISTK-b2+)
  \item PDYN_HUMAN.FLPSISTK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PDYN_HUMAN-FLPSISTK-y6+)
  \item PEDF_HUMAN.DTDTGALLFIGK.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PEDF_HUMAN-DTDTGALLFIGK-y4+)
  \item PEDF_HUMAN.DTDTGALLFIGK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PEDF_HUMAN-DTDTGALLFIGK-y8+)
  \item PEDF_HUMAN.SSFVAPLEK.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PEDF_HUMAN-SSFVAPLEK-y5+)
  \item PEDF_HUMAN.SSFVAPLEK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PEDF_HUMAN-SSFVAPLEK-y6+)
  \item PEDF_HUMAN.TVQAVLTVPK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PEDF_HUMAN-TVQAVLTVPK-y7+)
  \item PEDF_HUMAN.TVQAVLTVPK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PEDF_HUMAN-TVQAVLTVPK-y8+)
  \item PENK_HUMAN.ELLETGDNR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PENK_HUMAN-ELLETGDNR-y5+)
  \item PENK_HUMAN.ELLETGDNR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PENK_HUMAN-ELLETGDNR-y7+)
  \item PENK_HUMAN.FAEALPSDEEGESYSK.B4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PENK_HUMAN-FAEALPSDEEGESYSK-b4+)
  \item PENK_HUMAN.FAEALPSDEEGESYSK.B5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PENK_HUMAN-FAEALPSDEEGESYSK-b5+)
  \item PGRP2_HUMAN.AGLLRPDYALLGHR.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PGRP2_HUMAN-AGLLRPDYALLGHR-y3+)
  \item PGRP2_HUMAN.AGLLRPDYALLGHR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PGRP2_HUMAN-AGLLRPDYALLGHR-y4+)
  \item PGRP2_HUMAN.TFTLLDPK.Y2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PGRP2_HUMAN-TFTLLDPK-y2+)
  \item PGRP2_HUMAN.TFTLLDPK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PGRP2_HUMAN-TFTLLDPK-y6+)
  \item PIMT_HUMAN.VQLVVGDGR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PIMT_HUMAN-VQLVVGDGR-y5+)
  \item PIMT_HUMAN.VQLVVGDGR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PIMT_HUMAN-VQLVVGDGR-y7+)
  \item PLDX1_HUMAN.LYGPSEPHSR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PLDX1_HUMAN-LYGPSEPHSR-y4+)
  \item PLDX1_HUMAN.LYGPSEPHSR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PLDX1_HUMAN-LYGPSEPHSR-y8+)
  \item PLMN_HUMAN.EAQLPVIENK.B3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PLMN_HUMAN-EAQLPVIENK-b3+)
  \item PLMN_HUMAN.EAQLPVIENK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PLMN_HUMAN-EAQLPVIENK-y6+)
  \item PLMN_HUMAN.HSIFTPETNPR.B4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PLMN_HUMAN-HSIFTPETNPR-b4+)
  \item PLMN_HUMAN.HSIFTPETNPR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PLMN_HUMAN-HSIFTPETNPR-y6+)
  \item PLMN_HUMAN.LSSPAVITDK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PLMN_HUMAN-LSSPAVITDK-y7+)
  \item PLMN_HUMAN.LSSPAVITDK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PLMN_HUMAN-LSSPAVITDK-y8+)
  \item PON1_HUMAN.IFFYDSENPPASEVLR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PON1_HUMAN-IFFYDSENPPASEVLR-y8+)
  \item PON1_HUMAN.IFFYDSENPPASEVLR.Y9. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PON1_HUMAN-IFFYDSENPPASEVLR-y9+)
  \item PRDX1_HUMAN.ADEGISFR.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PRDX1_HUMAN-ADEGISFR-y3+)
  \item PRDX1_HUMAN.LVQAFQFTDK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PRDX1_HUMAN-LVQAFQFTDK-y7+)
  \item PRDX1_HUMAN.LVQAFQFTDK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PRDX1_HUMAN-LVQAFQFTDK-y8+)
  \item PRDX2_HUMAN.ATAVVDGAFK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PRDX2_HUMAN-ATAVVDGAFK-y6+)
  \item PRDX2_HUMAN.ATAVVDGAFK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PRDX2_HUMAN-ATAVVDGAFK-y8+)
  \item PRDX2_HUMAN.GLFIIDGK.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PRDX2_HUMAN-GLFIIDGK-y5+)
  \item PRDX2_HUMAN.GLFIIDGK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PRDX2_HUMAN-GLFIIDGK-y6+)
  \item PRDX2_HUMAN.IGKPAPDFK.Y2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PRDX2_HUMAN-IGKPAPDFK-y2+)
  \item PRDX2_HUMAN.IGKPAPDFK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PRDX2_HUMAN-IGKPAPDFK-y6+)
  \item PRDX3_HUMAN.GTAVVNGEFK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PRDX3_HUMAN-GTAVVNGEFK-y6+)
  \item PRDX3_HUMAN.GTAVVNGEFK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PRDX3_HUMAN-GTAVVNGEFK-y8+)
  \item PRDX6_HUMAN.LSILYPATTGR.B4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PRDX6_HUMAN-LSILYPATTGR-b4+)
  \item PRDX6_HUMAN.LSILYPATTGR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PRDX6_HUMAN-LSILYPATTGR-y7+)
  \item PTPRN_HUMAN.AEAPALFSR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PTPRN_HUMAN-AEAPALFSR-y5+)
  \item PTPRN_HUMAN.AEAPALFSR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PTPRN_HUMAN-AEAPALFSR-y7+)
  \item PTPRN_HUMAN.SELEAQTGLQILQTGVGQR.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PTPRN_HUMAN-SELEAQTGLQILQTGVGQR-y3+)
  \item PTPRN_HUMAN.SELEAQTGLQILQTGVGQR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PTPRN_HUMAN-SELEAQTGLQILQTGVGQR-y8+)
  \item PVRL1_HUMAN.ITQVTWQK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PVRL1_HUMAN-ITQVTWQK-y6+)
  \item PVRL1_HUMAN.ITQVTWQK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (PVRL1_HUMAN-ITQVTWQK-y7+)
  \item RET4_HUMAN.DPNGLPPEAQK.B5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (RET4_HUMAN-DPNGLPPEAQK-b5+)
  \item RET4_HUMAN.DPNGLPPEAQK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (RET4_HUMAN-DPNGLPPEAQK-y6+)
  \item RET4_HUMAN.YWGVASFLQK.B2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (RET4_HUMAN-YWGVASFLQK-b2+)
  \item RET4_HUMAN.YWGVASFLQK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (RET4_HUMAN-YWGVASFLQK-y8+)
  \item SAMP_HUMAN.IVLGQEQDSYGGK.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SAMP_HUMAN-IVLGQEQDSYGGK-y3+)
  \item SAMP_HUMAN.IVLGQEQDSYGGK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SAMP_HUMAN-IVLGQEQDSYGGK-y6+)
  \item SAMP_HUMAN.VGEYSLYIGR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SAMP_HUMAN-VGEYSLYIGR-y6+)
  \item SAMP_HUMAN.VGEYSLYIGR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SAMP_HUMAN-VGEYSLYIGR-y7+)
  \item SCG1_HUMAN.GEAGAPGEEDIQGPTK.B3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SCG1_HUMAN-GEAGAPGEEDIQGPTK-b3+)
  \item SCG1_HUMAN.GEAGAPGEEDIQGPTK.B5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SCG1_HUMAN-GEAGAPGEEDIQGPTK-b5+)
  \item SCG1_HUMAN.HLEEPGETQNAFLNER.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SCG1_HUMAN-HLEEPGETQNAFLNER-y5+)
  \item SCG1_HUMAN.HLEEPGETQNAFLNER.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SCG1_HUMAN-HLEEPGETQNAFLNER-y7+)
  \item SCG1_HUMAN.NYLNYGEEGAPGK.B3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SCG1_HUMAN-NYLNYGEEGAPGK-b3+)
  \item SCG1_HUMAN.NYLNYGEEGAPGK.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SCG1_HUMAN-NYLNYGEEGAPGK-y3+)
  \item SCG1_HUMAN.SSQGGSLPSEEK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SCG1_HUMAN-SSQGGSLPSEEK-y8+)
  \item SCG1_HUMAN.SSQGGSLPSEEK.Y9. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SCG1_HUMAN-SSQGGSLPSEEK-y9+)
  \item SCG3_HUMAN.ELSAERPLNEQIAEAEEDK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SCG3_HUMAN-ELSAERPLNEQIAEAEEDK-y7+)
  \item SCG3_HUMAN.ELSAERPLNEQIAEAEEDK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SCG3_HUMAN-ELSAERPLNEQIAEAEEDK-y8+)
  \item SCG3_HUMAN.LNVEDVDSTK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SCG3_HUMAN-LNVEDVDSTK-y7+)
  \item SCG3_HUMAN.LNVEDVDSTK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SCG3_HUMAN-LNVEDVDSTK-y8+)
  \item SE6L1_HUMAN.ETGTPIWTSR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SE6L1_HUMAN-ETGTPIWTSR-y4+)
  \item SE6L1_HUMAN.ETGTPIWTSR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SE6L1_HUMAN-ETGTPIWTSR-y6+)
  \item SHBG_HUMAN.IALGGLLFPASNLR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SHBG_HUMAN-IALGGLLFPASNLR-y6+)
  \item SHBG_HUMAN.IALGGLLFPASNLR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SHBG_HUMAN-IALGGLLFPASNLR-y7+)
  \item SHSA5_HUMAN.FVWSEER.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SHSA5_HUMAN-FVWSEER-y4+)
  \item SHSA5_HUMAN.FVWSEER.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SHSA5_HUMAN-FVWSEER-y5+)
  \item SLIK1_HUMAN.SLPVDVFAGVSLSK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SLIK1_HUMAN-SLPVDVFAGVSLSK-y7+)
  \item SLIK1_HUMAN.SLPVDVFAGVSLSK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SLIK1_HUMAN-SLPVDVFAGVSLSK-y8+)
  \item SODC_HUMAN.GDGPVQGIINFEQK.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SODC_HUMAN-GDGPVQGIINFEQK-y6+)
  \item SODC_HUMAN.GDGPVQGIINFEQK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SODC_HUMAN-GDGPVQGIINFEQK-y8+)
  \item SODC_HUMAN.HVGDLGNVTADK.Y10. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SODC_HUMAN-HVGDLGNVTADK-y10+)
  \item SODC_HUMAN.HVGDLGNVTADK.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SODC_HUMAN-HVGDLGNVTADK-y4+)
  \item SODC_HUMAN.TLVVHEK.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SODC_HUMAN-TLVVHEK-y3+)
  \item SODC_HUMAN.TLVVHEK.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SODC_HUMAN-TLVVHEK-y4+)
  \item SODE_HUMAN.AVVVHAGEDDLGR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SODE_HUMAN-AVVVHAGEDDLGR-y7+)
  \item SODE_HUMAN.AVVVHAGEDDLGR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SODE_HUMAN-AVVVHAGEDDLGR-y8+)
  \item SODE_HUMAN.VTGVVLFR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SODE_HUMAN-VTGVVLFR-y6+)
  \item SODE_HUMAN.VTGVVLFR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SODE_HUMAN-VTGVVLFR-y7+)
  \item SORC3_HUMAN.AGPELLPQQGGGR.B6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SORC3_HUMAN-AGPELLPQQGGGR-b6+)
  \item SORC3_HUMAN.AGPELLPQQGGGR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SORC3_HUMAN-AGPELLPQQGGGR-y7+)
  \item SORC3_HUMAN.AQPPITQER.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SORC3_HUMAN-AQPPITQER-y6+)
  \item SORC3_HUMAN.AQPPITQER.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (SORC3_HUMAN-AQPPITQER-y7+)
  \item THRB_HUMAN.YGFYTHVFR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (THRB_HUMAN-YGFYTHVFR-y5+)
  \item THRB_HUMAN.YGFYTHVFR.Y6. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (THRB_HUMAN-YGFYTHVFR-y6+)
  \item TIMP1_HUMAN.GFQALGDAADIR.Y2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (TIMP1_HUMAN-GFQALGDAADIR-y2+)
  \item TIMP1_HUMAN.GFQALGDAADIR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (TIMP1_HUMAN-GFQALGDAADIR-y7+)
  \item TIMP1_HUMAN.SEEFLIAGK.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (TIMP1_HUMAN-SEEFLIAGK-y3+)
  \item TIMP1_HUMAN.SEEFLIAGK.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (TIMP1_HUMAN-SEEFLIAGK-y7+)
  \item TPP1_HUMAN.LFGGNFAHQASVAR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (TPP1_HUMAN-LFGGNFAHQASVAR-y7+)
  \item TPP1_HUMAN.LFGGNFAHQASVAR.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (TPP1_HUMAN-LFGGNFAHQASVAR-y8+)
  \item TPP1_HUMAN.LSELVQAVSDPSSPQYGK.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (TPP1_HUMAN-LSELVQAVSDPSSPQYGK-y5+)
  \item TPP1_HUMAN.LSELVQAVSDPSSPQYGK.Y8. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (TPP1_HUMAN-LSELVQAVSDPSSPQYGK-y8+)
  \item TPP1_HUMAN.LYQQHGAGLFDVTR.Y3. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (TPP1_HUMAN-LYQQHGAGLFDVTR-y3+)
  \item TPP1_HUMAN.LYQQHGAGLFDVTR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (TPP1_HUMAN-LYQQHGAGLFDVTR-y5+)
  \item TREM2_HUMAN.VLVEVLADPLDHR.Y2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (TREM2_HUMAN-VLVEVLADPLDHR-y2+)
  \item TREM2_HUMAN.VLVEVLADPLDHR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (TREM2_HUMAN-VLVEVLADPLDHR-y5+)
  \item TREM2_HUMAN.VVSTHNLWLLSFLR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (TREM2_HUMAN-VVSTHNLWLLSFLR-y4+)
  \item TREM2_HUMAN.VVSTHNLWLLSFLR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (TREM2_HUMAN-VVSTHNLWLLSFLR-y5+)
  \item TTHY_HUMAN.VEIDTK.Y2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (TTHY_HUMAN-VEIDTK-y2+)
  \item TTHY_HUMAN.VEIDTK.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (TTHY_HUMAN-VEIDTK-y4+)
  \item UBB_HUMAN.ESTLHLVLR.Y4. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (UBB_HUMAN-ESTLHLVLR-y4+)
  \item UBB_HUMAN.ESTLHLVLR.Y5. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (UBB_HUMAN-ESTLHLVLR-y5+)
  \item VTDB_HUMAN.HLSLLTTLSNR.B2. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (VTDB_HUMAN-HLSLLTTLSNR-b2+)
  \item VTDB_HUMAN.HLSLLTTLSNR.Y7. -4 log2 scale/unitless Normalized semiquantitative data for MS transition (VTDB_HUMAN-HLSLLTTLSNR-y7+)
}
}
\examples{
\dontrun{
describe(csfproteomics)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/CSFPROTEOMICS/BC_MRM_Data_Primer_-_Final.pdf}
}
\keyword{datasets}
