% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{datadic}
\alias{datadic}
\title{Data Dictionary}
\format{
A data frame with 19343 rows and 12 variables
}
\usage{
data(datadic)
}
\description{
Data Dictionary
}
\details{
A dataset with variables as follows:

\itemize{
  \item 
}
}
\keyword{datasets}
