% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{desikanlab}
\alias{desikanlab}
\title{Desikan Lab Polygenic Hazard Score (PHS)}
\format{
A data frame with 1520 rows and 5 variables
}
\usage{
data(desikanlab)
}
\description{
Desikan Lab Polygenic Hazard Score (PHS)
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item USERDATE -4 -4 Date record created
  \item PHS -4 -4 Polygenic Hazard Score
  \item CIR -4 -4 Cummulative Incidence Rate
}
}
\examples{
\dontrun{
describe(desikanlab)
}
}
\keyword{datasets}
