% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{devlog}
\alias{devlog}
\title{Protocol Deviations Log}
\format{
A data frame with 1752 rows and 20 variables
}
\usage{
data(devlog)
}
\description{
Protocol Deviations Log
}
\details{
A dataset with variables as follows:

\itemize{
  \item PDDATE D YYYY-MM-DD Date of deviation
  \item PDREVIEW N NA Was deviation reviewed in advance by Coordinating Center or Project Director?
  \item PDRISK N NA Assessment of risk to participant:
  \item PDTYPE N NA Deviation Type:
  \item PDCOVID N NA Was deviation related to COVID-19
  \item PDCONSENT N NA If deviation related to Informed Consent, select details:
  \item PDRAND N NA If deviation related to Enrollment, select details:
  \item PDPROC N NA If deviation related to Protocol Procedures, select details:
  \item PDASSESS N NA If deviation related to other protocol procedure, select assessment/procedure impacted:
  \item PDEXMED N NA If deviation related to Excluded Concurrent Medications, select details:
  \item PDPERSON N NA If deviation related to Site Personnel, select details:
  \item PDDESC T NA Description of Event
  \item PDVISIT T NA Select visit(s) impacted
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
}
}
\examples{
\dontrun{
describe(devlog)
}
}
\keyword{datasets}
