% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{dtiroi_robustmean_v2}
\alias{dtiroi_robustmean_v2}
\title{USC - DTI ROI Summary Measures v2 (Robust Mean)}
\format{
A data frame with 972 rows and 304 variables
}
\usage{
data(dtiroi_robustmean_v2)
}
\description{
USC - DTI ROI Summary Measures v2 (Robust Mean)
}
\details{
A dataset with variables as follows:

\itemize{
  \item COLPROT -4 -4 Study protocol of data collection
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item VERSION -4 -4 -4
  \item MANUFACTURER -4 -4 Scanner Manufacturer
  \item VOLUMES -4 -4 Number of volumes prescribed by acquisition protocol
  \item IMAGEUID -4 -4 IMAGE ID of the raw diffusion scan
  \item RUNDATE -4 -4 -4
  \item STATUS -4 -4 -4
  \item QC -4 -4 Quality check notes
  \item AD_MCP -4 AD Middle cerebellar peduncle (robust mean)
  \item AD_PCT -4 AD Pontine crossing tract (robust mean)
  \item AD_GCC -4 AD Genu of corpus callosum (robust mean)
  \item AD_BCC -4 AD Body of corpus callosum (robust mean)
  \item AD_SCC -4 AD Splenium of corpus callosum (robust mean)
  \item AD_FX -4 AD Fornix (robust mean)
  \item AD_CST_R -4 AD Corticospinal tract right (robust mean)
  \item AD_CST_L -4 AD Corticospinal tract left (robust mean)
  \item AD_ML_R -4 AD Medial lemniscus right (robust mean)
  \item AD_ML_L -4 AD Medial lemniscus left (robust mean)
  \item AD_ICP_R -4 AD Inferior cerebellar peduncle right (robust mean)
  \item AD_ICP_L -4 AD Inferior cerebellar peduncle left (robust mean)
  \item AD_SCP_R -4 AD Superior cerebellar peduncle right (robust mean)
  \item AD_SCP_L -4 AD Superior cerebellar peduncle left (robust mean)
  \item AD_CP_R -4 AD Cerebral peduncle right (robust mean)
  \item AD_CP_L -4 AD Cerebral peduncle left (robust mean)
  \item AD_ALIC_R -4 AD Anterior limb of internal capsule right (robust mean)
  \item AD_ALIC_L -4 AD Anterior limb of internal capsule left (robust mean)
  \item AD_PLIC_R -4 AD Posterior limb of internal capsule right (robust mean)
  \item AD_PLIC_L -4 AD Posterior limb of internal capsule left (robust mean)
  \item AD_RLIC_R -4 AD Retrolenticular part of internal capsule right (robust mean)
  \item AD_RLIC_L -4 AD Retrolenticular part of internal capsule left (robust mean)
  \item AD_ACR_R -4 AD Anterior corona radiata right (robust mean)
  \item AD_ACR_L -4 AD Anterior corona radiata left (robust mean)
  \item AD_SCR_R -4 AD Superior corona radiata right (robust mean)
  \item AD_SCR_L -4 AD Superior corona radiata left (robust mean)
  \item AD_PCR_R -4 AD Posterior corona radiata right (robust mean)
  \item AD_PCR_L -4 AD Posterior corona radiata left (robust mean)
  \item AD_PTR_R -4 AD Posterior thalamic radiation right (robust mean)
  \item AD_PTR_L -4 AD Posterior thalamic radiation left (robust mean)
  \item AD_SS_R -4 AD Sagittal stratum right (robust mean)
  \item AD_SS_L -4 AD Sagittal stratum left (robust mean)
  \item AD_EC_R -4 AD External capsule right (robust mean)
  \item AD_EC_L -4 AD External capsule left (robust mean)
  \item AD_CGC_R -4 AD Cingulum right (robust mean)
  \item AD_CGC_L -4 AD Cingulum left (robust mean)
  \item AD_CGH_R -4 AD Cingulum right (robust mean)
  \item AD_CGH_L -4 AD Cingulum left (robust mean)
  \item AD_FXST_R -4 AD Fornix (cres) / Stria terminalis right (robust mean)
  \item AD_FXST_L -4 AD Fornix (cres) / Stria terminalis left (robust mean)
  \item AD_SLF_R -4 AD Superior longitudinal fasciculus right (robust mean)
  \item AD_SLF_L -4 AD Superior longitudinal fasciculus left (robust mean)
  \item AD_SFO_R -4 AD Superior fronto-occipital fasciculus right (robust mean)
  \item AD_SFO_L -4 AD Superior fronto-occipital fasciculus left (robust mean)
  \item AD_UNC_R -4 AD Uncinate fasciculus right (robust mean)
  \item AD_UNC_L -4 AD Uncinate fasciculus left (robust mean)
  \item AD_TAP_R -4 AD Tapetum right (robust mean)
  \item AD_TAP_L -4 AD Tapetum left (robust mean)
  \item AD_CST -4 AD Corticospinal tract Bilateral (robust mean)
  \item AD_ML -4 AD Medial lemniscus Bilateral (robust mean)
  \item AD_ICP -4 AD Inferior cerebellar peduncle Bilateral (robust mean)
  \item AD_SCP -4 AD Superior cerebellar peduncle Bilateral (robust mean)
  \item AD_CP -4 AD Cerebral peduncle Bilateral (robust mean)
  \item AD_ALIC -4 AD Anterior limb of internal capsule Bilateral (robust mean)
  \item AD_PLIC -4 AD Posterior limb of internal capsule Bilateral (robust mean)
  \item AD_RLIC -4 AD Retrolenticular part of internal capsule Bilateral (robust mean)
  \item AD_ACR -4 AD Anterior corona radiata Bilateral (robust mean)
  \item AD_SCR -4 AD Superior corona radiata Bilateral (robust mean)
  \item AD_PCR -4 AD Posterior corona radiata Bilateral (robust mean)
  \item AD_PTR -4 AD Posterior thalamic radiation Bilateral (robust mean)
  \item AD_SS -4 AD Sagittal stratum Bilateral (robust mean)
  \item AD_EC -4 AD External capsule Bilateral (robust mean)
  \item AD_CGC -4 AD Cingulum Bilateral (robust mean)
  \item AD_CGH -4 AD Cingulum Bilateral (robust mean)
  \item AD_FXST -4 AD Fornix (cres) / Stria terminalis Bilateral (robust mean)
  \item AD_SLF -4 AD Superior longitudinal fasciculus Bilateral (robust mean)
  \item AD_SFO -4 AD Superior fronto-occipital fasciculus Bilateral (robust mean)
  \item AD_UNC -4 AD Uncinate fasciculus Bilateral (robust mean)
  \item AD_TAP -4 AD Tapetum Bilateral (robust mean)
  \item AD_IC -4 AD Full internal capsule (robust mean)
  \item AD_CC -4 AD Full corpus callosum (robust mean)
  \item AD_CR -4 AD Full corona radiata (robust mean)
  \item AD_FULLWM -4 AD Full white matter (robust mean)
  \item FA_MCP -4 FA Middle cerebellar peduncle (robust mean)
  \item FA_PCT -4 FA Pontine crossing tract (robust mean)
  \item FA_GCC -4 FA Genu of corpus callosum (robust mean)
  \item FA_BCC -4 FA Body of corpus callosum (robust mean)
  \item FA_SCC -4 FA Splenium of corpus callosum (robust mean)
  \item FA_FX -4 FA Fornix (robust mean)
  \item FA_CST_R -4 FA Corticospinal tract right (robust mean)
  \item FA_CST_L -4 FA Corticospinal tract left (robust mean)
  \item FA_ML_R -4 FA Medial lemniscus right (robust mean)
  \item FA_ML_L -4 FA Medial lemniscus left (robust mean)
  \item FA_ICP_R -4 FA Inferior cerebellar peduncle right (robust mean)
  \item FA_ICP_L -4 FA Inferior cerebellar peduncle left (robust mean)
  \item FA_SCP_R -4 FA Superior cerebellar peduncle right (robust mean)
  \item FA_SCP_L -4 FA Superior cerebellar peduncle left (robust mean)
  \item FA_CP_R -4 FA Cerebral peduncle right (robust mean)
  \item FA_CP_L -4 FA Cerebral peduncle left (robust mean)
  \item FA_ALIC_R -4 FA Anterior limb of internal capsule right (robust mean)
  \item FA_ALIC_L -4 FA Anterior limb of internal capsule left (robust mean)
  \item FA_PLIC_R -4 FA Posterior limb of internal capsule right (robust mean)
  \item FA_PLIC_L -4 FA Posterior limb of internal capsule left (robust mean)
  \item FA_RLIC_R -4 FA Retrolenticular part of internal capsule right (robust mean)
  \item FA_RLIC_L -4 FA Retrolenticular part of internal capsule left (robust mean)
  \item FA_ACR_R -4 FA Anterior corona radiata right (robust mean)
  \item FA_ACR_L -4 FA Anterior corona radiata left (robust mean)
  \item FA_SCR_R -4 FA Superior corona radiata right (robust mean)
  \item FA_SCR_L -4 FA Superior corona radiata left (robust mean)
  \item FA_PCR_R -4 FA Posterior corona radiata right (robust mean)
  \item FA_PCR_L -4 FA Posterior corona radiata left (robust mean)
  \item FA_PTR_R -4 FA Posterior thalamic radiation right (robust mean)
  \item FA_PTR_L -4 FA Posterior thalamic radiation left (robust mean)
  \item FA_SS_R -4 FA Sagittal stratum right (robust mean)
  \item FA_SS_L -4 FA Sagittal stratum left (robust mean)
  \item FA_EC_R -4 FA External capsule right (robust mean)
  \item FA_EC_L -4 FA External capsule left (robust mean)
  \item FA_CGC_R -4 FA Cingulum right (robust mean)
  \item FA_CGC_L -4 FA Cingulum left (robust mean)
  \item FA_CGH_R -4 FA Cingulum right (robust mean)
  \item FA_CGH_L -4 FA Cingulum left (robust mean)
  \item FA_FXST_R -4 FA Fornix (cres) / Stria terminalis right (robust mean)
  \item FA_FXST_L -4 FA Fornix (cres) / Stria terminalis left (robust mean)
  \item FA_SLF_R -4 FA Superior longitudinal fasciculus right (robust mean)
  \item FA_SLF_L -4 FA Superior longitudinal fasciculus left (robust mean)
  \item FA_SFO_R -4 FA Superior fronto-occipital fasciculus right (robust mean)
  \item FA_SFO_L -4 FA Superior fronto-occipital fasciculus left (robust mean)
  \item FA_UNC_R -4 FA Uncinate fasciculus right (robust mean)
  \item FA_UNC_L -4 FA Uncinate fasciculus left (robust mean)
  \item FA_TAP_R -4 FA Tapetum right (robust mean)
  \item FA_TAP_L -4 FA Tapetum left (robust mean)
  \item FA_CST -4 FA Corticospinal tract Bilateral (robust mean)
  \item FA_ML -4 FA Medial lemniscus Bilateral (robust mean)
  \item FA_ICP -4 FA Inferior cerebellar peduncle Bilateral (robust mean)
  \item FA_SCP -4 FA Superior cerebellar peduncle Bilateral (robust mean)
  \item FA_CP -4 FA Cerebral peduncle Bilateral (robust mean)
  \item FA_ALIC -4 FA Anterior limb of internal capsule Bilateral (robust mean)
  \item FA_PLIC -4 FA Posterior limb of internal capsule Bilateral (robust mean)
  \item FA_RLIC -4 FA Retrolenticular part of internal capsule Bilateral (robust mean)
  \item FA_ACR -4 FA Anterior corona radiata Bilateral (robust mean)
  \item FA_SCR -4 FA Superior corona radiata Bilateral (robust mean)
  \item FA_PCR -4 FA Posterior corona radiata Bilateral (robust mean)
  \item FA_PTR -4 FA Posterior thalamic radiation Bilateral (robust mean)
  \item FA_SS -4 FA Sagittal stratum Bilateral (robust mean)
  \item FA_EC -4 FA External capsule Bilateral (robust mean)
  \item FA_CGC -4 FA Cingulum Bilateral (robust mean)
  \item FA_CGH -4 FA Cingulum Bilateral (robust mean)
  \item FA_FXST -4 FA Fornix (cres) / Stria terminalis Bilateral (robust mean)
  \item FA_SLF -4 FA Superior longitudinal fasciculus Bilateral (robust mean)
  \item FA_SFO -4 FA Superior fronto-occipital fasciculus Bilateral (robust mean)
  \item FA_UNC -4 FA Uncinate fasciculus Bilateral (robust mean)
  \item FA_TAP -4 FA Tapetum Bilateral (robust mean)
  \item FA_IC -4 FA Full internal capsule (robust mean)
  \item FA_CC -4 FA Full corpus callosum (robust mean)
  \item FA_CR -4 FA Full corona radiata (robust mean)
  \item FA_FULLWM -4 FA Full white matter (robust mean)
  \item MD_MCP -4 MD Middle cerebellar peduncle (robust mean)
  \item MD_PCT -4 MD Pontine crossing tract (robust mean)
  \item MD_GCC -4 MD Genu of corpus callosum (robust mean)
  \item MD_BCC -4 MD Body of corpus callosum (robust mean)
  \item MD_SCC -4 MD Splenium of corpus callosum (robust mean)
  \item MD_FX -4 MD Fornix (robust mean)
  \item MD_CST_R -4 MD Corticospinal tract right (robust mean)
  \item MD_CST_L -4 MD Corticospinal tract left (robust mean)
  \item MD_ML_R -4 MD Medial lemniscus right (robust mean)
  \item MD_ML_L -4 MD Medial lemniscus left (robust mean)
  \item MD_ICP_R -4 MD Inferior cerebellar peduncle right (robust mean)
  \item MD_ICP_L -4 MD Inferior cerebellar peduncle left (robust mean)
  \item MD_SCP_R -4 MD Superior cerebellar peduncle right (robust mean)
  \item MD_SCP_L -4 MD Superior cerebellar peduncle left (robust mean)
  \item MD_CP_R -4 MD Cerebral peduncle right (robust mean)
  \item MD_CP_L -4 MD Cerebral peduncle left (robust mean)
  \item MD_ALIC_R -4 MD Anterior limb of internal capsule right (robust mean)
  \item MD_ALIC_L -4 MD Anterior limb of internal capsule left (robust mean)
  \item MD_PLIC_R -4 MD Posterior limb of internal capsule right (robust mean)
  \item MD_PLIC_L -4 MD Posterior limb of internal capsule left (robust mean)
  \item MD_RLIC_R -4 MD Retrolenticular part of internal capsule right (robust mean)
  \item MD_RLIC_L -4 MD Retrolenticular part of internal capsule left (robust mean)
  \item MD_ACR_R -4 MD Anterior corona radiata right (robust mean)
  \item MD_ACR_L -4 MD Anterior corona radiata left (robust mean)
  \item MD_SCR_R -4 MD Superior corona radiata right (robust mean)
  \item MD_SCR_L -4 MD Superior corona radiata left (robust mean)
  \item MD_PCR_R -4 MD Posterior corona radiata right (robust mean)
  \item MD_PCR_L -4 MD Posterior corona radiata left (robust mean)
  \item MD_PTR_R -4 MD Posterior thalamic radiation right (robust mean)
  \item MD_PTR_L -4 MD Posterior thalamic radiation left (robust mean)
  \item MD_SS_R -4 MD Sagittal stratum right (robust mean)
  \item MD_SS_L -4 MD Sagittal stratum left (robust mean)
  \item MD_EC_R -4 MD External capsule right (robust mean)
  \item MD_EC_L -4 MD External capsule left (robust mean)
  \item MD_CGC_R -4 MD Cingulum right (robust mean)
  \item MD_CGC_L -4 MD Cingulum left (robust mean)
  \item MD_CGH_R -4 MD Cingulum right (robust mean)
  \item MD_CGH_L -4 MD Cingulum left (robust mean)
  \item MD_FXST_R -4 MD Fornix (cres) / Stria terminalis right (robust mean)
  \item MD_FXST_L -4 MD Fornix (cres) / Stria terminalis left (robust mean)
  \item MD_SLF_R -4 MD Superior longitudinal fasciculus right (robust mean)
  \item MD_SLF_L -4 MD Superior longitudinal fasciculus left (robust mean)
  \item MD_SFO_R -4 MD Superior fronto-occipital fasciculus right (robust mean)
  \item MD_SFO_L -4 MD Superior fronto-occipital fasciculus left (robust mean)
  \item MD_UNC_R -4 MD Uncinate fasciculus right (robust mean)
  \item MD_UNC_L -4 MD Uncinate fasciculus left (robust mean)
  \item MD_TAP_R -4 MD Tapetum right (robust mean)
  \item MD_TAP_L -4 MD Tapetum left (robust mean)
  \item MD_CST -4 MD Corticospinal tract Bilateral (robust mean)
  \item MD_ML -4 MD Medial lemniscus Bilateral (robust mean)
  \item MD_ICP -4 MD Inferior cerebellar peduncle Bilateral (robust mean)
  \item MD_SCP -4 MD Superior cerebellar peduncle Bilateral (robust mean)
  \item MD_CP -4 MD Cerebral peduncle Bilateral (robust mean)
  \item MD_ALIC -4 MD Anterior limb of internal capsule Bilateral (robust mean)
  \item MD_PLIC -4 MD Posterior limb of internal capsule Bilateral (robust mean)
  \item MD_RLIC -4 MD Retrolenticular part of internal capsule Bilateral (robust mean)
  \item MD_ACR -4 MD Anterior corona radiata Bilateral (robust mean)
  \item MD_SCR -4 MD Superior corona radiata Bilateral (robust mean)
  \item MD_PCR -4 MD Posterior corona radiata Bilateral (robust mean)
  \item MD_PTR -4 MD Posterior thalamic radiation Bilateral (robust mean)
  \item MD_SS -4 MD Sagittal stratum Bilateral (robust mean)
  \item MD_EC -4 MD External capsule Bilateral (robust mean)
  \item MD_CGC -4 MD Cingulum Bilateral (robust mean)
  \item MD_CGH -4 MD Cingulum Bilateral (robust mean)
  \item MD_FXST -4 MD Fornix (cres) / Stria terminalis Bilateral (robust mean)
  \item MD_SLF -4 MD Superior longitudinal fasciculus Bilateral (robust mean)
  \item MD_SFO -4 MD Superior fronto-occipital fasciculus Bilateral (robust mean)
  \item MD_UNC -4 MD Uncinate fasciculus Bilateral (robust mean)
  \item MD_TAP -4 MD Tapetum Bilateral (robust mean)
  \item MD_IC -4 MD Full internal capsule (robust mean)
  \item MD_CC -4 MD Full corpus callosum (robust mean)
  \item MD_CR -4 MD Full corona radiata (robust mean)
  \item MD_FULLWM -4 MD Full white matter (robust mean)
  \item RD_MCP -4 RD Middle cerebellar peduncle (robust mean)
  \item RD_PCT -4 RD Pontine crossing tract (robust mean)
  \item RD_GCC -4 RD Genu of corpus callosum (robust mean)
  \item RD_BCC -4 RD Body of corpus callosum (robust mean)
  \item RD_SCC -4 RD Splenium of corpus callosum (robust mean)
  \item RD_FX -4 RD Fornix (robust mean)
  \item RD_CST_R -4 RD Corticospinal tract right (robust mean)
  \item RD_CST_L -4 RD Corticospinal tract left (robust mean)
  \item RD_ML_R -4 RD Medial lemniscus right (robust mean)
  \item RD_ML_L -4 RD Medial lemniscus left (robust mean)
  \item RD_ICP_R -4 RD Inferior cerebellar peduncle right (robust mean)
  \item RD_ICP_L -4 RD Inferior cerebellar peduncle left (robust mean)
  \item RD_SCP_R -4 RD Superior cerebellar peduncle right (robust mean)
  \item RD_SCP_L -4 RD Superior cerebellar peduncle left (robust mean)
  \item RD_CP_R -4 RD Cerebral peduncle right (robust mean)
  \item RD_CP_L -4 RD Cerebral peduncle left (robust mean)
  \item RD_ALIC_R -4 RD Anterior limb of internal capsule right (robust mean)
  \item RD_ALIC_L -4 RD Anterior limb of internal capsule left (robust mean)
  \item RD_PLIC_R -4 RD Posterior limb of internal capsule right (robust mean)
  \item RD_PLIC_L -4 RD Posterior limb of internal capsule left (robust mean)
  \item RD_RLIC_R -4 RD Retrolenticular part of internal capsule right (robust mean)
  \item RD_RLIC_L -4 RD Retrolenticular part of internal capsule left (robust mean)
  \item RD_ACR_R -4 RD Anterior corona radiata right (robust mean)
  \item RD_ACR_L -4 RD Anterior corona radiata left (robust mean)
  \item RD_SCR_R -4 RD Superior corona radiata right (robust mean)
  \item RD_SCR_L -4 RD Superior corona radiata left (robust mean)
  \item RD_PCR_R -4 RD Posterior corona radiata right (robust mean)
  \item RD_PCR_L -4 RD Posterior corona radiata left (robust mean)
  \item RD_PTR_R -4 RD Posterior thalamic radiation right (robust mean)
  \item RD_PTR_L -4 RD Posterior thalamic radiation left (robust mean)
  \item RD_SS_R -4 RD Sagittal stratum right (robust mean)
  \item RD_SS_L -4 RD Sagittal stratum left (robust mean)
  \item RD_EC_R -4 RD External capsule right (robust mean)
  \item RD_EC_L -4 RD External capsule left (robust mean)
  \item RD_CGC_R -4 RD Cingulum right (robust mean)
  \item RD_CGC_L -4 RD Cingulum left (robust mean)
  \item RD_CGH_R -4 RD Cingulum right (robust mean)
  \item RD_CGH_L -4 RD Cingulum left (robust mean)
  \item RD_FXST_R -4 RD Fornix (cres) / Stria terminalis right (robust mean)
  \item RD_FXST_L -4 RD Fornix (cres) / Stria terminalis left (robust mean)
  \item RD_SLF_R -4 RD Superior longitudinal fasciculus right (robust mean)
  \item RD_SLF_L -4 RD Superior longitudinal fasciculus left (robust mean)
  \item RD_SFO_R -4 RD Superior fronto-occipital fasciculus right (robust mean)
  \item RD_SFO_L -4 RD Superior fronto-occipital fasciculus left (robust mean)
  \item RD_UNC_R -4 RD Uncinate fasciculus right (robust mean)
  \item RD_UNC_L -4 RD Uncinate fasciculus left (robust mean)
  \item RD_TAP_R -4 RD Tapetum right (robust mean)
  \item RD_TAP_L -4 RD Tapetum left (robust mean)
  \item RD_CST -4 RD Corticospinal tract Bilateral (robust mean)
  \item RD_ML -4 RD Medial lemniscus Bilateral (robust mean)
  \item RD_ICP -4 RD Inferior cerebellar peduncle Bilateral (robust mean)
  \item RD_SCP -4 RD Superior cerebellar peduncle Bilateral (robust mean)
  \item RD_CP -4 RD Cerebral peduncle Bilateral (robust mean)
  \item RD_ALIC -4 RD Anterior limb of internal capsule Bilateral (robust mean)
  \item RD_PLIC -4 RD Posterior limb of internal capsule Bilateral (robust mean)
  \item RD_RLIC -4 RD Retrolenticular part of internal capsule Bilateral (robust mean)
  \item RD_ACR -4 RD Anterior corona radiata Bilateral (robust mean)
  \item RD_SCR -4 RD Superior corona radiata Bilateral (robust mean)
  \item RD_PCR -4 RD Posterior corona radiata Bilateral (robust mean)
  \item RD_PTR -4 RD Posterior thalamic radiation Bilateral (robust mean)
  \item RD_SS -4 RD Sagittal stratum Bilateral (robust mean)
  \item RD_EC -4 RD External capsule Bilateral (robust mean)
  \item RD_CGC -4 RD Cingulum Bilateral (robust mean)
  \item RD_CGH -4 RD Cingulum Bilateral (robust mean)
  \item RD_FXST -4 RD Fornix (cres) / Stria terminalis Bilateral (robust mean)
  \item RD_SLF -4 RD Superior longitudinal fasciculus Bilateral (robust mean)
  \item RD_SFO -4 RD Superior fronto-occipital fasciculus Bilateral (robust mean)
  \item RD_UNC -4 RD Uncinate fasciculus Bilateral (robust mean)
  \item RD_TAP -4 RD Tapetum Bilateral (robust mean)
  \item RD_IC -4 RD Full internal capsule (robust mean)
  \item RD_CC -4 RD Full corpus callosum (robust mean)
  \item RD_CR -4 RD Full corona radiata (robust mean)
  \item RD_FULLWM -4 RD Full white matter (robust mean)
}
}
\examples{
\dontrun{
describe(dtiroi_robustmean_v2)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/DTIROI_ROBUSTMEAN_V2/USC_-_DTI_ROI_Summary_Measures_Methods_v2_[ADNI3].pdf}
}
\keyword{datasets}
