% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{dtiroi_v1}
\alias{dtiroi_v1}
\title{USC - DTI ROI Summary Measures v1}
\format{
A data frame with 976 rows and 241 variables
}
\usage{
data(dtiroi_v1)
}
\description{
USC - DTI ROI Summary Measures v1
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4  Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4  Examination Date
  \item VERSION -4  Version
  \item VISNAME -4 -4 -4
  \item VISITNO -4  Visit
  \item LONIUID_1 -4  LONI Image ID for FA
  \item LONIUID_2 -4  LONI Image ID for MD
  \item LONIUID_3 -4  LONI Image ID for RD
  \item LONIUID_4 -4  LONI Image ID for AD
  \item RUNDATE -4  Date
  \item STATUS -4  Status
  \item AD_CST_L -4 AD Corticospinal tract left 
  \item AD_CST_R -4 AD Corticospinal tract right 
  \item AD_ICP_L -4 AD Inferior cerebellar peduncle left 
  \item AD_ICP_R -4 AD Inferior cerebellar peduncle right 
  \item AD_ML_L -4 AD Medial lemniscus left 
  \item AD_ML_R -4 AD Medial lemniscus right 
  \item AD_SCP_L -4 AD Superior cerebellar peduncle left 
  \item AD_SCP_R -4 AD Superior cerebellar peduncle right 
  \item AD_CP_L -4 AD Cerebral peduncle left 
  \item AD_CP_R -4 AD Cerebral peduncle right 
  \item AD_ALIC_L -4 AD Anterior limb of internal capsule left 
  \item AD_ALIC_R -4 AD Anterior limb of internal capsule right 
  \item AD_PLIC_L -4 AD Posterior limb of internal capsule left 
  \item AD_PLIC_R -4 AD Posterior limb of internal capsule right 
  \item AD_PTR_L -4 AD Posterior thalamic radiation left 
  \item AD_PTR_R -4 AD Posterior thalamic radiation right 
  \item AD_ACR_L -4 AD Anterior corona radiata left 
  \item AD_ACR_R -4 AD Anterior corona radiata right 
  \item AD_SCR_L -4 AD Superior corona radiata left 
  \item AD_SCR_R -4 AD Superior corona radiata right 
  \item AD_PCR_L -4 AD Posterior corona radiata left 
  \item AD_PCR_R -4 AD Posterior corona radiata right 
  \item AD_CGC_L -4 AD Cingulum left 
  \item AD_CGC_R -4 AD Cingulum right 
  \item AD_CGH_L -4 AD Cingulum (hippocampus) left 
  \item AD_CGH_R -4 AD Cingulum (hippocampus) right 
  \item AD_FX_ST_L -4 AD Fornix (cres) / Stria terminalis left 
  \item AD_FX_ST_R -4 AD Fornix (cres) / Stria terminalis right 
  \item AD_SLF_L -4 AD Superior longitudinal fasciculus left 
  \item AD_SLF_R -4 AD Superior longitudinal fasciculus right 
  \item AD_SFO_L -4 AD Superior fronto-occipital fasciculus left 
  \item AD_SFO_R -4 AD Superior fronto-occipital fasciculus right 
  \item AD_IFO_L -4 AD Inferior fronto-occipital fasciculus left 
  \item AD_IFO_R -4 AD Inferior fronto-occipital fasciculus right 
  \item AD_SS_L -4 AD Sagittal stratum left 
  \item AD_SS_R -4 AD Sagittal stratum right 
  \item AD_EC_L -4 AD External capsule left 
  \item AD_EC_R -4 AD External capsule right 
  \item AD_UNC_L -4 AD Uncinate fasciculus left 
  \item AD_UNC_R -4 AD Uncinate fasciculus right 
  \item AD_FX_L -4 AD Fornix  left 
  \item AD_FX_R -4 AD Fornix  right 
  \item AD_GCC_L -4 AD Genu of corpus callosum left 
  \item AD_GCC_R -4 AD Genu of corpus callosum right 
  \item AD_BCC_L -4 AD Body of corpus callosum  left 
  \item AD_BCC_R -4 AD Body of corpus callosum  right 
  \item AD_SCC_L -4 AD Splenium of corpus callosum left 
  \item AD_SCC_R -4 AD Splenium of corpus callosum right 
  \item AD_RLIC_L -4 AD Retrolenticular part of internal capsule left 
  \item AD_RLIC_R -4 AD Retrolenticular part of internal capsule right 
  \item AD_TAP_L -4 AD Tapatum left 
  \item AD_TAP_R -4 AD Tapatum right 
  \item AD_SUMGCC -4 AD Bilateral genu of the corpus callosum 
  \item AD_SUMBCC -4 AD Bilateral body of  the corpus callosum 
  \item AD_SUMSCC -4 AD Bilateral splenium of  the corpus callosum 
  \item AD_SUMCC -4 AD Bilateral full corpus callosum 
  \item AD_SUMFX -4 AD Bilateral fornix 
  \item FA_CST_L -4 FA Corticospinal tract left 
  \item FA_CST_R -4 FA Corticospinal tract right 
  \item FA_ICP_L -4 FA Inferior cerebellar peduncle left 
  \item FA_ICP_R -4 FA Inferior cerebellar peduncle right 
  \item FA_ML_L -4 FA Medial lemniscus left 
  \item FA_ML_R -4 FA Medial lemniscus right 
  \item FA_SCP_L -4 FA Superior cerebellar peduncle left 
  \item FA_SCP_R -4 FA Superior cerebellar peduncle right 
  \item FA_CP_L -4 FA Cerebral peduncle left 
  \item FA_CP_R -4 FA Cerebral peduncle right 
  \item FA_ALIC_L -4 FA Anterior limb of internal capsule left 
  \item FA_ALIC_R -4 FA Anterior limb of internal capsule right 
  \item FA_PLIC_L -4 FA Posterior limb of internal capsule left 
  \item FA_PLIC_R -4 FA Posterior limb of internal capsule right 
  \item FA_PTR_L -4 FA Posterior thalamic radiation left 
  \item FA_PTR_R -4 FA Posterior thalamic radiation right 
  \item FA_ACR_L -4 FA Anterior corona radiata left 
  \item FA_ACR_R -4 FA Anterior corona radiata right 
  \item FA_SCR_L -4 FA Superior corona radiata left 
  \item FA_SCR_R -4 FA Superior corona radiata right 
  \item FA_PCR_L -4 FA Posterior corona radiata left 
  \item FA_PCR_R -4 FA Posterior corona radiata right 
  \item FA_CGC_L -4 FA Cingulum left 
  \item FA_CGC_R -4 FA Cingulum right 
  \item FA_CGH_L -4 FA Cingulum (hippocampus) left 
  \item FA_CGH_R -4 FA Cingulum (hippocampus) right 
  \item FA_FX_ST_L -4 FA Fornix (cres) / Stria terminalis left 
  \item FA_FX_ST_R -4 FA Fornix (cres) / Stria terminalis right 
  \item FA_SLF_L -4 FA Superior longitudinal fasciculus left 
  \item FA_SLF_R -4 FA Superior longitudinal fasciculus right 
  \item FA_SFO_L -4 FA Superior fronto-occipital fasciculus left 
  \item FA_SFO_R -4 FA Superior fronto-occipital fasciculus right 
  \item FA_IFO_L -4 FA Inferior fronto-occipital fasciculus left 
  \item FA_IFO_R -4 FA Inferior fronto-occipital fasciculus right 
  \item FA_SS_L -4 FA Sagittal stratum left 
  \item FA_SS_R -4 FA Sagittal stratum right 
  \item FA_EC_L -4 FA External capsule left 
  \item FA_EC_R -4 FA External capsule right 
  \item FA_UNC_L -4 FA Uncinate fasciculus left 
  \item FA_UNC_R -4 FA Uncinate fasciculus right 
  \item FA_FX_L -4 FA Fornix  left 
  \item FA_FX_R -4 FA Fornix  right 
  \item FA_GCC_L -4 FA Genu of corpus callosum left 
  \item FA_GCC_R -4 FA Genu of corpus callosum right 
  \item FA_BCC_L -4 FA Body of corpus callosum  left 
  \item FA_BCC_R -4 FA Body of corpus callosum  right 
  \item FA_SCC_L -4 FA Splenium of corpus callosum left 
  \item FA_SCC_R -4 FA Splenium of corpus callosum right 
  \item FA_RLIC_L -4 FA Retrolenticular part of internal capsule left 
  \item FA_RLIC_R -4 FA Retrolenticular part of internal capsule right 
  \item FA_TAP_L -4 FA Tapatum left 
  \item FA_TAP_R -4 FA Tapatum right 
  \item FA_SUMGCC -4 FA Bilateral genu of the corpus callosum 
  \item FA_SUMBCC -4 FA Bilateral body of  the corpus callosum 
  \item FA_SUMSCC -4 FA Bilateral splenium of  the corpus callosum 
  \item FA_SUMCC -4 FA Bilateral full corpus callosum 
  \item FA_SUMFX -4 FA Bilateral fornix 
  \item MD_CST_L -4 MD Corticospinal tract left 
  \item MD_CST_R -4 MD Corticospinal tract right 
  \item MD_ICP_L -4 MD Inferior cerebellar peduncle left 
  \item MD_ICP_R -4 MD Inferior cerebellar peduncle right 
  \item MD_ML_L -4 MD Medial lemniscus left 
  \item MD_ML_R -4 MD Medial lemniscus right 
  \item MD_SCP_L -4 MD Superior cerebellar peduncle left 
  \item MD_SCP_R -4 MD Superior cerebellar peduncle right 
  \item MD_CP_L -4 MD Cerebral peduncle left 
  \item MD_CP_R -4 MD Cerebral peduncle right 
  \item MD_ALIC_L -4 MD Anterior limb of internal capsule left 
  \item MD_ALIC_R -4 MD Anterior limb of internal capsule right 
  \item MD_PLIC_L -4 MD Posterior limb of internal capsule left 
  \item MD_PLIC_R -4 MD Posterior limb of internal capsule right 
  \item MD_PTR_L -4 MD Posterior thalamic radiation left 
  \item MD_PTR_R -4 MD Posterior thalamic radiation right 
  \item MD_ACR_L -4 MD Anterior corona radiata left 
  \item MD_ACR_R -4 MD Anterior corona radiata right 
  \item MD_SCR_L -4 MD Superior corona radiata left 
  \item MD_SCR_R -4 MD Superior corona radiata right 
  \item MD_PCR_L -4 MD Posterior corona radiata left 
  \item MD_PCR_R -4 MD Posterior corona radiata right 
  \item MD_CGC_L -4 MD Cingulum left 
  \item MD_CGC_R -4 MD Cingulum right 
  \item MD_CGH_L -4 MD Cingulum (hippocampus) left 
  \item MD_CGH_R -4 MD Cingulum (hippocampus) right 
  \item MD_FX_ST_L -4 MD Fornix (cres) / Stria terminalis left 
  \item MD_FX_ST_R -4 MD Fornix (cres) / Stria terminalis right 
  \item MD_SLF_L -4 MD Superior longitudinal fasciculus left 
  \item MD_SLF_R -4 MD Superior longitudinal fasciculus right 
  \item MD_SFO_L -4 MD Superior fronto-occipital fasciculus left 
  \item MD_SFO_R -4 MD Superior fronto-occipital fasciculus right 
  \item MD_IFO_L -4 MD Inferior fronto-occipital fasciculus left 
  \item MD_IFO_R -4 MD Inferior fronto-occipital fasciculus right 
  \item MD_SS_L -4 MD Sagittal stratum left 
  \item MD_SS_R -4 MD Sagittal stratum right 
  \item MD_EC_L -4 MD External capsule left 
  \item MD_EC_R -4 MD External capsule right 
  \item MD_UNC_L -4 MD Uncinate fasciculus left 
  \item MD_UNC_R -4 MD Uncinate fasciculus right 
  \item MD_FX_L -4 MD Fornix  left 
  \item MD_FX_R -4 MD Fornix  right 
  \item MD_GCC_L -4 MD Genu of corpus callosum left 
  \item MD_GCC_R -4 MD Genu of corpus callosum right 
  \item MD_BCC_L -4 MD Body of corpus callosum  left 
  \item MD_BCC_R -4 MD Body of corpus callosum  right 
  \item MD_SCC_L -4 MD Splenium of corpus callosum left 
  \item MD_SCC_R -4 MD Splenium of corpus callosum right 
  \item MD_RLIC_L -4 MD Retrolenticular part of internal capsule left 
  \item MD_RLIC_R -4 MD Retrolenticular part of internal capsule right 
  \item MD_TAP_L -4 MD Tapatum left 
  \item MD_TAP_R -4 MD Tapatum right 
  \item MD_SUMGCC -4 MD Bilateral genu of the corpus callosum 
  \item MD_SUMBCC -4 MD Bilateral body of  the corpus callosum 
  \item MD_SUMSCC -4 MD Bilateral splenium of  the corpus callosum 
  \item MD_SUMCC -4 MD Bilateral full corpus callosum 
  \item MD_SUMFX -4 MD Bilateral fornix 
  \item RD_CST_L -4 RD Corticospinal tract left 
  \item RD_CST_R -4 RD Corticospinal tract right 
  \item RD_ICP_L -4 RD Inferior cerebellar peduncle left 
  \item RD_ICP_R -4 RD Inferior cerebellar peduncle right 
  \item RD_ML_L -4 RD Medial lemniscus left 
  \item RD_ML_R -4 RD Medial lemniscus right 
  \item RD_SCP_L -4 RD Superior cerebellar peduncle left 
  \item RD_SCP_R -4 RD Superior cerebellar peduncle right 
  \item RD_CP_L -4 RD Cerebral peduncle left 
  \item RD_CP_R -4 RD Cerebral peduncle right 
  \item RD_ALIC_L -4 RD Anterior limb of internal capsule left 
  \item RD_ALIC_R -4 RD Anterior limb of internal capsule right 
  \item RD_PLIC_L -4 RD Posterior limb of internal capsule left 
  \item RD_PLIC_R -4 RD Posterior limb of internal capsule right 
  \item RD_PTR_L -4 RD Posterior thalamic radiation left 
  \item RD_PTR_R -4 RD Posterior thalamic radiation right 
  \item RD_ACR_L -4 RD Anterior corona radiata left 
  \item RD_ACR_R -4 RD Anterior corona radiata right 
  \item RD_SCR_L -4 RD Superior corona radiata left 
  \item RD_SCR_R -4 RD Superior corona radiata right 
  \item RD_PCR_L -4 RD Posterior corona radiata left 
  \item RD_PCR_R -4 RD Posterior corona radiata right 
  \item RD_CGC_L -4 RD Cingulum left 
  \item RD_CGC_R -4 RD Cingulum right 
  \item RD_CGH_L -4 RD Cingulum (hippocampus) left 
  \item RD_CGH_R -4 RD Cingulum (hippocampus) right 
  \item RD_FX_ST_L -4 RD Fornix (cres) / Stria terminalis left 
  \item RD_FX_ST_R -4 RD Fornix (cres) / Stria terminalis right 
  \item RD_SLF_L -4 RD Superior longitudinal fasciculus left 
  \item RD_SLF_R -4 RD Superior longitudinal fasciculus right 
  \item RD_SFO_L -4 RD Superior fronto-occipital fasciculus left 
  \item RD_SFO_R -4 RD Superior fronto-occipital fasciculus right 
  \item RD_IFO_L -4 RD Inferior fronto-occipital fasciculus left 
  \item RD_IFO_R -4 RD Inferior fronto-occipital fasciculus right 
  \item RD_SS_L -4 RD Sagittal stratum left 
  \item RD_SS_R -4 RD Sagittal stratum right 
  \item RD_EC_L -4 RD External capsule left 
  \item RD_EC_R -4 RD External capsule right 
  \item RD_UNC_L -4 RD Uncinate fasciculus left 
  \item RD_UNC_R -4 RD Uncinate fasciculus right 
  \item RD_FX_L -4 RD Fornix  left 
  \item RD_FX_R -4 RD Fornix  right 
  \item RD_GCC_L -4 RD Genu of corpus callosum left 
  \item RD_GCC_R -4 RD Genu of corpus callosum right 
  \item RD_BCC_L -4 RD Body of corpus callosum  left 
  \item RD_BCC_R -4 RD Body of corpus callosum  right 
  \item RD_SCC_L -4 RD Splenium of corpus callosum left 
  \item RD_SCC_R -4 RD Splenium of corpus callosum right 
  \item RD_RLIC_L -4 RD Retrolenticular part of internal capsule left 
  \item RD_RLIC_R -4 RD Retrolenticular part of internal capsule right 
  \item RD_TAP_L -4 RD Tapatum left 
  \item RD_TAP_R -4 RD Tapatum right 
  \item RD_SUMGCC -4 RD Bilateral genu of the corpus callosum 
  \item RD_SUMBCC -4 RD Bilateral body of  the corpus callosum 
  \item RD_SUMSCC -4 RD Bilateral splenium of  the corpus callosum 
  \item RD_SUMCC -4 RD Bilateral full corpus callosum 
  \item RD_SUMFX -4 RD Bilateral fornix 
}
}
\examples{
\dontrun{
describe(dtiroi_v1)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/DTIROI_V1/DTI-ADNI_Methods-Thompson-Oct2012.pdf}
}
\keyword{datasets}
