% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{dtiroi_v2}
\alias{dtiroi_v2}
\title{USC - DTI ROI Summary Measures v2 (Mean)}
\format{
A data frame with 972 rows and 304 variables
}
\usage{
data(dtiroi_v2)
}
\description{
USC - DTI ROI Summary Measures v2 (Mean)
}
\details{
A dataset with variables as follows:

\itemize{
  \item COLPROT -4 -4 Study protocol of data collection
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item VERSION -4 -4 -4
  \item MANUFACTURER -4 -4 Scanner Manufacturer
  \item VOLUMES -4 -4 Number of volumes prescribed by acquisition protocol
  \item IMAGEUID -4 -4 IMAGE ID of the raw diffusion scan
  \item RUNDATE -4 -4 -4
  \item STATUS -4 -4 -4
  \item QC -4 -4 Quality check notes
  \item AD_MCP -4 -4 -4
  \item AD_PCT -4 -4 -4
  \item AD_GCC -4 -4 -4
  \item AD_BCC -4 -4 -4
  \item AD_SCC -4 -4 -4
  \item AD_FX -4 -4 -4
  \item AD_CST_R -4 -4 -4
  \item AD_CST_L -4 -4 -4
  \item AD_ML_R -4 -4 -4
  \item AD_ML_L -4 -4 -4
  \item AD_ICP_R -4 -4 -4
  \item AD_ICP_L -4 -4 -4
  \item AD_SCP_R -4 -4 -4
  \item AD_SCP_L -4 -4 -4
  \item AD_CP_R -4 -4 -4
  \item AD_CP_L -4 -4 -4
  \item AD_ALIC_R -4 -4 -4
  \item AD_ALIC_L -4 -4 -4
  \item AD_PLIC_R -4 -4 -4
  \item AD_PLIC_L -4 -4 -4
  \item AD_RLIC_R -4 -4 -4
  \item AD_RLIC_L -4 -4 -4
  \item AD_ACR_R -4 -4 -4
  \item AD_ACR_L -4 -4 -4
  \item AD_SCR_R -4 -4 -4
  \item AD_SCR_L -4 -4 -4
  \item AD_PCR_R -4 -4 -4
  \item AD_PCR_L -4 -4 -4
  \item AD_PTR_R -4 -4 -4
  \item AD_PTR_L -4 -4 -4
  \item AD_SS_R -4 -4 -4
  \item AD_SS_L -4 -4 -4
  \item AD_EC_R -4 -4 -4
  \item AD_EC_L -4 -4 -4
  \item AD_CGC_R -4 -4 -4
  \item AD_CGC_L -4 -4 -4
  \item AD_CGH_R -4 -4 -4
  \item AD_CGH_L -4 -4 -4
  \item AD_FXST_R -4 -4 -4
  \item AD_FXST_L -4 -4 -4
  \item AD_SLF_R -4 -4 -4
  \item AD_SLF_L -4 -4 -4
  \item AD_SFO_R -4 -4 -4
  \item AD_SFO_L -4 -4 -4
  \item AD_UNC_R -4 -4 -4
  \item AD_UNC_L -4 -4 -4
  \item AD_TAP_R -4 -4 -4
  \item AD_TAP_L -4 -4 -4
  \item AD_CST -4 -4 -4
  \item AD_ML -4 -4 -4
  \item AD_ICP -4 -4 -4
  \item AD_SCP -4 -4 -4
  \item AD_CP -4 -4 -4
  \item AD_ALIC -4 -4 -4
  \item AD_PLIC -4 -4 -4
  \item AD_RLIC -4 -4 -4
  \item AD_ACR -4 -4 -4
  \item AD_SCR -4 -4 -4
  \item AD_PCR -4 -4 -4
  \item AD_PTR -4 -4 -4
  \item AD_SS -4 -4 -4
  \item AD_EC -4 -4 -4
  \item AD_CGC -4 -4 -4
  \item AD_CGH -4 -4 -4
  \item AD_FXST -4 -4 -4
  \item AD_SLF -4 -4 -4
  \item AD_SFO -4 -4 -4
  \item AD_UNC -4 -4 -4
  \item AD_TAP -4 -4 -4
  \item AD_IC -4 -4 -4
  \item AD_CC -4 -4 -4
  \item AD_CR -4 -4 -4
  \item AD_FULLWM -4 -4 -4
  \item FA_MCP -4 -4 -4
  \item FA_PCT -4 -4 -4
  \item FA_GCC -4 -4 -4
  \item FA_BCC -4 -4 -4
  \item FA_SCC -4 -4 -4
  \item FA_FX -4 -4 -4
  \item FA_CST_R -4 -4 -4
  \item FA_CST_L -4 -4 -4
  \item FA_ML_R -4 -4 -4
  \item FA_ML_L -4 -4 -4
  \item FA_ICP_R -4 -4 -4
  \item FA_ICP_L -4 -4 -4
  \item FA_SCP_R -4 -4 -4
  \item FA_SCP_L -4 -4 -4
  \item FA_CP_R -4 -4 -4
  \item FA_CP_L -4 -4 -4
  \item FA_ALIC_R -4 -4 -4
  \item FA_ALIC_L -4 -4 -4
  \item FA_PLIC_R -4 -4 -4
  \item FA_PLIC_L -4 -4 -4
  \item FA_RLIC_R -4 -4 -4
  \item FA_RLIC_L -4 -4 -4
  \item FA_ACR_R -4 -4 -4
  \item FA_ACR_L -4 -4 -4
  \item FA_SCR_R -4 -4 -4
  \item FA_SCR_L -4 -4 -4
  \item FA_PCR_R -4 -4 -4
  \item FA_PCR_L -4 -4 -4
  \item FA_PTR_R -4 -4 -4
  \item FA_PTR_L -4 -4 -4
  \item FA_SS_R -4 -4 -4
  \item FA_SS_L -4 -4 -4
  \item FA_EC_R -4 -4 -4
  \item FA_EC_L -4 -4 -4
  \item FA_CGC_R -4 -4 -4
  \item FA_CGC_L -4 -4 -4
  \item FA_CGH_R -4 -4 -4
  \item FA_CGH_L -4 -4 -4
  \item FA_FXST_R -4 -4 -4
  \item FA_FXST_L -4 -4 -4
  \item FA_SLF_R -4 -4 -4
  \item FA_SLF_L -4 -4 -4
  \item FA_SFO_R -4 -4 -4
  \item FA_SFO_L -4 -4 -4
  \item FA_UNC_R -4 -4 -4
  \item FA_UNC_L -4 -4 -4
  \item FA_TAP_R -4 -4 -4
  \item FA_TAP_L -4 -4 -4
  \item FA_CST -4 -4 -4
  \item FA_ML -4 -4 -4
  \item FA_ICP -4 -4 -4
  \item FA_SCP -4 -4 -4
  \item FA_CP -4 -4 -4
  \item FA_ALIC -4 -4 -4
  \item FA_PLIC -4 -4 -4
  \item FA_RLIC -4 -4 -4
  \item FA_ACR -4 -4 -4
  \item FA_SCR -4 -4 -4
  \item FA_PCR -4 -4 -4
  \item FA_PTR -4 -4 -4
  \item FA_SS -4 -4 -4
  \item FA_EC -4 -4 -4
  \item FA_CGC -4 -4 -4
  \item FA_CGH -4 -4 -4
  \item FA_FXST -4 -4 -4
  \item FA_SLF -4 -4 -4
  \item FA_SFO -4 -4 -4
  \item FA_UNC -4 -4 -4
  \item FA_TAP -4 -4 -4
  \item FA_IC -4 -4 -4
  \item FA_CC -4 -4 -4
  \item FA_CR -4 -4 -4
  \item FA_FULLWM -4 -4 -4
  \item MD_MCP -4 -4 -4
  \item MD_PCT -4 -4 -4
  \item MD_GCC -4 -4 -4
  \item MD_BCC -4 -4 -4
  \item MD_SCC -4 -4 -4
  \item MD_FX -4 -4 -4
  \item MD_CST_R -4 -4 -4
  \item MD_CST_L -4 -4 -4
  \item MD_ML_R -4 -4 -4
  \item MD_ML_L -4 -4 -4
  \item MD_ICP_R -4 -4 -4
  \item MD_ICP_L -4 -4 -4
  \item MD_SCP_R -4 -4 -4
  \item MD_SCP_L -4 -4 -4
  \item MD_CP_R -4 -4 -4
  \item MD_CP_L -4 -4 -4
  \item MD_ALIC_R -4 -4 -4
  \item MD_ALIC_L -4 -4 -4
  \item MD_PLIC_R -4 -4 -4
  \item MD_PLIC_L -4 -4 -4
  \item MD_RLIC_R -4 -4 -4
  \item MD_RLIC_L -4 -4 -4
  \item MD_ACR_R -4 -4 -4
  \item MD_ACR_L -4 -4 -4
  \item MD_SCR_R -4 -4 -4
  \item MD_SCR_L -4 -4 -4
  \item MD_PCR_R -4 -4 -4
  \item MD_PCR_L -4 -4 -4
  \item MD_PTR_R -4 -4 -4
  \item MD_PTR_L -4 -4 -4
  \item MD_SS_R -4 -4 -4
  \item MD_SS_L -4 -4 -4
  \item MD_EC_R -4 -4 -4
  \item MD_EC_L -4 -4 -4
  \item MD_CGC_R -4 -4 -4
  \item MD_CGC_L -4 -4 -4
  \item MD_CGH_R -4 -4 -4
  \item MD_CGH_L -4 -4 -4
  \item MD_FXST_R -4 -4 -4
  \item MD_FXST_L -4 -4 -4
  \item MD_SLF_R -4 -4 -4
  \item MD_SLF_L -4 -4 -4
  \item MD_SFO_R -4 -4 -4
  \item MD_SFO_L -4 -4 -4
  \item MD_UNC_R -4 -4 -4
  \item MD_UNC_L -4 -4 -4
  \item MD_TAP_R -4 -4 -4
  \item MD_TAP_L -4 -4 -4
  \item MD_CST -4 -4 -4
  \item MD_ML -4 -4 -4
  \item MD_ICP -4 -4 -4
  \item MD_SCP -4 -4 -4
  \item MD_CP -4 -4 -4
  \item MD_ALIC -4 -4 -4
  \item MD_PLIC -4 -4 -4
  \item MD_RLIC -4 -4 -4
  \item MD_ACR -4 -4 -4
  \item MD_SCR -4 -4 -4
  \item MD_PCR -4 -4 -4
  \item MD_PTR -4 -4 -4
  \item MD_SS -4 -4 -4
  \item MD_EC -4 -4 -4
  \item MD_CGC -4 -4 -4
  \item MD_CGH -4 -4 -4
  \item MD_FXST -4 -4 -4
  \item MD_SLF -4 -4 -4
  \item MD_SFO -4 -4 -4
  \item MD_UNC -4 -4 -4
  \item MD_TAP -4 -4 -4
  \item MD_IC -4 -4 -4
  \item MD_CC -4 -4 -4
  \item MD_CR -4 -4 -4
  \item MD_FULLWM -4 -4 -4
  \item RD_MCP -4 -4 -4
  \item RD_PCT -4 -4 -4
  \item RD_GCC -4 -4 -4
  \item RD_BCC -4 -4 -4
  \item RD_SCC -4 -4 -4
  \item RD_FX -4 -4 -4
  \item RD_CST_R -4 -4 -4
  \item RD_CST_L -4 -4 -4
  \item RD_ML_R -4 -4 -4
  \item RD_ML_L -4 -4 -4
  \item RD_ICP_R -4 -4 -4
  \item RD_ICP_L -4 -4 -4
  \item RD_SCP_R -4 -4 -4
  \item RD_SCP_L -4 -4 -4
  \item RD_CP_R -4 -4 -4
  \item RD_CP_L -4 -4 -4
  \item RD_ALIC_R -4 -4 -4
  \item RD_ALIC_L -4 -4 -4
  \item RD_PLIC_R -4 -4 -4
  \item RD_PLIC_L -4 -4 -4
  \item RD_RLIC_R -4 -4 -4
  \item RD_RLIC_L -4 -4 -4
  \item RD_ACR_R -4 -4 -4
  \item RD_ACR_L -4 -4 -4
  \item RD_SCR_R -4 -4 -4
  \item RD_SCR_L -4 -4 -4
  \item RD_PCR_R -4 -4 -4
  \item RD_PCR_L -4 -4 -4
  \item RD_PTR_R -4 -4 -4
  \item RD_PTR_L -4 -4 -4
  \item RD_SS_R -4 -4 -4
  \item RD_SS_L -4 -4 -4
  \item RD_EC_R -4 -4 -4
  \item RD_EC_L -4 -4 -4
  \item RD_CGC_R -4 -4 -4
  \item RD_CGC_L -4 -4 -4
  \item RD_CGH_R -4 -4 -4
  \item RD_CGH_L -4 -4 -4
  \item RD_FXST_R -4 -4 -4
  \item RD_FXST_L -4 -4 -4
  \item RD_SLF_R -4 -4 -4
  \item RD_SLF_L -4 -4 -4
  \item RD_SFO_R -4 -4 -4
  \item RD_SFO_L -4 -4 -4
  \item RD_UNC_R -4 -4 -4
  \item RD_UNC_L -4 -4 -4
  \item RD_TAP_R -4 -4 -4
  \item RD_TAP_L -4 -4 -4
  \item RD_CST -4 -4 -4
  \item RD_ML -4 -4 -4
  \item RD_ICP -4 -4 -4
  \item RD_SCP -4 -4 -4
  \item RD_CP -4 -4 -4
  \item RD_ALIC -4 -4 -4
  \item RD_PLIC -4 -4 -4
  \item RD_RLIC -4 -4 -4
  \item RD_ACR -4 -4 -4
  \item RD_SCR -4 -4 -4
  \item RD_PCR -4 -4 -4
  \item RD_PTR -4 -4 -4
  \item RD_SS -4 -4 -4
  \item RD_EC -4 -4 -4
  \item RD_CGC -4 -4 -4
  \item RD_CGH -4 -4 -4
  \item RD_FXST -4 -4 -4
  \item RD_SLF -4 -4 -4
  \item RD_SFO -4 -4 -4
  \item RD_UNC -4 -4 -4
  \item RD_TAP -4 -4 -4
  \item RD_IC -4 -4 -4
  \item RD_CC -4 -4 -4
  \item RD_CR -4 -4 -4
  \item RD_FULLWM -4 -4 -4
}
}
\examples{
\dontrun{
describe(dtiroi_v2)
}
}
\keyword{datasets}
