% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{dxsum}
\alias{dxsum}
\title{Diagnostic Summary}
\format{
A data frame with 13514 rows and 38 variables
}
\usage{
data(dxsum)
}
\description{
Diagnostic Summary
}
\details{
A dataset with variables as follows:

\itemize{
  \item DIAGNOSIS N NA 1.  Which best describes the participant's current diagnosis?
  \item RID N NA Participant roster ID
  \item SITEID N NA Site ID
  \item VISCODE T NA Visit code
  \item USERDATE S NA Date record created
  \item USERDATE2 S NA Date record last updated
  \item EXAMDATE D NA Examination Date
  \item DXCHANGE N NA 1.  Which best describes the participant's change in cognitive status from last visit to current visit:
  \item DXMDES T NA 2a.  MCI features (select all that apply):
  \item DXMPTR1 N NA If MCI - Memory features, complete the following (Petersen Criteria, see procedures manual for details):<p />i. Subjective memory complaint
  \item DXMPTR2 N NA ii. Informant memory complaint
  \item DXMPTR3 N NA iii. Normal general cognitive function
  \item DXMPTR4 N NA iv. Normal activities of daily living
  \item DXMPTR5 N NA v. Objective memory impairment for age and education
  \item DXMPTR6 N NA vi. Not demented by diagnostic criteria
  \item DXMDUE N NA 2b. Suspected cause of MCI:
  \item DXMOTHET T NA If MCI due to other etiology, select box(es) to indicate reason:
  \item DXDSEV N NA 3a.  Dementia Severity - Clinician's Impression
  \item DXDDUE N NA 3b.  Suspected cause of dementia
  \item DXAPP N NA If Dementia due to Alzheimer's Disease, indicate likelihood:
  \item DXAPOSS T NA If Possible AD, select box(es) to indicate reason:
  \item DXODES N NA If dementia due to other etiology, select best diagnosis:
  \item DXDEP N NA 4a. Depressive symptoms present?
  \item DXPARK N NA 4b.  Parkinsonism symptoms present?
}
}
\keyword{datasets}
