% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{ecg}
\alias{ecg}
\title{ECG}
\format{
A data frame with 119 rows and 20 variables
}
\usage{
data(ecg)
}
\description{
ECG
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
  \item DONE N  Was assessment/procedure done?
  \item NDREASON N  If No, reason not done:
  \item ECOMPDT D MM/DD/YYYY Date ECG Completed
  \item ECOMPTM T HHMM Time ECG Completed
  \item ECGPR N  Was there a PR interval?
  \item ECGPRI N ms PR Interval
  \item ECGQRS N ms QRS Duration
  \item ECGQTC N ms QT Interval
  \item ECGHRT N bpm Heart Rate
  \item ECGNORM N  ECG Status:
  \item ECGABNORM T  If abnormal, please describe:
  \item ECGREV N  Was the result reviewed prior to the AV-1451 dose?
  \item ECGLIMIT N  Was the result within limits specified in the addendum?
}
}
\examples{
\dontrun{
describe(ecg)
}
}
\keyword{datasets}
