% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adnimerger.R
\name{ecog.score}
\alias{ecog.score}
\title{Score Everyday Cognition E-Cog}
\usage{
ecog.score(x)
}
\arguments{
\item{x}{row of data to score}
}
\value{
scored value
}
\description{
This function scores the ECOG
following: http://www.sciencedirect.com/science/article/pii/S1552526011000896
}
\references{
\url{http://www.sciencedirect.com/science/article/pii/S1552526011000896}
}
\author{
Michael Donohue \email{mdonohue@usc.edu}
}
