% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{ecogpt}
\alias{ecogpt}
\title{Everyday Cognition - Participant Self-Report}
\format{
A data frame with 8055 rows and 59 variables
}
\usage{
data(ecogpt)
}
\description{
Everyday Cognition - Participant Self-Report
}
\details{
A dataset with variables as follows:

\itemize{
  \item COLDATE D NA Date collected
  \item EcogPtMem     N  Participant ECog - Mem         
  \item EcogPtLang    N  Participant ECog - Lang        
  \item EcogPtVisspat N  Participant ECog - Vis//Spat   
  \item EcogPtPlan    N  Participant ECog - Plan        
  \item EcogPtOrgan   N  Participant ECog - Organ       
  \item EcogPtDivatt  N  Participant ECog - Div atten   
  \item EcogPtTotal   N  Participant ECog - Total       
  \item SOURCE N  Information Source:
  \item RID N NA Participant roster ID
  \item SITEID N NA Site ID
  \item VISCODE T NA Visit code
  \item USERDATE S NA Date record created
  \item USERDATE2 S NA Date record last updated
  \item MEMORY1 N NA 1. Remembering a few shopping items without a list.
  \item MEMORY2 N NA 2. Remembering things that happened recently (such as recent outings, events in the news).
  \item MEMORY3 N NA 3. Recalling conversations a few days later.
  \item MEMORY4 N NA 4. Remembering where I have placed objects.
  \item MEMORY5 N NA 5. Repeating stories and/or questions.
  \item MEMORY6 N NA 6. Remembering the current date or day of the week.
  \item MEMORY7 N NA 7. Remembering I have already told someone something.
  \item MEMORY8 N NA 8. Remembering appointments, meetings, or engagements.
  \item LANG1 N NA 1. Forgetting the names of objects.
  \item LANG2 N NA 2. Verbally giving instructions to others.
  \item LANG3 N NA 3. Finding the right words to use in conversations.
  \item LANG4 N NA 4. Communicating thoughts in a conversation.
  \item LANG5 N NA 5. Following a story in a book or on TV.
  \item LANG6 N NA 6. Understanding the point of what other people are trying to say.
  \item LANG7 N NA 7. Remembering the meaning of common words.
  \item LANG8 N NA 8. Describing a program I have watched on TV.
  \item LANG9 N NA 9. Understanding spoken directions or instructions.
  \item VISSPAT1 N NA 1. Following a map to find a new location.
  \item VISSPAT2 N NA 2. Reading a map and helping with directions when someone else is driving.
  \item VISSPAT3 N NA 3. Finding my car in a parking lot.
  \item VISSPAT4 N NA 4. Finding my way back to a meeting spot in the mall or other location.
  \item VISSPAT5 N NA Duplicate Field Removed:  5. Following a story in a book or on TV.
  \item VISSPAT6 N NA 5. Finding my way around a familiar neighborhood.
  \item VISSPAT7 N NA 6. Finding my way around a familiar store.
  \item VISSPAT8 N NA 7. Finding my way around a house visited many times.
  \item PLAN1 N NA 1. Planning a sequence of stops on a shopping trip.
  \item PLAN2 N NA 2. The ability to anticipate weather changes and plan accordingly (i.e., bring a coat or umbrella)
  \item PLAN3 N NA 3. Developing a schedule in advance of anticipated events.
  \item PLAN4 N NA 4. Thinking things through before acting.
  \item PLAN5 N NA 5. Thinking ahead.
  \item ORGAN1 N NA 1. Keeping living and work space organized.
  \item ORGAN2 N NA 2. Balancing the checkbook without error.
  \item ORGAN3 N NA 3. Keeping financial records organized.
  \item ORGAN4 N NA 4. Prioritizing tasks by importance.
  \item ORGAN5 N NA 5. Keeping mail and papers organized.
  \item ORGAN6 N NA 6. Using an organized strategy to manage a medication schedule involving multiple medications.
  \item DIVATT1 N NA 1. The ability to do two things at once.
  \item DIVATT2 N NA 2. Returning to a task after being interrupted.
  \item DIVATT3 N NA 3. The ability to concentrate on a task without being distracted by external things in the environment.
  \item DIVATT4 N NA 4. Cooking or working and talking at the same time.
  \item STAFFASST N NA 1. How much assistance was needed from the research staff to complete this form?
  \item VALIDITY N NA 2. Validity of the information collected from the participant (based on research assistant's observations and interactions with the participant while they were completing the ECog):
  \item CONCERN N NA Are you concerned that you have a memory or other thinking problem?
}
}
\examples{
ecog.score <- function(x){ 
  x <- as.numeric(unlist(lapply(strsplit(as.character(x), '- ', fixed = TRUE),
      FUN = function(x) x[1])))
  x[x == 9] <- NA
  missing <- sum(is.na(x))
  if(missing/length(x) < 0.5){
    return(mean(x, na.rm = TRUE))
  }else{
    return(NA)
  }
}

mem <- c('MEMORY1', 'MEMORY2', 'MEMORY3', 'MEMORY4', 'MEMORY5', 'MEMORY6', 'MEMORY7', 'MEMORY8')
lang <-  c('LANG1', 'LANG2', 'LANG3', 'LANG4', 'LANG5', 'LANG6', 'LANG7', 'LANG8', 'LANG9')
visspat <- c('VISSPAT1', 'VISSPAT2', 'VISSPAT3', 'VISSPAT4', 
 'VISSPAT5', 'VISSPAT6', 'VISSPAT7', 'VISSPAT8')
plan <- c('PLAN1', 'PLAN2', 'PLAN3', 'PLAN4', 'PLAN5')
organ <- c('ORGAN1', 'ORGAN2', 'ORGAN3', 'ORGAN4', 'ORGAN5', 'ORGAN6')
divatt <- c('DIVATT1', 'DIVATT2', 'DIVATT3', 'DIVATT4')

ecogpt$EcogPtMem <- apply(ecogpt[, mem], 1, ecog.score)
ecogpt$EcogPtLang <- apply(ecogpt[, lang], 1, ecog.score)
ecogpt$EcogPtVisspat <- apply(ecogpt[, visspat], 1, ecog.score)
ecogpt$EcogPtPlan <- apply(ecogpt[, plan], 1, ecog.score)
ecogpt$EcogPtOrgan <- apply(ecogpt[, organ], 1, ecog.score)
ecogpt$EcogPtDivatt <- apply(ecogpt[, divatt], 1, ecog.score)
ecogpt$EcogPtTotal <- apply(ecogpt[, c(mem, lang, visspat, plan, organ, divatt)], 1, ecog.score)

label(ecogpt$EcogPtMem) <- 'Pt ECog - Mem'
label(ecogpt$EcogPtLang) <- 'Pt ECog - Lang'
label(ecogpt$EcogPtVisspat) <- 'Pt ECog - Vis/Spat'
label(ecogpt$EcogPtPlan) <- 'Pt ECog - Plan'
label(ecogpt$EcogPtOrgan) <- 'Pt ECog - Organ'
label(ecogpt$EcogPtDivatt) <- 'Pt ECog - Mem'
label(ecogpt$EcogPtTotal) <- 'Pt ECog - Div atten'


}
\references{
Farias T. S., Mungas, D., Harvey, D. J., Simmons, A., Reed, B. R., & DeCarli, C. (November 01, 2011). The measurement of everyday cognition: Development and validation of a short form of the Everyday Cognition scales. Alzheimer's and Dementia, 7, 6, 593-601. http://www.sciencedirect.com/science/article/pii/S1552526011000896
}
\keyword{datasets}
