% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{ecogsp}
\alias{ecogsp}
\title{Everyday Cognition - Study Partner Report}
\format{
A data frame with 8063 rows and 57 variables
}
\usage{
data(ecogsp)
}
\description{
Everyday Cognition - Study Partner Report
}
\details{
A dataset with variables as follows:

\itemize{
  \item COLDATE D NA Date collected
  \item SPID N NA Study Partner ID:
  \item EcogSPMem     N  Study Partner ECog - Mem       
  \item EcogSPLang    N  Study Partner ECog - Lang      
  \item EcogSPVisspat N  Study Partner ECog - Vis//Spat 
  \item EcogSPPlan    N  Study Partner ECog - Plan      
  \item EcogSPOrgan   N  Study Partner ECog - Organ     
  \item EcogSPDivatt  N  Study Partner ECog - Div atten 
  \item EcogSPTotal   N  Study Partner ECog - Total     
  \item SOURCE N  Information Source:
  \item RID N NA Participant roster ID
  \item SITEID N NA Site ID
  \item VISCODE T NA Visit code
  \item USERDATE S NA Date record created
  \item USERDATE2 S NA Date record last updated
  \item MEMORY1 N NA 1. Remembering a few shopping items without a list.
  \item MEMORY2 N NA 2. Remembering things that happened recently (such as recent outings, events in the news).
  \item MEMORY3 N NA 3. Recalling conversations a few days later.
  \item MEMORY4 N NA 4. Remembering where he/she has placed objects.
  \item MEMORY5 N NA 5. Repeating stories and/or questions.
  \item MEMORY6 N NA 6. Remembering the current date or day of the week.
  \item MEMORY7 N NA 7. Remembering he/she has already told someone something.
  \item MEMORY8 N NA 8. Remembering appointments, meetings, or engagements.
  \item LANG1 N NA 1. Forgetting the names of objects.
  \item LANG2 N NA 2. Verbally giving instructions to others.
  \item LANG3 N NA 3. Finding the right words to use in conversations.
  \item LANG4 N NA 4. Communicating thoughts in a conversation.
  \item LANG5 N NA 5. Following a story in a book or on TV.
  \item LANG6 N NA 6. Understanding the point of what other people are trying to say.
  \item LANG7 N NA 7. Remembering the meaning of common words.
  \item LANG8 N NA 8. Describing a program he/she has watched on TV.
  \item LANG9 N NA 9. Understanding spoken directions or instructions.
  \item VISSPAT1 N NA 1. Following a map to find a new location.
  \item VISSPAT2 N NA 2. Reading a map and helping with directions when someone else is driving.
  \item VISSPAT3 N NA 3. Finding one's car in a parking lot.
  \item VISSPAT4 N NA 4. Finding my way back to a meeting spot in the mall or other location.
  \item VISSPAT5 N NA Duplicate Field Removed:  5. Following a story in a book or on TV.
  \item VISSPAT6 N NA 5. Finding his/her way around a familiar neighborhood.
  \item VISSPAT7 N NA 6. Finding his/her way around a familiar store.
  \item VISSPAT8 N NA 7. Finding his/her way around a house visited many times.
  \item PLAN1 N NA 1. Planning a sequence of stops on a shopping trip.
  \item PLAN2 N NA 2. The ability to anticipate weather changes and plan accordingly (i.e., bring a coat or umbrella)
  \item PLAN3 N NA 3. Developing a schedule in advance of anticipated events.
  \item PLAN4 N NA 4. Thinking things through before acting.
  \item PLAN5 N NA 5. Thinking ahead.
  \item ORGAN1 N NA 1. Keeping living and work space organized.
  \item ORGAN2 N NA 2. Balancing the checkbook without error.
  \item ORGAN3 N NA 3. Keeping financial records organized.
  \item ORGAN4 N NA 4. Prioritizing tasks by importance.
  \item ORGAN5 N NA 5. Keeping mail and papers organized.
  \item ORGAN6 N NA 6. Using an organized strategy to manage a medication schedule involving multiple medications.
  \item DIVATT1 N NA 1. The ability to do two things at once.
  \item DIVATT2 N NA 2. Returning to a task after being interrupted.
  \item DIVATT3 N NA 3. The ability to concentrate on a task without being distracted by external things in the environment.
  \item DIVATT4 N NA 4. Cooking or working and talking at the same time.
}
}
\examples{

ecog.score <- function(x){ 
  x <- as.numeric(unlist(lapply(strsplit(as.character(x), '- ', fixed = TRUE),
      FUN = function(x) x[1])))
  x[x == 9] <- NA
  missing <- sum(is.na(x))
  if(missing/length(x) < 0.5){
    return(mean(x, na.rm = TRUE))
  }else{
    return(NA)
  }
}

mem <- c('MEMORY1', 'MEMORY2', 'MEMORY3', 'MEMORY4', 'MEMORY5', 'MEMORY6', 'MEMORY7', 'MEMORY8')
lang <-  c('LANG1', 'LANG2', 'LANG3', 'LANG4', 'LANG5', 'LANG6', 'LANG7', 'LANG8', 'LANG9')
visspat <- c('VISSPAT1', 'VISSPAT2', 'VISSPAT3', 'VISSPAT4', 
 'VISSPAT5', 'VISSPAT6', 'VISSPAT7', 'VISSPAT8')
plan <- c('PLAN1', 'PLAN2', 'PLAN3', 'PLAN4', 'PLAN5')
organ <- c('ORGAN1', 'ORGAN2', 'ORGAN3', 'ORGAN4', 'ORGAN5', 'ORGAN6')
divatt <- c('DIVATT1', 'DIVATT2', 'DIVATT3', 'DIVATT4')

ecogsp$EcogSPMem <- apply(ecogsp[, mem], 1, ecog.score)
ecogsp$EcogSPLang <- apply(ecogsp[, lang], 1, ecog.score)
ecogsp$EcogSPVisspat <- apply(ecogsp[, visspat], 1, ecog.score)
ecogsp$EcogSPPlan <- apply(ecogsp[, plan], 1, ecog.score)
ecogsp$EcogSPOrgan <- apply(ecogsp[, organ], 1, ecog.score)
ecogsp$EcogSPDivatt <- apply(ecogsp[, divatt], 1, ecog.score)
ecogsp$EcogSPTotal <- apply(ecogsp[, c(mem, lang, visspat, plan, organ, divatt)], 1, ecog.score)

label(ecogsp$EcogSPMem) <- 'SP ECog - Mem'
label(ecogsp$EcogSPLang) <- 'SP ECog - Lang'
label(ecogsp$EcogSPVisspat) <- 'SP ECog - Vis/Spat'
label(ecogsp$EcogSPPlan) <- 'SP ECog - Plan'
label(ecogsp$EcogSPOrgan) <- 'SP ECog - Organ'
label(ecogsp$EcogSPDivatt) <- 'SP ECog - Mem'
label(ecogsp$EcogSPTotal) <- 'SP ECog - Div atten'
}
\references{
Farias T. S., Mungas, D., Harvey, D. J., Simmons, A., Reed, B. R., & DeCarli, C. (November 01, 2011). The measurement of everyday cognition: Development and validation of a short form of the Everyday Cognition scales. Alzheimer's and Dementia, 7, 6, 593-601. http://www.sciencedirect.com/science/article/pii/S1552526011000896
}
\keyword{datasets}
