% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{eligconf}
\alias{eligconf}
\title{Eligibility Confirmation}
\format{
A data frame with 1386 rows and 10 variables
}
\usage{
data(eligconf)
}
\description{
Eligibility Confirmation
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item CHSTATUS N  Status of participant at this visit (check one):
  \item CHREASON T  Reason participant excluded from protocol:
  \item USERDATE2 S  Date record last updated
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
