% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{emory_peptideratios}
\alias{emory_peptideratios}
\title{Emory University CSF Targeted MS}
\format{
A data frame with 4782 rows and 17 variables
}
\usage{
data(emory_peptideratios)
}
\description{
Emory University CSF Targeted MS
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item VID -4 -4 -4
  \item PEPTIDE -4 NA Targeted peptide sequence
  \item PROTEIN -4 NA Protein Name from Skyline
  \item REPLICATE1 -4 -4 -4
  \item REPLICATE2 -4 -4 -4
  \item LINK.ID -4 -4 -4
  \item TOTAL.AREA.RATIO -4 NA Light/Heavy peptide ratio
  \item PROT -4 -4 -4
  \item SHIPPING.BOX -4 -4 -4
  \item POS -4 -4 -4
  \item DATABASE.ID -4 -4 -4
  \item DER -4 -4 -4
  \item VOL.ML -4 -4 -4
}
}
\examples{
\dontrun{
describe(emory_peptideratios)
}
}
\keyword{datasets}
