% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{enrolltr}
\alias{enrolltr}
\title{Enrollment Tracking}
\format{
A data frame with 961 rows and 14 variables
}
\usage{
data(enrolltr)
}
\description{
Enrollment Tracking
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item DATE D  Date
  \item CITY T  City
  \item STATE T  State/Province
  \item COUNTRY T  Country
  \item REFSOURCE T  Please check all that apply when asking the volunteer where they heard about the ADNI study:
  \item RADIOSPE T  Radio Station Specify:
  \item OTHERSPE T  Other Specify:
  \item USERDATE2 S  Date record last updated
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
