% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{euroimmun}
\alias{euroimmun}
\title{EUROIMMUN CSF Beta-Amyloid}
\format{
A data frame with 140 rows and 9 variables
}
\usage{
data(euroimmun)
}
\description{
EUROIMMUN CSF Beta-Amyloid
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item POOL -4 -4 -4
  \item DRAWTIME -4 -4 -4
  \item BETA.AMYLOID.1.40 -4 pg/ml Beta-Amyloid 1-40
  \item BETA.AMYLOID.1.42 -4 pg/ml Beta-Amyloid 1-42
  \item BETA.AMYLOID.42.40 -4  Beta-Amyloid 1-42/1-40 ratio
}
}
\examples{
\dontrun{
describe(euroimmun)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/EUROIMMUN/ADNI_Methods_EUROIMMUN.pdf}
}
\keyword{datasets}
