% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{exclusio}
\alias{exclusio}
\title{Exclusion Criteria}
\format{
A data frame with 1087 rows and 19 variables
}
\usage{
data(exclusio)
}
\description{
Exclusion Criteria
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item EXNEURO N  1. NL - Does the participant have a significant neurologic disease such as Parkinson's disease, multi-infarct dementia, Huntington's disease, normal pressure hydrocephalus, brain tumor, progressive supranuclear palsy, seizure disorder, subdural hematoma, multiple sclerosis, or history of significant head trauma followed by persistent neurologic defaults or known structural brain abnormalities.<br>MCI - Does the participant have a significant neurologic disease other than suspected incipient Alzheimer's disease such as...<br>AD - Does the participant have a significant neurologic disease other than Alzheimer's disease including...
  \item EXNIMAG N  2. Does the participant's screening/baseline MRI scans have evidence of infection, infarction, or other focal lesions?  Participants with multiple lacunes or lacunes in a critical memory structure are excluded.
  \item EXMRI N  3. Does the participant have a pacemaker, aneurysm clips, artificial heart valves, ear implants, metal fragments or foreign objects in the eyes, skin or body.
  \item EXPSYCH N  4. NL - Has the participant had major depression or bipolar disorder as described in DSM-IV within the past year or a history of schizophrenia (DSM IV criteria)?<br>MCI/AD - Does the participant have a history of major depression...or a history of psychotic features, agitation, or behavioral problems within the last 3 months which could lead to difficulty complying with the protocol?
  \item EXABUSE N  5. Does the participant have a history of alcohol or substance abuse or dependence within the past 2 years (DSM IV criteria)?
  \item EXILLNESS N  6. Does the participant have a significant systemic illness or unstable medical condition which could lead to difficulty complying with the protocol?
  \item EXLAB N  7. Does the participant have any clinically significant abnormalities in B12, RPR, or TFTs that might interfere with the study.
  \item EXRESID N  8. Does the participant reside in a skilled nursing facility?
  \item EXMEDS N  9. Is the participant currently taking, or has he/she taken in the last 4 weeks, any excluded medication(s) as described in the Procedures Manual?
  \item EXINVEST N  10. Has the participant used another investigational agent within one month prior to screening?
  \item EXPARTICIP N  11. Is the participant participating in a clinical study involving neuropsychological measures being collected more than one time per year?
  \item USERDATE2 S  Date record last updated
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
