% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{excptlog}
\alias{excptlog}
\title{Inclusion/Exclusion Exceptions Log}
\format{
A data frame with 1340 rows and 14 variables
}
\usage{
data(excptlog)
}
\description{
Inclusion/Exclusion Exceptions Log
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item RECNO N  Record Number
  \item EXAMDATE D  Examination Date
  \item EXCCRIT N  Exception applies to:
  \item EXCNUMB T  Inclusion/Exclusion Criteria Item number
  \item EXUNSCAN N  If requesting an unscheduled visit, specify which scans are being requested:
  \item EXCBIDT D  If Fluroscopy or LP spine film, indicate [anticipated] date performed:
  \item EXCDESC T  Description of exception
  \item USERDATE2 S  Date record last updated
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
