% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{extconsnt}
\alias{extconsnt}
\title{Extension Consent Form}
\format{
A data frame with 137 rows and 9 variables
}
\usage{
data(extconsnt)
}
\description{
Extension Consent Form
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
  \item INCEXT N  Has the participant signed the Informed Consent form for the ADNI Extension (Amendment 3)?
  \item INCEXTDT D  If Yes, date signed.
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
