% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{faganlab}
\alias{faganlab}
\title{Fagan lab summaries}
\format{
A data frame with 587 rows and 16 variables
}
\usage{
data(faganlab)
}
\description{
Fagan lab summaries
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item VILIP -4 pg/ml VILIP-1 
  \item VILIP_STDEV -4  Standard Deviation
  \item VILIP_CV -4  Coefficient of Variation (%)
  \item YKL -4 ng/ml YKL-40
  \item YKL_STDEV -4  Standard Deviation
  \item YKL_CV -4  Coefficient of Variation (%)
  \item SNAP -4 pg/ml SNAP-25 
  \item SNAP_STDEV -4  Standard Deviation
  \item SNAP_CV -4  Coefficient of Variation (%)
  \item NGRN -4 pg/ml NRGN
  \item NGRN_STDEV -4  Standard Deviation
  \item NGRN_CV -4  Coefficient of Variation (%)
}
}
\examples{
\dontrun{
describe(faganlab)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/FAGANLAB/ADNI_Methods_Fagan\%20Lab\%20Final.pdf}, \url{https://adni.bitbucket.io/reference/docs/FAGANLAB/ADNI_Methods_Fagan\%20Lab\%20VILIP.pdf}
}
\keyword{datasets}
