% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{famhxpar}
\alias{famhxpar}
\title{Family History - Parents}
\format{
A data frame with 1048 rows and 17 variables
}
\usage{
data(famhxpar)
}
\description{
Family History - Parents
}
\details{
A dataset with variables as follows:

\itemize{
  \item MOTHALIVE N NA Is mother living? 
  \item MOTHAGE N NA Current age or age at death? 
  \item MOTHDEM N NA Did/Does the biological mother have dementia?
  \item MOTHAD N NA If yes, did/does biological mother have Alzheimer's Disease?
  \item MOTHSXAGE N NA At what approximate age did mother's symptoms start?
  \item FATHALIVE N NA Is father living? 
  \item FATHAGE N NA Current age or age at death? 
  \item FATHDEM N NA Did/Does the biological father have dementia?
  \item FATHAD N NA If yes, did/does biological father have Alzheimer's Disease?
  \item FATHSXAGE N NA At what approximate age did father's symptoms start?
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
}
}
\examples{
\dontrun{
describe(famhxpar)
}
}
\keyword{datasets}
