% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{famhxsib}
\alias{famhxsib}
\title{Family History - Sibling Log}
\format{
A data frame with 2970 rows and 15 variables
}
\usage{
data(famhxsib)
}
\description{
Family History - Sibling Log
}
\details{
A dataset with variables as follows:

\itemize{
  \item SIBYOB N NA Biological sibling year of birth
  \item SIBRELAT N NA Specify relationship
  \item SIBGENDER N NA Gender
  \item SIBALIVE N NA Is sibling living?
  \item SIBAGE N NA Current age or age of death
  \item SIBDEMENT N NA Did/Does this sibling have dementia?
  \item SIBAD N NA If yes, did/does sibling have Alzheimer's Disease?
  \item SIBSXAGE N NA At what approximate age did symptoms start?
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
}
}
\examples{
\dontrun{
describe(famhxsib)
}
}
\keyword{datasets}
