% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{faq}
\alias{faq}
\title{Functional Assessment Questionnaire}
\format{
A data frame with 11952 rows and 23 variables
}
\usage{
data(faq)
}
\description{
Functional Assessment Questionnaire
}
\details{
A dataset with variables as follows:

\itemize{
  \item FAQDATE D NA Date collected
  \item SPID N NA Study Partner ID:
  \item SOURCE N NA Information Source
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item FAQSOURCE N  Information Source
  \item FAQFINAN N  1. Writing checks, paying bills, or balancing checkbook.
  \item FAQFORM N  2. Assembling tax records, business affairs, or other papers.
  \item FAQSHOP N  3. Shopping alone for clothes, household necessities, or groceries.
  \item FAQGAME N  4. Playing a game of skill such as bridge or chess, working on a hobby.
  \item FAQBEVG N  5. Heating water, making a cup of coffee, turing off the stove.
  \item FAQMEAL N  6. Preparing a balanced meal.
  \item FAQEVENT N  7. Keeping track of current events.
  \item FAQTV N  8. Paying attention to and understanding a TV program, book, or magazine.
  \item FAQREM N  9. Remembering appointments, family occasions, holidays, medications.
  \item FAQTRAVL N  10. Traveling out of the neighborhood, driving, or arranging to take public transportation.
  \item FAQTOTAL N  Total Score
  \item USERDATE2 S  Date record last updated
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
