% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{fci}
\alias{fci}
\title{Financial Capacity Instrument Short Form (FCI-SF)}
\format{
A data frame with 2588 rows and 66 variables
}
\usage{
data(fci)
}
\description{
Financial Capacity Instrument Short Form (FCI-SF)
}
\details{
A dataset with variables as follows:

\itemize{
  \item DONE N NA Was assessment/procedure done?
  \item NDREASON N NA If No, reason not done
  \item RATER T NA Rater Initials
  \item NICKELS N NA 1.  Nickels
  \item QUARTERS N NA 2.  Quarters
  \item BUDGET N NA 3. Budget
  \item EXPMEDINS N NA Experience with Medical Insurance Plan?
  \item INSURANCE N NA 4. Health Care Insurance Problem
  \item MEDINSTIME N seconds Time to complete Item 4
  \item EXPINCTAX N NA Experience with Income Taxes?
  \item TAXCREDA N NA 5. a. Tax Credit
  \item QUES5TIME N seconds Time to complete Item 5.a.
  \item TAXCREDB N NA 5. b. Tax Credit
  \item EXPCHECKREG N NA Experience with Checkbook/Check Register?
  \item PAYEE N NA 6. Payee
  \item SNGLCHK7 N NA 7. Payee section
  \item SNGLCHK8 N NA 8.  $ amount section (numerical)
  \item SNGLCHK9 N NA 9.  $ amount section (written)
  \item SNGLCHK10 N NA 10. Signature of payer
  \item SNGLCHK11 N NA 11. Date properly written
  \item SNGLCHK12 N NA 12. Number of check
  \item SNGLCHK13 N NA 13. Date properly entered
  \item SNGLCHK14 N NA 14. Payee section
  \item SNGLCHK15 N NA 15. Amount of check
  \item SNGLCHK16 N NA 16. New account balance
  \item SNGLCHKTIME N seconds Time to complete items 7-16
  \item COMPLXCHK17 N NA 17. Number of check
  \item COMPLXCHK18 N NA 18. Date properly entered
  \item COMPLXCHK19 N NA 19. Payee section
  \item COMPLXCHK20 N NA 20. Amount of check
  \item COMPLXCHK21 N NA 21. New Account Balance
  \item COMPLXCHK22 N NA 22. Number of check
  \item COMPLXCHK23 N NA 23. Date properly entered
  \item COMPLXCHK24 N NA 24. Payee section
  \item COMPLXCHK25 N NA 25. Amount of check
  \item COMPLXCHK26 N NA 26. New Account Balance
  \item COMPLXCHK27 N NA 27. Date properly entered
  \item COMPLXCHK28 N NA 28. Description of Transaction
  \item COMPLXCHK29 N NA 29. Amount of deposit
  \item COMPLXCHK30 N NA 30. New account balance
  \item COMPLXCHKTIME N seconds Time to complete Items 17-30
  \item BANKSTATA N NA 31. a. Bank Statement
  \item BANKSTATB N NA 31. b. Bank Statement
  \item EXPBANKSTATE N NA Experience with Bank Statement?
  \item BANKSTAT32 N NA 32. Interest Rate
  \item BANKSTAT33 N NA 33. Time Period
  \item BANKSTAT34 N NA 34. Checks Cleared
  \item BANKSTAT35 N NA 35. Quarterly Interest
  \item BANKSTAT36 N NA 36. Gaps in Check Sequence
  \item BANKSTAT37 N NA 37. Date of the Withdrawal Payment
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
  \item MC_SCORE N  Mental Calculation Score (Items 1-2)            
  \item FC_SCORE N  Financial Conceptual Knowledge Score (Items 3-6)
  \item SNGLCHKSCORE   N  Single/Checkbook Register Score (Items 7-16):       
  \item COMPLXCHKSCORE N  Complex Checkbook/Register Score (Items 17-30):
  \item CHKCOMPOSTIME  N seconds Checkbook/Register Composite Time (Items 7-16 & 17-30):
  \item CHKCOMPOSSCORE N  Combined Checkbook/Register Score (Items 7-16 + 17-30 ):
  \item BANKSTATSCORE  N  Bank Statement Management Score (Items 31-37):
  \item TOTCOMPTIME N seconds TOTAL Composite Time (Items 4, 5, 7-16, 17-30):
  \item FCISCORE N  TOTAL FCI-SF SCORE:
}
}
\examples{
\dontrun{
describe(fci)
}
}
\keyword{datasets}
