% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{fhq}
\alias{fhq}
\title{Family History Questionnaire}
\format{
A data frame with 2952 rows and 15 variables
}
\usage{
data(fhq)
}
\description{
Family History Questionnaire
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item FHQSOURCE N  Information Souce
  \item FHQPROV N  Indicate below who provided the information collected for this questionnaire:
  \item FHQMOM N  Dementia
  \item FHQMOMAD N  Alzheimer's Disease
  \item FHQDAD N  Dementia
  \item FHQDADAD N  Alzheimer's Disease
  \item FHQSIB N  3. Does the participant have any siblings?
  \item USERDATE2 S  Date record last updated
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
