% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{flairqc}
\alias{flairqc}
\title{FLAIR QC}
\format{
A data frame with 1128 rows and 14 variables
}
\usage{
data(flairqc)
}
\description{
FLAIR QC
}
\details{
A dataset with variables as follows:

\itemize{
  \item USERDATE -4 NA Date record created
  \item RID -4 NA Participant roster ID
  \item DATE -4 NA Date
  \item HEADCOVERAGE -4 NA Headcoverage
  \item SUSCEPTIBILITYARTIFACTS -4 NA Susceptibility artifacts
  \item INHOMOGENEITYARTIFACTS -4 NA Inhomogeneity artifacts
  \item MOTION -4 NA Motion
  \item TWODSERIESALIGNMENT -4 NA 2D series alignment
  \item CSFFLOWARTIFACTS -4 NA CSF Flow Artifacts
  \item CSFSUPPRESSION -4 NA CSF Suppression
  \item SNR -4 NA SNR
  \item PULSEARTIFACT -4 NA Pulse Artifact
  \item FLAIROVERALLQUALITY -4 NA FLAIR overallquality
}
}
\examples{
\dontrun{
describe(flairqc)
}
}
\keyword{datasets}
