% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{foxlabbsi}
\alias{foxlabbsi}
\title{Fox Lab BSI measures}
\format{
A data frame with 14148 rows and 26 variables
}
\usage{
data(foxlabbsi)
}
\description{
Fox Lab BSI measures
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4  Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4  Examination Date
  \item VERSION -4  Version date of results
  \item RUNDATE -4  Analysis run date
  \item STATUS -4  Analysis completeness
  \item LONIUID -4  UID of fully pre-processed scan for which results are provided
  \item LONIUID_BASE -4  UID of fully pre-processed baseline scan for longitudinal results
  \item BRAINVOL -4 ml Whole brain volume
  \item VENTVOL -4 ml Ventricular volume
  \item HIPPOVOL_R -4 ml Right side hippocampal volume
  \item HIPPOVOL_L -4 ml Left side hippocampal volume
  \item DBCBBSI -4 ml Whole brain classic BSI
  \item KMNDBCBBSI -4 ml Whole brain KN-BSI
  \item VBSI -4 ml Ventricular BSI
  \item HBSI_R -4 ml Right side hippocampal BSI
  \item HBSI_L -4 ml Left side hippocampal BSI
  \item REGRATING -4  Longitudinal QC rating
  \item VENTACCEPT -4  Ventricular QC rating
  \item HPACCEPT_R -4  Right side hippocampal QC rating
  \item HPACCEPT_L -4  Left side hippocampal QC rating
  \item MRSEQUENCE -4  MRI sequence
  \item MRFIELD -4 tesla MRI field strength
  \item PROTOCOL -4  The ADNI study under which the original scans were collected
  \item QC_PASS -4  Selected for cross-sectional analysis
}
}
\examples{
\dontrun{
describe(foxlabbsi)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/FOXLABBSI/ADNI2_Methods_BrainMAPS.pdf}, \url{https://adni.bitbucket.io/reference/docs/FOXLABBSI/ADNI2_Methods_STEPS.pdf}
}
\keyword{datasets}
