% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{fujirebioabeta}
\alias{fujirebioabeta}
\title{Fujirebio Beta-Amyloid Ratio}
\format{
A data frame with 442 rows and 17 variables
}
\usage{
data(fujirebioabeta)
}
\description{
Fujirebio Beta-Amyloid Ratio
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4  Visit code
  \item VISCODE2 -4  Translated visit code
  \item EXAMDATE -4  Examination Date
  \item GUSPECID -4  Specimen ID
  \item VID -4  Visit ID
  \item DRAWDTE -4 -4 -4
  \item DER -4 -4 -4
  \item SITE -4  Site
  \item VOL -4 mL Volume
  \item RECDTE -4  
  \item STORDTE -4  
  \item RUN -4  Run
  \item RUNDATE -4  Run date
  \item ABETA42 -4 pg/mL Beta-Amyloid 1-42
  \item ABETA40 -4 pg/mL Beta-Amyloid 1-40
  \item ABETA42_40 -4  Beta-Amyloid 1-42/1-40 ratio
}
}
\examples{
\dontrun{
describe(fujirebioabeta)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/FUJIREBIOABETA/ADNI_Methods_Fujirebio_Submission.pdf}
}
\keyword{datasets}
