% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{gdscale}
\alias{gdscale}
\title{Geriatric Depression Scale}
\format{
A data frame with 12116 rows and 28 variables
}
\usage{
data(gdscale)
}
\description{
Geriatric Depression Scale
}
\details{
A dataset with variables as follows:

\itemize{
  \item GDDATE D NA Date collected
  \item SOURCE N NA Information Source
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item GDSOURCE N  Information Source
  \item GDUNABL T  Check here if:
  \item GDSATIS N  1. Are you basically satisfied with your life?
  \item GDDROP N  2. Have you dropped many of your activities and interests?
  \item GDEMPTY N  3. Do you feel that your life is empty?
  \item GDBORED N  4. Do you often get bored?
  \item GDSPIRIT N  5. Are you in good spirits most of the time?
  \item GDAFRAID N  6. Are you afraid that something bad is going to happen to you?
  \item GDHAPPY N  7. Do you feel happy most of the time?
  \item GDHELP N  8. Do you often feel helpless?
  \item GDHOME N  9. Do you prefer to stay at home, rather than going out and doing new things?
  \item GDMEMORY N  10. Do you feel you have more problems with memory than most?
  \item GDALIVE N  11. Do you think its wonderful to be alive now?
  \item GDWORTH N  12. Do you feel pretty worthless the way you are now?
  \item GDENERGY N  13. Do you feel full of energy?
  \item GDHOPE N  14. Do you feel that your situation is hopeless?
  \item GDBETTER N  15. Do you think that most people are better off than you are?
  \item GDTOTAL N  Total Score
  \item USERDATE2 S  Date record last updated
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
