% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{genetic}
\alias{genetic}
\title{Genetic Sample Collection}
\format{
A data frame with 7241 rows and 37 variables
}
\usage{
data(genetic)
}
\description{
Genetic Sample Collection
}
\details{
A dataset with variables as follows:

\itemize{
  \item COLLDATE D YYYY-MM-DD Date of Collection
  \item RNAFROZT T HHMM If Frozen, Time RNA placed in freezer (24hr clock):
  \item PBMCCOLL N NA On this date, was PBMC sample collected?
  \item PBMCVOL N mL If PBMC collected, volume
  \item RBCCOLL N NA On this date, was RBC extracted from biomarker collection tube?
  \item RBCTUBE N NA If RBC collected, total number of aliquot tubes
  \item RBCALIQ1 N mL RBC aliquot 1 volume (2.0 mL tube)
  \item RBCALIQ2 N mL RBC aliquot 2 volume (4.0 mL tube)
  \item RBCFROZT T HHMM If RBC collected, time placed in freezer
  \item BCFROZT T HHMM If Buffy Coat frozen, time placed in freezer
  \item GENFEDDATE D YYYY-MM-DD Date FedExed
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
  \item RECNO N  
  \item APCOLLECT N  Was DNA sample collected (1 x 10 mL purple top EDTA tube)?
  \item EXAMDATE D  Examination Date
  \item APTIME T HHMM Time of DNA Collection (24hr clock)
  \item APVOLUME N mL Volume of blood drawn into 10 mL EDTA tube for DNA testing
  \item RNACOLL N  Was RNA sample collected (3 x 2.5 mL PAXgene RNA tubes)?
  \item RNADATE D  Date of RNA collection:
  \item RNATIME T HHMM Time of RNA collection (24hr clock):
  \item RNAVOL N mL Volume of blood drawn into 3 x 2.5 mL PAXgene RNA Tubes
  \item CLCOLL N  Was cell immortalization sample collected?
  \item CLDATE D  Date of cell immortalization collection
  \item CLTIME T HHMM Time of cell immortalization collection (24hr clock)
  \item CLVOLUME N mL Total volume of blood drawn for cell immortalization into 2 x 8.5 mL ACD-A (yellow top) tubes
  \item BCOAT N  Was Buffy Coat extracted from biomarker collection tube?
  \item BCREASON N  If No, please provide reason why the Buffy Coat was not extracted
  \item BCVOL N mL Volume of aliquot
  \item DNAREASON N  If no, indicate reason
  \item RNAREASON N  If no, indicate reason
  \item CLREASON N  If no, indicate reason
}
}
\keyword{datasets}
