% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{hcres}
\alias{hcres}
\title{Homocysteine - Results}
\format{
A data frame with 3390 rows and 14 variables
}
\usage{
data(hcres)
}
\description{
Homocysteine - Results
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item HCTESTDT D  Date Test Performed
  \item HCAMPLAS N  AM total plasma Homocysteine
  \item HCVOLUME N  Volume of Blood Shipped in Lavendar Top Tube
  \item HCRECEIVE N  Sample received within 24 hours of blood draw?
  \item HCFROZEN N  Sample received frozen and packed in dry ice?
  \item HCRESAMP N  Request Resample?
  \item HCUSABLE N  Sample Useable?
  \item USERDATE2 S  Date record last updated
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
