% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{hulab}
\alias{hulab}
\title{ADNI Data Hu Lab}
\format{
A data frame with 398 rows and 19 variables
}
\usage{
data(hulab)
}
\description{
ADNI Data Hu Lab
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item TNFR1 -4 -4 P19438
  \item TNFR2 -4 -4 P20333
  \item TGFBETA1 -4 -4 P01137
  \item TGFBETA2 -4 -4 P61812
  \item TGFBETA3 -4 -4 P10600
  \item IL.21 -4 -4 Q9HBE4
  \item IL.6 -4 -4 P05231
  \item IL.7 -4 -4 P13232
  \item IL.9 -4 -4 P15248
  \item IL.10 -4 -4 P22301
  \item TNFALPHA -4 -4  P01375
  \item IP.10 -4 -4 P02778
  \item IL.12.P40 -4 -4 P29460
  \item ICAM.1 -4 -4 P05362
  \item VCAM.1 -4 -4 P19320
}
}
\examples{
\dontrun{
describe(hulab)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/HULAB/ADNI_Methods_Hu_Lab.pdf}
}
\keyword{datasets}
