% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{inclusio}
\alias{inclusio}
\title{Inclusion Criteria}
\format{
A data frame with 1109 rows and 35 variables
}
\usage{
data(inclusio)
}
\description{
Inclusion Criteria
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item INCSIGN N  1. Have the participant and study partner signed the Informed Consent form?
  \item INCSGDT D  If Yes, date signed
  \item INCSG15T N  Check the following to indicate the participant is suitable for and consents to:
  \item INCSGPET N  
  \item INCSG3T N  
  \item INCSGLP N  
  \item INCMCOMP N  2. NL - Is participant free of memory complaints, verified by an informant, aside from those normal with age?<br>MCI - Does the subject have memory complaints and memory difficulties that are verified by an informant?<br>AD -  Does the subject have memory complaints that are verified by an informant?
  \item INCMFXN N  3. NL - Normal memory function documented by scoring at specific cutoffs on the Logical Memory II subscale (delayed Paragraph Recall) from the Wechsler Memory Scaled - Revised (the maximum score is 25)<br>MCI/AD - Abnormal memory function documented by scoring below the educationn adjusted cutoff on the Logical Memory II subscale (Delayed  Paragraph Recall) from the Wechsler Memory Scale - Revised (the maximum score is 25)
  \item INCMMSE N  4. NL/MCI - Does the participant have Mini-Mental State Exam score between 24 and 30 (inclusive)?  (Exceptions must be made for subjects with less than 8 years of education at the discretion of the project director).<br>AD - Does the participant have an MMSE score between 20 and 26 (inclusive)?
  \item INCCDR N  5. NL - Does the participant have a Clinical Dementia Rating of 0?  Memory Box score must be 0.<br>MCI - Does the participant have a Clinical Dementia Rating of 0.5?  Memory Box score must be at least 0.5.<br>AD - Does participant have a Clinical Dementia rating of 0.5 or 1.0?
  \item INCCOG N  6. NL - Is the participant cognitively normal based on an absence of significant impairment in cognitive functions or activities of daily living?<br>MCI - Is the participant's general cognition and functional performance sufficiently perserved such that a diagnosis of Alzheimer's disease cannot be made by the site physician at the time of the screening visit? <br>AD - Does the participant meet NINCDS/ADRDA criteria for probable AD?
  \item INCHACH N  7. Does the participant have a Modified Hachinski score less than or equal to 4?
  \item INCAGE N  8. Is the participant between 55 and 90 years of age inclusive?
  \item INCSTMED N  9. Has the participant been on stable doses of non-excluded medications for at least 4 weeks prior to screening?
  \item INCGDS N  10. Does the participant have a Geriatric Depression Scale score of <6?
  \item INCINFORM N  11. Does the participant have an informant available who they have frequent contact with (e.g. an average of 10 hours per week or more), and can accompany the participant to all clinic visits and imaging sessions for the duration of the protocol?
  \item INCVISUAL N  12. Does the participant have adequate visual and auditory acuity to allow neuropsychological testing?
  \item INCHEALTH N  13. Is the participant in good general health with no additional diseases expected to interfere with the study?
  \item INCFEM N  14. If female, is the participant not pregnant, lactating, or of childbearing potential (i.e. women must be two years post-menopausal or surgically sterile)?
  \item INCPART N  15. NL/MCI - Is the participant willing and able to complete all Baseline assessment and participate in a 3-year protocol?<br>AD - Is the participant willing and able to ...2-year protocol?
  \item INCIMAG N  16. Is the participant willing to undergo MRI 1.5 Tesla neuroimaging (PET and MRI 3 Tesla are optional) and provide DNA for ApoE assessments and banking as well as plasma samples at protocol specified time points?
  \item INCEDUC N  17. Has the participant completed 6 grades of education (or had a good work history sufficient to exclude mental retardation)?
  \item INCLANG N  18. Is the participant fluent in English or Spanish?
  \item INCHIST N  19a. Medical History
  \item INCPHYS N  19b. Physical Examination?
  \item INCNEUR N  19c. Neurological Examination?
  \item INCLAB N  19d. Laboratory Tests?
  \item USERDATE2 S  Date record last updated
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
