% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{inithealth}
\alias{inithealth}
\title{Initial Health Assessment}
\format{
A data frame with 16165 rows and 17 variables
}
\usage{
data(inithealth)
}
\description{
Initial Health Assessment
}
\details{
A dataset with variables as follows:

\itemize{
  \item IHNUM N NA Condition Number
  \item IHSYMPTOM N NA System/Category
  \item IHSURG N NA Is this a surgery?
  \item IHSURGDATE D YYYY-MM-DD If Yes, surgery date
  \item IHPRESENT N NA Is this a condition or symptom present in the three months prior to screening?
  \item IHCHRON N NA Chronicity
  \item IHSEVER N NA Severity
  \item IHDTONSET D YYYY-MM-DD Onset Date
  \item IHONGOING N NA Is symptom/condition ongoing?
  \item IHCEASE D YYYY-MM-DD Cease Date
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
}
}
\examples{
\dontrun{
describe(inithealth)
}
}
\keyword{datasets}
