% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{inventory}
\alias{inventory}
\title{An inventory of all ADNI assessments and samples}
\format{
A data frame with 76676 rows and 16 variables
}
\usage{
data(inventory)
}
\description{
An inventory of all ADNI assessments and samples
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID   Participant roster ID
  \item PTID   Participant ID
  \item VISCODE2   Translated visit code
  \item VISCODE   Visit code
  \item TABLE   Source data table or case report form
  \item ORIGPROT   Original study protocol
  \item COLPROT   Study protocol of data collection
  \item EVENT   Event type
  \item EXAMDATE   Exam date
  \item STATUS   Was the assessment conducted (Yes/No) or was CSF collected?
  \item LONI_VISCODE   VISCODE from LONI
  \item MODALITY_LONI   MODALITY from LONI
  \item STUDYID_LONI   STUDYID from LONI
  \item IMAGEUID_LONI   IMAGEUID from LONI
  \item USERDATE_LONI   Date image uploaded to LONI
  \item QCSTAT_LONI   QCstatus
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
