% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{isoprostane}
\alias{isoprostane}
\title{Baseleine isoprostanes data. ADNI Biomarker Core laboratory}
\format{
A data frame with 415 rows and 7 variables
}
\usage{
data(isoprostane)
}
\description{
Baseleine isoprostanes data. ADNI Biomarker Core laboratory
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N -4 Participant roster ID
  \item SITEID N -4 Site ID
  \item VISCODE T -4 Visit code
  \item USERDATE S -4 Date record created
  \item ISO8PGF2A N pg/ml 8-iso-PGF2alpha
  \item ISO812IPF2A N pg/ml 8,12-iso-iPF2alpha
}
}
\keyword{datasets}
