% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{labdata}
\alias{labdata}
\title{Laboratory Data}
\format{
A data frame with 2422 rows and 129 variables
}
\usage{
data(labdata)
}
\description{
Laboratory Data
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item RECNO N  
  \item ACCNO T  
  \item COVVIS T  
  \item EXAMDATE D  Examination Date
  \item AXT117 T  Test AXT117; Thyroid Stim. Hormone
  \item BAT126 T  Test BAT126; Vitamin B12
  \item CMT1 T  Test CMT1; Color
  \item CMT10 T  Test CMT10; Urine Nitrite
  \item CMT11 T  Test CMT11; Leukocyte Esterase
  \item CMT2 T  Test CMT2; Specific Gravity
  \item CMT3 T  Test CMT3; pH
  \item CMT43 T  Test CMT43; Blood (+)
  \item CMT49 T  Test CMT49; Urine Protein (3+)
  \item CMT5 T  Test CMT5; Urine Urobilinogen (new)
  \item CMT6 T  Test CMT6; Urine Ketones
  \item CMT7 T  Test CMT7; Urine Bilirubin
  \item HMT97 T  Test HMT97; Blasts
  \item HMT55 T  Test HMT55; Lymphoma Cell
  \item HMT56 T  Test HMT56; Prolymphocyte
  \item HMT57 T  Test HMT57; Sezary Cell
  \item HMT59 T  Test HMT59; Lymphocyte, Immunoblastic
  \item HMT60 T  Test HMT60; Lymphocyte, Plasmacytoid
  \item HMT61 T  Test HMT61; Lymphocyte, Reactive
  \item HMT62 T  Test HMT62; Lymphoblast
  \item HMT96 T  Test HMT96; Atypical Lymphocytes
  \item HMT63 T  Test HMT63; Promonocyte
  \item HMT64 T  Test HMT64; Monoblast
  \item HMT65 T  Test HMT65; Monocyte, Immature
  \item HMT49 T  Test HMT49; Eosinophil, Immature
  \item HMT48 T  Test HMT48; Basophil, Immature
  \item HMT50 T  Test HMT50; Malighant Cell, NOS
  \item HMT51 T  Test HMT51; WBC Comment:
  \item HMT52 T  Test HMT52; Plasma Cell Precursor
  \item HMT53 T  Test HMT53; Plasma Cell, Immature
  \item HMT54 T  Test HMT54; Plasma Cell, Mature
  \item HMT58 T  Test HMT58; Hairy Cell
  \item HMT70 T  Test HMT70; Nucleated Red Blood Cells
  \item HMT94 T  Test HMT94; Blasts
  \item HMT72 T  Test HMT72; Sezary Cell (Absolute)
  \item HMT80 T  Test HMT80; Lymphoma Cell
  \item HMT81 T  Test HMT81; Prolymphocyte
  \item HMT83 T  Test HMT83; Lymphocyte, Immunoblastic
  \item HMT84 T  Test HMT84; Lymphocyte, Plasmacytoid
  \item HMT85 T  Test HMT85; Lymphocyte, Reactive
  \item HMT86 T  Test HMT86; Lymphoblast
  \item HMT95 T  Test HMT95; Atypical Lymphocytes
  \item HMT87 T  Test HMT87; Promonocyte
  \item HMT88 T  Test HMT88; Monoblast
  \item HMT89 T  Test HMT89; Monocyte, Immature
  \item HMT75 T  Test HMT75; Eosinophil, Immature
  \item HMT74 T  Test HMT74; Basophil, Immature
  \item HMT76 T  Test HMT76; Malignant Cell, NOS
  \item HMT77 T  Test HMT77; Plasma Cell Precursor
  \item HMT78 T  Test HMT78; Plasma Cell, Immature
  \item HMT79 T  Test HMT79; Plasma Cell, Mature
  \item HMT82 T  Test HMT82; Hairy Cell
  \item HMT98 T  Test HMT98; Nucleated Red Blood Cells
  \item HMT99 T  Test HMT99; WBC Comment:
  \item HMT21 T  Test HMT21; Bands (%)
  \item HMT66 T  Test HMT66; Myeloblast
  \item HMT67 T  Test HMT67; Promyelocyte
  \item HMT68 T  Test HMT68; Myelocyte
  \item HMT69 T  Test HMT69; Metamyelocyte
  \item HMT20 T  Test HMT20; Bands
  \item HMT90 T  Test HMT90; Myeloblast
  \item HMT91 T  Test HMT91; Promyelocyte
  \item HMT92 T  Test HMT92; Myelocyte
  \item HMT93 T  Test HMT93; Metamyelocyte
  \item HMT10 T  Test HMT10; Monocytes
  \item HMT100 T  Test HMT100; MCH
  \item HMT102 T  Test HMT102; MCHC
  \item HMT11 T  Test HMT11; Eosinophils
  \item HMT12 T  Test HMT12; Basophils
  \item HMT13 T  Test HMT13; Platelets
  \item HMT15 T  Test HMT15; Neutrophils (%)
  \item HMT16 T  Test HMT16; Lymphocytes (%)
  \item HMT17 T  Test HMT17; Monocytes (%)
  \item HMT18 T  Test HMT18; Eosinophils (%)
  \item HMT19 T  Test HMT19; Basophils (%)
  \item HMT2 T  Test HMT2; Hematocrit
  \item HMT3 T  Test HMT3; RBC
  \item HMT4 T  Test HMT4; MCV
  \item HMT40 T  Test HMT40; Hemoglobin
  \item HMT7 T  Test HMT7; WBC
  \item HMT71 T  Test HMT71; RBC Morphology
  \item HMT8 T  Test HMT8; Neutrophils
  \item HMT9 T  Test HMT9; Lymphocytes
  \item ORT1859 T  Test ORT1859; Trep.Pall.Ab (FTA-ABS) (-70) -38
  \item RCT1 T  Test RCT1; Total Bilirubin
  \item RCT11 T  Test RCT11; Serum Glucose
  \item RCT12 T  Test RCT12; Total Protein
  \item RCT13 T  Test RCT13; Albumin
  \item RCT14 T  Test RCT14; Creatine Kinase
  \item RCT1407 T  Test RCT1407; Alkaline Phosphatase
  \item RCT1408 T  Test RCT1408; LDH
  \item RCT183 T  Test RCT183; Calcium (EDTA)
  \item RCT19 T  Test RCT19; Triglycerides (GPO)
  \item RCT20 T  Test RCT20; Cholesterol (High Performance)
  \item RCT29 T  Test RCT29; Direct Bilirubin
  \item RCT3 T  Test RCT3; GGT
  \item RCT392 T  Test RCT392; Creatinine (Rate Blanked)
  \item RCT4 T  Test RCT4; ALT (SGPT)
  \item RCT5 T  Test RCT5; AST (SGOT)
  \item RCT6 T  Test RCT6; Urea Nitrogen
  \item RCT8 T  Test RCT8; Serum Uric Acid
  \item RCT9 T  Test RCT9; Phosphorus
  \item SLT3 T  Test SLT3; Rapid Plasma Reagin
  \item CMT53 T  
  \item BAT324 T  
  \item UAT13 T  
  \item UAT10 T  
  \item UAT11 T  
  \item UAT2 T  
  \item UAT3 T  
  \item UAT43 T  
  \item UAT49 T  
  \item UAT59 T  
  \item UAT65 T  
  \item UAT7 T  
  \item UAT1 T  
  \item UAT5 T  
  \item UAT6 T  
  \item USERDATE2 S  Date record last updated
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/LABDATA/Cancellation\%20Codes\%2025-Mar-13-1.pdf}
}
\keyword{datasets}
