% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{labtests}
\alias{labtests}
\title{Clinical Laboratory Tests}
\format{
A data frame with 3360 rows and 11 variables
}
\usage{
data(labtests)
}
\description{
Clinical Laboratory Tests
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item LABLOOD N  1. Was blood drawn for safety labs?
  \item LAURINE N  2. Was a urine sample obtained for safety labs?
  \item LAABNORM N  3. Are there any clinically significant laboratory abnormalities that would exclude the participant from the study?
  \item USERDATE2 S  Date record last updated
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
