% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{latent_disease_time}
\alias{latent_disease_time}
\title{Latent Disease Time Estimates, Alzheimer's Therapeutic Research Institute (ATRI) of University of Southern California}
\format{
A data frame with 72310 rows and 9 variables
}
\usage{
data(latent_disease_time)
}
\description{
Latent Disease Time Estimates, Alzheimer's Therapeutic Research Institute (ATRI) of University of Southern California
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 NA Participant roster ID
  \item DX -4 NA Baseline diagnosis
  \item OUTCOME -4 NA Indicates the outcome for which the random_intercept or random_slope is estimated
  \item VARIABLE -4 NA Indicates the type of estimate (latent_time, random_intercept, or random_slope)
  \item ESTIMATE -4 NA The estimates. Note the scale of latent_time is years and the random intercepts and slopes are on a common standardized scale (the outcomes are modeled as Z-scores).
  \item LWR -4 NA Lower limit of 95% credible interval
  \item UPR -4 NA Upper limit of 95% credible interval
  \item RHAT -4 NA Convergence statistic that should be close to 1 (see e.g. https://www.jstatsoft.org/article/view/v076i01/v76i01.pdf)
}
}
\examples{
\dontrun{
describe(latent_disease_time)
}
}
\keyword{datasets}
