% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{lipidomicsdatabase_meiklelab}
\alias{lipidomicsdatabase_meiklelab}
\title{ADMC Lipidomics Meikle Lab Baseline Data Matrix [ADNI1]}
\format{
A data frame with 937 rows and 586 variables
}
\usage{
data(lipidomicsdatabase_meiklelab)
}
\description{
ADMC Lipidomics Meikle Lab Baseline Data Matrix [ADNI1]
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item SAMPLE_ID -4 -4 -4
  \item EID -4 -4 -4
  \item VIALID -4 -4 -4
  \item QCID -4 -4 -4
  \item ANALYSIS_START_DATE -4 -4 -4
  \item ACYLCARNITINE.12.0. -4 nM Measured Lipid
  \item ACYLCARNITINE.13.0. -4 nM Measured Lipid
  \item ACYLCARNITINE.14.0. -4 nM Measured Lipid
  \item ACYLCARNITINE.14.1. -4 nM Measured Lipid
  \item ACYLCARNITINE.14.2. -4 nM Measured Lipid
  \item ACYLCARNITINE.15.0...A. -4 nM Measured Lipid
  \item ACYLCARNITINE.15.0...B. -4 nM Measured Lipid
  \item ACYLCARNITINE.16.0. -4 nM Measured Lipid
  \item ACYLCARNITINE.16.1. -4 nM Measured Lipid
  \item ACYLCARNITINE.17.0...A. -4 nM Measured Lipid
  \item ACYLCARNITINE.17.0...B. -4 nM Measured Lipid
  \item ACYLCARNITINE.18.0. -4 nM Measured Lipid
  \item ACYLCARNITINE.18.1. -4 nM Measured Lipid
  \item ACYLCARNITINE.18.2. -4 nM Measured Lipid
  \item CE.14.0. -4 nM Measured Lipid
  \item CE.15.0. -4 nM Measured Lipid
  \item CE.16.0. -4 nM Measured Lipid
  \item CE.16.1. -4 nM Measured Lipid
  \item CE.16.2. -4 nM Measured Lipid
  \item CE.17.0. -4 nM Measured Lipid
  \item CE.17.1. -4 nM Measured Lipid
  \item CE.18.0. -4 nM Measured Lipid
  \item CE.18.1. -4 nM Measured Lipid
  \item CE.18.2. -4 nM Measured Lipid
  \item CE.18.3. -4 nM Measured Lipid
  \item CE.20.0. -4 nM Measured Lipid
  \item CE.20.1. -4 nM Measured Lipid
  \item CE.20.2. -4 nM Measured Lipid
  \item CE.20.3. -4 nM Measured Lipid
  \item CE.20.4. -4 nM Measured Lipid
  \item CE.20.5. -4 nM Measured Lipid
  \item CE.22.0. -4 nM Measured Lipid
  \item CE.22.1. -4 nM Measured Lipid
  \item CE.22.4. -4 nM Measured Lipid
  \item CE.22.5...N3. -4 nM Measured Lipid
  \item CE.22.5...N6. -4 nM Measured Lipid
  \item CE.22.6. -4 nM Measured Lipid
  \item CE.24.0. -4 nM Measured Lipid
  \item CE.24.1. -4 nM Measured Lipid
  \item CE.24.4. -4 nM Measured Lipid
  \item CE.24.5. -4 nM Measured Lipid
  \item CE.24.6. -4 nM Measured Lipid
  \item CER.D16.1.16.0. -4 nM Measured Lipid
  \item CER.D16.1.18.0. -4 nM Measured Lipid
  \item CER.D16.1.22.0. -4 nM Measured Lipid
  \item CER.D16.1.24.0. -4 nM Measured Lipid
  \item CER.D16.1.24.1. -4 nM Measured Lipid
  \item CER.D17.1.16.0. -4 nM Measured Lipid
  \item CER.D17.1.18.0. -4 nM Measured Lipid
  \item CER.D17.1.20.0. -4 nM Measured Lipid
  \item CER.D17.1.22.0. -4 nM Measured Lipid
  \item CER.D17.1.24.0. -4 nM Measured Lipid
  \item CER.D17.1.24.1. -4 nM Measured Lipid
  \item CER.D18.1.14.0. -4 nM Measured Lipid
  \item CER.D18.1.16.0. -4 nM Measured Lipid
  \item CER.D18.1.17.0. -4 nM Measured Lipid
  \item CER.D18.1.18.0. -4 nM Measured Lipid
  \item CER.D18.1.20.0. -4 nM Measured Lipid
  \item CER.D18.1.22.0. -4 nM Measured Lipid
  \item CER.D18.1.24.0. -4 nM Measured Lipid
  \item CER.D18.1.24.1. -4 nM Measured Lipid
  \item CER.D18.1.26.0. -4 nM Measured Lipid
  \item CER.D18.2.14.0. -4 nM Measured Lipid
  \item CER.D18.2.16.0. -4 nM Measured Lipid
  \item CER.D18.2.17.0. -4 nM Measured Lipid
  \item CER.D18.2.18.0. -4 nM Measured Lipid
  \item CER.D18.2.20.0. -4 nM Measured Lipid
  \item CER.D18.2.22.0. -4 nM Measured Lipid
  \item CER.D18.2.24.0. -4 nM Measured Lipid
  \item CER.D18.2.24.1. -4 nM Measured Lipid
  \item CER.D18.2.26.0. -4 nM Measured Lipid
  \item CER.D19.1.16.0. -4 nM Measured Lipid
  \item CER.D19.1.18.0. -4 nM Measured Lipid
  \item CER.D19.1.20.0. -4 nM Measured Lipid
  \item CER.D19.1.22.0. -4 nM Measured Lipid
  \item CER.D19.1.24.0. -4 nM Measured Lipid
  \item CER.D19.1.24.1. -4 nM Measured Lipid
  \item CER.D19.1.26.0. -4 nM Measured Lipid
  \item CER.D20.1.22.0. -4 nM Measured Lipid
  \item CER.D20.1.24.0. -4 nM Measured Lipid
  \item CER.D20.1.24.1. -4 nM Measured Lipid
  \item CER.D20.1.26.0. -4 nM Measured Lipid
  \item COH -4 nM Measured Lipid
  \item DE.16.0. -4 nM Measured Lipid
  \item DE.18.1. -4 nM Measured Lipid
  \item DE.18.2. -4 nM Measured Lipid
  \item DE.20.4. -4 nM Measured Lipid
  \item DE.20.5. -4 nM Measured Lipid
  \item DE.22.6. -4 nM Measured Lipid
  \item DG.14.0_16.0. -4 nM Measured Lipid
  \item DG.14.0_18.2. -4 nM Measured Lipid
  \item DG.16.0_16.0. -4 nM Measured Lipid
  \item DG.16.0_16.1. -4 nM Measured Lipid
  \item DG.16.0_18.1. -4 nM Measured Lipid
  \item DG.16.0_18.2. -4 nM Measured Lipid
  \item DG.16.0_20.4. -4 nM Measured Lipid
  \item DG.16.0_22.5. -4 nM Measured Lipid
  \item DG.16.0_22.6. -4 nM Measured Lipid
  \item DG.16.1_18.1. -4 nM Measured Lipid
  \item DG.18.0_18.1. -4 nM Measured Lipid
  \item DG.18.0_18.2. -4 nM Measured Lipid
  \item DG.18.0_20.4. -4 nM Measured Lipid
  \item DG.18.1_18.1. -4 nM Measured Lipid
  \item DG.18.1_18.2. -4 nM Measured Lipid
  \item DG.18.1_18.3. -4 nM Measured Lipid
  \item DG.18.1_20.3. -4 nM Measured Lipid
  \item DG.18.1_20.4. -4 nM Measured Lipid
  \item DG.18.2_18.2. -4 nM Measured Lipid
  \item DG.18.2_20.4. -4 nM Measured Lipid
  \item DHCER.D18.0.16.0. -4 nM Measured Lipid
  \item DHCER.D18.0.18.0. -4 nM Measured Lipid
  \item DHCER.D18.0.20.0. -4 nM Measured Lipid
  \item DHCER.D18.0.22.0. -4 nM Measured Lipid
  \item DHCER.D18.0.24.0. -4 nM Measured Lipid
  \item DHCER.D18.0.24.1. -4 nM Measured Lipid
  \item GM1.D18.1.16.0. -4 nM Measured Lipid
  \item GM3.D18.1.16.0. -4 nM Measured Lipid
  \item GM3.D18.1.18.0. -4 nM Measured Lipid
  \item GM3.D18.1.20.0. -4 nM Measured Lipid
  \item GM3.D18.1.22.0. -4 nM Measured Lipid
  \item GM3.D18.1.24.0. -4 nM Measured Lipid
  \item GM3.D18.1.24.1. -4 nM Measured Lipid
  \item HEX2CER.D18.1.16.0. -4 nM Measured Lipid
  \item HEX2CER.D18.1.18.0. -4 nM Measured Lipid
  \item HEX2CER.D18.1.20.0. -4 nM Measured Lipid
  \item HEX2CER.D18.1.22.0. -4 nM Measured Lipid
  \item HEX2CER.D18.1.24.0. -4 nM Measured Lipid
  \item HEX2CER.D18.1.24.1. -4 nM Measured Lipid
  \item HEX3CER.D18.1.16.0. -4 nM Measured Lipid
  \item HEX3CER.D18.1.18.0. -4 nM Measured Lipid
  \item HEX3CER.D18.1.20.0. -4 nM Measured Lipid
  \item HEX3CER.D18.1.22.0. -4 nM Measured Lipid
  \item HEX3CER.D18.1.24.0. -4 nM Measured Lipid
  \item HEX3CER.D18.1.24.1. -4 nM Measured Lipid
  \item HEXCER.D18.1.16.0. -4 nM Measured Lipid
  \item HEXCER.D18.1.18.0. -4 nM Measured Lipid
  \item HEXCER.D18.1.20.0. -4 nM Measured Lipid
  \item HEXCER.D18.1.22.0. -4 nM Measured Lipid
  \item HEXCER.D18.1.24.0. -4 nM Measured Lipid
  \item HEXCER.D18.1.24.1. -4 nM Measured Lipid
  \item LPC.14.0...SN1. -4 nM Measured Lipid
  \item LPC.14.0...SN2. -4 nM Measured Lipid
  \item LPC.15.0...SN1. -4 nM Measured Lipid
  \item LPC.15.0...SN2. -4 nM Measured Lipid
  \item LPC.15.MHDA...SN1....LPC.17.0...SN2. -4 nM Measured Lipid
  \item LPC.15.MHDA...SN1...104_SN1. -4 nM Measured Lipid
  \item LPC.15.MHDA...SN2. -4 nM Measured Lipid
  \item LPC.16.0...SN1. -4 nM Measured Lipid
  \item LPC.16.0...SN2. -4 nM Measured Lipid
  \item LPC.16.1...SN1. -4 nM Measured Lipid
  \item LPC.16.1...SN2. -4 nM Measured Lipid
  \item LPC.17.0...SN1. -4 nM Measured Lipid
  \item LPC.17.1....SN2...A. -4 nM Measured Lipid
  \item LPC.17.1...A...SN1...104_SN1. -4 nM Measured Lipid
  \item LPC.17.1...SN1...A....LPC.17.1...SN2...B. -4 nM Measured Lipid
  \item LPC.17.1...SN1...B. -4 nM Measured Lipid
  \item LPC.18.0...SN1. -4 nM Measured Lipid
  \item LPC.18.0...SN2. -4 nM Measured Lipid
  \item LPC.18.1...SN1. -4 nM Measured Lipid
  \item LPC.18.1...SN2. -4 nM Measured Lipid
  \item LPC.18.2...SN1. -4 nM Measured Lipid
  \item LPC.18.2...SN2. -4 nM Measured Lipid
  \item LPC.18.3...SN1...A..LPC.18.3...SN2...B. -4 nM Measured Lipid
  \item LPC.18.3...SN1...B. -4 nM Measured Lipid
  \item LPC.18.3...SN2...A. -4 nM Measured Lipid
  \item LPC.19.0...A...SN1...104_SN1. -4 nM Measured Lipid
  \item LPC.19.0...SN1...A....LPC.19.0...SN2...B. -4 nM Measured Lipid
  \item LPC.19.0...SN1...B. -4 nM Measured Lipid
  \item LPC.19.0...SN2...A. -4 nM Measured Lipid
  \item LPC.20.0...SN1. -4 nM Measured Lipid
  \item LPC.20.0...SN2. -4 nM Measured Lipid
  \item LPC.20.1...SN1. -4 nM Measured Lipid
  \item LPC.20.1...SN2. -4 nM Measured Lipid
  \item LPC.20.2...SN1. -4 nM Measured Lipid
  \item LPC.20.2...SN2. -4 nM Measured Lipid
  \item LPC.20.3...SN1. -4 nM Measured Lipid
  \item LPC.20.3...SN2. -4 nM Measured Lipid
  \item LPC.20.4...SN1. -4 nM Measured Lipid
  \item LPC.20.4...SN2. -4 nM Measured Lipid
  \item LPC.20.5...SN1. -4 nM Measured Lipid
  \item LPC.20.5...SN2. -4 nM Measured Lipid
  \item LPC.22.0...SN1. -4 nM Measured Lipid
  \item LPC.22.0...SN2. -4 nM Measured Lipid
  \item LPC.22.1...SN1. -4 nM Measured Lipid
  \item LPC.22.1...SN2. -4 nM Measured Lipid
  \item LPC.22.4...SN1. -4 nM Measured Lipid
  \item LPC.22.4...SN2. -4 nM Measured Lipid
  \item LPC.22.5...SN1...N3..LPC.22.5...SN2...N6. -4 nM Measured Lipid
  \item LPC.22.5...SN1...N6. -4 nM Measured Lipid
  \item LPC.22.5...SN2...N3. -4 nM Measured Lipid
  \item LPC.22.6...SN1. -4 nM Measured Lipid
  \item LPC.22.6...SN2. -4 nM Measured Lipid
  \item LPC.24.0...SN1. -4 nM Measured Lipid
  \item LPC.24.0...SN2. -4 nM Measured Lipid
  \item LPC.26.0...SN1. -4 nM Measured Lipid
  \item LPC.26.0...SN2. -4 nM Measured Lipid
  \item LPC.O.16.0. -4 nM Measured Lipid
  \item LPC.O.18.0. -4 nM Measured Lipid
  \item LPC.O.18.1. -4 nM Measured Lipid
  \item LPC.O.20.0. -4 nM Measured Lipid
  \item LPC.O.20.1. -4 nM Measured Lipid
  \item LPC.O.22.0. -4 nM Measured Lipid
  \item LPC.O.22.1. -4 nM Measured Lipid
  \item LPC.O.24.0. -4 nM Measured Lipid
  \item LPC.O.24.1. -4 nM Measured Lipid
  \item LPC.O.24.2. -4 nM Measured Lipid
  \item LPC.P.16.0. -4 nM Measured Lipid
  \item LPC.P.18.0. -4 nM Measured Lipid
  \item LPC.P.18.1. -4 nM Measured Lipid
  \item LPC.P.20.0. -4 nM Measured Lipid
  \item LPE.16.0...SN1. -4 nM Measured Lipid
  \item LPE.16.0...SN2. -4 nM Measured Lipid
  \item LPE.17.0...SN1. -4 nM Measured Lipid
  \item LPE.17.0...SN2. -4 nM Measured Lipid
  \item LPE.18.0...SN1. -4 nM Measured Lipid
  \item LPE.18.0...SN2. -4 nM Measured Lipid
  \item LPE.18.1...SN1. -4 nM Measured Lipid
  \item LPE.18.1...SN2. -4 nM Measured Lipid
  \item LPE.18.2...SN1. -4 nM Measured Lipid
  \item LPE.18.2...SN2. -4 nM Measured Lipid
  \item LPE.20.4...SN1. -4 nM Measured Lipid
  \item LPE.20.4...SN2. -4 nM Measured Lipid
  \item LPE.22.6...SN1. -4 nM Measured Lipid
  \item LPE.22.6...SN2. -4 nM Measured Lipid
  \item LPE.P.16.0. -4 nM Measured Lipid
  \item LPE.P.18.0. -4 nM Measured Lipid
  \item LPE.P.18.1. -4 nM Measured Lipid
  \item LPE.P.20.0. -4 nM Measured Lipid
  \item LPI.18.0...SN1. -4 nM Measured Lipid
  \item LPI.18.0...SN2. -4 nM Measured Lipid
  \item LPI.18.1...SN1. -4 nM Measured Lipid
  \item LPI.18.1...SN2. -4 nM Measured Lipid
  \item LPI.18.2...SN1. -4 nM Measured Lipid
  \item LPI.18.2...SN2. -4 nM Measured Lipid
  \item LPI.20.4...SN1. -4 nM Measured Lipid
  \item LPI.20.4...SN2. -4 nM Measured Lipid
  \item PC..39.5...A. -4 nM Measured Lipid
  \item PC..39.5...B. -4 nM Measured Lipid
  \item PC.14.0_16.0. -4 nM Measured Lipid
  \item PC.14.0_20.4. -4 nM Measured Lipid
  \item PC.14.0_22.6. -4 nM Measured Lipid
  \item PC.15.0_20.3. -4 nM Measured Lipid
  \item PC.15.0_20.4. -4 nM Measured Lipid
  \item PC.15.0_22.6. -4 nM Measured Lipid
  \item PC.15.MHDA_18.1. -4 nM Measured Lipid
  \item PC.15.MHDA_18.2. -4 nM Measured Lipid
  \item PC.15.MHDA_20.4. -4 nM Measured Lipid
  \item PC.15.MHDA_22.6. -4 nM Measured Lipid
  \item PC.16.0_16.0. -4 nM Measured Lipid
  \item PC.16.0_18.0. -4 nM Measured Lipid
  \item PC.16.0_18.1. -4 nM Measured Lipid
  \item PC.16.0_18.2. -4 nM Measured Lipid
  \item PC.16.0_18.3...A. -4 nM Measured Lipid
  \item PC.16.0_18.3...B. -4 nM Measured Lipid
  \item PC.16.0_20.3...A. -4 nM Measured Lipid
  \item PC.16.0_20.3...B. -4 nM Measured Lipid
  \item PC.16.0_20.4. -4 nM Measured Lipid
  \item PC.16.0_20.5. -4 nM Measured Lipid
  \item PC.16.0_22.6. -4 nM Measured Lipid
  \item PC.16.1_18.2. -4 nM Measured Lipid
  \item PC.16.1_20.4. -4 nM Measured Lipid
  \item PC.16.1_22.6. -4 nM Measured Lipid
  \item PC.17.0_18.1. -4 nM Measured Lipid
  \item PC.17.0_18.2. -4 nM Measured Lipid
  \item PC.17.0_20.4. -4 nM Measured Lipid
  \item PC.17.0_22.6. -4 nM Measured Lipid
  \item PC.17.1_18.2. -4 nM Measured Lipid
  \item PC.18.0_18.1. -4 nM Measured Lipid
  \item PC.18.0_18.2. -4 nM Measured Lipid
  \item PC.18.0_20.3. -4 nM Measured Lipid
  \item PC.18.0_20.4. -4 nM Measured Lipid
  \item PC.18.0_22.4. -4 nM Measured Lipid
  \item PC.18.0_22.5...N3..PC.20.1_20.4. -4 nM Measured Lipid
  \item PC.18.0_22.5...N6. -4 nM Measured Lipid
  \item PC.18.0_22.6. -4 nM Measured Lipid
  \item PC.18.1_18.1. -4 nM Measured Lipid
  \item PC.18.1_18.2. -4 nM Measured Lipid
  \item PC.18.1_20.3. -4 nM Measured Lipid
  \item PC.18.1_22.6...A. -4 nM Measured Lipid
  \item PC.18.1_22.6...B. -4 nM Measured Lipid
  \item PC.18.2_18.2. -4 nM Measured Lipid
  \item PC.18.2_20.5. -4 nM Measured Lipid
  \item PC.20.0_20.4. -4 nM Measured Lipid
  \item PC.28.0. -4 nM Measured Lipid
  \item PC.31.0...A. -4 nM Measured Lipid
  \item PC.31.0...B. -4 nM Measured Lipid
  \item PC.32.1. -4 nM Measured Lipid
  \item PC.32.2. -4 nM Measured Lipid
  \item PC.33.0...A. -4 nM Measured Lipid
  \item PC.33.0...B. -4 nM Measured Lipid
  \item PC.33.1. -4 nM Measured Lipid
  \item PC.33.2. -4 nM Measured Lipid
  \item PC.34.5. -4 nM Measured Lipid
  \item PC.35.5. -4 nM Measured Lipid
  \item PC.36.0. -4 nM Measured Lipid
  \item PC.36.6...A. -4 nM Measured Lipid
  \item PC.38.2. -4 nM Measured Lipid
  \item PC.38.4...B. -4 nM Measured Lipid
  \item PC.38.5...A. -4 nM Measured Lipid
  \item PC.38.5...B. -4 nM Measured Lipid
  \item PC.38.6...A. -4 nM Measured Lipid
  \item PC.38.7...C. -4 nM Measured Lipid
  \item PC.40.7...A. -4 nM Measured Lipid
  \item PC.40.8. -4 nM Measured Lipid
  \item PC.O.16.0.16.0. -4 nM Measured Lipid
  \item PC.O.16.0.20.3. -4 nM Measured Lipid
  \item PC.O.16.0.20.4. -4 nM Measured Lipid
  \item PC.O.16.0.22.6. -4 nM Measured Lipid
  \item PC.O.18.0.18.1. -4 nM Measured Lipid
  \item PC.O.18.0.18.2. -4 nM Measured Lipid
  \item PC.O.18.0.20.4. -4 nM Measured Lipid
  \item PC.O.18.0.22.6. -4 nM Measured Lipid
  \item PC.O.18.1.18.1. -4 nM Measured Lipid
  \item PC.O.18.1.18.2. -4 nM Measured Lipid
  \item PC.O.32.1. -4 nM Measured Lipid
  \item PC.O.32.2. -4 nM Measured Lipid
  \item PC.O.34.1. -4 nM Measured Lipid
  \item PC.O.34.2. -4 nM Measured Lipid
  \item PC.O.34.4. -4 nM Measured Lipid
  \item PC.O.35.4. -4 nM Measured Lipid
  \item PC.O.36.0. -4 nM Measured Lipid
  \item PC.O.36.5. -4 nM Measured Lipid
  \item PC.O.38.5. -4 nM Measured Lipid
  \item PC.O.40.5. -4 nM Measured Lipid
  \item PC.O.40.7...A. -4 nM Measured Lipid
  \item PC.O.40.7...B. -4 nM Measured Lipid
  \item PC.P.15.0.20.4...A. -4 nM Measured Lipid
  \item PC.P.15.0.20.4...B. -4 nM Measured Lipid
  \item PC.P.16.0.16.0. -4 nM Measured Lipid
  \item PC.P.16.0.16.1. -4 nM Measured Lipid
  \item PC.P.16.0.18.0. -4 nM Measured Lipid
  \item PC.P.16.0.18.1. -4 nM Measured Lipid
  \item PC.P.16.0.18.2. -4 nM Measured Lipid
  \item PC.P.16.0.18.3. -4 nM Measured Lipid
  \item PC.P.16.0.20.4. -4 nM Measured Lipid
  \item PC.P.16.0.22.6. -4 nM Measured Lipid
  \item PC.P.17.0.20.4...A. -4 nM Measured Lipid
  \item PC.P.17.0.20.4...B. -4 nM Measured Lipid
  \item PC.P.18.0.18.2. -4 nM Measured Lipid
  \item PC.P.18.0.20.4. -4 nM Measured Lipid
  \item PC.P.18.0.22.5. -4 nM Measured Lipid
  \item PC.P.18.0.22.6. -4 nM Measured Lipid
  \item PC.P.18.1.18.1. -4 nM Measured Lipid
  \item PC.P.18.1.22.6. -4 nM Measured Lipid
  \item PC.P.20.0.20.4. -4 nM Measured Lipid
  \item PC.P.36.3. -4 nM Measured Lipid
  \item PC.P.38.5...A. -4 nM Measured Lipid
  \item PC.P.38.5...B. -4 nM Measured Lipid
  \item PE.15.MHDA_18.1. -4 nM Measured Lipid
  \item PE.15.MHDA_18.2. -4 nM Measured Lipid
  \item PE.15.MHDA_20.4. -4 nM Measured Lipid
  \item PE.15.MHDA_22.6. -4 nM Measured Lipid
  \item PE.16.0_16.0. -4 nM Measured Lipid
  \item PE.16.0_16.1. -4 nM Measured Lipid
  \item PE.16.0_18.1. -4 nM Measured Lipid
  \item PE.16.0_18.2. -4 nM Measured Lipid
  \item PE.16.0_18.3...A. -4 nM Measured Lipid
  \item PE.16.0_18.3...B. -4 nM Measured Lipid
  \item PE.16.0_20.3. -4 nM Measured Lipid
  \item PE.16.0_20.4. -4 nM Measured Lipid
  \item PE.16.0_20.5. -4 nM Measured Lipid
  \item PE.16.0_22.6. -4 nM Measured Lipid
  \item PE.16.1_18.2. -4 nM Measured Lipid
  \item PE.16.1_20.4. -4 nM Measured Lipid
  \item PE.17.0_18.1. -4 nM Measured Lipid
  \item PE.17.0_18.2. -4 nM Measured Lipid
  \item PE.17.0_20.4. -4 nM Measured Lipid
  \item PE.17.0_22.6. -4 nM Measured Lipid
  \item PE.18.0_18.1. -4 nM Measured Lipid
  \item PE.18.0_18.2. -4 nM Measured Lipid
  \item PE.18.0_20.3...A. -4 nM Measured Lipid
  \item PE.18.0_20.3...B. -4 nM Measured Lipid
  \item PE.18.0_20.4. -4 nM Measured Lipid
  \item PE.18.0_22.4. -4 nM Measured Lipid
  \item PE.18.0_22.5...N3. -4 nM Measured Lipid
  \item PE.18.0_22.5...N6. -4 nM Measured Lipid
  \item PE.18.0_22.6. -4 nM Measured Lipid
  \item PE.18.1_18.1. -4 nM Measured Lipid
  \item PE.18.1_18.2. -4 nM Measured Lipid
  \item PE.18.1_22.6...A. -4 nM Measured Lipid
  \item PE.18.1_22.6...B. -4 nM Measured Lipid
  \item PE.20.0_20.4. -4 nM Measured Lipid
  \item PE.36.0. -4 nM Measured Lipid
  \item PE.38.5...A. -4 nM Measured Lipid
  \item PE.38.5...B. -4 nM Measured Lipid
  \item PE.O.16.0.18.2. -4 nM Measured Lipid
  \item PE.O.16.0.20.3. -4 nM Measured Lipid
  \item PE.O.16.0.20.4. -4 nM Measured Lipid
  \item PE.O.16.0.22.4. -4 nM Measured Lipid
  \item PE.O.16.0.22.6. -4 nM Measured Lipid
  \item PE.O.18.0.20.4. -4 nM Measured Lipid
  \item PE.O.18.0.22.5...A. -4 nM Measured Lipid
  \item PE.O.18.0.22.6. -4 nM Measured Lipid
  \item PE.O.18.1.18.2. -4 nM Measured Lipid
  \item PE.O.18.1.22.6. -4 nM Measured Lipid
  \item PE.O.34.1. -4 nM Measured Lipid
  \item PE.O.36.5. -4 nM Measured Lipid
  \item PE.O.38.5...A. -4 nM Measured Lipid
  \item PE.O.38.5...B. -4 nM Measured Lipid
  \item PE.P.15.0.20.4...A. -4 nM Measured Lipid
  \item PE.P.15.0.20.4...B. -4 nM Measured Lipid
  \item PE.P.15.0.22.6...A. -4 nM Measured Lipid
  \item PE.P.15.0.22.6...B. -4 nM Measured Lipid
  \item PE.P.16.0.18.1. -4 nM Measured Lipid
  \item PE.P.16.0.18.2. -4 nM Measured Lipid
  \item PE.P.16.0.18.3. -4 nM Measured Lipid
  \item PE.P.16.0.20.3...A. -4 nM Measured Lipid
  \item PE.P.16.0.20.3...B. -4 nM Measured Lipid
  \item PE.P.16.0.20.4. -4 nM Measured Lipid
  \item PE.P.16.0.20.5. -4 nM Measured Lipid
  \item PE.P.16.0.22.4. -4 nM Measured Lipid
  \item PE.P.16.0.22.5...N3. -4 nM Measured Lipid
  \item PE.P.16.0.22.5...N6. -4 nM Measured Lipid
  \item PE.P.16.0.22.6. -4 nM Measured Lipid
  \item PE.P.17.0.20.4...A. -4 nM Measured Lipid
  \item PE.P.17.0.20.4...B. -4 nM Measured Lipid
  \item PE.P.17.0.22.6...A. -4 nM Measured Lipid
  \item PE.P.17.0.22.6...B. -4 nM Measured Lipid
  \item PE.P.18.0.18.1. -4 nM Measured Lipid
  \item PE.P.18.0.18.2. -4 nM Measured Lipid
  \item PE.P.18.0.18.3. -4 nM Measured Lipid
  \item PE.P.18.0.20.3...A. -4 nM Measured Lipid
  \item PE.P.18.0.20.3...B. -4 nM Measured Lipid
  \item PE.P.18.0.20.4. -4 nM Measured Lipid
  \item PE.P.18.0.20.5. -4 nM Measured Lipid
  \item PE.P.18.0.22.4. -4 nM Measured Lipid
  \item PE.P.18.0.22.5...N3. -4 nM Measured Lipid
  \item PE.P.18.0.22.5...N6. -4 nM Measured Lipid
  \item PE.P.18.0.22.6. -4 nM Measured Lipid
  \item PE.P.18.1.18.1. -4 nM Measured Lipid
  \item PE.P.18.1.18.2...A. -4 nM Measured Lipid
  \item PE.P.18.1.18.2...B. -4 nM Measured Lipid
  \item PE.P.18.1.18.3. -4 nM Measured Lipid
  \item PE.P.18.1.20.3...A. -4 nM Measured Lipid
  \item PE.P.18.1.20.3...B. -4 nM Measured Lipid
  \item PE.P.18.1.20.4...A. -4 nM Measured Lipid
  \item PE.P.18.1.20.4...B. -4 nM Measured Lipid
  \item PE.P.18.1.20.5...A. -4 nM Measured Lipid
  \item PE.P.18.1.20.5...B. -4 nM Measured Lipid
  \item PE.P.18.1.22.4. -4 nM Measured Lipid
  \item PE.P.18.1.22.5...A. -4 nM Measured Lipid
  \item PE.P.18.1.22.5...B. -4 nM Measured Lipid
  \item PE.P.18.1.22.6...A. -4 nM Measured Lipid
  \item PE.P.18.1.22.6...B. -4 nM Measured Lipid
  \item PE.P.19.0.20.4...A. -4 nM Measured Lipid
  \item PE.P.19.0.20.4...B. -4 nM Measured Lipid
  \item PE.P.20.0.18.1. -4 nM Measured Lipid
  \item PE.P.20.0.18.2. -4 nM Measured Lipid
  \item PE.P.20.0.20.4. -4 nM Measured Lipid
  \item PE.P.20.0.22.6. -4 nM Measured Lipid
  \item PE.P.20.1.20.4. -4 nM Measured Lipid
  \item PE.P.20.1.22.6. -4 nM Measured Lipid
  \item PG.34.1. -4 nM Measured Lipid
  \item PG.34.2. -4 nM Measured Lipid
  \item PG.36.1. -4 nM Measured Lipid
  \item PG.36.2. -4 nM Measured Lipid
  \item PI..38.5...B. -4 nM Measured Lipid
  \item PI.15.MHDA_18.1..PI.17.0_18.1. -4 nM Measured Lipid
  \item PI.15.MHDA_18.2..PI.17.0_18.2. -4 nM Measured Lipid
  \item PI.15.MHDA_20.4..PI.17.0_20.4. -4 nM Measured Lipid
  \item PI.16.0.16.0. -4 nM Measured Lipid
  \item PI.16.0_16.1. -4 nM Measured Lipid
  \item PI.16.0_20.3...A. -4 nM Measured Lipid
  \item PI.16.0_20.3...B. -4 nM Measured Lipid
  \item PI.16.0_20.4. -4 nM Measured Lipid
  \item PI.18.0_18.1. -4 nM Measured Lipid
  \item PI.18.0_20.2. -4 nM Measured Lipid
  \item PI.18.0_20.3...A. -4 nM Measured Lipid
  \item PI.18.0_20.3...B. -4 nM Measured Lipid
  \item PI.18.0_20.4. -4 nM Measured Lipid
  \item PI.18.0_22.4. -4 nM Measured Lipid
  \item PI.18.0_22.5...N3. -4 nM Measured Lipid
  \item PI.18.0_22.5...N6. -4 nM Measured Lipid
  \item PI.18.0_22.6. -4 nM Measured Lipid
  \item PI.18.1_18.2. -4 nM Measured Lipid
  \item PI.20.0_20.4. -4 nM Measured Lipid
  \item PI.34.1. -4 nM Measured Lipid
  \item PI.36.2. -4 nM Measured Lipid
  \item PI.37.6. -4 nM Measured Lipid
  \item PI.38.5...A. -4 nM Measured Lipid
  \item PI.38.6. -4 nM Measured Lipid
  \item PI.39.6. -4 nM Measured Lipid
  \item PS.36.1. -4 nM Measured Lipid
  \item PS.36.2. -4 nM Measured Lipid
  \item PS.38.3. -4 nM Measured Lipid
  \item PS.38.4. -4 nM Measured Lipid
  \item PS.38.5. -4 nM Measured Lipid
  \item PS.40.5. -4 nM Measured Lipid
  \item PS.40.6. -4 nM Measured Lipid
  \item SM.34.3. -4 nM Measured Lipid
  \item SM.35.2...B. -4 nM Measured Lipid
  \item SM.37.2. -4 nM Measured Lipid
  \item SM.41.0. -4 nM Measured Lipid
  \item SM.41.1...A. -4 nM Measured Lipid
  \item SM.43.1. -4 nM Measured Lipid
  \item SM.43.2...B. -4 nM Measured Lipid
  \item SM.43.2...C. -4 nM Measured Lipid
  \item SM.D16.1.19.0. -4 nM Measured Lipid
  \item SM.D16.1.23.0..SM.D17.1.22.0. -4 nM Measured Lipid
  \item SM.D16.1.24.1. -4 nM Measured Lipid
  \item SM.D17.1.14.0. -4 nM Measured Lipid
  \item SM.D17.1.16.0. -4 nM Measured Lipid
  \item SM.D17.1.24.1. -4 nM Measured Lipid
  \item SM.D18.0.14.0. -4 nM Measured Lipid
  \item SM.D18.0.16.0. -4 nM Measured Lipid
  \item SM.D18.0.22.0. -4 nM Measured Lipid
  \item SM.D18.1.14.0..SM.D16.1.16.0. -4 nM Measured Lipid
  \item SM.D18.1.16.0. -4 nM Measured Lipid
  \item SM.D18.1.17.0..SM.D17.1.18.0. -4 nM Measured Lipid
  \item SM.D18.1.18.0..SM.D16.1.20.0. -4 nM Measured Lipid
  \item SM.D18.1.20.0..SM.D16.1.22.0. -4 nM Measured Lipid
  \item SM.D18.1.22.0..SM.D16.1.24.0. -4 nM Measured Lipid
  \item SM.D18.1.23.0..SM.D17.1.24.0. -4 nM Measured Lipid
  \item SM.D18.1.24.0. -4 nM Measured Lipid
  \item SM.D18.1.24.1. -4 nM Measured Lipid
  \item SM.D18.2.14.0. -4 nM Measured Lipid
  \item SM.D18.2.16.0. -4 nM Measured Lipid
  \item SM.D18.2.17.0. -4 nM Measured Lipid
  \item SM.D18.2.18.0. -4 nM Measured Lipid
  \item SM.D18.2.18.1. -4 nM Measured Lipid
  \item SM.D18.2.20.0. -4 nM Measured Lipid
  \item SM.D18.2.22.0. -4 nM Measured Lipid
  \item SM.D18.2.23.0. -4 nM Measured Lipid
  \item SM.D18.2.24.0. -4 nM Measured Lipid
  \item SM.D19.1.24.1. -4 nM Measured Lipid
  \item SULFATIDE..D18.1..16.0.OH.. -4 nM Measured Lipid
  \item SULFATIDE..D18.1..16.0. -4 nM Measured Lipid
  \item SULFATIDE..D18.1..24.0.OH.. -4 nM Measured Lipid
  \item SULFATIDE..D18.1..24.0. -4 nM Measured Lipid
  \item SULFATIDE..D18.1..24.1.OH.. -4 nM Measured Lipid
  \item SULFATIDE..D18.1..24.1. -4 nM Measured Lipid
  \item TG.48.0...NL.16.0. -4 nM Measured Lipid
  \item TG.48.1...NL.18.1. -4 nM Measured Lipid
  \item TG.48.2...NL.14.1. -4 nM Measured Lipid
  \item TG.48.2...NL.16.0. -4 nM Measured Lipid
  \item TG.48.2...NL.16.1. -4 nM Measured Lipid
  \item TG.48.2...NL.18.1. -4 nM Measured Lipid
  \item TG.48.3...NL.16.1. -4 nM Measured Lipid
  \item TG.48.3...NL.18.2. -4 nM Measured Lipid
  \item TG.49.1...NL.15.0. -4 nM Measured Lipid
  \item TG.49.1...NL.16.0. -4 nM Measured Lipid
  \item TG.49.1...NL.18.1. -4 nM Measured Lipid
  \item TG.50.0...NL.18.0. -4 nM Measured Lipid
  \item TG.50.1...NL.14.0. -4 nM Measured Lipid
  \item TG.50.1...NL.18.1. -4 nM Measured Lipid
  \item TG.50.2...NL.18.0. -4 nM Measured Lipid
  \item TG.50.2...NL.18.1. -4 nM Measured Lipid
  \item TG.50.2...NL.18.2. -4 nM Measured Lipid
  \item TG.50.3...NL.14.1. -4 nM Measured Lipid
  \item TG.50.3...NL.16.1. -4 nM Measured Lipid
  \item TG.50.3...NL.18.1. -4 nM Measured Lipid
  \item TG.50.4...NL.14.0. -4 nM Measured Lipid
  \item TG.51.0...NL.16.0. -4 nM Measured Lipid
  \item TG.51.1...NL.18.1. -4 nM Measured Lipid
  \item TG.51.2...NL.15.0. -4 nM Measured Lipid
  \item TG.51.2...NL.16.0. -4 nM Measured Lipid
  \item TG.51.2...NL.18.1. -4 nM Measured Lipid
  \item TG.52.1...NL.18.0. -4 nM Measured Lipid
  \item TG.52.2...NL.16.0. -4 nM Measured Lipid
  \item TG.52.3...NL.16.1. -4 nM Measured Lipid
  \item TG.52.3...NL.18.2. -4 nM Measured Lipid
  \item TG.52.4...NL.16.0. -4 nM Measured Lipid
  \item TG.52.4...NL.18.1. -4 nM Measured Lipid
  \item TG.53.2...NL.17.0. -4 nM Measured Lipid
  \item TG.54.0...NL.18.0. -4 nM Measured Lipid
  \item TG.54.1...NL.18.1. -4 nM Measured Lipid
  \item TG.54.2...NL.18.0. -4 nM Measured Lipid
  \item TG.54.3...NL.18.1. -4 nM Measured Lipid
  \item TG.54.4...NL.18.0. -4 nM Measured Lipid
  \item TG.54.4...NL.18.2. -4 nM Measured Lipid
  \item TG.54.5...NL.18.1. -4 nM Measured Lipid
  \item TG.54.6...NL.18.2. -4 nM Measured Lipid
  \item TG.56.6...NL.20.4. -4 nM Measured Lipid
  \item TG.56.8...NL.20.4. -4 nM Measured Lipid
  \item TG.58.8...NL.22.6. -4 nM Measured Lipid
  \item TG.O.50.1...NL.16.0. -4 nM Measured Lipid
  \item TG.O.52.0...NL.16.0. -4 nM Measured Lipid
  \item TG.O.52.2...NL.16.0. -4 nM Measured Lipid
  \item UBIQUINONE -4 nM Measured Lipid
}
}
\examples{
\dontrun{
describe(lipidomicsdatabase_meiklelab)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/LIPIDOMICSDATABASE_MEIKLELAB/adni_lipidomicsdatabase_meiklelab_file_versions.pdf}, \url{https://adni.bitbucket.io/reference/docs/LIPIDOMICSDATABASE_MEIKLELAB/meiklelab_methods.pdf}
}
\keyword{datasets}
