% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{loclab}
\alias{loclab}
\title{CSF - Local Lab Results}
\format{
A data frame with 7619 rows and 16 variables
}
\usage{
data(loclab)
}
\description{
CSF - Local Lab Results
}
\details{
A dataset with variables as follows:

\itemize{
  \item SAMPLECOLL N NA Was LP conducted?
  \item SAMPLEDATE D NA Date of Sample Collection
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item SAMPLETIME T  Time of Sample Collection
  \item SENTTIME T  Time sent to Local Lab
  \item CTWHITE N  White Blood Cell Count
  \item CTRED N  Red Blood Cell Count
  \item PROTEIN N  Protein Results
  \item GLUCOSE N  Glucose Results
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
