% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{m24csfcsnt}
\alias{m24csfcsnt}
\title{Month 24 CSF Extension Consent}
\format{
A data frame with 122 rows and 9 variables
}
\usage{
data(m24csfcsnt)
}
\description{
Month 24 CSF Extension Consent
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
  \item INCLPM24 N  Has the participant signed the Informed Consent form for the CSF Extension (2 year visit)?
  \item INCLPDTM24 D  If Yes, date signed.
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
