% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{mayoadirl_mri_fmri}
\alias{mayoadirl_mri_fmri}
\title{Jack Lab - fMRI Network Failure Quotient (NFQ)}
\format{
A data frame with 2996 rows and 33 variables
}
\usage{
data(mayoadirl_mri_fmri)
}
\description{
Jack Lab - fMRI Network Failure Quotient (NFQ)
}
\details{
A dataset with variables as follows:

\itemize{
  \item COLPROT -4 -4 Study protocol of data collection
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item SCANDATE -4 -4 -4
  \item SERIESDATE -4 NA SeriesDate
  \item SERIESTIME -4 NA SeriesTime
  \item MANUFACTURER -4 NA Manufacturer
  \item MANUFACTURERSMODELNAME -4 NA ManufacturersModelName
  \item REPETITIONTIME -4 NA RepetitionTime
  \item ECHOTIME -4 NA EchoTime
  \item SOFTWAREVERSIONS -4 NA SoftwareVersions
  \item SERIESNUMBER -4 NA SeriesNumber
  \item PATIENTSAGE -4 NA PatientsAge
  \item WADDMN_TO_WAVDMN -4 NA wadDMN_to_wavDMN
  \item WADDMN_TO_WPDMN -4 NA wadDMN_to_wpDMN
  \item WADDMN_TO_WVDMN -4 NA wadDMN_to_wvDMN
  \item WAVDMN_TO_WADDMN -4 NA wavDMN_to_wadDMN
  \item WAVDMN_TO_WPDMN -4 NA wavDMN_to_wpDMN
  \item WAVDMN_TO_WVDMN -4 NA wavDMN_to_wvDMN
  \item WPDMN_TO_WADDMN -4 NA wpDMN_to_wadDMN
  \item WPDMN_TO_WAVDMN -4 NA wpDMN_to_wavDMN
  \item WPDMN_TO_WVDMN -4 NA wpDMN_to_wvDMN
  \item WVDMN_TO_WADDMN -4 NA wvDMN_to_wadDMN
  \item WVDMN_TO_WAVDMN -4 NA wvDMN_to_wavDMN
  \item WVDMN_TO_WPDMN -4 NA wvDMN_to_wpDMN
  \item WPDMN_MEDIAN -4 NA wpDMN_median
  \item WADDMN_MEDIAN -4 NA wadDMN_median
  \item WAVDMN_MEDIAN -4 NA wavDMN_median
  \item WVDMN_MEDIAN -4 NA wvDMN_median
  \item NFQ -4 NA calculated network failure quotient
  \item OVERALLQC -4 NA manual QC assessment for the input data
  \item SLICETIMING -4 NA sliceTiming
  \item VERSION -4 -4 -4
  \item RUNDATE -4 CHAR(8) Run Date
  \item IMAGEUID -4 CHAR(var) LONI Image ID
  \item STATUS -4 CHAR(var) analysis status
  \item ADMNRV -4 CHAR(var) Anterior DMN RV
  \item PDMNRV -4 CHAR(var) Posterior DMN RV
  \item DMNRVR -4 CHAR(var) DMN RV-ratio
  \item MAGSTRENG -4 CHAR(var) Mag Field
  \item SERDESC -4 CHAR(var) Series Description
  \item PITCH -4 CHAR(var) Max Pitch Change
  \item ROLL -4 CHAR(var) Max Roll Change
  \item YAW -4 CHAR(var) Max Yaw Change
  \item MMX -4 CHAR(var) Max X change
  \item MMY -4 CHAR(var) Max Y change
  \item MMZ -4 CHAR(var) Max Z change
  \item SLICEORD -4 CHAR(var) temporal slice order
  \item COVVERMIS -4 CHAR(1) vermis coverage
  \item COVCEREBEL -4 CHAR(1) cerebellum coverage
  \item COVOCCIP -4 CHAR(1) occip lobe coverage
  \item COVSUPER -4 CHAR(1) superior aspect coverage
  \item COVTEMPOR -4 CHAR(1) temporal lobe coverage
  \item COVOTHER -4 CHAR(1) other coverage failure
  \item MEANTSNR -4 CHAR(var) mean temporal SNR in brain
  \item MEDTSNR -4 CHAR(var) median temporal SNR in brain
  \item SDTSNR -4 CHAR(var) std deviation of temporal SNR in brain
  \item PENCIL -4 CHAR(var) pencil artifact observed
  \item VENETIAN -4 CHAR(var) venetian blind artifact observed
  \item PHASEDIR -4 CHAR(var) observed phase direction
}
}
\examples{
\dontrun{
describe(mayoadirl_mri_fmri)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/MAYOADIRL_MRI_FMRI/ADNI_Methods_TF_fMRI_Oct31.pdf}
}
\keyword{datasets}
