% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{mayoadirl_mri_fmri_nfq}
\alias{mayoadirl_mri_fmri_nfq}
\title{Jack Lab - fMRI Network Failure Quotient (NFQ)}
\format{
A data frame with 2236 rows and 33 variables
}
\usage{
data(mayoadirl_mri_fmri_nfq)
}
\description{
Jack Lab - fMRI Network Failure Quotient (NFQ)
}
\details{
A dataset with variables as follows:

\itemize{
  \item COLPROT -4 -4 Study protocol of data collection
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item SCANDATE -4 -4 -4
  \item SERIESDATE -4 NA SeriesDate
  \item SERIESTIME -4 NA SeriesTime
  \item MANUFACTURER -4 NA Manufacturer
  \item MANUFACTURERSMODELNAME -4 NA ManufacturersModelName
  \item REPETITIONTIME -4 NA RepetitionTime
  \item ECHOTIME -4 NA EchoTime
  \item SOFTWAREVERSIONS -4 NA SoftwareVersions
  \item SERIESNUMBER -4 NA SeriesNumber
  \item PATIENTSAGE -4 NA PatientsAge
  \item WADDMN_TO_WAVDMN -4 NA wadDMN_to_wavDMN
  \item WADDMN_TO_WPDMN -4 NA wadDMN_to_wpDMN
  \item WADDMN_TO_WVDMN -4 NA wadDMN_to_wvDMN
  \item WAVDMN_TO_WADDMN -4 NA wavDMN_to_wadDMN
  \item WAVDMN_TO_WPDMN -4 NA wavDMN_to_wpDMN
  \item WAVDMN_TO_WVDMN -4 NA wavDMN_to_wvDMN
  \item WPDMN_TO_WADDMN -4 NA wpDMN_to_wadDMN
  \item WPDMN_TO_WAVDMN -4 NA wpDMN_to_wavDMN
  \item WPDMN_TO_WVDMN -4 NA wpDMN_to_wvDMN
  \item WVDMN_TO_WADDMN -4 NA wvDMN_to_wadDMN
  \item WVDMN_TO_WAVDMN -4 NA wvDMN_to_wavDMN
  \item WVDMN_TO_WPDMN -4 NA wvDMN_to_wpDMN
  \item WPDMN_MEDIAN -4 NA wpDMN_median
  \item WADDMN_MEDIAN -4 NA wadDMN_median
  \item WAVDMN_MEDIAN -4 NA wavDMN_median
  \item WVDMN_MEDIAN -4 NA wvDMN_median
  \item NFQ -4 NA calculated network failure quotient
  \item OVERALLQC -4 NA manual QC assessment for the input data
  \item SLICETIMING -4 NA sliceTiming
}
}
\examples{
\dontrun{
describe(mayoadirl_mri_fmri_nfq)
}
}
\keyword{datasets}
