% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{mayoadirl_mri_imageqc}
\alias{mayoadirl_mri_imageqc}
\title{Jack Lab - ADNI GO/2/3 MRI QC}
\format{
A data frame with 21268 rows and 31 variables
}
\usage{
data(mayoadirl_mri_imageqc)
}
\description{
Jack Lab - ADNI GO/2/3 MRI QC
}
\details{
A dataset with variables as follows:

\itemize{
  \item COLPROT -4 -4 Study protocol of data collection
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item SCANDATE -4 -4 -4
  \item SERIES_DATE -4 -4 Series Scan Date
  \item STUDYINSTANCEUID -4 -4 DICOM Study Instance UID
  \item STUDY_OVERALLPASS -4 -4 Study Overall Quality Assessment
  \item STUDY_COMMENTS -4 -4 Study Comments
  \item STUDY_PROTOCOL_STATUS -4 -4 Study Protocol Assessment 
  \item STUDY_PROTOCOL_COMMENT -4 -4 Study Protocol Comments
  \item STUDY_RESCAN_REQUESTED -4 -4 Study Rescan Requested
  \item SERIESINSTANCEUID -4 -4 DICOM Series Instance UID
  \item SERIES_DESCRIPTION -4 -4 Series Description
  \item SERIES_TYPE -4 -4 Series Type
  \item SERIES_QUALITY -4 -4 Series Quality Assessment
  \item SERIES_COMMENTS -4 -4 Series Comments
  \item PROTOCOL_STATUS -4 -4 Series Protocol Assessment
  \item PROTOCOL_COMMENTS -4 -4 Series Protocol Comments
  \item SERIES_SELECTED -4 -4 Series Selected
  \item LONI_STUDY -4 -4 LONI Study ID
  \item LONI_SERIES -4 -4 LONI Series ID
  \item LONI_IMAGE -4 -4 LONI Image ID
  \item STUDY_MEDICAL_ABNORMALITIES -4 -4 Study Medical Abnormalities
  \item SERIAL -4 -4 Serial Comparison
  \item MEDICAL_EXCLUSION -4 -4 Medical Exclusion Assessment
  \item RELEASE_FROM_QUARANTINE -4 -4 Release From Quarantine
  \item PAY_SITE -4 -4 Pay Site
  \item QUALIFICATION -4 -4 Qualification Scan
  \item FIELD_STRENGTH -4 -4 Field Strength
  \item T1_ACCELERATED -4 -4 T1 Accelerated
  \item SERIES_TIME -4 CHAR(12) Series Time
  \item SERIES_NUMBER -4 CHAR(4) Series Number
  \item SOFTWARE_VERSION -4 CHAR(45) Software Version
  \item COIL -4 CHAR(32) Coil
  \item SERIES_COVERAGE_OCCIPITAL -4 CHAR(1) Series Coverage Occipital
  \item SERIES_COVERAGE_VERMIS -4 CHAR(1) Series Coverage Vermis
  \item SERIES_COVERAGE_CEREBELLUM -4 CHAR(1) Series Coverage Cerebellum
  \item SERIES_COVERAGE_SUPERIOR -4 CHAR(1) Series Coverage Superior
  \item SERIES_COVERAGE_TEMPORAL -4 CHAR(1) Series Coverage Temporal
  \item SERIES_COVERAGE_OTHER -4 CHAR(1) Series Coverage Other
  \item FMRI_PHASE -4 CHAR(1) fMRI Phase
  \item FMRI_PENCIL -4 CHAR(1) fMRI Pencil
  \item FMRI_VENETIAN -4 CHAR(1) fMRI Venetian
  \item FMRI_DMN -4 CHAR(1) fMRI DMN
  \item FMRI_YAW -4 CHAR(7) fMRI Yaw
  \item FMRI_ROLL -4 CHAR(7) fMRI Roll
  \item FMRI_PITCH -4 CHAR(7) fMRI Pitch
  \item FMRI_MMZ -4 CHAR(7) fMRI MMZ
  \item FMRI_MMY -4 CHAR(7) fMRI MMY
  \item FMRI_MMX -4 CHAR(7) fMRI MMX
  \item FMRI_MOTION_DISPLAY -4 CHAR(1) fMRI Motion Display
  \item FMRI_TEMPORAL_SLICE_ORDER -4 CHAR(189) fMRI Temporal Slice Order
  \item FMRI_MEAN_SNR -4 CHAR(8) fMRI Mean SNR
  \item FMRI_PHASE_DIRECTION -4 CHAR(6) fMRI Phase Direction
}
}
\examples{
\dontrun{
describe(mayoadirl_mri_imageqc)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/MAYOADIRL_MRI_IMAGEQC/ADNI_data_file_specifications_quality.pdf}, \url{https://adni.bitbucket.io/reference/docs/MAYOADIRL_MRI_IMAGEQC/ADNI_Methods_Mayo_MRI_QC.pdf}
}
\keyword{datasets}
