% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{mayoadirl_mri_mch}
\alias{mayoadirl_mri_mch}
\title{Jack Lab - ADNI MRI MCH}
\format{
A data frame with 19229 rows and 21 variables
}
\usage{
data(mayoadirl_mri_mch)
}
\description{
Jack Lab - ADNI MRI MCH
}
\details{
A dataset with variables as follows:

\itemize{
  \item COLPROT -4 -4 Study protocol of data collection
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item SCANDATE -4 CHAR(8) Examination Date
  \item LONI_STUDY -4 CHAR(v) LONI Study ID
  \item LONI_SERIES -4 CHAR(v) LONI Series ID
  \item LONI_IMG_ID -4 CHAR(v) LONI Image ID
  \item SERIES_NUM -4 CHAR(v) Series Number
  \item CHOSEN -4 CHAR(1) Chosen
  \item STUDY_QUALITY -4 CHAR(2) Study Quality
  \item SERIES_QUALITY -4 CHAR(2) Series Quality
  \item UNIQUEID -4 CHAR(v) Unique ID of MCH
  \item TYPE -4 CHAR(48) Type of finding
  \item STATUS -4 CHAR(48) Status
  \item ATLASREGIONS -4 CHAR(200) Atlas Regions
  \item RASLOCATIONS -4 CHAR(300) RAS Location
  \item NOFINDINGS -4 CHAR(1) No Findings
  \item SCAN_COMMENT -4 CHAR(v) Scan Comments
  \item EVALDATE -4 CHAR(v) Date of last evalutation 
  \item SERIES_UID -4 CHAR(v) Series UID
}
}
\examples{
\dontrun{
describe(mayoadirl_mri_mch)
}
}
\keyword{datasets}
