% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{mayoadirl_mri_quality}
\alias{mayoadirl_mri_quality}
\title{Jack Lab - MRI Quality}
\format{
A data frame with 27366 rows and 31 variables
}
\usage{
data(mayoadirl_mri_quality)
}
\description{
Jack Lab - MRI Quality
}
\details{
A dataset with variables as follows:

\itemize{
  \item COLPROT -4 -4 Study protocol of data collection
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item SCANDATE -4 -4 -4
  \item SERIES_DATE -4 -4 Series Scan Date
  \item STUDYINSTANCEUID -4 -4 DICOM Study Instance UID
  \item STUDY_OVERALLPASS -4 -4 Study Overall Quality Assessment
  \item STUDY_COMMENTS -4 -4 Study Comments
  \item STUDY_PROTOCOL_STATUS -4 -4 Study Protocol Assessment 
  \item STUDY_PROTOCOL_COMMENT -4 -4 Study Protocol Comments
  \item STUDY_RESCAN_REQUESTED -4 -4 Study Rescan Requested
  \item SERIESINSTANCEUID -4 -4 DICOM Series Instance UID
  \item SERIES_DESCRIPTION -4 -4 Series Description
  \item SERIES_TYPE -4 -4 Series Type
  \item SERIES_QUALITY -4 -4 Series Quality Assessment
  \item SERIES_COMMENTS -4 -4 Series Comments
  \item PROTOCOL_STATUS -4 -4 Series Protocol Assessment
  \item PROTOCOL_COMMENTS -4 -4 Series Protocol Comments
  \item SERIES_SELECTED -4 -4 Series Selected
  \item LONI_STUDY -4 -4 LONI Study ID
  \item LONI_SERIES -4 -4 LONI Series ID
  \item LONI_IMAGE -4 -4 LONI Image ID
  \item STUDY_MEDICAL_ABNORMALITIES -4 -4 Study Medical Abnormalities
  \item SERIAL -4 -4 Serial Comparison
  \item MEDICAL_EXCLUSION -4 -4 Medical Exclusion Assessment
  \item RELEASE_FROM_QUARANTINE -4 -4 Release From Quarantine
  \item PAY_SITE -4 -4 Pay Site
  \item QUALIFICATION -4 -4 Qualification Scan
  \item FIELD_STRENGTH -4 -4 Field Strength
  \item T1_ACCELERATED -4 -4 T1 Accelerated
}
}
\examples{
\dontrun{
describe(mayoadirl_mri_quality)
}
}
\keyword{datasets}
