% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{mayoadirl_mri_tbmsyn}
\alias{mayoadirl_mri_tbmsyn}
\title{Jack Lab - TBM-SyN Based Scores}
\format{
A data frame with 2668 rows and 43 variables
}
\usage{
data(mayoadirl_mri_tbmsyn)
}
\description{
Jack Lab - TBM-SyN Based Scores
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 CHAR(10) Participant roster ID
  \item VISCODE -4 CHAR(3) Visit code
  \item VISCODE2 -4 CHAR(3) Translated visit code
  \item EXAMDATE -4 CHAR(10) Examination Date
  \item VERSION -4 CHAR(3) TBMSyN Version Number
  \item PHASEBL -4 CHAR(3) ADNI grant phase baseline
  \item PHASE -4 CHAR(3) ADNI grant phase
  \item VISCODEBL -4 CHAR(3) Visit code baseline
  \item VISCODE2BL -4 CHAR(3) Translated visit code baseline
  \item DXBL -4 CHAR(3) Participant diagnosis baseline
  \item DX -4 CHAR(3) Participant diagnosis
  \item EXAMDATEBL -4 CHAR(10) Examination Date baseline
  \item SERIESTIME_BL -4 CHAR(12) Series Time bseline
  \item SERIESTIME -4 CHAR(12) Series Time
  \item SERIESNUMBER_BL -4 CHAR(4) Series Number baseline
  \item SERIESNUMBER -4 CHAR(4) Series Number
  \item SERIESDESCRIPTION_BL -4 CHAR(34) Series Description baseline
  \item SERIESDESCRIPTION -4 CHAR(34) Series Description
  \item LONI_SERIESID_BL -4 CHAR(7) LONI Series ID baseline
  \item LONI_SERIESID -4 CHAR(7) LONI Series ID
  \item LONI_STUDYID_BL -4 CHAR(5) LONI Study ID baseline
  \item LONI_STUDYID -4 CHAR(5) LONI Study ID
  \item LONI_IMAGEID_BL -4 CHAR(7) LONI Image ID baseline
  \item LONI_IMAGEID -4 CHAR(7) LONI Image ID
  \item SERIESINSTANCEUID_BL -4 CHAR(48) DICOM SeriesInstanceUID baseline
  \item SERIESINSTANCEUID -4 CHAR(48) DICOM SeriesInstanceUID
  \item FIELDSTRENGTH_BL -4 CHAR(5) Field Strength
  \item FIELDSTRENGTH -4 CHAR(5) Field Strength
  \item MANUFACTURER_BL -4 CHAR(30) MRI Scanner Manufacturer baseline
  \item MANUFACTURER -4 CHAR(30) MRI Scanner Manufacturer
  \item MODELNAME_BL -4 CHAR(30) MRI Scanner Model Name baseline
  \item MODELNAME -4 CHAR(30) MRI Scanner Model Name
  \item COILNAME_BL -4 CHAR(30) Receive Coil Name baseline
  \item COILNAME -4 CHAR(30) Receive Coil Name
  \item SOFTWAREVERSIONS_BL -4 CHAR(30) MRI Scanner Software Version baseline
  \item SOFTWAREVERSIONS -4 CHAR(30) MRI Scanner Software Version
  \item QUALITY_BL -4 CHAR(1) Series Quality baseline
  \item QUALITY -4 CHAR(1) Series Quality
  \item STUDY_OVERALLPASS_BL -4 CHAR(1) Study Overall Quality Assessment baseline
  \item STUDY_OVERALLPASS -4 CHAR(1) Study Overall Quality Assessment
  \item RUNDATE -4 CHAR(10) Run Date
  \item STATUS -4 CHAR(8) Status
  \item TBMSYNSCOR -4 CHAR(var) Finding Comment
  \item ACCELERATED -4 -4 -4
  \item LONIUIDBL -4 CHAR(7) LONI Unique ID Baseline Scan
  \item LONIUID -4 CHAR(7) LONI Unique ID
}
}
\examples{
\dontrun{
describe(mayoadirl_mri_tbmsyn)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/MAYOADIRL_MRI_TBMSYN/ADNI_Mayo_Methods_TBM_SyN_20130206v1_1.pdf}
}
\keyword{datasets}
