% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{medhist}
\alias{medhist}
\title{Medical History}
\format{
A data frame with 3083 rows and 37 variables
}
\usage{
data(medhist)
}
\description{
Medical History
}
\details{
A dataset with variables as follows:

\itemize{
  \item MH15ADRUG N  15a.  Duration of abuse (years):
  \item MH15BDRUG N  15b.  Period of time since end of abuse (years):
  \item MH14AALCH N NA 14a. During period of alcohol abuse, estimate the average number of drinks per day
  \item MH14BALCH N NA 14b. Duration of abuse (years)
  \item MH14CALCH N NA 14c. Period of time since end of abuse (years)
  \item MH16ASMOK N NA 16a. During periods of smoking, the average number of packs/day
  \item MH16BSMOK N NA 16b. Duration (years)
  \item MH16CSMOK N NA 16c. If no longer smoking, provide period of time since stopped smoking (years)
  \item SOURCE N NA Information Source
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item MHSOURCE N  Information Source
  \item MHPSYCH N  1. Psychiatric
  \item MH2NEURL N  2. Neurologic (other than AD)
  \item MH3HEAD N  3. Head, Eyes, Ears, Nose and Throat
  \item MH4CARD N  4. Cardiovascular
  \item MH5RESP N  5. Respiratory
  \item MH6HEPAT N  6. Hepatic
  \item MH7DERM N  7. Dermatologic-Connective Tissue
  \item MH8MUSCL N  8. Musculoskeletal
  \item MH9ENDO N  9. Endocrine-Metabolic
  \item MH10GAST N  10. Gastrointestinal
  \item MH11HEMA N  11. Hematopoietic-Lymphatic
  \item MH12RENA N  12. Renal-Genitourinary
  \item MH13ALLE N  13. Allergies or Drug Sensitivities
  \item MH14ALCH N  14. Alcohol Abuse
  \item MH15DRUG N  15. Drug Abuse
  \item MH16SMOK N  16. Smoking
  \item MH17MALI N  17. Malignancy
  \item MH18SURG N  18. Major Surgical Procedures
  \item MH19OTHR N  19. Other (if none, select "No")
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
