% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{mesoscale}
\alias{mesoscale}
\title{Meso Scale Diagnostics Analysis of Residual CSF Samples}
\format{
A data frame with 250 rows and 11 variables
}
\usage{
data(mesoscale)
}
\description{
Meso Scale Diagnostics Analysis of Residual CSF Samples
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item DRAWDTE -4 -4 -4
  \item DRAWTIME -4 -4 -4
  \item RECDTE -4 -4 -4
  \item ABETA38 -4 pg/ml Amyloid beta pepetide (x-38)
  \item ABETA40 -4 pg/ml Amyloid peptide beta (x-40)
  \item ABETA42 -4 pg/ml Amyloid beta peptide (x-42)
  \item TAU -4 pg/ml total Tau protein
}
}
\examples{
\dontrun{
describe(mesoscale)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/MESOSCALE/ADNI_Methods_MSD.pdf}
}
\keyword{datasets}
