% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{mmse}
\alias{mmse}
\title{Mini Mental State Exam}
\format{
A data frame with 13039 rows and 62 variables
}
\usage{
data(mmse)
}
\description{
Mini Mental State Exam
}
\details{
A dataset with variables as follows:

\itemize{
  \item DONE N NA Was assessment/procedure done?
  \item NDREASON N NA If No, reason not done
  \item SOURCE N NA Information Source
  \item DATE D NA Date collected
  \item WORDLIST N NA Which list was used?
  \item WORD1 N NA Ball (alt: Apple)
  \item WORD2 N NA Flag (alt: Penny)
  \item WORD3 N NA Tree (alt: Table)
  \item MMLTR1 T NA Indicate 1st letter
  \item MMLTR2 T NA Indicate 2nd letter
  \item MMLTR3 T NA Indicate 3rd letter
  \item MMLTR4 T NA Indicate 4th letter
  \item MMLTR5 T NA Indicate 5th letter
  \item MMLTR6 T NA Indicate 6th letter (if given)
  \item MMLTR7 T NA Indicate 7th letter (if given)
  \item WORLDSCORE N NA Score:  World Backwards
  \item WORD1DL N NA Ball (alt: Apple)
  \item WORD2DL N NA Flag (alt: Penny)
  \item WORD3DL N NA Tree (alt: Table)
  \item MMRECALL N  Which list was used?
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item MMDATE N  1. What is today's date?
  \item MMYEAR N  2. What is the year?
  \item MMMONTH N  3. What is the month?
  \item MMDAY N  4. What day of the week is today?
  \item MMSEASON N  5. What season is it?
  \item MMHOSPIT N  6. What is the name of this hospital (clinic, place)?
  \item MMFLOOR N  7. What floor are we on?
  \item MMCITY N  8. What town or city are we in?
  \item MMAREA N  9. What county (district, borough, area) are we in?
  \item MMSTATE N  10. What state are we in?
  \item MMBALL N  11. Ball
  \item MMFLAG N  12. Flag
  \item MMTREE N  13. Tree
  \item MMTRIALS N  13a. Enter number of trials
  \item MMD N  14. D
  \item MML N  15. L
  \item MMR N  16. R
  \item MMO N  17. O
  \item MMW N  18. W
  \item MMBALLDL N  19. Ball
  \item MMFLAGDL N  20. Flag
  \item MMTREEDL N  21. Tree
  \item MMWATCH N  22. Show the participant a wrist watch and ask "What is this?"
  \item MMPENCIL N  23. Repeat for pencil.
  \item MMREPEAT N  24. Say, "Repeat after me: no ifs, ands, or buts."
  \item MMHAND N  25. Takes paper in right hand.
  \item MMFOLD N  26. Folds paper in half.
  \item MMONFLR N  27. Puts paper on floor.
  \item MMREAD N  28. Present the piece of paper which reads, "CLOSE YOUR EYES," and say: "Read this and do what it says."
  \item MMWRITE N  29. Give the participant a blank piece of paper and say: "Write a sentence."
  \item MMDRAW N  30. Present the participant with the Construction Stimulus page. Say, "Copy this design."
  \item MMSCORE N  MMSE TOTAL SCORE
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
