% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{moca}
\alias{moca}
\title{MoCA}
\format{
A data frame with 7696 rows and 52 variables
}
\usage{
data(moca)
}
\description{
MoCA
}
\details{
A dataset with variables as follows:

\itemize{
  \item MCDATE D YYYY-MM-DD Date collected
  \item SOURCE N NA Information Source
  \item MOCA          N  MOCA                           
  \item RID N NA Participant roster ID
  \item SITEID N NA Site ID
  \item VISCODE T NA Visit code
  \item USERDATE S NA Date record created
  \item USERDATE2 S NA Date record last updated
  \item TRAILS N NA Trails
  \item CUBE N NA Copy Cube
  \item CLOCKCON N NA Draw Clock - Contour
  \item CLOCKNO N NA Draw Clock - Number
  \item CLOCKHAN N NA Draw Clock - Hands
  \item LION N NA Lion
  \item RHINO N NA Rhinoceros
  \item CAMEL N NA Camel
  \item IMMT1W1 N NA Immediate (#1): Face
  \item IMMT1W2 N NA Immediate (#1): Velvet
  \item IMMT1W3 N NA Immediate (#1): Church
  \item IMMT1W4 N NA Immediate (#1): Daisy
  \item IMMT1W5 N NA Immediate (#1): Red
  \item IMMT2W1 N NA Immediate (#2): Face
  \item IMMT2W2 N NA Immediate (#2): Velvet
  \item IMMT2W3 N NA Immediate (#2): Church
  \item IMMT2W4 N NA Immediate (#2): Daisy
  \item IMMT2W5 N NA Immediate (#2): Red
  \item DIGFOR N NA Digits Forward
  \item DIGBACK N NA Digits Backward
  \item LETTERS N NA List of Letters/Tapping:  # Errors
  \item SERIAL1 N NA Serial 7: 1st Subtraction
  \item SERIAL2 N NA Serial 7: 2nd Subtraction
  \item SERIAL3 N NA Serial 7: 3rd Subtraction
  \item SERIAL4 N NA Serial 7: 4th Subtraction
  \item SERIAL5 N NA Serial 7: 5th Subtraction
  \item REPEAT1 N NA Repeat Sentence: I only know that John is the one to help today.
  \item REPEAT2 N NA Repeat Sentence: The cat always hid under the couch when dogs were in the room.
  \item FFLUENCY N NA Letter Fluency - F: Total number of correct words
  \item ABSTRAN N NA Abstraction: train-bicycle
  \item ABSMEAS N NA Abstraction: watch-ruler
  \item DELW1 N NA Delayed: Face
  \item DELW2 N NA Delayed: Velvet
  \item DELW3 N NA Delayed: Church
  \item DELW4 N NA Delayed: Daisy
  \item DELW5 N NA Delayed: Red
  \item DATE N NA Date
  \item MONTH N NA Month
  \item YEAR N NA Year
  \item DAY N NA Day
  \item PLACE N NA Place
  \item CITY N NA City
}
}
\keyword{datasets}
