% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{modhach}
\alias{modhach}
\title{Modified Hachinski}
\format{
A data frame with 3455 rows and 19 variables
}
\usage{
data(modhach)
}
\description{
Modified Hachinski
}
\details{
A dataset with variables as follows:

\itemize{
  \item SOURCE N NA Information Source
  \item DATE D NA Date Collected
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item HMONSET N  1. Abrupt Onset of Dementia
  \item HMSTEPWS N  2. Stepwise Deterioration of Dementia
  \item HMSOMATC N  3. Somatic Complaints
  \item HMEMOTIO N  4. Emotional Incontinence
  \item HMHYPERT N  5. History of Hypertension
  \item HMSTROKE N  6. History of Stroke
  \item HMNEURSM N  7. Focal Neurologic Symptoms
  \item HMNEURSG N  8. Focal Neurologic Signs
  \item HMSCORE N  TOTAL SCORE
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
