% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{monelg}
\alias{monelg}
\title{Monitor Eligibility}
\format{
A data frame with 4619 rows and 10 variables
}
\usage{
data(monelg)
}
\description{
Monitor Eligibility
}
\details{
A dataset with variables as follows:

\itemize{
  \item MPIN T NA I certify that I am using my own personal username and pin and understand that by using my username and pin, I am authorizing this electronic signature to be the legally binding equivalent of my handwritten signature executed on paper. - Enter Your PIN -
  \item RID N NA Participant roster ID
  \item SITEID N NA Site ID
  \item VISCODE T NA Visit code
  \item USERDATE S NA Date record created
  \item USERDATE2 S NA Date record last updated
  \item MINELG N NA Please confirm participant's eligibility for enrollment in the protocol.
  \item MELGDATE D NA Confirmation Date
}
}
\keyword{datasets}
