% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{monrev}
\alias{monrev}
\title{Monitor Review}
\format{
A data frame with 10733 rows and 12 variables
}
\usage{
data(monrev)
}
\description{
Monitor Review
}
\details{
A dataset with variables as follows:

\itemize{
  \item MONCONDUCT N NA Review conducted:
  \item SDV N NA Was this visit 100% Source Doc Verified (SDV)?
  \item RSDV N NA If No, were the minimum set of forms requiring SDV at this visit reviewed?
  \item RID N NA Participant roster ID
  \item SITEID N NA Site ID
  \item VISCODE T NA Visit code
  \item USERDATE S NA Date record created
  \item USERDATE2 S NA Date record last updated
  \item MONPASS N NA Please indicate below.
  \item MONDATE D NA Approval Date:
}
}
\keyword{datasets}
