% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{mri3gometa}
\alias{mri3gometa}
\title{3T MRI Scan Information}
\format{
A data frame with 417 rows and 23 variables
}
\usage{
data(mri3gometa)
}
\description{
3T MRI Scan Information
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N NA Participant roster ID
  \item SITEID N NA Site ID
  \item VISCODE T NA Visit code
  \item USERDATE S NA Date record created
  \item USERDATE2 S NA Date record last updated
  \item RECNO N NA 
  \item MMCONDCT N NA Was the scan conducted?
  \item MMREASON N NA Reason why the scan was not conducted:
  \item EXAMDATE D NA Examination Date
  \item MOTION N NA Patient Motion Problems:
  \item MALFUNC N NA Scanner Malfunction:
  \item OTRDEV N NA Other Protocol Variations
  \item MMTRNSFR N NA Was data transferred to LONI within 24 hours of scan?
  \item MMTRNDATE D NA Transfer Date
  \item MMARCHIVE N NA Data Archived Locally
  \item MMLPDONE N NA Was a Lumbar Puncture completed prior to the MRI scan?
  \item MMLPINTER N NA If Yes, What was the interval between LP and MRI?
  \item EYEASK4 N NA Was the subject instructed to open their eyes?
  \item EYEKEEP4 N NA Did the subject keep their eyes open (MRI Tech:  ask the subject right after the scan)
  \item EYEASK6 N NA Was the subject instructed to open their eyes?
  \item EYEKEEP6 N NA Did the subject keep their eyes open (MRI Tech:  ask the subject right after the scan)
}
}
\keyword{datasets}
