% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{mri3meta}
\alias{mri3meta}
\title{3T MRI Scan Information}
\format{
A data frame with 940 rows and 23 variables
}
\usage{
data(mri3meta)
}
\description{
3T MRI Scan Information
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item RECNO N  
  \item MMCONDCT N  Was the scan conducted?
  \item MMREASON N  Reason why the scan was not conducted:
  \item EXAMDATE D  Examination Date
  \item MMSCOUT N  1. Tri-Planar Scout (if available, otherwise use an axial scout)
  \item MMSMPRAGE N  2. Straight Sagittal MPRAGE Sequence
  \item MMRMPRAGE N  3. Repeat Straight Sagittal MPRAGE Sequence
  \item MMB1HEAD N  4. B1 Calibration Head Coil Scan (Only applicable for phased array head coil on GE and Siemens systems)
  \item MMB1BODY N  5. B1 Calibration Body Coil Scan (Only applicable for phased array head coil on GE and Siemens systems)
  \item MMECHO N  6. Straight Axial Fast or Turbo Spin Echo
  \item MMPHAN N  7. In new exam; Perform ADNI QC Scan. Localizer and Straight Sagittal MPRAGE (with increased slice thickness to ensure phantom coverage)
  \item MMTRNSFR N  Was data transferred to LONI within 24 hours of scan?
  \item MMTRNDATE D  Transfer Date
  \item MMARCHIVE N  Data Archived Locally
  \item MMLPDONE N  9. Was a Lumbar Puncture completed prior to the MRI scan?
  \item MMLPINTER N  If Yes, What was the interval between LP and MRI?
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
