% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{mri_infarcts}
\alias{mri_infarcts}
\title{MRI Infarcts}
\format{
A data frame with 8598 rows and 14 variables
}
\usage{
data(mri_infarcts)
}
\description{
MRI Infarcts
}
\details{
A dataset with variables as follows:

\itemize{
  \item COLPROT -4 -4 Study protocol of data collection
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4  Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item MRI.DATE2 -4 -4 -4
  \item MRI.DATE1 -4 -4 -4
  \item ANALYZE.DATE -4 -4 -4
  \item SIDE -4 Text Side of Infarct, if detected
  \item SIZE -4 Text Size of Infarct, if detected
  \item CVA.NUMBER -4 -4 -4
  \item CVA.LOCATION -4 -4 -4
  \item CVA.TYPE -4 -4 -4
  \item XYZ -4 -4 -4
  \item IMGFILE -4 -4 -4
  \item USERDATE -4  Date record created
  \item EXAMDATE -4  Examination Date
  \item STROKE_READ_DATE -4  Date Image Evaluated
  \item INFARCT_NUMBER -4 integer Number detected
  \item BRAIN_REGION -4 integer Anatomical Location
  \item STROKE_TYPE -4 -4 -4
  \item LOCATION_XYZ -4 Text Image Space Location
}
}
\examples{
\dontrun{
describe(mri_infarcts)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/MRI_INFARCTS/UCD\%20ADNI\%20MRI\%20Infarct\%20Assessment\%20Method.pdf}
}
\keyword{datasets}
