% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{mrib1calib}
\alias{mrib1calib}
\title{MRI B1 Calibration}
\format{
A data frame with 4480 rows and 14 variables
}
\usage{
data(mrib1calib)
}
\description{
MRI B1 Calibration
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item RECNO N  
  \item EXAMDATE D  Examination Date
  \item FLDSTRENG N  Field Strength
  \item COILTYPE T  Coil Type
  \item SCAN N  Scan
  \item LONISID N  LONI Study ID
  \item PRESENT N  
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
