% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{mrimppro}
\alias{mrimppro}
\title{MRI MPRAGE Process}
\format{
A data frame with 3269 rows and 14 variables
}
\usage{
data(mrimppro)
}
\description{
MRI MPRAGE Process
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item RECNO N  
  \item EXAMDATE D  Examination Date
  \item FLDSTRENG N  Field Strength
  \item COILTYPE T  Coil Type
  \item SCAN N  Scan
  \item LONISID N  LONI Study ID
  \item LONIUID T  LONI Unique Series ID selected for processing. This field identifies which scan the MRI Core has selected to be the preferred scan to be used for processing.  MPRAGE Rankings: The MPRAGE rankings are done in a qualitative manner.A scan with a 1 ranking is considered a high quality scan.  A 2 ranking is considered a medium quality scan.  3 ranking is a low quality scan.3 is the lowest ranking a scan can receive and still be considered a passing scan.  Anything below 3 would be a failure for quality. The rankings are found in the MRIMPRANK table in the RANK field.  Choosing MPRAGES: For scans ranked 2 and 3 the scan ranked 2 (whatever is the lower ranking of the two) should always be the "preferred/chosen" MPRAGE.  If the rankings are tied in ranking the QC-er would chose the "sharper of the two". If the two scans are identical the first MPRAGE was always chosen.
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
