% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{mrinclusio}
\alias{mrinclusio}
\title{MRI Subject Inclusion}
\format{
A data frame with 9799 rows and 34 variables
}
\usage{
data(mrinclusio)
}
\description{
MRI Subject Inclusion
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
  \item RECNO N  
  \item EXAMDATE D  Examination Date
  \item LONISID N  LONI Study ID
  \item FLDSTRENG N  Field Strength
  \item SCAN N  Scan
  \item QCTYPE N  QC Type
  \item NOMEDICAL N  Image Medical Issues (check all that apply):
  \item SURGERY N  
  \item HEMMORHAGE N  
  \item DEVANOMALY N  
  \item LESION N  
  \item ATROPHY N  
  \item INFARCTION N  
  \item TRAUMA N  
  \item METALLIC N  
  \item NPH N  
  \item EDEMA N  
  \item OTHER N  
  \item INCLUSION N  Inclusion
  \item STEREOMARK N  Stereotactic Marker present
  \item SERIAL N  Serial scan
  \item MEDEXCL N  Medical Exclusion (see Comments for details)
  \item PASS N  Pass:
  \item QUARANTINE N  Release from Quarantine:
  \item PAYPATIENT N  Pay Patient:
  \item PAYSITE N  Pay Site:
  \item RESCAN N  Rescan
}
}
\keyword{datasets}
