% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{mriphantom}
\alias{mriphantom}
\title{MRI Phantom}
\format{
A data frame with 4141 rows and 15 variables
}
\usage{
data(mriphantom)
}
\description{
MRI Phantom
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item RECNO N  
  \item EXAMDATE D  Examination Date
  \item FLDSTRENG N  Field Strength
  \item COILTYPE T  Coil Type
  \item SCAN N  Scan
  \item LONISID N  LONI Study ID
  \item LONIPSID N  Phantom - LONI Study ID
  \item PRESENT N  Phantom Present
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
