% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{mriquality}
\alias{mriquality}
\title{MRI Quality}
\format{
A data frame with 11720 rows and 28 variables
}
\usage{
data(mriquality)
}
\description{
MRI Quality
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item RECNO N  
  \item EXAMDATE D  Examination Date
  \item FLDSTRENG N  Field Strength
  \item COILTYPE T  Coil Type
  \item SCAN N  Scan
  \item LONISID N  LONI Study ID
  \item LONIUID T  LONI Unique Series ID
  \item SERIESNUMB N  
  \item SERIESDESC T  
  \item INBGR N  
  \item OUTBGR N  
  \item INFLOW N  
  \item OUTFLOW N  
  \item INOTHER N  
  \item OUTOTHER N  
  \item WRAP N  
  \item HEADCOVER N  
  \item INHOMOGEN N  
  \item IPMOTION N  
  \item MARKER N  
  \item PASS N  
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
