% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{mriread}
\alias{mriread}
\title{MRI Clinical Read}
\format{
A data frame with 4106 rows and 15 variables
}
\usage{
data(mriread)
}
\description{
MRI Clinical Read
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item MRREAD N  Was the 1.5T MRI done?
  \item MRUPLD I  Upload 1.5T MRI Clinical Read:
  \item MRCOMPAT N  Is the MRI compatible with the Inclusion/Exclusion Criteria?
  \item MRCOMSPE T  If No, explain:
  \item MR3TREAD N  Was a clinical read of the 3T MRI done?
  \item MR3TUPLD I  Upload 3T MRI Clinical Read :
  \item MRSUITABL N  Is the participant clinically suitable to remain in the study?
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
