% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{mriserial}
\alias{mriserial}
\title{MRI Serial}
\format{
A data frame with 3204 rows and 19 variables
}
\usage{
data(mriserial)
}
\description{
MRI Serial
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item RECNO N  
  \item EXAMDATE D  Examination Date
  \item FLDSTRENG N  Field Strength
  \item COILTYPE T  Coil Type
  \item SCAN N  Scan
  \item LONISID N  LONI Study ID
  \item LONIUID T  LONI Unique Series ID
  \item BLLONIUID T  LONI Unique Series ID for the baseline scan.
  \item BLSDATE D  Baseline scan date.
  \item SAMESUBJ N  
  \item REGISTRAT N  
  \item GROSSINHOM N  
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
