% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{neurobat}
\alias{neurobat}
\title{Neuropsychological Battery}
\format{
A data frame with 15490 rows and 82 variables
}
\usage{
data(neurobat)
}
\description{
The Neuropsychological Battery table contains data for multiple assessments, including:  Logical Memory, Rey Auditory Verbal Learning Test, Clock Drawing, Clock Copying, Category Fluency, Trail Making Test, Boston Naming Test, ANART, and Digit Span.  At any given visit, only a subset of those tests are administered.  The Digit Span was dropped after ADNI1.  The Category Fluency - Animals was retained, but Vegetables were dropped after ADNI1.  A Spanish language option for the Boston Naming Test became available for participants testing in Spanish beginning with ADNIGO.  The data dictionary for this table can be difficult to parse to identify which variables belong to which assessment.  For clarity, the field/variable names associated with each test are identified here:   

Logical Memory - Immediate Recall:  LMSTORY, LIMMTOTAL, LIMMEND

Logical Memory - Delayed Recall: LDELBEGIN, LDELTOTAL, LDELCUE.

Rey Auditory Verbal Learning Test: AVTOT1, AVERR1, AVTOT2, AVERR2, AVTOT3, AVERR3, AVTOT4, AVERR4, AVTOT5, AVERR5, AVTOT6, AVERR6, AVTOTB, AVERRB, AVENDED

Rey Auditory Verbal Learning Test - Delayed:  AVDELBEGAN, AVDEL30MIN, AVDELERR1, AVDELTOT, AVDELERR2 

Clock Drawing: CLOCKCIRC, CLOCKSYM, CLOCKNUM, CLOCKHAND, CLOCKTIME, CLOCKSCOR

Clock Copying: COPYCIRC, COPYSYM, COPYNUM, COPYHAND, COPYTIME, COPYSCOR

Category Fluency (Vegetables ADNI1 only): CATANIMSC, CATANPERS, CATANINTR, CATVEGESC, CATVGPERS, CATVGINTR

Trail Making Test: TRAASCOR, TRAAERRCOM, TRAAERROM, TRABSCOR, TRABERRCOM, TRABERROM

Boston Naming Test (Spanish version provided in ADNIGO/2): BNTND, BNTSPONT, BNTSTIM, BNTCSTIM, BNTPHON, BNTCPHON, BNTTOTAL

American National Adult Reading Test (ANART): ANARTND, ANARTERR

Digit Span (ADNI1 only): DSPANFOR, DSPANFLTH, DSPANBAC, DSPANBLTH
}
\details{
A dataset with variables as follows:

\itemize{
  \item SOURCE N NA Information Source
  \item DATE D NA Date collected
  \item MINTUNCUED N NA MINT Total Uncued Correct 
  \item MINTSEMCUE N NA MINT Total Correct - with Semantic Cue
  \item ANART N NA Was ANART conducted?  
  \item MINTTOTAL N NA MINT Total Correct (Uncued + Correct with Semantic cue)
  \item RAVLT.immediate       N  RAVLT Immediate (sum of 5 trials)    
  \item RAVLT.learning        N  RAVLT Learning (trial 5 - trial 1)   
  \item RAVLT.forgetting      N  RAVLT Forgetting (trial 5 - delayed) 
  \item RAVLT.perc.forgetting N  RAVLT Percent Forgetting             
  \item LIMMEND T HHMM Time Ended (24-hour clock)
  \item LDELBEGIN T HHMM Time Began (24-hour clock)
  \item AVENDED T HHMM AVLT Time Trial B Ended
  \item AVDELBEGAN T HHMM Time Began (24-hour clock)
  \item LMSTORY N NA Which Logical Memory story was used?
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item CLOCKCIRC N  1. Approximately circular face
  \item CLOCKSYM N  2. Symmetry of number placement
  \item CLOCKNUM N  3. Correctness of numbers
  \item CLOCKHAND N  4. Presence of the two hands
  \item CLOCKTIME N  5. Presence of the two hands, set to ten after eleven
  \item CLOCKSCOR N  Total Score
  \item COPYCIRC N  1. Approximately circular face
  \item COPYSYM N  2. Symmetry of number placement
  \item COPYNUM N  3. Correctness of numbers
  \item COPYHAND N  4. Presence of the two hands
  \item COPYTIME N  5. Presence of the two hands, set to ten after eleven
  \item COPYSCOR N  Total Score
  \item LIMMTOTAL N  Logical Memory - Immediate Recall <br />Total Number of Story Units Recalled:
  \item AVTOT1 N  Trial 1 Total
  \item AVERR1 N  Total Intrusions
  \item AVTOT2 N  Trial 2 Total
  \item AVERR2 N  Total Intrusions
  \item AVTOT3 N  Trial 3 Total
  \item AVERR3 N  Total Intrusions
  \item AVTOT4 N  Trial 4 Total
  \item AVERR4 N  Total Intrusions
  \item AVTOT5 N  Trial 5 Total
  \item AVERR5 N  Total Intrusions
  \item AVTOT6 N  Trial 6 Total
  \item AVERR6 N  Total Intrusions
  \item AVTOTB N  List B Total
  \item AVERRB N  Total Intrusions
  \item DSPANFOR N  Forward: Total Correct
  \item DSPANFLTH N  Forward: Length
  \item DSPANBAC N  Backward: Total Correct
  \item DSPANBLTH N  Backward: Length
  \item CATANIMSC N  Category Fluency (Animals) - Total Correct
  \item CATANPERS N  Category Fluency (Animals) - Perseverations
  \item CATANINTR N  Category Fluency (Animals) - Intrusions
  \item CATVEGESC N  Category Fluency (Vegetables) - Total Correct
  \item CATVGPERS N  Category Fluency (Vegetables) - Perseverations
  \item CATVGINTR N  Category Fluency (Vegetables) - Intrusions
  \item TRAASCOR N  Part A - Time to Complete
  \item TRAAERRCOM N  Errors of Commission
  \item TRAAERROM N  Errors of Omission
  \item TRABSCOR N  Part B - Time to complete
  \item TRABERRCOM N  Errors of Commission
  \item TRABERROM N  Errors of Omission
  \item DIGITSCOR N  Total Correct
  \item LDELTOTAL N  Logical Memory - Delayed Recall <br />Total Number of Story Units Recalled:
  \item LDELCUE N  Cue used?
  \item BNTND T  Check here if:
  \item BNTSPONT N  1. Number of spontaneously given correct responses
  \item BNTSTIM N  2. Number of semantic cues given
  \item BNTCSTIM N  3. Number of correct responses following a semantic cue
  \item BNTPHON N  4. Number of phonemic cues given
  \item BNTCPHON N  5. Number of correct responses following a phonemic cue
  \item BNTTOTAL N  Total Number Correct (1+3)
  \item AVDEL30MIN N  30 Minute Delay Total
  \item AVDELERR1 N  Total Intrusions
  \item AVDELTOT N  Recognition Score
  \item AVDELERR2 N  Recognition Errors
  \item ANARTND T  Check here if:
  \item ANARTERR N  ANART Total Score (Total # of errors)
  \item USERDATE2 S  Date record last updated
}
}
\examples{
neurobat$RAVLT.immediate <- apply(neurobat[, paste('AVTOT', 1:5, sep = '')], 
  1, sum, na.rm = FALSE)
neurobat$RAVLT.learning <- neurobat$AVTOT5-neurobat$AVTOT1
neurobat$RAVLT.forgetting <- neurobat$AVTOT5-neurobat$AVDEL30MIN
neurobat$RAVLT.perc.forgetting <- 100*neurobat$RAVLT.forgetting / neurobat$AVTOT5
neurobat$RAVLT.perc.forgetting <- ifelse(neurobat$AVTOT5 == 0, NA, neurobat$RAVLT.perc.forgetting)
neurobat$RAVLT <- neurobat$RAVLT.learning
label(neurobat$RAVLT) <- 'RAVLT Learning'

}
\references{
Estevez-Gonzalez, A., Kulisevsky, J., Boltes, A., Otermin, P., & Garcia-Sanchez, C. (2003). Rey verbal learning test is a useful tool for differential diagnosis in the preclinical phase of Alzheimer's disease: comparison with mild cognitive impairment and normal aging. \emph{International Journal of Geriatric Psychiatry}. 18 (11), 1021.
}
\keyword{datasets}
