% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{neuroexm}
\alias{neuroexm}
\title{Neurological Exam}
\format{
A data frame with 3674 rows and 22 variables
}
\usage{
data(neuroexm)
}
\description{
Neurological Exam
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item NXVISUAL N  1. Significant Visual Impairment
  \item NXAUDITO N  2. Significant Auditory Impairment
  \item NXTREMOR N  3. Tremor
  \item NXCONSCI N  4. Level of Consciousness
  \item NXNERVE N  5. Cranial Nerves
  \item NXMOTOR N  6. Motor Strength
  \item NXFINGER N  7a. Cerebellar - Finger to Nose
  \item NXHEEL N  7b. Cerebellar - Heel to Shin
  \item NXSENSOR N  8. Sensory
  \item NXTENDON N  9. Deep Tendon Reflexes
  \item NXPLANTA N  10. Plantar Reflexes
  \item NXGAIT N  11. Gait
  \item NXOTHER N  12. Other
  \item NXABNORM N  14. Based on Neurological Examination, clinician must check appropriate box below:
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
