% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{neuropath}
\alias{neuropath}
\title{NACC Neuropathology Data Form}
\format{
A data frame with 100 rows and 161 variables
}
\usage{
data(neuropath)
}
\description{
See \url{https://www.alz.washington.edu/NONMEMBER/NP/npded10.pdf} and \url{https://www.alz.washington.edu/NONMEMBER/NP/npded9.pdf}
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item NPFORMMO  NA 
  \item NPFORMDY  NA 
  \item NPFORMYR  NA 
  \item NPDAGE text NA Age at death (years)
  \item NPDODYR  NA 
  \item NPPMIH  NA 
  \item NPFIX dropdown NA 8. Fixative 
  \item NPFIXX text NA Specify:
  \item NPWBRWT text NA a. Whole brain weight (if half brain, multiply weight by two) (9999=unknown)
  \item NPWBRF dropdown NA b. Does the value in Question 9a represent fresh or fixed weight? (Check One)
  \item NPGRCCA dropdown NA c1. Cerebral cortex atrophy
  \item NPGRLA dropdown NA c2. Lobar atrophy (significant frontal and/or temporal atrophy)
  \item NPGRHA dropdown NA c3. Hippocampus atrophy
  \item NPGRSNH dropdown NA c4. Substantia nigra hypopigmentation
  \item NPGRLCH dropdown NA c5. L. ceruleus hypopigmentation
  \item NPAVAS dropdown NA c6. Atherosclerosis (of the circle of Willis)
  \item NPTAN dropdown NA a. Tau antibody (Check One)
  \item NPTANX text NA Specify:
  \item NPABAN dropdown NA b. Amyloid beta antibody (Check One)
  \item NPABANX text NA Specify:
  \item NPASAN dropdown NA c. Alpha synuclein antibody
  \item NPASANX text NA Specify:
  \item NPTDPAN dropdown NA d. TDP-43 antibody (Check One)
  \item NPTDPANX text NA Specify:
  \item NPHISMB dropdown NA e1. Modified Bielschowsky
  \item NPHISG dropdown NA e2. Gallyas
  \item NPHISSS dropdown NA e3. Other Silver stain
  \item NPHIST dropdown NA e4. Thioflavin
  \item NPHISO dropdown NA e5. Other
  \item NPHISOX text NA Specify:
  \item NPTHAL dropdown NA a. Thal phase for amyloid plaques by immunohisto-chemistry (IHC) (A score - Check One)
  \item NPBRAAK dropdown NA b. Braak stage for neurofibrillary degeneration (B score - Check One)
  \item NPNEUR dropdown NA c. CERAD score for density of neocortical neuritic plaque (plaques with argrophilic dystrophic neurities, with or without dense amyloid cores). Score without respect to age or diagnosis. (C score - Check One)
  \item NPADNC dropdown NA d. NIA-AA Alzheimer's disease neuropathologic change (ADNC) (Check One)
  \item NPDIFF dropdown NA e1. CERAD semi-quantitative score for diffuse plaques (plaques with non-compact amyloid and no apparent dystrophic neurites). Score from the neocortical field with the highest plaque density and without respect to age or diagnosis. (Check One)
  \item NPAMY dropdown NA e2. Cerebral amyloid angiopathy (Check One)
  \item NPINF dropdown NA a. Old infarcts observed grossly, including lacunes? (Check One)
  \item NPINF1A text NA Number
  \item NPINF1B text NA Size of largest (greatest dimension in cm)
  \item NPINF1D text NA Size of next (greatest dimension in cm)
  \item NPINF1F text NA Size of next (greatest dimension in cm)
  \item NPINF2A text NA Number
  \item NPINF2B text NA Size of largest (greatest dimension in cm)
  \item NPINF2D text NA Size of next (greatest dimension in cm)
  \item NPINF2F text NA Size of next (greatest dimension in cm)
  \item NPINF3A text NA Number
  \item NPINF3B text NA Size of largest (greatest dimension in cm)
  \item NPINF3D text NA Size of next (greatest dimension in cm)
  \item NPINF3F text NA Size of next (greatest dimension in cm)
  \item NPINF4A text NA Number
  \item NPINF4B text NA Size of largest (greatest dimension in cm)
  \item NPINF4D text NA Size of next (greatest dimension in cm)
  \item NPINF4F text NA Size of next (greatest dimension in cm)
  \item NPHEMO dropdown NA b. Were single or multiple old hemorrhages observed grossly?
  \item NPHEMO1 dropdown NA b1. Subdural or epidural hemorrhage
  \item NPHEMO2 dropdown NA b2. Primary parenchymal hemorrhage Include those >5mm. If ?5mm, include as microbleed; see Question 12d.
  \item NPHEMO3 dropdown NA b3. Secondary parenchymal hemorrhage (e.g., tumor, vascular malformation)
  \item NPOLD dropdown NA c. Old microinfarcts (not observed grossly)? (Check One)
  \item NPOLD1 dropdown NA c1. Number in screening sections of cerebral cortex (gray matter of cerebral cortex)
  \item NPOLD2 dropdown NA c2. Number in screening sections of subcortical white matter and periventricular white matter
  \item NPOLD3 dropdown NA c3. Number in screening sections of subcortical gray matter
  \item NPOLD4 dropdown NA c4. Number in brainstem and cerebellum
  \item NPOLDD dropdown NA d. Old cerebral microbleeds? (Check One) Include old hemorrahages that are greater than or equal to 5mn.
  \item NPOLDD1 dropdown NA d1. Number in screening sections of cerebral cortex
  \item NPOLDD2 dropdown NA d2. Number in screening sections of subcortical white matter and periventricular white matter
  \item NPOLDD3 dropdown NA d3. Number in screening sections of subcortical gray matter
  \item NPOLDD4 dropdown NA d4. Number in brainstem and cerebellum
  \item NPARTER dropdown NA e. Arteriolosclerosis? (Check One) (Assessed in subcortical white or gray matter)
  \item NPWMR dropdown NA f. White matter rarefaction? (Check One) (H&E or myelin stain may be used)
  \item NPPATH dropdown NA g. Other pathologic changes related to ischemic or vascular disease not previously specified?
  \item NPNEC dropdown NA g1. Laminar necrosis
  \item NPPATH2 dropdown NA g2. Acute neuronal necrosis
  \item NPPATH3 dropdown NA g3. Acute/subacute gross infarcts
  \item NPPATH4 dropdown NA g4. Acute/subacute microinfarcts
  \item NPPATH5 dropdown NA g5. Acute/subacute gross hemorrhage
  \item NPPATH6 dropdown NA g6. Acute/subacute microhemorrhage
  \item NPPATH7 dropdown NA g7. Vascular malformation of any type
  \item NPPATH8 dropdown NA g8. Aneurysm of any type
  \item NPPATH9 dropdown NA g9. Vasculitis of any type
  \item NPPATH10 dropdown NA g10. CADASIL
  \item NPPATH11 dropdown NA g11. Mineralization of blood vessels
  \item NPPATHO dropdown NA g12. Other
  \item NPPATHOX text NA Specify:
  \item NPLBOD dropdown NA Is there evidence of Lewy body pathology? (Check One)
  \item NPNLOSS dropdown NA (Check One)
  \item NPHIPSCL dropdown NA (Check One)
  \item NPTDPA dropdown NA a. Spinal cord
  \item NPTDPB dropdown NA b. Amygdala
  \item NPTDPC dropdown NA c. Hippocampus
  \item NPTDPD dropdown NA d. Entorhinal/inferior temporal cortex
  \item NPTDPE dropdown NA e. Neocortex
  \item NPFTDTAU dropdown NA a. FTLD with tau pathology (FTLD-tau) or other tauopathy (Check One)
  \item NPPICK dropdown NA b1. FTLD-tau (PiD)
  \item NPFTDT2 dropdown NA b2. Other 3R tauopathy (Includes MAPT mutation tauopathy)
  \item NPCORT dropdown NA b3. FTLD-tau (CBD)
  \item NPPROG dropdown NA b4. FTLD-tau (PSP)
  \item NPFTDT5 dropdown NA b5. Argyrophilic grains
  \item NPFTDT6 dropdown NA b6. Other 4R tauopathy (Includes sporadic multiple systems tauopathy, globular glial tauopathy, MAPT mutation tauopathy)
  \item NPFTDT7 dropdown NA b7. Chronic traumatic encephalopathy
  \item NPFTDT8 dropdown NA b8. Amyotrophic lateral sclerosis (ALS)/Parkinsonism-dementia complex of Guam
  \item NPFTDT9 dropdown NA b9. Tangle dominant disease
  \item NPFTDT10 dropdown NA b10. Other 3R + 4F tauopathy (Includes unclassifiable, focal, glial only, MAPT mutation tauopathy, NOS)
  \item NPFTDTDP dropdown NA c. FTLD with TDP-43 pathology (FTLD-TDP)? (Check One)
  \item NPALSMND dropdown NA d. ALS/motor neuron disease (MND) present? (Check One) 
  \item NPOFTD dropdown NA e. Other FTLD? (Check One)
  \item NPOFTD1 dropdown NA f1. Atypical FTLD-U (aFTLD-U)
  \item NPOFTD2 dropdown NA f2. NIFID (neuronal intermediate filament inclusions disease)
  \item NPOFTD3 dropdown NA f3. BIBD (basophilic inclusion body disease)
  \item NPOFTD4 dropdown NA f4. FTLD-UPS (ubiquitin-proteasome system [ubiquitin or p62 positive, tau/TDP-43/FUS negative inclusions])
  \item NPOFTD5 dropdown NA f5. FTLD-NOS (includes dementia lacking distinctive histology (DLDH) and FTLD with no inclusions (FTLD-NI) detected by tau, TDP-43, or ubiquitin/p62 IHC)
  \item NPPDXA dropdown NA a. Pigment-spheroid degeneration/NBIA
  \item NPPDXB dropdown NA b. Multiple system atrophy
  \item NPPDXC dropdown NA c. Prion disease
  \item NPPDXD dropdown NA d. Trinucleotide disease (Huntington disease, SCA, other)
  \item NPPDXE dropdown NA e. Malformation of cortical development
  \item NPPDXF dropdown NA f. Metabolic/storage disorder of any type
  \item NPPDXG dropdown NA g. WM disease, leukodystrophy
  \item NPPDXH dropdown NA h. WM disease, multiple sclerosis or other demyelinating disease
  \item NPPDXI dropdown NA i. Contusion/traumatic brain injury of any type, acute
  \item NPPDXJ dropdown NA j. Contusion/traumatic brain injury of any type, chronic
  \item NPPDXK dropdown NA k. Neoplasm, primary
  \item NPPDXL dropdown NA l. Neoplasm, metastatic
  \item NPPDXM dropdown NA m. Infectious process of any type (encephalitis, abscess, etc.)
  \item NPPDXN dropdown NA n. Herniation, any site
  \item NPPDXO dropdown NA o. Trisomy 21/Down syndrome
  \item NPPDXP dropdown NA p. AD-related genes (dominantly inherited); do not include APOE or other polymorphisms or genetic risk factors.
  \item NPPDXQ dropdown NA q. FTLD-related genes (dominantly inherited); do not include polymorphisms or genetic risk factors.
  \item NPPDXR dropdown NA r. Other (Specify)
  \item NPPDXRX text NA Specify: 
  \item NPPDXS dropdown NA s. Other (Specify)
  \item NPPDXSX text NA Specify:
  \item NPPDXT dropdown NA t. Other (Specify)
  \item NPPDXTX text NA Specify:
  \item NPBNKA dropdown NA a. Banked frozen brain or half brain
  \item NPBNKB dropdown NA b. Banked frozen wedge of cerebellum or other sample for future DNA prep
  \item NPBNKC dropdown NA c. Formalin- or paraformaldehyde-fixed brain
  \item NPBNKD dropdown NA d. Paraffin-embedded blocks of brain regions
  \item NPBNKE dropdown NA e. Banked postmortem CSF
  \item NPBNKF dropdown NA f. Banked postmortem blood or serum
  \item NPBNKG dropdown NA g. Banked DNA
  \item NPFAUT dropdown NA h. Full autopsy performed?
  \item NPFAUT1 text NA h1. 
  \item NPFAUT2 text NA h2.
  \item NPFAUT3 text NA h3.
  \item NPFAUT4 text NA h4.
  \item FORMVER  NA 
  \item NPARTAG dropdown NA 
  \item NPATGSEV dropdown NA 
  \item NPATGAMY dropdown NA 
  \item NPATGAM1 dropdown NA 
  \item NPATGAM2 dropdown NA 
  \item NPATGAM3 dropdown NA 
  \item NPATGAM4 dropdown NA 
  \item NPATGAM5 dropdown NA 
  \item NPATGFRN dropdown NA 
  \item NPATGFR1 dropdown NA 
  \item NPATGFR2 dropdown NA 
  \item NPATGFR3 dropdown NA 
  \item NPATGFR4 dropdown NA 
  \item NPSEX N -4 Subject's sex 
  \item ADCID N -4 center id
  \item NPFORMVER N -4 NP form version
  \item NACCMOD N -4 Month of death
  \item NACCYOD N -4 year of death
  \item NACCNBRN N -4 Normal brain/No major pathology
  \item NACCAVAS N -4 Severity of gross findings-atherosclerosis of the circle of Willis
  \item NACCBRAA N -4 Braak stage for neurofibrillary degeneration (B score)
  \item NACCNEUR N -4 Density of neocortical neuritic plaques (CERAD score) (C score) 
  \item NACCDIFF N -4 Density of diffuse plaques (CERAD semiquantitative score)
  \item NACCAMY N -4 Cerebral amyloid angiopathy
  \item NACCINF N -4 Infarcts and lacunes
  \item NACCMICR N -4 Microinfarcts
  \item NACCHEM N -4 Hemorrhages and microbleeds
  \item NACCARTE N -4 Arteriolosclerosis
  \item NACCNEC N -4 Laminar necrosis
  \item NACCVASC N -4 Ischemic, hemorrhagic, or vascular pathology present
  \item NACCLEWY N -4 Lewy body pathology
  \item NACCPICK N -4 FTLD-tau subtype Pick's (PID) 
  \item NACCCBD N -4 FTLD-tau subtype corticobasal degeneration (CBD)
  \item NACCPROG N -4 FTLD-tau subtype progressive supranuclear palsy (PSP)
  \item NACCPRIO N -4 Prion disease
  \item NACCOTHP N -4 Other pathologic diagnosis
  \item NACCWRI1 N -4 First other pathologic diagnosis write-in
  \item NACCWRI2 N -4 Second other pathologic diagnosis write-in
  \item NACCWRI3 N -4 Third other pathologic diagnosis write-in
  \item NACCDOWN N -4 Down syndrome
  \item NACCBNKF N -4 Banked frozen brain
  \item NACCFORM N -4 Formalin or paraformaldehyde -fixed brain
  \item NACCPARA N -4 Paraffin-embedded blocks or brain regions
  \item NACCCSFP N -4 Banked postmortem CSF
  \item NPLINF N -4 Are one or more large artery cerebral infarcts present?
  \item NPMICRO N -4 Are one or more cortical microinfarcts (including "granular atrophy") present?
  \item NPLAC N -4 Are one or more lacunes (small artery infarcts and/or hemmorrhages) present?
  \item NPHEM N -4 Are single or multiple hemmorhages present?
  \item NPART N -4 Is subcortical arteriosclerotic leukoencephalopathy present?
  \item NPSCL N -4 Is medial temporal lobe sclerosis (including hippocampal sclerosis) present?
  \item NPOANG N -4 Is another type of angiopathy (e.g., CADASIL or arteritis) present?
  \item NPFRONT N -4 Frontotemporal dementia and Parkinsonism with tau-positive or argyrophiilic inclusions
  \item NPTAU N -4 Tauopathy, other (e.g., tangle-only dementia and argyrophilic grain dementia):
  \item NPFTD N -4 FTD with ubiquitin-positive (tau-negative) inclusions
  \item NPFTDNO N -4 Is ther FTD with no distinctive histopathology (tau-negative, ubiquitin-negative, and no agryrophilic inclusions)?
  \item NPFTDSPC N -4 Was FTD "not otherwise specified" present (e.g., "immunostaining for ubiquitin and tau not done")?
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/NEUROPATH/ADNI_Methods_Neuropathology_Core_03-06-2018.pdf}, \url{https://adni.bitbucket.io/reference/docs/NEUROPATH/ADNI_Neuropathology_Core_Methods_5_5_2021.pdf}, \url{https://adni.bitbucket.io/reference/docs/NEUROPATH/ADNI_Neuropathology_Methods_9_23_2020-final.pdf}, \url{https://adni.bitbucket.io/reference/docs/NEUROPATH/np11-ded.pdf}, \url{https://adni.bitbucket.io/reference/docs/NEUROPATH/Perrin_ADNI_Methods_10_13_2022_FINAL.pdf}
}
\author{
\email{betsy@wubios.wustl.edu}
}
\keyword{datasets}
