% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{npconsent}
\alias{npconsent}
\title{Neuropathology Extension Consent}
\format{
A data frame with 50 rows and 14 variables
}
\usage{
data(npconsent)
}
\description{
Neuropathology Extension Consent
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
  \item NPINFO N  Has participant been given information regarding autopsy/brain donation?
  \item NPAPPROVE N  Has the ADNI provisional autopsy consent form been signed?
  \item NPAPDT D  If "Yes", date signed.
  \item NPREASON T  If "No", please give reason
  \item NPOTHERAP N  Has provisional autopsy consent form been signed as part of a study other than ADNI?
  \item NPOTHERDT D  If "Yes", date signed.
  \item NPOTHERCOM T  If "Yes", please provide additional information regarding the "other" autopsy consent, including the name of the other study.
}
}
\keyword{datasets}
