% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{npi}
\alias{npi}
\title{Neuropsychiatric Inventory Examination}
\format{
A data frame with 6940 rows and 162 variables
}
\usage{
data(npi)
}
\description{
Neuropsychiatric Inventory Examination
}
\details{
A dataset with variables as follows:

\itemize{
  \item SPID N NA Study Partner ID:
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
  \item EXAMDATE T  Examination Date
  \item NPISOURC N  Information source
  \item NPIA N  A. Delusions
  \item NPIA1 N  1. Does {P} believe that he/she is in danger - that others are planning to hurt him/her?
  \item NPIA2 N  2. Does {P} believe that others are stealing from him/her?
  \item NPIA3 N  3. Does {P} believe that his/her spouse is having an affair?
  \item NPIA4 N  4. Does {P} believe that unwelcome guests are living in his/her house?
  \item NPIA5 N  5. Does {P} believe that his/her spouse or others are not who they claim to be?
  \item NPIA6 N  6. Does {P} believe that his/her house is not his/her home?
  \item NPIA7 N  7. Does {P} believe that family members plan to abandon him/her?
  \item NPIA8 N  8. Does {P} believe that television or magazine figures are actually present in the home?  [Does he/she try to talk to or interact with them?]
  \item NPIA9 N  9. Does {P} believe any other unusual things that I haven't asked about?
  \item NPIA10A N  10a. Frequency Ratings
  \item NPIA10B N  10b. Severity Ratings
  \item NPIA10C N  10c. Caregiver Distress
  \item NPIATOT N  A. Delusions: Item score
  \item NPIB N  B.  Hallucinations
  \item NPIB1 N  1. Does {P} describe hearing voices or act as if he/she hears voices?
  \item NPIB2 N  2. Does {P} talk to people who are not there?
  \item NPIB3 N  3. Does {P} describe seeing things not seen by others or behave as if he/she is seeing things not seen by others (people, animals, lights, etc.)?
  \item NPIB4 N  4. Does {P} report smelling odors not smelled by others?
  \item NPIB5 N  5. Does {P} describe feeling things on his/her skin or otherwise appear to be feeling things crawling or touching him/her?
  \item NPIB6 N  6. Does {P} describe tastes that are without any known cause?
  \item NPIB7 N  7. Does {P} describe any other unusual sensory experiences?
  \item NPIB8A N  8a.  Frequency Ratings
  \item NPIB8B N  8b. Severity Ratings
  \item NPIB8C N  8c. Caregiver Distress
  \item NPIBTOT N  B. Hallucinations: Item score
  \item NPIC N  C: Agitation/Aggression
  \item NPIC1 N  1. Does {P} get upset with those trying to care for him/her or resist activities such as bathing or changing clothes?
  \item NPIC2 N  2. Is {P} stubborn, having to have things his/her way?
  \item NPIC3 N  3. Is {P} uncooperative, resistive to help from others?
  \item NPIC4 N  4. Does {P} have any other behaviors that make him/her hard to handle?
  \item NPIC5 N  5. Does {P} shout or curse angrily?
  \item NPIC6 N  6. Does {P} slam doors, kick furniture, throw things?
  \item NPIC7 N  7. Does {P} attempt to hurt or hit others?
  \item NPIC8 N  8. Does {P} have any other aggressive or agitated behaviors?
  \item NPIC9A N  9a. Frequency Ratings
  \item NPIC9B N  9b. Severity Ratings
  \item NPIC9C N  9c. Caregiver Distress
  \item NPICTOT N  C. Agitation/Aggression: Item score
  \item NPID N  D: Depression/Dysphoria
  \item NPID1 N  1. Does {P} have periods of tearfulness or sobbing that seem to indicate sadness?
  \item NPID2 N  2. Does {P} say or act as if he/she is sad or low in spirits?
  \item NPID3 N  3. Does {P} put him/herself down or say that he/she feels like a failure?
  \item NPID4 N  4. Does {P} say that he/she is a bad person or deserves to be punished?
  \item NPID5 N  5. Does {P} seem very discouraged or say that he/ she has not future?
  \item NPID6 N  6. Does {P} say he/she is a burden to the family or that the family would be better off without him/her?
  \item NPID7 N  7. Does {P} express a wish for death or talk about killing him/herself?
  \item NPID8 N  8. Does {P} show any other signs of depression or sadness?
  \item NPID9A N  9a. Frequency Ratings
  \item NPID9B N  9b. Severity Ratings
  \item NPID9C N  9c. Caregiver Distress
  \item NPIDTOT N  D. Depression/Dysphoria: Item score
  \item NPIE N  E. Anxiety
  \item NPIE1 N  1. Does {P} say that he/she is worried about planned events?
  \item NPIE2 N  2. Does {P} have periods of feeling shaky, unable to relax, or feeling excessively tense?
  \item NPIE3 N  3. Does {P} have periods of [or complain of] shortness of breath, gasping, or sighing for no apparent reason other than nervousness?
  \item NPIE4 N  4. Does {P} complain of butterflies in his/her stomach, or of racing or pounding of the heart in association with nervousness? [Symptoms not explained by ill health]?
  \item NPIE5 N  5. Does {P} avoid certain places or situations that make him.her more nervous such as riding in the car, meeting with friends, or being in crowds?
  \item NPIE6 N  6. Does {P} become nervous or upset when separated from you [or his/her caregiver]? [Does he/she cling to you to keep from being separated?]
  \item NPIE7 N  7. Does {P} show any other signs of anxiety?
  \item NPIE8A N  8a. Frequency Ratings
  \item NPIE8B N  8b. Severity Ratings
  \item NPIE8C N  8c. Caregiver Distress
  \item NPIETOT N  E. Anxiety: Item score
  \item NPIF N  F. Elation/Euphoria
  \item NPIF1 N  1. Does {P} appear to feel too good or too happy, different from his/her usual self?
  \item NPIF2 N  2. Does {P} find humor and laugh at things that others do not find funny?
  \item NPIF3 N  3. Does {P} seem to have a childish sense of humor with a tendency to giggle or laugh inappropriately (such as when something unfortunate happens to others)?
  \item NPIF4 N  4. Does {P} tell jokes or make remarks that have little humor for others but seem funny to him/her?
  \item NPIF5 N  5. Does he/she play childish pranks such as pinching or playing "keep away" for the fun of it?
  \item NPIF6 N  6. Does {P} "talk big" or claim to have more abilities or wealth than is true?
  \item NPIF7 N  7. Does {P} show any other signs of feeling too good or being too happy?
  \item NPIF8A N  8a. Frequency Ratings
  \item NPIF8B N  8b. Severity Ratings
  \item NPIF8C N  8c. Caregiver Distress
  \item NPIFTOT N  F. Elation/Euphoria: Item score
  \item NPIG N  G. Apathy/Indifference
  \item NPIG1 N  1. Does {P} seem less spontaneous and less active than usual?
  \item NPIG2 N  2. Is {P} less likely to initiate a conversation?
  \item NPIG3 N  3. Is {P} less affectionate or lacking in emotions when compared to his/her usual self?
  \item NPIG4 N  4. Does {P} contribute less to household chores?
  \item NPIG5 N  5. Does {P} seem less interested in the activities and plans of others?
  \item NPIG6 N  6. Has {P} lost interest in friends and family members?
  \item NPIG7 N  7. Is {P} less enthusiastic about his/her usual interests?
  \item NPIG8 N  8. Does {P} show any other signs that he/she doesn't care about doing new things?
  \item NPIG9A N  9a. Frequency Ratings
  \item NPIG9B N  9b. Severity Ratings
  \item NPIG9C N  9c. Caregiver Distress
  \item NPIGTOT N  G. Apathy/Indifference: Item score
  \item NPIH N  H. Disinhibition
  \item NPIH1 N  1. Does {P} act impulsively without appearing to consider the consequences?
  \item NPIH2 N  2. Does {P} talk to total strangers as if he/she knew them?
  \item NPIH3 N  3. Does {P} say things to people that are insensitive or hurt their feelings?
  \item NPIH4 N  4. Does {P} say crude things or make sexual remarks that they would not usually have said?
  \item NPIH5 N  5. Does {P} talk openly about very personal or private matters not usually discussed in public?
  \item NPIH6 N  6. Does {P} take liberties or touch or hug others in a way that is out of character for him/her?
  \item NPIH7 N  7. Does {P} show any other signs of loss of control of his/her impulses?
  \item NPIH8A N  8a. Frequency Ratings
  \item NPIH8B N  8b. Severity Ratings
  \item NPIH8C N  8c. Caregiver Distress
  \item NPIHTOT N  H. Disinhibition: Item score
  \item NPII N  I. Irritability/Lability
  \item NPII1 N  1. Does {P} have a bad temper, flying "off the handle" easily over little things?
  \item NPII2 N  2. Does {P} rapidly change moods from one to another, being fine one minute and angry the next?
  \item NPII3 N  3. Does {P} have sudden flashes of anger?
  \item NPII4 N  4. Is {P} impatient, having trouble coping with delays or waiting for planned activities?
  \item NPII5 N  5. Is {P} cranky and irritable?
  \item NPII6 N  6. Is {P} argumentative and difficult to get along with?
  \item NPII7 N  7. Does {P} show any other signs of irritability?
  \item NPII8A N  8a. Frequency Ratings
  \item NPII8B N  8b. Severity Ratings
  \item NPII8C N  8c. Caregiver Distress
  \item NPIITOT N  I. Irritability/Lability: Item score
  \item NPIJ N  J. Aberrant Motor Behavior
  \item NPIJ1 N  1. Does {P} pace around the house without apparent purpose?
  \item NPIJ2 N  2. Does {P} rummage around opening and unpacking drawers or closets?
  \item NPIJ3 N  3. Does {P} repeatedly put on and take off clothing?
  \item NPIJ4 N  4. Does {P} have repetitive activities or "habits" that he/she performs over and over?
  \item NPIJ5 N  5. Does {P} engage in repetitive activities such as handling buttons, picking, wrapping string, etc?
  \item NPIJ6 N  6. Has {P} fidget excessively, seem unable to sit still, or bounce his/her feet or tap his/her fingers a lot?
  \item NPIJ7 N  7. Does {P} do any other activities over and over?
  \item NPIJ8A N  8a. Frequency Ratings
  \item NPIJ8B N  8b. Severity Ratings
  \item NPIJ8C N  8c. Caregiver Distress
  \item NPIJTOT N  J. Aberrant Motor Behavior: Item score
  \item NPIK N  K. Sleep
  \item NPIK1 N  1. Does {P} have difficulty falling asleep?
  \item NPIK2 N  2. Does {P} get up during the night (do not count if the patient gets up once or twice per night only to go to the bathroom and falls back asleep immediately)?
  \item NPIK3 N  3. Does {P} wander, pace, or get involved in inappropriate activities at night?
  \item NPIK4 N  4. Does {P} awaken you during the night?
  \item NPIK5 N  5. Does {P} awaken at night, dress, and plan to go out thinking that it is morning and time to start the day?
  \item NPIK6 N  6. Does {P} awaken too early in the morning (earlier than was his/her habit)?
  \item NPIK7 N  7. Does {P} sleep excessively during the day?
  \item NPIK8 N  8. Does {P} have any other night-time behaviors that bother you that we haven't talked about?
  \item NPIK9A N  9a. Frequency Ratings
  \item NPIK9B N  9b. Severity Ratings
  \item NPIK9C N  9c. Caregiver Distress
  \item NPIKTOT N  K. Sleep: Item score
  \item NPIL N  L. Appetite and eating disorders:
  \item NPIL1 N  1. Has he/she had a loss of appetite?
  \item NPIL2 N  2. Has he/she had an increase in appetite?
  \item NPIL3 N  3. Has he/she had a loss of weight?
  \item NPIL4 N  4. Has he/she gained weight?
  \item NPIL5 N  5. Has he/she had a change in eating behavior such as putting too much food in his/her mouth at once?
  \item NPIL6 N  6. Has he/she had a change in the kind of food he/she likes such as eating too many sweets or other specific types of food?
  \item NPIL7 N  7. Has he/she developed eating behaviors such as eating exactly the same types of food each day or eating the food in exactly the same order?
  \item NPIL8 N  8. Have there been any other changes in appetite or eating that I haven't asked about?
  \item NPIL9A N  9a. Frequency Ratings
  \item NPIL9B N  9b. Severity Ratings
  \item NPIL9C N  9c. Caregiver Distress
  \item NPILTOT N  L. Appetite and eating disorders: Item score
  \item NPITOTAL N  NPI Total Score
}
}
\keyword{datasets}
