% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{npiq}
\alias{npiq}
\title{Neuropsychiatric Inventory Q}
\format{
A data frame with 7091 rows and 37 variables
}
\usage{
data(npiq)
}
\description{
Neuropsychiatric Inventory Q
}
\details{
A dataset with variables as follows:

\itemize{
  \item NPDATE D NA Date collected
  \item SPID N NA Study Partner ID:
  \item SOURCE N NA Information Source
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item NPISOURCE N  Information Source
  \item NPIA N  Does {P} believe that others are stealing from him/her, or planning to harm him/her in some way?
  \item NPIASEV N  Severity Ratings <!--A. Delusions-->
  \item NPIB N  Does {P} act as if he/she hears voices? Does he/she talk to people who are not there?
  \item NPIBSEV N  Severity Ratings <!--B. Hallucinations-->
  \item NPIC N  Is {P} stubborn and resistive to help from others?
  \item NPICSEV N  Severity Ratings <!--C. Agitation/Aggression-->
  \item NPID N  Does {P} act as if he/she is sad or in low spirits? Does he/she cry?
  \item NPIDSEV N  Severity Ratings <!--D. Depression/Dysphoria-->
  \item NPIE N  Does {P} become upset when separated from you? Does he/she have any other signs of nervousness, such as shortness of breath, sighing, being unable to relax, or feeling excessively tense?
  \item NPIESEV N  Severity Ratings <!--E. Anxiety-->
  \item NPIF N  Does {P} appear to feel too good or act excessively happy?
  \item NPIFSEV N  Severity Ratings <!--F. Elation/Euphoria-->
  \item NPIG N  Does {P} seem less interested in his/her usual activities and in the activities and plans of others?
  \item NPIGSEV N  Severity Ratings <!--G. Apathy/Indifference-->
  \item NPIH N  Does {P} seem to act impulsively? For example, does {P} talk to strangers as if he/she knows them, or does {P} say things that may hurt people's feelings?
  \item NPIHSEV N  Severity Ratings <!--H. Disinhibition-->
  \item NPII N  Is {P} impatient or cranky? Does he/she have difficulty coping with delays or waiting for planned activities?
  \item NPIISEV N  Severity Ratings <!--I. Irritability/Lability-->
  \item NPIJ N  Does {P} engage in repetitive activities, such as pacing around the house, handling buttons, wrapping strings, or doing other things repeatedly?
  \item NPIJSEV N  Severity Ratings <!--J. Aberrant Motor Behavior-->
  \item NPIK N  Does {P} awaken you during the night, rise too early in the morning, or take excessive naps during the day?
  \item NPIKSEV N  Severity Ratings <!--K. Sleep-->
  \item NPIL N  Has {P} lost or gained weight, or had a change in the food he/she likes?
  \item NPILSEV N  Severity Ratings <!--L. Appetite-->
  \item NPISCORE N  Total Score
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
