% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{npstatus}
\alias{npstatus}
\title{Neuropathology Status}
\format{
A data frame with 9734 rows and 16 variables
}
\usage{
data(npstatus)
}
\description{
Neuropathology Status
}
\details{
A dataset with variables as follows:

\itemize{
  \item NPSDATE D NA Date collected
  \item NPSOURC N NA Information source
  \item NPCONS N NA Was Neuropath consent discussed at this visit?
  \item NPCONSNO N NA If no, why not:
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
  \item NPDECIDE N  Has a definitive decision about brain donation been made?
  \item NPUNDEC N  If no decision has been made, has participant been given information (brochure, discussion with clinician) regarding autopsy/brain donation?
  \item NPDEC N  If a decision has been made, is there agreement to participate in brain donation?
  \item NPDECOTH T  If other, specify:
  \item NPDECCF N  If there is agreement to participate, is the provisional autopsy consent part of a study other than ADNI?
}
}
\keyword{datasets}
