% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{nyufdghip}
\alias{nyufdghip}
\title{NYU FDG-PET Hippocampus (pons normalized)}
\format{
A data frame with 612 rows and 16 variables
}
\usage{
data(nyufdghip)
}
\description{
NYU FDG-PET Hippocampus (pons normalized)
}
\details{
A dataset with variables as follows:

\itemize{
  \item PHASE -4 NA ADNI phase
  \item SITEID -4 NA Site ID
  \item PTID -4 NA Participant ID
  \item RID -4 NA Participant roster ID
  \item EXAMDATE -4 NA Examination Date
  \item VISCODE -4 NA Visit code
  \item LONIUID -4 NA LONI Image ID
  \item STUDYID -4 NA LONI PET study ID
  \item SERIESID -4 NA LONI PET series ID
  \item SCANDATE -4 NA PET scan date
  \item RUNDATE -4 NA Run date of analysis result
  \item STATUS -4 NA Result status
  \item VERSION -4 NA Version of results
  \item HIPRPONS -4 NA Right Hip / Pons normalized
  \item HIPLPONS -4 NA Left Hip / Pons normalized
}
}
\examples{
\dontrun{
describe(nyufdghip)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/NYUFDGHIP/NYUFDGHip.pdf}
}
\keyword{datasets}
