% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{otconsent}
\alias{otconsent}
\title{Addenda Consent Tracking}
\format{
A data frame with 252 rows and 14 variables
}
\usage{
data(otconsent)
}
\description{
Addenda Consent Tracking
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
  \item RECNO N  
  \item ADDENDUM N  Indicate addendum:
  \item INCDATE D MM/DD/YYYY Date consent(s) signed/updated:
  \item INCVERSION N  ADCS template version signed:
  \item OTIRB D MM/DD/YYYY Consent IRB approval date:
  \item OTEXPDT D MM/DD/YYYY Consent IRB expiration date:
  \item OTVRS T  Site version of consent:
}
}
\examples{
\dontrun{
describe(otconsent)
}
}
\keyword{datasets}
