% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{otdisc}
\alias{otdisc}
\title{Other Discontinuations and Withdrawals}
\format{
A data frame with 174 rows and 13 variables
}
\usage{
data(otdisc)
}
\description{
Other Discontinuations and Withdrawals
}
\details{
A dataset with variables as follows:

\itemize{
  \item OTCOMP N NA Study Component
  \item OTSPEC T NA If other, specify
  \item OTDATE D YYYY-MM-DD Date of Discontinuation/Withdrawn Consent
  \item OTREASON N NA Reason(s) for Discontinuation/Withdrawn Consent
  \item OTOTHER T NA If other, specify
  \item OTAE T NA If Adverse Event, provide Adverse Event number(s)
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
}
}
\examples{
\dontrun{
describe(otdisc)
}
}
\keyword{datasets}
