% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{otelgtau}
\alias{otelgtau}
\title{Tau AV-1451 PET Eligibility}
\format{
A data frame with 141 rows and 19 variables
}
\usage{
data(otelgtau)
}
\description{
Tau AV-1451 PET Eligibility
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
  \item RECNO N  
  \item TAUELG N  Please confirm if participant meets eligibility criteria:
  \item EXCLUSION N  If participant should be excluded, select reason:
  \item NVROT N  
  \item NVRSPECIFY T  If Other, specify:
  \item EXC1 N  If Exclusion Criteria, which criteria does the participant meet:
  \item EXC2 N  
  \item EXC3 N  
  \item EXC4 N  
  \item EXC5 N  
  \item EXC6 N  
  \item EXC7 N  
}
}
\examples{
\dontrun{
describe(otelgtau)
}
}
\keyword{datasets}
