% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{petmeta}
\alias{petmeta}
\title{PET Scan Information}
\format{
A data frame with 6034 rows and 119 variables
}
\usage{
data(petmeta)
}
\description{
PET Scan Information
}
\details{
A dataset with variables as follows:

\itemize{
  \item DONE N NA Was the scan conducted?
  \item NDREASON N NA If No, reason not done:
  \item SCANDATE D YYYY-MM-DD Scan Date/Dose Date
  \item QCTIME T HHMM Time of today's Scanner QC
  \item ASSAYTIME T HHMM Time of FDG dose assay
  \item FDGDOS N mCi Injected FDG dose assay
  \item INJTIME T HHMM Time of FDG injection
  \item SCANTIME T HHMM Emission Scan Start Time
  \item VARIAT N NA Any variations from protocol during FDG uptake?
  \item PROTID T NA Predefined Acquisition Protocol Name
  \item MOTION N NA Participant motion problems
  \item OTHERVAR N NA Other protocol variations
  \item RECON N NA Select reconstruction used
  \item SHARP N NA For Phillips scanners:  was "smooth" parameter set to "sharp"?
  \item ARCHIVE N NA Data archived locally?
  \item LONI N NA Data transferred to LONI?
  \item TRANDATE D YYYY-MM-DD Transfer Date
  \item RADTRACER N  If no, was radiotracer administered?
  \item PMSHARP T NA Was "Smooth" parameter set to "Sharp"?
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item RECNO N  
  \item PMCONDCT N  Was the scan conducted?
  \item PMREASON N  Reason why the scan was not conducted:
  \item PMCONSENT N  Did the participant consent to quantitative scan?
  \item PMTYPE N  Type of scan conducted
  \item EXAMDATE D  Examination Date
  \item PMSCANNER T  GE
  \item PMGEMODEL N  If GE, Scanner Model:
  \item PMSIEMENS T  Siemens/CTI
  \item PMSIEMODEL N  If Siemens/CTI, Scanner Model:
  \item PMPHILLIPS T  Phillips
  \item PMPHMODEL N  If Phillips, Scanner Model:
  \item PMQCTIME T  Time of today's Scanner QC
  \item PMBLTIME T  Time of blood glucose measurement
  \item PMBLGLUC N  Blood Glucose (pre-FDG)
  \item PMFDGTIME T  Time of FDG dose assay
  \item PMFDGDOS N  FDG dose assay
  \item PMFDGVOL N  FDG Volume
  \item PMINJTIME T  Time of FDG injection
  \item PMSCTIME T  Time scan started (emission)
  \item PMVARIAT N  Any variations from protocol during FDG uptake?
  \item PMFRAME N  Which framing rate was used?
  \item PMMOTION N  Subject motion problems:
  \item PMMALFUN N  Scanner malfunction
  \item PMOTHER N  Other protocol variations:
  \item PMRECON N  Check which of the following reconstructions was used:
  \item PMSUBSET N  # subsets:
  \item PMSUBSPE N  If Other, specify
  \item PMITERAT N  # iterations:
  \item PMITERSPE N  If Other, specify:
  \item PMRAMLA T  If Ramla, Lambda=0.016?
  \item PMFILTER T  If Back Projection, Ramp filter?
  \item PMMODEON T  If FORE/2D-OSEM, Brain Mode "ON" for PET only scanners or TRIM "ON" for PET/CT scanners?
  \item PMSMOOTH T  No post-process smoothing:
  \item PMDECAY N  Decay Correction
  \item PMSCATTR N  Scatter Correction:
  \item PMATTEN N  Attenuation Correction:
  \item PMWITHDRAW N  Clock for blood sample withdrawal time
  \item PMWITHTIME T  If No, provide the time difference
  \item PMCOUNT N  Clock for blood sample count time
  \item PMCOUNTIME T  If No, provide the time difference
  \item PMBG1TIME T  Sample Count Time (24h) <!--Background #1-->
  \item PMBG1VOL N  Sample Plasma Volume Counted <!--Background #1-->
  \item PMBG1DUR N  Sample Count Duration <!--Background #1-->
  \item PMBG1COUNT N  Sample Count Rate <!--Background #1-->
  \item PMP1DTIME T  Sample Draw Time (24h) <!--Plasma 1-->
  \item PMP1CTIME T  Sample Count Time (24h) <!--Plasma 1-->
  \item PMP1BGL N  Sample BGL <!--Plasma 1-->
  \item PMP1VOL N  Sample Plasma Volume Counted <!--Plasma 1-->
  \item PMP1DUR N  Sample Count Duration <!--Plasma 1-->
  \item PMP1COUNT N  Sample Count Rate <!--Plasma 1-->
  \item PMP2DTIME T  Sample Draw Time (24h) <!--Plasma 2-->
  \item PMP2CTIME T  Sample Count Time (24h) <!--Plasma 2-->
  \item PMP2BGL N  Sample BGL <!--Plasma 2-->
  \item PMP2VOL N  Sample Plasma Volume Counted <!--Plasma 2-->
  \item PMP2DUR N  Sample Count Duration <!--Plasma 2-->
  \item PMP2COUNT N  Sample Count Rate <!--Plasma 2-->
  \item PMP3DTIME T  Sample Draw Time (24h) <!--Plasma 3-->
  \item PMP3CTIME T  Sample Count Time (24h) <!--Plasma 3-->
  \item PMP3BGL N  Sample BGL <!--Plasma 3-->
  \item PMP3VOL N  Sample Plasma Volume Counted <!--Plasma 3-->
  \item PMP3DUR N  Sample Count Duration <!--Plasma 3-->
  \item PMP3COUNT N  Sample Count Rate <!--Plasma 3-->
  \item PMP4DTIME T  Sample Draw Time (24h) <!--Plasma 4-->
  \item PMP4CTIME T  Sample Count Time (24h) <!--Plasma 4-->
  \item PMP4BGL N  Sample BGL <!--Plasma 4-->
  \item PMP4VOL N  Sample Plasma Volume Counted <!--Plasma 4-->
  \item PMP4DUR N  Sample Count Duration <!--Plasma 4-->
  \item PMP4COUNT N  Sample Count Rate <!--Plasma 4-->
  \item PMP5DTIME T  Sample Draw Time (24h) <!--Plasma 5-->
  \item PMP5CTIME T  Sample Count Time (24h) <!--Plasma 5-->
  \item PMP5BGL N  Sample BGL <!--Plasma 5-->
  \item PMP5VOL N  Sample Plasma Volume Counted <!--Plasma 5-->
  \item PMP5DUR N  Sample Count Duration <!--Plasma 5-->
  \item PMP5COUNT N  Sample Count Rate <!--Plasma 5-->
  \item PMBG2CTIME T  Sample Count Time (24h) <!--Background 2-->
  \item PMBG2VOL N  Sample Plasma Volume Counted <!--Background 2-->
  \item PMBG2DUR N  Sample Count Duration <!--Background 2-->
  \item PMBG2COUNT N  Sample Count Rate <!--Background 2-->
  \item PMPIPVOL N  Was the pipetted plasma volume 200 uL?
  \item PMPIPVOLSP T  If No, denote volume used:
  \item PMCTIME N  Was the plasma sample count time 1 minute?
  \item PMCTIMESP T  If No, denote count time used
  \item PMPHACT N  Phantom Activity at Time of Scan
  \item PMPHVOL N  Phantom Volume
  \item PMPHCOUNT N  Average Counts from Phantom Image ROI
  \item PMALIQVOL N  Aliquot Volume
  \item PMALIQACT N  Aliquot Count Rate
  \item PMTRNSFR N  Was data transferred to LONI within 24 hours of scan?
  \item PMTRNDATE D  Transfer Date
  \item PMARCHIVE N  Data Archived Locally
  \item PMLPDONE N  Was a Lumbar Puncture completed prior to the PET scan?
  \item PMLPINTER N  If Yes, What was the interval between LP and PET?
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
