% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{petqc}
\alias{petqc}
\title{PET QC Tracking}
\format{
A data frame with 4385 rows and 46 variables
}
\usage{
data(petqc)
}
\description{
PET QC Tracking
}
\details{
A dataset with variables as follows:

\itemize{
  \item SCANDATE D YYYY-MM-DD Scan Date
  \item LONIUPDT D YYYY-MM-DD Date uploaded to LONI
  \item REVDT D YYYY-MM-DD Date Reviewed
  \item ALLFRAME N NA Were all frames acceptable?
  \item UNUSABL T NA If No, indicate which frames were unacceptable
  \item UNRSN N NA If No, indicate why frames were unacceptable
  \item UNRSNSPEC T NA If Other, Specify
  \item SCANQLTY N NA Scan Pass QC?
  \item REPROCREQ N NA Reprocessing requested?
  \item PROCERR T NA If Yes, select processing error(s)
  \item PROCERRSPEC T NA If Other, Specify
  \item RESCANREQ N NA Rescan requested?
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item RECNO N  
  \item SCAN N  Scan
  \item EXAMDATE D  Examination Date
  \item PQFILENO N  Upload Number
  \item PQDATTRA D  Date images uploaded to LONI
  \item LONIUID T  LONI Unique Series ID
  \item PQDATE D  Date reviewed
  \item PQACQFR N  Acquisition
  \item PQCORALG N  Correct Reconstruction Algorithm?
  \item PQRECON N  Check which of the following reconstruction algorithms was used
  \item PQCORNUM N  If OSEM or Ramla, correct # iterations and subsets?
  \item PQCORFIL N  Correct Filters/TRIM?
  \item PQCORTHI N  Correct Slice Thickness (PET/CT systems)?
  \item PQVOXEL N  Acceptable Voxel Size?
  \item PQALLFRAME N  Were all frames acceptable?
  \item PQUNUSABLE T  Indicate which frames were unacceptable:
  \item PQREASON T  Indicate why frames were unacceptable:
  \item PQREASSP T  If Other, Specify:
  \item PASS N  Pass QC?
  \item PQPROERR T  A. Processing Error(s)
  \item PQERRSP T  Specify
  \item PQACTION N  B. QC outcome
  \item PQNONESP T  Reason for not rescanning
  \item PQISSUES T  Imaging issues
  \item PQISSOTH T  Other, specify
  \item QUARANTINE N  Release from quarantine?
  \item PAYSITE N  Pay Site?
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
