% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{physical}
\alias{physical}
\title{Physical Exam}
\format{
A data frame with 3673 rows and 22 variables
}
\usage{
data(physical)
}
\description{
Physical Exam
}
\details{
A dataset with variables as follows:

\itemize{
  \item PXBACK N NA 12. Back
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item PXGENAPP N  1. General Appearance
  \item PXHEADEY N  2. Head, Eyes, Ears, Nose and Throat
  \item PXNECK N  3. Neck
  \item PXCHEST N  4. Chest
  \item PXHEART N  5. Heart
  \item PXABDOM N  6. Abdomen
  \item PXEXTREM N  7. Extremities
  \item PXEDEMA N  8. Edema
  \item PXPERIPH N  9. Peripheral Vascular
  \item PXSKIN N  10. Skin and Appendages
  \item PXMUSCUL N  11. Musculoskeletal
  \item PXOTHER N  12. Other
  \item PXABNORM N  14. Based on the Physical Examination, clinician must check appropriate box below:
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
