% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{pibapprove}
\alias{pibapprove}
\title{PIB Approval}
\format{
A data frame with 110 rows and 7 variables
}
\usage{
data(pibapprove)
}
\description{
PIB Approval
}
\details{
A dataset with variables as follows:

\itemize{
  \item USERDATE S  Date record created
  \item RID N  Participant roster ID
  \item PIBAPPROVE N  Is the participant approved to participate in PIB sub-study?
  \item PIBAPDT D  If Yes, date approved
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
