% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{pibconsent}
\alias{pibconsent}
\title{PIB Consent}
\format{
A data frame with 126 rows and 10 variables
}
\usage{
data(pibconsent)
}
\description{
PIB Consent
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item PIBSIGN N  Have the participant and study partner signed the PIB Informed Consent form?
  \item PIBSGDT D  If Yes, date signed
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
