% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{pibmeta}
\alias{pibmeta}
\title{PIB Scan Information}
\format{
A data frame with 257 rows and 75 variables
}
\usage{
data(pibmeta)
}
\description{
PIB Scan Information
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item PBCONDCT N  Was the scan conducted?
  \item PBREASON N  Reason why the scan was not conducted:
  \item PBOTHSPE T  If Other, specify:
  \item EXAMDATE D  Examination Date
  \item PBGE T  GE
  \item PBGEMODEL N  If GE, Scanner Model:
  \item PBSIEMENS T  Siemens/CTI
  \item PBSIEMODEL N  If Siemens/CTI, Scanner Model:
  \item PBPHILLIPS T  Phillips
  \item PBPHMODEL N  If Phillips, Scanner Model:
  \item PBQCTIME T  Time of today's Scanner QC
  \item PBPIBTIME T  Time of PIB dose assay
  \item PBPIBDOS N  PIB dose assay
  \item PBRESTIME T  Time of residual PIB assay
  \item PBRESID T  Residual left in syringe
  \item PBINJDOS N  Net injected dose of PIB
  \item PBPIBVOL N  PIB volume
  \item PBINJTIME T  Time of PIB injection
  \item PBSCTIME T  Time scan started (emission)
  \item PBSCANCOM T  Provide an explanation if start time is not within the allowable window
  \item PBVARIAT N  Any variations from protocol during PIB uptake?
  \item PBVARSP T  If Yes, describe:
  \item PBPROTID T  Predefined acquisition protocol ID
  \item PBSTATDYN N  Indicate whether scan was static or dynamic:
  \item PBNMFRAM1 N  1. No. of Frames
  \item PBDURATN1 N  Duration:
  \item PBNMFRAM2 N  2. No. of Frames:
  \item PBDURATN2 N  Duration:
  \item PBNMFRAM3 N  3. No. of Frames:
  \item PBDURATN3 N  Duration:
  \item PBNMFRAM4 N  4. No. of Frames:
  \item PBDURATN4 N  Duration:
  \item PBNMFRAM5 N  5. No. of Frames:
  \item PBDURATN5 N  Duration:
  \item PBNMFRAM6 N  6. No. of Frames:
  \item PBDURATN6 N  Duration:
  \item PBNMFRAM7 N  7. No. of Frames:
  \item PBDURATN7 N  Duration:
  \item PBNMFRAM8 N  8. No. of Frames:
  \item PBDURATN8 N  Duration:
  \item PBNMFRAM9 N  9. No. of Frames:
  \item PBDURATN9 N  Duration:
  \item PBNMFRAM10 N  10. No. of Frames:
  \item PBDURATN10 N  Duration:
  \item PBMOTION N  Subject motion problems:
  \item PBMOTSP T  If yes, describe:
  \item PBMALFUN N  Scanner malfunction
  \item PBMALSP T  If yes, describe:
  \item PBOTHER N  Other protocol variations:
  \item PBOTHSP T  If yes, describe:
  \item PBRECON N  Check which of the following reconstructions was used:
  \item PBSUBSET N  # subsets:
  \item PBSUBSPE N  If Other, specify
  \item PBITERAT N  # iterations:
  \item PBITERSPE N  If Other, specify:
  \item PBRAMLA N  If Ramla, Lambda=0.016?
  \item PBFILTER N  If Back Projection, Ramp filter?
  \item PBMODEON N  If FORE/2D-OSEM, Brain Mode "ON" for PET only scanners or TRIM "ON" for PET/CT scanners?
  \item PBSMOOTH N  No post-process smoothing:
  \item PBDECAY N  Decay Correction
  \item PBSCATTR N  Scatter Correction:
  \item PBMATTEN N  Attenuation Correction:
  \item PBTRSNFR N  Was data transferred to LONI within 24 hours of scan?
  \item PBTRNDATE D  Transfer Date
  \item PBTRNCOM T  Comments
  \item PBARCHIVE N  Data Archived Locally
  \item PBARCMED T  Archive Medium
  \item PBARCCOM T  Comments
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
