% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{pibpetsuvr}
\alias{pibpetsuvr}
\title{U Pitt PIB PET Analysis}
\format{
A data frame with 224 rows and 21 variables
}
\usage{
data(pibpetsuvr)
}
\description{
U Pitt PIB PET Analysis
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N -4 Participant roster ID
  \item VISCODE T -4 Visit code
  \item EXAMDATE D -4 Examination Date
  \item IMAGEUID T -4 -4
  \item LONISID T -4 -4
  \item LONIUID T -4 -4
  \item ACG N -4 Anterior cingulate 
  \item FRC N -4 Frontal cortex 
  \item LTC N -4 Lateral temporal cortex 
  \item PAR N -4 Parietal cortex 
  \item PRC N -4 Precuneus cortex 
  \item MTC N -4 Mesial temporal cortex 
  \item OCC N -4 Occipital cortex 
  \item OCP N -4 Occipital pole 
  \item PON N -4 Pons
  \item AVS N -4 Anterior ventral striatum
  \item CER N -4 Cerebellum 
  \item SMC N -4 Sensory motor cortex
  \item SWM N -4 Sub-cortical white matter 
  \item THL N -4 Thalamus
}
}
\examples{
pibpetsuvr$PIBp = apply(pibpetsuvr[, c('FRC', 'ACG', 'PRC', 'PAR')], 1, mean) > 1.5
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/PIBPETSUVR/UPitt_PIBPET_AD_ROI.pdf}, \url{https://adni.bitbucket.io/reference/docs/PIBPETSUVR/UPitt_PIBPET_MCI_ROI.pdf}, \url{https://adni.bitbucket.io/reference/docs/PIBPETSUVR/UPitt_PIBPET_NL_ROI.pdf}
}
\keyword{datasets}
