% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{pibqc}
\alias{pibqc}
\title{PIB QC Tracking}
\format{
A data frame with 294 rows and 32 variables
}
\usage{
data(pibqc)
}
\description{
PIB QC Tracking
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item RECNO N  
  \item EXAMDATE D  Examination Date
  \item PQFILENO N  Upload Number
  \item PQDATTRA D  Date images uploaded to LONI
  \item LONIUID T  LONI Unique Series ID
  \item PQDATE D  Date reviewed
  \item PQACQFR N  Acquisition
  \item PQCORALG N  Correct Reconstruction Algorithm?
  \item PQRECON N  Check which of the following reconstruction algorithms was used
  \item PQCORNUM N  If OSEM or Ramla, correct # iterations and subsets?
  \item PQCORFIL N  Correct Filters/TRIM?
  \item PQCORTHI N  Correct Slice Thickness (PET/CT systems)?
  \item PQVOXEL N  Acceptable Voxel Size?
  \item PQALLFRAME N  Were all frames acceptable?
  \item PQUNUSABLE T  Indicate which frames were unacceptable:
  \item PQREASON T  Indicate why frames were unacceptable:
  \item PQREASSP T  If Other, Specify:
  \item PASS N  Pass QC?
  \item PQPROERR T  A. Processing Error(s)
  \item PQERRSP T  Specify
  \item PQACTION N  B. QC outcome
  \item PQISSUES T  Imaging issues
  \item PQISSOTH T  Other, specify
  \item QUARANTINE N  Release from quarantine?
  \item PAYSITE N  Pay Site?
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
