% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{picslashs}
\alias{picslashs}
\title{ASHS volume data}
\format{
A data frame with 1978 rows and 57 variables
}
\usage{
data(picslashs)
}
\description{
ASHS volume data
}
\details{
A dataset with variables as follows:

\itemize{
  \item COLPROT -4 -4 Study protocol of data collection
  \item RID -4  Participant roster ID
  \item VISCODE -4  Visit code
  \item VISCODE2 -4  Translated visit code
  \item EXAMDATE -4  Examination Date
  \item ICV -4  Intracranial volume computed by ASHS
  \item SLICE_THICKNESS -4  Coronal slice thickness
  \item LEFT_CA1_VOL -4 mm^3 Volume of left CA1 ROI
  \item LEFT_CA1_NS -4  Number of coroinal slices in which left CA1 appears
  \item LEFT_CA2_VOL -4 mm^3 Volume of left CA2 ROI
  \item LEFT_CA2_NS -4  Number of coroinal slices in which left CA2 appears
  \item LEFT_CA3_VOL -4 mm^3 Volume of left CA3 ROI
  \item LEFT_CA3_NS -4  Number of coroinal slices in which left CA3 appears
  \item LEFT_DG_VOL -4 mm^3 Volume of left DG ROI
  \item LEFT_DG_NS -4  Number of coroinal slices in which left DG appears
  \item LEFT_MISC_VOL -4 mm^3 Volume of left MISC ROI
  \item LEFT_MISC_NS -4  Number of coroinal slices in which left MISC appears
  \item LEFT_SUB_VOL -4 mm^3 Volume of left subiculum ROI
  \item LEFT_SUB_NS -4  Number of coroinal slices in which left subiculum appears
  \item LEFT_ERC_VOL -4 mm^3 Volume of left entorhinal cortex ROI
  \item LEFT_ERC_NS -4  Number of coroinal slices in which left entorhinal cortex  appears
  \item LEFT_BA35_VOL -4 mm^3 Volume of left BA35 ROI
  \item LEFT_BA35_NS -4  Number of coroinal slices in which left BA35 appears
  \item LEFT_BA36_VOL -4 mm^3 Volume of left BA36 ROI
  \item LEFT_BA36_NS -4  Number of coroinal slices in which left BA36 appears
  \item LEFT_PHC_VOL -4 mm^3 Volume of left parahippocampal cortex ROI
  \item LEFT_PHC_NS -4  Number of coroinal slices in which left parahippocampal appears
  \item LEFT_SULCUS_VOL -4 mm^3 Volume of left sulcus ROI
  \item LEFT_SULCUS_NS -4  Number of coroinal slices in which left sulcus appears
  \item LEFT_CA_VOL -4 mm^3 Volume of left CA (CA1+CA2+CA3) ROI
  \item LEFT_CA_NS -4  Number of coroinal slices in which left CA (CA1+CA2+CA3) appears
  \item RIGHT_CA1_VOL -4 mm^3 Volume of right CA1 ROI
  \item RIGHT_CA1_NS -4  Number of coroinal slices in which right CA1 appears
  \item RIGHT_CA2_VOL -4 mm^3 Volume of right CA2 ROI
  \item RIGHT_CA2_NS -4  Number of coroinal slices in which right CA2 appears
  \item RIGHT_CA3_VOL -4 mm^3 Volume of right CA3 ROI
  \item RIGHT_CA3_NS -4  Number of coroinal slices in which right CA3 appears
  \item RIGHT_DG_VOL -4 mm^3 Volume of right DG ROI
  \item RIGHT_DG_NS -4  Number of coroinal slices in which right DG appears
  \item RIGHT_MISC_VOL -4 mm^3 Volume of right MISC ROI
  \item RIGHT_MISC_NS -4  Number of coroinal slices in which right MISC appears
  \item RIGHT_SUB_VOL -4 mm^3 Volume of right subiculum ROI
  \item RIGHT_SUB_NS -4  Number of coroinal slices in which right subiculum appears
  \item RIGHT_ERC_VOL -4 mm^3 Volume of right entorhinal cortex ROI
  \item RIGHT_ERC_NS -4  Number of coroinal slices in which right entorhinal cortex  appears
  \item RIGHT_BA35_VOL -4 mm^3 Volume of right BA35 ROI
  \item RIGHT_BA35_NS -4  Number of coroinal slices in which right BA35 appears
  \item RIGHT_BA36_VOL -4 mm^3 Volume of right BA36 ROI
  \item RIGHT_BA36_NS -4  Number of coroinal slices in which right BA36 appears
  \item RIGHT_PHC_VOL -4 mm^3 Volume of right parahippocampal cortex ROI
  \item RIGHT_PHC_NS -4  Number of coroinal slices in which right parahippocampal appears
  \item RIGHT_SULCUS_VOL -4 mm^3 Volume of right sulcus ROI
  \item RIGHT_SULCUS_NS -4  Number of coroinal slices in which right sulcus appears
  \item RIGHT_CA_VOL -4 mm^3 Volume of right CA (CA1+CA2+CA3) ROI
  \item RIGHT_CA_NS -4  Number of coroinal slices in which right CA (CA1+CA2+CA3) appears
  \item T2_QC_R -4  Image quality right hemisphere
  \item T2_QC_L -4  Image quality left hemisphere
}
}
\examples{
\dontrun{
describe(picslashs)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/PICSLASHS/PICSL_ASHS_methods_10-15-2018.pdf}
}
\keyword{datasets}
