% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{pifinal}
\alias{pifinal}
\title{Site PI Final Confirmation}
\format{
A data frame with 1583 rows and 9 variables
}
\usage{
data(pifinal)
}
\description{
Site PI Final Confirmation
}
\details{
A dataset with variables as follows:

\itemize{
  \item PIFINAL T NA I certify that I am using my own personal username and pin and understand that by using my username and pin, I am authorizing this electronic signature to be the legally binding equivalent of my handwritten signature executed on paper. - Enter Your PIN -
  \item FINALCOMM T NA Comments
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
}
}
\examples{
\dontrun{
describe(pifinal)
}
}
\keyword{datasets}
