% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{plasma_abeta_project_adx_vumc}
\alias{plasma_abeta_project_adx_vumc}
\title{FNIH Biomarkers Consortium Plasma Abeta Project: ADx VUmc}
\format{
A data frame with 130 rows and 14 variables
}
\usage{
data(plasma_abeta_project_adx_vumc)
}
\description{
FNIH Biomarkers Consortium Plasma Abeta Project: ADx VUmc
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item GUSPECID -4 -4 -4
  \item VID -4 -4 -4
  \item DER -4 -4 -4
  \item DRAWTIME -4 -4 -4
  \item RUN -4 -4 Indicates the run in which the sample was analyzed
  \item SEQ -4 -4 Order or sequence of sample measurement
  \item NF.LIGHT -4 -4 Calibrated plasma Neurofilament light (NF-light) concentration in pg/mL
  \item ABETA42 -4 -4 Calibrated plasma Abeta 42 concentration in pg/mL
  \item ABETA40 -4 -4 Calibrated plasma Abeta 40 concentration in pg/mL. Measurements > 146.4 pg/mL were extrapolated above the standard curve because there was a Simoa-error on the highest calibrator point.
  \item GFAP -4 -4 Calibrated plasma glial fibrillary acidic protein (GFAP) concentration in pg/mL
}
}
\examples{
\dontrun{
describe(plasma_abeta_project_adx_vumc)
}
}
\keyword{datasets}
