% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{plasma_abeta_project_quanterix}
\alias{plasma_abeta_project_quanterix}
\title{FNIH Biomarkers Consortium Plasma Abeta Project: Quanterix}
\format{
A data frame with 131 rows and 29 variables
}
\usage{
data(plasma_abeta_project_quanterix)
}
\description{
FNIH Biomarkers Consortium Plasma Abeta Project: Quanterix
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item GUSPECID -4 -4 -4
  \item VID -4 -4 -4
  \item DER -4 -4 -4
  \item DRAWTIME -4 -4 -4
  \item SAMPLE.ID -4 -4 -4
  \item AB1.40.RUN -4 -4 Indicates the analytical run in which the sample was analyzed for AB1-40
  \item AB1.40.AEB.REP1 -4 -4 AB1-40 average enzymes per bead, replicate 1 value
  \item AB1.40.AEB.REP2 -4 -4 AB1-40 average enzymes per bead, replicate 2 value
  \item AB1.40.AVE..AEB -4 -4 AB1-40 average of the AEB replicate values
  \item AB1.40.CONC..REP1 -4 -4 AB1-40  calibrated concentration in pg/mL for replicate 1
  \item AB1.40.CONC..REP2 -4 -4 AB1-40 calibrated concentration in pg/mL for replicate 2
  \item AB1.40.AVE..CONC. -4 -4 AB1-40 average of the calibrated concentrations
  \item AB1.40.DILUTION.FACTOR -4 -4 AB1-40 dilution factor. Plasma samples were diluted 4-fold.
  \item AB1.40.DILUTION.CORRECTED.CONC. -4 -4 Dilution corrected AB1-40 concentration of analyte in pg/mL
  \item AB1.40.RUN.NOTES -4 -4 Notes on acceptance criteria or retesting for AB1-40 assay
  \item AB1.42.RUN -4 -4 Indicates the analytical run in which the sample was analyzed for AB1-42
  \item AB1.42.AEB.REP1 -4 -4 AB1-42 average enzymes per bead, replicate 1 value
  \item AB1.42.AEB.REP2 -4 -4 AB1-42  average enzymes per bead, replicate 2 value
  \item AB1.42.AVE..AEB -4 -4 AB1-42  average of the AEB replicate values
  \item AB1.42.CONC..REP1 -4 -4 AB1-42  calibrated concentration in pg/mL for replicate 1
  \item AB1.42.CONC..REP2 -4 -4 AB1-42 calibrated concentration in pg/mL for replicate 2
  \item AB1.42.AVE..CONC. -4 -4 AB1-42 average of the calibrated concentrations
  \item AB1.42.DILUTION.FACTOR -4 -4 AB1-42  dilution factor. Plasma samples were diluted 4-fold.
  \item AB1.42.DILUTION.CORRECTED.CONC. -4 -4 Dilution corrected AB1-42 concentration in pg/mL
  \item AB1.42.RUN.NOTES -4 -4 Notes on acceptance criteria or retesting for AB1-40 assay
}
}
\examples{
\dontrun{
describe(plasma_abeta_project_quanterix)
}
}
\keyword{datasets}
