% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{plasma_abeta_project_roche}
\alias{plasma_abeta_project_roche}
\title{FNIH Biomarkers Consortium Plasma Abeta Project: Roche}
\format{
A data frame with 260 rows and 12 variables
}
\usage{
data(plasma_abeta_project_roche)
}
\description{
FNIH Biomarkers Consortium Plasma Abeta Project: Roche
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item GUSPECID -4 -4 -4
  \item VID -4 -4 -4
  \item DER -4 -4 -4
  \item DRAWTIME -4 -4 -4
  \item STUDYID -4 -4 -4
  \item RBARCODE -4 -4 Identifer for run sequence
  \item ASSAY -4 -4 Plasma Abeta1-40 or Plasma Abeta1-42 immunoassay
  \item RESULT -4 -4 Calibrated concentration of analyte in pg/mL. Samples were run in singlicate and were not diluted.
}
}
\examples{
\dontrun{
describe(plasma_abeta_project_roche)
}
}
\keyword{datasets}
