% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{plasma_abeta_project_shimadzu}
\alias{plasma_abeta_project_shimadzu}
\title{FNIH Biomarkers Consortium Plasma Abeta Project: Shimadzu}
\format{
A data frame with 130 rows and 13 variables
}
\usage{
data(plasma_abeta_project_shimadzu)
}
\description{
FNIH Biomarkers Consortium Plasma Abeta Project: Shimadzu
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item GUSPECID -4 -4 -4
  \item VID -4 -4 -4
  \item DER -4 -4 -4
  \item DRAWTIME -4 -4 -4
  \item DATE -4 -4 Date of sample analysis in month/date/year format
  \item AB1.42 -4 -4 Normalized intensity of Abeta 1-42 in arbitrary units
  \item AB1.40 -4 -4 Normalized intensity of Abeta 1-40 in arbitrary units
  \item APP.669.711 -4 -4 Normalized intensity APP 669-711 in arbitrary units
  \item COMPOSITE.BIOMARKER -4 -4 Composite biomarker value in arbitrary units generated by combining normalized scores of APP669711/AB 142 and AB 140/AB 142 with a pre-determined weight of 1:1. Each biomarker contributed equally to the composite biomarker score.
}
}
\examples{
\dontrun{
describe(plasma_abeta_project_shimadzu)
}
}
\keyword{datasets}
