% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{plasma_abeta_project_u_of_gothenburg}
\alias{plasma_abeta_project_u_of_gothenburg}
\title{FNIH Biomarkers Consortium Plasma Abeta Project: U of Gothenburg}
\format{
A data frame with 130 rows and 14 variables
}
\usage{
data(plasma_abeta_project_u_of_gothenburg)
}
\description{
FNIH Biomarkers Consortium Plasma Abeta Project: U of Gothenburg
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item GUSPECID -4 -4 -4
  \item VID -4 -4 -4
  \item DER -4 -4 -4
  \item DRAWTIME -4 -4 -4
  \item NUMBER -4 -4 -4
  \item PLATE -4 -4 Indicates the grouping of sample analysis on an assay plate
  \item AB.1.37 -4 -4 Calibrated concentrations of Abeta 1-37 determined using standard curve/calibrators and reported in pg/mL
  \item AB.1.38 -4 -4 Calibrated concentrations of Abeta 1-38 determined using standard curve/calibrators and reported in pg/mL
  \item AB.1.40 -4 -4 Calibrated concentrations of Abeta 1-40 determined using standard curve/calibrators and reported in pg/mL
  \item AB.1.42 -4 -4 Calibrated concentrations of Abeta 1-42 determined using standard curve/calibrators and reported in pg/mL
}
}
\examples{
\dontrun{
describe(plasma_abeta_project_u_of_gothenburg)
}
}
\keyword{datasets}
