% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{plasma_abeta_project_wash_u}
\alias{plasma_abeta_project_wash_u}
\title{FNIH Biomarkers Consortium Plasma Abeta Project: Wash U}
\format{
A data frame with 130 rows and 27 variables
}
\usage{
data(plasma_abeta_project_wash_u)
}
\description{
FNIH Biomarkers Consortium Plasma Abeta Project: Wash U
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item GUSPECID -4 -4 -4
  \item VID -4 -4 -4
  \item DER -4 -4 -4
  \item DRAWTIME -4 -4 -4
  \item FILE_NAME_NE -4 -4 -4
  \item LC_MS_SAMPLE_ID -4 -4 Internal ID of LS_MS sample to control for technical replicates. It serves as internal blinding ID and for QC purpocess
  \item SAMPLE_TYPE -4 -4 Type of LC_MS sample (FNIH Clinical Plasma; FNIH QC; FNIH CSF; QC Calibrator)
  \item INJECTION -4 -4 injection (technical replicate)  1-st denoted as 'a'; second as 'b' ei.
  \item INSTRUMENT -4 -4 Masspectrometer (Lumos; Altis)
  \item BATCH -4 -4 Indicates the batch in which the sample was processed (prepped) for LC-MSMS
  \item ASSAY -4 -4 Experiment file name
  \item ASSAY_IP_PROTOCOL -4 -4 Sample processing protocol
  \item QC_STATUS -4 -4 Status of samples assign by quality control comitte (Passed; Failed). If failed, samples should not be used for further analysis
  \item QC_FLAGS -4 -4 None or one or many QC flags assigned by QC comitte when sample does not meet respective quality criterias
  \item QC_COMMENT -4 -4 Detailed description of QC decision if applicable
  \item MS_RUN_DATE -4 -4 date when sequense get started to run on LC-MSMS
  \item AB40 -4 -4 raw plasma A40 concentration, ng/ml (those are not standardized and should not be used for analysis)
  \item AB42 -4 -4 raw plasma A42 concentration, ng/ml (those are not standardized and should not be used for analysis)
  \item PLASMAAB4240 -4 -4 Concentration ratio of 42/40 as measured by Di-oxidized C-terminal peptides, based on its relation to respective known concentrations of N15 internal standard
  \item STANDARDIZED_PLASMAAB42 -4 -4 Standardized plasma A42 concentration, ng/ml
  \item STANDARDIZED_PLASMAAB40 -4 -4 Standardized plasma A40 concentration, ng/ml
  \item STANDARDIZEAB42_STANDARDIZEDAB40 -4 -4 -4
  \item STANDARDIZED_PLASMAAB4240 -4 -4 The standardized_plasmaAB4240 variable should be used for analyses.  This value is the measured abeta 42/40 by our lab.
}
}
\examples{
\dontrun{
describe(plasma_abeta_project_wash_u)
}
}
\keyword{datasets}
