% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{plasmaabeta}
\alias{plasmaabeta}
\title{ABeta40 and ABeta42 quantification in plasma using ABtest40 and ABtest42}
\format{
A data frame with 784 rows and 7 variables
}
\usage{
data(plasmaabeta)
}
\description{
ABeta40 and ABeta42 quantification in plasma using ABtest40 and ABtest42
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item FP40 -4 (pg/ml) Free ABeta40 in plasma
  \item TP40 -4 (pg/ml) Total ABeta40 in plasma
  \item FP42 -4 (pg/ml) Free ABeta42 in plasma
  \item TP42 -4 (pg/ml) Total ABeta42 in plasma
}
}
\examples{
\dontrun{
describe(plasmaabeta)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/PLASMAABETA/ADNI_Araclon\%20Methods.pdf}
}
\keyword{datasets}
