% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{ptdemog}
\alias{ptdemog}
\title{Participant Demographic Information}
\format{
A data frame with 4857 rows and 29 variables
}
\usage{
data(ptdemog)
}
\description{
Participant Demographic Information
}
\details{
A dataset with variables as follows:

\itemize{
  \item PTCOGBEG N  11a. Year of onset of cognitive symptoms (best estimate)
  \item PTADDX N  11b. Year of Alzheimer's Disease diagnosis
  \item PTMCIBEG N NA 11a. Year of onset of Mild Cognitive Impairment symptoms (best estimate)
  \item PTDOBMM N NA 2a. Participant Month of Birth
  \item PTDOBYY N NA 2b. Participant Year of Birth
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item PTSOURCE N  Information Source
  \item PTGENDER N  1. Participant Gender
  \item PTDOB D  2. Participant Date of Birth
  \item PTHAND N  3. Participant Handedness
  \item PTMARRY N  4. Participant Marital Status
  \item PTEDUCAT N  5. Participant Education
  \item PTWORKHS N  5a. Does the participant have a work history sufficient to exclude mental retardation? <!--Participant Education-->
  \item PTNOTRT N  7. Participant Retired?
  \item PTRTYR D  Retirement Date
  \item PTHOME N  8. Type of Participant residence
  \item PTOTHOME T  If Other, specify:
  \item PTTLANG N  9. Language to be used for testing the Participant
  \item PTPLANG N  10. Participant's Primary Language
  \item PTPSPEC T  If Other, specify:
  \item PTADBEG N  11. Year of onset of Alzheimer's disease symptoms (best estimate)
  \item PTETHCAT N  12. Ethnic Category
  \item PTRACCAT N  13. Racial Categories
  \item USERDATE2 S  Date record last updated
  \item AGE   
}
}
\keyword{datasets}
