% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{publicity}
\alias{publicity}
\title{Publicity Tracking}
\format{
A data frame with 2619 rows and 39 variables
}
\usage{
data(publicity)
}
\description{
Publicity Tracking
}
\details{
A dataset with variables as follows:

\itemize{
  \item REFERRAL N NA Do you have a referral code for this participant?
  \item REFERCODE T NA If Yes, enter referral code:
  \item HEARD N NA 1.  How did you hear about the study?  
  \item HRDSPCFY T NA Please provide specifics here for how you heard about  the study.
  \item WHOMHRD N NA 2.  From whom did you hear about the study?  
  \item WHOMSPCFY T NA If Other, specify:
  \item RID N NA Participant roster ID
  \item SITEID N NA Site ID
  \item VISCODE T NA Visit code
  \item USERDATE S NA Date record created
  \item USERDATE2 S NA Date record last updated
  \item INTERNAL T NA Internal
  \item DOCTOR T NA Doctor's Office
  \item MAJORNEWS T NA Major Newspaper
  \item MAGAZINE T NA Magazine
  \item OTHERPUB T NA Other Publication
  \item OTRPUBSO T NA Specify Publication
  \item OTHPUBTYP T NA Indicate:
  \item MOUTH T NA Word of Mouth
  \item MOUTHSO T NA Specify Other:
  \item NEWSLETTER T NA Organization Newsletter
  \item NESWSLORG T NA Specify Organization:
  \item NEWSLSO T NA Specify Other:
  \item TELEVISION T NA Television
  \item HEALTHFAIR T NA Health Fair
  \item ALZASSOC T NA Alzheimer's Association
  \item WEBSITE T NA Website
  \item WEBSITESO T NA Specify Other:
  \item SNAILMAIL T NA Direct Mail
  \item SNMAILORG T NA Specify Organization:
  \item EMAIL T NA E-Mail
  \item EMAILORG T NA Specify Organization:
  \item PRESENT T NA Presentation
  \item PRESENTORG T NA Specify Organization:
  \item OTHERORG T NA Other Organization<p>Specify:
  \item RADIO T NA Radio<p>Specify Station:
  \item PRIORTRIAL T NA Involved in Prior Clinical Trials<p>Specify Trials:
}
}
\keyword{datasets}
