% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{rbmqc}
\alias{rbmqc}
\title{Rules Based Medicine Plasma Multiplex QC Data}
\format{
A data frame with 1063 rows and 195 variables
}
\usage{
data(rbmqc)
}
\description{
Rules Based Medicine Plasma Multiplex QC Data
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4  Participant roster ID
  \item VISCODE -4  Visit code
  \item RBMID -4  RBM Sample ID
  \item RECDATE D  Sample_Received_Date
  \item A1MICRO -4 ug/ml Alpha-1-Microglobulin 
  \item A2MACRO -4 mg/mL Alpha-2-Macroglobulin 
  \item AACT -4 ug/ml Alpha-1-Antichymotrypsin 
  \item AAT -4 mg/mL Alpha-1-Antitrypsin 
  \item ACE -4 ng/ml Angiotensin-Converting Enzyme 
  \item ACTH -4 ng/mL Adrenocorticotropic Hormone 
  \item ADIPNCTN -4 ug/mL Adiponectin 
  \item AFP -4 ng/mL Alpha-Fetoprotein 
  \item AGRP -4 pg/mL Agouti-Related Protein 
  \item ANG2 -4 ng/mL Angiopoietin-2 
  \item ANGTNSNG -4 ng/mL Angiotensinogen 
  \item APOAI -4 mg/mL Apolipoprotein A-I 
  \item APOAII -4 ng/ml Apolipoprotein A-II 
  \item APOAIV -4 ug/ml Apolipoprotein A-IV 
  \item APOB -4 ug/ml Apolipoprotein B 
  \item APOCI -4 ng/ml Apolipoprotein C-I 
  \item APOCIII -4 ug/mL Apolipoprotein C-III 
  \item APOD -4 ug/ml Apolipoprotein D 
  \item APOE -4 ug/ml Apolipoprotein E 
  \item APOH -4 ug/mL Apolipoprotein H 
  \item AR -4 pg/mL Amphiregulin 
  \item AXL -4 ng/mL AXL Receptor Tyrosine Kinase 
  \item B2M -4 ug/mL Beta-2-Microglobulin 
  \item BDNF -4 ng/mL Brain-Derived Neurotrophic Factor 
  \item BLC -4 pg/ml B Lymphocyte Chemoattractant 
  \item BMP6 -4 ng/mL Bone Morphogenetic Protein 6 
  \item BNP -4 pg/ml Brain Natriuretic Peptide  
  \item BTC -4 pg/mL Betacellulin 
  \item C3 -4 mg/mL Complement C3 
  \item CA125 -4 U/mL Cancer Antigen 125 
  \item CA199 -4 U/mL Cancer Antigen 19-9 
  \item CALBINDN -4 ng/ml Calbindin 
  \item CALCITNN -4 pg/mL Calcitonin 
  \item CD40 -4 ng/mL CD 40 antigen 
  \item CD40L -4 ng/mL CD40 Ligand 
  \item CD5L -4 ng/ml CD5 
  \item CEA -4 ng/mL Carcinoembryonic Antigen 
  \item CGA -4 ng/mL Chromogranin-A 
  \item CKMB -4 ng/mL Creatine Kinase-MB 
  \item CLU -4 ug/ml Clusterin 
  \item CNTF -4 pg/mL Ciliary Neurotrophic Factor 
  \item CMPLMNFH -4 ug/ml Complement Factor H 
  \item CORTISOL -4 ng/ml Cortisol 
  \item CPEPTIDE -4 ng/ml C-peptide 
  \item CRP -4 ug/mL C-Reactive Protein 
  \item CTGF -4 ng/ml Connective Tissue Growth Factor 
  \item CYSTATNC -4 ng/ml Cystatin-C 
  \item EGF -4 pg/mL Epidermal Growth Factor 
  \item EGFR -4 ng/mL Epidermal Growth Factor Receptor 
  \item EPTHLDNA -4 ng/mL Epithelial-Derived Neutrophil-Activating 
  \item ENRAGE -4 ng/mL EN-RAGE 
  \item EOTAXIN1 -4 pg/mL Eotaxin-1 
  \item EOTAXIN3 -4 pg/mL Eotaxin-3 
  \item EPO -4 pg/mL Erythropoietin 
  \item EPR -4 pg/mL Epiregulin 
  \item ESELECTN -4 ng/mL E-Selectin 
  \item ET1 -4 pg/mL Endothelin-1 
  \item FABP -4 ng/mL Fatty Acid-Binding Protein- heart  
  \item FACTRVII -4 ng/mL Factor VII 
  \item FAS -4 ng/mL FASLG Receptor 
  \item FASL -4 pg/mL Fas Ligand 
  \item FETUINA -4 ug/ml Fetuin-A 
  \item FGF4 -4 pg/mL Fibroblast Growth Factor 4 
  \item FGFBASI -4 pg/mL Fibroblast Growth Factor basic 
  \item FIBRINGN -4 mg/mL Fibrinogen 
  \item FRTN -4 ng/mL Ferritin 
  \item FSH -4 mIU/mL Follicle-Stimulating Hormone 
  \item G -4 pg/mL Granulocyte Colony-Stimulating Factor 
  \item GH -4 ng/mL Growth Hormone 
  \item GLP1TO -4 pg/ml Glucagon-like Peptide 1- total 
  \item GLUCAGON -4 pg/ml Glucagon 
  \item GRNLCMCS -4 pg/mL Granulocyte-Macrophage Colony-Stimulatin 
  \item GROALPH -4 pg/mL Growth-Regulated alpha protein 
  \item GSTALP -4 ng/ml Glutathione S-Transferase alpha 
  \item HAPTGLBN -4 mg/mL Haptoglobin 
  \item HBEGFGF -4 pg/mL Heparin-Binding EGF-Like Growth Factor 
  \item HCC4 -4 ng/mL Chemokine CC-4 
  \item HGF -4 ng/mL Hepatocyte Growth Factor 
  \item HSP60 -4 ng/ml Heat Shock Protein 60 
  \item I3 -4 pg/mL T Lymphocyte-Secreted Protein I-309 
  \item ICAM -4 ng/mL Intercellular Adhesion Molecule 1 
  \item IFNGAMMA -4 pg/mL Interferon gamma 
  \item IGA -4 mg/mL Immunoglobulin A 
  \item IGE -4 ng/mL Immunoglobulin E 
  \item INSLGFBP -4 ng/mL Insulin-like Growth Factor-Binding Prote 
  \item IGFI -4 ng/mL Insulin-like Growth Factor I 
  \item IGM -4 mg/mL Immunoglobulin M 
  \item IL1ALPHA -4 ng/mL Interleukin-1 alpha 
  \item IL1BETA -4 pg/mL Interleukin-1 beta 
  \item IL10 -4 pg/mL Interleukin-10 
  \item IL11 -4 pg/mL Interleukin-11 
  \item IL12P40 -4 ng/mL Interleukin-12 Subunit p40 
  \item IL12P70 -4 pg/mL Interleukin-12 Subunit p70 
  \item IL13 -4 pg/mL Interleukin-13 
  \item IL15 -4 ng/mL Interleukin-15 
  \item IL16 -4 pg/mL Interleukin-16 
  \item IL18 -4 pg/mL Interleukin-18 
  \item IL1R -4 pg/mL Interleukin-1 receptor antagonist 
  \item IL2 -4 pg/mL Interleukin-2 
  \item IL25 -4 pg/mL Interleukin-25 
  \item IL3 -4 ng/mL Interleukin-3 
  \item IL4 -4 pg/mL Interleukin-4 
  \item IL5 -4 pg/mL Interleukin-5 
  \item IL6 -4 pg/mL Interleukin-6 
  \item IL6R -4 ng/mL Interleukin-6 receptor 
  \item IL7 -4 pg/mL Interleukin-7 
  \item IL8 -4 pg/mL Interleukin-8 
  \item INSULIN -4 uIU/mL Insulin 
  \item IP -4 pg/ml Interferon gamma Induced Protein 10 
  \item KIM1 -4 ng/ml Kidney Injury Molecule-1  
  \item LEPTIN -4 ng/mL Leptin 
  \item LH -4 mIU/mL Luteinizing Hormone  
  \item LOX -4 ng/mL Lectin-Like Oxidized LDL Receptor 1 
  \item A -4 ug/mL Apolipoprotein
  \item LYMPHTCT -4 ng/mL Lymphotactin 
  \item MCP1 -4 pg/mL Monocyte Chemotactic Protein 1 
  \item MCP2 -4 pg/ml Monocyte Chemotactic Protein 2 
  \item MCP3 -4 pg/mL Monocyte Chemotactic Protein 3 
  \item MCP4 -4 pg/ml Monocyte Chemotactic Protein 4 
  \item MCLNSF1 -4 ng/mL Macrophage Colony-Stimulating Factor 1 
  \item MLNDMLDL -4 ng/mL Malondialdehyde-Modified Low-Density Lip 
  \item MDC -4 pg/mL Macrophage-Derived Chemokine 
  \item MCRPHMIF -4 ng/mL Macrophage Migration Inhibitory Factor 
  \item MI -4 pg/ml Monokine Induced by Gamma Interferon 
  \item MCRPHIP1 -4 pg/mL Macrophage Inflammatory Protein-1 alpha 
  \item MIPRT1B -4 pg/mL Macrophage Inflammatory Protein-1 beta 
  \item MCRPHIP3 -4 pg/ml Macrophage Inflammatory Protein-3 alpha 
  \item MMP1 -4 ng/ml Matrix Metalloproteinase-1 
  \item MMP10 -4 ng/ml Matrix Metalloproteinase-10 
  \item MMP2 -4 ng/mL Matrix Metalloproteinase-2 
  \item MMP3 -4 ng/mL Matrix Metalloproteinase-3 
  \item MMP7 -4 ng/ml Matrix Metalloproteinase-7 
  \item MMP9 -4 ng/mL Matrix Metalloproteinase-9 
  \item MM9T -4 ng/ml Matrix Metalloproteinase-9- total 
  \item MYLDPIF1 -4 ng/mL Myeloid Progenitor Inhibitory Factor 1 
  \item MPO -4 ng/mL Myeloperoxidase 
  \item MYOGLOBN -4 ng/mL Myoglobin 
  \item NTRPHGAL -4 ng/ml Neutrophil Gelatinase-Associated Lipocal 
  \item NGFBETA -4 ng/mL Nerve Growth Factor beta 
  \item NRCAM -4 ng/mL Neuronal Cell Adhesion Molecule 
  \item OSTEPNTN -4 ng/ml Osteopontin 
  \item PAI1 -4 ng/mL Plasminogen Activator Inhibitor 1 
  \item PAP -4 ng/mL Prostatic Acid Phosphatase 
  \item P -4 mIU/mL Pregnancy-Associated Plasma Protein A 
  \item PLMNRARC -4 ng/mL Pulmonary and Activation-Regulated Chemo 
  \item PDGF -4 pg/ml Platelet-Derived Growth Factor BB 
  \item PLGF -4 pg/ml Placenta Growth Factor 
  \item PPP -4 pg/ml Pancreatic Polypeptide 
  \item PRL -4 ng/ml Prolactin 
  \item PROGSTRN -4 ng/ml Progesterone 
  \item PRNSLNIN -4 pM Proinsulin- Intact 
  \item PRNSLNTT -4 pM Proinsulin- Total 
  \item PSAF -4 ng/mL Prostate-Specific Antigen- Free 
  \item PYY -4 pg/mL Peptide YY 
  \item RCPTRFRD -4 ng/mL Receptor for advanced glycosylation end 
  \item RANTES -4 ng/mL T-Cell-Specific Protein RANTES 
  \item RESISTIN -4 ng/ml Resistin 
  \item S100B -4 ng/mL S100 calcium-binding protein B 
  \item SAP -4 ug/mL Serum Amyloid P-Component 
  \item SCF -4 pg/mL Stem Cell Factor 
  \item SECRETIN -4 ng/mL Secretin 
  \item SRMGLTOT -4 ug/mL Serum Glutamic Oxaloacetic Transaminase 
  \item SHBG -4 nmol/L Sex Hormone-Binding Globulin 
  \item SOD1 -4 ng/mL Superoxide Dismutase 1- Soluble 
  \item SORTILIN -4 ng/mL Sortilin 
  \item TBG -4 ug/mL Thyroxine-Binding Globulin 
  \item TECK -4 ng/mL Thymus-Expressed Chemokine 
  \item TSTSTRNT -4 ng/ml Testosterone- Total 
  \item TF -4 ng/mL Tissue Factor 
  \item TFF3 -4 ug/ml Trefoil Factor 3 
  \item TGFAL -4 pg/mL Transforming Growth Factor alpha 
  \item TGFB -4 pg/mL Transforming Growth Factor beta-3 
  \item THP -4 ug/ml Tamm-Horsfall Urinary Glycoprotein 
  \item THRMBSP1 -4 ng/mL Thrombospondin-1 
  \item TSSINHM1 -4 ng/mL Tissue Inhibitor of Metalloproteinases 1 
  \item TM -4 ng/ml Thrombomodulin 
  \item TNC -4 ng/mL Tenascin-C 
  \item TNFALPHA -4 pg/mL Tumor Necrosis Factor alpha 
  \item TNFBETA -4 pg/mL Tumor Necrosis Factor beta 
  \item T -4 ng/mL Tumor Necrosis Factor Receptor-Like 2 
  \item THRMBPTN -4 ng/mL Thrombopoietin 
  \item TNFRAILR -4 ng/mL TNF-Related Apoptosis-Inducing Ligand Re 
  \item TRNSFRRN -4 mg/dl Serotransferrin 
  \item TSH -4 uIU/mL Thyroid-Stimulating Hormone  
  \item TTR -4 mg/dl Transthyretin 
  \item VCAM -4 ng/mL Vascular Cell Adhesion Molecule-1 
  \item VEGF -4 pg/mL Vascular Endothelial Growth Factor 
  \item VITRNCTN -4 ug/ml Vitronectin 
  \item VKDPS -4 ug/ml Vitamin K-Dependent Protein S 
  \item VWF -4 ug/mL von Willebrand Factor 
}
}
\keyword{datasets}
