% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{rbmraw}
\alias{rbmraw}
\title{Rules Based Medicine Plasma Multiplex Raw Data}
\format{
A data frame with 201970 rows and 12 variables
}
\usage{
data(rbmraw)
}
\description{
Rules Based Medicine Plasma Multiplex Raw Data
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item RBMID -4 -4 Sample ID from RBM 
  \item VISCODE -4 -4 Visit code
  \item ANALYTE -4 -4 Name of Analyte with Units 
  \item LDD -4 -4 Least Detectable Dose (see primer for details) 
  \item AVALUE -4 -4 Recorded Value 
  \item ANALVAL -4 -4 Numeric Value after possible imputation (see primer for details; note that some analytes are all blank due to too many <LOW> values) 
  \item BELOWLDD -4 -4 Is analval < LDD? Note: this flag pertains to both recorded value and imputed value
  \item READLOW -4 -4 Is recorded value <LOW> or numeric? (see primer for details) 
  \item LOGTRANS -4 -4 Is analval log10 transformed?
  \item OUTLIER -4 -4 Is analval considered an outlier (see primer for details)
}
}
\keyword{datasets}
