% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{recadsx}
\alias{recadsx}
\title{Diagnosis and Symptoms Log}
\format{
A data frame with 881 rows and 15 variables
}
\usage{
data(recadsx)
}
\description{
Diagnosis and Symptoms Log
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N NA Participant roster ID
  \item SITEID N NA Site ID
  \item VISCODE T NA Visit code
  \item USERDATE S NA Date record created
  \item USERDATE2 S NA Date record last updated
  \item RECNO N NA 
  \item SXSYMNO N NA Symptom Number
  \item SXSYMP T NA Description of Symptom
  \item SXSEVER N NA Severity
  \item SXCHRON N NA Chronicity
  \item SXONSET D NA Date of Onset
  \item SXCONTD N NA Is the symptom ongoing?
  \item SXCEASE D NA Date Ceased
}
}
\keyword{datasets}
