% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{recadv}
\alias{recadv}
\title{Adverse Events/Hospitalizations - Log}
\format{
A data frame with 19407 rows and 36 variables
}
\usage{
data(recadv)
}
\description{
Adverse Events/Hospitalizations - Log
}
\details{
A dataset with variables as follows:

\itemize{
  \item AEHCSHR T HHMM Estimated Cease Time:
  \item AENUMBER N  Adverse Event Number
  \item AERELATSP N  Related to other study procedure(s)
  \item AEHTAU N  Related to AV-1451 (Tau) Imaging
  \item AEHAV45 N NA Investigator Judgment of Relatedness to F-AV-45
  \item AEHONSHR N NA Estimated Onset Time:
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item RECNO N  
  \item EXAMDATE D  Examination Date
  \item AEHBLSYM T  Check here if:
  \item AEHONSDT D  Onset Date
  \item AEHONGNG N  Is the event ongoing?
  \item AEHCSDT D  Cease Date
  \item AEHCHRON N  Chronicity
  \item AEHSEVR N  Severity
  \item AEHSERIO N  Serious?
  \item AEHSAE T  Check here if:
  \item AEHLIFE N  Life-Threatening?
  \item AEHIMG N  Related to Imaging Procedure
  \item AEHLUMB N  Related to Lumbar Puncture
  \item AEHCMEDS N  Concurrent Medication Prescribed or Changed
  \item AEHINHOS N  Did this event occur while the participant was being hospitalized for another event?
  \item AEHPRO N  If Yes, did this event prolong hospitalization?
  \item AEHHOSP N  If No, did this event require hospitalization?
  \item AEHOUTDT D  If Outpatient, provide the date of visit
  \item AEHADMDT D  Admit Date
  \item AEHDISDT D  Discharge Date
  \item AEHDEATH N  Did this event result in death?
  \item AEHDTHDT D  Date of death
  \item AEHALZHI N  Was diagnosis of Alzheimer's confirmed at autopsy?
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
