% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{recbllog}
\alias{recbllog}
\title{Documentation of Baseline Symptoms Log}
\format{
A data frame with 12848 rows and 15 variables
}
\usage{
data(recbllog)
}
\description{
Documentation of Baseline Symptoms Log
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item RECNO N  
  \item EXAMDATE D  Examination Date
  \item BSXSYMNO N  Symptom Number
  \item BSXSEVER N  Severity
  \item BSXCHRON N  Chronicity
  \item BSXONSET D  Date of Onset
  \item BSXCONTD N  Is the symptom ongoing?
  \item BSXCEASE D  Date Ceased
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
