% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{recfhq}
\alias{recfhq}
\title{Family History Questionnaire Subtable}
\format{
A data frame with 7407 rows and 12 variables
}
\usage{
data(recfhq)
}
\description{
Family History Questionnaire Subtable
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item RECNO N  
  \item EXAMDATE D  Examination Date
  \item FHQGNDR N  Gender
  \item FHQSIB N  Dementia
  \item FHQSIBAD N  Alzheimer's Disease
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
