% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{recmhist}
\alias{recmhist}
\title{Medical History}
\format{
A data frame with 30810 rows and 13 variables
}
\usage{
data(recmhist)
}
\description{
Medical History
}
\details{
A dataset with variables as follows:

\itemize{
  \item MHDTONSET D  Date of onset:
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item RECNO N  
  \item EXAMDATE D  Examination Date
  \item MHNUM N  System Number
  \item MHCUR N  Is the problem current?
  \item MHSTAB N  If Yes, is the problem stable?
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
