% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{redox}
\alias{redox}
\title{Redox Reactive Reagents LLC ADNI R-RAA ELISA}
\format{
A data frame with 90 rows and 11 variables
}
\usage{
data(redox)
}
\description{
Redox Reactive Reagents LLC ADNI R-RAA ELISA
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID NA  Participant roster ID
  \item VISCODE NA  Visit code
  \item PSG NA Interpolated OD PSG 10% Hemin Treated minus Untreated
  \item PSGSD NA  PSG 10% Hemin Treated minus Untreated SD
  \item CLG NA Interpolated OD CLG 10% Hemin Treated minus Untreated
  \item CLGSD NA  CLG 10% Hemin Treated minus Untreated SD
  \item PEG NA Interpolated OD PEG 10% Hemin Treated minus Untreated
  \item PEGSD NA  PEG 10% Hemin Treated minus Untreated SD
  \item PCG NA Interpolated OD PCG 1% Hemin Treated minus Untreated
  \item PCGSD NA  PCG 1% Hemin Treated minus Untreated SD
}
}
\keyword{datasets}
