% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{registry}
\alias{registry}
\title{Registry}
\format{
A data frame with 26262 rows and 19 variables
}
\usage{
data(registry)
}
\description{
Registry
}
\details{
A dataset with variables as follows:

\itemize{
  \item RGSTATUS N  Was a screening visit/ADNI2 initial visit performed?
  \item RGSOURCE N  Who participated in this visit?
  \item PTSTATUS N NA Participant Status:
  \item VISTYPE N NA Visit Type:
  \item CHANGTR N NA Would you like to change this visit's track or the participant's visit schedule?
  \item CGTRACK N NA Please select from the following visit's tracks:
  \item CGTRACK2 N NA Please select from the following visit schedule:
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item RGCONDCT N  Was this visit conducted?
  \item RGREASON N  Reason why the visit was not conducted:
  \item EXAMDATE D  Examination Date
  \item RGRESCRN N  Is this a rescreen?
  \item RGPREVID T  If Yes, what was the participant's initial ID number?
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
