% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{residual_aliquot_list}
\alias{residual_aliquot_list}
\title{Residual aliquot count in the LDMS database}
\format{
A data frame with 1661 rows and 11 variables
}
\usage{
data(residual_aliquot_list)
}
\description{
Residual aliquot count in the LDMS database
}
\details{
A dataset with variables as follows:

\itemize{
  \item COLPROT -4 -4 Study protocol of data collection
  \item RID -4  Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item SITE -4  Hospital/Site ID
  \item VID -4  Visit ID from sample tube
  \item VOL -4 mL Volume of residual CSF left after analysis
  \item GUSPECID -4  Global Specimen ID identifying individual aliquots
  \item RECDTE -4 YYYYMMDD Date of sample receipt at UPenn
  \item STORDTE -4 YYYYMMDD Date of aliquoting and storing
}
}
\examples{
\dontrun{
describe(residual_aliquot_list)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/RESIDUAL_ALIQUOT_LIST/ADNI_Methods_Template_Biomarker\%20Core\%20Residual\%20CSF\%20aliquot\%20samples\%20v1.pdf}
}
\keyword{datasets}
