% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{rorr}
\alias{rorr}
\title{Return of Research Results}
\format{
A data frame with 34 rows and 24 variables
}
\usage{
data(rorr)
}
\description{
Return of Research Results
}
\details{
A dataset with variables as follows:

\itemize{
  \item RECIPPT N NA 7a. Were the research results returned to the Participant?
  \item RECIPFAM N NA 7b. Were the research results returned to a Family Member?  
  \item RECIPFR N NA 7c.  Were the research results returned to a Friend?
  \item RECIPSP N NA 7d.  Were the research results returned to a Study Partner?
  \item RECIPCLIN N NA 7e.  Were the research results returned to a Treating Clinician?
  \item RECIPOTH N NA 7f.  Were the research results returned to another individual (Other, Specify)
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
  \item RECNO N  
  \item TYPE N  1. What type of research results were disclosed?
  \item REQUESTED N  2. Who requested the return of research results?
  \item CONSENT N  4. Who gave informed consent for the decision to return research results?
  \item IRBPLAN N  5. Was the plan to return research results presented to your institution's IRB?
  \item WAIVER N  c) Was a waiver obtained?
  \item EXCEPOBT N  d) Was an exception obtained?
  \item COGSTATUS N  6. What was the participant's cognitive status at the time these results were returned?
  \item RECIPIENT T  7. Who were the research results returned to? (check all that apply)
  \item RETURNMETH N  How were results returned (verbally or in writing)?
  \item RETURNROLE N  9. What is this person's role in ADNI?
}
}
\examples{
\dontrun{
describe(rorr)
}
}
\keyword{datasets}
