% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{roster}
\alias{roster}
\title{Roster}
\format{
A data frame with 5306 rows and 7 variables
}
\usage{
data(roster)
}
\description{
Roster
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item USERDATE S  Date record created
  \item PTID T  Participant ID
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
