% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adnimerger.R
\name{scoreMOCA}
\alias{scoreMOCA}
\title{Score Montreal Cognitive Assessment (MOCA) without Education Adjustment.}
\usage{
scoreMOCA(
  dd,
  qlist = list(visuo = c(i1 = "TRAILS", i2 = "CUBE", i3Contour = "CLOCKCON", i3Numbers =
    "CLOCKNO", iHands = "CLOCKHAN"), naming = c(i4Lion = "LION", i4Rhino = "RHINO",
    i4Camel = "CAMEL"), memory = paste("IMMT", rep(1:2, c(5, 5)), paste("W", 1:5, sep =
    ""), sep = ""), attention = c(i6ForDigitSpan = "DIGFOR", i6BackDigitSpan = "DIGBACK",
    i6Vigilance = "LETTERS", i6Serial1 = "SERIAL1", i6Serial2 = "SERIAL2", i6Serial3 =
    "SERIAL3", i6Serial4 = "SERIAL4", i6Serial5 = "SERIAL5"), language =
    c(i7SentenceReptition = paste("REPEAT", 
     1:2, sep = ""), i8VerbalFluency =
    "FFLUENCY"), abstraction = c(i9Abstraction1 = "ABSTRAN", i9Abstraction2 = "ABSMEAS"),
    delayedRecall = paste("DELW", 1:5, sep = ""), orientation = c(i11Date = "DATE",
    i11Month = "MONTH", i11Year = "YEAR", i11Day = "DAY", i11Place = "PLACE", i11City =
    "CITY")),
  trueCoding = "Correct",
  delayedRecallTrueCoding = "Correct with No Cue",
  qScore = TRUE
)
}
\arguments{
\item{dd}{data frame to score.}

\item{qlist}{Item list}

\item{trueCoding}{String for correct responses (default is `Correct`)}

\item{delayedRecallTrueCoding}{defaults to `Correct with No Cue``}

\item{qScore}{defaults to `TRUE`}
}
\value{
a data.frame with MOCA scores.
}
\description{
This function scores the MOCA.
}
\references{
\url{http://www.parkinsons.va.gov/consortium/MOCA.asp}
}
\author{
Chung-Kai Sun \email{chungkas@usc.edu}
}
