% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{sourceup}
\alias{sourceup}
\title{Source Doc Upload}
\format{
A data frame with 21151 rows and 11 variables
}
\usage{
data(sourceup)
}
\description{
Source Doc Upload
}
\details{
A dataset with variables as follows:

\itemize{
  \item SRCDOC N NA Source Documents to Upload:
  \item SRCDOCOTH T NA If Other, Specify: 
  \item SRCSTAT N NA Initial or Updated Version:
  \item SRCVFY N NA Has all PHI been removed from the documents to be uploaded?
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
}
}
\examples{
\dontrun{
describe(sourceup)
}
}
\keyword{datasets}
