% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{spconsent}
\alias{spconsent}
\title{Study Partner Consent}
\format{
A data frame with 2058 rows and 13 variables
}
\usage{
data(spconsent)
}
\description{
Study Partner Consent
}
\details{
A dataset with variables as follows:

\itemize{
  \item SPID N NA Study Partner ID Number:
  \item SPINCDATE D YYYY-MM-DD Date of consent:
  \item SPPROTVERSN N NA Indicate Version of Protocol
  \item SPCONSVERSN N NA Indicate Coordinating Center consent template version signed:  
  \item SPIRB D YYYY-MM-DD Consent IRB approval date:
  \item SPIRBEXPDT D YYYY-MM-DD Consent IRB expiration date:
  \item SPSITEVRS T NA Site version of consent:  
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
}
}
\examples{
\dontrun{
describe(spconsent)
}
}
\keyword{datasets}
