% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{spinfo}
\alias{spinfo}
\title{Study Partner Information}
\format{
A data frame with 1518 rows and 16 variables
}
\usage{
data(spinfo)
}
\description{
Study Partner Information
}
\details{
A dataset with variables as follows:

\itemize{
  \item INFSPID N NA Study Partner ID Number:
  \item INFRELAT N NA Relationship of the study partner to the participant
  \item INFOTHER T NA If Paid caregiver or Other, specify:
  \item INFGENDER N NA Gender of study partner
  \item INFAGE N NA Is study partner 90 or older?
  \item INFAGESP N NA If No, age of study partner:
  \item INFLIVE N NA Do the study partner and participant live together?
  \item INFHRS N NA How many hours per week does the study partner spend with the participant in-person?
  \item INFHRSOT N NA Other than in-person, how many hours per week does the study partner spend interacting with the participant?
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
}
}
\examples{
\dontrun{
describe(spinfo)
}
}
\keyword{datasets}
