% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{strokesum}
\alias{strokesum}
\title{Stroke Summary}
\format{
A data frame with 3648 rows and 11 variables
}
\usage{
data(strokesum)
}
\description{
Stroke Summary
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N -4 Participant roster ID
  \item VISCODE T -4 Visit code
  \item EXAMDATE D -4 Examination Date
  \item SIDE N -4 Side
  \item SIZE N -4 Size
  \item STROKEFORM N -4 Stroke number (1-100)
  \item ANATOMTER N -4 Anatomical Territory
  \item STROKETYPE T -4 Stoke type
  \item LOCATIONXYZ T -4 Location (xyz coordinate)
  \item WHITMATHYP T -4 White matter hyperintensivity volume -- whole brain.
}
}
\keyword{datasets}
