% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{studysum}
\alias{studysum}
\title{Study Summary}
\format{
A data frame with 731 rows and 24 variables
}
\usage{
data(studysum)
}
\description{
Study Summary
}
\details{
A dataset with variables as follows:

\itemize{
  \item SDSTATUS N NA Final disposition
  \item SDDATE D YYYY-MM-DD Final ADNI3 disposition date
  \item INCLUSION N NA Never enrolled due to: Failing to meet Inclusion Criteria
  \item EXCLUSION N NA Never enrolled due to: Meeting Exclusion Criteria
  \item VERSION N NA Indicate version of criteria evaluated:
  \item INCROLL T NA If Never enrolled due to failing Inclusion Criteria select criteria:
  \item INCNEWPT T NA If Never enrolled due to failing Inclusion Criteria select criteria:
  \item EXCCRIT T NA If Never enrolled due to meeting Exclusion Criteria select criteria:
  \item MRIFIND N NA Never enrolled due to: MRI finding(s) other than those listed in Exclusion Criteria
  \item NVRDISC N NA Never enrolled due to: Discontinuation (e.g. consent withdrawal, lost to follow-up)
  \item NVROT N NA Never enrolled due to Other reason(s)
  \item NVRSPECIFY T NA If Other, specify
  \item SDPRIMARY N NA Primary reason for early discontinuation of study visits
  \item SDPRIMAE N NA If Adverse Event, provide Adverse Event number
  \item SDSECOND N NA Were there other contributing reason(s) for early discontinuation of study visits?
  \item SDSECREAS N NA If Yes, select other contributing reason(s) for early discontinuation of study visits:
  \item AENUM T NA If Adverse Event, provide Adverse Event number
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
