% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{subject}
\alias{subject}
\title{Participant Registration}
\format{
A data frame with 1915 rows and 8 variables
}
\usage{
data(subject)
}
\description{
Participant Registration
}
\details{
A dataset with variables as follows:

\itemize{
  \item PTTYPE N NA Indicate Participant Type
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
}
}
\examples{
\dontrun{
describe(subject)
}
}
\keyword{datasets}
