% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{taumeta}
\alias{taumeta}
\title{Tau AV-1451 PET Scan Information}
\format{
A data frame with 2039 rows and 45 variables
}
\usage{
data(taumeta)
}
\description{
Tau AV-1451 PET Scan Information
}
\details{
A dataset with variables as follows:

\itemize{
  \item QCTIME T HHMM Time of today's Scanner QC
  \item ASSAYTIME T HHMM Time of AV-1451 dose assay
  \item INJTIME T HHMM Time of AV-1451 injection
  \item SCANTIME T HHMM Emission Scan Start Time
  \item VARIAT N NA Any variations from protocol during AV-1451 uptake?
  \item PROTID T NA Predefined Acquisition Protocol Name
  \item MOTION N NA Participant motion problems
  \item OTHERVAR N NA Other protocol variations
  \item RECON N NA Select reconstruction used
  \item SHARP N NA For Phillips scanners:  was "smooth" parameter set to "sharp"?
  \item ARCHIVE N NA Data archived locally?
  \item LONI N NA Data transferred to LONI?
  \item TRANDATE D YYYY-MM-DD Transfer Date
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
  \item RECNO N  
  \item DONE N  Was scan conducted?
  \item NDREASON N  If No, reason not done:
  \item RADTRACER N  If No, was radiotracer administered?
  \item SCANDATE D  Scan Date/Dose Date
  \item TAUQCTIME T HHMM Time of today's Scanner QC
  \item TAUDOSE N mCi AV-1451 dose assay
  \item TAUTIME T HHMM Time of AV-1451 dose assay
  \item TAUINJTIME T HHMM Time of AV-1451 injection
  \item TAUSCANTM T HHMM Emission Scan Start Time
  \item TAUVARIAT N  Any variations from protocol during AV-1451 uptake?
  \item TAUVARSP T  If Yes, describe:
  \item TAUSCANDIF T  If scan start time is not between 73 and 77 min post-injection, provide explanation
  \item TAUPROTID T  Predefined Acquisition Protocol Name
  \item TAUFRAME N  For the acquisition, which framing rate was used?
  \item TAUFRAMESP T  If other, specify
  \item TAUMOTION N  Participant motion problems:
  \item TAUOTHER N  Other protocol variations:
  \item TAUARCHIVE N  Data archived locally?
  \item TAULONI N  Data transferred to LONI?
  \item TAUTRANDT D  Transfer Date
  \item TAURECON N  Check which of the following reconstructions was used:
  \item TAURAMLA N  Lambda
  \item TAUSHARP T  Was "Smooth" parameter set to "Sharp"?
  \item TAUSMOOTH T  Was post-process smoothing performed:
}
}
\examples{
\dontrun{
describe(taumeta)
}
}
\keyword{datasets}
