% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{tauphone}
\alias{tauphone}
\title{Tau AV-1451 Follow-up Phone Contact}
\format{
A data frame with 124 rows and 12 variables
}
\usage{
data(tauphone)
}
\description{
Tau AV-1451 Follow-up Phone Contact
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
  \item DONE N  Was assessment/procedure done?
  \item NDREASON N  If No, reason not done:
  \item TAUPHONEDT D MM/DD/YYYY Date Completed:
  \item TAUPHONETM T HHMM Time Completed:
  \item TAUPHONEAE N  Were any Advere Events reported?
}
}
\examples{
\dontrun{
describe(tauphone)
}
}
\keyword{datasets}
