% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{tauqc}
\alias{tauqc}
\title{Tau AV-1451 PET QC}
\format{
A data frame with 1698 rows and 29 variables
}
\usage{
data(tauqc)
}
\description{
Tau AV-1451 PET QC
}
\details{
A dataset with variables as follows:

\itemize{
  \item REVDT D YYYY-MM-DD Date Reviewed
  \item ALLFRAME N NA Were all frames acceptable?
  \item UNUSABL T NA If No, indicate which frames were unacceptable
  \item UNRSN N NA If No, indicate why frames were unacceptable
  \item UNRSNSPEC T NA If Other, Specify
  \item PROCERR T NA If Yes, select processing error(s)
  \item PROCERRSPEC T NA If Other, Specify
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
  \item RECNO N  
  \item SCANDATE D MM/DD/YYYY Scan Date:
  \item LONIUPDT D MM/DD/YYYY Date uploaded to LONI
  \item LONIUID T  LONI Unique Series ID
  \item TAUREVDT D MM/DD/YYYY Date Reviewed:
  \item TAUALLFRM N  Were all frames acceptable?
  \item TAUUNUSABL T  If No, indicate which frames were unacceptable?
  \item TAUUNRSN N  If No, indicate why frames were unacceptable:
  \item TAUUNRSNSP T  If Other, Specify:
  \item SCANQLTY N  Scan Pass QC?
  \item TECHERR N  If No, was this due to technologist error?
  \item RESCANREQ N  Rescan requested?
  \item REPROCREQ N  Reprocessing requested?
  \item TAUPROERR T  If Yes, select processing error(s):
  \item TAUERRSP T  If Other, Specify:
}
}
\examples{
\dontrun{
describe(tauqc)
}
}
\keyword{datasets}
