% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{tbm}
\alias{tbm}
\title{Paul Thompson's Lab, Imaging Genetics Center (IGC), University of Southern California - Cross-sectional and longitudinal tensor-based morphometry Versions 2.0 and 2.1}
\format{
A data frame with 9354 rows and 18 variables
}
\usage{
data(tbm)
}
\description{
Paul Thompson's Lab, Imaging Genetics Center (IGC), University of Southern California - Cross-sectional and longitudinal tensor-based morphometry Versions 2.0 and 2.1
}
\details{
A dataset with variables as follows:

\itemize{
  \item COLPROT -4 -4 Study protocol of data collection
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item EXAMDATE -4 -4 Examination Date
  \item VERSION -4 -4 -4
  \item IMAGEUID_1 -4 -4 IMAGE ID of the screening scan
  \item IMAGEUID_2 -4 -4 IMAGE ID of the follow-up scan; see VISCODE
  \item RUNDATE -4 -4 -4
  \item STATUS -4 -4 -4
  \item MEASURE_1 -4 -4 Numerical summary of cumulative temporal lobe atrophy  average within a statistically defined region-of-interest (p<0.00001) inside the temporal lobes; see VISCODE; summaries are scaled by 1000 (e.g. 1000: no change, 1200: 20% increase, 800: 20% loss)
  \item MEASURE_2 -4 -4 Numerical summary of cumulative temporal lobe atrophy  average within an anatomically defined region-of-interest including bilateral temporal lobes; see VISCODE; summaries are scaled by 1000 (e.g. 1000: no change, 1200: 20% increase, 800: 20% loss)
  \item SEQUENCE -4 -4 -4
  \item ACCEL_NONACCEL -4 -4 -4
  \item DESCRIPTION -4 -4 -4
  \item FIELD_STRENGTH -4 -4 -4
  \item EXTRA_NOTES_ON_PROCESSED_IMAGES -4 -4 -4
  \item update_stamp -4 -4 -4
  \item FIELD.STRENGTH -4 -4 -4
  \item ACCEL.NONACCEL -4 -4 -4
  \item ACQUISITION -4 -4 -4
}
}
\examples{
\dontrun{
describe(tbm)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/TBM/ADNI_Methods_TBM_IGC_LONI_Oct2012_2.pdf}, \url{https://adni.bitbucket.io/reference/docs/TBM/ADNI_Methods_TBM_IGC_LONI_Oct2012.pdf}
}
\keyword{datasets}
