% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{tcv}
\alias{tcv}
\title{Total Cranial Vault Segmentation}
\format{
A data frame with 810 rows and 10 variables
}
\usage{
data(tcv)
}
\description{
Total Cranial Vault Segmentation
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4  Participant roster ID
  \item VISCODE -4  Visit code
  \item EXAMDATE -4  Examination Date
  \item LONIUID -4  Study ID
  \item RUNDATE -4  Run Date
  \item STATUS -4  Status
  \item VERSION -4  Version Number
  \item T2TCV -4 cc T2 Total Intracranial Volume
  \item GRADE -4  Grade
}
}
\examples{
\dontrun{
describe(tcv)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/TCV/TCVGradeSheet.pdf}
}
\keyword{datasets}
