% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{tomm40}
\alias{tomm40}
\title{TOMM40 PolyT variable length polymorphism}
\format{
A data frame with 757 rows and 4 variables
}
\usage{
data(tomm40)
}
\description{
TOMM40 PolyT variable length polymorphism
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N NA Participant roster ID
  \item TOMM40ALLELE1 N NA 
  \item TOMM40ALLELE2 N NA 
}
}
\keyword{datasets}
