% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{transfer}
\alias{transfer}
\title{Participant Transfers}
\format{
A data frame with 40 rows and 15 variables
}
\usage{
data(transfer)
}
\description{
Participant Transfers
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
  \item RECNO N  
  \item PTTYPE N  Type of transfer
  \item SITEIDA T  Site conducting visits PRIOR to transfer:
  \item PTPVISIT T  Visits conducted prior to transfer:
  \item SITEIDB T  Site conducting visits AFTER transfer:
  \item SITEIDTMP T  Temporary site:
  \item PTTVISIT T  Visit(s) conducted at site:
  \item PTREASON T  Reason for Transfer:
}
}
\examples{
\dontrun{
describe(transfer)
}
}
\keyword{datasets}
