% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{treatdis}
\alias{treatdis}
\title{Early Discontinuation and Withdrawal}
\format{
A data frame with 1895 rows and 13 variables
}
\usage{
data(treatdis)
}
\description{
Early Discontinuation and Withdrawal
}
\details{
A dataset with variables as follows:

\itemize{
  \item RECNO N  
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item WDRAWTYPE N  1. Is this a Full or Partial Withdrawal or Completion of ADNI 1, not continuing to extension protocol?
  \item WDPARTIAL T  If Partial, what is the participant withdrawing from?
  \item WDREASON T  Reason for Withdrawal
  \item WDFOLLOWUP T  2. Follow-up:  (check all that apply)
  \item USERDATE2 S  Date record last updated
}
}
\keyword{datasets}
