% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{uaspmvbm}
\alias{uaspmvbm}
\title{UA (Gene Alexander) MRI SPM voxel based morphometry (VBM) analysis}
\format{
A data frame with 4091 rows and 126 variables
}
\usage{
data(uaspmvbm)
}
\description{
UA (Gene Alexander) MRI SPM voxel based morphometry (VBM) analysis
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item RECNO N  
  \item EXAMDATE D  Examination Date
  \item IMAGEUID N  imageUID
  \item FLDSTRENG N  Field Strength
  \item PRECENTL N Mean Gray Matter Precentral_L. 1) Mean gray matter values extracted from 1.5T ADNI MPRAGE MRI segmented gray matter maps using longitudinal voxel based morphometry (VBM) processing steps (Chetelat et al., Neuroimage, 2005, 27:934-946) adapted for SPM5 with customized tissue-specific priors derived from a representative sample of 300 ADNI subjects (ADNI300). The regions of interest (ROI) were defined by Anatomical Automatic Labeling (AAL, Tzourio-Mazoyer et al., Neuroimage, 2002, 15:273-89, http://www.cyceron.fr/freeware/) after warping the standard AAL ROI template into the customized sample-specific ADNI300 template space. The ADNI300-customized AAL ROI template was applied to spatially normalized and segmented SPM5 gray matter maps for the longitudinal VBM generated baseline and 6 month (indicated by RECNO=2) or 12 month (indicated by RECNO=3) follow up scans without Gaussian smoothing to reduce averaging across regional borders. 2) The ADNI300 template was created from AD, MCI, and Normal subjects proportional to the distribution of these subject groups enrolled in the whole ADNI study and the groups were matched in age and gender.  3) Regional values will be subject to change due to study alterations in image pre- or post-processing, inclusion of additional analysis time points, and/or modifications or updates in image analysis procedures. 4) SPM analyses were performed by ADNI Analysis Team Members: Gene E. Alexander, Ph.D., Kewei Chen, Ph.D., and Eric M. Reiman, M.D.
  \item PRECENTR N  Precentral_R
  \item FRONTSUPL N  Frontal_Sup_L
  \item FRONTSUPR N  Frontal_Sup_R
  \item FRONTSORBL N  Frontal_Sup_Orb_L
  \item FRONTSORBR N  Frontal_Sup_Orb_R
  \item FRONTMIDL N  Frontal_Mid_L
  \item FRONTMIDR N  Frontal_Mid_R
  \item FRTMIDORBL N  Frontal_Mid_Orb_L
  \item FRTMIDORBR N  Frontal_Mid_Orb_R
  \item FRONTINOPL N  Frontal_Inf_Oper_L
  \item FRONTINOPR N  Frontal_Inf_Oper_R
  \item FRONTINTRL N  Frontal_Inf_Tri_L
  \item FRONTINTRR N  Frontal_Inf_Tri_R
  \item FRONTINOBL N  Frontal_Inf_Orb_L
  \item FRONTINOBR N  Frontal_Inf_Orb_R
  \item ROLANDOPL N  Rolandic_Oper_L
  \item ROLANDOPR N  Rolandic_Oper_R
  \item SUPMOTORL N  Supp_Motor_Area_L
  \item SUPMOTORR N  Supp_Motor_Area_R
  \item OLFACTL N  Olfactory_L
  \item OLFACTR N  Olfactory_R
  \item FRONTSMEDL N  Frontal_Sup_Medial_L
  \item FRONTSMEDR N  Frontal_Sup_Medial_R
  \item FRTMEDORBL N  Frontal_Med_Orb_L
  \item FRTMEDORBR N  Frontal_Med_Orb_R
  \item RECTUSL N  Rectus_L
  \item RECTUSR N  Rectus_R
  \item INSULAL N  Insula_L
  \item INSULAR N  Insula_R
  \item CINGANTL N  Cingulum_Ant_L
  \item CINGANTR N  Cingulum_Ant_R
  \item CINGMIDL N  Cingulum_Mid_L
  \item CINGMIDR N  Cingulum_Mid_R
  \item CINGPOSTL N  Cingulum_Post_L
  \item CINGPOSTR N  Cingulum_Post_R
  \item HIPPL N  Hippocampus_L
  \item HIPPR N  Hippocampus_R
  \item PARAHIPPL N  ParaHippocampal_L
  \item PARAHIPPR N  
  \item AMYGDL N  
  \item AMYGDR N  Amygdala_R
  \item CALCARINEL N  Calcarine_L
  \item CALCARINER N  Calcarine_R
  \item CUNEUSL N  
  \item CUNEUSR N  
  \item LINGUALL N  
  \item LINGUALR N  
  \item OCCSUPL N  
  \item OCCSUPR N  
  \item OCCMIDL N  
  \item OCCMIDR N  
  \item OCCINFL N  
  \item OCCINFR N  
  \item FUSIFORML N  
  \item FUSIFORMR N  
  \item POSTCENTL N  
  \item POSTCENTR N  
  \item PARIETSUPL N  
  \item PARIETSUPR N  
  \item PARIETINFL N  
  \item PARIETINFR N  
  \item SUPRAMARGL N  
  \item SUPRAMARGR N  
  \item ANGULARL N  
  \item ANGULARR N  
  \item PRECUNEUSL N  
  \item PRECUNEUSR N  
  \item PARCENTLBL N  
  \item PARCENTLBR N  
  \item CAUDATEL N  
  \item CAUDATER N  
  \item PUTAMENL N  
  \item PUTAMENR N  
  \item PALLIDUML N  
  \item PALLIDUMR N  
  \item THALAMUSL N  
  \item THALAMUSR N  
  \item HESCHLL N  
  \item HESCHLR N  
  \item TEMPSUPL N  
  \item TEMPSUPR N  
  \item TEMPPLSUPL N  
  \item TEMPPLSUPR N  
  \item TEMPMIDL N  
  \item TEMPMIDR N  
  \item TEMPPLMIDL N  
  \item TEMPPLMIDR N  
  \item TEMPINFL N  
  \item TEMPINFR N  
  \item CEREBCR1L N  
  \item CEREBCR1R N  
  \item CEREBCR2L N  
  \item CEREBCR2R N  
  \item CEREB3L N  
  \item CEREB3R N  
  \item CEREB45L N  
  \item CEREB45R N  
  \item CEREB6L N  
  \item CEREB6R N  
  \item CEREB7BL N  
  \item CEREB7BR N  
  \item CEREB8L N  
  \item CEREB8R N  
  \item CEREB9L N  
  \item CEREB9R N  
  \item CEREB10L N  
  \item CEREB10R N  
  \item VERMIS12 N  
  \item VERMIS3 N  
  \item VERMIS45 N  
  \item VERMIS6 N  
  \item VERMIS7 N  
  \item VERMIS8 N  
  \item VERMIS9 N  
  \item VERMIS10 N  
  \item ETIV N  
}
}
\keyword{datasets}
