% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{ucberkeleyav1451_pvc}
\alias{ucberkeleyav1451_pvc}
\title{UC Berkeley - AV1451 PVC analysis}
\format{
A data frame with 1450 rows and 170 variables
}
\usage{
data(ucberkeleyav1451_pvc)
}
\description{
UC Berkeley - AV1451 PVC analysis
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID T  Participant roster ID
  \item VISCODE T  Visit code
  \item VISCODE2 T  Translated visit code
  \item EXAMDATE D  Examination Date
  \item INFERIOR_CEREBGM_SUVR N av1451 uptake - reference region Reference region - av1451 mean of inferior cerebellar grey matter
  \item INFERIOR_CEREBGM_VOLUME N mm^3 Inferior Cerebellar Grey Matter ROI size in mm^3
  \item HEMIWM_SUVR N av1451 uptake - reference region Reference region - av1451 mean of hemispheric white matter
  \item HEMIWM_VOLUME N mm^3 Hemispheric white matter ROI size in mm^3
  \item BRAAK1_SUVR N av1451 uptake - composite Weighted av1451 mean of Braak region 1
  \item BRAAK1_VOLUME N mm^3 Braak region 1 size in mm^3
  \item BRAAK34_SUVR N av1451 uptake - composite Weighted av1451 mean of Braak region 3 and 4 
  \item BRAAK34_VOLUME N mm^3 Braak region 3+4 size in mm^3
  \item META_TEMPORAL_SUVR N av1451 uptake - subregion Weighted av1451 mean of meta-temporal region
  \item META_TEMPORAL_VOLUME N mm^3 Meta-temporal size in mm^3
  \item BRAAK56_SUVR N av1451 uptake - composite Weighted av1451 mean of Braak region 5 and 6
  \item BRAAK56_VOLUME N mm^3 Braak region 5+6 size in mm^3
  \item BRAAK2_SUVR N av1451 uptake - subregion Weighted av1451 mean of Braak region 2
  \item BRAAK2_VOLUME N mm^3 Braak region 2 size in mm^3
  \item AMYGDALA_SUVR N av1451 uptake - subregion Weighted av1451 mean of L/R amygdala
  \item AMYGDALA_VOLUME N mm^3 Amygdala size in mm^3
  \item BRAIN_STEM_SUVR N av1451 uptake - subregion Brain stem ROI
  \item BRAIN_STEM_VOLUME N mm^3 Brain stem ROI size in mm^3
  \item LEFT_MIDDLEFR_SUVR N av1451 uptake - subregion Left middle frontal ROI
  \item LEFT_MIDDLEFR_VOLUME N mm^3 Left middle frontal ROI size in mm^3
  \item LEFT_ORBITOFR_SUVR N av1451 uptake - subregion Left orbitofrontal ROI
  \item LEFT_ORBITOFR_VOLUME N mm^3 Left orbitofrontal ROI size in mm^3
  \item LEFT_PARSFR_SUVR N av1451 uptake - subregion Left pars frontal ROI
  \item LEFT_PARSFR_VOLUME N mm^3 Left pars frontal ROI size in mm^3
  \item LEFT_ACCUMBENS_AREA_SUVR N av1451 uptake - subregion left-accumbens-area
  \item LEFT_ACCUMBENS_AREA_VOLUME N mm^3 left-accumbens-area ROI size in mm^3
  \item LEFT_AMYGDALA_SUVR N av1451 uptake - subregion left-amygdala
  \item LEFT_AMYGDALA_VOLUME N mm^3 left-amygdala ROI size in mm^3
  \item LEFT_CAUDATE_SUVR N av1451 uptake - subregion left-caudate
  \item LEFT_CAUDATE_VOLUME N mm^3 left-caudate ROI size in mm^3
  \item LEFT_HIPPOCAMPUS_SUVR N av1451 uptake - subregion left-hippocampus
  \item LEFT_HIPPOCAMPUS_VOLUME N mm^3 left-hippocampus ROI size in mm^3
  \item LEFT_PALLIDUM_SUVR N av1451 uptake - subregion left-pallidum
  \item LEFT_PALLIDUM_VOLUME N mm^3 left-pallidum ROI size in mm^3
  \item LEFT_PUTAMEN_SUVR N av1451 uptake - subregion left-putamen
  \item LEFT_PUTAMEN_VOLUME N mm^3 left-putamen ROI size in mm^3
  \item LEFT_THALAMUS_PROPER_SUVR N av1451 uptake - subregion left-thalamus-proper
  \item LEFT_THALAMUS_PROPER_VOLUME N mm^3 left-thalamus-proper ROI size in mm^3
  \item RIGHT_MIDDLEFR_SUVR N av1451 uptake - subregion Right middle frontal ROI
  \item RIGHT_MIDDLEFR_VOLUME N mm^3 Right middle frontal ROI size in mm^3
  \item RIGHT_ORBITOFR_SUVR N av1451 uptake - subregion Right orbitofrontal ROI
  \item RIGHT_ORBITOFR_VOLUME N mm^3 Right orbitofrontal ROI size in mm^3
  \item RIGHT_PARSFR_SUVR N av1451 uptake - subregion Right pars frontal ROI
  \item RIGHT_PARSFR_VOLUME N mm^3 Right pars frontal ROI size in mm^3
  \item RIGHT_ACCUMBENS_AREA_SUVR N av1451 uptake - subregion right-accumbens-area
  \item RIGHT_ACCUMBENS_AREA_VOLUME N mm^3 right-accumbens-area ROI size in mm^3
  \item RIGHT_AMYGDALA_SUVR N av1451 uptake - subregion right-amygdala
  \item RIGHT_AMYGDALA_VOLUME N mm^3 right-amygdala ROI size in mm^3
  \item RIGHT_CAUDATE_SUVR N av1451 uptake - subregion right-caudate
  \item RIGHT_CAUDATE_VOLUME N mm^3 right-caudate ROI size in mm^3
  \item RIGHT_HIPPOCAMPUS_SUVR N av1451 uptake - subregion right-hippocampus
  \item RIGHT_HIPPOCAMPUS_VOLUME N mm^3 right-hippocampus ROI size in mm^3
  \item RIGHT_PALLIDUM_SUVR N av1451 uptake - subregion right-pallidum
  \item RIGHT_PALLIDUM_VOLUME N mm^3 right-pallidum ROI size in mm^3
  \item RIGHT_PUTAMEN_SUVR N av1451 uptake - subregion right-putamen
  \item RIGHT_PUTAMEN_VOLUME N mm^3 right-putamen ROI size in mm^3
  \item RIGHT_THALAMUS_PROPER_SUVR N av1451 uptake - subregion right-thalamus-proper
  \item RIGHT_THALAMUS_PROPER_VOLUME N mm^3 right-thalamus-proper ROI size in mm^3
  \item CHOROID_SUVR N av1451 uptake - subregion Bilateral choroid plexus
  \item CHOROID_VOLUME N mm^3 Bilateral choroid plexus ROI size in mm^3
  \item CTX_LH_BANKSSTS_SUVR N av1451 uptake - subregion ctx-lh-bankssts
  \item CTX_LH_BANKSSTS_VOLUME N mm^3 ctx-lh-bankssts ROI size in mm^3
  \item CTX_LH_CAUDALANTERIORCINGULATE_SUVR N av1451 uptake - subregion ctx-lh-caudalanteriorcingulate
  \item CTX_LH_CAUDALANTERIORCINGULATE_VOLUME N mm^3 ctx-lh-caudalanteriorcingulate ROI size in mm^3
  \item CTX_LH_CUNEUS_SUVR N av1451 uptake - subregion ctx-lh-cuneus
  \item CTX_LH_CUNEUS_VOLUME N mm^3 ctx-lh-cuneus ROI size in mm^3
  \item CTX_LH_ENTORHINAL_SUVR N av1451 uptake - subregion ctx-lh-entorhinal
  \item CTX_LH_ENTORHINAL_VOLUME N mm^3 ctx-lh-entorhinal ROI size in mm^3
  \item CTX_LH_FUSIFORM_SUVR N av1451 uptake - subregion ctx-lh-fusiform
  \item CTX_LH_FUSIFORM_VOLUME N mm^3 ctx-lh-fusiform ROI size in mm^3
  \item CTX_LH_INFERIORPARIETAL_SUVR N av1451 uptake - subregion ctx-lh-inferiorparietal
  \item CTX_LH_INFERIORPARIETAL_VOLUME N mm^3 ctx-lh-inferiorparietal ROI size in mm^3
  \item CTX_LH_INFERIORTEMPORAL_SUVR N av1451 uptake - subregion ctx-lh-inferiortemporal
  \item CTX_LH_INFERIORTEMPORAL_VOLUME N mm^3 ctx-lh-inferiortemporal ROI size in mm^3
  \item CTX_LH_INSULA_SUVR N av1451 uptake - subregion ctx-lh-insula
  \item CTX_LH_INSULA_VOLUME N mm^3 ctx-lh-insula ROI size in mm^3
  \item CTX_LH_ISTHMUSCINGULATE_SUVR N av1451 uptake - subregion ctx-lh-isthmuscingulate
  \item CTX_LH_ISTHMUSCINGULATE_VOLUME N mm^3 ctx-lh-isthmuscingulate ROI size in mm^3
  \item CTX_LH_LATERALOCCIPITAL_SUVR N av1451 uptake - subregion ctx-lh-lateraloccipital
  \item CTX_LH_LATERALOCCIPITAL_VOLUME N mm^3 ctx-lh-lateraloccipital ROI size in mm^3
  \item CTX_LH_LINGUAL_SUVR N av1451 uptake - subregion ctx-lh-lingual
  \item CTX_LH_LINGUAL_VOLUME N mm^3 ctx-lh-lingual ROI size in mm^3
  \item CTX_LH_MIDDLETEMPORAL_SUVR N av1451 uptake - subregion ctx-lh-middletemporal
  \item CTX_LH_MIDDLETEMPORAL_VOLUME N mm^3 ctx-lh-middletemporal ROI size in mm^3
  \item CTX_LH_PARACENTRAL_SUVR N av1451 uptake - subregion ctx-lh-paracentral
  \item CTX_LH_PARACENTRAL_VOLUME N mm^3 ctx-lh-paracentral ROI size in mm^3
  \item CTX_LH_PARAHIPPOCAMPAL_SUVR N av1451 uptake - subregion ctx-lh-parahippocampal
  \item CTX_LH_PARAHIPPOCAMPAL_VOLUME N mm^3 ctx-lh-parahippocampal ROI size in mm^3
  \item CTX_LH_PERICALCARINE_SUVR N av1451 uptake - subregion ctx-lh-pericalcarine
  \item CTX_LH_PERICALCARINE_VOLUME N mm^3 ctx-lh-pericalcarine ROI size in mm^3
  \item CTX_LH_POSTCENTRAL_SUVR N av1451 uptake - subregion ctx-lh-postcentral
  \item CTX_LH_POSTCENTRAL_VOLUME N mm^3 ctx-lh-postcentral ROI size in mm^3
  \item CTX_LH_POSTERIORCINGULATE_SUVR N av1451 uptake - subregion ctx-lh-posteriorcingulate
  \item CTX_LH_POSTERIORCINGULATE_VOLUME N mm^3 ctx-lh-posteriorcingulate ROI size in mm^3
  \item CTX_LH_PRECENTRAL_SUVR N av1451 uptake - subregion ctx-lh-precentral
  \item CTX_LH_PRECENTRAL_VOLUME N mm^3 ctx-lh-precentral ROI size in mm^3
  \item CTX_LH_PRECUNEUS_SUVR N av1451 uptake - subregion ctx-lh-precuneus
  \item CTX_LH_PRECUNEUS_VOLUME N mm^3 ctx-lh-precuneus ROI size in mm^3
  \item CTX_LH_ROSTRALANTERIORCINGULATE_SUVR N av1451 uptake - subregion ctx-lh-rostralanteriorcingulate
  \item CTX_LH_ROSTRALANTERIORCINGULATE_VOLUME N mm^3 ctx-lh-rostralanteriorcingulate ROI size in mm^3
  \item CTX_LH_SUPERIORFRONTAL_SUVR N av1451 uptake - subregion ctx-lh-superiorfrontal
  \item CTX_LH_SUPERIORFRONTAL_VOLUME N mm^3 ctx-lh-superiorfrontal ROI size in mm^3
  \item CTX_LH_SUPERIORPARIETAL_SUVR N av1451 uptake - subregion ctx-lh-superiorparietal
  \item CTX_LH_SUPERIORPARIETAL_VOLUME N mm^3 ctx-lh-superiorparietal ROI size in mm^3
  \item CTX_LH_SUPERIORTEMPORAL_SUVR N av1451 uptake - subregion ctx-lh-superiortemporal
  \item CTX_LH_SUPERIORTEMPORAL_VOLUME N mm^3 ctx-lh-superiortemporal ROI size in mm^3
  \item CTX_LH_SUPRAMARGINAL_SUVR N av1451 uptake - subregion ctx-lh-supramarginal
  \item CTX_LH_SUPRAMARGINAL_VOLUME N mm^3 ctx-lh-supramarginal ROI size in mm^3
  \item CTX_LH_TEMPORALPOLE_SUVR N av1451 uptake - subregion ctx-lh-temporalpole
  \item CTX_LH_TEMPORALPOLE_VOLUME N mm^3 ctx-lh-temporalpole ROI size in mm^3
  \item CTX_LH_TRANSVERSETEMPORAL_SUVR N av1451 uptake - subregion ctx-lh-transversetemporal
  \item CTX_LH_TRANSVERSETEMPORAL_VOLUME N mm^3 ctx-lh-transversetemporal ROI size in mm^3
  \item CTX_RH_BANKSSTS_SUVR N av1451 uptake - subregion ctx-rh-bankssts
  \item CTX_RH_BANKSSTS_VOLUME N mm^3 ctx-rh-bankssts ROI size in mm^3
  \item CTX_RH_CAUDALANTERIORCINGULATE_SUVR N av1451 uptake - subregion ctx-rh-caudalanteriorcingulate
  \item CTX_RH_CAUDALANTERIORCINGULATE_VOLUME N mm^3 ctx-rh-caudalanteriorcingulate ROI size in mm^3
  \item CTX_RH_CUNEUS_SUVR N av1451 uptake - subregion ctx-rh-cuneus
  \item CTX_RH_CUNEUS_VOLUME N mm^3 ctx-rh-cuneus ROI size in mm^3
  \item CTX_RH_ENTORHINAL_SUVR N av1451 uptake - subregion ctx-rh-entorhinal
  \item CTX_RH_ENTORHINAL_VOLUME N mm^3 ctx-rh-entorhinal ROI size in mm^3
  \item CTX_RH_FUSIFORM_SUVR N av1451 uptake - subregion ctx-rh-fusiform
  \item CTX_RH_FUSIFORM_VOLUME N mm^3 ctx-rh-fusiform ROI size in mm^3
  \item CTX_RH_INFERIORPARIETAL_SUVR N av1451 uptake - subregion ctx-rh-inferiorparietal
  \item CTX_RH_INFERIORPARIETAL_VOLUME N mm^3 ctx-rh-inferiorparietal ROI size in mm^3
  \item CTX_RH_INFERIORTEMPORAL_SUVR N av1451 uptake - subregion ctx-rh-inferiortemporal
  \item CTX_RH_INFERIORTEMPORAL_VOLUME N mm^3 ctx-rh-inferiortemporal ROI size in mm^3
  \item CTX_RH_INSULA_SUVR N av1451 uptake - subregion ctx-rh-insula
  \item CTX_RH_INSULA_VOLUME N mm^3 ctx-rh-insula ROI size in mm^3
  \item CTX_RH_ISTHMUSCINGULATE_SUVR N av1451 uptake - subregion ctx-rh-isthmuscingulate
  \item CTX_RH_ISTHMUSCINGULATE_VOLUME N mm^3 ctx-rh-isthmuscingulate ROI size in mm^3
  \item CTX_RH_LATERALOCCIPITAL_SUVR N av1451 uptake - subregion ctx-rh-lateraloccipital
  \item CTX_RH_LATERALOCCIPITAL_VOLUME N mm^3 ctx-rh-lateraloccipital ROI size in mm^3
  \item CTX_RH_LINGUAL_SUVR N av1451 uptake - subregion ctx-rh-lingual
  \item CTX_RH_LINGUAL_VOLUME N mm^3 ctx-rh-lingual ROI size in mm^3
  \item CTX_RH_MIDDLETEMPORAL_SUVR N av1451 uptake - subregion ctx-rh-middletemporal
  \item CTX_RH_MIDDLETEMPORAL_VOLUME N mm^3 ctx-rh-middletemporal ROI size in mm^3
  \item CTX_RH_PARACENTRAL_SUVR N av1451 uptake - subregion ctx-rh-paracentral
  \item CTX_RH_PARACENTRAL_VOLUME N mm^3 ctx-rh-paracentral ROI size in mm^3
  \item CTX_RH_PARAHIPPOCAMPAL_SUVR N av1451 uptake - subregion ctx-rh-parahippocampal
  \item CTX_RH_PARAHIPPOCAMPAL_VOLUME N mm^3 ctx-rh-parahippocampal ROI size in mm^3
  \item CTX_RH_PERICALCARINE_SUVR N av1451 uptake - subregion ctx-rh-pericalcarine
  \item CTX_RH_PERICALCARINE_VOLUME N mm^3 ctx-rh-pericalcarine ROI size in mm^3
  \item CTX_RH_POSTCENTRAL_SUVR N av1451 uptake - subregion ctx-rh-postcentral
  \item CTX_RH_POSTCENTRAL_VOLUME N mm^3 ctx-rh-postcentral ROI size in mm^3
  \item CTX_RH_POSTERIORCINGULATE_SUVR N av1451 uptake - subregion ctx-rh-posteriorcingulate
  \item CTX_RH_POSTERIORCINGULATE_VOLUME N mm^3 ctx-rh-posteriorcingulate ROI size in mm^3
  \item CTX_RH_PRECENTRAL_SUVR N av1451 uptake - subregion ctx-rh-precentral
  \item CTX_RH_PRECENTRAL_VOLUME N mm^3 ctx-rh-precentral ROI size in mm^3
  \item CTX_RH_PRECUNEUS_SUVR N av1451 uptake - subregion ctx-rh-precuneus
  \item CTX_RH_PRECUNEUS_VOLUME N mm^3 ctx-rh-precuneus ROI size in mm^3
  \item CTX_RH_ROSTRALANTERIORCINGULATE_SUVR N av1451 uptake - subregion ctx-rh-rostralanteriorcingulate
  \item CTX_RH_ROSTRALANTERIORCINGULATE_VOLUME N mm^3 ctx-rh-rostralanteriorcingulate ROI size in mm^3
  \item CTX_RH_SUPERIORFRONTAL_SUVR N av1451 uptake - subregion ctx-rh-superiorfrontal
  \item CTX_RH_SUPERIORFRONTAL_VOLUME N mm^3 ctx-rh-superiorfrontal ROI size in mm^3
  \item CTX_RH_SUPERIORPARIETAL_SUVR N av1451 uptake - subregion ctx-rh-superiorparietal
  \item CTX_RH_SUPERIORPARIETAL_VOLUME N mm^3 ctx-rh-superiorparietal ROI size in mm^3
  \item CTX_RH_SUPERIORTEMPORAL_SUVR N av1451 uptake - subregion ctx-rh-superiortemporal
  \item CTX_RH_SUPERIORTEMPORAL_VOLUME N mm^3 ctx-rh-superiortemporal ROI size in mm^3
  \item CTX_RH_SUPRAMARGINAL_SUVR N av1451 uptake - subregion ctx-rh-supramarginal
  \item CTX_RH_SUPRAMARGINAL_VOLUME N mm^3 ctx-rh-supramarginal ROI size in mm^3
  \item CTX_RH_TEMPORALPOLE_SUVR N av1451 uptake - subregion ctx-rh-temporalpole
  \item CTX_RH_TEMPORALPOLE_VOLUME N mm^3 ctx-rh-temporalpole ROI size in mm^3
  \item CTX_RH_TRANSVERSETEMPORAL_SUVR N av1451 uptake - subregion ctx-rh-transversetemporal
  \item CTX_RH_TRANSVERSETEMPORAL_VOLUME N mm^3 ctx-rh-transversetemporal ROI size in mm^3
  \item OTHER_SUVR N av1451 uptake - subregion Other regions: Left-VentralDC    Left-vessel     Right-VentralDC     Right-vessel     WM-hypointensities     non-WM-hypointensities     Optic-Chiasm     ctx-lh-unknown     ctx-lh-corpuscallosum    ctx-rh-unknown  ctx-rh-corpuscallosum
  \item OTHER_VOLUME N mm^3 Other regions size in mm^3
  \item BRAAK1 N av1451 uptake - reference region Braak region 1
  \item BRAAK1_SIZE N mm^3 Braak region 1 ROI size in mm^3
  \item BRAAK2 N av1451 uptake - reference region Braak region 2
  \item BRAAK2_SIZE N mm^3 Braak region 2 ROI size in mm^3
  \item BRAAK3 N av1451 uptake - subregion Braak region 3
  \item BRAAK3_SIZE N mm^3 Braak region 3 ROI size in mm^3
  \item BRAAK4 N av1451 uptake - subregion Braak region 4
  \item BRAAK4_SIZE N mm^3 Braak region 4 ROI size in mm^3
  \item BRAAK5 N av1451 uptake - subregion Braak region 5
  \item BRAAK5_SIZE N mm^3 Braak region 5 ROI size in mm^3
  \item BRAAK6 N av1451 uptake - subregion Braak region 6
  \item BRAAK6_SIZE N mm^3 Braak region 6 ROI size in mm^3
  \item CEREBGM N av1451 uptake - subregion Cerebellar grey matter
  \item CEREBGM_SIZE N mm^3 Cerebellar grey matter ROI size in mm^3
  \item CEREBWM N av1451 uptake - subregion Cerebellar white matter
  \item CEREBWM_SIZE N mm^3 Cerebellar white matter ROI size in mm^3
  \item BRAIN_STEM N av1451 uptake - subregion Brain stem 
  \item BRAIN_STEM_SIZE N mm^3 Brain stem ROI size in mm^3
  \item HEMIWM N av1451 uptake - subregion Hemispheric white matter
  \item HEMIWM_SIZE N mm^3 Hemispheric white matter ROI size in mm^3
  \item BONE N av1451 uptake - subregion Bone
  \item BONE_SIZE N mm^3 Bone ROI size in mm^3
  \item CSF_SA N av1451 uptake - subregion CSF 
  \item CSF_SA_SIZE N mm^3 CSF ROI size in mm^3
  \item SOFT_TISSUE N av1451 uptake - subregion Soft tissue
  \item SOFT_TISSUE_SIZE N mm^3 Soft tissue ROI size in mm^3
  \item LEFT_MIDDLEFR N av1451 uptake - subregion Left middle frontal ROI
  \item LEFT_MIDDLEFR_SIZE N mm^3 Left middle frontal ROI size in mm^3
  \item LEFT_ORBITOFR N av1451 uptake - subregion Left orbitofrontal ROI
  \item LEFT_ORBITOFR_SIZE N mm^3 Left orbitofrontal ROI size in mm^3
  \item LEFT_PARSFR N av1451 uptake - subregion Left pars frontal ROI
  \item LEFT_PARSFR_SIZE N mm^3 Left pars frontal ROI size in mm^3
  \item LEFT_ACCUMBENS_AREA N av1451 uptake - subregion left-accumbens-area
  \item LEFT_ACCUMBENS_AREA_SIZE N mm^3 left-accumbens-area ROI size in mm^3
  \item LEFT_AMYGDALA N av1451 uptake - subregion left-amygdala
  \item LEFT_AMYGDALA_SIZE N mm^3 left-amygdala ROI size in mm^3
  \item LEFT_CAUDATE N av1451 uptake - subregion left-caudate
  \item LEFT_CAUDATE_SIZE N mm^3 left-caudate ROI size in mm^3
  \item LEFT_HIPPOCAMPUS N av1451 uptake - subregion left-hippocampus
  \item LEFT_HIPPOCAMPUS_SIZE N mm^3 left-hippocampus ROI size in mm^3
  \item LEFT_PALLIDUM N av1451 uptake - subregion left-pallidum
  \item LEFT_PALLIDUM_SIZE N mm^3 left-pallidum ROI size in mm^3
  \item LEFT_PUTAMEN N av1451 uptake - subregion left-putamen
  \item LEFT_PUTAMEN_SIZE N mm^3 left-putamen ROI size in mm^3
  \item LEFT_THALAMUS_PROPER N av1451 uptake - subregion left-thalamus-proper
  \item LEFT_THALAMUS_PROPER_SIZE N mm^3 left-thalamus-proper ROI size in mm^3
  \item RIGHT_MIDDLEFR N av1451 uptake - subregion Right middle frontal ROI
  \item RIGHT_MIDDLEFR_SIZE N mm^3 Right middle frontal ROI size in mm^3
  \item RIGHT_ORBITOFR N av1451 uptake - subregion Right orbitofrontal ROI
  \item RIGHT_ORBITOFR_SIZE N mm^3 Right orbitofrontal ROI size in mm^3
  \item RIGHT_PARSFR N av1451 uptake - subregion Right pars frontal ROI
  \item RIGHT_PARSFR_SIZE N mm^3 Right pars frontal ROI size in mm^3
  \item RIGHT_ACCUMBENS_AREA N av1451 uptake - subregion right-accumbens-area
  \item RIGHT_ACCUMBENS_AREA_SIZE N mm^3 right-accumbens-area ROI size in mm^3
  \item RIGHT_AMYGDALA N av1451 uptake - subregion right-amygdala
  \item RIGHT_AMYGDALA_SIZE N mm^3 right-amygdala ROI size in mm^3
  \item RIGHT_CAUDATE N av1451 uptake - subregion right-caudate
  \item RIGHT_CAUDATE_SIZE N mm^3 right-caudate ROI size in mm^3
  \item RIGHT_HIPPOCAMPUS N av1451 uptake - subregion right-hippocampus
  \item RIGHT_HIPPOCAMPUS_SIZE N mm^3 right-hippocampus ROI size in mm^3
  \item RIGHT_PALLIDUM N av1451 uptake - subregion right-pallidum
  \item RIGHT_PALLIDUM_SIZE N mm^3 right-pallidum ROI size in mm^3
  \item RIGHT_PUTAMEN N av1451 uptake - subregion right-putamen
  \item RIGHT_PUTAMEN_SIZE N mm^3 right-putamen ROI size in mm^3
  \item RIGHT_THALAMUS_PROPER N av1451 uptake - subregion right-thalamus-proper
  \item RIGHT_THALAMUS_PROPER_SIZE N mm^3 right-thalamus-proper ROI size in mm^3
  \item CHOROID N av1451 uptake - subregion Bilateral choroid plexus
  \item CHOROID_SIZE N mm^3 Bilateral choroid plexus ROI size in mm^3
  \item CTX_LH_BANKSSTS N av1451 uptake - subregion ctx-lh-bankssts
  \item CTX_LH_BANKSSTS_SIZE N mm^3 ctx-lh-bankssts ROI size in mm^3
  \item CTX_LH_CAUDALANTERIORCINGULATE N av1451 uptake - subregion ctx-lh-caudalanteriorcingulate
  \item CTX_LH_CAUDALANTERIORCINGULATE_SIZE N mm^3 ctx-lh-caudalanteriorcingulate ROI size in mm^3
  \item CTX_LH_CUNEUS N av1451 uptake - subregion ctx-lh-cuneus
  \item CTX_LH_CUNEUS_SIZE N mm^3 ctx-lh-cuneus ROI size in mm^3
  \item CTX_LH_ENTORHINAL N av1451 uptake - subregion ctx-lh-entorhinal
  \item CTX_LH_ENTORHINAL_SIZE N mm^3 ctx-lh-entorhinal ROI size in mm^3
  \item CTX_LH_FUSIFORM N av1451 uptake - subregion ctx-lh-fusiform
  \item CTX_LH_FUSIFORM_SIZE N mm^3 ctx-lh-fusiform ROI size in mm^3
  \item CTX_LH_INFERIORPARIETAL N av1451 uptake - subregion ctx-lh-inferiorparietal
  \item CTX_LH_INFERIORPARIETAL_SIZE N mm^3 ctx-lh-inferiorparietal ROI size in mm^3
  \item CTX_LH_INFERIORTEMPORAL N av1451 uptake - subregion ctx-lh-inferiortemporal
  \item CTX_LH_INFERIORTEMPORAL_SIZE N mm^3 ctx-lh-inferiortemporal ROI size in mm^3
  \item CTX_LH_INSULA N av1451 uptake - subregion ctx-lh-insula
  \item CTX_LH_INSULA_SIZE N mm^3 ctx-lh-insula ROI size in mm^3
  \item CTX_LH_ISTHMUSCINGULATE N av1451 uptake - subregion ctx-lh-isthmuscingulate
  \item CTX_LH_ISTHMUSCINGULATE_SIZE N mm^3 ctx-lh-isthmuscingulate ROI size in mm^3
  \item CTX_LH_LATERALOCCIPITAL N av1451 uptake - subregion ctx-lh-lateraloccipital
  \item CTX_LH_LATERALOCCIPITAL_SIZE N mm^3 ctx-lh-lateraloccipital ROI size in mm^3
  \item CTX_LH_LINGUAL N av1451 uptake - subregion ctx-lh-lingual
  \item CTX_LH_LINGUAL_SIZE N mm^3 ctx-lh-lingual ROI size in mm^3
  \item CTX_LH_MIDDLETEMPORAL N av1451 uptake - subregion ctx-lh-middletemporal
  \item CTX_LH_MIDDLETEMPORAL_SIZE N mm^3 ctx-lh-middletemporal ROI size in mm^3
  \item CTX_LH_PARACENTRAL N av1451 uptake - subregion ctx-lh-paracentral
  \item CTX_LH_PARACENTRAL_SIZE N mm^3 ctx-lh-paracentral ROI size in mm^3
  \item CTX_LH_PARAHIPPOCAMPAL N av1451 uptake - subregion ctx-lh-parahippocampal
  \item CTX_LH_PARAHIPPOCAMPAL_SIZE N mm^3 ctx-lh-parahippocampal ROI size in mm^3
  \item CTX_LH_PERICALCARINE N av1451 uptake - subregion ctx-lh-pericalcarine
  \item CTX_LH_PERICALCARINE_SIZE N mm^3 ctx-lh-pericalcarine ROI size in mm^3
  \item CTX_LH_POSTCENTRAL N av1451 uptake - subregion ctx-lh-postcentral
  \item CTX_LH_POSTCENTRAL_SIZE N mm^3 ctx-lh-postcentral ROI size in mm^3
  \item CTX_LH_POSTERIORCINGULATE N av1451 uptake - subregion ctx-lh-posteriorcingulate
  \item CTX_LH_POSTERIORCINGULATE_SIZE N mm^3 ctx-lh-posteriorcingulate ROI size in mm^3
  \item CTX_LH_PRECENTRAL N av1451 uptake - subregion ctx-lh-precentral
  \item CTX_LH_PRECENTRAL_SIZE N mm^3 ctx-lh-precentral ROI size in mm^3
  \item CTX_LH_PRECUNEUS N av1451 uptake - subregion ctx-lh-precuneus
  \item CTX_LH_PRECUNEUS_SIZE N mm^3 ctx-lh-precuneus ROI size in mm^3
  \item CTX_LH_ROSTRALANTERIORCINGULATE N av1451 uptake - subregion ctx-lh-rostralanteriorcingulate
  \item CTX_LH_ROSTRALANTERIORCINGULATE_SIZE N mm^3 ctx-lh-rostralanteriorcingulate ROI size in mm^3
  \item CTX_LH_SUPERIORFRONTAL N av1451 uptake - subregion ctx-lh-superiorfrontal
  \item CTX_LH_SUPERIORFRONTAL_SIZE N mm^3 ctx-lh-superiorfrontal ROI size in mm^3
  \item CTX_LH_SUPERIORPARIETAL N av1451 uptake - subregion ctx-lh-superiorparietal
  \item CTX_LH_SUPERIORPARIETAL_SIZE N mm^3 ctx-lh-superiorparietal ROI size in mm^3
  \item CTX_LH_SUPERIORTEMPORAL N av1451 uptake - subregion ctx-lh-superiortemporal
  \item CTX_LH_SUPERIORTEMPORAL_SIZE N mm^3 ctx-lh-superiortemporal ROI size in mm^3
  \item CTX_LH_SUPRAMARGINAL N av1451 uptake - subregion ctx-lh-supramarginal
  \item CTX_LH_SUPRAMARGINAL_SIZE N mm^3 ctx-lh-supramarginal ROI size in mm^3
  \item CTX_LH_TEMPORALPOLE N av1451 uptake - subregion ctx-lh-temporalpole
  \item CTX_LH_TEMPORALPOLE_SIZE N mm^3 ctx-lh-temporalpole ROI size in mm^3
  \item CTX_LH_TRANSVERSETEMPORAL N av1451 uptake - subregion ctx-lh-transversetemporal
  \item CTX_LH_TRANSVERSETEMPORAL_SIZE N mm^3 ctx-lh-transversetemporal ROI size in mm^3
  \item CTX_RH_BANKSSTS N av1451 uptake - subregion ctx-rh-bankssts
  \item CTX_RH_BANKSSTS_SIZE N mm^3 ctx-rh-bankssts ROI size in mm^3
  \item CTX_RH_CAUDALANTERIORCINGULATE N av1451 uptake - subregion ctx-rh-caudalanteriorcingulate
  \item CTX_RH_CAUDALANTERIORCINGULATE_SIZE N mm^3 ctx-rh-caudalanteriorcingulate ROI size in mm^3
  \item CTX_RH_CUNEUS N av1451 uptake - subregion ctx-rh-cuneus
  \item CTX_RH_CUNEUS_SIZE N mm^3 ctx-rh-cuneus ROI size in mm^3
  \item CTX_RH_ENTORHINAL N av1451 uptake - subregion ctx-rh-entorhinal
  \item CTX_RH_ENTORHINAL_SIZE N mm^3 ctx-rh-entorhinal ROI size in mm^3
  \item CTX_RH_FUSIFORM N av1451 uptake - subregion ctx-rh-fusiform
  \item CTX_RH_FUSIFORM_SIZE N mm^3 ctx-rh-fusiform ROI size in mm^3
  \item CTX_RH_INFERIORPARIETAL N av1451 uptake - subregion ctx-rh-inferiorparietal
  \item CTX_RH_INFERIORPARIETAL_SIZE N mm^3 ctx-rh-inferiorparietal ROI size in mm^3
  \item CTX_RH_INFERIORTEMPORAL N av1451 uptake - subregion ctx-rh-inferiortemporal
  \item CTX_RH_INFERIORTEMPORAL_SIZE N mm^3 ctx-rh-inferiortemporal ROI size in mm^3
  \item CTX_RH_INSULA N av1451 uptake - subregion ctx-rh-insula
  \item CTX_RH_INSULA_SIZE N mm^3 ctx-rh-insula ROI size in mm^3
  \item CTX_RH_ISTHMUSCINGULATE N av1451 uptake - subregion ctx-rh-isthmuscingulate
  \item CTX_RH_ISTHMUSCINGULATE_SIZE N mm^3 ctx-rh-isthmuscingulate ROI size in mm^3
  \item CTX_RH_LATERALOCCIPITAL N av1451 uptake - subregion ctx-rh-lateraloccipital
  \item CTX_RH_LATERALOCCIPITAL_SIZE N mm^3 ctx-rh-lateraloccipital ROI size in mm^3
  \item CTX_RH_LINGUAL N av1451 uptake - subregion ctx-rh-lingual
  \item CTX_RH_LINGUAL_SIZE N mm^3 ctx-rh-lingual ROI size in mm^3
  \item CTX_RH_MIDDLETEMPORAL N av1451 uptake - subregion ctx-rh-middletemporal
  \item CTX_RH_MIDDLETEMPORAL_SIZE N mm^3 ctx-rh-middletemporal ROI size in mm^3
  \item CTX_RH_PARACENTRAL N av1451 uptake - subregion ctx-rh-paracentral
  \item CTX_RH_PARACENTRAL_SIZE N mm^3 ctx-rh-paracentral ROI size in mm^3
  \item CTX_RH_PARAHIPPOCAMPAL N av1451 uptake - subregion ctx-rh-parahippocampal
  \item CTX_RH_PARAHIPPOCAMPAL_SIZE N mm^3 ctx-rh-parahippocampal ROI size in mm^3
  \item CTX_RH_PERICALCARINE N av1451 uptake - subregion ctx-rh-pericalcarine
  \item CTX_RH_PERICALCARINE_SIZE N mm^3 ctx-rh-pericalcarine ROI size in mm^3
  \item CTX_RH_POSTCENTRAL N av1451 uptake - subregion ctx-rh-postcentral
  \item CTX_RH_POSTCENTRAL_SIZE N mm^3 ctx-rh-postcentral ROI size in mm^3
  \item CTX_RH_POSTERIORCINGULATE N av1451 uptake - subregion ctx-rh-posteriorcingulate
  \item CTX_RH_POSTERIORCINGULATE_SIZE N mm^3 ctx-rh-posteriorcingulate ROI size in mm^3
  \item CTX_RH_PRECENTRAL N av1451 uptake - subregion ctx-rh-precentral
  \item CTX_RH_PRECENTRAL_SIZE N mm^3 ctx-rh-precentral ROI size in mm^3
  \item CTX_RH_PRECUNEUS N av1451 uptake - subregion ctx-rh-precuneus
  \item CTX_RH_PRECUNEUS_SIZE N mm^3 ctx-rh-precuneus ROI size in mm^3
  \item CTX_RH_ROSTRALANTERIORCINGULATE N av1451 uptake - subregion ctx-rh-rostralanteriorcingulate
  \item CTX_RH_ROSTRALANTERIORCINGULATE_SIZE N mm^3 ctx-rh-rostralanteriorcingulate ROI size in mm^3
  \item CTX_RH_SUPERIORFRONTAL N av1451 uptake - subregion ctx-rh-superiorfrontal
  \item CTX_RH_SUPERIORFRONTAL_SIZE N mm^3 ctx-rh-superiorfrontal ROI size in mm^3
  \item CTX_RH_SUPERIORPARIETAL N av1451 uptake - subregion ctx-rh-superiorparietal
  \item CTX_RH_SUPERIORPARIETAL_SIZE N mm^3 ctx-rh-superiorparietal ROI size in mm^3
  \item CTX_RH_SUPERIORTEMPORAL N av1451 uptake - subregion ctx-rh-superiortemporal
  \item CTX_RH_SUPERIORTEMPORAL_SIZE N mm^3 ctx-rh-superiortemporal ROI size in mm^3
  \item CTX_RH_SUPRAMARGINAL N av1451 uptake - subregion ctx-rh-supramarginal
  \item CTX_RH_SUPRAMARGINAL_SIZE N mm^3 ctx-rh-supramarginal ROI size in mm^3
  \item CTX_RH_TEMPORALPOLE N av1451 uptake - subregion ctx-rh-temporalpole
  \item CTX_RH_TEMPORALPOLE_SIZE N mm^3 ctx-rh-temporalpole ROI size in mm^3
  \item CTX_RH_TRANSVERSETEMPORAL N av1451 uptake - subregion ctx-rh-transversetemporal
  \item CTX_RH_TRANSVERSETEMPORAL_SIZE N mm^3 ctx-rh-transversetemporal ROI size in mm^3
  \item OTHER N av1451 uptake - subregion Other regions
  \item OTHER_SIZE N mm^3 Other regions size in mm^3
}
}
\examples{
\dontrun{
describe(ucberkeleyav1451_pvc)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UCBERKELEYAV1451_PVC/UCBERKELEY_AV1451_Methods_04.25.2022.pdf}
}
\keyword{datasets}
