% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{ucberkeleyav45}
\alias{ucberkeleyav45}
\alias{AV45,}
\alias{Amyloid,}
\alias{PET}
\title{UC Berkeley - AV45 analysis}
\format{
A data frame with 3086 rows and 246 variables
}
\usage{
data(ucberkeleyav45)
}
\description{
UC Berkeley - AV45 analysis
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID T  Participant roster ID
  \item VISCODE T  Visit code
  \item VISCODE2 T  Translated visit code
  \item EXAMDATE D  Examination Date
  \item TP -4 -4 -4
  \item CEREBELLUMGREYMATTER_SUVR N av45 uptake - reference region Reference region - florbetapir mean of cerebellar grey matter.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details
  \item CEREBELLUMGREYMATTER_VOLUME N mm^3 Reference region - volume of cerebellar grey matter.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details
  \item WHOLECEREBELLUM_SUVR N av45 uptake - reference region Reference region - florbetapir mean of whole cerebellum.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details
  \item WHOLECEREBELLUM_VOLUME N mm^3 Reference region - volume of whole cerebellum.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details
  \item ERODED_SUBCORTICALWM_SUVR N av45 uptake - reference region Reference region - florbetapir mean of eroded subcortical white matter.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details
  \item COMPOSITE_REF_SUVR N av45 uptake - reference region Reference region - florbetapir mean of composite ref region (avg of whole cerebellum, pons/brainstem, eroded WM).  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details
  \item FRONTAL_SUVR N av45 uptake - cortical subregion Weighted florbetapir mean in frontal regions.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details
  \item CINGULATE_SUVR N av45 uptake - cortical subregion Weighted florbetapir mean in anterior/posterior cingulate regions.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details
  \item PARIETAL_SUVR N av45 uptake - cortical subregion Weighted florbetapir mean in lateral parietal regions.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details
  \item TEMPORAL_SUVR N av45 uptake - cortical subregion Weighted florbetapir mean in lateral temporal regions.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details
  \item COMPOSITE_SUVR N av45 uptake - reference region Summary florbetapir cortical SUVR. Weighted florbetapir mean in frontal, cingulate, parietal and temporal regions.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details
  \item SUMMARYSUVR_WHOLECEREBNORM N av45 summary suvr (cortical grey matter/whole cerebellum) Summary florbetapir cortical SUVR normalized by whole cerebellum.  See Jagust lab PDF on LONI for details and cutoff info
  \item SUMMARYSUVR_WHOLECEREBNORM_1.11CUTOFF N binary av45-positivity threshold (cortical grey matter/whole cerebellum) Amyloid positivity determined by applying a cutoff value of 1.11 to the summary florbetapir cortical SUVR normalized by whole cerebellum
  \item SUMMARYSUVR_COMPOSITE_REFNORM N av45 summary suvr (cortical grey matter/composite ref region) Summary florbetapir cortical SUVR normalized by composite ref region.  See Jagust lab PDF on LONI for details
  \item SUMMARYSUVR_COMPOSITE_REFNORM_0.78CUTOFF N binary av45-positivity threshold (cortical grey matter/composite ref region) Amyloid positivity determined by applying a cutoff value of 0.79 to the summary florbetapir cortical SUVR normalized by composite ref region
  \item BRAINSTEM_SUVR N av45 uptake - reference region Reference region - florbetapir mean of brainstem.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details
  \item BRAINSTEM_VOLUME N mm^3 brain-stem ROI size in mm^3
  \item VENTRICLE_3RD_SUVR N av45 uptake - subregion 3rd-ventricle
  \item VENTRICLE_3RD_VOLUME N mm^3 3rd-ventricle ROI size in mm^3
  \item VENTRICLE_4TH_SUVR N av45 uptake - subregion 4th-ventricle
  \item VENTRICLE_4TH_VOLUME N mm^3 4th-ventricle ROI size in mm^3
  \item VENTRICLE_5TH_SUVR N av45 uptake - subregion 5th-ventricle
  \item VENTRICLE_5TH_VOLUME N mm^3 5th-ventricle ROI size in mm^3
  \item CC_ANTERIOR_SUVR N av45 uptake - subregion cc-anterior
  \item CC_ANTERIOR_VOLUME N mm^3 cc-anterior ROI size in mm^3
  \item CC_CENTRAL_SUVR N av45 uptake - subregion cc-central
  \item CC_CENTRAL_VOLUME N mm^3 cc-central ROI size in mm^3
  \item CC_MID_ANTERIOR_SUVR N av45 uptake - subregion cc-mid-anterior
  \item CC_MID_ANTERIOR_VOLUME N mm^3 cc-mid-anterior ROI size in mm^3
  \item CC_MID_POSTERIOR_SUVR N av45 uptake - subregion cc-mid-posterior
  \item CC_MID_POSTERIOR_VOLUME N mm^3 cc-mid-posterior ROI size in mm^3
  \item CC_POSTERIOR_SUVR N av45 uptake - subregion cc-posterior
  \item CC_POSTERIOR_VOLUME N mm^3 cc-posterior ROI size in mm^3
  \item CSF_SUVR N av45 uptake - subregion csf
  \item CSF_VOLUME N mm^3 csf ROI size in mm^3
  \item CTX_LH_BANKSSTS_SUVR N av45 uptake - subregion ctx-lh-bankssts
  \item CTX_LH_BANKSSTS_VOLUME N mm^3 ctx-lh-bankssts ROI size in mm^3
  \item CTX_LH_CAUDALANTERIORCINGULATE_SUVR N av45 uptake - subregion ctx-lh-caudalanteriorcingulate
  \item CTX_LH_CAUDALANTERIORCINGULATE_VOLUME N mm^3 ctx-lh-caudalanteriorcingulate ROI size in mm^3
  \item CTX_LH_CAUDALMIDDLEFRONTAL_SUVR N av45 uptake - subregion ctx-lh-caudalmiddlefrontal
  \item CTX_LH_CAUDALMIDDLEFRONTAL_VOLUME N mm^3 ctx-lh-caudalmiddlefrontal ROI size in mm^3
  \item CTX_LH_CUNEUS_SUVR N av45 uptake - subregion ctx-lh-cuneus
  \item CTX_LH_CUNEUS_VOLUME N mm^3 ctx-lh-cuneus ROI size in mm^3
  \item CTX_LH_ENTORHINAL_SUVR N av45 uptake - subregion ctx-lh-entorhinal
  \item CTX_LH_ENTORHINAL_VOLUME N mm^3 ctx-lh-entorhinal ROI size in mm^3
  \item CTX_LH_FRONTALPOLE_SUVR N av45 uptake - subregion ctx-lh-frontalpole
  \item CTX_LH_FRONTALPOLE_VOLUME N mm^3 ctx-lh-frontalpole ROI size in mm^3
  \item CTX_LH_FUSIFORM_SUVR N av45 uptake - subregion ctx-lh-fusiform
  \item CTX_LH_FUSIFORM_VOLUME N mm^3 ctx-lh-fusiform ROI size in mm^3
  \item CTX_LH_INFERIORPARIETAL_SUVR N av45 uptake - subregion ctx-lh-inferiorparietal
  \item CTX_LH_INFERIORPARIETAL_VOLUME N mm^3 ctx-lh-inferiorparietal ROI size in mm^3
  \item CTX_LH_INFERIORTEMPORAL_SUVR N av45 uptake - subregion ctx-lh-inferiortemporal
  \item CTX_LH_INFERIORTEMPORAL_VOLUME N mm^3 ctx-lh-inferiortemporal ROI size in mm^3
  \item CTX_LH_INSULA_SUVR N av45 uptake - subregion ctx-lh-insula
  \item CTX_LH_INSULA_VOLUME N mm^3 ctx-lh-insula ROI size in mm^3
  \item CTX_LH_ISTHMUSCINGULATE_SUVR N av45 uptake - subregion ctx-lh-isthmuscingulate
  \item CTX_LH_ISTHMUSCINGULATE_VOLUME N mm^3 ctx-lh-isthmuscingulate ROI size in mm^3
  \item CTX_LH_LATERALOCCIPITAL_SUVR N av45 uptake - subregion ctx-lh-lateraloccipital
  \item CTX_LH_LATERALOCCIPITAL_VOLUME N mm^3 ctx-lh-lateraloccipital ROI size in mm^3
  \item CTX_LH_LATERALORBITOFRONTAL_SUVR N av45 uptake - subregion ctx-lh-lateralorbitofrontal
  \item CTX_LH_LATERALORBITOFRONTAL_VOLUME N mm^3 ctx-lh-lateralorbitofrontal ROI size in mm^3
  \item CTX_LH_LINGUAL_SUVR N av45 uptake - subregion ctx-lh-lingual
  \item CTX_LH_LINGUAL_VOLUME N mm^3 ctx-lh-lingual ROI size in mm^3
  \item CTX_LH_MEDIALORBITOFRONTAL_SUVR N av45 uptake - subregion ctx-lh-medialorbitofrontal
  \item CTX_LH_MEDIALORBITOFRONTAL_VOLUME N mm^3 ctx-lh-medialorbitofrontal ROI size in mm^3
  \item CTX_LH_MIDDLETEMPORAL_SUVR N av45 uptake - subregion ctx-lh-middletemporal
  \item CTX_LH_MIDDLETEMPORAL_VOLUME N mm^3 ctx-lh-middletemporal ROI size in mm^3
  \item CTX_LH_PARACENTRAL_SUVR N av45 uptake - subregion ctx-lh-paracentral
  \item CTX_LH_PARACENTRAL_VOLUME N mm^3 ctx-lh-paracentral ROI size in mm^3
  \item CTX_LH_PARAHIPPOCAMPAL_SUVR N av45 uptake - subregion ctx-lh-parahippocampal
  \item CTX_LH_PARAHIPPOCAMPAL_VOLUME N mm^3 ctx-lh-parahippocampal ROI size in mm^3
  \item CTX_LH_PARSOPERCULARIS_SUVR N av45 uptake - subregion ctx-lh-parsopercularis
  \item CTX_LH_PARSOPERCULARIS_VOLUME N mm^3 ctx-lh-parsopercularis ROI size in mm^3
  \item CTX_LH_PARSORBITALIS_SUVR N av45 uptake - subregion ctx-lh-parsorbitalis
  \item CTX_LH_PARSORBITALIS_VOLUME N mm^3 ctx-lh-parsorbitalis ROI size in mm^3
  \item CTX_LH_PARSTRIANGULARIS_SUVR N av45 uptake - subregion ctx-lh-parstriangularis
  \item CTX_LH_PARSTRIANGULARIS_VOLUME N mm^3 ctx-lh-parstriangularis ROI size in mm^3
  \item CTX_LH_PERICALCARINE_SUVR N av45 uptake - subregion ctx-lh-pericalcarine
  \item CTX_LH_PERICALCARINE_VOLUME N mm^3 ctx-lh-pericalcarine ROI size in mm^3
  \item CTX_LH_POSTCENTRAL_SUVR N av45 uptake - subregion ctx-lh-postcentral
  \item CTX_LH_POSTCENTRAL_VOLUME N mm^3 ctx-lh-postcentral ROI size in mm^3
  \item CTX_LH_POSTERIORCINGULATE_SUVR N av45 uptake - subregion ctx-lh-posteriorcingulate
  \item CTX_LH_POSTERIORCINGULATE_VOLUME N mm^3 ctx-lh-posteriorcingulate ROI size in mm^3
  \item CTX_LH_PRECENTRAL_SUVR N av45 uptake - subregion ctx-lh-precentral
  \item CTX_LH_PRECENTRAL_VOLUME N mm^3 ctx-lh-precentral ROI size in mm^3
  \item CTX_LH_PRECUNEUS_SUVR N av45 uptake - subregion ctx-lh-precuneus
  \item CTX_LH_PRECUNEUS_VOLUME N mm^3 ctx-lh-precuneus ROI size in mm^3
  \item CTX_LH_ROSTRALANTERIORCINGULATE_SUVR N av45 uptake - subregion ctx-lh-rostralanteriorcingulate
  \item CTX_LH_ROSTRALANTERIORCINGULATE_VOLUME N mm^3 ctx-lh-rostralanteriorcingulate ROI size in mm^3
  \item CTX_LH_ROSTRALMIDDLEFRONTAL_SUVR N av45 uptake - subregion ctx-lh-rostralmiddlefrontal
  \item CTX_LH_ROSTRALMIDDLEFRONTAL_VOLUME N mm^3 ctx-lh-rostralmiddlefrontal ROI size in mm^3
  \item CTX_LH_SUPERIORFRONTAL_SUVR N av45 uptake - subregion ctx-lh-superiorfrontal
  \item CTX_LH_SUPERIORFRONTAL_VOLUME N mm^3 ctx-lh-superiorfrontal ROI size in mm^3
  \item CTX_LH_SUPERIORPARIETAL_SUVR N av45 uptake - subregion ctx-lh-superiorparietal
  \item CTX_LH_SUPERIORPARIETAL_VOLUME N mm^3 ctx-lh-superiorparietal ROI size in mm^3
  \item CTX_LH_SUPERIORTEMPORAL_SUVR N av45 uptake - subregion ctx-lh-superiortemporal
  \item CTX_LH_SUPERIORTEMPORAL_VOLUME N mm^3 ctx-lh-superiortemporal ROI size in mm^3
  \item CTX_LH_SUPRAMARGINAL_SUVR N av45 uptake - subregion ctx-lh-supramarginal
  \item CTX_LH_SUPRAMARGINAL_VOLUME N mm^3 ctx-lh-supramarginal ROI size in mm^3
  \item CTX_LH_TEMPORALPOLE_SUVR N av45 uptake - subregion ctx-lh-temporalpole
  \item CTX_LH_TEMPORALPOLE_VOLUME N mm^3 ctx-lh-temporalpole ROI size in mm^3
  \item CTX_LH_TRANSVERSETEMPORAL_SUVR N av45 uptake - subregion ctx-lh-transversetemporal
  \item CTX_LH_TRANSVERSETEMPORAL_VOLUME N mm^3 ctx-lh-transversetemporal ROI size in mm^3
  \item CTX_LH_UNKNOWN_SUVR N av45 uptake - subregion ctx-lh-unknown
  \item CTX_LH_UNKNOWN_VOLUME N mm^3 ctx-lh-unknown ROI size in mm^3
  \item CTX_RH_BANKSSTS_SUVR N av45 uptake - subregion ctx-rh-bankssts
  \item CTX_RH_BANKSSTS_VOLUME N mm^3 ctx-rh-bankssts ROI size in mm^3
  \item CTX_RH_CAUDALANTERIORCINGULATE_SUVR N av45 uptake - subregion ctx-rh-caudalanteriorcingulate
  \item CTX_RH_CAUDALANTERIORCINGULATE_VOLUME N mm^3 ctx-rh-caudalanteriorcingulate ROI size in mm^3
  \item CTX_RH_CAUDALMIDDLEFRONTAL_SUVR N av45 uptake - subregion ctx-rh-caudalmiddlefrontal
  \item CTX_RH_CAUDALMIDDLEFRONTAL_VOLUME N mm^3 ctx-rh-caudalmiddlefrontal ROI size in mm^3
  \item CTX_RH_CUNEUS_SUVR N av45 uptake - subregion ctx-rh-cuneus
  \item CTX_RH_CUNEUS_VOLUME N mm^3 ctx-rh-cuneus ROI size in mm^3
  \item CTX_RH_ENTORHINAL_SUVR N av45 uptake - subregion ctx-rh-entorhinal
  \item CTX_RH_ENTORHINAL_VOLUME N mm^3 ctx-rh-entorhinal ROI size in mm^3
  \item CTX_RH_FRONTALPOLE_SUVR N av45 uptake - subregion ctx-rh-frontalpole
  \item CTX_RH_FRONTALPOLE_VOLUME N mm^3 ctx-rh-frontalpole ROI size in mm^3
  \item CTX_RH_FUSIFORM_SUVR N av45 uptake - subregion ctx-rh-fusiform
  \item CTX_RH_FUSIFORM_VOLUME N mm^3 ctx-rh-fusiform ROI size in mm^3
  \item CTX_RH_INFERIORPARIETAL_SUVR N av45 uptake - subregion ctx-rh-inferiorparietal
  \item CTX_RH_INFERIORPARIETAL_VOLUME N mm^3 ctx-rh-inferiorparietal ROI size in mm^3
  \item CTX_RH_INFERIORTEMPORAL_SUVR N av45 uptake - subregion ctx-rh-inferiortemporal
  \item CTX_RH_INFERIORTEMPORAL_VOLUME N mm^3 ctx-rh-inferiortemporal ROI size in mm^3
  \item CTX_RH_INSULA_SUVR N av45 uptake - subregion ctx-rh-insula
  \item CTX_RH_INSULA_VOLUME N mm^3 ctx-rh-insula ROI size in mm^3
  \item CTX_RH_ISTHMUSCINGULATE_SUVR N av45 uptake - subregion ctx-rh-isthmuscingulate
  \item CTX_RH_ISTHMUSCINGULATE_VOLUME N mm^3 ctx-rh-isthmuscingulate ROI size in mm^3
  \item CTX_RH_LATERALOCCIPITAL_SUVR N av45 uptake - subregion ctx-rh-lateraloccipital
  \item CTX_RH_LATERALOCCIPITAL_VOLUME N mm^3 ctx-rh-lateraloccipital ROI size in mm^3
  \item CTX_RH_LATERALORBITOFRONTAL_SUVR N av45 uptake - subregion ctx-rh-lateralorbitofrontal
  \item CTX_RH_LATERALORBITOFRONTAL_VOLUME N mm^3 ctx-rh-lateralorbitofrontal ROI size in mm^3
  \item CTX_RH_LINGUAL_SUVR N av45 uptake - subregion ctx-rh-lingual
  \item CTX_RH_LINGUAL_VOLUME N mm^3 ctx-rh-lingual ROI size in mm^3
  \item CTX_RH_MEDIALORBITOFRONTAL_SUVR N av45 uptake - subregion ctx-rh-medialorbitofrontal
  \item CTX_RH_MEDIALORBITOFRONTAL_VOLUME N mm^3 ctx-rh-medialorbitofrontal ROI size in mm^3
  \item CTX_RH_MIDDLETEMPORAL_SUVR N av45 uptake - subregion ctx-rh-middletemporal
  \item CTX_RH_MIDDLETEMPORAL_VOLUME N mm^3 ctx-rh-middletemporal ROI size in mm^3
  \item CTX_RH_PARACENTRAL_SUVR N av45 uptake - subregion ctx-rh-paracentral
  \item CTX_RH_PARACENTRAL_VOLUME N mm^3 ctx-rh-paracentral ROI size in mm^3
  \item CTX_RH_PARAHIPPOCAMPAL_SUVR N av45 uptake - subregion ctx-rh-parahippocampal
  \item CTX_RH_PARAHIPPOCAMPAL_VOLUME N mm^3 ctx-rh-parahippocampal ROI size in mm^3
  \item CTX_RH_PARSOPERCULARIS_SUVR N av45 uptake - subregion ctx-rh-parsopercularis
  \item CTX_RH_PARSOPERCULARIS_VOLUME N mm^3 ctx-rh-parsopercularis ROI size in mm^3
  \item CTX_RH_PARSORBITALIS_SUVR N av45 uptake - subregion ctx-rh-parsorbitalis
  \item CTX_RH_PARSORBITALIS_VOLUME N mm^3 ctx-rh-parsorbitalis ROI size in mm^3
  \item CTX_RH_PARSTRIANGULARIS_SUVR N av45 uptake - subregion ctx-rh-parstriangularis
  \item CTX_RH_PARSTRIANGULARIS_VOLUME N mm^3 ctx-rh-parstriangularis ROI size in mm^3
  \item CTX_RH_PERICALCARINE_SUVR N av45 uptake - subregion ctx-rh-pericalcarine
  \item CTX_RH_PERICALCARINE_VOLUME N mm^3 ctx-rh-pericalcarine ROI size in mm^3
  \item CTX_RH_POSTCENTRAL_SUVR N av45 uptake - subregion ctx-rh-postcentral
  \item CTX_RH_POSTCENTRAL_VOLUME N mm^3 ctx-rh-postcentral ROI size in mm^3
  \item CTX_RH_POSTERIORCINGULATE_SUVR N av45 uptake - subregion ctx-rh-posteriorcingulate
  \item CTX_RH_POSTERIORCINGULATE_VOLUME N mm^3 ctx-rh-posteriorcingulate ROI size in mm^3
  \item CTX_RH_PRECENTRAL_SUVR N av45 uptake - subregion ctx-rh-precentral
  \item CTX_RH_PRECENTRAL_VOLUME N mm^3 ctx-rh-precentral ROI size in mm^3
  \item CTX_RH_PRECUNEUS_SUVR N av45 uptake - subregion ctx-rh-precuneus
  \item CTX_RH_PRECUNEUS_VOLUME N mm^3 ctx-rh-precuneus ROI size in mm^3
  \item CTX_RH_ROSTRALANTERIORCINGULATE_SUVR N av45 uptake - subregion ctx-rh-rostralanteriorcingulate
  \item CTX_RH_ROSTRALANTERIORCINGULATE_VOLUME N mm^3 ctx-rh-rostralanteriorcingulate ROI size in mm^3
  \item CTX_RH_ROSTRALMIDDLEFRONTAL_SUVR N av45 uptake - subregion ctx-rh-rostralmiddlefrontal
  \item CTX_RH_ROSTRALMIDDLEFRONTAL_VOLUME N mm^3 ctx-rh-rostralmiddlefrontal ROI size in mm^3
  \item CTX_RH_SUPERIORFRONTAL_SUVR N av45 uptake - subregion ctx-rh-superiorfrontal
  \item CTX_RH_SUPERIORFRONTAL_VOLUME N mm^3 ctx-rh-superiorfrontal ROI size in mm^3
  \item CTX_RH_SUPERIORPARIETAL_SUVR N av45 uptake - subregion ctx-rh-superiorparietal
  \item CTX_RH_SUPERIORPARIETAL_VOLUME N mm^3 ctx-rh-superiorparietal ROI size in mm^3
  \item CTX_RH_SUPERIORTEMPORAL_SUVR N av45 uptake - subregion ctx-rh-superiortemporal
  \item CTX_RH_SUPERIORTEMPORAL_VOLUME N mm^3 ctx-rh-superiortemporal ROI size in mm^3
  \item CTX_RH_SUPRAMARGINAL_SUVR N av45 uptake - subregion ctx-rh-supramarginal
  \item CTX_RH_SUPRAMARGINAL_VOLUME N mm^3 ctx-rh-supramarginal ROI size in mm^3
  \item CTX_RH_TEMPORALPOLE_SUVR N av45 uptake - subregion ctx-rh-temporalpole
  \item CTX_RH_TEMPORALPOLE_VOLUME N mm^3 ctx-rh-temporalpole ROI size in mm^3
  \item CTX_RH_TRANSVERSETEMPORAL_SUVR N av45 uptake - subregion ctx-rh-transversetemporal
  \item CTX_RH_TRANSVERSETEMPORAL_VOLUME N mm^3 ctx-rh-transversetemporal ROI size in mm^3
  \item CTX_RH_UNKNOWN_SUVR N av45 uptake - subregion ctx-rh-unknown
  \item CTX_RH_UNKNOWN_VOLUME N mm^3 ctx-rh-unknown ROI size in mm^3
  \item LEFT_ACCUMBENS_AREA_SUVR N av45 uptake - subregion left-accumbens-area
  \item LEFT_ACCUMBENS_AREA_VOLUME N mm^3 left-accumbens-area ROI size in mm^3
  \item LEFT_AMYGDALA_SUVR N av45 uptake - subregion left-amygdala
  \item LEFT_AMYGDALA_VOLUME N mm^3 left-amygdala ROI size in mm^3
  \item LEFT_CAUDATE_SUVR N av45 uptake - subregion left-caudate
  \item LEFT_CAUDATE_VOLUME N mm^3 left-caudate ROI size in mm^3
  \item LEFT_CEREBELLUM_CORTEX_SUVR N av45 uptake - subregion left-cerebellum-cortex
  \item LEFT_CEREBELLUM_CORTEX_VOLUME N mm^3 left-cerebellum-cortex ROI size in mm^3
  \item LEFT_CEREBELLUM_WHITE_MATTER_SUVR N av45 uptake - subregion left-cerebellum-white-matter
  \item LEFT_CEREBELLUM_WHITE_MATTER_VOLUME N mm^3 left-cerebellum-white-matter ROI size in mm^3
  \item LEFT_CEREBRAL_WHITE_MATTER_SUVR N av45 uptake - subregion left-cerebral-white-matter
  \item LEFT_CEREBRAL_WHITE_MATTER_VOLUME N mm^3 left-cerebral-white-matter ROI size in mm^3
  \item LEFT_CHOROID_PLEXUS_SUVR N av45 uptake - subregion left-choroid-plexus
  \item LEFT_CHOROID_PLEXUS_VOLUME N mm^3 left-choroid-plexus ROI size in mm^3
  \item LEFT_HIPPOCAMPUS_SUVR N av45 uptake - subregion left-hippocampus
  \item LEFT_HIPPOCAMPUS_VOLUME N mm^3 left-hippocampus ROI size in mm^3
  \item LEFT_INF_LAT_VENT_SUVR N av45 uptake - subregion left-inf-lat-vent
  \item LEFT_INF_LAT_VENT_VOLUME N mm^3 left-inf-lat-vent ROI size in mm^3
  \item LEFT_LATERAL_VENTRICLE_SUVR N av45 uptake - subregion left-lateral-ventricle
  \item LEFT_LATERAL_VENTRICLE_VOLUME N mm^3 left-lateral-ventricle ROI size in mm^3
  \item LEFT_PALLIDUM_SUVR N av45 uptake - subregion left-pallidum
  \item LEFT_PALLIDUM_VOLUME N mm^3 left-pallidum ROI size in mm^3
  \item LEFT_PUTAMEN_SUVR N av45 uptake - subregion left-putamen
  \item LEFT_PUTAMEN_VOLUME N mm^3 left-putamen ROI size in mm^3
  \item LEFT_THALAMUS_PROPER_SUVR N av45 uptake - subregion left-thalamus-proper
  \item LEFT_THALAMUS_PROPER_VOLUME N mm^3 left-thalamus-proper ROI size in mm^3
  \item LEFT_VENTRALDC_SUVR N av45 uptake - subregion left-ventraldc
  \item LEFT_VENTRALDC_VOLUME N mm^3 left-ventraldc ROI size in mm^3
  \item LEFT_VESSEL_SUVR N av45 uptake - subregion left-vessel
  \item LEFT_VESSEL_VOLUME N mm^3 left-vessel ROI size in mm^3
  \item NON_WM_HYPOINTENSITIES_SUVR N av45 uptake - subregion non-wm-hypointensities
  \item NON_WM_HYPOINTENSITIES_VOLUME N mm^3 non-wm-hypointensities ROI size in mm^3
  \item OPTIC_CHIASM_SUVR N av45 uptake - subregion optic-chiasm
  \item OPTIC_CHIASM_VOLUME N mm^3 optic-chiasm ROI size in mm^3
  \item RIGHT_ACCUMBENS_AREA_SUVR N av45 uptake - subregion right-accumbens-area
  \item RIGHT_ACCUMBENS_AREA_VOLUME N mm^3 right-accumbens-area ROI size in mm^3
  \item RIGHT_AMYGDALA_SUVR N av45 uptake - subregion right-amygdala
  \item RIGHT_AMYGDALA_VOLUME N mm^3 right-amygdala ROI size in mm^3
  \item RIGHT_CAUDATE_SUVR N av45 uptake - subregion right-caudate
  \item RIGHT_CAUDATE_VOLUME N mm^3 right-caudate ROI size in mm^3
  \item RIGHT_CEREBELLUM_CORTEX_SUVR N av45 uptake - subregion right-cerebellum-cortex
  \item RIGHT_CEREBELLUM_CORTEX_VOLUME N mm^3 right-cerebellum-cortex ROI size in mm^3
  \item RIGHT_CEREBELLUM_WHITE_MATTER_SUVR N av45 uptake - subregion right-cerebellum-white-matter
  \item RIGHT_CEREBELLUM_WHITE_MATTER_VOLUME N mm^3 right-cerebellum-white-matter ROI size in mm^3
  \item RIGHT_CEREBRAL_WHITE_MATTER_SUVR N av45 uptake - subregion right-cerebral-white-matter
  \item RIGHT_CEREBRAL_WHITE_MATTER_VOLUME N mm^3 right-cerebral-white-matter ROI size in mm^3
  \item RIGHT_CHOROID_PLEXUS_SUVR N av45 uptake - subregion right-choroid-plexus
  \item RIGHT_CHOROID_PLEXUS_VOLUME N mm^3 right-choroid-plexus ROI size in mm^3
  \item RIGHT_HIPPOCAMPUS_SUVR N av45 uptake - subregion right-hippocampus
  \item RIGHT_HIPPOCAMPUS_VOLUME N mm^3 right-hippocampus ROI size in mm^3
  \item RIGHT_INF_LAT_VENT_SUVR N av45 uptake - subregion right-inf-lat-vent
  \item RIGHT_INF_LAT_VENT_VOLUME N mm^3 right-inf-lat-vent ROI size in mm^3
  \item RIGHT_LATERAL_VENTRICLE_SUVR N av45 uptake - subregion right-lateral-ventricle
  \item RIGHT_LATERAL_VENTRICLE_VOLUME N mm^3 right-lateral-ventricle ROI size in mm^3
  \item RIGHT_PALLIDUM_SUVR N av45 uptake - subregion right-pallidum
  \item RIGHT_PALLIDUM_VOLUME N mm^3 right-pallidum ROI size in mm^3
  \item RIGHT_PUTAMEN_SUVR N av45 uptake - subregion right-putamen
  \item RIGHT_PUTAMEN_VOLUME N mm^3 right-putamen ROI size in mm^3
  \item RIGHT_THALAMUS_PROPER_SUVR N av45 uptake - subregion right-thalamus-proper
  \item RIGHT_THALAMUS_PROPER_VOLUME N mm^3 right-thalamus-proper ROI size in mm^3
  \item RIGHT_VENTRALDC_SUVR N av45 uptake - subregion right-ventraldc
  \item RIGHT_VENTRALDC_VOLUME N mm^3 right-ventraldc ROI size in mm^3
  \item RIGHT_VESSEL_SUVR N av45 uptake - subregion right-vessel
  \item RIGHT_VESSEL_VOLUME N mm^3 right-vessel ROI size in mm^3
  \item WM_HYPOINTENSITIES_SUVR N av45 uptake - subregion wm-hypointensities
  \item WM_HYPOINTENSITIES_VOLUME N mm^3 wm-hypointensities ROI size in mm^3
  \item CEREBELLUMGREYMATTER N av45 uptake - reference region Reference region - florbetapir mean of cerebellar grey matter.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details
  \item WHOLECEREBELLUM N av45 uptake - reference region Reference region - florbetapir mean of whole cerebellum.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details
  \item ERODED_SUBCORTICALWM N av45 uptake - reference region Reference region - florbetapir mean of eroded subcortical white matter.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details
  \item COMPOSITE_REF N av45 uptake - reference region Reference region - florbetapir mean of composite ref region (avg of whole cerebellum, pons/brainstem, eroded WM).  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details
  \item FRONTAL N av45 uptake in cortical grey matter Weighted florbetapir mean in frontal regions.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details
  \item CINGULATE N av45 uptake in cortical grey matter Weighted florbetapir mean in anterior/posterior cingulate regions.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details
  \item PARIETAL N av45 uptake in cortical grey matter Weighted florbetapir mean in lateral parietal regions.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details
  \item TEMPORAL N av45 uptake in cortical grey matter Weighted florbetapir mean in lateral temporal regions.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details
  \item SUMMARYSUVR_COMPOSITE_REFNORM_0.79CUTOFF N binary Amyloid positivity determined by applying a cutoff value of 0.79 to the summary florbetapir cortical SUVR normalized by composite ref region
  \item BRAINSTEM N av45 uptake - reference region Reference region - florbetapir mean of brainstem.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details
  \item BRAINSTEM_SIZE N mm^3 brain-stem ROI size in mm^3
  \item VENTRICLE_3RD N av45 uptake - subregion 3rd-ventricle
  \item VENTRICLE_3RD_SIZE N mm^3 3rd-ventricle ROI size in mm^3
  \item VENTRICLE_4TH N av45 uptake - subregion 4th-ventricle
  \item VENTRICLE_4TH_SIZE N mm^3 4th-ventricle ROI size in mm^3
  \item VENTRICLE_5TH N av45 uptake - subregion 5th-ventricle
  \item VENTRICLE_5TH_SIZE N mm^3 5th-ventricle ROI size in mm^3
  \item CC_ANTERIOR N av45 uptake - subregion cc-anterior
  \item CC_ANTERIOR_SIZE N mm^3 cc-anterior ROI size in mm^3
  \item CC_CENTRAL N av45 uptake - subregion cc-central
  \item CC_CENTRAL_SIZE N mm^3 cc-central ROI size in mm^3
  \item CC_MID_ANTERIOR N av45 uptake - subregion cc-mid-anterior
  \item CC_MID_ANTERIOR_SIZE N mm^3 cc-mid-anterior ROI size in mm^3
  \item CC_MID_POSTERIOR N av45 uptake - subregion cc-mid-posterior
  \item CC_MID_POSTERIOR_SIZE N mm^3 cc-mid-posterior ROI size in mm^3
  \item CC_POSTERIOR N av45 uptake - subregion cc-posterior
  \item CC_POSTERIOR_SIZE N mm^3 cc-posterior ROI size in mm^3
  \item CSF N av45 uptake - subregion csf
  \item CSF_SIZE N mm^3 csf ROI size in mm^3
  \item CTX_LH_BANKSSTS N av45 uptake - subregion ctx-lh-bankssts
  \item CTX_LH_BANKSSTS_SIZE N mm^3 ctx-lh-bankssts ROI size in mm^3
  \item CTX_LH_CAUDALANTERIORCINGULATE N av45 uptake - subregion ctx-lh-caudalanteriorcingulate
  \item CTX_LH_CAUDALANTERIORCINGULATE_SIZE N mm^3 ctx-lh-caudalanteriorcingulate ROI size in mm^3
  \item CTX_LH_CAUDALMIDDLEFRONTAL N av45 uptake - subregion ctx-lh-caudalmiddlefrontal
  \item CTX_LH_CAUDALMIDDLEFRONTAL_SIZE N mm^3 ctx-lh-caudalmiddlefrontal ROI size in mm^3
  \item CTX_LH_CUNEUS N av45 uptake - subregion ctx-lh-cuneus
  \item CTX_LH_CUNEUS_SIZE N mm^3 ctx-lh-cuneus ROI size in mm^3
  \item CTX_LH_ENTORHINAL N av45 uptake - subregion ctx-lh-entorhinal
  \item CTX_LH_ENTORHINAL_SIZE N mm^3 ctx-lh-entorhinal ROI size in mm^3
  \item CTX_LH_FRONTALPOLE N av45 uptake - subregion ctx-lh-frontalpole
  \item CTX_LH_FRONTALPOLE_SIZE N mm^3 ctx-lh-frontalpole ROI size in mm^3
  \item CTX_LH_FUSIFORM N av45 uptake - subregion ctx-lh-fusiform
  \item CTX_LH_FUSIFORM_SIZE N mm^3 ctx-lh-fusiform ROI size in mm^3
  \item CTX_LH_INFERIORPARIETAL N av45 uptake - subregion ctx-lh-inferiorparietal
  \item CTX_LH_INFERIORPARIETAL_SIZE N mm^3 ctx-lh-inferiorparietal ROI size in mm^3
  \item CTX_LH_INFERIORTEMPORAL N av45 uptake - subregion ctx-lh-inferiortemporal
  \item CTX_LH_INFERIORTEMPORAL_SIZE N mm^3 ctx-lh-inferiortemporal ROI size in mm^3
  \item CTX_LH_INSULA N av45 uptake - subregion ctx-lh-insula
  \item CTX_LH_INSULA_SIZE N mm^3 ctx-lh-insula ROI size in mm^3
  \item CTX_LH_ISTHMUSCINGULATE N av45 uptake - subregion ctx-lh-isthmuscingulate
  \item CTX_LH_ISTHMUSCINGULATE_SIZE N mm^3 ctx-lh-isthmuscingulate ROI size in mm^3
  \item CTX_LH_LATERALOCCIPITAL N av45 uptake - subregion ctx-lh-lateraloccipital
  \item CTX_LH_LATERALOCCIPITAL_SIZE N mm^3 ctx-lh-lateraloccipital ROI size in mm^3
  \item CTX_LH_LATERALORBITOFRONTAL N av45 uptake - subregion ctx-lh-lateralorbitofrontal
  \item CTX_LH_LATERALORBITOFRONTAL_SIZE N mm^3 ctx-lh-lateralorbitofrontal ROI size in mm^3
  \item CTX_LH_LINGUAL N av45 uptake - subregion ctx-lh-lingual
  \item CTX_LH_LINGUAL_SIZE N mm^3 ctx-lh-lingual ROI size in mm^3
  \item CTX_LH_MEDIALORBITOFRONTAL N av45 uptake - subregion ctx-lh-medialorbitofrontal
  \item CTX_LH_MEDIALORBITOFRONTAL_SIZE N mm^3 ctx-lh-medialorbitofrontal ROI size in mm^3
  \item CTX_LH_MIDDLETEMPORAL N av45 uptake - subregion ctx-lh-middletemporal
  \item CTX_LH_MIDDLETEMPORAL_SIZE N mm^3 ctx-lh-middletemporal ROI size in mm^3
  \item CTX_LH_PARACENTRAL N av45 uptake - subregion ctx-lh-paracentral
  \item CTX_LH_PARACENTRAL_SIZE N mm^3 ctx-lh-paracentral ROI size in mm^3
  \item CTX_LH_PARAHIPPOCAMPAL N av45 uptake - subregion ctx-lh-parahippocampal
  \item CTX_LH_PARAHIPPOCAMPAL_SIZE N mm^3 ctx-lh-parahippocampal ROI size in mm^3
  \item CTX_LH_PARSOPERCULARIS N av45 uptake - subregion ctx-lh-parsopercularis
  \item CTX_LH_PARSOPERCULARIS_SIZE N mm^3 ctx-lh-parsopercularis ROI size in mm^3
  \item CTX_LH_PARSORBITALIS N av45 uptake - subregion ctx-lh-parsorbitalis
  \item CTX_LH_PARSORBITALIS_SIZE N mm^3 ctx-lh-parsorbitalis ROI size in mm^3
  \item CTX_LH_PARSTRIANGULARIS N av45 uptake - subregion ctx-lh-parstriangularis
  \item CTX_LH_PARSTRIANGULARIS_SIZE N mm^3 ctx-lh-parstriangularis ROI size in mm^3
  \item CTX_LH_PERICALCARINE N av45 uptake - subregion ctx-lh-pericalcarine
  \item CTX_LH_PERICALCARINE_SIZE N mm^3 ctx-lh-pericalcarine ROI size in mm^3
  \item CTX_LH_POSTCENTRAL N av45 uptake - subregion ctx-lh-postcentral
  \item CTX_LH_POSTCENTRAL_SIZE N mm^3 ctx-lh-postcentral ROI size in mm^3
  \item CTX_LH_POSTERIORCINGULATE N av45 uptake - subregion ctx-lh-posteriorcingulate
  \item CTX_LH_POSTERIORCINGULATE_SIZE N mm^3 ctx-lh-posteriorcingulate ROI size in mm^3
  \item CTX_LH_PRECENTRAL N av45 uptake - subregion ctx-lh-precentral
  \item CTX_LH_PRECENTRAL_SIZE N mm^3 ctx-lh-precentral ROI size in mm^3
  \item CTX_LH_PRECUNEUS N av45 uptake - subregion ctx-lh-precuneus
  \item CTX_LH_PRECUNEUS_SIZE N mm^3 ctx-lh-precuneus ROI size in mm^3
  \item CTX_LH_ROSTRALANTERIORCINGULATE N av45 uptake - subregion ctx-lh-rostralanteriorcingulate
  \item CTX_LH_ROSTRALANTERIORCINGULATE_SIZE N mm^3 ctx-lh-rostralanteriorcingulate ROI size in mm^3
  \item CTX_LH_ROSTRALMIDDLEFRONTAL N av45 uptake - subregion ctx-lh-rostralmiddlefrontal
  \item CTX_LH_ROSTRALMIDDLEFRONTAL_SIZE N mm^3 ctx-lh-rostralmiddlefrontal ROI size in mm^3
  \item CTX_LH_SUPERIORFRONTAL N av45 uptake - subregion ctx-lh-superiorfrontal
  \item CTX_LH_SUPERIORFRONTAL_SIZE N mm^3 ctx-lh-superiorfrontal ROI size in mm^3
  \item CTX_LH_SUPERIORPARIETAL N av45 uptake - subregion ctx-lh-superiorparietal
  \item CTX_LH_SUPERIORPARIETAL_SIZE N mm^3 ctx-lh-superiorparietal ROI size in mm^3
  \item CTX_LH_SUPERIORTEMPORAL N av45 uptake - subregion ctx-lh-superiortemporal
  \item CTX_LH_SUPERIORTEMPORAL_SIZE N mm^3 ctx-lh-superiortemporal ROI size in mm^3
  \item CTX_LH_SUPRAMARGINAL N av45 uptake - subregion ctx-lh-supramarginal
  \item CTX_LH_SUPRAMARGINAL_SIZE N mm^3 ctx-lh-supramarginal ROI size in mm^3
  \item CTX_LH_TEMPORALPOLE N av45 uptake - subregion ctx-lh-temporalpole
  \item CTX_LH_TEMPORALPOLE_SIZE N mm^3 ctx-lh-temporalpole ROI size in mm^3
  \item CTX_LH_TRANSVERSETEMPORAL N av45 uptake - subregion ctx-lh-transversetemporal
  \item CTX_LH_TRANSVERSETEMPORAL_SIZE N mm^3 ctx-lh-transversetemporal ROI size in mm^3
  \item CTX_LH_UNKNOWN N av45 uptake - subregion ctx-lh-unknown
  \item CTX_LH_UNKNOWN_SIZE N mm^3 ctx-lh-unknown ROI size in mm^3
  \item CTX_RH_BANKSSTS N av45 uptake - subregion ctx-rh-bankssts
  \item CTX_RH_BANKSSTS_SIZE N mm^3 ctx-rh-bankssts ROI size in mm^3
  \item CTX_RH_CAUDALANTERIORCINGULATE N av45 uptake - subregion ctx-rh-caudalanteriorcingulate
  \item CTX_RH_CAUDALANTERIORCINGULATE_SIZE N mm^3 ctx-rh-caudalanteriorcingulate ROI size in mm^3
  \item CTX_RH_CAUDALMIDDLEFRONTAL N av45 uptake - subregion ctx-rh-caudalmiddlefrontal
  \item CTX_RH_CAUDALMIDDLEFRONTAL_SIZE N mm^3 ctx-rh-caudalmiddlefrontal ROI size in mm^3
  \item CTX_RH_CUNEUS N av45 uptake - subregion ctx-rh-cuneus
  \item CTX_RH_CUNEUS_SIZE N mm^3 ctx-rh-cuneus ROI size in mm^3
  \item CTX_RH_ENTORHINAL N av45 uptake - subregion ctx-rh-entorhinal
  \item CTX_RH_ENTORHINAL_SIZE N mm^3 ctx-rh-entorhinal ROI size in mm^3
  \item CTX_RH_FRONTALPOLE N av45 uptake - subregion ctx-rh-frontalpole
  \item CTX_RH_FRONTALPOLE_SIZE N mm^3 ctx-rh-frontalpole ROI size in mm^3
  \item CTX_RH_FUSIFORM N av45 uptake - subregion ctx-rh-fusiform
  \item CTX_RH_FUSIFORM_SIZE N mm^3 ctx-rh-fusiform ROI size in mm^3
  \item CTX_RH_INFERIORPARIETAL N av45 uptake - subregion ctx-rh-inferiorparietal
  \item CTX_RH_INFERIORPARIETAL_SIZE N mm^3 ctx-rh-inferiorparietal ROI size in mm^3
  \item CTX_RH_INFERIORTEMPORAL N av45 uptake - subregion ctx-rh-inferiortemporal
  \item CTX_RH_INFERIORTEMPORAL_SIZE N mm^3 ctx-rh-inferiortemporal ROI size in mm^3
  \item CTX_RH_INSULA N av45 uptake - subregion ctx-rh-insula
  \item CTX_RH_INSULA_SIZE N mm^3 ctx-rh-insula ROI size in mm^3
  \item CTX_RH_ISTHMUSCINGULATE N av45 uptake - subregion ctx-rh-isthmuscingulate
  \item CTX_RH_ISTHMUSCINGULATE_SIZE N mm^3 ctx-rh-isthmuscingulate ROI size in mm^3
  \item CTX_RH_LATERALOCCIPITAL N av45 uptake - subregion ctx-rh-lateraloccipital
  \item CTX_RH_LATERALOCCIPITAL_SIZE N mm^3 ctx-rh-lateraloccipital ROI size in mm^3
  \item CTX_RH_LATERALORBITOFRONTAL N av45 uptake - subregion ctx-rh-lateralorbitofrontal
  \item CTX_RH_LATERALORBITOFRONTAL_SIZE N mm^3 ctx-rh-lateralorbitofrontal ROI size in mm^3
  \item CTX_RH_LINGUAL N av45 uptake - subregion ctx-rh-lingual
  \item CTX_RH_LINGUAL_SIZE N mm^3 ctx-rh-lingual ROI size in mm^3
  \item CTX_RH_MEDIALORBITOFRONTAL N av45 uptake - subregion ctx-rh-medialorbitofrontal
  \item CTX_RH_MEDIALORBITOFRONTAL_SIZE N mm^3 ctx-rh-medialorbitofrontal ROI size in mm^3
  \item CTX_RH_MIDDLETEMPORAL N av45 uptake - subregion ctx-rh-middletemporal
  \item CTX_RH_MIDDLETEMPORAL_SIZE N mm^3 ctx-rh-middletemporal ROI size in mm^3
  \item CTX_RH_PARACENTRAL N av45 uptake - subregion ctx-rh-paracentral
  \item CTX_RH_PARACENTRAL_SIZE N mm^3 ctx-rh-paracentral ROI size in mm^3
  \item CTX_RH_PARAHIPPOCAMPAL N av45 uptake - subregion ctx-rh-parahippocampal
  \item CTX_RH_PARAHIPPOCAMPAL_SIZE N mm^3 ctx-rh-parahippocampal ROI size in mm^3
  \item CTX_RH_PARSOPERCULARIS N av45 uptake - subregion ctx-rh-parsopercularis
  \item CTX_RH_PARSOPERCULARIS_SIZE N mm^3 ctx-rh-parsopercularis ROI size in mm^3
  \item CTX_RH_PARSORBITALIS N av45 uptake - subregion ctx-rh-parsorbitalis
  \item CTX_RH_PARSORBITALIS_SIZE N mm^3 ctx-rh-parsorbitalis ROI size in mm^3
  \item CTX_RH_PARSTRIANGULARIS N av45 uptake - subregion ctx-rh-parstriangularis
  \item CTX_RH_PARSTRIANGULARIS_SIZE N mm^3 ctx-rh-parstriangularis ROI size in mm^3
  \item CTX_RH_PERICALCARINE N av45 uptake - subregion ctx-rh-pericalcarine
  \item CTX_RH_PERICALCARINE_SIZE N mm^3 ctx-rh-pericalcarine ROI size in mm^3
  \item CTX_RH_POSTCENTRAL N av45 uptake - subregion ctx-rh-postcentral
  \item CTX_RH_POSTCENTRAL_SIZE N mm^3 ctx-rh-postcentral ROI size in mm^3
  \item CTX_RH_POSTERIORCINGULATE N av45 uptake - subregion ctx-rh-posteriorcingulate
  \item CTX_RH_POSTERIORCINGULATE_SIZE N mm^3 ctx-rh-posteriorcingulate ROI size in mm^3
  \item CTX_RH_PRECENTRAL N av45 uptake - subregion ctx-rh-precentral
  \item CTX_RH_PRECENTRAL_SIZE N mm^3 ctx-rh-precentral ROI size in mm^3
  \item CTX_RH_PRECUNEUS N av45 uptake - subregion ctx-rh-precuneus
  \item CTX_RH_PRECUNEUS_SIZE N mm^3 ctx-rh-precuneus ROI size in mm^3
  \item CTX_RH_ROSTRALANTERIORCINGULATE N av45 uptake - subregion ctx-rh-rostralanteriorcingulate
  \item CTX_RH_ROSTRALANTERIORCINGULATE_SIZE N mm^3 ctx-rh-rostralanteriorcingulate ROI size in mm^3
  \item CTX_RH_ROSTRALMIDDLEFRONTAL N av45 uptake - subregion ctx-rh-rostralmiddlefrontal
  \item CTX_RH_ROSTRALMIDDLEFRONTAL_SIZE N mm^3 ctx-rh-rostralmiddlefrontal ROI size in mm^3
  \item CTX_RH_SUPERIORFRONTAL N av45 uptake - subregion ctx-rh-superiorfrontal
  \item CTX_RH_SUPERIORFRONTAL_SIZE N mm^3 ctx-rh-superiorfrontal ROI size in mm^3
  \item CTX_RH_SUPERIORPARIETAL N av45 uptake - subregion ctx-rh-superiorparietal
  \item CTX_RH_SUPERIORPARIETAL_SIZE N mm^3 ctx-rh-superiorparietal ROI size in mm^3
  \item CTX_RH_SUPERIORTEMPORAL N av45 uptake - subregion ctx-rh-superiortemporal
  \item CTX_RH_SUPERIORTEMPORAL_SIZE N mm^3 ctx-rh-superiortemporal ROI size in mm^3
  \item CTX_RH_SUPRAMARGINAL N av45 uptake - subregion ctx-rh-supramarginal
  \item CTX_RH_SUPRAMARGINAL_SIZE N mm^3 ctx-rh-supramarginal ROI size in mm^3
  \item CTX_RH_TEMPORALPOLE N av45 uptake - subregion ctx-rh-temporalpole
  \item CTX_RH_TEMPORALPOLE_SIZE N mm^3 ctx-rh-temporalpole ROI size in mm^3
  \item CTX_RH_TRANSVERSETEMPORAL N av45 uptake - subregion ctx-rh-transversetemporal
  \item CTX_RH_TRANSVERSETEMPORAL_SIZE N mm^3 ctx-rh-transversetemporal ROI size in mm^3
  \item CTX_RH_UNKNOWN N av45 uptake - subregion ctx-rh-unknown
  \item CTX_RH_UNKNOWN_SIZE N mm^3 ctx-rh-unknown ROI size in mm^3
  \item LEFT_ACCUMBENS_AREA N av45 uptake - subregion left-accumbens-area
  \item LEFT_ACCUMBENS_AREA_SIZE N mm^3 left-accumbens-area ROI size in mm^3
  \item LEFT_AMYGDALA N av45 uptake - subregion left-amygdala
  \item LEFT_AMYGDALA_SIZE N mm^3 left-amygdala ROI size in mm^3
  \item LEFT_CAUDATE N av45 uptake - subregion left-caudate
  \item LEFT_CAUDATE_SIZE N mm^3 left-caudate ROI size in mm^3
  \item LEFT_CEREBELLUM_CORTEX N av45 uptake - subregion left-cerebellum-cortex
  \item LEFT_CEREBELLUM_CORTEX_SIZE N mm^3 left-cerebellum-cortex ROI size in mm^3
  \item LEFT_CEREBELLUM_WHITE_MATTER N av45 uptake - subregion left-cerebellum-white-matter
  \item LEFT_CEREBELLUM_WHITE_MATTER_SIZE N mm^3 left-cerebellum-white-matter ROI size in mm^3
  \item LEFT_CEREBRAL_WHITE_MATTER N av45 uptake - subregion left-cerebral-white-matter
  \item LEFT_CEREBRAL_WHITE_MATTER_SIZE N mm^3 left-cerebral-white-matter ROI size in mm^3
  \item LEFT_CHOROID_PLEXUS N av45 uptake - subregion left-choroid-plexus
  \item LEFT_CHOROID_PLEXUS_SIZE N mm^3 left-choroid-plexus ROI size in mm^3
  \item LEFT_HIPPOCAMPUS N av45 uptake - subregion left-hippocampus
  \item LEFT_HIPPOCAMPUS_SIZE N mm^3 left-hippocampus ROI size in mm^3
  \item LEFT_INF_LAT_VENT N av45 uptake - subregion left-inf-lat-vent
  \item LEFT_INF_LAT_VENT_SIZE N mm^3 left-inf-lat-vent ROI size in mm^3
  \item LEFT_LATERAL_VENTRICLE N av45 uptake - subregion left-lateral-ventricle
  \item LEFT_LATERAL_VENTRICLE_SIZE N mm^3 left-lateral-ventricle ROI size in mm^3
  \item LEFT_PALLIDUM N av45 uptake - subregion left-pallidum
  \item LEFT_PALLIDUM_SIZE N mm^3 left-pallidum ROI size in mm^3
  \item LEFT_PUTAMEN N av45 uptake - subregion left-putamen
  \item LEFT_PUTAMEN_SIZE N mm^3 left-putamen ROI size in mm^3
  \item LEFT_THALAMUS_PROPER N av45 uptake - subregion left-thalamus-proper
  \item LEFT_THALAMUS_PROPER_SIZE N mm^3 left-thalamus-proper ROI size in mm^3
  \item LEFT_VENTRALDC N av45 uptake - subregion left-ventraldc
  \item LEFT_VENTRALDC_SIZE N mm^3 left-ventraldc ROI size in mm^3
  \item LEFT_VESSEL N av45 uptake - subregion left-vessel
  \item LEFT_VESSEL_SIZE N mm^3 left-vessel ROI size in mm^3
  \item NON_WM_HYPOINTENSITIES N av45 uptake - subregion non-wm-hypointensities
  \item NON_WM_HYPOINTENSITIES_SIZE N mm^3 non-wm-hypointensities ROI size in mm^3
  \item OPTIC_CHIASM N av45 uptake - subregion optic-chiasm
  \item OPTIC_CHIASM_SIZE N mm^3 optic-chiasm ROI size in mm^3
  \item RIGHT_ACCUMBENS_AREA N av45 uptake - subregion right-accumbens-area
  \item RIGHT_ACCUMBENS_AREA_SIZE N mm^3 right-accumbens-area ROI size in mm^3
  \item RIGHT_AMYGDALA N av45 uptake - subregion right-amygdala
  \item RIGHT_AMYGDALA_SIZE N mm^3 right-amygdala ROI size in mm^3
  \item RIGHT_CAUDATE N av45 uptake - subregion right-caudate
  \item RIGHT_CAUDATE_SIZE N mm^3 right-caudate ROI size in mm^3
  \item RIGHT_CEREBELLUM_CORTEX N av45 uptake - subregion right-cerebellum-cortex
  \item RIGHT_CEREBELLUM_CORTEX_SIZE N mm^3 right-cerebellum-cortex ROI size in mm^3
  \item RIGHT_CEREBELLUM_WHITE_MATTER N av45 uptake - subregion right-cerebellum-white-matter
  \item RIGHT_CEREBELLUM_WHITE_MATTER_SIZE N mm^3 right-cerebellum-white-matter ROI size in mm^3
  \item RIGHT_CEREBRAL_WHITE_MATTER N av45 uptake - subregion right-cerebral-white-matter
  \item RIGHT_CEREBRAL_WHITE_MATTER_SIZE N mm^3 right-cerebral-white-matter ROI size in mm^3
  \item RIGHT_CHOROID_PLEXUS N av45 uptake - subregion right-choroid-plexus
  \item RIGHT_CHOROID_PLEXUS_SIZE N mm^3 right-choroid-plexus ROI size in mm^3
  \item RIGHT_HIPPOCAMPUS N av45 uptake - subregion right-hippocampus
  \item RIGHT_HIPPOCAMPUS_SIZE N mm^3 right-hippocampus ROI size in mm^3
  \item RIGHT_INF_LAT_VENT N av45 uptake - subregion right-inf-lat-vent
  \item RIGHT_INF_LAT_VENT_SIZE N mm^3 right-inf-lat-vent ROI size in mm^3
  \item RIGHT_LATERAL_VENTRICLE N av45 uptake - subregion right-lateral-ventricle
  \item RIGHT_LATERAL_VENTRICLE_SIZE N mm^3 right-lateral-ventricle ROI size in mm^3
  \item RIGHT_PALLIDUM N av45 uptake - subregion right-pallidum
  \item RIGHT_PALLIDUM_SIZE N mm^3 right-pallidum ROI size in mm^3
  \item RIGHT_PUTAMEN N av45 uptake - subregion right-putamen
  \item RIGHT_PUTAMEN_SIZE N mm^3 right-putamen ROI size in mm^3
  \item RIGHT_THALAMUS_PROPER N av45 uptake - subregion right-thalamus-proper
  \item RIGHT_THALAMUS_PROPER_SIZE N mm^3 right-thalamus-proper ROI size in mm^3
  \item RIGHT_VENTRALDC N av45 uptake - subregion right-ventraldc
  \item RIGHT_VENTRALDC_SIZE N mm^3 right-ventraldc ROI size in mm^3
  \item RIGHT_VESSEL N av45 uptake - subregion right-vessel
  \item RIGHT_VESSEL_SIZE N mm^3 right-vessel ROI size in mm^3
  \item WM_HYPOINTENSITIES N av45 uptake - subregion wm-hypointensities
  \item WM_HYPOINTENSITIES_SIZE N mm^3 wm-hypointensities ROI size in mm^3
}
}
\examples{
ucberkeleyav45$SUVR   <- rowMeans(ucberkeleyav45[, c('FRONTAL', 'CINGULATE', 'TEMPORAL', 'PARIETAL')]) /
  ucberkeleyav45$WHOLECEREBELLUM

}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UCBERKELEYAV45/ADNI_AV45_Methods_JagustLab_01.20.15.pdf}, \url{https://adni.bitbucket.io/reference/docs/UCBERKELEYAV45/ADNI_AV45_Methods_JagustLab_04.29.14.pdf}, \url{https://adni.bitbucket.io/reference/docs/UCBERKELEYAV45/ADNI_AV45_Methods_JagustLab_06.25.15.pdf}, \url{https://adni.bitbucket.io/reference/docs/UCBERKELEYAV45/florbetapir_dataset_correction_2.pdf}, \url{https://adni.bitbucket.io/reference/docs/UCBERKELEYAV45/UCBERKELEY_AV45_Methods_01.14.21.pdf}, \url{https://adni.bitbucket.io/reference/docs/UCBERKELEYAV45/UCBERKELEY_AV45_Methods_04.25.2022.pdf}, \url{https://adni.bitbucket.io/reference/docs/UCBERKELEYAV45/UCBERKELEY_AV45_Methods_11.15.2021.pdf}, \url{https://adni.bitbucket.io/reference/docs/UCBERKELEYAV45/UCBERKELEY_AV45_Methods_11.20.15.pdf}, \url{https://adni.bitbucket.io/reference/docs/UCBERKELEYAV45/UCBERKELEY_AV45_Methods_12.03.15.pdf}
}
\author{
Susan Landau \email{slandau@berkeley.edu} & William Jagust
Helen Wills Neuroscience Institute, UC Berkeley and Lawrence Berkeley National Laboratory
}
\keyword{datasets}
