% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{ucberkeleyfbb}
\alias{ucberkeleyfbb}
\title{UC Berkeley - FBB analysis}
\format{
A data frame with 521 rows and 246 variables
}
\usage{
data(ucberkeleyfbb)
}
\description{
UC Berkeley - FBB analysis
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID Participant roster ID  Participant roster ID
  \item VISCODE Visit code  Visit code
  \item VISCODE2 Translated visit code  Translated visit code
  \item EXAMDATE Examination Date  Examination Date
  \item TP -4 -4 -4
  \item CEREBELLUMGREYMATTER_SUVR Reference region - florbetaben mean of cerebellar grey matter.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details fbb uptake - reference region -4
  \item CEREBELLUMGREYMATTER_VOLUME Reference region - volume of cerebellar grey matter.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details mm^3 -4
  \item WHOLECEREBELLUM_SUVR Reference region - florbetaben mean of whole cerebellum.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details fbb uptake - reference region -4
  \item WHOLECEREBELLUM_VOLUME Reference region - volume of whole cerebellum.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details mm^3 -4
  \item ERODED_SUBCORTICALWM_SUVR Reference region - florbetaben mean of eroded subcortical white matter.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details fbb uptake - reference region -4
  \item COMPOSITE_REF_SUVR Reference region - florbetaben mean of composite ref region (avg of whole cerebellum, pons/brainstem, eroded WM).  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details fbb uptake - reference region -4
  \item FRONTAL_SUVR Weighted florbetaben mean in frontal regions.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details fbb uptake - cortical subregion -4
  \item CINGULATE_SUVR Weighted florbetaben mean in anterior/posterior cingulate regions.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details fbb uptake - cortical subregion -4
  \item PARIETAL_SUVR Weighted florbetaben mean in lateral parietal regions.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details fbb uptake - cortical subregion -4
  \item TEMPORAL_SUVR Weighted florbetaben mean in lateral temporal regions.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details fbb uptake - cortical subregion -4
  \item COMPOSITE_SUVR Summary florbetaben cortical SUVR. Weighted florbetaben mean in frontal, cingulate, parietal and temporal regions.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details fbb uptake - reference region -4
  \item SUMMARYSUVR_WHOLECEREBNORM Summary florbetaben cortical SUVR normalized by whole cerebellum.  See Jagust lab PDF on LONI for details and cutoff info fbb summary suvr (cortical grey matter/whole cerebellum) -4
  \item SUMMARYSUVR_WHOLECEREBNORM_1.08CUTOFF Amyloid positivity determined by applying a cutoff value of 1.11 to the summary florbetaben cortical SUVR normalized by whole cerebellum binary fbb-positivity threshold (cortical grey matter/whole cerebellum) -4
  \item SUMMARYSUVR_COMPOSITE_REFNORM Summary florbetaben cortical SUVR normalized by composite ref region.  See Jagust lab PDF on LONI for details fbb summary suvr (cortical grey matter/composite ref region) -4
  \item SUMMARYSUVR_COMPOSITE_REFNORM_0.74CUTOFF Amyloid positivity determined by applying a cutoff value of 0.79 to the summary florbetaben cortical SUVR normalized by composite ref region binary fbb-positivity threshold (cortical grey matter/composite ref region) -4
  \item BRAINSTEM_SUVR Reference region - florbetaben mean of brainstem.  Regions defined by Freesurfer; see Jagust lab PDF on LONI for details fbb uptake - subregion -4
  \item BRAINSTEM_VOLUME brain-stem ROI size in mm^3 mm^3 -4
  \item VENTRICLE_3RD_SUVR 3rd-ventricle fbb uptake - subregion -4
  \item VENTRICLE_3RD_VOLUME 3rd-ventricle ROI size in mm^3 mm^3 -4
  \item VENTRICLE_4TH_SUVR 4th-ventricle fbb uptake - subregion -4
  \item VENTRICLE_4TH_VOLUME 4th-ventricle ROI size in mm^3 mm^3 -4
  \item VENTRICLE_5TH_SUVR 5th-ventricle fbb uptake - subregion -4
  \item VENTRICLE_5TH_VOLUME 5th-ventricle ROI size in mm^3 mm^3 -4
  \item CC_ANTERIOR_SUVR cc-anterior fbb uptake - subregion -4
  \item CC_ANTERIOR_VOLUME cc-anterior ROI size in mm^3 mm^3 -4
  \item CC_CENTRAL_SUVR cc-central fbb uptake - subregion -4
  \item CC_CENTRAL_VOLUME cc-central ROI size in mm^3 mm^3 -4
  \item CC_MID_ANTERIOR_SUVR cc-mid-anterior fbb uptake - subregion -4
  \item CC_MID_ANTERIOR_VOLUME cc-mid-anterior ROI size in mm^3 mm^3 -4
  \item CC_MID_POSTERIOR_SUVR cc-mid-posterior fbb uptake - subregion -4
  \item CC_MID_POSTERIOR_VOLUME cc-mid-posterior ROI size in mm^3 mm^3 -4
  \item CC_POSTERIOR_SUVR cc-posterior fbb uptake - subregion -4
  \item CC_POSTERIOR_VOLUME cc-posterior ROI size in mm^3 mm^3 -4
  \item CSF_SUVR csf fbb uptake - subregion -4
  \item CSF_VOLUME csf ROI size in mm^3 mm^3 -4
  \item CTX_LH_BANKSSTS_SUVR ctx-lh-bankssts fbb uptake - subregion -4
  \item CTX_LH_BANKSSTS_VOLUME ctx-lh-bankssts ROI size in mm^3 mm^3 -4
  \item CTX_LH_CAUDALANTERIORCINGULATE_SUVR ctx-lh-caudalanteriorcingulate fbb uptake - subregion -4
  \item CTX_LH_CAUDALANTERIORCINGULATE_VOLUME ctx-lh-caudalanteriorcingulate ROI size in mm^3 mm^3 -4
  \item CTX_LH_CAUDALMIDDLEFRONTAL_SUVR ctx-lh-caudalmiddlefrontal fbb uptake - subregion -4
  \item CTX_LH_CAUDALMIDDLEFRONTAL_VOLUME ctx-lh-caudalmiddlefrontal ROI size in mm^3 mm^3 -4
  \item CTX_LH_CUNEUS_SUVR ctx-lh-cuneus fbb uptake - subregion -4
  \item CTX_LH_CUNEUS_VOLUME ctx-lh-cuneus ROI size in mm^3 mm^3 -4
  \item CTX_LH_ENTORHINAL_SUVR ctx-lh-entorhinal fbb uptake - subregion -4
  \item CTX_LH_ENTORHINAL_VOLUME ctx-lh-entorhinal ROI size in mm^3 mm^3 -4
  \item CTX_LH_FRONTALPOLE_SUVR ctx-lh-frontalpole fbb uptake - subregion -4
  \item CTX_LH_FRONTALPOLE_VOLUME ctx-lh-frontalpole ROI size in mm^3 mm^3 -4
  \item CTX_LH_FUSIFORM_SUVR ctx-lh-fusiform fbb uptake - subregion -4
  \item CTX_LH_FUSIFORM_VOLUME ctx-lh-fusiform ROI size in mm^3 mm^3 -4
  \item CTX_LH_INFERIORPARIETAL_SUVR ctx-lh-inferiorparietal fbb uptake - subregion -4
  \item CTX_LH_INFERIORPARIETAL_VOLUME ctx-lh-inferiorparietal ROI size in mm^3 mm^3 -4
  \item CTX_LH_INFERIORTEMPORAL_SUVR ctx-lh-inferiortemporal fbb uptake - subregion -4
  \item CTX_LH_INFERIORTEMPORAL_VOLUME ctx-lh-inferiortemporal ROI size in mm^3 mm^3 -4
  \item CTX_LH_INSULA_SUVR ctx-lh-insula fbb uptake - subregion -4
  \item CTX_LH_INSULA_VOLUME ctx-lh-insula ROI size in mm^3 mm^3 -4
  \item CTX_LH_ISTHMUSCINGULATE_SUVR ctx-lh-isthmuscingulate fbb uptake - subregion -4
  \item CTX_LH_ISTHMUSCINGULATE_VOLUME ctx-lh-isthmuscingulate ROI size in mm^3 mm^3 -4
  \item CTX_LH_LATERALOCCIPITAL_SUVR ctx-lh-lateraloccipital fbb uptake - subregion -4
  \item CTX_LH_LATERALOCCIPITAL_VOLUME ctx-lh-lateraloccipital ROI size in mm^3 mm^3 -4
  \item CTX_LH_LATERALORBITOFRONTAL_SUVR ctx-lh-lateralorbitofrontal fbb uptake - subregion -4
  \item CTX_LH_LATERALORBITOFRONTAL_VOLUME ctx-lh-lateralorbitofrontal ROI size in mm^3 mm^3 -4
  \item CTX_LH_LINGUAL_SUVR ctx-lh-lingual fbb uptake - subregion -4
  \item CTX_LH_LINGUAL_VOLUME ctx-lh-lingual ROI size in mm^3 mm^3 -4
  \item CTX_LH_MEDIALORBITOFRONTAL_SUVR ctx-lh-medialorbitofrontal fbb uptake - subregion -4
  \item CTX_LH_MEDIALORBITOFRONTAL_VOLUME ctx-lh-medialorbitofrontal ROI size in mm^3 mm^3 -4
  \item CTX_LH_MIDDLETEMPORAL_SUVR ctx-lh-middletemporal fbb uptake - subregion -4
  \item CTX_LH_MIDDLETEMPORAL_VOLUME ctx-lh-middletemporal ROI size in mm^3 mm^3 -4
  \item CTX_LH_PARACENTRAL_SUVR ctx-lh-paracentral fbb uptake - subregion -4
  \item CTX_LH_PARACENTRAL_VOLUME ctx-lh-paracentral ROI size in mm^3 mm^3 -4
  \item CTX_LH_PARAHIPPOCAMPAL_SUVR ctx-lh-parahippocampal fbb uptake - subregion -4
  \item CTX_LH_PARAHIPPOCAMPAL_VOLUME ctx-lh-parahippocampal ROI size in mm^3 mm^3 -4
  \item CTX_LH_PARSOPERCULARIS_SUVR ctx-lh-parsopercularis fbb uptake - subregion -4
  \item CTX_LH_PARSOPERCULARIS_VOLUME ctx-lh-parsopercularis ROI size in mm^3 mm^3 -4
  \item CTX_LH_PARSORBITALIS_SUVR ctx-lh-parsorbitalis fbb uptake - subregion -4
  \item CTX_LH_PARSORBITALIS_VOLUME ctx-lh-parsorbitalis ROI size in mm^3 mm^3 -4
  \item CTX_LH_PARSTRIANGULARIS_SUVR ctx-lh-parstriangularis fbb uptake - subregion -4
  \item CTX_LH_PARSTRIANGULARIS_VOLUME ctx-lh-parstriangularis ROI size in mm^3 mm^3 -4
  \item CTX_LH_PERICALCARINE_SUVR ctx-lh-pericalcarine fbb uptake - subregion -4
  \item CTX_LH_PERICALCARINE_VOLUME ctx-lh-pericalcarine ROI size in mm^3 mm^3 -4
  \item CTX_LH_POSTCENTRAL_SUVR ctx-lh-postcentral fbb uptake - subregion -4
  \item CTX_LH_POSTCENTRAL_VOLUME ctx-lh-postcentral ROI size in mm^3 mm^3 -4
  \item CTX_LH_POSTERIORCINGULATE_SUVR ctx-lh-posteriorcingulate fbb uptake - subregion -4
  \item CTX_LH_POSTERIORCINGULATE_VOLUME ctx-lh-posteriorcingulate ROI size in mm^3 mm^3 -4
  \item CTX_LH_PRECENTRAL_SUVR ctx-lh-precentral fbb uptake - subregion -4
  \item CTX_LH_PRECENTRAL_VOLUME ctx-lh-precentral ROI size in mm^3 mm^3 -4
  \item CTX_LH_PRECUNEUS_SUVR ctx-lh-precuneus fbb uptake - subregion -4
  \item CTX_LH_PRECUNEUS_VOLUME ctx-lh-precuneus ROI size in mm^3 mm^3 -4
  \item CTX_LH_ROSTRALANTERIORCINGULATE_SUVR ctx-lh-rostralanteriorcingulate fbb uptake - subregion -4
  \item CTX_LH_ROSTRALANTERIORCINGULATE_VOLUME ctx-lh-rostralanteriorcingulate ROI size in mm^3 mm^3 -4
  \item CTX_LH_ROSTRALMIDDLEFRONTAL_SUVR ctx-lh-rostralmiddlefrontal fbb uptake - subregion -4
  \item CTX_LH_ROSTRALMIDDLEFRONTAL_VOLUME ctx-lh-rostralmiddlefrontal ROI size in mm^3 mm^3 -4
  \item CTX_LH_SUPERIORFRONTAL_SUVR ctx-lh-superiorfrontal fbb uptake - subregion -4
  \item CTX_LH_SUPERIORFRONTAL_VOLUME ctx-lh-superiorfrontal ROI size in mm^3 mm^3 -4
  \item CTX_LH_SUPERIORPARIETAL_SUVR ctx-lh-superiorparietal fbb uptake - subregion -4
  \item CTX_LH_SUPERIORPARIETAL_VOLUME ctx-lh-superiorparietal ROI size in mm^3 mm^3 -4
  \item CTX_LH_SUPERIORTEMPORAL_SUVR ctx-lh-superiortemporal fbb uptake - subregion -4
  \item CTX_LH_SUPERIORTEMPORAL_VOLUME ctx-lh-superiortemporal ROI size in mm^3 mm^3 -4
  \item CTX_LH_SUPRAMARGINAL_SUVR ctx-lh-supramarginal fbb uptake - subregion -4
  \item CTX_LH_SUPRAMARGINAL_VOLUME ctx-lh-supramarginal ROI size in mm^3 mm^3 -4
  \item CTX_LH_TEMPORALPOLE_SUVR ctx-lh-temporalpole fbb uptake - subregion -4
  \item CTX_LH_TEMPORALPOLE_VOLUME ctx-lh-temporalpole ROI size in mm^3 mm^3 -4
  \item CTX_LH_TRANSVERSETEMPORAL_SUVR ctx-lh-transversetemporal fbb uptake - subregion -4
  \item CTX_LH_TRANSVERSETEMPORAL_VOLUME ctx-lh-transversetemporal ROI size in mm^3 mm^3 -4
  \item CTX_LH_UNKNOWN_SUVR ctx-lh-unknown fbb uptake - subregion -4
  \item CTX_LH_UNKNOWN_VOLUME ctx-lh-unknown ROI size in mm^3 mm^3 -4
  \item CTX_RH_BANKSSTS_SUVR ctx-rh-bankssts fbb uptake - subregion -4
  \item CTX_RH_BANKSSTS_VOLUME ctx-rh-bankssts ROI size in mm^3 mm^3 -4
  \item CTX_RH_CAUDALANTERIORCINGULATE_SUVR ctx-rh-caudalanteriorcingulate fbb uptake - subregion -4
  \item CTX_RH_CAUDALANTERIORCINGULATE_VOLUME ctx-rh-caudalanteriorcingulate ROI size in mm^3 mm^3 -4
  \item CTX_RH_CAUDALMIDDLEFRONTAL_SUVR ctx-rh-caudalmiddlefrontal fbb uptake - subregion -4
  \item CTX_RH_CAUDALMIDDLEFRONTAL_VOLUME ctx-rh-caudalmiddlefrontal ROI size in mm^3 mm^3 -4
  \item CTX_RH_CUNEUS_SUVR ctx-rh-cuneus fbb uptake - subregion -4
  \item CTX_RH_CUNEUS_VOLUME ctx-rh-cuneus ROI size in mm^3 mm^3 -4
  \item CTX_RH_ENTORHINAL_SUVR ctx-rh-entorhinal fbb uptake - subregion -4
  \item CTX_RH_ENTORHINAL_VOLUME ctx-rh-entorhinal ROI size in mm^3 mm^3 -4
  \item CTX_RH_FRONTALPOLE_SUVR ctx-rh-frontalpole fbb uptake - subregion -4
  \item CTX_RH_FRONTALPOLE_VOLUME ctx-rh-frontalpole ROI size in mm^3 mm^3 -4
  \item CTX_RH_FUSIFORM_SUVR ctx-rh-fusiform fbb uptake - subregion -4
  \item CTX_RH_FUSIFORM_VOLUME ctx-rh-fusiform ROI size in mm^3 mm^3 -4
  \item CTX_RH_INFERIORPARIETAL_SUVR ctx-rh-inferiorparietal fbb uptake - subregion -4
  \item CTX_RH_INFERIORPARIETAL_VOLUME ctx-rh-inferiorparietal ROI size in mm^3 mm^3 -4
  \item CTX_RH_INFERIORTEMPORAL_SUVR ctx-rh-inferiortemporal fbb uptake - subregion -4
  \item CTX_RH_INFERIORTEMPORAL_VOLUME ctx-rh-inferiortemporal ROI size in mm^3 mm^3 -4
  \item CTX_RH_INSULA_SUVR ctx-rh-insula fbb uptake - subregion -4
  \item CTX_RH_INSULA_VOLUME ctx-rh-insula ROI size in mm^3 mm^3 -4
  \item CTX_RH_ISTHMUSCINGULATE_SUVR ctx-rh-isthmuscingulate fbb uptake - subregion -4
  \item CTX_RH_ISTHMUSCINGULATE_VOLUME ctx-rh-isthmuscingulate ROI size in mm^3 mm^3 -4
  \item CTX_RH_LATERALOCCIPITAL_SUVR ctx-rh-lateraloccipital fbb uptake - subregion -4
  \item CTX_RH_LATERALOCCIPITAL_VOLUME ctx-rh-lateraloccipital ROI size in mm^3 mm^3 -4
  \item CTX_RH_LATERALORBITOFRONTAL_SUVR ctx-rh-lateralorbitofrontal fbb uptake - subregion -4
  \item CTX_RH_LATERALORBITOFRONTAL_VOLUME ctx-rh-lateralorbitofrontal ROI size in mm^3 mm^3 -4
  \item CTX_RH_LINGUAL_SUVR ctx-rh-lingual fbb uptake - subregion -4
  \item CTX_RH_LINGUAL_VOLUME ctx-rh-lingual ROI size in mm^3 mm^3 -4
  \item CTX_RH_MEDIALORBITOFRONTAL_SUVR ctx-rh-medialorbitofrontal fbb uptake - subregion -4
  \item CTX_RH_MEDIALORBITOFRONTAL_VOLUME ctx-rh-medialorbitofrontal ROI size in mm^3 mm^3 -4
  \item CTX_RH_MIDDLETEMPORAL_SUVR ctx-rh-middletemporal fbb uptake - subregion -4
  \item CTX_RH_MIDDLETEMPORAL_VOLUME ctx-rh-middletemporal ROI size in mm^3 mm^3 -4
  \item CTX_RH_PARACENTRAL_SUVR ctx-rh-paracentral fbb uptake - subregion -4
  \item CTX_RH_PARACENTRAL_VOLUME ctx-rh-paracentral ROI size in mm^3 mm^3 -4
  \item CTX_RH_PARAHIPPOCAMPAL_SUVR ctx-rh-parahippocampal fbb uptake - subregion -4
  \item CTX_RH_PARAHIPPOCAMPAL_VOLUME ctx-rh-parahippocampal ROI size in mm^3 mm^3 -4
  \item CTX_RH_PARSOPERCULARIS_SUVR ctx-rh-parsopercularis fbb uptake - subregion -4
  \item CTX_RH_PARSOPERCULARIS_VOLUME ctx-rh-parsopercularis ROI size in mm^3 mm^3 -4
  \item CTX_RH_PARSORBITALIS_SUVR ctx-rh-parsorbitalis fbb uptake - subregion -4
  \item CTX_RH_PARSORBITALIS_VOLUME ctx-rh-parsorbitalis ROI size in mm^3 mm^3 -4
  \item CTX_RH_PARSTRIANGULARIS_SUVR ctx-rh-parstriangularis fbb uptake - subregion -4
  \item CTX_RH_PARSTRIANGULARIS_VOLUME ctx-rh-parstriangularis ROI size in mm^3 mm^3 -4
  \item CTX_RH_PERICALCARINE_SUVR ctx-rh-pericalcarine fbb uptake - subregion -4
  \item CTX_RH_PERICALCARINE_VOLUME ctx-rh-pericalcarine ROI size in mm^3 mm^3 -4
  \item CTX_RH_POSTCENTRAL_SUVR ctx-rh-postcentral fbb uptake - subregion -4
  \item CTX_RH_POSTCENTRAL_VOLUME ctx-rh-postcentral ROI size in mm^3 mm^3 -4
  \item CTX_RH_POSTERIORCINGULATE_SUVR ctx-rh-posteriorcingulate fbb uptake - subregion -4
  \item CTX_RH_POSTERIORCINGULATE_VOLUME ctx-rh-posteriorcingulate ROI size in mm^3 mm^3 -4
  \item CTX_RH_PRECENTRAL_SUVR ctx-rh-precentral fbb uptake - subregion -4
  \item CTX_RH_PRECENTRAL_VOLUME ctx-rh-precentral ROI size in mm^3 mm^3 -4
  \item CTX_RH_PRECUNEUS_SUVR ctx-rh-precuneus fbb uptake - subregion -4
  \item CTX_RH_PRECUNEUS_VOLUME ctx-rh-precuneus ROI size in mm^3 mm^3 -4
  \item CTX_RH_ROSTRALANTERIORCINGULATE_SUVR ctx-rh-rostralanteriorcingulate fbb uptake - subregion -4
  \item CTX_RH_ROSTRALANTERIORCINGULATE_VOLUME ctx-rh-rostralanteriorcingulate ROI size in mm^3 mm^3 -4
  \item CTX_RH_ROSTRALMIDDLEFRONTAL_SUVR ctx-rh-rostralmiddlefrontal fbb uptake - subregion -4
  \item CTX_RH_ROSTRALMIDDLEFRONTAL_VOLUME ctx-rh-rostralmiddlefrontal ROI size in mm^3 mm^3 -4
  \item CTX_RH_SUPERIORFRONTAL_SUVR ctx-rh-superiorfrontal fbb uptake - subregion -4
  \item CTX_RH_SUPERIORFRONTAL_VOLUME ctx-rh-superiorfrontal ROI size in mm^3 mm^3 -4
  \item CTX_RH_SUPERIORPARIETAL_SUVR ctx-rh-superiorparietal fbb uptake - subregion -4
  \item CTX_RH_SUPERIORPARIETAL_VOLUME ctx-rh-superiorparietal ROI size in mm^3 mm^3 -4
  \item CTX_RH_SUPERIORTEMPORAL_SUVR ctx-rh-superiortemporal fbb uptake - subregion -4
  \item CTX_RH_SUPERIORTEMPORAL_VOLUME ctx-rh-superiortemporal ROI size in mm^3 mm^3 -4
  \item CTX_RH_SUPRAMARGINAL_SUVR ctx-rh-supramarginal fbb uptake - subregion -4
  \item CTX_RH_SUPRAMARGINAL_VOLUME ctx-rh-supramarginal ROI size in mm^3 mm^3 -4
  \item CTX_RH_TEMPORALPOLE_SUVR ctx-rh-temporalpole fbb uptake - subregion -4
  \item CTX_RH_TEMPORALPOLE_VOLUME ctx-rh-temporalpole ROI size in mm^3 mm^3 -4
  \item CTX_RH_TRANSVERSETEMPORAL_SUVR ctx-rh-transversetemporal fbb uptake - subregion -4
  \item CTX_RH_TRANSVERSETEMPORAL_VOLUME ctx-rh-transversetemporal ROI size in mm^3 mm^3 -4
  \item CTX_RH_UNKNOWN_SUVR ctx-rh-unknown fbb uptake - subregion -4
  \item CTX_RH_UNKNOWN_VOLUME ctx-rh-unknown ROI size in mm^3 mm^3 -4
  \item LEFT_ACCUMBENS_AREA_SUVR left-accumbens-area fbb uptake - subregion -4
  \item LEFT_ACCUMBENS_AREA_VOLUME left-accumbens-area ROI size in mm^3 mm^3 -4
  \item LEFT_AMYGDALA_SUVR left-amygdala fbb uptake - subregion -4
  \item LEFT_AMYGDALA_VOLUME left-amygdala ROI size in mm^3 mm^3 -4
  \item LEFT_CAUDATE_SUVR left-caudate fbb uptake - subregion -4
  \item LEFT_CAUDATE_VOLUME left-caudate ROI size in mm^3 mm^3 -4
  \item LEFT_CEREBELLUM_CORTEX_SUVR left-cerebellum-cortex fbb uptake - subregion -4
  \item LEFT_CEREBELLUM_CORTEX_VOLUME left-cerebellum-cortex ROI size in mm^3 mm^3 -4
  \item LEFT_CEREBELLUM_WHITE_MATTER_SUVR left-cerebellum-white-matter fbb uptake - subregion -4
  \item LEFT_CEREBELLUM_WHITE_MATTER_VOLUME left-cerebellum-white-matter ROI size in mm^3 mm^3 -4
  \item LEFT_CEREBRAL_WHITE_MATTER_SUVR left-cerebral-white-matter fbb uptake - subregion -4
  \item LEFT_CEREBRAL_WHITE_MATTER_VOLUME left-cerebral-white-matter ROI size in mm^3 mm^3 -4
  \item LEFT_CHOROID_PLEXUS_SUVR left-choroid-plexus fbb uptake - subregion -4
  \item LEFT_CHOROID_PLEXUS_VOLUME left-choroid-plexus ROI size in mm^3 mm^3 -4
  \item LEFT_HIPPOCAMPUS_SUVR left-hippocampus fbb uptake - subregion -4
  \item LEFT_HIPPOCAMPUS_VOLUME left-hippocampus ROI size in mm^3 mm^3 -4
  \item LEFT_INF_LAT_VENT_SUVR left-inf-lat-vent fbb uptake - subregion -4
  \item LEFT_INF_LAT_VENT_VOLUME left-inf-lat-vent ROI size in mm^3 mm^3 -4
  \item LEFT_LATERAL_VENTRICLE_SUVR left-lateral-ventricle fbb uptake - subregion -4
  \item LEFT_LATERAL_VENTRICLE_VOLUME left-lateral-ventricle ROI size in mm^3 mm^3 -4
  \item LEFT_PALLIDUM_SUVR left-pallidum fbb uptake - subregion -4
  \item LEFT_PALLIDUM_VOLUME left-pallidum ROI size in mm^3 mm^3 -4
  \item LEFT_PUTAMEN_SUVR left-putamen fbb uptake - subregion -4
  \item LEFT_PUTAMEN_VOLUME left-putamen ROI size in mm^3 mm^3 -4
  \item LEFT_THALAMUS_PROPER_SUVR left-thalamus-proper fbb uptake - subregion -4
  \item LEFT_THALAMUS_PROPER_VOLUME left-thalamus-proper ROI size in mm^3 mm^3 -4
  \item LEFT_VENTRALDC_SUVR left-ventraldc fbb uptake - subregion -4
  \item LEFT_VENTRALDC_VOLUME left-ventraldc ROI size in mm^3 mm^3 -4
  \item LEFT_VESSEL_SUVR left-vessel fbb uptake - subregion -4
  \item LEFT_VESSEL_VOLUME left-vessel ROI size in mm^3 mm^3 -4
  \item NON_WM_HYPOINTENSITIES_SUVR non-wm-hypointensities fbb uptake - subregion -4
  \item NON_WM_HYPOINTENSITIES_VOLUME non-wm-hypointensities ROI size in mm^3 mm^3 -4
  \item OPTIC_CHIASM_SUVR optic-chiasm fbb uptake - subregion -4
  \item OPTIC_CHIASM_VOLUME optic-chiasm ROI size in mm^3 mm^3 -4
  \item RIGHT_ACCUMBENS_AREA_SUVR right-accumbens-area fbb uptake - subregion -4
  \item RIGHT_ACCUMBENS_AREA_VOLUME right-accumbens-area ROI size in mm^3 mm^3 -4
  \item RIGHT_AMYGDALA_SUVR right-amygdala fbb uptake - subregion -4
  \item RIGHT_AMYGDALA_VOLUME right-amygdala ROI size in mm^3 mm^3 -4
  \item RIGHT_CAUDATE_SUVR right-caudate fbb uptake - subregion -4
  \item RIGHT_CAUDATE_VOLUME right-caudate ROI size in mm^3 mm^3 -4
  \item RIGHT_CEREBELLUM_CORTEX_SUVR right-cerebellum-cortex fbb uptake - subregion -4
  \item RIGHT_CEREBELLUM_CORTEX_VOLUME right-cerebellum-cortex ROI size in mm^3 mm^3 -4
  \item RIGHT_CEREBELLUM_WHITE_MATTER_SUVR right-cerebellum-white-matter fbb uptake - subregion -4
  \item RIGHT_CEREBELLUM_WHITE_MATTER_VOLUME right-cerebellum-white-matter ROI size in mm^3 mm^3 -4
  \item RIGHT_CEREBRAL_WHITE_MATTER_SUVR right-cerebral-white-matter fbb uptake - subregion -4
  \item RIGHT_CEREBRAL_WHITE_MATTER_VOLUME right-cerebral-white-matter ROI size in mm^3 mm^3 -4
  \item RIGHT_CHOROID_PLEXUS_SUVR right-choroid-plexus fbb uptake - subregion -4
  \item RIGHT_CHOROID_PLEXUS_VOLUME right-choroid-plexus ROI size in mm^3 mm^3 -4
  \item RIGHT_HIPPOCAMPUS_SUVR right-hippocampus fbb uptake - subregion -4
  \item RIGHT_HIPPOCAMPUS_VOLUME right-hippocampus ROI size in mm^3 mm^3 -4
  \item RIGHT_INF_LAT_VENT_SUVR right-inf-lat-vent fbb uptake - subregion -4
  \item RIGHT_INF_LAT_VENT_VOLUME right-inf-lat-vent ROI size in mm^3 mm^3 -4
  \item RIGHT_LATERAL_VENTRICLE_SUVR right-lateral-ventricle fbb uptake - subregion -4
  \item RIGHT_LATERAL_VENTRICLE_VOLUME right-lateral-ventricle ROI size in mm^3 mm^3 -4
  \item RIGHT_PALLIDUM_SUVR right-pallidum fbb uptake - subregion -4
  \item RIGHT_PALLIDUM_VOLUME right-pallidum ROI size in mm^3 mm^3 -4
  \item RIGHT_PUTAMEN_SUVR right-putamen fbb uptake - subregion -4
  \item RIGHT_PUTAMEN_VOLUME right-putamen ROI size in mm^3 mm^3 -4
  \item RIGHT_THALAMUS_PROPER_SUVR right-thalamus-proper fbb uptake - subregion -4
  \item RIGHT_THALAMUS_PROPER_VOLUME right-thalamus-proper ROI size in mm^3 mm^3 -4
  \item RIGHT_VENTRALDC_SUVR right-ventraldc fbb uptake - subregion -4
  \item RIGHT_VENTRALDC_VOLUME right-ventraldc ROI size in mm^3 mm^3 -4
  \item RIGHT_VESSEL_SUVR right-vessel fbb uptake - subregion -4
  \item RIGHT_VESSEL_VOLUME right-vessel ROI size in mm^3 mm^3 -4
  \item WM_HYPOINTENSITIES_SUVR wm-hypointensities fbb uptake - subregion -4
  \item WM_HYPOINTENSITIES_VOLUME wm-hypointensities ROI size in mm^3 mm^3 -4
}
}
\examples{
\dontrun{
describe(ucberkeleyfbb)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UCBERKELEYFBB/UCBerkeley_FBB_Methods_01.14.21.pdf}, \url{https://adni.bitbucket.io/reference/docs/UCBERKELEYFBB/UCBerkeley_FBB_Methods_04.11.19.pdf}, \url{https://adni.bitbucket.io/reference/docs/UCBERKELEYFBB/UCBerkeley_FBB_Methods_04.25.2022.pdf}, \url{https://adni.bitbucket.io/reference/docs/UCBERKELEYFBB/UCBerkeley_FBB_Methods_11.15.2021.pdf}
}
\keyword{datasets}
