% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{ucberkeleyfdg}
\alias{ucberkeleyfdg}
\alias{FDG,}
\alias{Glucose}
\alias{Metabolism,}
\alias{PET}
\title{UC Berkeley - FDG analysis}
\format{
A data frame with 7398 rows and 11 variables
}
\usage{
data(ucberkeleyfdg)
}
\description{
UC Berkeley - FDG analysis
}
\details{
A dataset with variables as follows:

\itemize{
  \item COLPROT -4 -4 Study protocol of data collection
  \item RID -4 -4 Participant roster ID
  \item VISCODE T  Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE D  Examination Date
  \item UID T  Unique image ID
  \item ROINAME T  Region of interest or reference region (for methods, see Landau et al., Neurology, 2010 or email slandau@berkeley.edu)
  \item MEAN N glucose metabolism normalized to pons mean of FDG-PET within ROI
  \item MAX N glucose metabolism normalized to pons max of FDG-PET within ROI
  \item STDEV N glucose metabolism normalized to pons standard deviation of FDG-PET within ROI
  \item TOTVOX N voxels total number of voxels within ROI
  \item ROILAT T  Laterality of ROI
  \item MEDIAN N glucose metabolism normalized to pons median of FDG-PET within ROI
  \item MODE N glucose metabolism normalized to pons mode of FDG-PET within ROI
  \item MIN N glucose metabolism normalized to pons min of FDG-PET within ROI
  \item NANVOX N voxels total number of zero or NaN voxels within ROI
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UCBERKELEYFDG/ADNI_UC_Berkeley_FDG_Methods_20220323.pdf}
}
\author{
Susan Landau \email{slandau@berkeley.edu} & William Jagust
Helen Wills Neuroscience Institute, UC Berkeley and Lawrence Berkeley National Laboratory
}
\keyword{datasets}
