% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{ucd_adni1_wmh}
\alias{ucd_adni1_wmh}
\title{Univerity of California Davis White Matter Hyperintensity Volumes (ADNI1)}
\format{
A data frame with 3469 rows and 10 variables
}
\usage{
data(ucd_adni1_wmh)
}
\description{
Univerity of California Davis White Matter Hyperintensity Volumes (ADNI1)
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID T  Participant roster ID
  \item EXAMDATE D  Examination Date
  \item VISCODE -4 -4 Visit code
  \item WHITMATHYP N cc Adjusted WMH
  \item MANUFACTURER T  MRI Machine Manufacturer
  \item MODEL T  MRI Machine Model
  \item MAGSTRENGTH T  MRI Machine Magnetic Field Strength
  \item SEGPROCIMG T  Images used for segmentation
  \item SEGPROC T  Analysis method used for segmentation
}
}
\examples{
\dontrun{
describe(ucd_adni1_wmh)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UCD_ADNI1_WMH/UCD\%20ADNI\%20WMH\%20Segmentation\%20Method.pdf}
}
\keyword{datasets}
