% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{ucd_adni2_wmh}
\alias{ucd_adni2_wmh}
\title{Univerity of California Davis White Matter Hyperintensity Volumes (ADNI2)}
\format{
A data frame with 3398 rows and 14 variables
}
\usage{
data(ucd_adni2_wmh)
}
\description{
Univerity of California Davis White Matter Hyperintensity Volumes (ADNI2)
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE D  Examination Date
  \item CSF N cc intracranial CSF volume
  \item GRAY N cc Intracranial gray matter volume
  \item WHITE N cc Intracranial white matter volume
  \item WMH N cc WMH volume
  \item TOTAL -4 -4 -4
  \item MAGNETICFIELDSTRENGTH T  MRI Machine Magnetic Field Strength
  \item MANUFACTURERSMODELNAME T  MRI Machine Model
  \item MANUFACTURER T  MRI Machine Manufacturer
  \item SEGPROC T  Images used for segmentation
  \item SEGPROCIMG T  Analysis method used for segmentation
  \item WHITMATHYP N cc WMH volume
  \item ICV N cc Intracranial volume
  \item STATUS N  ADNI Participant status
  \item MAGSTRENGTH T  MRI Machine Magnetic Field Strength
  \item MANUFACTURERSMODEL T  MRI Machine Model
}
}
\examples{
\dontrun{
describe(ucd_adni2_wmh)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UCD_ADNI2_WMH/UCD\%20ADNI\%20II\%204\%20tissue\%20segmentation\%20Method.pdf}
}
\keyword{datasets}
