% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{ucd_wmh}
\alias{ucd_wmh}
\title{UCD - White Matter Hyperintensity Volumes}
\format{
A data frame with 6429 rows and 23 variables
}
\usage{
data(ucd_wmh)
}
\description{
UCD - White Matter Hyperintensity Volumes
}
\details{
A dataset with variables as follows:

\itemize{
  \item COLPROT -4 -4 Study protocol of data collection
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item MANUFACTURER -4  MRI Scanner Manufacturers
  \item MANUFACTURERSMODELNAME -4  MRI Scanner Model Names
  \item MAGNETICFIELDSTRENGTH -4  MRI Scanner Magnetic Field Strength (1.5 or 3.0 Tesla)
  \item MRACQUISITIONTYPEFLAIR -4  Flair Acquisition Type (2D or 3D)
  \item CEREBRUM_TCV -4 Centimeters Cubed Total Cerebrum Cranial Volume (supratentorial portion)
  \item CEREBRUM_TCB -4 Centimeters Cubed Total Cerebrum Brain Volume
  \item CEREBRUM_TCC -4 Centimeters Cubed Total Cerebrum Cerebrospinal Fluid Volume
  \item CEREBRUM_GRAY -4 Centimeters Cubed Total Cerebrum Gray Matter Volume
  \item CEREBRUM_WHITE -4 Centimeters Cubed Total Cerebrum White Matter Volume
  \item LEFT_HIPPO -4 Centimeters Cubed Segmented Left Hippocampus Volume
  \item RIGHT_HIPPO -4 Centimeters Cubed Segmented Right Hippocampus Volume
  \item TOTAL_HIPPO -4 Centimeters Cubed Segmented Total Hippocampi Volume
  \item TOTAL_CSF -4 Centimeters Cubed Total Brain Cerebrospinal Fluid Volume
  \item TOTAL_GRAY -4 Centimeters Cubed Total Brain Gray Matter Volume 
  \item TOTAL_WHITE -4 Centimeters Cubed Total Brain White Matter Volume 
  \item TOTAL_WMH -4 Centimeters Cubed Total White Matter Hyperintensity Volume
  \item TOTAL_BRAIN -4 Centimeters Cubed Total Brain Volume (includes both cerebrum and infratentorial region)
  \item STATUS -4  Analysis Status
}
}
\examples{
\dontrun{
describe(ucd_wmh)
}
}
\keyword{datasets}
