% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{ucsdvol}
\alias{ucsdvol}
\title{Anders Dale Lab (UCSD) - Derived Volumes}
\format{
A data frame with 3212 rows and 22 variables
}
\usage{
data(ucsdvol)
}
\description{
Anders Dale Lab (UCSD) - Derived Volumes
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N -4 Participant roster ID
  \item VISITNUM N -4 
  \item VISCODE T -4 Visit code
  \item EXAMDATE D -4 Examination Date
  \item LONISID T -4 LONISID
  \item QCPASS N -4 
  \item BRAIN N mm3 Whole Brain. A summary measure of total brain parenchyma, including the Cerebral-Cortex, Cerebellum-Cortex, Thalamus-Proper, CaudatePutamen, Pallidum, Hippocampus, Amygdala, Accumbens-area, VentralDC, Cerebral-White-Matter, Cerebellum-White-Matter, and WM-hypointensities. As such, it does not include ventricles or other CSF spaces, and is not intended to be used as a measure of intracranial volume.
  \item EICV N mm3 Estimated Intracranial Volume, aka eTIV. To generate the mask, the baseline image is automatically segmented; all thus-defined brain and ventricular voxels are given the value 1 with all other voxels 0. This binary mask is then repeatedly smoothed with a Gaussian kernel to produce a simply connected uniform mask, covering all sulci, whose boundary tapers smoothly from 1 to 0 over the length of a few voxels. Ideally, the mask would end on the skull and include the brain stem down to where it starts to bend with the neck. The smoothing can be controlled fairly nicely to begin tapering at the skull, so that voxels with a mask value less than 1 can  be considered outside the ICV and therefore ignored.
  \item VENTRICLES N mm3 Ventricles
  \item LHIPPOC N mm3 Left Hippocampus
  \item RHIPPOC N mm3 Right Hippocampus
  \item LINFLATVEN N mm3 Left inferior lateral ventricle
  \item RINFLATVEN N mm3 Right inferior lateral ventricle
  \item LMIDTEMP N mm Left Middle Temporal
  \item RMIDTEMP N mm Right Middle Temporal
  \item LINFTEMP N mm Left Inferior Temporal
  \item RINFTEMP N mm Right Inferior Temporal
  \item LFUSIFORM N mm Left Fusiform
  \item RFUSIFORM N mm Right Fusiform
  \item LENTORHIN N mm Left Entorhinal
  \item RENTORHIN N mm Right Entorhinal
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UCSDVOL/ADpower.jpg}, \url{https://adni.bitbucket.io/reference/docs/UCSDVOL/MCIpower.jpg}
}
\keyword{datasets}
