% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{ucsfaslfscbf}
\alias{ucsfaslfscbf}
\title{ASL Perfusion CBF by FreeSurfer ROI}
\format{
A data frame with 250 rows and 437 variables
}
\usage{
data(ucsfaslfscbf)
}
\description{
ASL Perfusion CBF by FreeSurfer ROI
}
\details{
A dataset with variables as follows:

\itemize{
  \item COLPROT -4 -4 Study protocol of data collection
  \item RID N  Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE N  Examination Date
  \item VERSION N  
  \item LONIUID N  
  \item IMAGEUID N  
  \item RUNDATE N  
  \item ASLFLAVOR T  The dimensionality and pulse sequence of the ASL acquisition
  \item CBFQC N  Checks the qCBF overlaid on the PD image
  \item SEGQC N  Checks the PD (M0) to T1 registration
  \item REGQC N  Checks the FreeSurfer aparc+aseg in T1 space
  \item FSOVERALLQC N  An overall quality rating. Refer to additional QC variables for Partial rating.
  \item FSTEMPQC N  QC rating of temporal lobe. Fail affects the following regions: LeftTemporalPole (ST60); RightTemporalPole (ST119); LeftFusiform (ST26); RightFusiform (ST85); LeftSuperiorTemporal (ST58); RightSuperiorTemporal (ST117); LeftInferiorTemporal (ST32); RightInferiorTemporal (ST91); LeftMiddleTemporal (ST40); RightMiddleTemporal (ST99)
  \item FSFRONTQC N  QC rating of the frontal lobe. Fail affects the following regions: LeftFrontalPole (ST25); RightFrontalPole (ST84); LeftPrecentral (ST51); RightPrecentral (ST110); LeftSuperiorFrontal (ST56); RightSuperiorFrontal (ST115); LeftCaudalMiddleFrontal (ST15); RightCaudalMiddleFrontal (ST74); LeftRostralMiddleFrontal (ST55); RightRostralMiddleFrontal (ST114); LeftMedialOrbitofrontal (ST39); RightMedialOrbitofrontal (ST98)
  \item FSPARQC N  QC rating of the parietal lobe. Fail affects the following regions: LeftPostcentral (ST49); RightPostcentral (ST108); LeftSuperiorParietal (ST57); RightSuperiorParietal (ST116); LeftSupraMarginal (ST59); RightSupraMarginal (ST118); LeftParacentral (ST43); RightParacentral (ST102); LeftInferiorParietal (ST31); RightInferiorParietal (ST90)
  \item FSINSULAQC N  QC rating of the insula. Fail affects the following regions: LeftInsula (ST129); RightInsula (ST130)
  \item FSOCCQC N  QC rating of the occipital lobe. Fail affects the following regions: LeftLingual (ST38); RightLingual (ST97); LeftLateralOccipital (ST35); RightLateralOccipital (ST94); LeftPericalcarine (ST48); RightPericalcarine (ST107); LeftCuneus (ST23); RightCuneus (ST82)
  \item FSBGQC N  QC rating of the basal ganglia. Fail affects the following regions: LeftPutamen (ST53); RightPutamen (ST112); LeftCaudate (ST16); RightCaudate (ST75); LeftPallidum (ST42); RightPallidum (ST101)
  \item FSCWMQC N  QC rating of the cerebral WM. Fail affects the following regions: LeftCerebralWM (ST20); RightCerebralWM (ST79); LeftCorticalWM (ST150); RightCorticalWM (ST151); CorticalWM (ST152); LeftHemisphereWM (ST28); RightHemisphereWM (ST87)
  \item FSVENTQC N  QC rating of the ventricles. Fail affects the following regions: LeftLateralVentricle (ST37); RightLateralVentricle (ST96); LeftInferiorLateralVentricle (ST30); RightInferiorLateralVentricle (ST89); ThirdVentricle (ST127); FourthVentricle (ST9); FifthVentrical (ST8); LeftChoroidPlexus (ST21); RightChoroidPlexus (ST80)
  \item ST1MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: Brain-Stem
  \item ST1MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: Brain-Stem
  \item ST1AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: Brain-Stem
  \item ST1SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: Brain-Stem
  \item ST1CT N  Count of CBF voxels in FreeSurfer ROI: Brain-Stem
  \item ST11MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: Left-Accumbens-area
  \item ST11MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: Left-Accumbens-area
  \item ST11AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: Left-Accumbens-area
  \item ST11SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Accumbens-area
  \item ST11CT N  Count of CBF voxels in FreeSurfer ROI: Left-Accumbens-area
  \item ST12MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: Left-Amygdala
  \item ST12MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: Left-Amygdala
  \item ST12AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: Left-Amygdala
  \item ST12SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Amygdala
  \item ST12CT N  Count of CBF voxels in FreeSurfer ROI: Left-Amygdala
  \item ST13MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-bankssts
  \item ST13MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-bankssts
  \item ST13AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-bankssts
  \item ST13SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-bankssts
  \item ST13CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-bankssts
  \item ST14MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-caudalanteriorcingulate
  \item ST14MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-caudalanteriorcingulate
  \item ST14AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-caudalanteriorcingulate
  \item ST14SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-caudalanteriorcingulate
  \item ST14CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-caudalanteriorcingulate
  \item ST15MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-caudalmiddlefrontal
  \item ST15MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-caudalmiddlefrontal
  \item ST15AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-caudalmiddlefrontal
  \item ST15SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-caudalmiddlefrontal
  \item ST15CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-caudalmiddlefrontal
  \item ST16MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: Left-Caudate
  \item ST16MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: Left-Caudate
  \item ST16AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: Left-Caudate
  \item ST16SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Caudate
  \item ST16CT N  Count of CBF voxels in FreeSurfer ROI: Left-Caudate
  \item ST23MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-cuneus
  \item ST23MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-cuneus
  \item ST23AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-cuneus
  \item ST23SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-cuneus
  \item ST23CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-cuneus
  \item ST24MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-entorhinal
  \item ST24MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-entorhinal
  \item ST24AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-entorhinal
  \item ST24SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-entorhinal
  \item ST24CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-entorhinal
  \item ST25MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-frontalpole
  \item ST25MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-frontalpole
  \item ST25AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-frontalpole
  \item ST25SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-frontalpole
  \item ST25CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-frontalpole
  \item ST26MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-fusiform
  \item ST26MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-fusiform
  \item ST26AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-fusiform
  \item ST26SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-fusiform
  \item ST26CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-fusiform
  \item ST29MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: Left-Hippocampus
  \item ST29MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: Left-Hippocampus
  \item ST29AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: Left-Hippocampus
  \item ST29SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Hippocampus
  \item ST29CT N  Count of CBF voxels in FreeSurfer ROI: Left-Hippocampus
  \item ST31MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-inferiorparietal
  \item ST31MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-inferiorparietal
  \item ST31AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-inferiorparietal
  \item ST31SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-inferiorparietal
  \item ST31CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-inferiorparietal
  \item ST32MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-inferiortemporal
  \item ST32MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-inferiortemporal
  \item ST32AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-inferiortemporal
  \item ST32SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-inferiortemporal
  \item ST32CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-inferiortemporal
  \item ST34MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-isthmuscingulate
  \item ST34MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-isthmuscingulate
  \item ST34AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-isthmuscingulate
  \item ST34SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-isthmuscingulate
  \item ST34CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-isthmuscingulate
  \item ST35MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-lateraloccipital
  \item ST35MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-lateraloccipital
  \item ST35AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-lateraloccipital
  \item ST35SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-lateraloccipital
  \item ST35CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-lateraloccipital
  \item ST36MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-lateralorbitofrontal
  \item ST36MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-lateralorbitofrontal
  \item ST36AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-lateralorbitofrontal
  \item ST36SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-lateralorbitofrontal
  \item ST36CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-lateralorbitofrontal
  \item ST38MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-lingual
  \item ST38MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-lingual
  \item ST38AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-lingual
  \item ST38SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-lingual
  \item ST38CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-lingual
  \item ST39MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-medialorbitofrontal
  \item ST39MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-medialorbitofrontal
  \item ST39AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-medialorbitofrontal
  \item ST39SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-medialorbitofrontal
  \item ST39CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-medialorbitofrontal
  \item ST40MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-middletemporal
  \item ST40MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-middletemporal
  \item ST40AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-middletemporal
  \item ST40SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-middletemporal
  \item ST40CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-middletemporal
  \item ST42MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: Left-Pallidum
  \item ST42MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: Left-Pallidum
  \item ST42AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: Left-Pallidum
  \item ST42SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Pallidum
  \item ST42CT N  Count of CBF voxels in FreeSurfer ROI: Left-Pallidum
  \item ST43MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-paracentral
  \item ST43MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-paracentral
  \item ST43AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-paracentral
  \item ST43SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-paracentral
  \item ST43CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-paracentral
  \item ST44MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-parahippocampal
  \item ST44MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-parahippocampal
  \item ST44AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-parahippocampal
  \item ST44SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-parahippocampal
  \item ST44CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-parahippocampal
  \item ST45MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-parsopercularis
  \item ST45MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-parsopercularis
  \item ST45AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-parsopercularis
  \item ST45SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-parsopercularis
  \item ST45CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-parsopercularis
  \item ST46MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-parsorbitalis
  \item ST46MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-parsorbitalis
  \item ST46AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-parsorbitalis
  \item ST46SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-parsorbitalis
  \item ST46CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-parsorbitalis
  \item ST47MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-parstriangularis
  \item ST47MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-parstriangularis
  \item ST47AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-parstriangularis
  \item ST47SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-parstriangularis
  \item ST47CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-parstriangularis
  \item ST48MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-pericalcarine
  \item ST48MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-pericalcarine
  \item ST48AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-pericalcarine
  \item ST48SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-pericalcarine
  \item ST48CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-pericalcarine
  \item ST49MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-postcentral
  \item ST49MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-postcentral
  \item ST49AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-postcentral
  \item ST49SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-postcentral
  \item ST49CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-postcentral
  \item ST50MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-posteriorcingulate
  \item ST50MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-posteriorcingulate
  \item ST50AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-posteriorcingulate
  \item ST50SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-posteriorcingulate
  \item ST50CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-posteriorcingulate
  \item ST51MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-precentral
  \item ST51MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-precentral
  \item ST51AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-precentral
  \item ST51SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-precentral
  \item ST51CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-precentral
  \item ST52MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-precuneus
  \item ST52MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-precuneus
  \item ST52AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-precuneus
  \item ST52SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-precuneus
  \item ST52CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-precuneus
  \item ST53MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: Left-Putamen
  \item ST53MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: Left-Putamen
  \item ST53AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: Left-Putamen
  \item ST53SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Putamen
  \item ST53CT N  Count of CBF voxels in FreeSurfer ROI: Left-Putamen
  \item ST54MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-rostralanteriorcingulate
  \item ST54MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-rostralanteriorcingulate
  \item ST54AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-rostralanteriorcingulate
  \item ST54SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-rostralanteriorcingulate
  \item ST54CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-rostralanteriorcingulate
  \item ST55MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-rostralmiddlefrontal
  \item ST55MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-rostralmiddlefrontal
  \item ST55AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-rostralmiddlefrontal
  \item ST55SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-rostralmiddlefrontal
  \item ST55CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-rostralmiddlefrontal
  \item ST56MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-superiorfrontal
  \item ST56MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-superiorfrontal
  \item ST56AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-superiorfrontal
  \item ST56SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-superiorfrontal
  \item ST56CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-superiorfrontal
  \item ST57MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-superiorparietal
  \item ST57MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-superiorparietal
  \item ST57AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-superiorparietal
  \item ST57SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-superiorparietal
  \item ST57CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-superiorparietal
  \item ST58MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-superiortemporal
  \item ST58MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-superiortemporal
  \item ST58AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-superiortemporal
  \item ST58SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-superiortemporal
  \item ST58CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-superiortemporal
  \item ST59MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-supramarginal
  \item ST59MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-supramarginal
  \item ST59AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-supramarginal
  \item ST59SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-supramarginal
  \item ST59CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-supramarginal
  \item ST60MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-temporalpole
  \item ST60MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-temporalpole
  \item ST60AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-temporalpole
  \item ST60SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-temporalpole
  \item ST60CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-temporalpole
  \item ST61MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: Left-Thalamus-Proper
  \item ST61MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: Left-Thalamus-Proper
  \item ST61AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: Left-Thalamus-Proper
  \item ST61SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: Left-Thalamus-Proper
  \item ST61CT N  Count of CBF voxels in FreeSurfer ROI: Left-Thalamus-Proper
  \item ST62MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-transversetemporal
  \item ST62MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-transversetemporal
  \item ST62AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-transversetemporal
  \item ST62SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-transversetemporal
  \item ST62CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-transversetemporal
  \item ST70MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: Right-Accumbens-area
  \item ST70MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: Right-Accumbens-area
  \item ST70AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: Right-Accumbens-area
  \item ST70SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Accumbens-area
  \item ST70CT N  Count of CBF voxels in FreeSurfer ROI: Right-Accumbens-area
  \item ST71MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: Right-Amygdala
  \item ST71MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: Right-Amygdala
  \item ST71AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: Right-Amygdala
  \item ST71SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Amygdala
  \item ST71CT N  Count of CBF voxels in FreeSurfer ROI: Right-Amygdala
  \item ST72MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-bankssts
  \item ST72MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-bankssts
  \item ST72AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-bankssts
  \item ST72SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-bankssts
  \item ST72CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-bankssts
  \item ST73MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-caudalanteriorcingulate
  \item ST73MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-caudalanteriorcingulate
  \item ST73AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-caudalanteriorcingulate
  \item ST73SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-caudalanteriorcingulate
  \item ST73CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-caudalanteriorcingulate
  \item ST74MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-caudalmiddlefrontal
  \item ST74MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-caudalmiddlefrontal
  \item ST74AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-caudalmiddlefrontal
  \item ST74SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-caudalmiddlefrontal
  \item ST74CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-caudalmiddlefrontal
  \item ST75MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: Right-Caudate
  \item ST75MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: Right-Caudate
  \item ST75AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: Right-Caudate
  \item ST75SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Caudate
  \item ST75CT N  Count of CBF voxels in FreeSurfer ROI: Right-Caudate
  \item ST82MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-cuneus
  \item ST82MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-cuneus
  \item ST82AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-cuneus
  \item ST82SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-cuneus
  \item ST82CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-cuneus
  \item ST83MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-entorhinal
  \item ST83MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-entorhinal
  \item ST83AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-entorhinal
  \item ST83SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-entorhinal
  \item ST83CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-entorhinal
  \item ST84MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-frontalpole
  \item ST84MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-frontalpole
  \item ST84AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-frontalpole
  \item ST84SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-frontalpole
  \item ST84CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-frontalpole
  \item ST85MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-fusiform
  \item ST85MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-fusiform
  \item ST85AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-fusiform
  \item ST85SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-fusiform
  \item ST85CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-fusiform
  \item ST88MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: Right-Hippocampus
  \item ST88MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: Right-Hippocampus
  \item ST88AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: Right-Hippocampus
  \item ST88SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Hippocampus
  \item ST88CT N  Count of CBF voxels in FreeSurfer ROI: Right-Hippocampus
  \item ST90MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-inferiorparietal
  \item ST90MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-inferiorparietal
  \item ST90AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-inferiorparietal
  \item ST90SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-inferiorparietal
  \item ST90CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-inferiorparietal
  \item ST91MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-inferiortemporal
  \item ST91MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-inferiortemporal
  \item ST91AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-inferiortemporal
  \item ST91SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-inferiortemporal
  \item ST91CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-inferiortemporal
  \item ST93MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-isthmuscingulate
  \item ST93MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-isthmuscingulate
  \item ST93AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-isthmuscingulate
  \item ST93SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-isthmuscingulate
  \item ST93CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-isthmuscingulate
  \item ST94MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-lateraloccipital
  \item ST94MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-lateraloccipital
  \item ST94AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-lateraloccipital
  \item ST94SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-lateraloccipital
  \item ST94CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-lateraloccipital
  \item ST95MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-lateralorbitofrontal
  \item ST95MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-lateralorbitofrontal
  \item ST95AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-lateralorbitofrontal
  \item ST95SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-lateralorbitofrontal
  \item ST95CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-lateralorbitofrontal
  \item ST97MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-lingual
  \item ST97MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-lingual
  \item ST97AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-lingual
  \item ST97SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-lingual
  \item ST97CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-lingual
  \item ST98MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-medialorbitofrontal
  \item ST98MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-medialorbitofrontal
  \item ST98AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-medialorbitofrontal
  \item ST98SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-medialorbitofrontal
  \item ST98CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-medialorbitofrontal
  \item ST99MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-middletemporal
  \item ST99MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-middletemporal
  \item ST99AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-middletemporal
  \item ST99SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-middletemporal
  \item ST99CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-middletemporal
  \item ST101MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: Right-Pallidum
  \item ST101MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: Right-Pallidum
  \item ST101AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: Right-Pallidum
  \item ST101SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Pallidum
  \item ST101CT N  Count of CBF voxels in FreeSurfer ROI: Right-Pallidum
  \item ST102MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-paracentral
  \item ST102MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-paracentral
  \item ST102AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-paracentral
  \item ST102SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-paracentral
  \item ST102CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-paracentral
  \item ST103MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-parahippocampal
  \item ST103MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-parahippocampal
  \item ST103AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-parahippocampal
  \item ST103SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-parahippocampal
  \item ST103CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-parahippocampal
  \item ST104MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-parsopercularis
  \item ST104MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-parsopercularis
  \item ST104AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-parsopercularis
  \item ST104SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-parsopercularis
  \item ST104CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-parsopercularis
  \item ST105MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-parsorbitalis
  \item ST105MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-parsorbitalis
  \item ST105AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-parsorbitalis
  \item ST105SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-parsorbitalis
  \item ST105CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-parsorbitalis
  \item ST106MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-parstriangularis
  \item ST106MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-parstriangularis
  \item ST106AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-parstriangularis
  \item ST106SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-parstriangularis
  \item ST106CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-parstriangularis
  \item ST107MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-pericalcarine
  \item ST107MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-pericalcarine
  \item ST107AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-pericalcarine
  \item ST107SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-pericalcarine
  \item ST107CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-pericalcarine
  \item ST108MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-postcentral
  \item ST108MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-postcentral
  \item ST108AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-postcentral
  \item ST108SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-postcentral
  \item ST108CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-postcentral
  \item ST109MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-posteriorcingulate
  \item ST109MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-posteriorcingulate
  \item ST109AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-posteriorcingulate
  \item ST109SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-posteriorcingulate
  \item ST109CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-posteriorcingulate
  \item ST110MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-precentral
  \item ST110MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-precentral
  \item ST110AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-precentral
  \item ST110SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-precentral
  \item ST110CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-precentral
  \item ST111MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-precuneus
  \item ST111MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-precuneus
  \item ST111AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-precuneus
  \item ST111SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-precuneus
  \item ST111CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-precuneus
  \item ST112MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: Right-Putamen
  \item ST112MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: Right-Putamen
  \item ST112AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: Right-Putamen
  \item ST112SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Putamen
  \item ST112CT N  Count of CBF voxels in FreeSurfer ROI: Right-Putamen
  \item ST113MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-rostralanteriorcingulate
  \item ST113MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-rostralanteriorcingulate
  \item ST113AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-rostralanteriorcingulate
  \item ST113SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-rostralanteriorcingulate
  \item ST113CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-rostralanteriorcingulate
  \item ST114MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-rostralmiddlefrontal
  \item ST114MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-rostralmiddlefrontal
  \item ST114AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-rostralmiddlefrontal
  \item ST114SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-rostralmiddlefrontal
  \item ST114CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-rostralmiddlefrontal
  \item ST115MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-superiorfrontal
  \item ST115MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-superiorfrontal
  \item ST115AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-superiorfrontal
  \item ST115SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-superiorfrontal
  \item ST115CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-superiorfrontal
  \item ST116MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-superiorparietal
  \item ST116MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-superiorparietal
  \item ST116AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-superiorparietal
  \item ST116SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-superiorparietal
  \item ST116CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-superiorparietal
  \item ST117MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-superiortemporal
  \item ST117MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-superiortemporal
  \item ST117AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-superiortemporal
  \item ST117SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-superiortemporal
  \item ST117CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-superiortemporal
  \item ST118MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-supramarginal
  \item ST118MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-supramarginal
  \item ST118AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-supramarginal
  \item ST118SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-supramarginal
  \item ST118CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-supramarginal
  \item ST119MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-temporalpole
  \item ST119MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-temporalpole
  \item ST119AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-temporalpole
  \item ST119SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-temporalpole
  \item ST119CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-temporalpole
  \item ST120MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: Right-Thalamus-Proper
  \item ST120MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: Right-Thalamus-Proper
  \item ST120AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: Right-Thalamus-Proper
  \item ST120SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: Right-Thalamus-Proper
  \item ST120CT N  Count of CBF voxels in FreeSurfer ROI: Right-Thalamus-Proper
  \item ST121MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-transversetemporal
  \item ST121MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-transversetemporal
  \item ST121AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-transversetemporal
  \item ST121SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-transversetemporal
  \item ST121CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-transversetemporal
  \item ST129MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-lh-insula
  \item ST129MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-lh-insula
  \item ST129AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-lh-insula
  \item ST129SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-lh-insula
  \item ST129CT N  Count of CBF voxels in FreeSurfer ROI: ctx-lh-insula
  \item ST130MIN N ml / 100g tissue / min Min CBF value for FreeSurfer ROI: ctx-rh-insula
  \item ST130MAX N ml / 100g tissue / min Max CBF value for FreeSurfer ROI: ctx-rh-insula
  \item ST130AVG F ml / 100g tissue / min Mean CBF value for FreeSurfer ROI: ctx-rh-insula
  \item ST130SD F ml / 100g tissue / min Standard Deviation of Mean CBF value for FreeSurfer ROI: ctx-rh-insula
  \item ST130CT N  Count of CBF voxels in FreeSurfer ROI: ctx-rh-insula
}
}
\examples{
\dontrun{
describe(ucsfaslfscbf)
}
}
\keyword{datasets}
