% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{ucsfatrphy}
\alias{ucsfatrphy}
\title{UCSF (Colin Studholme) Regional Atrophy Rates}
\format{
A data frame with 1721 rows and 13 variables
}
\usage{
data(ucsfatrphy)
}
\description{
UCSF (Colin Studholme) Regional Atrophy Rates
}
\details{
A dataset with variables as follows:

\itemize{
  \item USERDATE S -4 Date record created
  \item RECNO N -4 
  \item RID N -4 Participant roster ID
  \item SITEID N -4 Site ID
  \item VISCODE1 T -4 
  \item EXAMDATE1 D -4 
  \item IMAGEUID1 D -4 
  \item VISCODE2 T -4 Translated visit code
  \item EXAMDATE2 D -4 
  \item IMAGEUID2 D -4 
  \item FLDSTRENG D -4 Field Strength
  \item PCTCHANGE N -4 Summaries of regional atrophy rates between first and last scan (i.e. 12 month interval). There is one percent change for each subject summarizing changes in both temporal lobes together (minus superior temporal gyrus).
}
}
\keyword{datasets}
