% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{ucsffsx}
\alias{ucsffsx}
\title{Cross-Sectional FreeSurfer}
\format{
A data frame with 4486 rows and 366 variables
}
\usage{
data(ucsffsx)
}
\description{
Cross-Sectional FreeSurfer
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID T  Participant roster ID
  \item VISCODE T  Visit code
  \item EXAMDATE D  Examination Date
  \item VERSION D  
  \item LONISID T  The LONI study UID for the image processed
  \item FLDSTRENG T  The field strength. 1.5 or 3
  \item LONIUID T  The LONI series UID for the image processed.
  \item IMAGEUID T  The ImageUID for the image processed.
  \item RUNDATE D  
  \item STATUS T  
  \item OVERALLQC T  An overall quality rating. Refer to additional QC variables for Partial rating.
  \item TEMPQC T  QC rating of temporal lobe. Fail affects the following regions: LeftTemporalPole (ST60); RightTemporalPole (ST119); LeftFusiform (ST26); RightFusiform (ST85); LeftSuperiorTemporal (ST58); RightSuperiorTemporal (ST117); LeftInferiorTemporal (ST32); RightInferiorTemporal (ST91); LeftMiddleTemporal (ST40); RightMiddleTemporal (ST99)
  \item FRONTQC T  QC rating of the frontal lobe. Fail affects the following regions: LeftFrontalPole (ST25); RightFrontalPole (ST84); LeftPrecentral (ST51); RightPrecentral (ST110); LeftSuperiorFrontal (ST56); RightSuperiorFrontal (ST115); LeftCaudalMiddleFrontal (ST15); RightCaudalMiddleFrontal (ST74); LeftRostralMiddleFrontal (ST55); RightRostralMiddleFrontal (ST114); LeftMedialOrbitofrontal (ST39); RightMedialOrbitofrontal (ST98)
  \item PARQC T  QC rating of the parietal lobe. Fail affects the following regions: LeftPostcentral (ST49); RightPostcentral (ST108); LeftSuperiorParietal (ST57); RightSuperiorParietal (ST116); LeftSupraMarginal (ST59); RightSupraMarginal (ST118); LeftParacentral (ST43); RightParacentral (ST102); LeftInferiorParietal (ST31); RightInferiorParietal (ST90)
  \item INSULAQC T  QC rating of the insula. Fail affects the following regions: LeftInsula (ST129); RightInsula (ST130)
  \item OCCQC T  QC rating of the occipital lobe. Fail affects the following regions: LeftLingual (ST38); RightLingual (ST97); LeftLateralOccipital (ST35); RightLateralOccipital (ST94); LeftPericalcarine (ST48); RightPericalcarine (ST107); LeftCuneus (ST23); RightCuneus (ST82)
  \item BGQC T  QC rating of the basal ganglia. Fail affects the following regions: LeftPutamen (ST53); RightPutamen (ST112); LeftCaudate (ST16); RightCaudate (ST75); LeftPallidum (ST42); RightPallidum (ST101)
  \item CWMQC T  QC rating of the cerebral WM. Fail affects the following regions: LeftCerebralWM (ST20); RightCerebralWM (ST79)
  \item VENTQC T  QC rating of the ventricles. Fail affects the following regions: LeftLateralVentricle (ST37); RightLateralVentricle (ST96); LeftInferiorLateralVentricle (ST30); RightInferiorLateralVentricle (ST89); ThirdVentricle (ST127); FourthVentricle (ST9); FifthVentrical (ST8); LeftChoroidPlexus (ST21); RightChoroidPlexus (ST80)
  \item ST100SV N mm3 Volume (WM Parcellation) of RightNonWMHypoIntensities
  \item ST101SV N mm3 Volume (WM Parcellation) of RightPallidum
  \item ST102CV N mm3 Volume (Cortical Parcellation) of RightParacentral
  \item ST102SA N mm2 Surface Area of RightParacentral
  \item ST102TA N mm Cortical Thickness Average of RightParacentral
  \item ST102TS N mm Cortical Thickness Standard Deviation of RightParacentral
  \item ST103CV N mm3 Volume (Cortical Parcellation) of RightParahippocampal
  \item ST103SA N mm2 Surface Area of RightParahippocampal
  \item ST103TA N mm Cortical Thickness Average of RightParahippocampal
  \item ST103TS N mm Cortical Thickness Standard Deviation of RightParahippocampal
  \item ST104CV N mm3 Volume (Cortical Parcellation) of RightParsOpercularis
  \item ST104SA N mm2 Surface Area of RightParsOpercularis
  \item ST104TA N mm Cortical Thickness Average of RightParsOpercularis
  \item ST104TS N mm Cortical Thickness Standard Deviation of RightParsOpercularis
  \item ST105CV N mm3 Volume (Cortical Parcellation) of RightParsOrbitalis
  \item ST105SA N mm2 Surface Area of RightParsOrbitalis
  \item ST105TA N mm Cortical Thickness Average of RightParsOrbitalis
  \item ST105TS N mm Cortical Thickness Standard Deviation of RightParsOrbitalis
  \item ST106CV N mm3 Volume (Cortical Parcellation) of RightParsTriangularis
  \item ST106SA N mm2 Surface Area of RightParsTriangularis
  \item ST106TA N mm Cortical Thickness Average of RightParsTriangularis
  \item ST106TS N mm Cortical Thickness Standard Deviation of RightParsTriangularis
  \item ST107CV N mm3 Volume (Cortical Parcellation) of RightPericalcarine
  \item ST107SA N mm2 Surface Area of RightPericalcarine
  \item ST107TA N mm Cortical Thickness Average of RightPericalcarine
  \item ST107TS N mm Cortical Thickness Standard Deviation of RightPericalcarine
  \item ST108CV N mm3 Volume (Cortical Parcellation) of RightPostcentral
  \item ST108SA N mm2 Surface Area of RightPostcentral
  \item ST108TA N mm Cortical Thickness Average of RightPostcentral
  \item ST108TS N mm Cortical Thickness Standard Deviation of RightPostcentral
  \item ST109CV N mm3 Volume (Cortical Parcellation) of RightPosteriorCingulate
  \item ST109SA N mm2 Surface Area of RightPosteriorCingulate
  \item ST109TA N mm Cortical Thickness Average of RightPosteriorCingulate
  \item ST109TS N mm Cortical Thickness Standard Deviation of RightPosteriorCingulate
  \item ST10CV N mm3 Volume (Cortical Parcellation) of Icv
  \item ST110CV N mm3 Volume (Cortical Parcellation) of RightPrecentral
  \item ST110SA N mm2 Surface Area of RightPrecentral
  \item ST110TA N mm Cortical Thickness Average of RightPrecentral
  \item ST110TS N mm Cortical Thickness Standard Deviation of RightPrecentral
  \item ST111CV N mm3 Volume (Cortical Parcellation) of RightPrecuneus
  \item ST111SA N mm2 Surface Area of RightPrecuneus
  \item ST111TA N mm Cortical Thickness Average of RightPrecuneus
  \item ST111TS N mm Cortical Thickness Standard Deviation of RightPrecuneus
  \item ST112SV N mm3 Volume (WM Parcellation) of RightPutamen
  \item ST113CV N mm3 Volume (Cortical Parcellation) of RightRostralAnteriorCingulate
  \item ST113SA N mm2 Surface Area of RightRostralAnteriorCingulate
  \item ST113TA N mm Cortical Thickness Average of RightRostralAnteriorCingulate
  \item ST113TS N mm Cortical Thickness Standard Deviation of RightRostralAnteriorCingulate
  \item ST114CV N mm3 Volume (Cortical Parcellation) of RightRostralMiddleFrontal
  \item ST114SA N mm2 Surface Area of RightRostralMiddleFrontal
  \item ST114TA N mm Cortical Thickness Average of RightRostralMiddleFrontal
  \item ST114TS N mm Cortical Thickness Standard Deviation of RightRostralMiddleFrontal
  \item ST115CV N mm3 Volume (Cortical Parcellation) of RightSuperiorFrontal
  \item ST115SA N mm2 Surface Area of RightSuperiorFrontal
  \item ST115TA N mm Cortical Thickness Average of RightSuperiorFrontal
  \item ST115TS N mm Cortical Thickness Standard Deviation of RightSuperiorFrontal
  \item ST116CV N mm3 Volume (Cortical Parcellation) of RightSuperiorParietal
  \item ST116SA N mm2 Surface Area of RightSuperiorParietal
  \item ST116TA N mm Cortical Thickness Average of RightSuperiorParietal
  \item ST116TS N mm Cortical Thickness Standard Deviation of RightSuperiorParietal
  \item ST117CV N mm3 Volume (Cortical Parcellation) of RightSuperiorTemporal
  \item ST117SA N mm2 Surface Area of RightSuperiorTemporal
  \item ST117TA N mm Cortical Thickness Average of RightSuperiorTemporal
  \item ST117TS N mm Cortical Thickness Standard Deviation of RightSuperiorTemporal
  \item ST118CV N mm3 Volume (Cortical Parcellation) of RightSupramarginal
  \item ST118SA N mm2 Surface Area of RightSupramarginal
  \item ST118TA N mm Cortical Thickness Average of RightSupramarginal
  \item ST118TS N mm Cortical Thickness Standard Deviation of RightSupramarginal
  \item ST119CV N mm3 Volume (Cortical Parcellation) of RightTemporalPole
  \item ST119SA N mm2 Surface Area of RightTemporalPole
  \item ST119TA N mm Cortical Thickness Average of RightTemporalPole
  \item ST119TS N mm Cortical Thickness Standard Deviation of RightTemporalPole
  \item ST11SV N mm3 Volume (WM Parcellation) of LeftAccumbensArea
  \item ST120SV N mm3 Volume (WM Parcellation) of RightThalamus
  \item ST121CV N mm3 Volume (Cortical Parcellation) of RightTransverseTemporal
  \item ST121SA N mm2 Surface Area of RightTransverseTemporal
  \item ST121TA N mm Cortical Thickness Average of RightTransverseTemporal
  \item ST121TS N mm Cortical Thickness Standard Deviation of RightTransverseTemporal
  \item ST122SV N mm3 Volume (WM Parcellation) of RightUndetermined
  \item ST123CV N mm3 Volume (Cortical Parcellation) of RightUnknown
  \item ST123SA N mm2 Surface Area of RightUnknown
  \item ST123TA N mm Cortical Thickness Average of RightUnknown
  \item ST123TS N mm Cortical Thickness Standard Deviation of RightUnknown
  \item ST124SV N mm3 Volume (WM Parcellation) of RightVentralDC
  \item ST125SV N mm3 Volume (WM Parcellation) of RightVessel
  \item ST126SV N mm3 Volume (WM Parcellation) of RightWMHypoIntensities
  \item ST127SV N mm3 Volume (WM Parcellation) of ThirdVentricle
  \item ST128SV N mm3 Volume (WM Parcellation) of WMHypoIntensities
  \item ST129CV N mm3 Volume (Cortical Parcellation) of LeftInsula
  \item ST129SA N mm2 Surface Area of LeftInsula
  \item ST129TA N mm Cortical Thickness Average of LeftInsula
  \item ST129TS N mm Cortical Thickness Standard Deviation of LeftInsula
  \item ST12SV N mm3 Volume (WM Parcellation) of LeftAmygdala
  \item ST130CV N mm3 Volume (Cortical Parcellation) of RightInsula
  \item ST130SA N mm2 Surface Area of RightInsula
  \item ST130TA N mm Cortical Thickness Average of RightInsula
  \item ST130TS N mm Cortical Thickness Standard Deviation of RightInsula
  \item ST13CV N mm3 Volume (Cortical Parcellation) of LeftBankssts
  \item ST13SA N mm2 Surface Area of LeftBankssts
  \item ST13TA N mm Cortical Thickness Average of LeftBankssts
  \item ST13TS N mm Cortical Thickness Standard Deviation of LeftBankssts
  \item ST14CV N mm3 Volume (Cortical Parcellation) of LeftCaudalAnteriorCingulate
  \item ST14SA N mm2 Surface Area of LeftCaudalAnteriorCingulate
  \item ST14TA N mm Cortical Thickness Average of LeftCaudalAnteriorCingulate
  \item ST14TS N mm Cortical Thickness Standard Deviation of LeftCaudalAnteriorCingulate
  \item ST15CV N mm3 Volume (Cortical Parcellation) of LeftCaudalMiddleFrontal
  \item ST15SA N mm2 Surface Area of LeftCaudalMiddleFrontal
  \item ST15TA N mm Cortical Thickness Average of LeftCaudalMiddleFrontal
  \item ST15TS N mm Cortical Thickness Standard Deviation of LeftCaudalMiddleFrontal
  \item ST16SV N mm3 Volume (WM Parcellation) of LeftCaudate
  \item ST17SV N mm3 Volume (WM Parcellation) of LeftCerebellumCortex
  \item ST18SV N mm3 Volume (WM Parcellation) of LeftCerebellumWM
  \item ST19SV N mm3 Volume (WM Parcellation) of LeftCerebralCortex
  \item ST1SV N mm3 Volume (WM Parcellation) of Brainstem
  \item ST20SV N mm3 Volume (WM Parcellation) of LeftCerebralWM
  \item ST21SV N mm3 Volume (WM Parcellation) of LeftChoroidPlexus
  \item ST22CV N mm3 Volume (Cortical Parcellation) of LeftCorpusCallosum
  \item ST22SA N mm2 Surface Area of LeftCorpusCallosum
  \item ST22TA N mm Cortical Thickness Average of LeftCorpusCallosum
  \item ST22TS N mm Cortical Thickness Standard Deviation of LeftCorpusCallosum
  \item ST23CV N mm3 Volume (Cortical Parcellation) of LeftCuneus
  \item ST23SA N mm2 Surface Area of LeftCuneus
  \item ST23TA N mm Cortical Thickness Average of LeftCuneus
  \item ST23TS N mm Cortical Thickness Standard Deviation of LeftCuneus
  \item ST24CV N mm3 Volume (Cortical Parcellation) of LeftEntorhinal
  \item ST24SA N mm2 Surface Area of LeftEntorhinal
  \item ST24TA N mm Cortical Thickness Average of LeftEntorhinal
  \item ST24TS N mm Cortical Thickness Standard Deviation of LeftEntorhinal
  \item ST25CV N mm3 Volume (Cortical Parcellation) of LeftFrontalPole
  \item ST25SA N mm2 Surface Area of LeftFrontalPole
  \item ST25TA N mm Cortical Thickness Average of LeftFrontalPole
  \item ST25TS N mm Cortical Thickness Standard Deviation of LeftFrontalPole
  \item ST26CV N mm3 Volume (Cortical Parcellation) of LeftFusiform
  \item ST26SA N mm2 Surface Area of LeftFusiform
  \item ST26TA N mm Cortical Thickness Average of LeftFusiform
  \item ST26TS N mm Cortical Thickness Standard Deviation of LeftFusiform
  \item ST27SA N mm2 Surface Area of LeftHemisphere
  \item ST28CV N mm3 Volume (Cortical Parcellation) of LeftHemisphereWM
  \item ST29SV N mm3 Volume (WM Parcellation) of LeftHippocampus
  \item ST2SV N mm3 Volume (WM Parcellation) of CorpusCallosumAnterior
  \item ST30SV N mm3 Volume (WM Parcellation) of LeftInferiorLateralVentricle
  \item ST31CV N mm3 Volume (Cortical Parcellation) of LeftInferiorParietal
  \item ST31SA N mm2 Surface Area of LeftInferiorParietal
  \item ST31TA N mm Cortical Thickness Average of LeftInferiorParietal
  \item ST31TS N mm Cortical Thickness Standard Deviation of LeftInferiorParietal
  \item ST32CV N mm3 Volume (Cortical Parcellation) of LeftInferiorTemporal
  \item ST32SA N mm2 Surface Area of LeftInferiorTemporal
  \item ST32TA N mm Cortical Thickness Average of LeftInferiorTemporal
  \item ST32TS N mm Cortical Thickness Standard Deviation of LeftInferiorTemporal
  \item ST33SV N mm3 Volume (WM Parcellation) of LeftInterior
  \item ST34CV N mm3 Volume (Cortical Parcellation) of LeftIsthmusCingulate
  \item ST34SA N mm2 Surface Area of LeftIsthmusCingulate
  \item ST34TA N mm Cortical Thickness Average of LeftIsthmusCingulate
  \item ST34TS N mm Cortical Thickness Standard Deviation of LeftIsthmusCingulate
  \item ST35CV N mm3 Volume (Cortical Parcellation) of LeftLateralOccipital
  \item ST35SA N mm2 Surface Area of LeftLateralOccipital
  \item ST35TA N mm Cortical Thickness Average of LeftLateralOccipital
  \item ST35TS N mm Cortical Thickness Standard Deviation of LeftLateralOccipital
  \item ST36CV N mm3 Volume (Cortical Parcellation) of LeftLateralOrbitofrontal
  \item ST36SA N mm2 Surface Area of LeftLateralOrbitofrontal
  \item ST36TA N mm Cortical Thickness Average of LeftLateralOrbitofrontal
  \item ST36TS N mm Cortical Thickness Standard Deviation of LeftLateralOrbitofrontal
  \item ST37SV N mm3 Volume (WM Parcellation) of LeftLateralVentricle
  \item ST38CV N mm3 Volume (Cortical Parcellation) of LeftLingual
  \item ST38SA N mm2 Surface Area of LeftLingual
  \item ST38TA N mm Cortical Thickness Average of LeftLingual
  \item ST38TS N mm Cortical Thickness Standard Deviation of LeftLingual
  \item ST39CV N mm3 Volume (Cortical Parcellation) of LeftMedialOrbitofrontal
  \item ST39SA N mm2 Surface Area of LeftMedialOrbitofrontal
  \item ST39TA N mm Cortical Thickness Average of LeftMedialOrbitofrontal
  \item ST39TS N mm Cortical Thickness Standard Deviation of LeftMedialOrbitofrontal
  \item ST3SV N mm3 Volume (WM Parcellation) of CorpusCallosumCentral
  \item ST40CV N mm3 Volume (Cortical Parcellation) of LeftMiddleTemporal
  \item ST40SA N mm2 Surface Area of LeftMiddleTemporal
  \item ST40TA N mm Cortical Thickness Average of LeftMiddleTemporal
  \item ST40TS N mm Cortical Thickness Standard Deviation of LeftMiddleTemporal
  \item ST41SV N mm3 Volume (WM Parcellation) of LeftNonWMHypoIntensities
  \item ST42SV N mm3 Volume (WM Parcellation) of LeftPallidum
  \item ST43CV N mm3 Volume (Cortical Parcellation) of LeftParacentral
  \item ST43SA N mm2 Surface Area of LeftParacentral
  \item ST43TA N mm Cortical Thickness Average of LeftParacentral
  \item ST43TS N mm Cortical Thickness Standard Deviation of LeftParacentral
  \item ST44CV N mm3 Volume (Cortical Parcellation) of LeftParahippocampal
  \item ST44SA N mm2 Surface Area of LeftParahippocampal
  \item ST44TA N mm Cortical Thickness Average of LeftParahippocampal
  \item ST44TS N mm Cortical Thickness Standard Deviation of LeftParahippocampal
  \item ST45CV N mm3 Volume (Cortical Parcellation) of LeftParsOpercularis
  \item ST45SA N mm2 Surface Area of LeftParsOpercularis
  \item ST45TA N mm Cortical Thickness Average of LeftParsOpercularis
  \item ST45TS N mm Cortical Thickness Standard Deviation of LeftParsOpercularis
  \item ST46CV N mm3 Volume (Cortical Parcellation) of LeftParsOrbitalis
  \item ST46SA N mm2 Surface Area of LeftParsOrbitalis
  \item ST46TA N mm Cortical Thickness Average of LeftParsOrbitalis
  \item ST46TS N mm Cortical Thickness Standard Deviation of LeftParsOrbitalis
  \item ST47CV N mm3 Volume (Cortical Parcellation) of LeftParsTriangularis
  \item ST47SA N mm2 Surface Area of LeftParsTriangularis
  \item ST47TA N mm Cortical Thickness Average of LeftParsTriangularis
  \item ST47TS N mm Cortical Thickness Standard Deviation of LeftParsTriangularis
  \item ST48CV N mm3 Volume (Cortical Parcellation) of LeftPericalcarine
  \item ST48SA N mm2 Surface Area of LeftPericalcarine
  \item ST48TA N mm Cortical Thickness Average of LeftPericalcarine
  \item ST48TS N mm Cortical Thickness Standard Deviation of LeftPericalcarine
  \item ST49CV N mm3 Volume (Cortical Parcellation) of LeftPostcentral
  \item ST49SA N mm2 Surface Area of LeftPostcentral
  \item ST49TA N mm Cortical Thickness Average of LeftPostcentral
  \item ST49TS N mm Cortical Thickness Standard Deviation of LeftPostcentral
  \item ST4SV N mm3 Volume (WM Parcellation) of CorpusCallosumMidAnterior
  \item ST50CV N mm3 Volume (Cortical Parcellation) of LeftPosteriorCingulate
  \item ST50SA N mm2 Surface Area of LeftPosteriorCingulate
  \item ST50TA N mm Cortical Thickness Average of LeftPosteriorCingulate
  \item ST50TS N mm Cortical Thickness Standard Deviation of LeftPosteriorCingulate
  \item ST51CV N mm3 Volume (Cortical Parcellation) of LeftPrecentral
  \item ST51SA N mm2 Surface Area of LeftPrecentral
  \item ST51TA N mm Cortical Thickness Average of LeftPrecentral
  \item ST51TS N mm Cortical Thickness Standard Deviation of LeftPrecentral
  \item ST52CV N mm3 Volume (Cortical Parcellation) of LeftPrecuneus
  \item ST52SA N mm2 Surface Area of LeftPrecuneus
  \item ST52TA N mm Cortical Thickness Average of LeftPrecuneus
  \item ST52TS N mm Cortical Thickness Standard Deviation of LeftPrecuneus
  \item ST53SV N mm3 Volume (WM Parcellation) of LeftPutamen
  \item ST54CV N mm3 Volume (Cortical Parcellation) of LeftRostralAnteriorCingulate
  \item ST54SA N mm2 Surface Area of LeftRostralAnteriorCingulate
  \item ST54TA N mm Cortical Thickness Average of LeftRostralAnteriorCingulate
  \item ST54TS N mm Cortical Thickness Standard Deviation of LeftRostralAnteriorCingulate
  \item ST55CV N mm3 Volume (Cortical Parcellation) of LeftRostralMiddleFrontal
  \item ST55SA N mm2 Surface Area of LeftRostralMiddleFrontal
  \item ST55TA N mm Cortical Thickness Average of LeftRostralMiddleFrontal
  \item ST55TS N mm Cortical Thickness Standard Deviation of LeftRostralMiddleFrontal
  \item ST56CV N mm3 Volume (Cortical Parcellation) of LeftSuperiorFrontal
  \item ST56SA N mm2 Surface Area of LeftSuperiorFrontal
  \item ST56TA N mm Cortical Thickness Average of LeftSuperiorFrontal
  \item ST56TS N mm Cortical Thickness Standard Deviation of LeftSuperiorFrontal
  \item ST57CV N mm3 Volume (Cortical Parcellation) of LeftSuperiorParietal
  \item ST57SA N mm2 Surface Area of LeftSuperiorParietal
  \item ST57TA N mm Cortical Thickness Average of LeftSuperiorParietal
  \item ST57TS N mm Cortical Thickness Standard Deviation of LeftSuperiorParietal
  \item ST58CV N mm3 Volume (Cortical Parcellation) of LeftSuperiorTemporal
  \item ST58SA N mm2 Surface Area of LeftSuperiorTemporal
  \item ST58TA N mm Cortical Thickness Average of LeftSuperiorTemporal
  \item ST58TS N mm Cortical Thickness Standard Deviation of LeftSuperiorTemporal
  \item ST59CV N mm3 Volume (Cortical Parcellation) of LeftSupramarginal
  \item ST59SA N mm2 Surface Area of LeftSupramarginal
  \item ST59TA N mm Cortical Thickness Average of LeftSupramarginal
  \item ST59TS N mm Cortical Thickness Standard Deviation of LeftSupramarginal
  \item ST5SV N mm3 Volume (WM Parcellation) of CorpusCallosumMidPosterior
  \item ST60CV N mm3 Volume (Cortical Parcellation) of LeftTemporalPole
  \item ST60SA N mm2 Surface Area of LeftTemporalPole
  \item ST60TA N mm Cortical Thickness Average of LeftTemporalPole
  \item ST60TS N mm Cortical Thickness Standard Deviation of LeftTemporalPole
  \item ST61SV N mm3 Volume (WM Parcellation) of LeftThalamus
  \item ST62CV N mm3 Volume (Cortical Parcellation) of LeftTransverseTemporal
  \item ST62SA N mm2 Surface Area of LeftTransverseTemporal
  \item ST62TA N mm Cortical Thickness Average of LeftTransverseTemporal
  \item ST62TS N mm Cortical Thickness Standard Deviation of LeftTransverseTemporal
  \item ST63SV N mm3 Volume (WM Parcellation) of LeftUndetermined
  \item ST64CV N mm3 Volume (Cortical Parcellation) of LeftUnknown
  \item ST64SA N mm2 Surface Area of LeftUnknown
  \item ST64TA N mm Cortical Thickness Average of LeftUnknown
  \item ST64TS N mm Cortical Thickness Standard Deviation of LeftUnknown
  \item ST65SV N mm3 Volume (WM Parcellation) of LeftVentralDC
  \item ST66SV N mm3 Volume (WM Parcellation) of LeftVessel
  \item ST67SV N mm3 Volume (WM Parcellation) of LeftWMHypoIntensities
  \item ST68SV N mm3 Volume (WM Parcellation) of NonWMHypoIntensities
  \item ST69SV N mm3 Volume (WM Parcellation) of OpticChiasm
  \item ST6SV N mm3 Volume (WM Parcellation) of CorpusCallosumPosterior
  \item ST70SV N mm3 Volume (WM Parcellation) of RightAccumbensArea
  \item ST71SV N mm3 Volume (WM Parcellation) of RightAmygdala
  \item ST72CV N mm3 Volume (Cortical Parcellation) of RightBankssts
  \item ST72SA N mm2 Surface Area of RightBankssts
  \item ST72TA N mm Cortical Thickness Average of RightBankssts
  \item ST72TS N mm Cortical Thickness Standard Deviation of RightBankssts
  \item ST73CV N mm3 Volume (Cortical Parcellation) of RightCaudalAnteriorCingulate
  \item ST73SA N mm2 Surface Area of RightCaudalAnteriorCingulate
  \item ST73TA N mm Cortical Thickness Average of RightCaudalAnteriorCingulate
  \item ST73TS N mm Cortical Thickness Standard Deviation of RightCaudalAnteriorCingulate
  \item ST74CV N mm3 Volume (Cortical Parcellation) of RightCaudalMiddleFrontal
  \item ST74SA N mm2 Surface Area of RightCaudalMiddleFrontal
  \item ST74TA N mm Cortical Thickness Average of RightCaudalMiddleFrontal
  \item ST74TS N mm Cortical Thickness Standard Deviation of RightCaudalMiddleFrontal
  \item ST75SV N mm3 Volume (WM Parcellation) of RightCaudate
  \item ST76SV N mm3 Volume (WM Parcellation) of RightCerebellumCortex
  \item ST77SV N mm3 Volume (WM Parcellation) of RightCerebellumWM
  \item ST78SV N mm3 Volume (WM Parcellation) of RightCerebralCortex
  \item ST79SV N mm3 Volume (WM Parcellation) of RightCerebralWM
  \item ST7SV N mm3 Volume (WM Parcellation) of Csf
  \item ST80SV N mm3 Volume (WM Parcellation) of RightChoroidPlexus
  \item ST81CV N mm3 Volume (Cortical Parcellation) of RightCorpusCallosum
  \item ST81SA N mm2 Surface Area of RightCorpusCallosum
  \item ST81TA N mm Cortical Thickness Average of RightCorpusCallosum
  \item ST81TS N mm Cortical Thickness Standard Deviation of RightCorpusCallosum
  \item ST82CV N mm3 Volume (Cortical Parcellation) of RightCuneus
  \item ST82SA N mm2 Surface Area of RightCuneus
  \item ST82TA N mm Cortical Thickness Average of RightCuneus
  \item ST82TS N mm Cortical Thickness Standard Deviation of RightCuneus
  \item ST83CV N mm3 Volume (Cortical Parcellation) of RightEntorhinal
  \item ST83SA N mm2 Surface Area of RightEntorhinal
  \item ST83TA N mm Cortical Thickness Average of RightEntorhinal
  \item ST83TS N mm Cortical Thickness Standard Deviation of RightEntorhinal
  \item ST84CV N mm3 Volume (Cortical Parcellation) of RightFrontalPole
  \item ST84SA N mm2 Surface Area of RightFrontalPole
  \item ST84TA N mm Cortical Thickness Average of RightFrontalPole
  \item ST84TS N mm Cortical Thickness Standard Deviation of RightFrontalPole
  \item ST85CV N mm3 Volume (Cortical Parcellation) of RightFusiform
  \item ST85SA N mm2 Surface Area of RightFusiform
  \item ST85TA N mm Cortical Thickness Average of RightFusiform
  \item ST85TS N mm Cortical Thickness Standard Deviation of RightFusiform
  \item ST86SA N mm2 Surface Area of RightHemisphere
  \item ST87CV N mm3 Volume (Cortical Parcellation) of RightHemisphereWM
  \item ST88SV N mm3 Volume (WM Parcellation) of RightHippocampus
  \item ST89SV N mm3 Volume (WM Parcellation) of RightInferiorLateralVentricle
  \item ST8SV N mm3 Volume (WM Parcellation) of FifthVentricle
  \item ST90CV N mm3 Volume (Cortical Parcellation) of RightInferiorParietal
  \item ST90SA N mm2 Surface Area of RightInferiorParietal
  \item ST90TA N mm Cortical Thickness Average of RightInferiorParietal
  \item ST90TS N mm Cortical Thickness Standard Deviation of RightInferiorParietal
  \item ST91CV N mm3 Volume (Cortical Parcellation) of RightInferiorTemporal
  \item ST91SA N mm2 Surface Area of RightInferiorTemporal
  \item ST91TA N mm Cortical Thickness Average of RightInferiorTemporal
  \item ST91TS N mm Cortical Thickness Standard Deviation of RightInferiorTemporal
  \item ST92SV N mm3 Volume (WM Parcellation) of RightInterior
  \item ST93CV N mm3 Volume (Cortical Parcellation) of RightIsthmusCingulate
  \item ST93SA N mm2 Surface Area of RightIsthmusCingulate
  \item ST93TA N mm Cortical Thickness Average of RightIsthmusCingulate
  \item ST93TS N mm Cortical Thickness Standard Deviation of RightIsthmusCingulate
  \item ST94CV N mm3 Volume (Cortical Parcellation) of RightLateralOccipital
  \item ST94SA N mm2 Surface Area of RightLateralOccipital
  \item ST94TA N mm Cortical Thickness Average of RightLateralOccipital
  \item ST94TS N mm Cortical Thickness Standard Deviation of RightLateralOccipital
  \item ST95CV N mm3 Volume (Cortical Parcellation) of RightLateralOrbitofrontal
  \item ST95SA N mm2 Surface Area of RightLateralOrbitofrontal
  \item ST95TA N mm Cortical Thickness Average of RightLateralOrbitofrontal
  \item ST95TS N mm Cortical Thickness Standard Deviation of RightLateralOrbitofrontal
  \item ST96SV N mm3 Volume (WM Parcellation) of RightLateralVentricle
  \item ST97CV N mm3 Volume (Cortical Parcellation) of RightLingual
  \item ST97SA N mm2 Surface Area of RightLingual
  \item ST97TA N mm Cortical Thickness Average of RightLingual
  \item ST97TS N mm Cortical Thickness Standard Deviation of RightLingual
  \item ST98CV N mm3 Volume (Cortical Parcellation) of RightMedialOrbitofrontal
  \item ST98SA N mm2 Surface Area of RightMedialOrbitofrontal
  \item ST98TA N mm Cortical Thickness Average of RightMedialOrbitofrontal
  \item ST98TS N mm Cortical Thickness Standard Deviation of RightMedialOrbitofrontal
  \item ST99CV N mm3 Volume (Cortical Parcellation) of RightMiddleTemporal
  \item ST99SA N mm2 Surface Area of RightMiddleTemporal
  \item ST99TA N mm Cortical Thickness Average of RightMiddleTemporal
  \item ST99TS N mm Cortical Thickness Standard Deviation of RightMiddleTemporal
  \item ST9SV N mm3 Volume (WM Parcellation) of FourthVentricle
}
}
\keyword{datasets}
