% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{ucsffsx51final}
\alias{ucsffsx51final}
\title{Longitudinal FreeSurfer (5.1) - Final Run w/ All A}
\format{
A data frame with 3312 rows and 363 variables
}
\usage{
data(ucsffsx51final)
}
\description{
Longitudinal FreeSurfer (5.1) - Final Run w/ All A
}
\details{
A dataset with variables as follows:

\itemize{
  \item COLPROT -4 -4 Study protocol of data collection
  \item RID N  Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE N  Examination Date
  \item VERSION N  
  \item LONISID N  The LONI study UID for the image processed
  \item LONIUID N  The LONI series UID for the image processed.
  \item IMAGEUID N  The ImageUID for the image processed.
  \item RUNDATE N  
  \item STATUS N  
  \item OVERALLQC N  An overall quality rating. Refer to additional QC variables for Partial rating.
  \item TEMPQC N  QC rating of temporal lobe. Fail affects the following regions: LeftTemporalPole (ST60); RightTemporalPole (ST119); LeftFusiform (ST26); RightFusiform (ST85); LeftSuperiorTemporal (ST58); RightSuperiorTemporal (ST117); LeftInferiorTemporal (ST32); RightInferiorTemporal (ST91); LeftMiddleTemporal (ST40); RightMiddleTemporal (ST99)
  \item FRONTQC N  QC rating of the frontal lobe. Fail affects the following regions: LeftFrontalPole (ST25); RightFrontalPole (ST84); LeftPrecentral (ST51); RightPrecentral (ST110); LeftSuperiorFrontal (ST56); RightSuperiorFrontal (ST115); LeftCaudalMiddleFrontal (ST15); RightCaudalMiddleFrontal (ST74); LeftRostralMiddleFrontal (ST55); RightRostralMiddleFrontal (ST114); LeftMedialOrbitofrontal (ST39); RightMedialOrbitofrontal (ST98)
  \item PARQC N  QC rating of the parietal lobe. Fail affects the following regions: LeftPostcentral (ST49); RightPostcentral (ST108); LeftSuperiorParietal (ST57); RightSuperiorParietal (ST116); LeftSupraMarginal (ST59); RightSupraMarginal (ST118); LeftParacentral (ST43); RightParacentral (ST102); LeftInferiorParietal (ST31); RightInferiorParietal (ST90)
  \item INSULAQC N  QC rating of the insula. Fail affects the following regions: LeftInsula (ST129); RightInsula (ST130)
  \item OCCQC N  QC rating of the occipital lobe. Fail affects the following regions: LeftLingual (ST38); RightLingual (ST97); LeftLateralOccipital (ST35); RightLateralOccipital (ST94); LeftPericalcarine (ST48); RightPericalcarine (ST107); LeftCuneus (ST23); RightCuneus (ST82)
  \item BGQC N  QC rating of the basal ganglia. Fail affects the following regions: LeftPutamen (ST53); RightPutamen (ST112); LeftCaudate (ST16); RightCaudate (ST75); LeftPallidum (ST42); RightPallidum (ST101)
  \item CWMQC N  QC rating of the cerebral WM. Fail affects the following regions: LeftCerebralWM (ST20); RightCerebralWM (ST79); LeftCorticalWM (ST150); RightCorticalWM (ST151); CorticalWM (ST152); LeftHemisphereWM (ST28); RightHemisphereWM (ST87)
  \item VENTQC N  QC rating of the ventricles. Fail affects the following regions: LeftLateralVentricle (ST37); RightLateralVentricle (ST96); LeftInferiorLateralVentricle (ST30); RightInferiorLateralVentricle (ST89); ThirdVentricle (ST127); FourthVentricle (ST9); FifthVentrical (ST8); LeftChoroidPlexus (ST21); RightChoroidPlexus (ST80)
  \item LHIPQC N  QC rating of the left hippocampal subfields. Fail affects the following regions: Left CA1 (ST131HS); Left CA2-3 (ST132HS); Left CA4-DG (ST133HS); Left Fimbria (ST134HS); Left Hippocampal Fissure (ST135HS); Left Presubiculum (ST136HS); Left Subiculum (ST137HS); Left Tail (ST138HS) 
  \item RHIPQC N  QC rating of the right hippocampal subfields. Fail affects the following regions: Right CA1 (ST139HS); Right CA2-3 (ST140HS); Right CA4-DG (ST141HS); Right Fimbria (ST142HS); Right Hippocampal Fissure (ST143HS); Right Presubiculum (ST144HS); Right Subiculum (ST145HS); Right Tail (ST146HS) 
  \item ST149SV N mm3 Subcortical Volume (aseg.stats) of CorticalGM
  \item ST28SA N mm2 Surface Area (aparc.stats) of LeftHemisphereWM
  \item ST87SA N mm2 Surface Area (aparc.stats) of RightHemisphereWM
  \item ST1SV N mm3 Subcortical Volume (aseg.stats) of Brainstem
  \item ST2SV N mm3 Subcortical Volume (aseg.stats) of CorpusCallosumAnterior
  \item ST3SV N mm3 Subcortical Volume (aseg.stats) of CorpusCallosumCentral
  \item ST4SV N mm3 Subcortical Volume (aseg.stats) of CorpusCallosumMidAnterior
  \item ST5SV N mm3 Subcortical Volume (aseg.stats) of CorpusCallosumMidPosterior
  \item ST6SV N mm3 Subcortical Volume (aseg.stats) of CorpusCallosumPosterior
  \item ST7SV N mm3 Subcortical Volume (aseg.stats) of Csf
  \item ST8SV N mm3 Subcortical Volume (aseg.stats) of FifthVentricle
  \item ST9SV N mm3 Subcortical Volume (aseg.stats) of FourthVentricle
  \item ST10CV N mm3 Cortical Volume (aparc.stats) of Icv
  \item ST11SV N mm3 Subcortical Volume (aseg.stats) of LeftAccumbensArea
  \item ST12SV N mm3 Subcortical Volume (aseg.stats) of LeftAmygdala
  \item ST13SA N mm2 Surface Area (aparc.stats) of LeftBankssts
  \item ST13TA N mm Thickness Average (aparc.stats) of LeftBankssts
  \item ST13TS N mm Thickness Stardard Deviation (aparc.stats) of LeftBankssts
  \item ST13CV N mm3 Cortical Volume (aparc.stats) of LeftBankssts
  \item ST14SA N mm2 Surface Area (aparc.stats) of LeftCaudalAnteriorCingulate
  \item ST14TA N mm Thickness Average (aparc.stats) of LeftCaudalAnteriorCingulate
  \item ST14TS N mm Thickness Stardard Deviation (aparc.stats) of LeftCaudalAnteriorCingulate
  \item ST14CV N mm3 Cortical Volume (aparc.stats) of LeftCaudalAnteriorCingulate
  \item ST15SA N mm2 Surface Area (aparc.stats) of LeftCaudalMiddleFrontal
  \item ST15TA N mm Thickness Average (aparc.stats) of LeftCaudalMiddleFrontal
  \item ST15TS N mm Thickness Stardard Deviation (aparc.stats) of LeftCaudalMiddleFrontal
  \item ST15CV N mm3 Cortical Volume (aparc.stats) of LeftCaudalMiddleFrontal
  \item ST16SV N mm3 Subcortical Volume (aseg.stats) of LeftCaudate
  \item ST17SV N mm3 Subcortical Volume (aseg.stats) of LeftCerebellumCortex
  \item ST18SV N mm3 Subcortical Volume (aseg.stats) of LeftCerebellumWM
  \item ST21SV N mm3 Subcortical Volume (aseg.stats) of LeftChoroidPlexus
  \item ST23SA N mm2 Surface Area (aparc.stats) of LeftCuneus
  \item ST23TA N mm Thickness Average (aparc.stats) of LeftCuneus
  \item ST23TS N mm Thickness Stardard Deviation (aparc.stats) of LeftCuneus
  \item ST23CV N mm3 Cortical Volume (aparc.stats) of LeftCuneus
  \item ST24SA N mm2 Surface Area (aparc.stats) of LeftEntorhinal
  \item ST24TA N mm Thickness Average (aparc.stats) of LeftEntorhinal
  \item ST24TS N mm Thickness Stardard Deviation (aparc.stats) of LeftEntorhinal
  \item ST24CV N mm3 Cortical Volume (aparc.stats) of LeftEntorhinal
  \item ST25SA N mm2 Surface Area (aparc.stats) of LeftFrontalPole
  \item ST25TA N mm Thickness Average (aparc.stats) of LeftFrontalPole
  \item ST25TS N mm Thickness Stardard Deviation (aparc.stats) of LeftFrontalPole
  \item ST25CV N mm3 Cortical Volume (aparc.stats) of LeftFrontalPole
  \item ST26SA N mm2 Surface Area (aparc.stats) of LeftFusiform
  \item ST26TA N mm Thickness Average (aparc.stats) of LeftFusiform
  \item ST26TS N mm Thickness Stardard Deviation (aparc.stats) of LeftFusiform
  \item ST26CV N mm3 Cortical Volume (aparc.stats) of LeftFusiform
  \item ST29SV N mm3 Subcortical Volume (aseg.stats) of LeftHippocampus
  \item ST30SV N mm3 Subcortical Volume (aseg.stats) of LeftInferiorLateralVentricle
  \item ST31SA N mm2 Surface Area (aparc.stats) of LeftInferiorParietal
  \item ST31TA N mm Thickness Average (aparc.stats) of LeftInferiorParietal
  \item ST31TS N mm Thickness Stardard Deviation (aparc.stats) of LeftInferiorParietal
  \item ST31CV N mm3 Cortical Volume (aparc.stats) of LeftInferiorParietal
  \item ST32SA N mm2 Surface Area (aparc.stats) of LeftInferiorTemporal
  \item ST32TA N mm Thickness Average (aparc.stats) of LeftInferiorTemporal
  \item ST32TS N mm Thickness Stardard Deviation (aparc.stats) of LeftInferiorTemporal
  \item ST32CV N mm3 Cortical Volume (aparc.stats) of LeftInferiorTemporal
  \item ST34SA N mm2 Surface Area (aparc.stats) of LeftIsthmusCingulate
  \item ST34TA N mm Thickness Average (aparc.stats) of LeftIsthmusCingulate
  \item ST34TS N mm Thickness Stardard Deviation (aparc.stats) of LeftIsthmusCingulate
  \item ST34CV N mm3 Cortical Volume (aparc.stats) of LeftIsthmusCingulate
  \item ST35SA N mm2 Surface Area (aparc.stats) of LeftLateralOccipital
  \item ST35TA N mm Thickness Average (aparc.stats) of LeftLateralOccipital
  \item ST35TS N mm Thickness Stardard Deviation (aparc.stats) of LeftLateralOccipital
  \item ST35CV N mm3 Cortical Volume (aparc.stats) of LeftLateralOccipital
  \item ST36SA N mm2 Surface Area (aparc.stats) of LeftLateralOrbitofrontal
  \item ST36TA N mm Thickness Average (aparc.stats) of LeftLateralOrbitofrontal
  \item ST36TS N mm Thickness Stardard Deviation (aparc.stats) of LeftLateralOrbitofrontal
  \item ST36CV N mm3 Cortical Volume (aparc.stats) of LeftLateralOrbitofrontal
  \item ST37SV N mm3 Subcortical Volume (aseg.stats) of LeftLateralVentricle
  \item ST38SA N mm2 Surface Area (aparc.stats) of LeftLingual
  \item ST38TA N mm Thickness Average (aparc.stats) of LeftLingual
  \item ST38TS N mm Thickness Stardard Deviation (aparc.stats) of LeftLingual
  \item ST38CV N mm3 Cortical Volume (aparc.stats) of LeftLingual
  \item ST39SA N mm2 Surface Area (aparc.stats) of LeftMedialOrbitofrontal
  \item ST39TA N mm Thickness Average (aparc.stats) of LeftMedialOrbitofrontal
  \item ST39TS N mm Thickness Stardard Deviation (aparc.stats) of LeftMedialOrbitofrontal
  \item ST39CV N mm3 Cortical Volume (aparc.stats) of LeftMedialOrbitofrontal
  \item ST40SA N mm2 Surface Area (aparc.stats) of LeftMiddleTemporal
  \item ST40TA N mm Thickness Average (aparc.stats) of LeftMiddleTemporal
  \item ST40TS N mm Thickness Stardard Deviation (aparc.stats) of LeftMiddleTemporal
  \item ST40CV N mm3 Cortical Volume (aparc.stats) of LeftMiddleTemporal
  \item ST42SV N mm3 Subcortical Volume (aseg.stats) of LeftPallidum
  \item ST43SA N mm2 Surface Area (aparc.stats) of LeftParacentral
  \item ST43TA N mm Thickness Average (aparc.stats) of LeftParacentral
  \item ST43TS N mm Thickness Stardard Deviation (aparc.stats) of LeftParacentral
  \item ST43CV N mm3 Cortical Volume (aparc.stats) of LeftParacentral
  \item ST44SA N mm2 Surface Area (aparc.stats) of LeftParahippocampal
  \item ST44TA N mm Thickness Average (aparc.stats) of LeftParahippocampal
  \item ST44TS N mm Thickness Stardard Deviation (aparc.stats) of LeftParahippocampal
  \item ST44CV N mm3 Cortical Volume (aparc.stats) of LeftParahippocampal
  \item ST45SA N mm2 Surface Area (aparc.stats) of LeftParsOpercularis
  \item ST45TA N mm Thickness Average (aparc.stats) of LeftParsOpercularis
  \item ST45TS N mm Thickness Stardard Deviation (aparc.stats) of LeftParsOpercularis
  \item ST45CV N mm3 Cortical Volume (aparc.stats) of LeftParsOpercularis
  \item ST46SA N mm2 Surface Area (aparc.stats) of LeftParsOrbitalis
  \item ST46TA N mm Thickness Average (aparc.stats) of LeftParsOrbitalis
  \item ST46TS N mm Thickness Stardard Deviation (aparc.stats) of LeftParsOrbitalis
  \item ST46CV N mm3 Cortical Volume (aparc.stats) of LeftParsOrbitalis
  \item ST47SA N mm2 Surface Area (aparc.stats) of LeftParsTriangularis
  \item ST47TA N mm Thickness Average (aparc.stats) of LeftParsTriangularis
  \item ST47TS N mm Thickness Stardard Deviation (aparc.stats) of LeftParsTriangularis
  \item ST47CV N mm3 Cortical Volume (aparc.stats) of LeftParsTriangularis
  \item ST48SA N mm2 Surface Area (aparc.stats) of LeftPericalcarine
  \item ST48TA N mm Thickness Average (aparc.stats) of LeftPericalcarine
  \item ST48TS N mm Thickness Stardard Deviation (aparc.stats) of LeftPericalcarine
  \item ST48CV N mm3 Cortical Volume (aparc.stats) of LeftPericalcarine
  \item ST49SA N mm2 Surface Area (aparc.stats) of LeftPostcentral
  \item ST49TA N mm Thickness Average (aparc.stats) of LeftPostcentral
  \item ST49TS N mm Thickness Stardard Deviation (aparc.stats) of LeftPostcentral
  \item ST49CV N mm3 Cortical Volume (aparc.stats) of LeftPostcentral
  \item ST50SA N mm2 Surface Area (aparc.stats) of LeftPosteriorCingulate
  \item ST50TA N mm Thickness Average (aparc.stats) of LeftPosteriorCingulate
  \item ST50TS N mm Thickness Stardard Deviation (aparc.stats) of LeftPosteriorCingulate
  \item ST50CV N mm3 Cortical Volume (aparc.stats) of LeftPosteriorCingulate
  \item ST51SA N mm2 Surface Area (aparc.stats) of LeftPrecentral
  \item ST51TA N mm Thickness Average (aparc.stats) of LeftPrecentral
  \item ST51TS N mm Thickness Stardard Deviation (aparc.stats) of LeftPrecentral
  \item ST51CV N mm3 Cortical Volume (aparc.stats) of LeftPrecentral
  \item ST52SA N mm2 Surface Area (aparc.stats) of LeftPrecuneus
  \item ST52TA N mm Thickness Average (aparc.stats) of LeftPrecuneus
  \item ST52TS N mm Thickness Stardard Deviation (aparc.stats) of LeftPrecuneus
  \item ST52CV N mm3 Cortical Volume (aparc.stats) of LeftPrecuneus
  \item ST53SV N mm3 Subcortical Volume (aseg.stats) of LeftPutamen
  \item ST54SA N mm2 Surface Area (aparc.stats) of LeftRostralAnteriorCingulate
  \item ST54TA N mm Thickness Average (aparc.stats) of LeftRostralAnteriorCingulate
  \item ST54TS N mm Thickness Stardard Deviation (aparc.stats) of LeftRostralAnteriorCingulate
  \item ST54CV N mm3 Cortical Volume (aparc.stats) of LeftRostralAnteriorCingulate
  \item ST55SA N mm2 Surface Area (aparc.stats) of LeftRostralMiddleFrontal
  \item ST55TA N mm Thickness Average (aparc.stats) of LeftRostralMiddleFrontal
  \item ST55TS N mm Thickness Stardard Deviation (aparc.stats) of LeftRostralMiddleFrontal
  \item ST55CV N mm3 Cortical Volume (aparc.stats) of LeftRostralMiddleFrontal
  \item ST56SA N mm2 Surface Area (aparc.stats) of LeftSuperiorFrontal
  \item ST56TA N mm Thickness Average (aparc.stats) of LeftSuperiorFrontal
  \item ST56TS N mm Thickness Stardard Deviation (aparc.stats) of LeftSuperiorFrontal
  \item ST56CV N mm3 Cortical Volume (aparc.stats) of LeftSuperiorFrontal
  \item ST57SA N mm2 Surface Area (aparc.stats) of LeftSuperiorParietal
  \item ST57TA N mm Thickness Average (aparc.stats) of LeftSuperiorParietal
  \item ST57TS N mm Thickness Stardard Deviation (aparc.stats) of LeftSuperiorParietal
  \item ST57CV N mm3 Cortical Volume (aparc.stats) of LeftSuperiorParietal
  \item ST58SA N mm2 Surface Area (aparc.stats) of LeftSuperiorTemporal
  \item ST58TA N mm Thickness Average (aparc.stats) of LeftSuperiorTemporal
  \item ST58TS N mm Thickness Stardard Deviation (aparc.stats) of LeftSuperiorTemporal
  \item ST58CV N mm3 Cortical Volume (aparc.stats) of LeftSuperiorTemporal
  \item ST59SA N mm2 Surface Area (aparc.stats) of LeftSupramarginal
  \item ST59TA N mm Thickness Average (aparc.stats) of LeftSupramarginal
  \item ST59TS N mm Thickness Stardard Deviation (aparc.stats) of LeftSupramarginal
  \item ST59CV N mm3 Cortical Volume (aparc.stats) of LeftSupramarginal
  \item ST60SA N mm2 Surface Area (aparc.stats) of LeftTemporalPole
  \item ST60TA N mm Thickness Average (aparc.stats) of LeftTemporalPole
  \item ST60TS N mm Thickness Stardard Deviation (aparc.stats) of LeftTemporalPole
  \item ST60CV N mm3 Cortical Volume (aparc.stats) of LeftTemporalPole
  \item ST61SV N mm3 Subcortical Volume (aseg.stats) of LeftThalamus
  \item ST62SA N mm2 Surface Area (aparc.stats) of LeftTransverseTemporal
  \item ST62TA N mm Thickness Average (aparc.stats) of LeftTransverseTemporal
  \item ST62TS N mm Thickness Stardard Deviation (aparc.stats) of LeftTransverseTemporal
  \item ST62CV N mm3 Cortical Volume (aparc.stats) of LeftTransverseTemporal
  \item ST65SV N mm3 Subcortical Volume (aseg.stats) of LeftVentralDC
  \item ST66SV N mm3 Subcortical Volume (aseg.stats) of LeftVessel
  \item ST68SV N mm3 Subcortical Volume (aseg.stats) of NonWMHypoIntensities
  \item ST69SV N mm3 Subcortical Volume (aseg.stats) of OpticChiasm
  \item ST70SV N mm3 Subcortical Volume (aseg.stats) of RightAccumbensArea
  \item ST71SV N mm3 Subcortical Volume (aseg.stats) of RightAmygdala
  \item ST72SA N mm2 Surface Area (aparc.stats) of RightBankssts
  \item ST72TA N mm Thickness Average (aparc.stats) of RightBankssts
  \item ST72TS N mm Thickness Stardard Deviation (aparc.stats) of RightBankssts
  \item ST72CV N mm3 Cortical Volume (aparc.stats) of RightBankssts
  \item ST73SA N mm2 Surface Area (aparc.stats) of RightCaudalAnteriorCingulate
  \item ST73TA N mm Thickness Average (aparc.stats) of RightCaudalAnteriorCingulate
  \item ST73TS N mm Thickness Stardard Deviation (aparc.stats) of RightCaudalAnteriorCingulate
  \item ST73CV N mm3 Cortical Volume (aparc.stats) of RightCaudalAnteriorCingulate
  \item ST74SA N mm2 Surface Area (aparc.stats) of RightCaudalMiddleFrontal
  \item ST74TA N mm Thickness Average (aparc.stats) of RightCaudalMiddleFrontal
  \item ST74TS N mm Thickness Stardard Deviation (aparc.stats) of RightCaudalMiddleFrontal
  \item ST74CV N mm3 Cortical Volume (aparc.stats) of RightCaudalMiddleFrontal
  \item ST75SV N mm3 Subcortical Volume (aseg.stats) of RightCaudate
  \item ST76SV N mm3 Subcortical Volume (aseg.stats) of RightCerebellumCortex
  \item ST77SV N mm3 Subcortical Volume (aseg.stats) of RightCerebellumWM
  \item ST80SV N mm3 Subcortical Volume (aseg.stats) of RightChoroidPlexus
  \item ST82SA N mm2 Surface Area (aparc.stats) of RightCuneus
  \item ST82TA N mm Thickness Average (aparc.stats) of RightCuneus
  \item ST82TS N mm Thickness Stardard Deviation (aparc.stats) of RightCuneus
  \item ST82CV N mm3 Cortical Volume (aparc.stats) of RightCuneus
  \item ST83SA N mm2 Surface Area (aparc.stats) of RightEntorhinal
  \item ST83TA N mm Thickness Average (aparc.stats) of RightEntorhinal
  \item ST83TS N mm Thickness Stardard Deviation (aparc.stats) of RightEntorhinal
  \item ST83CV N mm3 Cortical Volume (aparc.stats) of RightEntorhinal
  \item ST84SA N mm2 Surface Area (aparc.stats) of RightFrontalPole
  \item ST84TA N mm Thickness Average (aparc.stats) of RightFrontalPole
  \item ST84TS N mm Thickness Stardard Deviation (aparc.stats) of RightFrontalPole
  \item ST84CV N mm3 Cortical Volume (aparc.stats) of RightFrontalPole
  \item ST85SA N mm2 Surface Area (aparc.stats) of RightFusiform
  \item ST85TA N mm Thickness Average (aparc.stats) of RightFusiform
  \item ST85TS N mm Thickness Stardard Deviation (aparc.stats) of RightFusiform
  \item ST85CV N mm3 Cortical Volume (aparc.stats) of RightFusiform
  \item ST88SV N mm3 Subcortical Volume (aseg.stats) of RightHippocampus
  \item ST89SV N mm3 Subcortical Volume (aseg.stats) of RightInferiorLateralVentricle
  \item ST90SA N mm2 Surface Area (aparc.stats) of RightInferiorParietal
  \item ST90TA N mm Thickness Average (aparc.stats) of RightInferiorParietal
  \item ST90TS N mm Thickness Stardard Deviation (aparc.stats) of RightInferiorParietal
  \item ST90CV N mm3 Cortical Volume (aparc.stats) of RightInferiorParietal
  \item ST91SA N mm2 Surface Area (aparc.stats) of RightInferiorTemporal
  \item ST91TA N mm Thickness Average (aparc.stats) of RightInferiorTemporal
  \item ST91TS N mm Thickness Stardard Deviation (aparc.stats) of RightInferiorTemporal
  \item ST91CV N mm3 Cortical Volume (aparc.stats) of RightInferiorTemporal
  \item ST93SA N mm2 Surface Area (aparc.stats) of RightIsthmusCingulate
  \item ST93TA N mm Thickness Average (aparc.stats) of RightIsthmusCingulate
  \item ST93TS N mm Thickness Stardard Deviation (aparc.stats) of RightIsthmusCingulate
  \item ST93CV N mm3 Cortical Volume (aparc.stats) of RightIsthmusCingulate
  \item ST94SA N mm2 Surface Area (aparc.stats) of RightLateralOccipital
  \item ST94TA N mm Thickness Average (aparc.stats) of RightLateralOccipital
  \item ST94TS N mm Thickness Stardard Deviation (aparc.stats) of RightLateralOccipital
  \item ST94CV N mm3 Cortical Volume (aparc.stats) of RightLateralOccipital
  \item ST95SA N mm2 Surface Area (aparc.stats) of RightLateralOrbitofrontal
  \item ST95TA N mm Thickness Average (aparc.stats) of RightLateralOrbitofrontal
  \item ST95TS N mm Thickness Stardard Deviation (aparc.stats) of RightLateralOrbitofrontal
  \item ST95CV N mm3 Cortical Volume (aparc.stats) of RightLateralOrbitofrontal
  \item ST96SV N mm3 Subcortical Volume (aseg.stats) of RightLateralVentricle
  \item ST97SA N mm2 Surface Area (aparc.stats) of RightLingual
  \item ST97TA N mm Thickness Average (aparc.stats) of RightLingual
  \item ST97TS N mm Thickness Stardard Deviation (aparc.stats) of RightLingual
  \item ST97CV N mm3 Cortical Volume (aparc.stats) of RightLingual
  \item ST98SA N mm2 Surface Area (aparc.stats) of RightMedialOrbitofrontal
  \item ST98TA N mm Thickness Average (aparc.stats) of RightMedialOrbitofrontal
  \item ST98TS N mm Thickness Stardard Deviation (aparc.stats) of RightMedialOrbitofrontal
  \item ST98CV N mm3 Cortical Volume (aparc.stats) of RightMedialOrbitofrontal
  \item ST99SA N mm2 Surface Area (aparc.stats) of RightMiddleTemporal
  \item ST99TA N mm Thickness Average (aparc.stats) of RightMiddleTemporal
  \item ST99TS N mm Thickness Stardard Deviation (aparc.stats) of RightMiddleTemporal
  \item ST99CV N mm3 Cortical Volume (aparc.stats) of RightMiddleTemporal
  \item ST101SV N mm3 Subcortical Volume (aseg.stats) of RightPallidum
  \item ST102SA N mm2 Surface Area (aparc.stats) of RightParacentral
  \item ST102TA N mm Thickness Average (aparc.stats) of RightParacentral
  \item ST102TS N mm Thickness Stardard Deviation (aparc.stats) of RightParacentral
  \item ST102CV N mm3 Cortical Volume (aparc.stats) of RightParacentral
  \item ST103SA N mm2 Surface Area (aparc.stats) of RightParahippocampal
  \item ST103TA N mm Thickness Average (aparc.stats) of RightParahippocampal
  \item ST103TS N mm Thickness Stardard Deviation (aparc.stats) of RightParahippocampal
  \item ST103CV N mm3 Cortical Volume (aparc.stats) of RightParahippocampal
  \item ST104SA N mm2 Surface Area (aparc.stats) of RightParsOpercularis
  \item ST104TA N mm Thickness Average (aparc.stats) of RightParsOpercularis
  \item ST104TS N mm Thickness Stardard Deviation (aparc.stats) of RightParsOpercularis
  \item ST104CV N mm3 Cortical Volume (aparc.stats) of RightParsOpercularis
  \item ST105SA N mm2 Surface Area (aparc.stats) of RightParsOrbitalis
  \item ST105TA N mm Thickness Average (aparc.stats) of RightParsOrbitalis
  \item ST105TS N mm Thickness Stardard Deviation (aparc.stats) of RightParsOrbitalis
  \item ST105CV N mm3 Cortical Volume (aparc.stats) of RightParsOrbitalis
  \item ST106SA N mm2 Surface Area (aparc.stats) of RightParsTriangularis
  \item ST106TA N mm Thickness Average (aparc.stats) of RightParsTriangularis
  \item ST106TS N mm Thickness Stardard Deviation (aparc.stats) of RightParsTriangularis
  \item ST106CV N mm3 Cortical Volume (aparc.stats) of RightParsTriangularis
  \item ST107SA N mm2 Surface Area (aparc.stats) of RightPericalcarine
  \item ST107TA N mm Thickness Average (aparc.stats) of RightPericalcarine
  \item ST107TS N mm Thickness Stardard Deviation (aparc.stats) of RightPericalcarine
  \item ST107CV N mm3 Cortical Volume (aparc.stats) of RightPericalcarine
  \item ST108SA N mm2 Surface Area (aparc.stats) of RightPostcentral
  \item ST108TA N mm Thickness Average (aparc.stats) of RightPostcentral
  \item ST108TS N mm Thickness Stardard Deviation (aparc.stats) of RightPostcentral
  \item ST108CV N mm3 Cortical Volume (aparc.stats) of RightPostcentral
  \item ST109SA N mm2 Surface Area (aparc.stats) of RightPosteriorCingulate
  \item ST109TA N mm Thickness Average (aparc.stats) of RightPosteriorCingulate
  \item ST109TS N mm Thickness Stardard Deviation (aparc.stats) of RightPosteriorCingulate
  \item ST109CV N mm3 Cortical Volume (aparc.stats) of RightPosteriorCingulate
  \item ST110SA N mm2 Surface Area (aparc.stats) of RightPrecentral
  \item ST110TA N mm Thickness Average (aparc.stats) of RightPrecentral
  \item ST110TS N mm Thickness Stardard Deviation (aparc.stats) of RightPrecentral
  \item ST110CV N mm3 Cortical Volume (aparc.stats) of RightPrecentral
  \item ST111SA N mm2 Surface Area (aparc.stats) of RightPrecuneus
  \item ST111TA N mm Thickness Average (aparc.stats) of RightPrecuneus
  \item ST111TS N mm Thickness Stardard Deviation (aparc.stats) of RightPrecuneus
  \item ST111CV N mm3 Cortical Volume (aparc.stats) of RightPrecuneus
  \item ST112SV N mm3 Subcortical Volume (aseg.stats) of RightPutamen
  \item ST113SA N mm2 Surface Area (aparc.stats) of RightRostralAnteriorCingulate
  \item ST113TA N mm Thickness Average (aparc.stats) of RightRostralAnteriorCingulate
  \item ST113TS N mm Thickness Stardard Deviation (aparc.stats) of RightRostralAnteriorCingulate
  \item ST113CV N mm3 Cortical Volume (aparc.stats) of RightRostralAnteriorCingulate
  \item ST114SA N mm2 Surface Area (aparc.stats) of RightRostralMiddleFrontal
  \item ST114TA N mm Thickness Average (aparc.stats) of RightRostralMiddleFrontal
  \item ST114TS N mm Thickness Stardard Deviation (aparc.stats) of RightRostralMiddleFrontal
  \item ST114CV N mm3 Cortical Volume (aparc.stats) of RightRostralMiddleFrontal
  \item ST115SA N mm2 Surface Area (aparc.stats) of RightSuperiorFrontal
  \item ST115TA N mm Thickness Average (aparc.stats) of RightSuperiorFrontal
  \item ST115TS N mm Thickness Stardard Deviation (aparc.stats) of RightSuperiorFrontal
  \item ST115CV N mm3 Cortical Volume (aparc.stats) of RightSuperiorFrontal
  \item ST116SA N mm2 Surface Area (aparc.stats) of RightSuperiorParietal
  \item ST116TA N mm Thickness Average (aparc.stats) of RightSuperiorParietal
  \item ST116TS N mm Thickness Stardard Deviation (aparc.stats) of RightSuperiorParietal
  \item ST116CV N mm3 Cortical Volume (aparc.stats) of RightSuperiorParietal
  \item ST117SA N mm2 Surface Area (aparc.stats) of RightSuperiorTemporal
  \item ST117TA N mm Thickness Average (aparc.stats) of RightSuperiorTemporal
  \item ST117TS N mm Thickness Stardard Deviation (aparc.stats) of RightSuperiorTemporal
  \item ST117CV N mm3 Cortical Volume (aparc.stats) of RightSuperiorTemporal
  \item ST118SA N mm2 Surface Area (aparc.stats) of RightSupramarginal
  \item ST118TA N mm Thickness Average (aparc.stats) of RightSupramarginal
  \item ST118TS N mm Thickness Stardard Deviation (aparc.stats) of RightSupramarginal
  \item ST118CV N mm3 Cortical Volume (aparc.stats) of RightSupramarginal
  \item ST119SA N mm2 Surface Area (aparc.stats) of RightTemporalPole
  \item ST119TA N mm Thickness Average (aparc.stats) of RightTemporalPole
  \item ST119TS N mm Thickness Stardard Deviation (aparc.stats) of RightTemporalPole
  \item ST119CV N mm3 Cortical Volume (aparc.stats) of RightTemporalPole
  \item ST120SV N mm3 Subcortical Volume (aseg.stats) of RightThalamus
  \item ST121SA N mm2 Surface Area (aparc.stats) of RightTransverseTemporal
  \item ST121TA N mm Thickness Average (aparc.stats) of RightTransverseTemporal
  \item ST121TS N mm Thickness Stardard Deviation (aparc.stats) of RightTransverseTemporal
  \item ST121CV N mm3 Cortical Volume (aparc.stats) of RightTransverseTemporal
  \item ST124SV N mm3 Subcortical Volume (aseg.stats) of RightVentralDC
  \item ST125SV N mm3 Subcortical Volume (aseg.stats) of RightVessel
  \item ST127SV N mm3 Subcortical Volume (aseg.stats) of ThirdVentricle
  \item ST128SV N mm3 Subcortical Volume (aseg.stats) of WMHypoIntensities
  \item ST129SA N mm2 Surface Area (aparc.stats) of LeftInsula
  \item ST129TA N mm Thickness Average (aparc.stats) of LeftInsula
  \item ST129TS N mm Thickness Stardard Deviation (aparc.stats) of LeftInsula
  \item ST129CV N mm3 Cortical Volume (aparc.stats) of LeftInsula
  \item ST130SA N mm2 Surface Area (aparc.stats) of RightInsula
  \item ST130TA N mm Thickness Average (aparc.stats) of RightInsula
  \item ST130TS N mm Thickness Stardard Deviation (aparc.stats) of RightInsula
  \item ST130CV N mm3 Cortical Volume (aparc.stats) of RightInsula
  \item ST131HS N mm3 Hippocampal Subfields Volume of LeftCA1
  \item ST132HS N mm3 Hippocampal Subfields Volume of LeftCA2_3
  \item ST133HS N mm3 Hippocampal Subfields Volume of LeftCA4_DG
  \item ST134HS N mm3 Hippocampal Subfields Volume of LeftFimbria
  \item ST135HS N mm3 Hippocampal Subfields Volume of LeftHippocampalFissure
  \item ST136HS N mm3 Hippocampal Subfields Volume of LeftPresubiculum
  \item ST137HS N mm3 Hippocampal Subfields Volume of LeftSubiculum
  \item ST138HS N mm3 Hippocampal Subfields Volume of LeftTail
  \item ST139HS N mm3 Hippocampal Subfields Volume of RightCA1
  \item ST140HS N mm3 Hippocampal Subfields Volume of RightCA2_3
  \item ST141HS N mm3 Hippocampal Subfields Volume of RightCA4_DG
  \item ST142HS N mm3 Hippocampal Subfields Volume of RightFimbria
  \item ST143HS N mm3 Hippocampal Subfields Volume of RightHippocampalFissure
  \item ST144HS N mm3 Hippocampal Subfields Volume of RightPresubiculum
  \item ST145HS N mm3 Hippocampal Subfields Volume of RightSubiculum
  \item ST146HS N mm3 Hippocampal Subfields Volume of RightTail
  \item ST147SV N mm3 Subcortical Volume (aseg.stats) of LeftCorticalGM
  \item ST148SV N mm3 Subcortical Volume (aseg.stats) of RightCorticalGM
  \item ST150SV N mm3 Subcortical Volume (aseg.stats) of LeftCorticalWM
  \item ST151SV N mm3 Subcortical Volume (aseg.stats) of RightCorticalWM
  \item ST152SV N mm3 Subcortical Volume (aseg.stats) of CorticalWM
  \item ST153SV N mm3 Subcortical Volume (aseg.stats) of SubcorticalGM
  \item ST154SV N mm3 Subcortical Volume (aseg.stats) of TotalGM
  \item ST155SV N mm3 Subcortical Volume (aseg.stats) of SupraTentorial
}
}
\examples{
\dontrun{
describe(ucsffsx51final)
}
}
\keyword{datasets}
