% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{ucsfsntvol}
\alias{ucsfsntvol}
\title{UCSF SNT Hippocampal Volumes}
\format{
A data frame with 1881 rows and 15 variables
}
\usage{
data(ucsfsntvol)
}
\description{
UCSF SNT Hippocampal Volumes
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item RECNO N  
  \item EXAMDATE D  Examination Date
  \item LONISID T  LONI Study ID
  \item LONIUID T  LONI Unique Series ID
  \item IMAGEUID T  imageUID
  \item FLDSTRENG T  Field Strength
  \item SNTDATE D  Date Volume Generated
  \item ATLAS T  Name of Atlas Used
  \item LEFTHIPPO N mm3 Left Hippocampus Volume
  \item RIGHTHIPPO N mm3 Right Hippocampus Volume
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UCSFSNTVOL/UCSFMRI_Analysis.pdf}
}
\keyword{datasets}
