% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{ugotptau181}
\alias{ugotptau181}
\title{University of Gothenburg Longitudinal plasma P-tau181}
\format{
A data frame with 3758 rows and 6 variables
}
\usage{
data(ugotptau181)
}
\description{
University of Gothenburg Longitudinal plasma P-tau181
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item VID -4 -4 UPENN Visit Code
  \item PLASMAPTAU181 -4 pg/mL Plasma Tau 181 Measurements
}
}
\examples{
\dontrun{
describe(ugotptau181)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UGOTPTAU181/UGOT_Lab_-_ADNI_1-Go-2_-_Method_-_Plasma_P-tau181_longitudinal.pdf}
}
\keyword{datasets}
