% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{upennbiomk}
\alias{upennbiomk}
\title{UPENN Biomarker Data}
\format{
A data frame with 416 rows and 10 variables
}
\usage{
data(upennbiomk)
}
\description{
UPENN Biomarker Data
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item USERDATE2 S  Date record last updated
  \item TAU N  
  \item ABETA142 N  
  \item PTAU181P N  CSF pTau181p concentration data for BASELINE, YEAR1 and YEAR 2 will bequarantined temporarily due to our discovery that for some individualsamples there appears to be greater analytical variability than shouldbe the case. We are working with the manufacturer of the immunoassayreagents we use for the CSF biomarker measurements, Innogenetics, andhave worked out a procedural change that eliminates this extraneousnoise. Once the retesting analyses are completed the new data will beuploaded with accompanying description.ADNI Biomarker Core laboratory at UPenn
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UPENNBIOMK/ADNI_Biomarker_Core_News_April_2009_2.pdf}
}
\keyword{datasets}
