% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{upennbiomk10}
\alias{upennbiomk10}
\title{UPENN CSF Biomarkers Elecsys [ADNI3, batch I]}
\format{
A data frame with 498 rows and 10 variables
}
\usage{
data(upennbiomk10)
}
\description{
UPENN CSF Biomarkers Elecsys [ADNI3, batch I]
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4  Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item DRAWDATE -4  Date of CSF draw
  \item RUNDATE -4  Date of analytical run
  \item ABETA40 -4 pg/ml Result of Abeta40
  \item ABETA42 -4 pg/ml Result of Abeta42
  \item PTAU -4 pg/ml Result of p-Tau
  \item TAU -4 pg/ml Result of Tau
  \item NOTE -4  Comment on missing data
}
}
\examples{
\dontrun{
describe(upennbiomk10)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UPENNBIOMK10/ADNI_Methods_Template_Shaw_2019_Roche_Elecsys_ADNI3_CSFs_Batch_1_v1.pdf}
}
\keyword{datasets}
