% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{upennbiomk12_2020}
\alias{upennbiomk12_2020}
\title{UPENN CSF Biomarkers Elecsys}
\format{
A data frame with 216 rows and 12 variables
}
\usage{
data(upennbiomk12_2020)
}
\description{
UPENN CSF Biomarkers Elecsys
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4  Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4  Examination Date
  \item PROT -4  Study Phase
  \item RUNDATE -4  Date of analytical run
  \item AB40 -4 pg/ml Result: Abeta1-40
  \item ABETA -4 pg/ml Result: Abeta1-42
  \item PTAU -4 pg/ml Result: p-Tau181
  \item TAU -4 pg/ml Result: total Tau
  \item AB4240 -4 - Calculated result: Abeta42/Abeta40
  \item NOTE -4  Reason for missing result
}
}
\examples{
\dontrun{
describe(upennbiomk12_2020)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UPENNBIOMK12_2020/ADNI_UPENNBIOMK12_Methods_2020_Roche_Elecsys_ADNI3_CSFs.pdf}
}
\keyword{datasets}
