% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{upennbiomk2}
\alias{upennbiomk2}
\title{UPENN Longitudinal Biomarker Data}
\format{
A data frame with 834 rows and 7 variables
}
\usage{
data(upennbiomk2)
}
\description{
UPENN Longitudinal Biomarker Data
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N -4 Participant roster ID
  \item SITEID N -4 Site ID
  \item VISCODE T -4 Visit code
  \item USERDATE S -4 Date record created
  \item TTAU N -4 -4
  \item ABETA142 N -4 -4
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UPENNBIOMK2/ADNI_Biomarker_Core_News_April_2009_2.pdf}
}
\keyword{datasets}
