% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{upennbiomk3}
\alias{upennbiomk3}
\title{UPENN Longitudinal Biomarker Data (3 yr)}
\format{
A data frame with 314 rows and 10 variables
}
\usage{
data(upennbiomk3)
}
\description{
UPENN Longitudinal Biomarker Data (3 yr)
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N -4 Participant roster ID
  \item SITEID N -4 Site ID
  \item VISCODE T -4 Visit code
  \item USERDATE S -4 Date record created
  \item ABETA N -4 -4
  \item TAU N -4 -4
  \item PTAU N -4 -4
  \item TAUAB N -4 -4
  \item PTAUAB N -4 -4
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UPENNBIOMK3/ADNI_1_Longitudinal_dataset_overview_12_21_2011_for_ADNI_ADCS_website_upload.pdf}
}
\keyword{datasets}
