% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{upennbiomk4}
\alias{upennbiomk4}
\title{UPENN Longitudinal Biomarker Data (4 yr)}
\format{
A data frame with 495 rows and 10 variables
}
\usage{
data(upennbiomk4)
}
\description{
UPENN Longitudinal Biomarker Data (4 yr)
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item LP.DATE D -4 LP Date
  \item TAU -4 (pg/mL) TAU
  \item ABETA -4 (pg/mL) ABETA
  \item PTAU -4 (pg/mL) PTAU
  \item TAU.ABETA -4 -4 TAU/ABETA
  \item PTAU.ABETA -4 -4 PTAU/ABETA
  \item UPDATE.STAMP -4 -4 Update stamp
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UPENNBIOMK4/ADNI_1_Longitudinal_dataset_overview_12_21_2011_for_ADNI_ADCS_website_upload.pdf}
}
\keyword{datasets}
