% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{upennbiomk5}
\alias{upennbiomk5}
\title{ADNI GO and ADNI 2: first batch analysis of CSF biomarkers}
\format{
A data frame with 389 rows and 10 variables
}
\usage{
data(upennbiomk5)
}
\description{
ADNI GO and ADNI 2: first batch analysis of CSF biomarkers
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item RUNDATE D -4 Run date
  \item ABETA -4 pg/mL ABETA
  \item TAU -4 pg/mL TAU
  \item PTAU -4 pg/mL PTAU
  \item ABETA.RAW -4 -4 -4
  \item TAU.RAW -4 -4 -4
  \item PTAU.RAW -4 -4 -4
}
}
\examples{
\dontrun{
describe(upennbiomk5)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UPENNBIOMK5/ADNI-GO+2_analytical_signed_3_6_28_2012.pdf}, \url{https://adni.bitbucket.io/reference/docs/UPENNBIOMK5/ADNI_Methods_Template_Shaw_Trojanowski_6_30_2012.pdf}, \url{https://adni.bitbucket.io/reference/docs/UPENNBIOMK5/ADNI_Methods_Template_Shaw\%20Trojanowski\%20Corrected\%20Anchoring\%20of\%202012\%20dataset\%2010\%2031\%2013.pdf}
}
\keyword{datasets}
