% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{upennbiomk6}
\alias{upennbiomk6}
\title{ADNI 2: second batch analysis of CSF biomarkers}
\format{
A data frame with 692 rows and 8 variables
}
\usage{
data(upennbiomk6)
}
\description{
ADNI 2: second batch analysis of CSF biomarkers
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID   Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item EXAMDATE D  Examination Date
  \item RUNDATE D  Run date
  \item ABETA  pg/mL ABETA
  \item TAU  pg/mL TAU
  \item PTAU  pg/mL PTAU
}
}
\examples{
\dontrun{
describe(upennbiomk6)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UPENNBIOMK6/ADNI\%20LT2013\%20analytical\%20signed\%20all.pdf}, \url{https://adni.bitbucket.io/reference/docs/UPENNBIOMK6/ADNI_Methods_Shaw\%20Trojanowski\%20for\%202013\%20ADNI\%20II\%207\%201\%202013.pdf}, \url{https://adni.bitbucket.io/reference/docs/UPENNBIOMK6/ADNI_Methods_Template_Shaw\%20Trojanowski\%20Corrected\%20Anchoring\%20of\%202012\%20dataset\%2010\%2031\%2013.pdf}
}
\keyword{datasets}
