% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{upennbiomk7}
\alias{upennbiomk7}
\title{UPENN-Third batch analysis of CSF biomarkers}
\format{
A data frame with 428 rows and 11 variables
}
\usage{
data(upennbiomk7)
}
\description{
UPENN-Third batch analysis of CSF biomarkers
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID   Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE D  Examination Date
  \item RUNDATE D  Run date
  \item ABETA  pg/mL ABETA
  \item TAU  pg/mL TAU
  \item PTAU  pg/mL PTAU
  \item ABETA.RAW  pg/mL ABETA
  \item TAU.RAW  pg/mL TAU
  \item PTAU.RAW  pg/mL PTAU
}
}
\examples{
\dontrun{
describe(upennbiomk7)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UPENNBIOMK7/ADNI\%20LT2014\%20analytical\%20signed.pdf}, \url{https://adni.bitbucket.io/reference/docs/UPENNBIOMK7/ADNI_Methods_Template_Shaw\%20Trojanowski\%20for\%202014\%20ADNI\%20II\%206\%209\%202014.pdf}
}
\keyword{datasets}
