% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{upennbiomk8}
\alias{upennbiomk8}
\title{UPENN ADNI2 Fourth Batch Analyses of CSF Biomarkers}
\format{
A data frame with 230 rows and 11 variables
}
\usage{
data(upennbiomk8)
}
\description{
UPENN ADNI2 Fourth Batch Analyses of CSF Biomarkers
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID   Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE D  Examination Date
  \item RUNDATE D  Run date
  \item ABETA  pg/mL ABETA
  \item TAU  pg/mL TAU
  \item PTAU  pg/mL PTAU
  \item ABETA.RAW  pg/mL ABETA
  \item TAU.RAW  pg/mL TAU
  \item PTAU.RAW  pg/mL PTAU
}
}
\examples{
\dontrun{
describe(upennbiomk8)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UPENNBIOMK8/ADNI_Methods_Template_Shaw\%20Trojanowski\%20for\%202015\%20ADNI\%20II\%204\%2010\%202015.pdf}
}
\keyword{datasets}
