% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{upennbiomk9}
\alias{upennbiomk9}
\title{UPENN CSF Biomarkers Elecsys}
\format{
A data frame with 2401 rows and 14 variables
}
\usage{
data(upennbiomk9)
}
\description{
UPENN CSF Biomarkers Elecsys
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4  Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4  Examination Date
  \item PHASE -4  Study Phase
  \item VID -4  Visit ID
  \item BARCODE -4  Sample Barcode ID
  \item BATCH -4  Analytical batch of results
  \item KIT -4  Reagents lot number
  \item STDS -4  Calibrators and Quality Controls lot number
  \item RUNDATE -4  Date of analytical run
  \item ABETA -4 pg/ml Result
  \item TAU -4 pg/ml Result
  \item PTAU -4 pg/ml Result
}
}
\examples{
\dontrun{
describe(upennbiomk9)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UPENNBIOMK9/ADNI\%20METHODS\%20doc\%20for\%20Roche\%20Elecsys\%20CSF\%20immunoassays\%20vfinal.pdf}
}
\keyword{datasets}
