% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{upennbiomk_adnidian_es_2017}
\alias{upennbiomk_adnidian_es_2017}
\title{UPENN - ADNI/DIAN study CSF Elecsys results}
\format{
A data frame with 422 rows and 12 variables
}
\usage{
data(upennbiomk_adnidian_es_2017)
}
\description{
UPENN - ADNI/DIAN study CSF Elecsys results
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4  Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4  Examination Date
  \item STUDY -4  Identifies ADNI and DIAN samples
  \item RUNDATE -4  Run date
  \item ABETA -4 pg/mL Abeta 1-42
  \item AB40 -4 pg/mL Abeta 1-40
  \item TAU -4 pg/mL TAU
  \item PTAU -4 pg/mL PTAU
  \item A4240 -4  Ratio of Abeta1-42 to Abeta1-40
  \item NOTE -4  Reason for missing result
}
}
\examples{
\dontrun{
describe(upennbiomk_adnidian_es_2017)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UPENNBIOMK_ADNIDIAN_ES_2017/ADNI_Methods_Template_Shaw_2017_2018_Roche_Elecsys_ADNI_CSFs_in_DIAN_ADNI_study_3_29_2018_v3.pdf}
}
\keyword{datasets}
