% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{upennbiomk_master}
\alias{upennbiomk_master}
\title{UPENN CSF Biomarker Master}
\format{
A data frame with 5876 rows and 14 variables
}
\usage{
data(upennbiomk_master)
}
\description{
UPENN CSF Biomarker Master
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4  Participant roster ID
  \item VISCODE -4  Visit code
  \item BATCH -4  Name of LONI table, corresponding to analytical batch of results
  \item KIT -4  Reagents lot number
  \item STDS -4  Calibrators and Quality Controls lot number
  \item DRAWDTE -4  Date of CSF draw
  \item RUNDATE -4  Date of analytical run
  \item ABETA -4 pg/ml Result rescaled to UPENNBIOMK
  \item TAU -4 pg/ml Result rescaled to UPENNBIOMK
  \item PTAU -4 pg/ml Result rescaled to UPENNBIOMK
  \item ABETA_RAW -4 pg/ml Original (raw) result, before rescaling
  \item TAU_RAW -4 pg/ml Original (raw) result, before rescaling
  \item PTAU_RAW -4 pg/ml Original (raw) result, before rescaling
}
}
\examples{
\dontrun{
describe(upennbiomk_master)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UPENNBIOMK_MASTER/ADNI_Methods_Template_Shaw\%20Figurski\%20Waligorska\%20Trojanowski\%20overview\%20for\%20CSF\%20Ab1-42\%20tau\%20and\%20ptau181\%20AlzBio3\%20immunoassay\%20datav3\%20(5).pdf}
}
\keyword{datasets}
