% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{upennmsmsabeta}
\alias{upennmsmsabeta}
\title{2D-UPLC tandem mass spectrometry measurement of ABeta1-42, ABeta1-40 and ABeta1-38 in ADNI1 BASELINE CSF}
\format{
A data frame with 400 rows and 8 variables
}
\usage{
data(upennmsmsabeta)
}
\description{
2D-UPLC tandem mass spectrometry measurement of ABeta1-42, ABeta1-40 and ABeta1-38 in ADNI1 BASELINE CSF
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID NA  Participant roster ID
  \item VISCODE NA  Visit code
  \item DRAWDATE NA  Draw date
  \item RUNDATE NA  Run date
  \item ABETA42 NA pg/mL Amyloid Beta 1-42
  \item ABETA40 NA pg/mL Amyloid Beta 1-40
  \item ABETA38 NA pg/mL Amyloid Beta 1-38
}
}
\examples{
\dontrun{
describe(upennmsmsabeta)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UPENNMSMSABETA/ADNI_Methods_Template_Biomarker\%20Core\%202D\%20UPLC\%20tandem\%20mass\%20spectrometry\%20analyses\%20of\%20ADNI1\%20BASELINE\%20CSF.pdf}
}
\keyword{datasets}
