% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{upennmsmsabeta2}
\alias{upennmsmsabeta2}
\title{-4}
\format{
A data frame with 1445 rows and 9 variables
}
\usage{
data(upennmsmsabeta2)
}
\description{
-4
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4  Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4  Examination Date
  \item VID -4  Visit number from sample tube
  \item RUNDATE -4  Run date
  \item ABETA42 -4 pg/mL Amyloid Beta 1-42
  \item ABETA40 -4 pg/mL Amyloid Beta 1-40
  \item ABETA38 -4 pg/mL Amyloid Beta 1-38
}
}
\examples{
\dontrun{
describe(upennmsmsabeta2)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UPENNMSMSABETA2/Methods_report_ADNI2_GO__2D_UPLC_tandem_mass_spectrometry_analyses_of_ADNI2_GO_BASELINE_CSF_Final_06_11_2018_v3.pdf}
}
\keyword{datasets}
