% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{upennplasma}
\alias{upennplasma}
\title{UPENN Plasma Biomarker Data}
\format{
A data frame with 2454 rows and 5 variables
}
\usage{
data(upennplasma)
}
\description{
UPENN Plasma Biomarker Data
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N NA Participant roster ID
  \item VISCODE T NA Visit code
  \item AB40 N NA Abeta1-40 result in plasma
  \item AB42 N NA Abeta1-42 result in plasma
}
}
\keyword{datasets}
