% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{upennspare_ad}
\alias{upennspare_ad}
\title{Spatial Pattern of Abnormalities for Recognition of Early AD (SPARE-AD). C. Davatzikos. U. Penn.}
\format{
A data frame with 840 rows and 12 variables
}
\usage{
data(upennspare_ad)
}
\description{
Spatial Pattern of Abnormalities for Recognition of Early AD (SPARE-AD). C. Davatzikos. U. Penn.
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item EXAMDATE -4 -4 Examination Date
  \item VERSION -4 -4 Version Date
  \item IMAGE_UID -4 -4 
  \item RUNDATE -4 -4 
  \item STATUS -4 -4 
  \item CBICA_ID -4 -4 Subject ID used at CBICA
  \item DATE -4 -4 
  \item TrueLabel -4 -4 Diagnostic Group
  \item SPARE_AD -4 -4 Spatial Pattern of Abnormalities for Recognition of Early AD (SPARE-AD)
}
}
\examples{
\dontrun{
describe(upennspare_ad)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UPENNSPARE_AD/ADNI1_SPARE_AD.pdf}
}
\keyword{datasets}
