% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{upennspare_mci}
\alias{upennspare_mci}
\title{Spatial Pattern of Abnormalities for Recognition of Early MCI to AD conversion (SPARE-MCI). C. Davatzikos. U. Penn.}
\format{
A data frame with 330 rows and 13 variables
}
\usage{
data(upennspare_mci)
}
\description{
Spatial Pattern of Abnormalities for Recognition of Early MCI to AD conversion (SPARE-MCI). C. Davatzikos. U. Penn.
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item EXAMDATE -4 -4 Examination Date
  \item VERSION -4 -4 Version Date
  \item IMAGE_UID -4 -4 
  \item RUNDATE -4 -4 
  \item STATUS -4 -4 
  \item CBICA_ID -4 -4 Subject ID used at CBICA
  \item DATE -4 -4 
  \item TimeToConversion -4 -4 Time to Conversion from MCI to AD
  \item TrueLabel -4 -4 Diagnostic Group (LS - Long term Stable, SC - Short Converter)
  \item SPARE_MCI -4 -4 Spatial Pattern of Abnormalities for Recognition of Early MCI to AD conversion (SPARE-MCI)
}
}
\examples{
\dontrun{
describe(upennspare_mci)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UPENNSPARE_MCI/ADNI1_SPARE_MCI.pdf}
}
\keyword{datasets}
