% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{uucacir}
\alias{uucacir}
\title{UU PET Analysis (Norman Foster)}
\format{
A data frame with 1822 rows and 15 variables
}
\usage{
data(uucacir)
}
\description{
UU PET Analysis (Norman Foster)
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE D  Examination Date
  \item IMAGEUID T  ImageUID
  \item PETTYPE T  Type of PET scan
  \item REFCTRLPOP T  Control population used for comparison
  \item REFREGION T  Region used for normalization
  \item AVEREF N  Average value of reference region
  \item AVEASSOC N  Average frontal, parietal and temporal cortices, norm to reference region
  \item AVEFRONT N  Average frontal cortex, norm to reference region
  \item X2SDSIGPXL N pixels Number of pixels with Zscores in the range [2, 3) SD
  \item X3SDSIGPXL N pixels Number of pixels with Zscores  in the range  [3, inf) SD
  \item SUMZ2 N  Sum of pixels with Zscores in the range [2, inf) SD
  \item SUMZ3 N  Sum of pixels with Zscores in the range [3, inf) SD
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UUCACIR/ADNI_UUCACIR_Methods_20120305.pdf}, \url{https://adni.bitbucket.io/reference/docs/UUCACIR/UUCACIR\%20-\%20PET\%20Numeric\%20Summaries\%20Methods_08_10_15.pdf}, \url{https://adni.bitbucket.io/reference/docs/UUCACIR/UUCACIR\%20-\%20PET\%20Numeric\%20Summaries\%20Methods-2016-Jul-08.pdf}, \url{https://adni.bitbucket.io/reference/docs/UUCACIR/UUtah_Analysis.pdf}, \url{https://adni.bitbucket.io/reference/docs/UUCACIR/UUtah\%20Description\%20of\%20Data\%20on\%20LONI.pdf}, \url{https://adni.bitbucket.io/reference/docs/UUCACIR/UUtah\%20-\%20PET\%20Numeric\%20Summaries\%20Methods_20130307.pdf}, \url{https://adni.bitbucket.io/reference/docs/UUCACIR/UUtah\%20-\%20PET\%20Numeric\%20Summaries\%20Methods.pdf}
}
\keyword{datasets}
