% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{uwnpsychsum}
\alias{uwnpsychsum}
\title{Crane Lab (UW) Neuropsych Summary Scores}
\format{
A data frame with 11882 rows and 11 variables
}
\usage{
data(uwnpsychsum)
}
\description{
Crane Lab (UW) Neuropsych Summary Scores
}
\details{
A dataset with variables as follows:

\itemize{
  \item COLPROT -4 -4 Study protocol of data collection
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item USERDATE -4 -4 Date record created
  \item EXAMDATE T -4 Examination Date
  \item ADNI_MEM -4 -4 Composite score for memory
  \item ADNI_EF -4 -4 Composite score for executive functioning
  \item ADNI_LAN -4 -4 Composite score for language cognitive items
  \item ADNI_VS -4 -4 Composite score for visuo-spatial skills
  \item ADNI_EF2 -4 -4 Composite score for executive functioning, version 2
  \item PHASE -4 -4 -4
}
}
\examples{
\dontrun{
describe(uwnpsychsum)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UWNPSYCHSUM/ADNI_Methods_UWNPSYCHSUM_May_2018.pdf}, \url{https://adni.bitbucket.io/reference/docs/UWNPSYCHSUM/ADNI_Methods_UWNPSYCHSUM.pdf}, \url{https://adni.bitbucket.io/reference/docs/UWNPSYCHSUM/adni_uwnpsychsum_doc_20200326.pdf}
}
\keyword{datasets}
