% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{uwovent}
\alias{uwovent}
\title{University of Western Ontario (Bartha Lab) Ventricular Volumes}
\format{
A data frame with 2207 rows and 9 variables
}
\usage{
data(uwovent)
}
\description{
University of Western Ontario (Bartha Lab) Ventricular Volumes
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item RECNO N  
  \item EXAMDATE D mmddyyyy Examination Date
  \item IMAGEUID N  
  \item VENTVOL N cm3 Segmented Ventricular Volume
}
}
\keyword{datasets}
