% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{uwplasmaevapogee}
\alias{uwplasmaevapogee}
\title{University of Washington - Plasma EV Apogee}
\format{
A data frame with 300 rows and 11 variables
}
\usage{
data(uwplasmaevapogee)
}
\description{
University of Washington - Plasma EV Apogee
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item EXAMDATE -4 -4 Examination Date
  \item GUSPECID -4 -4 -4
  \item VID -4 -4 UPENN Visit Code
  \item UW_ASSAY_ID -4 -4 University of Washington Sample (Assay) ID
  \item ASSAY_START_DATE -4 -4 Sample (Apogee) assay start date
  \item L1CAM -4 -4 Proportion of L1CAM+ EVs (%)
  \item ABETA40 -4 -4 Proportion of Abeta40+ EVs (%)
  \item ABETA42 -4 -4 Proportion of Abeta42+ EVs (%)
}
}
\examples{
\dontrun{
describe(uwplasmaevapogee)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/UWPLASMAEVAPOGEE/Plasma_EV_Apogee_data_UW_Methods_20200804.pdf}
}
\keyword{datasets}
