% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{violog}
\alias{violog}
\title{Protocol Deviations Log}
\format{
A data frame with 4087 rows and 16 variables
}
\usage{
data(violog)
}
\description{
Protocol Deviations Log
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N NA Participant roster ID
  \item SITEID N NA Site ID
  \item VISCODE T NA Visit code
  \item USERDATE S NA Date record created
  \item USERDATE2 S NA Date record last updated
  \item RECNO N NA 
  \item VIOTYPE N NA Deviation applies to: (check one)
  \item VIODES N NA Please select the most appropriate description: (check one)
  \item VIODESO T NA If Other specify:
  \item INEXNUM T NA If Inclusion/Exclusion Criteria: Item number
  \item IRB N NA Was IRB informed of Protocol Deviation?
  \item IRBDATE D NA If yes, indicate date reported:
  \item RIGHTS N NA Have the rights, safety or well-being of participant been compromised?
  \item EVENTDES T NA Description of Event
}
}
\keyword{datasets}
