% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{vitals}
\alias{vitals}
\title{Vital Signs}
\format{
A data frame with 15133 rows and 19 variables
}
\usage{
data(vitals)
}
\description{
Vital Signs
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID N  Participant roster ID
  \item SITEID N  Site ID
  \item VISCODE T  Visit code
  \item USERDATE S  Date record created
  \item EXAMDATE D  Examination Date
  \item VSWEIGHT N  1a. Weight
  \item VSWTUNIT N  1b. Weight Units
  \item VSHEIGHT N  2a. Height
  \item VSHTUNIT N  2b. Height Units
  \item VSBPSYS N  Systolic - mmHg
  \item VSBPDIA N  Diastolic - mmHg
  \item VSPULSE N  4. Seated Pulse Rate (per minute)
  \item VSRESP N  5. Respirations (per minute)
  \item VSTEMP N  6a. Temperature
  \item VSTMPSRC N  6b. Temperature Source
  \item VSTMPUNT N  6c. Temperature Units
  \item USERDATE2 S  Date record last updated
}
}
\author{
\email{adni-data@googlegroups.com}
}
\keyword{datasets}
