% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{zhang}
\alias{zhang}
\title{CSF Hemoglobin ELISA and PS129 Luminex Assays (Zhang Lab, University of Washington)}
\format{
A data frame with 805 rows and 13 variables
}
\usage{
data(zhang)
}
\description{
CSF Hemoglobin ELISA and PS129 Luminex Assays (Zhang Lab, University of Washington)
}
\details{
A dataset with variables as follows:

\itemize{
  \item RID -4 -4 Participant roster ID
  \item VISCODE -4 -4 Visit code
  \item VISCODE2 -4 -4 Translated visit code
  \item EXAMDATE -4 -4 Examination Date
  \item BOX_POS -4  Aliquot position in shipping box
  \item VOLUME -4  aliquot volume
  \item SHIPMENT -4  Respository shipment identifier
  \item RUNDATE_ASYN -4  Date of alpha-synuclein assay
  \item FINAL_ASYN -4 ng/ml Final reported concentration of alpha-synuclein
  \item RUNDATE_PS129 -4  Date of ps129 assay
  \item FINAL_PS129 -4 ng/ml Final reported concentration of ps129
  \item RUNDATE_HGB -4  Date of hemoglobin assay
  \item FINAL_HGB -4 ng/ml Final reported concentration of hemoglobin
}
}
\examples{
\dontrun{
describe(zhang)
}
}
\seealso{
\url{https://adni.bitbucket.io/reference/docs/ZHANG/UW\%20-\%20Zhang\%20Lab,\%20CSF\%20Hgb_v160914.pdf}, \url{https://adni.bitbucket.io/reference/docs/ZHANG/UW\%20-\%20Zhang\%20Lab,\%20CSF\%20ps129_v160914.pdf}
}
\keyword{datasets}
